% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-methods.R
\name{aggregate}
\alias{aggregate}
\alias{aggregate,Matrix-method}
\title{Aggregate}
\usage{
aggregate(x, ...)

\S4method{aggregate}{Matrix}(x, by, fun = c("sum", "mean", "n"))
}
\arguments{
\item{x}{Object.}

\item{by}{\code{factor}.
Aggregation groupings. The new aggregate names are defined as the \code{factor}
\link[base:levels]{levels}, and the original, unaggregated names are defined as
the \link[base:names]{names}.}

\item{fun}{\code{character(1)}.
Name of the aggregation function to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Aggregate
}
\note{
Updated 2020-01-30.
}
\examples{
## Example data ====
counts <- matrix(
    data = c(
        0L, 1L, 1L, 1L,
        1L, 0L, 1L, 1L,
        1L, 1L, 0L, 1L,
        1L, 1L, 1L, 0L
    ),
    nrow = 4L,
    ncol = 4L,
    byrow = TRUE,
    dimnames = list(
        paste0("transcript", seq_len(4L)),
        paste(
            paste0("sample", rep(seq_len(2L), each = 2L)),
            paste0("replicate", rep(seq_len(2L), times = 2L)),
            sep = "_"
        )
    )
)
class(counts)
print(counts)

genes <- factor(paste0("gene", rep(seq_len(2L), each = 2L)))
names(genes) <- rownames(counts)
print(genes)

samples <- factor(paste0("sample", rep(seq_len(2L), each = 2L)))
names(samples) <- colnames(counts)
print(samples)

## Matrix
sparse <- as(counts, "sparseMatrix")
class(sparse)
print(sparse)

## SummarizedExperiment
se <- SummarizedExperiment::SummarizedExperiment(
    assays = SimpleList(counts = counts),
    colData = DataFrame(
        sampleName = as.factor(names(samples)),
        aggregate = samples
    ),
    rowData = DataFrame(aggregate = genes)
)
print(se)

## aggregateRows ====
aggregateRows(counts, groupings = genes)
aggregateRows(sparse, groupings = genes)
aggregateRows(se)

## aggregateCols ====
aggregateCols(counts, groupings = samples)
aggregateCols(sparse, groupings = samples)
aggregateCols(se)
}
\seealso{
\itemize{
\item \code{stats::aggregate()}.
\item \code{S4Vectors::aggregate()}.
\item \code{Matrix.utils::aggregate.Matrix()} (defunct).
\item \code{muscat::aggregateData()}.
}
}
