% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregate-methods.R
\docType{methods}
\name{aggregateCellsToSamples}
\alias{aggregateCellsToSamples}
\alias{aggregateCellsToSamples,SingleCellExperiment-method}
\title{Aggregate cells to samples}
\usage{
aggregateCellsToSamples(object, ...)

\S4method{aggregateCellsToSamples}{SingleCellExperiment}(object,
  fun = c("sum", "mean"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{fun}{\code{character(1)}.
Name of the aggregation function. Uses \code{match.arg}.}
}
\value{
\code{SummarizedExperiment}. Object with cell-level counts aggregated
to sample-level.
}
\description{
Utilty function that factilites cell-to-sample aggregation. By default, this
function will sum the counts across cells to sample level.
}
\details{
This function is intended primarily for quality control analysis.

Internally it automatically obtains the cell-to-sample groupings and then
performs aggregation with the \code{aggregateCols} function.
}
\examples{
data(sce)
x <- aggregateCellsToSamples(sce)
print(x)
}
