% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCols-methods.R
\name{aggregateCols}
\alias{aggregateCols}
\alias{aggregateCols,matrix-method}
\alias{aggregateCols,Matrix-method}
\alias{aggregateCols,SummarizedExperiment-method}
\alias{aggregateCols,SingleCellExperiment-method}
\title{Aggregate columns}
\usage{
aggregateCols(x, ...)

\S4method{aggregateCols}{matrix}(x, by, fun = c("sum", "mean", "median", "geometricMean"))

\S4method{aggregateCols}{Matrix}(x, by, fun = c("sum", "mean"))

\S4method{aggregateCols}{SummarizedExperiment}(x, col = "aggregate", fun = "sum")

\S4method{aggregateCols}{SingleCellExperiment}(x, fun)
}
\arguments{
\item{x}{Object.}
}
\value{
Modified object of same class, with columns aggregated.
}
\description{
Aggregate columns
}
\examples{
class(aggregateCols)
}
