% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateFeatures.R
\docType{methods}
\name{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateFeatures,dgCMatrix-method}
\alias{aggregateFeatures,dgTMatrix-method}
\title{Aggregate Features}
\usage{
\S4method{aggregateFeatures}{dgCMatrix}(object, features)

\S4method{aggregateFeatures}{dgTMatrix}(object, features)
}
\arguments{
\item{object}{Object.}

\item{features}{Feature identifiers (e.g. gene or transcript IDs). These are
the rownames of the counts matrix.}
}
\value{
Object of same class with aggregated features (rows).
}
\description{
Aggregate Features
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Data Management Utilities}
