% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,dgCMatrix-method}
\alias{aggregateReplicates,dgTMatrix-method}
\title{Aggregate Replicates}
\usage{
aggregateReplicates(object, ...)

\S4method{aggregateReplicates}{dgCMatrix}(object, cells)

\S4method{aggregateReplicates}{dgTMatrix}(object, cells)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{cells}{Cellular barcode identifiers. These are the colnames of the
counts matrix.}
}
\value{
Object of same class with aggregated counts per pooled sample
(columns).
}
\description{
Aggregate Replicates
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
