% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateRows-methods.R
\name{aggregateRows}
\alias{aggregateRows}
\alias{aggregateRows,matrix-method}
\alias{aggregateRows,Matrix-method}
\alias{aggregateRows,SummarizedExperiment-method}
\title{Aggregate rows}
\usage{
aggregateRows(x, ...)

\S4method{aggregateRows}{matrix}(x, by, fun = c("sum", "mean", "median", "geometricMean"))

\S4method{aggregateRows}{Matrix}(x, by, fun = c("sum", "mean"))

\S4method{aggregateRows}{SummarizedExperiment}(x, col = "aggregate", fun = "sum")
}
\arguments{
\item{x}{Object.}
}
\value{
Modified object of same class, with rows aggregated.
}
\description{
Aggregate rows
}
\examples{
class(aggregateRows)
}
