% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/load.R, R/math.R, R/names.R,
%   R/project.R, R/read.R, R/reexports.R, R/rmarkdown.R, R/save.R,
%   R/tidyverse.R, R/toString.R, R/utils.R
\name{aliases}
\alias{aliases}
\alias{load_data_raw}
\alias{load_remote}
\alias{log_ratio_to_fold_change}
\alias{dot_notation}
\alias{sanitizeNames}
\alias{sanitize_names}
\alias{sanitiseNames}
\alias{sanitise_names}
\alias{title_case}
\alias{first_case}
\alias{create_project_dirs}
\alias{detect_hpc}
\alias{package_project}
\alias{render_project}
\alias{clear_warnings}
\alias{read_data_raw}
\alias{read_yaml}
\alias{setColnames}
\alias{setRownames}
\alias{md_list}
\alias{save_data}
\alias{save_data_raw}
\alias{import_tidy_verbs}
\alias{to_string_unique}
\alias{to_string_sort_unique}
\alias{to_string_summarize}
\alias{toStringSummarise}
\alias{to_string_summarise}
\alias{assign_multiple}
\alias{fix_na}
\alias{grep_string}
\alias{sort_unique}
\title{Function aliases}
\usage{
sanitizeNames(x)
}
\description{
Name variants supporting \code{snake_case} and British English (e.g. \code{\link[=summarise]{summarise()}}
instead of \code{\link[=summarize]{summarize()}}).
}
\examples{
# camelCase (default)
logRatioToFoldChange(4)

# snake_case
log_ratio_to_fold_change(4)
}
\keyword{internal}
