% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{AllClasses}
\alias{AllClasses}
\title{S4 Classes}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending upon the method.
}
\description{
S4 Classes
}
\seealso{
\itemize{
\item \code{\link[methods:getClass]{methods::getClass()}}.
\item \code{\link[methods:isClass]{methods::isClass()}}.
}
}
\keyword{internal}
