% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\title{S4 Generics}
\arguments{
\item{object}{Object.}

\item{x}{Primary object.}

\item{y}{Secondary object.}

\item{value}{Value to assign.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\description{
S4 Generics
}
\keyword{internal}
