% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-methods.R
\docType{methods}
\name{alpha}
\alias{alpha}
\alias{alpha<-}
\alias{alpha,Annotated-method}
\alias{alpha<-,Annotated,numeric-method}
\title{Alpha level}
\usage{
alpha(object, ...)

alpha(object, ...) <- value

\S4method{alpha}{Annotated}(object)

\S4method{alpha}{Annotated,numeric}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value.}

\item{...}{Additional arguments.}
}
\description{
Alpha level
}
\examples{
data(rse, package = "acidtest")

## Annotated
alpha(rse) <- 0.05
alpha(rse)
}
