% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotables.R
\name{annotable}
\alias{annotable}
\title{\href{http://www.ensembl.org/}{Ensembl} annotations}
\usage{
annotable(genomeBuild, format = "gene")
}
\arguments{
\item{genomeBuild}{Genome build. Consult the \link{annotables} documentation
for a list of currently supported genomes.}

\item{format}{Desired table format, either \code{gene}, \code{tx2gene}, \code{gene2symbol},
or \code{gene2entrez}.}
}
\value{
\link{data.frame} with unique rows per gene/transcript, or grouped
\link{tibble} for \code{gene2entrez} format argument with non-unique rows.
}
\description{
Quickly access gene annotations and transcript-to-gene (tx2gene) mappings
pre-compiled from \href{http://www.ensembl.org/}{Ensembl} with the
\href{http://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
package. For gene annotables, the Entrez identifier is removed, to allow
for unique Ensembl gene identifiers.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
