% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-annotables.R
\docType{methods}
\name{annotable}
\alias{annotable}
\alias{gene2entrez}
\alias{gene2symbol}
\alias{tx2gene}
\alias{.annotable}
\alias{annotable,character-method}
\alias{gene2entrez,character-method}
\alias{gene2symbol,character-method}
\alias{tx2gene,character-method}
\title{\href{http://www.ensembl.org/}{Ensembl} Annotations}
\usage{
annotable(object)

gene2entrez(object)

gene2symbol(object)

tx2gene(object)

\S4method{annotable}{character}(object)

\S4method{gene2entrez}{character}(object)

\S4method{gene2symbol}{character}(object)

\S4method{tx2gene}{character}(object)
}
\arguments{
\item{object}{Object. Default usage is to provide Ensembl genome build as a
string.}
}
\value{
\link{data.frame} with unique rows per gene or transcript.
}
\description{
Quickly access gene annotations and transcript-to-gene (tx2gene) mappings
pre-compiled from \href{http://www.ensembl.org/}{Ensembl} with the
\href{http://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
package. For gene annotables, the Entrez identifier is removed, to allow
for unique Ensembl gene identifiers.
}
\note{
If the \code{format} argument is set to \code{gene2entrez}, \code{\link[=annotable]{annotable()}} returns
a \link{tibble} with non-unique rows grouped by \code{ensgene}, instead of a
\link{data.frame}.
}
\examples{
annotable("hg38") \%>\% glimpse
detectOrganism("ENSMUSG00000000001") \%>\% gene2symbol \%>\% head
}
\seealso{
Consult the annotables package documentation (\code{help("annotables")})
for a list of currently supported genomes.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
