% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antilog2fc.R
\name{antilog2fc}
\alias{antilog2fc}
\title{Log base2 fold change antilog conversion}
\usage{
antilog2fc(log2fc)
}
\arguments{
\item{log2fc}{Numeric vector of log normalized fold change values.}
}
\value{
Antilog numeric vector of fold change values with proper sign.
}
\description{
Convert a string of log2 normalized fold change values for better
readability.
}

