% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-methods.R
\docType{methods}
\name{as}
\alias{as}
\alias{coerce}
\alias{as.data.frame,sparseMatrix-method}
\alias{coerce,sparseMatrix,data.frame-method}
\alias{coerce,sparseMatrix,DataFrame-method}
\alias{coerce,tbl_df,DataFrame-method}
\alias{coerce,data.frame,tbl_df-method}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GRanges,tbl_df-method}
\title{Force an Object to Belong to a Class}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Object of new class.
}
\description{
Coerce an object to a given class.
}
\section{list}{


It is often useful to coerce an S4 object to a flat \code{list} for archival
storage. Here we are providing the \code{coerceS4ToList()} function, which
consistently coerces the slots in any S4 to a standard \code{list}. Additionally,
here we have improved support for \code{SummarizedExperiment} to \code{list} coercion,
returning the slots as a \code{list}.
}

\section{tibble}{


Coerce an object to a tibble (\code{tbl_df}, \code{data.frame}) using either S3 or S4
methods:
\itemize{
\item S3: \code{as_tibble(x)} (or \code{as_tibble()} alias).
\item S4: \code{as(object, Class = "tbl_df")}.
}

Tibbles don't support row name assignment, so here we are ensuring they are
kept by moving them to a column named \code{rowname} by default. This helps avoid
downstream unexpected data loss when using the dplyr chain of single table
verbs, such as \code{dplyr::arrange()}, \code{dplyr::filter()}, or \code{dplyr::mutate()}.

This behavior can be overriden in the S3 method by setting \code{rowname = NULL}
instead, which is the current default in the tibble package. The S4 coercion
method doesn't support arguments, and therefore always attempts to move
rownames automatically, if defined.

Conversely, when coercing a \code{tibble} back to an S4 \code{DataFrame}, our
\code{as(tbl_df, Class = "DataFrame")} method looks for the \code{rowname} column and
will attempt to move it back to \code{base::rownames()} automatically, unless
there are duplicates present.
}

\examples{
data(rse)

## DataFrame to tbl_df ====
data <- SummarizedExperiment::colData(rse)
print(data)

x <- as(data, "tbl_df")
print(x)

x <- tibble::as_tibble(data)
print(x)

## GRanges to tbl_df ====
data <- SummarizedExperiment::rowRanges(rse)

x <- as(data, "tbl_df")

x <- tibble::as_tibble(data)
colnames(x)

## sparseMatrix to data.frame ====
data(sparse)

x <- BiocGenerics::as.data.frame(sparse)
class(x)

x <- as(sparse, "data.frame")
class(x)

## tbl_df to DataFrame ====
data <- tibble::as_tibble(datasets::iris)
x <- as(data, "DataFrame")
}
\references{
Chambers, John M. (2016)
 \emph{Extending R},
  Chapman & Hall.
(Chapters 9 and 10.)
}
\seealso{
\itemize{
\item \code{methods::as()}.
\item \code{methods::canCoerce()}.
\item \code{methods::showMethods()}.
\item \code{tibble::tibble()}.
\item \code{utils::methods()}.
}
}
