% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert Checks}
\value{
Abort on mismatch.
}
\description{
Assert Checks
}
\seealso{
Other Assert Checks: \code{\link{assert_formal_annotation_col}},
  \code{\link{assert_formal_color_function}},
  \code{\link{assert_formal_compress}},
  \code{\link{assert_formal_gene2symbol}},
  \code{\link{assert_formal_header_level}},
  \code{\link{assert_has_rownames}},
  \code{\link{assert_is_a_number_or_null}},
  \code{\link{assert_is_a_string_or_null}},
  \code{\link{assert_is_annotable}},
  \code{\link{assert_is_character_or_null}},
  \code{\link{assert_is_gene2symbol}},
  \code{\link{assert_is_implicit integer}},
  \code{\link{assert_is_tx2gene}}
}
\keyword{internal}
