% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_all_are_non_existing}
\alias{assert_all_are_non_existing}
\title{All Variables Do Not Exist Assert Check}
\usage{
assert_all_are_non_existing(x, envir = parent.frame(), inherits = FALSE)
}
\arguments{
\item{x}{Character vector of variable names to check in environment.}

\item{envir}{Environment.}

\item{inherits}{Should the enclosing frames of the environment be searched?}
}
\value{
Abort on mismatch.
}
\description{
All Variables Do Not Exist Assert Check
}
\examples{
a <- 1L
b <- 2L
assert_all_are_non_existing(c("a", "b", "c"))
}
\seealso{
Other Assert Checks: \code{\link{assert_formal_annotation_col}},
  \code{\link{assert_formal_color_function}},
  \code{\link{assert_formal_compress}},
  \code{\link{assert_formal_gene2symbol}},
  \code{\link{assert_formal_header_level}},
  \code{\link{assert_has_rownames}},
  \code{\link{assert_is_a_number_or_null}},
  \code{\link{assert_is_a_string_or_null}},
  \code{\link{assert_is_an_implicit_integer}},
  \code{\link{assert_is_an_integer_or_null}},
  \code{\link{assert_is_annotable}},
  \code{\link{assert_is_character_or_null}},
  \code{\link{assert_is_data.frame_or_null}},
  \code{\link{assert_is_gene2symbol}},
  \code{\link{assert_is_tx2gene}}, \code{\link{assert}}
}
\keyword{internal}
