% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_formal_gene2symbol}
\alias{assert_formal_gene2symbol}
\title{Gene to Symbol Mappings Formal Assert Check}
\usage{
assert_formal_gene2symbol(x, genes, gene2symbol, severity = "stop")
}
\arguments{
\item{x}{Object containing gene identifiers in the rownames.}

\item{genes}{Character vector of gene identifiers.}

\item{gene2symbol}{Gene-to-symbol mappings data frame.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Gene to Symbol Mappings Formal Assert Check
}
\seealso{
Other Assert Checks: \code{\link{assert_all_are_non_existing}},
  \code{\link{assert_formal_annotation_col}},
  \code{\link{assert_formal_color_function}},
  \code{\link{assert_formal_compress}},
  \code{\link{assert_formal_header_level}},
  \code{\link{assert_has_rownames}},
  \code{\link{assert_is_a_number_or_null}},
  \code{\link{assert_is_a_string_or_null}},
  \code{\link{assert_is_an_implicit_integer}},
  \code{\link{assert_is_an_integer_or_null}},
  \code{\link{assert_is_annotable}},
  \code{\link{assert_is_character_or_null}},
  \code{\link{assert_is_data.frame_or_null}},
  \code{\link{assert_is_gene2symbol}},
  \code{\link{assert_is_tx2gene}}, \code{\link{assert}}
}
\keyword{internal}
