% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_formal_header_level.R
\name{assert_formal_header_level}
\alias{assert_formal_header_level}
\title{Markdown Header Level Assert Check}
\usage{
assert_formal_header_level(x)
}
\arguments{
\item{x}{Primary object.}
}
\value{
Abort on mismatch.
}
\description{
Markdown Header Level Assert Check
}
\seealso{
Other Assert Checks: \code{\link{assert_formal_annotation_col}},
  \code{\link{assert_formal_color_function}},
  \code{\link{assert_formal_compress}},
  \code{\link{assert_formal_gene2symbol}},
  \code{\link{assert_has_rownames}},
  \code{\link{assert_is_a_string_or_null}},
  \code{\link{assert_is_annotable}},
  \code{\link{assert_is_character_or_null}},
  \code{\link{assert_is_gene2symbol}},
  \code{\link{assert_is_implicit_integer}},
  \code{\link{assert_is_numeric_scalar_or_null}},
  \code{\link{assert_is_tx2gene}}, \code{\link{assert}}
}
