% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_is_an_integer}
\alias{assert_is_an_integer}
\title{Is an Integer Assert Check}
\usage{
assert_is_an_integer(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Is an Integer Assert Check
}
\examples{
# Explicit integer
assert_is_an_integer(1L)

# Implicit integer
assert_is_an_integer(1, severity = "message")
}
\seealso{
Other Assert Checks: \code{\link{assert_formal_annotation_col}},
  \code{\link{assert_formal_color_function}},
  \code{\link{assert_formal_compress}},
  \code{\link{assert_formal_gene2symbol}},
  \code{\link{assert_formal_header_level}},
  \code{\link{assert_has_rownames}},
  \code{\link{assert_is_a_number_or_null}},
  \code{\link{assert_is_a_string_or_null}},
  \code{\link{assert_is_an_implicit_integer}},
  \code{\link{assert_is_an_integer_or_null}},
  \code{\link{assert_is_annotable}},
  \code{\link{assert_is_character_or_null}},
  \code{\link{assert_is_gene2symbol}},
  \code{\link{assert_is_tx2gene}}, \code{\link{assert}}
}
