% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_is_annotable.R
\name{assert_is_annotable}
\alias{assert_is_annotable}
\title{Ensembl Annotations Assert Check}
\usage{
assert_is_annotable(x)
}
\arguments{
\item{x}{\link{data.frame} containing Ensembl gene annotations returned
from the \code{\link[=annotable]{annotable()}} function.}
}
\value{
Error on mismatch.
}
\description{
Ensembl Annotations Assert Check
}
\examples{
annotable <- annotable("Homo sapiens")
assert_is_annotable(annotable)
}
