% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_is_implicit_integer.R
\name{assert_is_implicit_integer}
\alias{assert_is_implicit_integer}
\alias{is_implicit_integer}
\title{Implict Integer Assert Check}
\usage{
assert_is_implicit_integer(x)

is_implicit_integer(x)
}
\arguments{
\item{x}{Primary object.}
}
\value{
Abort on mismatch.
}
\description{
Implict Integer Assert Check
}
\seealso{
\code{\link[base:all.equal]{base::all.equal()}} \code{tolerance} parameter.

Other Assert Checks: \code{\link{assert_formal_annotation_col}},
  \code{\link{assert_formal_color_function}},
  \code{\link{assert_formal_compress}},
  \code{\link{assert_formal_gene2symbol}},
  \code{\link{assert_formal_header_level}},
  \code{\link{assert_has_rownames}},
  \code{\link{assert_is_a_string_or_null}},
  \code{\link{assert_is_annotable}},
  \code{\link{assert_is_character_or_null}},
  \code{\link{assert_is_gene2symbol}},
  \code{\link{assert_is_numeric_scalar_or_null}},
  \code{\link{assert_is_tx2gene}}, \code{\link{assert}}
}
