% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_is_tx2gene.R
\name{assert_is_tx2gene}
\alias{assert_is_tx2gene}
\title{Check Transcript to Gene Mapping Data}
\usage{
assert_is_tx2gene(object)
}
\arguments{
\item{object}{\link{data.frame} containing Ensembl transcript to gene identifier
mappings. Must be structured as a two column \link{data.frame} with "enstxp" and
"ensgene" columns.}
}
\value{
Invisible \code{TRUE} on pass, \code{\link[=stop]{stop()}} on error.
}
\description{
Check Transcript to Gene Mapping Data
}
\examples{
tx2gene <- annotable("Homo sapiens", format = "tx2gene")
assert_is_tx2gene(tx2gene)
}
