% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertAllAreNonExisting}
\alias{assertAllAreNonExisting}
\title{Assert All Variables Are Non-Existing}
\usage{
assertAllAreNonExisting(names, envir = parent.frame(),
  inherits = FALSE)
}
\arguments{
\item{names}{\code{character}. Names (i.e. corresponding to rows or columns).
See \code{\link[base:make.names]{base::make.names()}} for more information on valid names.}

\item{envir}{\code{environment} to use for assignment. Defaults to
\code{parent.frame()}, which will assign into the calling environment.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}
}
\value{
Always stop on error.
}
\description{
Assert All Variables Are Non-Existing
}
\examples{
assertAllAreNonExisting(c("XXX", "YYY"))
}
