% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertAllAreUniqueGeneNames}
\alias{assertAllAreUniqueGeneNames}
\title{Assert All Are Unique Gene Names}
\usage{
assertAllAreUniqueGeneNames(object, genes)
}
\arguments{
\item{object}{\code{SummarizedExperiment}.}

\item{genes}{\code{character}. Input vector to check against definitions in the
corresponding \code{SummarizedExperiment}.}
}
\value{
Always stop on error.
}
\description{
This assert check determines if a user-defined gene name query is using only
unique (non-amgibuous) symbols. It is designed to be used for gene plotting
particularly when performing single-cell RNA-seq marker analysis.
}
\examples{
data(rse)

object <- rse
print(object)

genes <- object \%>\%
    SummarizedExperiment::rowData(.) \%>\%
    .[["geneName"]] \%>\%
    as.character() \%>\%
    head()
print(genes)

assertAllAreUniqueGeneNames(object = object, genes = genes)
}
