% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertAreGeneAnnotations}
\alias{assertAreGeneAnnotations}
\title{Assert Are Gene Annotations}
\usage{
assertAreGeneAnnotations(object)
}
\arguments{
\item{object}{Object that can be coerced to \code{data.frame}.}
}
\value{
Always stop on error.
}
\description{
Must contain \code{geneID} and \code{geneName} columns. Does not need to contain
rownames, so \code{tibble} class is supported.
}
\examples{
object <- tibble(
    geneID = "ENSG00000000003",
    geneName = "TSPAN6"
)
assertAreGeneAnnotations(object)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2Symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDir}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
