% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertAreValidNames}
\alias{assertAreValidNames}
\alias{validNames}
\alias{assertHasValidNames}
\alias{validDimnames}
\alias{assertHasValidDimnames}
\title{Assert Are Valid Names}
\usage{
validNames(names)

assertAreValidNames(names)

assertHasValidNames(object)

validDimnames(object)

assertHasValidDimnames(object)
}
\arguments{
\item{names}{\code{character}. Names (i.e. corresponding to rows or columns).
See \code{\link[base:make.names]{base::make.names()}} for more information on valid names.}

\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Are Valid Names
}
\examples{
data(rse)
gr <- SummarizedExperiment::rowRanges(rse)

# Dots (periods) and underscores are valid.
validNames(c("sample.1", "sample_1"))

# Can't begin with a number.
validNames("293cells")

# Spaces, dashes (hyphens), and other non-alphanumerics aren't valid.
validNames("sample 1")
validNames("cell-AAAAAAAA")
validNames("GFP+")

assertAreValidNames(c("sample1", "sample2"))
assertHasValidNames(gr)

# Check rows and columns (dimnames) in a single call.
validDimnames(datasets::iris)    # TRUE
validDimnames(datasets::mtcars)  # FALSE
assertHasValidDimnames(rse)
}
