% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertFormalCompress}
\alias{assertFormalCompress}
\title{Check Compression Formal}
\usage{
assertFormalCompress(x)
}
\arguments{
\item{x}{Object.}
}
\value{
No value.
}
\description{
Check Compression Formal
}
\examples{
assertFormalCompress("xz")
}
