% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertFormalInterestingGroups}
\alias{assertFormalInterestingGroups}
\title{Interesting Groups Formal Assert Check}
\usage{
assertFormalInterestingGroups(object, interestingGroups)
}
\arguments{
\item{object}{S4 class object.}

\item{interestingGroups}{\code{character}. Interesting groups.}
}
\value{
Always stop on error.
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
data(rse)
assertFormalInterestingGroups(rse, "treatment")
assertFormalInterestingGroups(rse, NULL)
}
