% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertHasAggregateInfo}
\alias{assertHasAggregateInfo}
\alias{hasAggregateInfo}
\title{Assert Has Aggregate Information}
\usage{
assertHasAggregateInfo(object)

hasAggregateInfo(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Determine whether an object can be aggregated automatically.
}
