% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertHasRownames.R
\name{assertHasRownames}
\alias{assertHasRownames}
\alias{hasRownames}
\title{Has Rownames Assert Check}
\usage{
assertHasRownames(x, severity = "stop")

hasRownames(x)
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
A stricter alternative to the assertive version that works properply with
data frames.
}
\note{
\code{tibble::has_rownames()} works better than \code{assertive::has_rownames()}
for data frames and tibbles.
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
