% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertions}
\alias{assertions}
\title{Assertions}
\arguments{
\item{names}{\code{character}. Names (i.e. corresponding to rows or columns).
See \code{\link[base:make.names]{base::make.names()}} for more information on valid names.}

\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assertions
}
\keyword{internal}
