% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsAHeaderLevel}
\alias{assertIsAHeaderLevel}
\title{Assert Is a Header Level}
\usage{
assertIsAHeaderLevel(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Stop on mismatch.
}
\description{
Markdown supports header levels 1-7 (\code{<H1>}-\code{<H7>}).
}
\examples{
assertIsAHeaderLevel(1L)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
