% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsAlpha}
\alias{assertIsAlpha}
\title{Assert Is Alpha Level}
\usage{
assertIsAlpha(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
An alpha level must be between 0 and 1, but not equal either 0 or 1.
}
\examples{
assertIsAlpha(0.05)
}
\author{
Michael Steinbaugh
}
