% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsAnIntegerOrNULL.R
\name{assertIsAnIntegerOrNULL}
\alias{assertIsAnIntegerOrNULL}
\title{Is an Integer or NULL Assert Check}
\usage{
assertIsAnIntegerOrNULL(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Is an Integer or NULL Assert Check
}
\examples{
# Success
assertIsAnIntegerOrNULL(1L)
assertIsAnIntegerOrNULL(NULL)

# Failure
tryCatch(
    assertIsAnIntegerOrNULL(c(1L, 2L)),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
