% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsColorScaleContinuousOrNULL.R
\name{assertIsColorScaleContinuousOrNULL}
\alias{assertIsColorScaleContinuousOrNULL}
\alias{assert_is_color_scale_continuous_or_null}
\title{Is Color Palette Scale Continuous or NULL Assert Check}
\usage{
assertIsColorScaleContinuousOrNULL(x, severity = "stop")

assert_is_color_scale_continuous_or_null(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Is Color Palette Scale Continuous or NULL Assert Check
}
\examples{
# Success
color <- scale_color_viridis(discrete = TRUE)
class(color)
assertIsColorScaleContinuousOrNULL(color)
assertIsColorScaleContinuousOrNULL(NULL)

# Failure
color <- scale_color_viridis(discrete = FALSE)
class(color)
tryCatch(
    assertIsColorScaleContinuousOrNULL(color),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
