% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsColorScaleDiscreteOrNULL}
\alias{assertIsColorScaleDiscreteOrNULL}
\title{Assert Is Color Palette Scale Discrete or NULL}
\usage{
assertIsColorScaleDiscreteOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Color Palette Scale Discrete or NULL
}
\examples{
color <- ggplot2::scale_color_manual(values = c("red", "blue"))
class(color)
assertIsColorScaleDiscreteOrNULL(color)
assertIsColorScaleDiscreteOrNULL(NULL)
}
\author{
Michael Steinbaugh
}
