% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsDataFrameOrNULL.R
\name{assertIsDataFrameOrNULL}
\alias{assertIsDataFrameOrNULL}
\alias{assert_is_data.frame_or_null}
\title{Data Frame or NULL Assert Check}
\usage{
assertIsDataFrameOrNULL(x, severity = "stop")

assert_is_data.frame_or_null(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Data Frame or NULL Assert Check
}
\examples{
# Success
assertIsDataFrameOrNULL(mtcars)
assertIsDataFrameOrNULL(NULL)

# Failure
tryCatch(
    assertIsDataFrameOrNULL(1L),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
