% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsDir}
\alias{assertIsDir}
\title{Assert Is Directory}
\usage{
assertIsDir(dir)
}
\arguments{
\item{dir}{\code{string}. Directory path.}
}
\value{
Always stop on error.
}
\description{
Assert Is Directory
}
\examples{
assertIsDir("~")
}
