% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsGene2symbol.R
\name{assertIsGene2symbol}
\alias{assertIsGene2symbol}
\title{Assert Is Gene to Symbol Mapping Data Frame}
\usage{
assertIsGene2symbol(x, severity = getOption("assertive.severity", "stop"))
}
\arguments{
\item{x}{\code{data.frame} containing Ensembl gene-to-symbol mappings. Must be
structured as a two column \code{data.frame} with "geneID" and "geneName"
columns.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Assert Is Gene to Symbol Mapping Data Frame
}
\examples{
gene2symbol <- gene2symbol("Homo sapiens", genomeBuild = "GRCh37")
assertIsGene2symbol(gene2symbol)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
