% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsHeaderLevel}
\alias{assertIsHeaderLevel}
\title{Assert Is Markdown Header Level}
\usage{
assertIsHeaderLevel(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Markdown supports header levels 1-7 (\code{<H1>}-\code{<H7>}).
}
\examples{
assertIsHeaderLevel(1L)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2Symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDir}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
