% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsHexColorFunctionOrNULL.R
\name{assertIsHexColorFunctionOrNULL}
\alias{assertIsHexColorFunctionOrNULL}
\alias{assert_is_hex_color_function_or_null}
\title{Is Hex Color Function or NULL Assert Check}
\usage{
assertIsHexColorFunctionOrNULL(x, severity = "stop")

assert_is_hex_color_function_or_null(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Is Hex Color Function or NULL Assert Check
}
\examples{
assertIsHexColorFunctionOrNULL(viridis)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
