% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsImplicitInteger}
\alias{assertIsImplicitInteger}
\alias{isImplicitInteger}
\alias{assertIsImplicitIntegerOrNULL}
\alias{isAnImplicitInteger}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\title{Assert Is Implicit Integer}
\usage{
isImplicitInteger(object)

assertIsImplicitInteger(object)

assertIsImplicitIntegerOrNULL(object)

isAnImplicitInteger(object)

assertIsAnImplicitInteger(object)

assertIsAnImplicitIntegerOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Implicit Integer
}
\examples{
isImplicitInteger(list(1, 1L, 1.1, "XXX"))
assertIsImplicitInteger(c(1, 2))
assertIsImplicitIntegerOrNULL(c(1, 2))
assertIsImplicitIntegerOrNULL(NULL)

isAnImplicitInteger(1)
assertIsAnImplicitInteger(1)
assertIsAnImplicitIntegerOrNULL(1)
assertIsAnImplicitIntegerOrNULL(NULL)
}
