% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsTx2gene}
\alias{assertIsTx2gene}
\title{Assert Is Transcript-to-Gene Mapping Data}
\usage{
assertIsTx2gene(object)
}
\arguments{
\item{object}{\code{DataFrame} or \code{tbl_df} containing transcript-to-gene
identifier mappings. Must contain the columns "\code{transcriptID}" and
"\code{geneID}". If \code{DataFrame}, must also have \code{rownames} set.}
}
\value{
Always stop on error.
}
\description{
Assert Is Transcript-to-Gene Mapping Data
}
\examples{
# DataFrame ====
object <- DataFrame(
    transcriptID = "ENST00000000233",
    geneID = "ENSG00000004059",
    row.names = "ENST00000000233"
)

# tibble ====
object <- tibble(
    transcriptID = "ENST00000000233",
    geneID = "ENSG00000004059"
)
assertIsTx2gene(object)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDir}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
