% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign_data}
\alias{assign_data}
\title{Assign and save data}
\usage{
assign_data(name, object, compress = "xz")
}
\arguments{
\item{name}{Desired variable name.}

\item{object}{Object.}

\item{compress}{File compression argument from \code{\link[=save]{save()}}.}
}
\description{
Assign and save data
}
