% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-assignAndSaveData.R
\docType{methods}
\name{assignAndSaveData}
\alias{assignAndSaveData}
\alias{assignAndSaveData,character-method}
\title{Assign and Save Data}
\usage{
assignAndSaveData(name, object, ...)

\S4method{assignAndSaveData}{character}(name, object, dir = "data",
  compress = TRUE)
}
\arguments{
\item{name}{Desired variable name.}

\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Assign and Save Data
}
\details{
\itemize{
\item \code{\link[=assignAndSaveData]{assignAndSaveData()}}: Assigns a new object by name to the
current working environment then saves the newly assigned object, specified
by \code{dir}.
}
}
