% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assignData}
\alias{assignData}
\title{Assign data}
\usage{
assignData(name, object, dir = "data", compress = "xz")
}
\arguments{
\item{name}{Desired variable name.}

\item{object}{Object.}

\item{dir}{Output directory.}

\item{compress}{File compression argument from \code{\link[=save]{save()}}.}
}
\description{
Also automatically saves newly assigned object to \code{data} directory.
}
