% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/assignIntoNewEnv-methods.R
\docType{methods}
\name{assignIntoNewEnv}
\alias{assignIntoNewEnv}
\alias{assignIntoNewEnv,ANY-method}
\title{Assign Multiple Objects to a New Environment}
\usage{
assignIntoNewEnv(..., envName, parentEnv = parent.frame(99L))

\S4method{assignIntoNewEnv}{ANY}(..., envName, parentEnv = parent.frame(99L))
}
\arguments{
\item{...}{Objects as dots.}

\item{envName}{New environment name.}

\item{parentEnv}{Parent environment of new environment.}
}
\value{
No value.
}
\description{
Assign Multiple Objects to a New Environment
}
\examples{
assignIntoNewEnv(mtcars, starwars, envName = "test")
}
\seealso{
\code{\link[Biobase:multiassign]{Biobase::multiassign()}}.

Other Assign Utilities: \code{\link{assignAndSaveData}}
}
