% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/autopadZeros-methods.R
\docType{methods}
\name{autopadZeros}
\alias{autopadZeros}
\alias{autopadZeros,character-method}
\alias{autopadZeros,matrix-method}
\alias{autopadZeros,SummarizedExperiment-method}
\title{Autopad Zeros in Vector}
\usage{
autopadZeros(object, ...)

\S4method{autopadZeros}{character}(object)

\S4method{autopadZeros}{matrix}(object, rownames = TRUE,
  colnames = TRUE)

\S4method{autopadZeros}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Autopad Zeros in Vector
}
\examples{
data(rse)

## character ====
autopadZeros(c("A1", "B10"))
autopadZeros(c("A1", "B10", "C100"))

## SummarizedExperiment ====
autopadZeros(rse)
}
