% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\title{Barcode ranks per sample}
\usage{
barcodeRanksPerSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
}
\description{
Barcode ranks per sample
}
\examples{
isS4(barcodeRanksPerSample)
}
\seealso{
\itemize{
\item \href{https://bioconductor.org/packages/DropletUtils/}{DropletUtils}.
\item \href{https://bioconductor.org/packages/SingleCellExperiment/}{SingleCellExperiment}.
\item \href{http://bioinformatics.sph.harvard.edu/bcbioSingleCell/}{bcbioSingleCell}.
\item \href{https://github.com/steinbaugh/Chromium/}{Chromium}.
}
}
