% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{basejump-package}
\alias{basejump}
\alias{basejump-package}
\title{basejump}
\description{
Toolkit for bioinformatics and R package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/basejump/}
  \item \url{https://github.com/acidgenomics/r-basejump/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-basejump/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
  \item Rory Kirchner \email{roryk@alum.mit.edu} (\href{https://orcid.org/0000-0003-4814-5885}{ORCID}) [contributor]
  \item Mary Piper \email{mary.piper@gmail.com} (\href{https://orcid.org/0000-0003-2699-3840}{ORCID}) [contributor]
}

}
\keyword{internal}
