% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\docType{package}
\name{basejump}
\alias{basejump}
\alias{basejump-package}
\title{\code{basejump} package}
\description{
Base functions for bioinformatics and R package development.
}
\details{
See the README on \href{https://github.com/steinbaugh/basejump}{GitHub}.
}
