% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{basejumpTestsURL}
\alias{basejumpTestsURL}
\title{acid.test data URL}
\format{An object of class \code{character} of length 1.}
\usage{
basejumpTestsURL
}
\description{
acid.test data URL
}
\examples{
basejumpTestsURL
}
\keyword{internal}
