% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioData.R
\name{bcbioData}
\alias{bcbioData}
\title{Import a bcbio project data file}
\usage{
bcbioData(project, file, input = "tsv", output = "data.frame",
  rownames = NULL)
}
\arguments{
\item{project}{bcbio project list}

\item{file}{File name}

\item{input}{Input format}

\item{output}{Desired output format}

\item{rownames}{Column identifier to use for rownames}
}
\value{
bcbio data
}
\description{
Import a bcbio project data file
}
\examples{
\dontrun{
bcbioData(project,
          file = "combined.counts",
          output = "matrix",
          rownames = "id")
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{import}
