% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioFile.R
\name{bcbioFile}
\alias{bcbioFile}
\title{Import a bcbio project file}
\usage{
bcbioFile(project, file, input = "tsv", output = "data.frame",
  rownames = NULL)
}
\arguments{
\item{project}{bcbio project list}

\item{file}{File name}

\item{input}{Input format}

\item{output}{Desired output format}

\item{rownames}{Column identifier to use for rownames}
}
\description{
Import a bcbio project file
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{import}
