% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{bcbioGenerics}
\alias{bcbioGenerics}
\alias{bcbio}
\alias{bcbio<-}
\alias{flatFiles}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{metrics}
\alias{plotDot}
\alias{plotGene}
\alias{plotQC}
\alias{plotViolin}
\alias{sampleMetadata}
\alias{selectSamples}
\alias{tpm}
\title{bcbio S4 Generics}
\usage{
bcbio(object, ...)

bcbio(object, ...) <- value

flatFiles(object, ...)

interestingGroups(object, ...)

interestingGroups(object, ...) <- value

metrics(object, ...)

plotDot(object, ...)

plotGene(object, ...)

plotQC(object, ...)

plotViolin(object, ...)

sampleMetadata(object, ...)

selectSamples(object, ...)

tpm(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{value}{Value to assign.}
}
\description{
These generics are currently currently in use for both the bcbioRNASeq and
bcbioSingleCell packages.
}
