% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioMetadata.R
\name{bcbioMetadata}
\alias{bcbioMetadata}
\title{Metadata for bcbio run}
\usage{
bcbioMetadata(project, group = "genotype", laneSplit = FALSE)
}
\arguments{
\item{project}{bcbio project}

\item{group}{Sample grouping. Used later in pipline for graph colors.}

\item{laneSplit}{Whether samples were split across flow cell lanes.}
}
\value{
Metadata \code{data.frame}
}
\description{
Metadata for bcbio run
}
\examples{
\dontrun{
bcbioMetadata(project,
              group = "treatment",
              laneSplit = TRUE)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{rnaseq}
