% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioRnaseqData.R
\name{bcbioRnaseqData}
\alias{bcbioRnaseqData}
\title{Import RNA-Seq data from a bcbio project}
\usage{
bcbioRnaseqData(project)
}
\arguments{
\item{project}{bcbio project}
}
\value{
List containing RNA-Seq data
}
\description{
Import RNA-Seq data from a bcbio project
}
\examples{
\dontrun{
rnaseq <- bcbioRnaseqData(project)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{import}
