% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSailfish.R
\name{bcbioSailfish}
\alias{bcbioSailfish}
\title{Sailfish data}
\usage{
bcbioSailfish(summaryDir, summary)
}
\arguments{
\item{summaryDir}{Summary directory}

\item{summary}{Summary \code{data.frame}}
}
\value{
txi \code{tximport} count data from \code{sailfish}
}
\description{
Import sailfish data from \code{bcbio-rnaseq} run
}
\examples{
\dontrun{
bcbioSailfish(summary_dir, summary)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{rnaseq}
