% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioScrnaseqData.R
\name{bcbioScrnaseqData}
\alias{bcbioScrnaseqData}
\title{Import scRNA-Seq data from a bcbio project}
\usage{
bcbioScrnaseqData(project)
}
\arguments{
\item{project}{bcbio project}
}
\value{
List of scRNA-Seq data
}
\description{
Import scRNA-Seq data from a bcbio project
}
\examples{
\dontrun{
bcbioTagcounts(project)
}
}
\author{
Michael Steinbaugh
}
