% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSummary.R
\name{bcbioSummary}
\alias{bcbioSummary}
\title{Import a bcbio project summary}
\usage{
bcbioSummary(summaryDir)
}
\arguments{
\item{summaryDir}{Summary directory}
}
\description{
Import a bcbio project summary
}
\examples{
\dontrun{
bcbioSummary()
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{rnaseq}
