% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{bioc}
\alias{bioc}
\title{Install or update Bioconductor packages}
\usage{
bioc(...)
}
\arguments{
\item{...}{Passthrough to \code{biocLite()}}
}
\description{
This is function simply a wrapper for \code{biocLite}. It first sources the
code from the Bioconductor website then calls \code{biocLite}.
}
\examples{
\dontrun{
# Load Bioconductor and update packages
bioc()

# Install `limma`
bioc("limma")
}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
