% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{biocLite}
\alias{biocLite}
\title{Install or update Bioconductor and CRAN packages}
\usage{
biocLite(pkgs = NULL, ...)
}
\arguments{
\item{pkgs}{Character vector of package names.}

\item{...}{Additional arguments passed to \code{\link[BiocInstaller:biocLite]{BiocInstaller::biocLite()}}.}
}
\description{
This function is a convenience wrapper for \code{\link[BiocInstaller:biocLite]{BiocInstaller::biocLite()}}. It
sources the required \href{https://bioconductor.org/biocLite.R}{biocLite.R} file
from the Bioconductor website automatically then calls \code{\link[=biocLite]{biocLite()}}.
}
\details{
Supports Bioconductor, CRAN, and GitHub repositories.
}
\examples{
\dontrun{
# Load Bioconductor and update packages
biocLite()

# Install packages
biocLite("limma")            # Bioconductor
biocLite("dplyr")            # CRAN
biocLite("tidyverse/dplyr")  # GitHub
}
}
\seealso{
Modified variant of \code{rafalib::install_bioc()}.
}
\keyword{internal}
