% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/collapse.R, R/deprecated.R,
%   R/names.R
\name{british_aliases}
\alias{british_aliases}
\alias{summariseRows}
\alias{summarise_rows}
\alias{toStringSummarise}
\alias{to_string_summarize}
\alias{sanitiseNames}
\alias{sanitise_names}
\title{British spelling variants}
\description{
British English function name variants (e.g. \code{\link[=summarise]{summarise()}} instead of
\code{\link[=summarize]{summarize()}}).
}
\keyword{internal}
