% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/broadClass-methods.R
\docType{methods}
\name{broadClass}
\alias{broadClass}
\alias{broadClass,GRanges-method}
\alias{broadClass,SummarizedExperiment-method}
\title{Broad Class Definitions}
\usage{
broadClass(object, ...)

\S4method{broadClass}{GRanges}(object)

\S4method{broadClass}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object that can be coerced to \code{data.frame}, containing gene or
transcript annotations.}

\item{...}{Additional arguments.}
}
\value{
Named \code{factor} containing broad class definitions.
}
\description{
Broad Class Definitions
}
\examples{
# GRanges ====
x <- broadClass(makeGRangesFromEnsembl("Homo sapiens"))
table(x)

# SummarizedExperiment ====
x <- broadClass(rse_bcb)
table(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Rory Kirchner and Michael Steinbaugh
}
\concept{Gene Annotation Functions}
