% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheURL.R
\name{cacheURL}
\alias{cacheURL}
\title{Download and cache a data file from DepMap into BiocFileCache}
\usage{
cacheURL(url, fileName = basename(url), update = FALSE, verbose = TRUE)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{fileName}{\code{character(1)}.
File name to store internally in \code{BiocFileCache}.
Defaults to basename of URL.}

\item{update}{\code{logical(1)}.
Call \code{bfcneedsupdate} internally to see if URL needs an update.
Doesn't work reliably for all servers, so disabled by default.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}
}
\value{
\code{character(1)}.
Cached file path on disk.
}
\description{
Download and cache a data file from DepMap into BiocFileCache
}
\note{
Updated 2020-10-02.
}
\examples{
url <- pasteURL(
    basejumpTestsURL,
    "biocfilecache-test.txt",
    protocol = "none"
)
file <- cacheURL(url)
print(file)
}
