% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetrics-methods.R
\docType{methods}
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,DelayedArray-method}
\alias{calculateMetrics,Matrix-method}
\alias{calculateMetrics,SingleCellExperiment-method}
\title{Calculate quality control metrics}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{DelayedArray}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{calculateMetrics}{Matrix}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{calculateMetrics}{SingleCellExperiment}(object,
  prefilter = FALSE)
}
\arguments{
\item{object}{Object.}

\item{rowRanges}{\code{GRanges} or \code{GRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{prefilter}{\code{logical(1)}.
Apply prefiltering. Remove zero count genes.}
}
\value{
\code{DataFrame}.
}
\description{
Calculate quality control metrics
}
\note{
Input a raw count matrix. Do not use size factor adjusted or log
normalized counts here.

Updated 2019-08-07.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
x <- calculateMetrics(object)
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
