% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel.R
\name{camel}
\alias{camel}
\title{camelCase}
\usage{
camel(string)
}
\arguments{
\item{string}{String}
}
\value{
String with camelCase formatting
}
\description{
camelCase
}
\examples{
camel("RNAi clone")
}
\author{
Michael Steinbaugh
}
