% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/names.R, R/reexports.R
\name{camel_aliases}
\alias{camel_aliases}
\alias{sanitizeNames}
\alias{setColnames}
\alias{setRownames}
\title{camelCase aliases}
\usage{
sanitizeNames(x, rownames = TRUE)
}
\description{
Function name variants formatted in \code{camelCase}.
}
\keyword{internal}
