% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelCase-methods.R
\docType{methods}
\name{camelCase}
\alias{camelCase}
\alias{camelCase,atomic-method}
\alias{camelCase,factor-method}
\alias{camelCase,list-method}
\alias{camelCase,matrix-method}
\alias{camelCase,data.frame-method}
\alias{camelCase,Vector-method}
\alias{camelCase,DataTable-method}
\alias{camelCase,Ranges-method}
\alias{camelCase,Matrix-method}
\alias{camelCase,SummarizedExperiment-method}
\title{Camel case}
\usage{
\S4method{camelCase}{atomic}(object, names = TRUE, strict = FALSE)

\S4method{camelCase}{factor}(object, names = TRUE, strict = FALSE)

\S4method{camelCase}{list}(object, names = TRUE, strict = FALSE)

\S4method{camelCase}{matrix}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camelCase}{data.frame}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{camelCase}{Vector}(object, names = TRUE, mcols = TRUE,
  metadata = TRUE, strict = FALSE)

\S4method{camelCase}{DataTable}(object, rownames = FALSE,
  colnames = TRUE, mcols = TRUE, metadata = TRUE, strict = FALSE)

\S4method{camelCase}{Ranges}(object, names = FALSE, mcols = TRUE,
  metadata = TRUE, strict = FALSE)

\S4method{camelCase}{Matrix}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camelCase}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE, assayNames = TRUE, rowData = TRUE,
  colData = TRUE, metadata = TRUE, strict = FALSE)
}
\arguments{
\item{object}{Object.}

\item{names}{\code{logical(1)}.
Sanitize names.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
}
\description{
Format character strings to use (lower) camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{thisIsCamelCase}).
}
\details{
Camel case is recommended by Bioconductor for variable and function names.
}
\note{
Updated 2019-09-25.
}
\examples{
data(syntactic, package = "acidtest")
lapply(syntactic, camelCase)
}
