% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnnotable.R
\name{checkAnnotable}
\alias{checkAnnotable}
\title{Check Ensembl Annotations}
\usage{
checkAnnotable(object)
}
\arguments{
\item{object}{\link{data.frame} containing Ensembl gene annotations returned
from the \code{\link[=annotable]{annotable()}} function.}
}
\value{
Silent on pass, stop on error.
}
\description{
Check Ensembl Annotations
}
\examples{
annotable <- annotable("Homo sapiens")
checkAnnotable(annotable)
}
