% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{checkClasses}
\alias{checkClasses}
\title{Check Classes}
\usage{
checkClasses(object, expected, subset = FALSE)
}
\arguments{
\item{object}{Object.}

\item{expected}{\code{list}. Named list of expected classes per slot.}

\item{subset}{\code{boolean}. Only check a subset of slots in the object.}
}
\description{
Check Classes
}
\examples{
data(rse)
object <- rse
metadata <- S4Vectors::metadata(object)
checkClasses(
    object = metadata,
    expected = list(
        version = c("package_version", "numeric_version"),
        date = "Date",
        interestingGroups = "character"
    )
)
}
