% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGene2symbol.R
\name{checkGene2symbol}
\alias{checkGene2symbol}
\title{Check Gene to Symbol Mapping Data}
\usage{
checkGene2symbol(object)
}
\arguments{
\item{object}{\link{data.frame} containing Ensembl gene identifier to gene symbol
mappings. Must be structured as a two column \link{data.frame} with "ensgene"
and "symbol" columns.}
}
\value{
Silent on pass, stop on error.
}
\description{
Check Gene to Symbol Mapping Data
}
\examples{
# Success
gene2symbol <- annotable("Homo sapiens", format = "gene2symbol")
checkGene2symbol(gene2symbol)

# Failure
\dontrun{
checkGene2symbol(mtcars)
}
}
