% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInstall.R
\name{checkInstall}
\alias{checkInstall}
\title{Check to see if packages are installed}
\usage{
checkInstall(packages)
}
\arguments{
\item{packages}{Character vector of packages}
}
\description{
Check to see if packages are installed
}
