% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInterestingGroups.R
\name{checkInterestingGroups}
\alias{checkInterestingGroups}
\title{Check Interesting Groups}
\usage{
checkInterestingGroups(object, interestingGroups, warnOnNULL = FALSE)
}
\arguments{
\item{object}{Object supporting \code{\link[=colnames]{colnames()}}, typically a \link{data.frame}.}

\item{interestingGroups}{Interesting groups character vector.}

\item{warnOnNULL}{Warn the user on \code{NULL} argument.}
}
\value{
Valid character of defined interesting groups. Stop on failure.
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
