% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTx2gene.R
\name{checkTx2gene}
\alias{checkTx2gene}
\title{Check Transcript to Gene Mapping Data}
\usage{
checkTx2gene(object)
}
\arguments{
\item{object}{\link{data.frame} containing Ensembl transcript to gene identifier
mappings. Must be structured as a two column \link{data.frame} with "enstxp" and
"ensgene" columns.}
}
\value{
Silent on pass, stop on error.
}
\description{
Check Transcript to Gene Mapping Data
}
\examples{
tx2gene <- annotable("Homo sapiens", format = "tx2gene")
checkTx2gene(tx2gene)
}
