% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{import}
\name{classes}
\alias{classes}
\alias{EggNOG-class}
\alias{Ensembl2Entrez-class}
\alias{Gene2Symbol-class}
\alias{HGNC2Ensembl-class}
\alias{MGI2Ensembl-class}
\alias{PANTHER-class}
\alias{Tx2Gene-class}
\title{Classes exported from other packages}
\description{
These classes are imported from other packages.
Follow the links below to see their documentation.
}
\details{
\itemize{
\item \link[basejump.classes:`EggNOG-class`]{basejump.classes::EggNOG-class}.
\item \link[basejump.classes:`Ensembl2Entrez-class`]{basejump.classes::Ensembl2Entrez-class}.
\item \link[basejump.classes:`Gene2Symbol-class`]{basejump.classes::Gene2Symbol-class}.
\item \link[basejump.classes:`HGNC2Ensembl-class`]{basejump.classes::HGNC2Ensembl-class}.
\item \link[basejump.classes:`MGI2Ensembl-class`]{basejump.classes::MGI2Ensembl-class}.
\item \link[basejump.classes:`PANTHER-class`]{basejump.classes::PANTHER-class}.
\item \link[basejump.classes:`Tx2Gene-class`]{basejump.classes::Tx2Gene-class}.
}
}
\keyword{internal}
