% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCells.R
\name{cleanCells}
\alias{cleanCells}
\title{Cruft removal for data frame cells}
\usage{
cleanCells(df)
}
\arguments{
\item{df}{Messy data frame -- leading/trailing spaces, dashes}
}
\value{
A reformatted, clean data frame
}
\description{
Cruft removal for data frame cells
}

