% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSystemLibrary.R
\name{cleanSystemLibrary}
\alias{cleanSystemLibrary}
\title{Check for clean system library}
\usage{
cleanSystemLibrary()
}
\value{
\code{logical(1)}.
Is the system library clean?
}
\description{
Determine whether a user has installed packages into the system library.
}
\examples{
cleanSystemLibrary()
}
