% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearWarnings.R
\name{clearWarnings}
\alias{clearWarnings}
\alias{clear_warnings}
\title{Clear warnings.}
\usage{
clearWarnings()

clear_warnings()
}
\description{
Clear warnings.
}
\examples{
\dontrun{
clearWarnings()
}
}
