% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-list-methods.R
\name{coerceToList}
\alias{coerceToList}
\alias{coerceToList,Annotated-method}
\title{Coerce to list}
\usage{
\S4method{coerceToList}{Annotated}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
}
\description{
Coerce to list
}
\note{
Don't attempt to defined methods for \code{as.list}.

Updated 2020-01-18.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
x <- RangedSummarizedExperiment
x <- coerceToList(x)
class(x)
names(x)
}
\seealso{
Refer to the S4Vectors package for details.
\itemize{
\item \code{help(topic = "Annotated-class", package = "S4Vectors")}
\item \code{help(topic = "List-class", package = "S4Vectors")}
\item \code{showClass("Annnotated")}.
\item \code{showClass("List")}.
\item \code{methods::coerce}.
}
}
