% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R, R/generics.R
\name{coerce}
\alias{coerce}
\alias{coerce,DataFrame,data.table-method}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GenomicRanges,data.table-method}
\alias{coerce,GenomicRanges,tbl_df-method}
\alias{coerce,IntegerRanges,data.frame-method}
\alias{coerce,IntegerRanges,data.table-method}
\alias{coerce,IntegerRanges,tbl_df-method}
\alias{coerce,Matrix,DataFrame-method}
\alias{coerce,Matrix,data.frame-method}
\alias{coerce,data.frame,data.table-method}
\alias{coerce,data.frame,tbl_df-method}
\alias{coerce,data.table,DataFrame-method}
\alias{coerce,tbl_df,DataFrame-method}
\title{Coercion methods}
\value{
Modified object, of desired conversion class.
}
\description{
Coercion methods
}
\details{
These conversion methods are primarily intended to interconvert between
popular tabular formats in R, including \code{data.frame}, \code{data.table}, \code{tbl_df},
and the Bioconductor \code{DataFrame} classes.
}
\note{
Updated 2022-03-03.
}
