% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceS4ToList.R
\docType{methods}
\name{coerceS4ToList}
\alias{coerceS4ToList}
\alias{flatFiles,SummarizedExperiment-method}
\title{Coerce any S4 object to a list}
\usage{
coerceS4ToList(from)

\S4method{flatFiles}{SummarizedExperiment}(object)
}
\arguments{
\item{from}{The classes between which the coerce methods
    \code{def} and \code{replace} perform coercion.
  }

\item{object}{Object.}
}
\value{
\code{list}.
}
\description{
Coerce any S4 object to a list
}
\examples{
data(rse)

x <- coerceS4ToList(rse)
class(x)
names(x)

x <- flatFiles(rse)
class(x)
names(x)
}
\seealso{
\code{methods::coerce}.
}
