% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerceToList-methods.R
\docType{methods}
\name{coerceToList}
\alias{coerceToList}
\alias{coerceToList,ANY-method}
\title{Coerce S4 Object to List}
\usage{
coerceToList(object, ...)

\S4method{coerceToList}{ANY}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
}
\description{
Extract the slots inside an S4 object for archival storage.
}
\examples{
x <- coerceToList(rse_small)
class(x)
names(x)
}
\author{
Michael Steinbaugh
}
\concept{Coercion Methods}
