% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/collapse.R
\docType{methods}
\name{toStringUnique}
\alias{toStringUnique}
\alias{toStringSortUnique}
\alias{summarizeRows}
\alias{toStringUnique,character-method}
\alias{toStringSortUnique,character-method}
\alias{.summarizeRows}
\alias{summarizeRows,data.frame-method}
\alias{summarizeRows,matrix-method}
\title{Collapse Utilities}
\usage{
toStringUnique(object, ...)

toStringSortUnique(object, ...)

summarizeRows(object, ...)

\S4method{toStringUnique}{character}(object, sep = ", ")

\S4method{toStringSortUnique}{character}(object, sep = ", ")

\S4method{summarizeRows}{data.frame}(object, sep = ", ")

\S4method{summarizeRows}{matrix}(object, sep = ", ")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{sep}{Separator used for collapse.}

\item{x}{The object to be converted.}
}
\value{
A character vector of length 1 is returned.
}
\description{
Helper functions for to produce collapsed output for vectors
(\code{\link[=toString]{toString()}}) and data frames (\code{\link[=summarizeRows]{summarizeRows()}}).
}
\examples{
groceries <- c("milk", "eggs", "eggs", NA)

toStringUnique(groceries)
toStringSortUnique(groceries)

mtcars \%>\% head \%>\% summarizeRows
}
