% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseToString-methods.R
\docType{methods}
\name{collapseToString}
\alias{collapseToString}
\alias{collapseToString,atomic-method}
\alias{collapseToString,matrix-method}
\alias{collapseToString,data.frame-method}
\alias{collapseToString,DataFrame-method}
\title{Collapse to string}
\usage{
\S4method{collapseToString}{atomic}(object, sep = ", ", sort = FALSE,
  removeNA = FALSE, unique = FALSE)

\S4method{collapseToString}{matrix}(object, sep = ", ", sort = FALSE,
  removeNA = FALSE, unique = FALSE)

\S4method{collapseToString}{data.frame}(object, sep = ", ",
  sort = FALSE, removeNA = FALSE, unique = FALSE)

\S4method{collapseToString}{DataFrame}(object, sep = ", ",
  sort = FALSE, removeNA = FALSE, unique = FALSE)
}
\arguments{
\item{object}{Object.}

\item{sep}{\code{character(1)}.
Separator. Defaults to comma.}

\item{sort}{\code{logical(1)}.
Sort values.}

\item{removeNA}{\code{logical(1)}.
Remove NA values.}

\item{unique}{\code{logical(1)}.
Unique values.}
}
\value{
\itemize{
\item \code{atomic}: \code{character(1)}.
\item \code{dim}: Object of same class, collapsed to a single row.
}
}
\description{
Collapse to string
}
\examples{
## character ====
groceries <- c(NA, NA, "milk", "eggs", "eggs", "veggies")
collapseToString(
    groceries,
    unique = TRUE,
    sort = TRUE,
    removeNA = TRUE
)
collapseToString(
    groceries,
    unique = FALSE,
    sort = FALSE,
    removeNA = FALSE
)

## numeric ====
collapseToString(seq(1:5))

## logical ====
collapseToString(c(TRUE, FALSE))
collapseToString(c(NA, NaN))

## data.frame ====
datasets::iris \%>\%
    head() \%>\%
    collapseToString(sort = TRUE, unique = TRUE) \%>\%
    t()
}
\seealso{
\code{toString}.
}
