% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\name{combine}
\alias{combine}
\alias{combine,SummarizedExperiment,SummarizedExperiment-method}
\alias{combine,SingleCellExperiment,SingleCellExperiment-method}
\title{Combining or merging different Bioconductor data structures}
\usage{
combine(x, y, ...)

\S4method{combine}{SummarizedExperiment,SummarizedExperiment}(x, y)

\S4method{combine}{SingleCellExperiment,SingleCellExperiment}(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{SummarizedExperiment}.
}
\description{
Combining or merging different Bioconductor data structures
}
\note{
Updated 2020-01-20.

We're attempting to make this as strict as possible, requiring:
\itemize{
\item Rows (genes) across objects must be identical.
\item \link[SummarizedExperiment:RangedSummarizedExperiment-class]{rowRanges} and/or
\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}
\link[S4Vectors:Annotated-class]{metadata} must be identical.
\item \link[SummarizedExperiment:SummarizedExperiment-class]{colData} must contain the same columns.
\item Specific metadata must be identical (see \code{metadata} argument).
}
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)
str_pad <- stringr::str_pad

## SummarizedExperiment ====
x <- RangedSummarizedExperiment
colnames(x) <- paste0(
    "sample",
    str_pad(
        string = seq_len(ncol(x)),
        width = 2L,
        pad = "0"
    )
)

y <- x
colnames(y) <- paste0(
    "sample",
    str_pad(
        string = seq(from = ncol(y) + 1L, to = ncol(y) * 2L),
        width = 2L,
        pad = "0"
    )
)

## Combine the two objects.
c <- combine(x, y)
sampleData(c)
print(c)

## SingleCellExperiment ====
x <- SingleCellExperiment
colnames(x) <- paste0(
    "cell",
    str_pad(
        string = seq_len(ncol(x)),
        width = 4L,
        pad = "0"
    )
)

y <- x
colnames(y) <- paste0(
    "cell",
    str_pad(
        string = seq_len(ncol(y)) + ncol(y),
        width = 4L,
        pad = "0"
    )
)

## Combine the two objects.
c <- combine(x, y)
sampleData(c)
print(c)
}
\seealso{
\itemize{
\item \code{BiocGenerics::combine()}.
\item \code{help("merge.Matrix", "Matrix.utils")}.
}
}
