% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{contrastName}
\alias{contrastName}
\title{Contrast name}
\usage{
contrastName(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
Contrast name.
}
\description{
Contrast name
}
\examples{
isS4(contrastName)
}
\seealso{
\href{https://steinbaugh.com/DESeqAnalysis/}{DESeqAnalysis}.
}
