% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertGenesToSymbols-methods.R
\docType{methods}
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertSymbolsToGenes}
\alias{convertGenesToSymbols,character-method}
\alias{convertGenesToSymbols,matrix-method}
\alias{convertGenesToSymbols,sparseMatrix-method}
\alias{convertGenesToSymbols,SummarizedExperiment-method}
\alias{convertSymbolsToGenes,SummarizedExperiment-method}
\title{Convert Ensembl Identifiers to Gene Symbols}
\usage{
convertGenesToSymbols(object, ...)

convertSymbolsToGenes(object, ...)

\S4method{convertGenesToSymbols}{character}(object, gene2symbol)

\S4method{convertGenesToSymbols}{matrix}(object, gene2symbol)

\S4method{convertGenesToSymbols}{sparseMatrix}(object, gene2symbol)

\S4method{convertGenesToSymbols}{SummarizedExperiment}(object)

\S4method{convertSymbolsToGenes}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{gene2symbol}{\code{Gene2Symbol}. Gene-to-symbol mappings.}
}
\value{
Object with gene IDs converted to names (symbols).
}
\description{
Convert Ensembl Identifiers to Gene Symbols
}
\examples{
data(rse_small)
object <- rse_small

g2s <- Gene2Symbol(object)
print(g2s)
genes <- head(g2s[["geneID"]])
print(genes)

## character ====
x <- convertGenesToSymbols(genes, gene2symbol = g2s)
print(x)

## matrix ====
samples <- head(colnames(object))
counts <- matrix(
    data = seq_len(length(genes) * length(samples)),
    byrow = TRUE,
    nrow = length(genes),
    ncol = length(samples),
    dimnames = list(genes, samples)
)
print(counts)
x <- convertGenesToSymbols(counts, gene2symbol = g2s)
print(x)

## SummarizedExperiment ====
x <- convertGenesToSymbols(rse_small)
print(x)
## Interconvert back to gene IDs.
y <- convertSymbolsToGenes(x)
print(y)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{Ensembl2Entrez}},
  \code{\link{Gene2Symbol}}, \code{\link{HGNC2Ensembl}},
  \code{\link{MGI2Ensembl}}, \code{\link{Tx2Gene}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{mapGenes}}
}
\concept{Identifier Mapping Functions}
