% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUCSCBuildToEnsembl.R
\name{convertUCSCBuildToEnsembl}
\alias{convertUCSCBuildToEnsembl}
\title{Convert UCSC build to Ensembl}
\usage{
convertUCSCBuildToEnsembl(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character}. Ensembl genome build as the value, UCSC build as the
name. Stops on match failure.
}
\description{
Convert UCSC build to Ensembl
}
\examples{
from <- c("hg19", "hg38")
to <- convertUCSCBuildToEnsembl(from)
print(to)
}
\seealso{
\itemize{
\item \href{https://genome.ucsc.edu/cgi-bin/hgGateway}{UCSC hgGateway}
\item \href{https://github.com/bcbio/bcbio-nextgen/tree/master/config/genomes}{bcbio genome recipes}
}
}
