% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corHeatmap.R
\name{corHeatmap}
\alias{corHeatmap}
\title{Correlation matrix heatmap}
\usage{
corHeatmap(counts, method = "pearson", annotation, ...)
}
\arguments{
\item{counts}{Counts matrix}
}
\value{
Heatmap
}
\description{
Correlation matrix heatmap
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
