% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-methods.R
\docType{methods}
\name{correlation}
\alias{correlation}
\alias{correlation,numeric,numeric-method}
\alias{correlation,matrix,missingOrNULL-method}
\alias{correlation,matrix,matrix-method}
\title{Correlation}
\usage{
correlation(x, y, ...)

\S4method{correlation}{numeric,numeric}(x, y, method = c("pearson",
  "kendall", "spearman"))

\S4method{correlation}{matrix,missingOrNULL}(x, y, method = c("pearson",
  "kendall", "spearman"))

\S4method{correlation}{matrix,matrix}(x, y, method = c("pearson",
  "kendall", "spearman"))
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{i}{\code{integer(1)} or \code{character(1)}.
For \code{SummarizedExperiment}, primary assay.}

\item{j}{\code{integer(1)}, \code{character(1)}, or \code{NULL}.
For \code{SummarizedExperiment}, optional secondary assay.
If \code{NULL}, calculates correlation matrix only on the primary assay.}
}
\value{
\code{numeric(1)} or \code{matrix}.
}
\description{
Correlation
}
\note{
Updated 2019-11-08.
}
\examples{
data(correlation, package = "acidtest")
list <- correlation
rm(correlation)

## vector ====
x <- list[["vector_x"]]
y <- list[["vector_y"]]

head(x)
head(y)

correlation(x = x, y = y)

## matrix ====
x <- list[["matrix_x"]]
y <- list[["matrix_y"]]

head(x)
head(y)

cor(x)
correlation(x)

cor(x = c(x), y = c(y))
correlation(x, y)

## SummarizedExperiment ====
x <- list[["SummarizedExperiment_x"]]
y <- list[["SummarizedExperiment_y"]]

correlation(x = x, i = 1L, j = 2L)
}
\seealso{
\itemize{
\item \code{stats::cor()}.
\item \code{stats::cor.test()}.
\item \code{stats::sd()}.
\item \code{S4Vectors::cor()}.
\item https://stats.stackexchange.com/questions/24980
}
}
