% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts,SummarizedExperiment-method}
\alias{counts<-,SummarizedExperiment,ANY-method}
\title{Counts}
\usage{
\S4method{counts}{SummarizedExperiment}(object)

\S4method{counts}{SummarizedExperiment,ANY}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Matrix. Typically \code{matrix} or \code{sparseMatrix} class.
}
\description{
Count matrix.
}
\note{
For a \code{SummarizedExperiment} object, \code{"counts"} must be explicitly
defined in \code{assayNames()}.
}
\examples{
data(rse)
x <- counts(rse)
summary(x)
}
