% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsPerGene.R
\name{countsPerGene}
\alias{countsPerGene}
\title{Counts per gene plot}
\usage{
countsPerGene(rawCounts, metadata)
}
\arguments{
\item{rawCounts}{Raw counts \code{matrix}}

\item{metadata}{Metadata \code{data.frame}}
}
\value{
Boxplot
}
\description{
Counts per gene plot
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
