% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/createProjectDirs-methods.R
\docType{methods}
\name{createProjectDirs}
\alias{createProjectDirs}
\alias{createProjectDirs,missing-method}
\title{Create Project Directory Structure}
\usage{
createProjectDirs(object, ...)

\S4method{createProjectDirs}{missing}()
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Create necessary directory structure for an
\href{http://rmarkdown.rstudio.com}{RMarkdown} report in
\href{https://www.rstudio.com}{RStudio}.
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadData}},
  \code{\link{loadRemoteData}}, \code{\link{packageSE}},
  \code{\link{readDataRaw}},
  \code{\link{readFileByExtension}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
