% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentGenomeVersions.R
\name{currentGenomeVersion}
\alias{currentGenomeVersion}
\alias{currentEnsemblVersion}
\alias{currentGencodeVersion}
\alias{currentRefseqVersion}
\alias{currentFlybaseVersion}
\alias{currentWormbaseVersion}
\title{Current genome version}
\usage{
currentEnsemblVersion()

currentGencodeVersion(organism)

currentRefseqVersion()

currentFlybaseVersion(dmel = FALSE)

currentWormbaseVersion()
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{dmel}{\code{logical(1)}.
Return \emph{Drosophila melanogaster} genome release version.}
}
\value{
\code{character(1)}.
Release version.
}
\description{
Obtain the latest release version from various genome annotation sources.
}
\note{
Updated 2020-07-24.
}
\examples{
## Protect against Ensembl timeouts causing build checks to fail.
## > if (goalie::hasInternet(url = "ftp://ftp.ensembl.org/")) {
## >     currentEnsemblVersion()
## > }
}
\seealso{
Refer to the koopa package for shell variants:
\itemize{
\item \code{current-ensembl-version}.
\item \code{current-flybase-version}.
\item \code{current-gencode-version}.
\item \code{current-refseq-version}.
\item \code{current-wormbase-version}.
}
}
