% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david.R
\name{david}
\alias{david}
\title{RDAVIDWebService query}
\usage{
david(email, fg, bg, id)
}
\arguments{
\item{email}{Registered email address}

\item{fg}{foregroundGenes identifier vector}

\item{bg}{backgroundGenes identifier vector}

\item{id}{Gene identifier type (see DAVID website)}
}
\value{
Functional annotation chart data frame
}
\description{
RDAVIDWebService query
}

