% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/davidQuery.R
\name{davidQuery}
\alias{davidQuery}
\title{RDAVIDWebService query}
\usage{
davidQuery(email, fg, bg, id, save = NULL)
}
\arguments{
\item{email}{Registered email address}

\item{fg}{foregroundGenes identifier vector}

\item{bg}{backgroundGenes identifier vector}

\item{id}{Gene identifier type (see DAVID website)}
}
\value{
Functional annotation chart data frame
}
\description{
RDAVIDWebService query
}

