% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rle-methods.R
\docType{methods}
\name{decode}
\alias{decode}
\alias{decode,DataFrame-method}
\title{Rle objects}
\usage{
\S4method{decode}{DataFrame}(x)
}
\arguments{
\item{x}{Object.}
}
\description{
The Rle class is a general container for storing an atomic vector
  that is stored in a run-length encoding format. It is based on the
  \code{\link[base]{rle}} function from the base package.
}
\examples{
data(rse)
mcols <- S4Vectors::mcols(rse)
lapply(mcols, class)
x <- decode(mcols)
lapply(x, class)
}
