% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{theme_midnight}
\alias{theme_paperwhite}
\alias{tx2geneFromGFF}
\alias{separatorBar}
\alias{readFileByExtension}
\alias{readGFF}
\alias{readGTF}
\alias{readJSON}
\alias{readYAML}
\alias{relevelColData}
\alias{relevelRowData}
\alias{relevelRowRanges}
\alias{readSampleData}
\alias{readTx2Gene}
\alias{coerceS4ToList}
\alias{matchEnsemblReleaseToURL}
\alias{matchHumanOrthologs}
\alias{multiassignAsEnvir}
\alias{sortUnique}
\alias{toStringUnique}
\title{Defunct functions}
\usage{
theme_midnight(...)

theme_paperwhite(...)

tx2geneFromGFF(...)

separatorBar(...)

readFileByExtension(...)

readGFF(...)

readGTF(...)

readJSON(...)

readYAML(...)

relevelColData(...)

relevelRowData(...)

relevelRowRanges(...)

readSampleData(...)

readTx2Gene(...)

coerceS4ToList(...)

matchEnsemblReleaseToURL(...)

matchHumanOrthologs(...)

multiassignAsEnvir(...)

sortUnique(...)

toStringUnique(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\examples{
a <- function(...) {
    .Defunct("b")
}

withCallingHandlers(
    expr = tryCatch(
        expr = a(1L),
        error = function(e) {
            print(e)
            invisible()
        }
    )
)
}
\seealso{
\link[base:Defunct]{Defunct}.
}
\keyword{internal}
