% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{deg}
\alias{deg}
\title{Differentially expressed genes}
\usage{
deg(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Gene identifiers.
}
\description{
Differentially expressed genes
}
\examples{
isS4(deg)
}
\seealso{
\href{https://steinbaugh.com/DESeqAnalysis/}{DESeqAnalysis}.
}
