% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{assertIsURL}
\alias{aggregateFeatures}
\alias{aggregateReplicates}
\alias{aggregateSamples}
\alias{assertFormalGene2symbol}
\alias{assertIsAHeaderLevel}
\alias{assertIsCharacterOrNULL}
\alias{assertIsDataFrameOrNULL}
\alias{eggnog}
\alias{ensembl2entrez}
\alias{fixNA}
\alias{flatFiles}
\alias{gene2symbol}
\alias{grepString}
\alias{hgnc2gene}
\alias{hgnc2ensembl}
\alias{initializeDirectory}
\alias{kables}
\alias{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGTF}
\alias{makeTx2geneFromEnsembl}
\alias{makeTx2geneFromGFF}
\alias{makeTx2geneFromGTF}
\alias{mgi2gene}
\alias{mgi2ensembl}
\alias{panther}
\alias{readFileByExtension}
\alias{readGFF}
\alias{readGTF}
\alias{readJSON}
\alias{readYAML}
\alias{tx2gene}
\title{Deprecated Functions}
\usage{
assertIsURL(...)

aggregateFeatures(...)

aggregateReplicates(...)

aggregateSamples(...)

assertFormalGene2symbol(...)

assertIsAHeaderLevel(...)

assertIsCharacterOrNULL(object, ...)

assertIsDataFrameOrNULL(object, ...)

eggnog(...)

ensembl2entrez(...)

fixNA(...)

flatFiles(object)

gene2symbol(...)

grepString(...)

hgnc2gene(...)

hgnc2ensembl(...)

initializeDirectory(...)

kables(...)

makeGene2symbolFromEnsembl(...)

makeGene2symbolFromGFF(...)

makeGene2symbolFromGTF(...)

makeTx2geneFromEnsembl(...)

makeTx2geneFromGFF(...)

makeTx2geneFromGTF(...)

mgi2gene(...)

mgi2ensembl(...)

panther(...)

readFileByExtension(...)

readGFF(...)

readGTF(...)

readJSON(...)

readYAML(...)

tx2gene(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\keyword{internal}
