% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseqPca.R
\name{deseqPca}
\alias{deseqPca}
\title{Quickly plot principal component analysis (PCA)}
\usage{
deseqPca(dt, factor = c("group"), ...)
}
\arguments{
\item{dt}{\code{DESeqTransform} object generated from \code{rlog()} or
\code{vst()} on a \code{DESeqDataSet} object. We prefer to pipe the
transformed data in to this function because the operation is CPU
intensive, and is slow when applied to multiple PCA plot operations.}

\item{factor}{PCA factor groups for \code{DESeq2::plotPCA()}. Shapes are
defined by the first value.}

\item{...}{Passthrough to \code{DESeq2::plotPCA()}.}
}
\value{
PCA plot
}
\description{
Quickly plot principal component analysis (PCA)
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{rnaseq}
