% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_setup.R
\name{dev_setup}
\alias{dev_setup}
\title{Set up developer environment}
\usage{
dev_setup()
}
\value{
Installs devtools, roxygen2, and lintr.
}
\description{
Installs devtools, roxygen2, and lintr.
}
\examples{
dev_setup()
}

