% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-dna.R
\docType{methods}
\name{comp}
\alias{comp}
\alias{revcomp}
\alias{dna}
\alias{comp,character-method}
\alias{comp,numeric-method}
\alias{revcomp,character-method}
\alias{revcomp,numeric-method}
\title{DNA Sequence Utilites}
\usage{
comp(object, ...)

revcomp(object, ...)

\S4method{comp}{character}(object)

\S4method{comp}{numeric}(object)

\S4method{revcomp}{character}(object)

\S4method{revcomp}{numeric}(object)
}
\arguments{
\item{object}{DNA sequence (\code{ATGC} nucleotides).}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Matching DNA sequence.
}
\description{
Generate complement or reverse complement sequences.
}
\examples{
dna <- "ATGCATGC"
comp(dna)
revcomp(dna)
}
\concept{DNA Sequence Utilities}
