% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-arrangeMetadataByPriorityCols.R
\name{.arrangeMetadataByPriorityCols}
\alias{.arrangeMetadataByPriorityCols}
\title{Arrange Metadata by Priority Columns}
\usage{
.arrangeMetadataByPriorityCols(meta)
}
\arguments{
\item{meta}{Sample metadata \link{data.frame}.}
}
\value{
\link{data.frame}.
}
\description{
Arrange Metadata by Priority Columns
}
\seealso{
Other Exported Constructor Functions: \code{\link{.SummarizedExperiment}},
  \code{\link{.interestingColData}},
  \code{\link{.interestingGroups}},
  \code{\link{.interestingGroup}},
  \code{\link{.sampleYAMLMetadata}},
  \code{\link{.sampleYAMLMetrics}},
  \code{\link{.sampleYAML}},
  \code{\link{.setMetadataFactors}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
