% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-dataVersions.R
\name{.dataVersions}
\alias{.dataVersions}
\title{Read Data Versions}
\usage{
.dataVersions(projectDir)
}
\arguments{
\item{projectDir}{Project directory.}
}
\value{
\link{data.frame}.
}
\description{
Read Data Versions
}
\seealso{
Other bcbio Project Directory File Utilities: \code{\link{.logFile}},
  \code{\link{.programs}}
}
\concept{bcbio Project Directory File Utilities}
\keyword{internal}
