% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-interestingGroup.R
\name{.interestingGroup}
\alias{.interestingGroup}
\title{Primary Interesting Group}
\usage{
.interestingGroup(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{string}.
}
\description{
Primary Interesting Group
}
\seealso{
Other Exported Constructor Functions: \code{\link{.interestingColData}},
  \code{\link{.interestingGroups}},
  \code{\link{.sampleMetadata}},
  \code{\link{.sampleYAMLMetadata}},
  \code{\link{.sampleYAMLMetrics}},
  \code{\link{.sampleYAML}},
  \code{\link{prepareSummarizedExperiment}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
