% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-localOrRemoteFile.R
\name{.localOrRemoteFile}
\alias{.localOrRemoteFile}
\title{Dynamically Handle a Local or Remote File Path}
\usage{
.localOrRemoteFile(object, quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Named character vector containing the original file name as the
name and local file path as the string. Returns \code{NULL} on a misisng
local file.
}
\description{
Dynamically Handle a Local or Remote File Path
}
\examples{
\dontrun{
file.path(testDataURL, "mtcars.csv") \%>\%
    .localOrRemoteFile()
}
}
\keyword{internal}
