% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-logFile.R
\name{.logFile}
\alias{.logFile}
\title{Read Log File}
\usage{
.logFile(file)
}
\arguments{
\item{file}{Log file.}
}
\value{
Character vector.
}
\description{
Read Log File
}
\seealso{
Other bcbio Project Directory File Utilities: \code{\link{.dataVersions}},
  \code{\link{.programs}}
}
\concept{bcbio Project Directory File Utilities}
\keyword{internal}
