% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-prepareSampleMetadata.R
\name{.prepareSampleMetadata}
\alias{.prepareSampleMetadata}
\title{Sample Metadata Constructor}
\usage{
.prepareSampleMetadata(metadata, factors = TRUE)
}
\arguments{
\item{metadata}{Metadata \link{data.frame}.}

\item{factors}{Set columns that don't apply to sample names as factors.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata Constructor
}
\keyword{internal}
