% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-programs.R
\name{.programs}
\alias{.programs}
\title{Read Program Versions}
\usage{
.programs(projectDir)
}
\arguments{
\item{projectDir}{Project directory.}
}
\value{
\link{data.frame}.
}
\description{
Read Program Versions
}
\seealso{
Other bcbio Project Directory File Utilities: \code{\link{.dataVersions}},
  \code{\link{.logFile}}
}
\concept{bcbio Project Directory File Utilities}
\keyword{internal}
