% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-returnInterestingGroups.R
\name{.returnInterestingGroups}
\alias{.returnInterestingGroups}
\title{Return Interesting Groups}
\usage{
.returnInterestingGroups(object, interestingGroups)
}
\value{
\code{character}.
}
\description{
Return Interesting Groups
}
\examples{
.returnInterestingGroups(rse_bcb)
}
\author{
Michael Steinbaugh
}
\keyword{internal}
