% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-sampleMetadata.R
\name{.sampleMetadata}
\alias{.sampleMetadata}
\title{Sample Metadata}
\usage{
.sampleMetadata(meta)
}
\arguments{
\item{meta}{Metadata \link{data.frame}.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata
}
\seealso{
Other Exported Constructor Functions: \code{\link{.SummarizedExperiment}},
  \code{\link{.interestingColData}},
  \code{\link{.interestingGroups}},
  \code{\link{.interestingGroup}},
  \code{\link{.sampleYAMLMetadata}},
  \code{\link{.sampleYAMLMetrics}},
  \code{\link{.sampleYAML}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
