% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-sampleYAML.R
\name{.sampleYAML}
\alias{.sampleYAML}
\title{Sample YAML Metadata Utilities}
\usage{
.sampleYAML(yaml, ...)
}
\arguments{
\item{yaml}{Project summary YAML.}

\item{...}{Nested operator keys supplied as dot objects.}
}
\value{
\link{tibble}.
}
\description{
Sample YAML Metadata Utilities
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\seealso{
Other Exported Constructor Functions: \code{\link{.SummarizedExperiment}},
  \code{\link{.interestingColData}},
  \code{\link{.interestingGroups}},
  \code{\link{.interestingGroup}},
  \code{\link{.sampleMetadata}},
  \code{\link{.sampleYAMLMetadata}},
  \code{\link{.sampleYAMLMetrics}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
