% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-sampleYAMLMetadata.R
\name{.sampleYAMLMetadata}
\alias{.sampleYAMLMetadata}
\title{Sample Metadata from YAML}
\usage{
.sampleYAMLMetadata(yaml)
}
\arguments{
\item{yaml}{YAML \link{list}.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata from YAML
}
\seealso{
Other Exported Constructor Functions: \code{\link{.SummarizedExperiment}},
  \code{\link{.arrangeMetadataByPriorityCols}},
  \code{\link{.interestingColData}},
  \code{\link{.interestingGroups}},
  \code{\link{.interestingGroup}},
  \code{\link{.sampleYAMLMetrics}},
  \code{\link{.sampleYAML}},
  \code{\link{.setMetadataFactors}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
