% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-sampleYAMLMetrics.R
\name{.sampleYAMLMetrics}
\alias{.sampleYAMLMetrics}
\title{Sample Metrics from YAML File}
\usage{
.sampleYAMLMetrics(yaml, characterCols)
}
\arguments{
\item{yaml}{YAML \link{list}.}

\item{characterCols}{Columns to set as character.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metrics from YAML File
}
\seealso{
Other Exported Constructor Functions: \code{\link{.SummarizedExperiment}},
  \code{\link{.interestingColData}},
  \code{\link{.interestingGroups}},
  \code{\link{.interestingGroup}},
  \code{\link{.sampleMetadata}},
  \code{\link{.sampleYAMLMetadata}},
  \code{\link{.sampleYAML}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
