% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-setMetadataFactors.R
\name{.setMetadataFactors}
\alias{.setMetadataFactors}
\title{Set Metadata Columns as Factors}
\usage{
.setMetadataFactors(meta)
}
\arguments{
\item{meta}{Metadata \link{data.frame}.}
}
\value{
\link{data.frame}.
}
\description{
Set Metadata Columns as Factors
}
\seealso{
Other Exported Constructor Functions: \code{\link{.SummarizedExperiment}},
  \code{\link{.arrangeMetadataByPriorityCols}},
  \code{\link{.interestingColData}},
  \code{\link{.interestingGroups}},
  \code{\link{.interestingGroup}},
  \code{\link{.sampleYAMLMetadata}},
  \code{\link{.sampleYAMLMetrics}},
  \code{\link{.sampleYAML}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
