% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{dotNotation}
\alias{dotNotation}
\title{Dot notation}
\usage{
dotNotation(string)
}
\arguments{
\item{string}{String}
}
\value{
String with dot notation formatting
}
\description{
Dot notation
}
\examples{
dotNotation("RNAi clone")
}
\author{
Michael Steinbaugh
}
