% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots}
\alias{dots}
\alias{get_objs_from_dots}
\alias{getObjsFromDots}
\title{Objects as dots utilities.}
\usage{
dots(...)

getObjsFromDots(.dots)
}
\arguments{
\item{...}{Object as ellipsis.}

\item{.dots}{Objects as dots.}
}
\description{
Handle object names as dots instead of character vector.
}
\seealso{
\code{browseVignettes(package = "dplyr")}.
}
