% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-dotted.R
\docType{methods}
\name{dotted}
\alias{dotted}
\alias{dotted,ANY-method}
\alias{dotted,character-method}
\alias{dotted,data.frame-method}
\alias{dotted,list-method}
\alias{dotted,matrix-method}
\alias{dotted,tbl_df-method}
\title{Dotted Case}
\usage{
dotted(object, ...)

\S4method{dotted}{ANY}(object, rownames = FALSE)

\S4method{dotted}{character}(object)

\S4method{dotted}{data.frame}(object, rownames = FALSE)

\S4method{dotted}{list}(object)

\S4method{dotted}{matrix}(object, rownames = FALSE)

\S4method{dotted}{tbl_df}(object)
}
\arguments{
\item{object}{Character vector or an object for which \code{\link[=names]{names()}} assignment
will be meaningful.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{rownames}{Apply sanitization on row names. This is not recommended
by default, since rownames commonly contain gene identifiers that should
not be modified.}
}
\value{
Object with syntatically valid names. For objects supporting
\code{\link[base:names]{base::names()}}, the underlying data returns unchanged.
}
\description{
Dotted Case
}
\details{
For unnamed character vectors, this function will sanitize the
underlying values. Otherwise, it will set \code{\link[=names]{names()}} and/or \code{\link[=rownames]{rownames()}} on
objects supporting name assignments. They return the object without
modification of the underlying data.
}
\note{
\code{dotted.case} support is provided for matching against base R
parameters, but we strongly advise against using it for object and/or
argument name assignments.
}
\examples{
loadRemoteData(file.path(testDataURL, "makeNames.rda"))

# Character vector
print(makeNames$vec)
dotted(makeNames$vec)

# data.frame
print(makeNames$df)
dotted(makeNames$df, rownames = TRUE)
}
\concept{Make Names Utilities}
