% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFile.R
\name{downloadFile}
\alias{downloadFile}
\title{Match and download a file from a remote directory}
\usage{
downloadFile(remoteDir, string, rename = NULL, localDir = "data-raw")
}
\arguments{
\item{remoteDir}{Remote directory URL}

\item{rename}{Rename the file, if desired}

\item{localDir}{Directory where to save file locally}

\item{subset}{Match \code{string}}
}
\description{
Match and download a file from a remote directory
}
\keyword{web}
