% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/EggNOG.R
\docType{class}
\name{EggNOG-class}
\alias{EggNOG-class}
\alias{EggNOG}
\title{EggNOG database annotations}
\usage{
EggNOG()
}
\value{
\code{EggNOG}.
This class extends \code{list} and contains:
\enumerate{
\item "\code{cogFunctionalCategories}": \strong{C}luster of \strong{O}rthologous \strong{G}roups
(COG) functional category information.
\item "\code{annotations}": up-to-date functional descriptions and categories
for \strong{Eu}karyotes \strong{N}on-supervised \strong{O}rthologous \strong{G}roups (euNOG)
and \strong{N}on-supervised \strong{O}rthologous \strong{G}roups (NOG) protein
identifiers.
}

The \href{http://eggnogdb.embl.de/download/latest/README.txt}{EggNOG README}
contains additional useful reference information.
}
\description{
\href{http://eggnogdb.embl.de}{EggNOG} is a database of biological information
hosted by the EMBL. It is based on the original idea of COGs (\strong{c}lusters of
\strong{o}rthologous \strong{g}roups) and expands that idea to non-supervised
orthologous groups constructed from numerous organisms. eggNOG stands for
\strong{e}volutionary \strong{g}enealogy of \strong{g}enes: \strong{N}on-supervised
\strong{O}rthologous \strong{G}roups.
}
\examples{
options(acid.test = TRUE)
x <- EggNOG()
print(x)
}
\author{
Michael Steinbaugh
}
