% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EggNOG.R
\name{EggNOG}
\alias{EggNOG}
\title{EggNOG Database Annotations}
\usage{
EggNOG()
}
\value{
\code{EggNOG}.
}
\description{
\href{http://eggnogdb.embl.de}{EggNOG} is a database of biological information
hosted by the EMBL. It is based on the original idea of COGs (\strong{c}lusters of
\strong{o}rthologous \strong{g}roups) and expands that idea to non-supervised
orthologous groups constructed from numerous organisms. eggNOG stands for
\strong{e}volutionary \strong{g}enealogy of \strong{g}enes: \strong{N}on-supervised
\strong{O}rthologous \strong{G}roups.
}
\examples{
options(basejump.test = TRUE)
x <- EggNOG()
print(x)
}
\seealso{
\code{\link[=EggNOG]{EggNOG()}}.

The \href{http://eggnogdb.embl.de/download/latest/README.txt}{EggNOG README file}
contains additional useful reference information.

Other S4 classes: \code{\link{Ensembl2Entrez-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
