% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyAsNA.R
\name{emptyAsNA}
\alias{emptyAsNA}
\title{Convert empty values to NA.}
\usage{
emptyAsNA(x)
}
\arguments{
\item{x}{Values missing \code{NA}.}
}
\value{
Values containing \code{NA}.
}
\description{
Convert empty values to NA.
}

