% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyRanges.R
\name{emptyRanges}
\alias{emptyRanges}
\title{Generate Empty Genomic Ranges}
\usage{
emptyRanges(names, seqname = c("unknown", "transgene", "spike"),
  mcolsNames = NULL)
}
\arguments{
\item{names}{\emph{Required.} \code{character} vector of gene or transcript names.}

\item{seqname}{\emph{Required.} \code{character} string that defines the alternative
chromosome name where these ranges will be grouped. Defaults to "\code{unknown}"
but "\code{transgene}" (for transgenes) and "\code{spike}" (for spike-ins) are also
supported.}

\item{mcolsNames}{\emph{Optional.} Metadata columns to be defined in the \code{mcols()}
of the \code{GRanges} object. Normally this does not need to be defined; useful
when combining with another defined \code{GRanges} object containing metadata.}
}
\value{
\code{GRanges}.
}
\description{
Utility function that provides support for creating internal \code{GRanges} for
transgene and FASTA spike-in sequences.
}
\examples{
# Unknown/dead genes
emptyRanges("ENSG00000000000", seqname = "unknown")

# Transgenes
emptyRanges(c("EGFP", "TDTOMATO", "GAL4"), seqname = "transgene")

# Spike-ins
emptyRanges("ERCC", seqname = "spike")
}
\seealso{
\itemize{
\item \code{help("seqinfo", "GenomeInfoDb")}.
\item \code{DESeq2::makeExampleDESeqDataSet()}.
}

Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}}, \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
