% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode-methods.R
\docType{methods}
\name{encode}
\alias{encode}
\alias{encode,DataFrame-method}
\title{Encode using run-length encoding}
\usage{
\S4method{encode}{DataFrame}(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Modified object.
All \code{atomic} columns will be encoded to \code{Rle} S4 class.
}
\description{
Encode using run-length encoding
}
\examples{
binary <- seq(from = 0L, to = 1L)
df <- DataFrame(
    a = rep(x = binary, times = 50L),
    b = rep(x = binary, each = 50L)
)
lapply(df, class)
x <- encode(df)
lapply(x, class)
}
\seealso{
\code{S4Vectors::Rle()}.
}
