% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichDavid.R
\name{enrichDavid}
\alias{enrichDavid}
\title{DAVID enrichment analysis}
\usage{
enrichDavid(foreground, background = NULL, idType = "ENSEMBL_GENE_ID",
  saveFiles = TRUE, saveDir = "results/david", count = 3, fdr = 0.1)
}
\arguments{
\item{foreground}{Foreground identifier vector}

\item{background}{Background identifier vector}

\item{idType}{Identifier type (see DAVID website)}

\item{saveFiles}{Save files to disk (\code{TRUE} or \code{FALSE})}

\item{saveDir}{Directory where to save TSV files}

\item{count}{Minimum hit count}

\item{fdr}{False discovery rate cutoff (alpha)}
}
\value{
List of \code{RDAVIDWebService()} report objects
}
\description{
Wrapper function that performs gene set enrichment analysis (GSEA) with
RDAVIDWebService, using simplified input options.
}
\author{
Michael Steinbaugh
}
\keyword{enrichment}
\keyword{gsea}
\keyword{rnaseq}
