% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichStringdb.R
\name{enrichStringdb}
\alias{enrichStringdb}
\title{STRINGdb enrichment analysis}
\usage{
enrichStringdb(gene, background, species, cache = TRUE, csv = FALSE,
  png = TRUE, pdf = FALSE, version = "10")
}
\arguments{
\item{gene}{Gene identifier vector}

\item{background}{Background identifier vector}

\item{species}{Species (e.g. "Homo sapiens")}

\item{cache}{Cache files to disk (TRUE/FALSE)}

\item{csv}{Export to CSV (TRUE/FALSE)}

\item{png}{Export network as PNG file (TRUE/FALSE)}

\item{pdf}{Export network as high resolution PDF file (TRUE/FALSE)}

\item{version}{STRING version to use}
}
\value{
STRINGdb output list
}
\description{
Wrapper function that performs gene set enrichment analysis (GSEA) with
STRINGdb, using simplified input options.
}
