% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ensembl2Entrez.R
\docType{methods}
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{Ensembl2Entrez,DataFrame-method}
\alias{Ensembl2Entrez,GRanges-method}
\alias{Ensembl2Entrez,SummarizedExperiment-method}
\title{Ensembl-to-Entrez Gene Identifier Mappings}
\usage{
Ensembl2Entrez(object, ...)

\S4method{Ensembl2Entrez}{DataFrame}(object, format = c("1:1", "long",
  "list"))

\S4method{Ensembl2Entrez}{GRanges}(object, format = c("1:1", "long",
  "list"))

\S4method{Ensembl2Entrez}{SummarizedExperiment}(object, format = c("1:1",
  "long", "list"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{format}{\code{string}. Formatting method to apply:
\itemize{
\item \code{"1:1"}: Return with 1:1 mappings. For Ensembl genes that don't map 1:1
with Entrez, pick the oldest Entrez ID. Genes that don't map to Entrez
will contain \code{NA} in \code{entrezID} column.
\item \code{"long"}: Return 1:many in long format.
\item \code{"list"}: Return 1:many with \code{entrezID} as a \code{list} column.
}}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Defines 1:1 mappings from Ensembl gene IDs to Entrez IDs. Uses the oldest
Entrez ID if there are multiple identifiers that map to an Ensembl gene ID.
}
\examples{
data(rse_small)
x <- Ensembl2Entrez(rse_small)
print(x)
}
\seealso{
\code{\link[=Ensembl2Entrez]{Ensembl2Entrez()}}.

Other S4 classes: \code{\link{EggNOG-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
