% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblAnnotations.R
\name{ensemblAnnotations}
\alias{ensemblAnnotations}
\title{Ensembl Annotations}
\usage{
ensemblAnnotations(organism, format = "genes", genomeBuild = NULL,
  release = NULL, return = "GRanges")
}
\arguments{
\item{organism}{Default recommended usage is to provide the full latin
organism name as a string.}

\item{format}{Desired annotation data format, either "\code{genes}",
"\code{transcripts}", "\code{gene2symbol}"., or "\code{tx2gene}".}

\item{genomeBuild}{\emph{Optional.} Check to see if a specific genome build is
supported.}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{return}{Class of the returned object. Can be "\code{data.frame}",
"\code{DataFrame}" or "\code{GRanges}. See \code{help("genes", "ensembldb")} for
additional information.}
}
\value{
Gene or transcript annotations.
}
\description{
Quickly obtain gene and transcript annotations from
\href{http://www.ensembl.org/}{Ensembl} using \link{AnnotationHub}.
}
\section{Using AnnotationHub}{

The default recommended approach is to simply specify the desired organism,
using the full latin name. For example, we can obtain human annotations with
\code{Homo sapiens} and mouse annotations with \code{Mus musculus}. Under the hood, the
function fetches these annotations from AnnotationHub using the ensembldb
package. AnnotationHub supports versioned Ensembl releases, back to version
87.
}

\examples{
ensemblAnnotations("Homo sapiens") \%>\% glimpse()
}
\seealso{
\itemize{
\item \href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
\item \href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.
}

Other Gene Annotation Utilities: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
