% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ensemblVersion.R
\docType{methods}
\name{ensemblVersion}
\alias{ensemblVersion}
\alias{ensemblVersion}
\alias{ensemblVersion,missing-method}
\title{Ensembl Build Version}
\usage{
ensemblVersion(object, ...)

\S4method{ensemblVersion}{missing}()
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
String containing Ensembl build version of annotables.
}
\description{
Ensembl Build Version
}
\examples{
ensemblVersion()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}}, \code{\link{gene2entrez}},
  \code{\link{gene2symbol}}, \code{\link{tx2geneFromGTF}},
  \code{\link{tx2gene}}
}
\keyword{internal}
