% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Entrez2Ensembl-class}
\alias{Entrez2Ensembl-class}
\title{Entrez-to-Ensembl gene identifier mappings}
\value{
\code{Entrez2Ensembl}.
}
\description{
Entrez-to-Ensembl gene identifier mappings
}
\details{
Contains a \code{DataFrame} with \code{ensembl} and \code{entrez} columns.
}
\note{
Updated 2020-10-05.
}
