% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonicMappingRate.R
\name{exonicMappingRate}
\alias{exonicMappingRate}
\title{Exonic mapping rate plot}
\usage{
exonicMappingRate(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Exonic mapping rate plot
}
\examples{
\dontrun{
exonicMappingRate(summary)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
