% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,ANY-method}
\alias{export,sparseMatrix-method}
\alias{export,SummarizedExperiment-method}
\alias{export,SingleCellExperiment-method}
\title{Export}
\usage{
export(x, ...)

\S4method{export}{ANY}(x, file, format, ...)

\S4method{export}{sparseMatrix}(x, file, format)

\S4method{export}{SummarizedExperiment}(x, name = NULL, dir = ".",
  compress = FALSE, human = FALSE, assays = TRUE, rowData = TRUE,
  colData = TRUE)

\S4method{export}{SingleCellExperiment}(x, name = NULL, dir = ".",
  compress = FALSE, human = FALSE, assays = TRUE, rowData = TRUE,
  colData = TRUE)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{file}{\code{string}. File path.}

\item{format}{An optional character string containing the file format, which can be used to override the format inferred from \code{file} or, in lieu of specifying \code{file}, a file with the symbol name of \code{x} and the specified file extension will be created. Must specify \code{file} and/or \code{format}. Shortcuts include: \dQuote{,} (for comma-separated values), \dQuote{;} (for semicolon-separated values), \dQuote{|} (for pipe-separated values), and \dQuote{dump} for \code{\link[base]{dump}}.}

\item{name}{\code{string}. Name to use on disk. If left \code{NULL}, will use the name
of the object instead.}

\item{dir}{\code{string}. Directory path.}

\item{compress}{\code{boolean}. Apply gzip compression to all files.}

\item{human}{\code{boolean}. Automatically convert gene IDs to gene symbols in the
\code{rownames()} and sample IDs to sample names in the \code{colnames()}.}

\item{assays}{\code{boolean}. Write \code{assays()} to disk. All assays in object
will be written into an \code{assays/} subdirectory.}

\item{rowData}{\code{boolean}. Write \code{rowData()} to disk.}

\item{colData}{\code{boolean}. Write \code{colData()} to disk.}
}
\value{
Invisible \code{character}. File path(s).
}
\description{
Export data out of R and write to disk.
}
\details{
This is a wrapper for \code{rio::export()} that adds support for additional S4
classes in Bioconductor.
}
\section{Row names}{


The standard \code{rio::export()} call will drop rownames when exporting to CSV.
The \code{readr::write_*()} family of functions also never write rownames. This is
a \emph{really poor} default setting for handling genomic data, which often
contain gene identifiers in the rownames. Here we're performing any internal
tibble coercion step to ensure rownames are always moved to a "\code{rowname}"
column in the CSV export.
}

\examples{
library(SummarizedExperiment)
library(SingleCellExperiment)
data(rse, sce)

## matrix ====
rnaseq_counts <- assay(rse)
export(rnaseq_counts, format = "csv")

## sparseMatrix ====
single_cell_counts <- assay(sce)
export(single_cell_counts, format = "mtx")

## Clean up.
file.remove(c(
    "rnaseq_counts.csv",
    "single_cell_counts.mtx",
    "single_cell_counts.mtx.colnames",
    "single_cell_counts.mtx.rownames"
))
}
\seealso{
\code{rio::export()}.
}
