% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCsv.R
\name{exportCsv}
\alias{exportCsv}
\title{Export CSV file}
\usage{
exportCsv(..., dir = "results")
}
\arguments{
\item{...}{Unquoted names of existing objects to export}

\item{dir}{Target directory}
}
\description{
Quickly export a comma separated values (CSV) file to the \code{results}
folder.
}
\examples{
\dontrun{
x <- data.frame(col1 = 1:10, col2 = 11:20)
y <- data.frame(col3 = 21:30, col4 = 31:40)
exportCsv(x, y)
}
}
