% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-externalFile.R
\docType{methods}
\name{externalFile}
\alias{externalFile}
\alias{externalFile}
\alias{externalFile,missing-method}
\alias{externalFile,character-method}
\title{Download External Dependency File}
\usage{
externalFile(object, ...)

\S4method{externalFile}{missing}(object, package)

\S4method{externalFile}{character}(object, package)
}
\arguments{
\item{object}{\emph{Optional}. File name. If \code{NULL} (default), download the
default dependency files for a new experiment.}

\item{...}{Additional arguments.}

\item{package}{Package name.}
}
\value{
No value.
}
\description{
If the required dependency file isn't present, download latest version from
the package website.
}
\details{
File download utility for RMarkdown knit reports.
}
\examples{
\dontrun{
externalFile("setup.R", package = "bcbioSinglecell")
}
}
\seealso{
Other Infrastructure Utilities: \code{\link{detectHPC}}
}
