% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{extPattern}
\alias{extPattern}
\title{File extension pattern
Note optional matching of gzip.}
\format{An object of class \code{character} of length 1.}
\usage{
extPattern
}
\description{
File extension pattern
Note optional matching of gzip.
}
\examples{
extPattern
}
\keyword{datasets}
