% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fixNA-methods.R
\docType{methods}
\name{fixNA}
\alias{fixNA}
\alias{fixNA}
\alias{fixNA,ANY-method}
\alias{fixNA,character-method}
\title{Fix Character Strings Missing \code{NA}}
\usage{
fixNA(object)

\S4method{fixNA}{ANY}(object)

\S4method{fixNA}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Object containing proper \code{NA} values.
}
\description{
Fix Character Strings Missing \code{NA}
}
\examples{
fixNA(c(1, "x", "", "NA"))
}
\seealso{
Other Cleanup Utilities: \code{\link{camel}},
  \code{\link{removeNA}}, \code{\link{sortUnique}},
  \code{\link{wash}}
}
