% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/flatFiles-methods.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{flatFiles,SummarizedExperiment-method}
\title{Flat Files from S4 Object}
\usage{
flatFiles(object, ...)

\S4method{flatFiles}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list} containing the S4 object slots.
}
\description{
Extract the slots inside an S4 object for archival storage.
}
\examples{
# SummarizedExperiment ====
x <- flatFiles(rse_dds)
class(x)
names(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}}, \code{\link{counts}},
  \code{\link{gene2symbol}}, \code{\link{sampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
