% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{formalsList}
\alias{formalsList}
\title{Shared list of optional default formals}
\format{
An object of class \code{list} of length 12.
}
\usage{
formalsList
}
\description{
Shared list of optional default formals
}
\note{
Updated 2019-11-19.
}
\examples{
formalsList
}
\keyword{datasets}
