% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather-methods.R
\docType{methods}
\name{gather}
\alias{gather}
\alias{gather,matrix-method}
\alias{gather,data.frame-method}
\alias{gather,SummarizedExperiment-method}
\alias{gather,SingleCellExperiment-method}
\title{Gather columns into key-value pairs}
\usage{
gather(object, ...)

\S4method{gather}{matrix}(object, min = NULL, minMethod = c("perRow",
  "absolute"), trans = c("identity", "log2", "log10"))

\S4method{gather}{data.frame}(object, ...)

\S4method{gather}{SummarizedExperiment}(object, assay = 1L, min = NULL,
  minMethod = c("perRow", "absolute"), trans = c("identity", "log2",
  "log10"))

\S4method{gather}{SingleCellExperiment}(object, assay = 1L, min = NULL,
  minMethod = c("perRow", "absolute"), trans = c("identity", "log2",
  "log10"))
}
\arguments{
\item{object}{Object.}

\item{min}{\code{integer(1)} or \code{NULL}.
Minimum count threshold to apply. Filters using "greater than or equal to"
logic internally. Note that this threshold gets applied prior to
logarithmic transformation, when \code{trans} argument applies.}

\item{minMethod}{\code{character(1)}.
Uses \code{\link[base:match.arg]{match.arg()}}.
\itemize{
\item \code{perRow}: \emph{Recommended}. Applies cutoff per row (i.e. gene).
Internally, \code{\link[base:rowSums]{rowSums()}} values are checked against this
cutoff threshold prior to the melt operation.
\item \code{absolute}: Applies hard cutoff to \code{counts} column after the melt
operation. This applies to all counts, not per feature.
}}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}

\item{...}{Additional arguments.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}
}
\value{
\code{DataFrame}.
}
\description{
Gather columns into key-value pairs
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Passes to \code{tidyr::gather()}.
}}

\note{
Updated 2019-08-26.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment
dim(rse)
x <- gather(rse, min = NULL)
nrow(x)
print(x)
}
\seealso{
tidyr (recommended):\preformatted{methods("gather")
methods("gather_")
getS3method("gather", "data.frame", envir = asNamespace("tidyr"))
getS3method("gather_", "data.frame", envir = asNamespace("tidyr"))
}

reshape2 (deprecated):\preformatted{help(topic = "melt.array", package = "reshape2")
methods("melt")
getS3method("melt", "data.array", envir = asNamespace("tidyr"))
getS3method("melt", "data.frame", envir = asNamespace("tidyr"))
}
}
