% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_markers}
\alias{gender_markers}
\title{Sexually Dimorphic Gender Marker Genes}
\format{An object of class \code{list} of length 2.}
\usage{
gender_markers
}
\description{
\emph{Homo sapiens} and \emph{Mus musculus} are currently defined.
}
\examples{
names(gender_markers)
}
\author{
Michael Steinbaugh
}
\keyword{internal}
