% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderMarkers.R
\name{genderMarkers}
\alias{genderMarkers}
\title{Graph sexually dimorphic marker genes}
\usage{
genderMarkers(counts, organism)
}
\arguments{
\item{counts}{Counts matrix. Transcripts per million (TPM) or normalized
counts are preferred.}

\item{organism}{Organism. \code{hsapiens} and \code{mmusculus} are supported.}
}
\value{
Scatterplot
}
\description{
Graph sexually dimorphic marker genes
}
\examples{
\dontrun{
genderMarkers(tpm, "mmusculus")
}
}
\author{
Michael Steinbaugh
}
