% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2entrez.R
\docType{methods}
\name{gene2entrez}
\alias{gene2entrez}
\alias{gene2entrez,character-method}
\title{Convert Ensembl Identifier to Entrez}
\usage{
gene2entrez(object, ...)

\S4method{gene2entrez}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Grouped \link{tibble}.
}
\description{
Convert Ensembl Identifier to Entrez
}
\examples{
gene2entrez("grcm38")
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}},
  \code{\link{ensemblVersion}}, \code{\link{gene2symbol}},
  \code{\link{tx2geneFromGTF}}, \code{\link{tx2gene}}
}
