% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Gene2Symbol-methods.R
\docType{methods}
\name{Gene2Symbol}
\alias{Gene2Symbol}
\alias{Gene2Symbol,DataFrame-method}
\alias{Gene2Symbol,GRanges-method}
\alias{Gene2Symbol,SummarizedExperiment-method}
\title{Gene-to-Symbol Mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{DataFrame}(object, format = c("makeUnique", "1:1",
  "long"))

\S4method{Gene2Symbol}{GRanges}(object, format = c("makeUnique", "1:1",
  "long"))

\S4method{Gene2Symbol}{SummarizedExperiment}(object,
  format = c("makeUnique", "1:1", "long"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{format}{\code{string}. Formatting method to apply:
\itemize{
\item \code{"makeUnique"}: \emph{Recommended.} Apply \code{\link[base:make.unique]{base::make.unique()}} to the
\code{geneName} column. Gene symbols are made unique, while the gene IDs remain
unmodified.
\item \code{"1:1"}: For gene symbols that map to multiple gene IDs, select only the
first annotated gene ID.
\item \code{"long"}: Return \code{geneID} and \code{geneName} columns unmodified in long format.
}}
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-Symbol Mappings
}
\note{
For some organisms, gene names and gene symbols do not map 1:1
(e.g. \emph{Homo sapiens} and \emph{Mus musculus}). Refer to the \code{format} argument here
in the documentation for approaches that deal with this issue.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:metadata]{S4Vectors::metadata()}}.
}

\examples{
data(rse_small)
x <- Gene2Symbol(rse_small)
print(x)
}
\seealso{
\code{\link[=Gene2Symbol]{Gene2Symbol()}}, \link{makeGene2Symbol}.

Other S4 classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
