% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbolFromGFF.R
\docType{methods}
\name{gene2symbolFromGFF}
\alias{gene2symbolFromGFF}
\alias{gene2symbolFromGFF}
\alias{gene2symbolFromGFF,character-method}
\alias{gene2symbolFromGFF,data.frame-method}
\alias{gene2symbolFromGTF}
\title{Define Gene to Symbol Mappings from GFF/GTF File}
\usage{
gene2symbolFromGFF(object, ...)

\S4method{gene2symbolFromGFF}{character}(object, uniqueSymbol = FALSE)

\S4method{gene2symbolFromGFF}{data.frame}(object, uniqueSymbol = FALSE)

gene2symbolFromGTF(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{uniqueSymbol}{Make gene symbols unique.}
}
\value{
\link{data.frame}.
}
\description{
Define Gene to Symbol Mappings from GFF/GTF File
}
\details{
The GFF (General Feature Format) format consists of one line per
feature, each containing 9 columns of data, plus optional track definition
lines. The GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
# From URL (recommended)
url <- "http://basejump.seq.cloud/mmusculus.gtf"
gene2symbolFromGFF(url) \%>\% glimpse()

# GFF data.frame
gff <- readGFF(url)
gene2symbolFromGFF(gff) \%>\% glimpse()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{gene2symbol}},
  \code{\link{tx2geneFromGFF}}, \code{\link{tx2gene}}
}
