% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gene2symbolFromGTF}
\alias{gene2symbolFromGTF}
\title{Generate gene2symbol from GTF File}
\usage{
gene2symbolFromGTF(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Generate gene2symbol from GTF File
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}}, \code{\link{gene2symbol}},
  \code{\link{tx2geneFromGTF}}, \code{\link{tx2gene}}
}
