% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDetectionSaturation.R
\name{geneDetectionSaturation}
\alias{geneDetectionSaturation}
\title{Genes detection saturation plot}
\usage{
geneDetectionSaturation(counts, summary)
}
\arguments{
\item{counts}{Counts matrix}

\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Genes detection saturation plot
}
\examples{
\dontrun{
geneDetectionSaturation(counts, summary)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
