% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneNames-methods.R
\docType{methods}
\name{geneNames}
\alias{geneNames}
\alias{geneNames,GRanges-method}
\alias{geneNames,SummarizedExperiment-method}
\title{Gene names}
\usage{
geneNames(object, ...)

\S4method{geneNames}{GRanges}(object)

\S4method{geneNames}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Convenience function for quickly returning unique gene symbols that map to
the stable gene IDs.
}
\note{
Updated 2019-07-28.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
head(geneNames(rse))
}
