% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}

\item{envir}{\code{environment} to use for assignment. Defaults to
\code{parent.frame()}, which will assign into the calling environment.}

\item{file}{\code{string}. File path.}

\item{headerLevel}{\code{scalar integer}. Markdown header level (1-7).}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{plotlist}{\code{list} containing \code{ggplot} objects.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\description{
General Arguments
}
\author{
Michael Steinbaugh
}
\keyword{internal}
