% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{arrange}
\alias{arrange}
\alias{as_tibble}
\alias{filter}
\alias{mutate}
\title{S4 generics}
\usage{
arrange(object, ...)

as_tibble(object, ...)

filter(object, ...)

mutate(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\description{
S4 generics
}
