% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesDetected.R
\name{genesDetected}
\alias{genesDetected}
\title{Genes detected plot}
\usage{
genesDetected(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Genes detected plot
}
\examples{
\dontrun{
genesDetected(summary)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
