% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSynonyms.R
\name{geneSynonyms}
\alias{geneSynonyms}
\title{Gene synonyms}
\usage{
geneSynonyms(organism)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}
}
\value{
\code{SplitDataFrameList}.
Split by \code{geneID} column.
Handles genes with duplicate entries (e.g. ENSG00000004866).
}
\description{
Look up gene synonyms from NCBI.
}
\note{
Synonym support for \emph{Caenorhabditis elegans} is poor on NCBI.
Use the \href{https://wormbase.acidgenomics.com/}{wormbase} package instead.

Updated 2020-05-11.
}
\examples{
options(acid.test = TRUE)
x <- geneSynonyms(organism = "Homo sapiens")
print(x)
}
