% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeVersions.R
\name{genomeVersions}
\alias{genomeVersions}
\alias{ensemblVersion}
\alias{flybaseVersion}
\alias{gencodeVersion}
\alias{refseqVersion}
\alias{wormbaseVersion}
\title{Genome versions}
\usage{
ensemblVersion()

flybaseVersion(dmel = FALSE)

gencodeVersion(organism)

refseqVersion()

wormbaseVersion()
}
\arguments{
\item{dmel}{\code{logical(1)}.
Return \emph{Drosophila melanogaster} genome release version.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}
}
\value{
\code{character(1)}.
Release version.
}
\description{
Obtain the latest release version from various genome annotation sources.
}
\note{
Updated 2019-10-18.
}
\examples{
if (hasInternet()) {
    ensemblVersion()
    flybaseVersion()
    gencodeVersion(organism = "Homo sapiens")
    refseqVersion()
    wormbaseVersion()
}
}
\seealso{
Refer to the koopa package for shell variants:
\itemize{
\item \code{ensembl-version}.
\item \code{flybase-version}.
\item \code{gencode-version}.
\item \code{refseq-version}.
\item \code{wormbase-version}.
}
}
