% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicMappingRate.R
\name{genomicMappingRate}
\alias{genomicMappingRate}
\title{Genomic mapping rate plot}
\usage{
genomicMappingRate(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Genomic mapping rate plot
}
\examples{
\dontrun{
genomicMappingRate(summary)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
