% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{geomean}
\alias{geomean}
\title{Geometric Mean}
\usage{
geomean(x)
}
\arguments{
\item{x}{Vector or column data (data frame, matrix).}
}
\value{
Geometric means.
}
\description{
The geometric mean is the nth root of n products or e to the mean log of \code{x}.
Useful for describing non-normal (i.e. geometric) distributions.
}
\details{
Modified version of \code{psych::geometric.mean()}.
}
\note{
Not particularly useful if there are elements that are <= 0.
}
\examples{
# Vector
vec <- seq(1L, 5L, 1L)
geomean(vec)

vec2 <- vec ^ 2L
geomean(vec2)

# Data frame
df <- data.frame(vec, vec2)
geomean(df)
}
\seealso{
Other Math Utilities: \code{\link{fc2lr}},
  \code{\link{pct}}
}
