% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsDb.R
\name{getEnsDb}
\alias{getEnsDb}
\title{Get EnsDb from AnnotationHub}
\usage{
getEnsDb(organism, genomeBuild = NULL, release = NULL)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). If set \code{NULL}, defaults to the most
recent release available.}
}
\value{
\code{EnsDb}.
}
\description{
Get EnsDb from AnnotationHub
}
\note{
Updated 2020-09-24.
}
\examples{
edb <- getEnsDbFromAnnotationHub(
    organism = "Homo sapiens",
    release = 100L,
    build = "GRCh38"
)
print(edb)
}
