% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{ggplot2-geoms}
\alias{ggplot2-geoms}
\alias{basejump_geom_abline}
\alias{basejump_geom_label}
\alias{basejump_geom_label_average}
\alias{basejump_geom_label_repel}
\title{ggplot2 geometric objects}
\usage{
basejump_geom_abline(xintercept = NULL, yintercept = NULL)

basejump_geom_label(data = NULL, mapping = NULL, ...)

basejump_geom_label_average(data, col, fun = c("mean", "median"),
  digits = 0L, ...)

basejump_geom_label_repel(data = NULL, mapping = NULL, color = NULL,
  size = 4L, ...)
}
\arguments{
\item{xintercept, yintercept}{\code{numeric(1)}.
Value denoting x- or y-axis cutoff. Specify one but not both.}

\item{data}{\code{data.frame}.
Data.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{col}{\code{character(1)}.
Column name.}

\item{fun}{\code{character(1)}.
Function name to use for average calculation. Currently supports "\code{mean}"
or "\code{median}".}

\item{digits}{\code{integer(1)}.
Number of significant digits to use. Defaults to rounded.}

\item{color}{\code{character(1)}.
Text color (e.g. \code{"orange"}).}

\item{size}{\code{integer(1)}.
Font size.}
}
\value{
\code{ggproto}.
}
\description{
Convenience functions with modified defaults for
\href{http://ggplot2.org}{ggplot2}.
}
\section{basejump_geom_abline}{


Horizontal or vertical cutoff line.
}

\section{basejump_geom_label}{


Modified version of \code{\link[ggplot2:geom_label]{ggplot2::geom_label()}}.
}

\section{basejump_geom_label_average}{


Add average labels to a plot. For example, \code{col} can be \code{nGene}. Median or
mean values are always calculated per sample (\code{sampleName}).
}

\section{basejump_geom_label_repel}{


Repulsive textual annotations. Modified basejump version of
\code{\link[ggrepel:geom_label_repel]{ggrepel::geom_label_repel()}}. If advanced customization of the text labels
is required, simply use the ggrepel version instead.
}

\examples{
## basejump_geom_abline ====
## x-axis line
geom <- basejump_geom_abline(xintercept = 1L)
geom

## y-axis line
geom <- basejump_geom_abline(yintercept = 1L)
geom

## basejump_geom_label ====
geom <- basejump_geom_label()
geom

## basejump_geom_label_average ====
geom <- basejump_geom_label_average(
    data = tibble::tibble(
        sampleName = rep(c("sample1", "sample2"), times = 4L),
        counts = seq_len(8L)
    ),
    col = "counts",
    fun = "mean"
)
geom

## basejump_geom_label_repel ====
geom <- basejump_geom_label_repel()
geom
}
\seealso{
\itemize{
\item \code{ggplot2::geom_label()}.
\item \code{ggrepel::geom_label_repel()}.
}
}
