% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.R
\name{ggplot2_themes}
\alias{ggplot2_themes}
\alias{theme_paperwhite}
\alias{theme_midnight}
\title{ggplot2 Themes}
\usage{
theme_paperwhite(base_size = 14L, base_family = "", face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("right", "bottom",
  "top", "none"), grid = FALSE, minimal = FALSE)

theme_midnight(base_size = 14L, base_family = "", face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("right", "bottom",
  "top", "none"), grid = FALSE, minimal = FALSE)
}
\arguments{
\item{base_size}{\code{scalar numeric}. Base font size.}

\item{base_family}{\code{string}. Base font family.}

\item{face}{\code{string}. Font face ("\code{bold}", "\code{plain}").}

\item{aspect_ratio}{\code{scalar numeric}. Aspect ratio, specifying the plot
proportions. Use \code{1} for a perfectly square plot (including the axis
labels).}

\item{legend_position}{\code{string}. Legend key position. We're being a little
more restrictive here, only allowing "\code{bottom}", "\code{right}", or "\code{none}".
Including the legend at the top or the left side of the plot rarely makes
sense and is discouraged.}

\item{grid}{\code{boolean}. Label the major panel grids with a gray accent.}

\item{minimal}{\code{boolean}. Remove all axis lines, axis ticks, and
panel borders.}
}
\value{
\code{theme}.
}
\description{
Complete \href{http://ggplot2.tidyverse.org}{ggplot2} themes.
}
\details{
All themes are based off of \code{\link[ggplot2:theme_linedraw]{ggplot2::theme_linedraw()}}, but with
modifications and extra user-definable parameters.
}
\section{Paperwhite}{


High contrast black and white theme optimized for print. Recommended for
scientific manuscripts and website tutorials.
}

\section{Midnight}{


Blackout theme that sets the plot background as black, with white text.
Inspired by \code{Seurat::DarkTheme()}, with some color modifications. Useful
for visualizing many points with a high dynamic color range, such as t-SNE
expression plots.
}

\examples{
library(ggplot2)
p <- ggplot(
    data = mpg,
    mapping = aes(
        x = manufacturer,
        y = displ,
        color = manufacturer,
        fill = manufacturer
    )
)

# Paperwhite theme.
p +
    geom_point() +
    theme_paperwhite(legend_position = "none")

# Midnight theme.
p +
    geom_point() +
    theme_midnight(legend_position = "none")

# Minimal bar plot
p +
    geom_bar(stat = "identity") +
    # This will collapse the extra margin at y-intercept.
    scale_y_continuous(expand = c(0, 0)) +
    theme_paperwhite(minimal = TRUE, legend_position = "none")
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.

Other ggplot2 Functions: \code{\link{ggplot2_geoms}}
}
\author{
Michael Steinbaugh
}
\concept{ggplot2 Functions}
