% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\docType{data}
\name{globals}
\alias{globals}
\title{Global variables.}
\format{An object of class \code{character} of length 6.}
\usage{
globals
}
\description{
Global variables.
}
\seealso{
\itemize{
\item \code{\link[=globalVariables]{globalVariables()}}.
\item \code{\link[=import_tidy_verbs]{import_tidy_verbs()}}.
\item \link{reexports}.
}
}
\keyword{internal}
