% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{grch37}
\alias{grch37}
\title{GRCh37 (hg19) Genome Annotations}
\format{An object of class \code{data.frame} with 63677 rows and 10 columns.}
\usage{
grch37
}
\description{
GRCh37 (hg19) Genome Annotations
}
\author{
Michael Steinbaugh
}
\keyword{internal}
