% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{grch37Tx2gene}
\alias{grch37Tx2gene}
\title{GRCh37 (hg19) Transcript to Gene Mappings}
\format{An object of class \code{data.frame} with 215170 rows and 2 columns.}
\usage{
grch37Tx2gene
}
\description{
GRCh37 (hg19) Transcript to Gene Mappings
}
\author{
Michael Steinbaugh
}
\keyword{internal}
