% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepString.R
\name{grepString}
\alias{grepString}
\title{\code{grep} String}
\usage{
grepString(string, sep = ", ")
}
\arguments{
\item{string}{\code{character} string.}

\item{sep}{Separator. Defaults to comma.}
}
\value{
\code{character} string, containing a regular expression pattern.
}
\description{
Generate a \code{grep} string for pattern matching against a delimited \code{character}
string returned by \code{\link[base:toString]{base::toString()}} or \code{\link[base:paste]{base::paste()}} using the \code{collapse}
argument.
}
\examples{
grepString("gene")
}
\seealso{
\itemize{
\item \code{\link[base:grep]{base::grep()}}.
\item \code{\link[base:toString]{base::toString()}}.
}

Other Atomic Vector Functions: \code{\link{sortUnique}},
  \code{\link{toStringUnique}}
}
\author{
Michael Steinbaugh
}
