% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepToString.R
\name{grepToString}
\alias{grepToString}
\alias{grep_to_string}
\title{Grep against toString formatted data.}
\usage{
grepToString(identifier)
}
\arguments{
\item{identifier}{Identifier.}
}
\value{
Comma separated string for matching against \code{toString} return.
}
\description{
Generates a grep pattern that will match an identifier in a comma separated
string.
}
\examples{
grepToString("gene")
}
