% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDavid.R
\name{gseaDavid}
\alias{gseaDavid}
\title{Gene set enrichment analysis (GSEA) with DAVID}
\usage{
gseaDavid(gene, background, format, file = FALSE)
}
\arguments{
\item{gene}{Gene identifier vector}

\item{background}{Background identifier vector}

\item{format}{Identifier format (see DAVID website)}

\item{file}{Export to file (\code{TRUE} or \code{FALSE})}
}
\value{
List of DAVID chart and cluster
}
\description{
Gene set enrichment analysis (GSEA) with DAVID
}

