% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDavidCutoff.R
\name{gseaDavidCutoff}
\alias{gseaDavidCutoff}
\title{Reformat RDAVIDWebService functionalAnnotationChart to a tibble with significance cutoffs}
\usage{
gseaDavidCutoff(gseaDavid, count = getOption("count"), p = getOption("p"),
  fdr = getOption("fdr"))
}
\arguments{
\item{gseaDavid}{RDAVIDWebService list from `gseaDavid()`.}

\item{count}{Minimum count of hits.}

\item{p}{P value cutoff value.}

\item{fdr}{False discovery rate cutoff value.}
}
\value{
Tibble or kable depending on the call.
}
\description{
Reformat RDAVIDWebService functionalAnnotationChart to a tibble with significance cutoffs
}
