% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaString.R
\name{gseaString}
\alias{gseaString}
\title{Gene set enrichment analysis (GSEA) with STRING}
\usage{
gseaString(gene, background, species, cache = TRUE, csv = FALSE,
  png = TRUE, pdf = FALSE, version = "10")
}
\arguments{
\item{gene}{Gene identifier vector.}

\item{background}{Background identifier vector.}

\item{species}{Species (e.g. "Homo sapiens")}

\item{cache}{Cache files to disk (TRUE/FALSE).}

\item{csv}{Export to CSV (TRUE/FALSE).}

\item{png}{Export network as PNG file (TRUE/FALSE).}

\item{pdf}{Export network as high resolution PDF file (TRUE/FALSE).}

\item{version}{STRING version to use.}
}
\value{
STRING list.
}
\description{
Gene set enrichment analysis (GSEA) with STRING
}
