% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaStringCutoff.R
\name{gseaStringCutoff}
\alias{gseaStringCutoff}
\title{Reformat STRINGdb to a tibble with significance cutoffs}
\usage{
gseaStringCutoff(gseaString, count = getOption("count"), p = getOption("p"),
  fdr = getOption("fdr"))
}
\arguments{
\item{gseaString}{STRINGdb list from `gseaString()`.}

\item{count}{Minimum count of hits.}

\item{p}{P value cutoff value.}

\item{fdr}{False discovery rate cutoff value.}
}
\value{
Tibble or kable depending on the call.
}
\description{
Reformat STRINGdb to a tibble with significance cutoffs
}

