% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headtail-methods.R
\name{headtail}
\alias{headtail}
\alias{headtail,atomic-method}
\alias{headtail,matrix-method}
\alias{headtail,sparseMatrix-method}
\alias{headtail,data.frame-method}
\alias{headtail,DataFrame-method}
\alias{headtail,GRanges-method}
\alias{headtail,SummarizedExperiment-method}
\title{Return the first and last parts of an object}
\usage{
headtail(x, ...)

\S4method{headtail}{atomic}(x, n = 2L)

\S4method{headtail}{matrix}(x, n = 2L, unicode = TRUE)

\S4method{headtail}{sparseMatrix}(x, n = 2L, unicode = TRUE)

\S4method{headtail}{data.frame}(x, n = 2L, unicode = TRUE)

\S4method{headtail}{DataFrame}(x, n = 2L, unicode = TRUE)

\S4method{headtail}{GRanges}(x, n = 2L, unicode = TRUE)

\S4method{headtail}{SummarizedExperiment}(x, n = 2L, unicode = TRUE)
}
\arguments{
\item{x}{Object.}

\item{n}{\code{integer(1)}.
Positive integer denoting the number of first and last items to include.}

\item{unicode}{\code{logical(1)}.
Allow Unicode characters in console output.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Applies to both rows and columns, enabling quick inspection during
interactive use.
}
\details{
Inspired by the \code{\link[base:print]{print()}} method for \code{DataFrame} class
objects.
}
\section{Functions}{
\itemize{
\item \code{headtail,atomic-method}: Paste collapse to a \code{character(1)}.

\item \code{headtail,matrix-method}: Show first and last rows.

\item \code{headtail,sparseMatrix-method}: Same method as \code{matrix}.

\item \code{headtail,data.frame-method}: Same method as \code{matrix}.

\item \code{headtail,DataFrame-method}: Same method as \code{data.frame}.

\item \code{headtail,GRanges-method}: Summarize the ranges.

\item \code{headtail,SummarizedExperiment-method}: Summarize the primary \code{assay}.
}}

\note{
Updated 2020-01-20.
}
\examples{
data(mtcars, package = "datasets")
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## data.frame ====
headtail(mtcars, unicode = TRUE)

## SummarizedExperiment ====
headtail(rse, unicode = TRUE)
}
\seealso{
\itemize{
\item \code{utils::head()}, \code{utils::tail()}.
\item \code{cat()}.
\item \code{getMethod("show", "DataTable")}.
}
}
