% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC2Ensembl.R
\name{HGNC2Ensembl}
\alias{HGNC2Ensembl}
\title{HGNC-to-Ensembl gene identifier mappings}
\usage{
HGNC2Ensembl()
}
\value{
\code{HGNC2Ensembl}.
Contains a \code{DataFrame} with \code{hgncID} and \code{geneID} columns.
}
\description{
HGNC-to-Ensembl gene identifier mappings
}
\examples{
options(basejump.test = TRUE)
x <- HGNC2Ensembl()
print(x)
}
