% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc2gene.R
\name{hgnc2gene}
\alias{hgnc2gene}
\title{HGNC to Ensembl Gene ID Mappings}
\usage{
hgnc2gene()
}
\value{
\code{data.frame}
}
\description{
HGNC to Ensembl Gene ID Mappings
}
\examples{
x <- hgnc2gene()
glimpse(x)
}
\author{
Michael Steinbaugh
}
