% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/humanize-methods.R
\name{humanize}
\alias{humanize}
\title{Humanize an R Object}
\usage{
humanize(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with human-friendly rownames (e.g. gene symbols
instead of stable gene IDs) and colnames (e.g. sample names instead of
sample IDs).
}
\description{
Humanize an R Object
}
\note{
This can make dimnames invalid (see \code{make.names()}) and should only
be called prior to writing files to disk.
}
