% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import}
\usage{
import(file, ..., dataFrame = getOption(x = "basejump.data.frame",
  default = "data.frame"))
}
\arguments{
\item{file}{\code{string}. File path.}

\item{...}{Additional arguments.}

\item{dataFrame}{\code{string}. Data frame class to return. Can be set globally
using the \code{basejump.data.frame} option.

Current recommendations (by priority):
\itemize{
\item \code{data.frame}: Base R default. Generally recommended.
\itemize{
\item S3 class.
\item Allows rownames, but they're required and can't be set \code{NULL}.
\item See \code{help(topic = "data.frame", package = "base")} for details.
}
\item \code{DataFrame}: Recommended when working with Bioconductor packages.
\itemize{
\item S4 class; inherits \code{DataTable} virtual class.
\item Allows rownames, but they're optional and can be set \code{NULL}.
\item See \code{help(topic = "DataFrame", package = "S4Vectors")} for details.
}
\item \code{tibble}: Recommended when working with tidyverse packages.
\itemize{
\item S3 class (\code{tbl_df}); inherits \code{data.frame}.
\item Does not allow rownames.
\item See \code{help(topic = "tibble", package = "tibble")} for details.
}
\item \code{data.table}: Recommended when working with the data.table package.
\itemize{
\item S3 class; inherits \code{data.frame}.
\item Does not allow rownames.
\item See \code{help(topic = "data.table", package = "data.table")} for details.
}
}}
}
\value{
Varies, depending on the file extension:
\itemize{
\item Default: \code{DataFrame}.
\item \code{MTX}: \code{sparseMatrix}.
\item \code{GTF}/\code{GFF}: \code{GRanges}.
\item \code{JSON}/\code{YAML}: \code{list}.
\item Source code or log: \code{character}.
}
}
\description{
Read file by extension into R.
}
\details{
This is a wrapper for \code{rio::import()} that adds support for additional common
genomic data files. Remote URLs and compressed files are supported. All
extensions are case insensitive.

This function supports automatic loading of common file types:
\itemize{
\item \code{CSV}: Comma Separated Values.
\item \code{TSV} Tab Separated Values.
\item \code{XLSX}: Excel workbook.
\item \code{MTV}: MatrixMarket sparse matrix.
\item \code{GTF}/\code{GFF}/\code{GFF3}: General Feature Format.
\item \code{RDA}/\code{RDATA}: R Data.
\itemize{
\item Must contain a single object.
\item Doesn't require internal object name to match, unlike \code{\link[=loadData]{loadData()}}.
}
\item \code{RDS}: R Data Serialized.
\item \code{JSON}: JSON.
\item \code{YAML}/\code{YML}: YAML.
}

These file formats will be read as (source code) lines:
\code{LOG}, \code{MD}, \code{PY}, \code{R}, \code{RMD}, \code{SH}.

These file formats are blacklisted, and intentionally not supported:
\code{DOC}, \code{DOCX}, \code{PDF}, \code{PPT}, \code{PPTX}, \code{TXT}.

If the file format isn't supported natively (or blacklisted), the
\href{https://cran.r-project.org/web/packages/rio/index.html}{rio} package will
be used as a fallback attempt.
}
\section{Matrix Market Exchange (MTX/MEX)}{


Reading a Matrix Market Exchange (\code{MTX}) file now requires \code{COLNAMES} and
\code{ROWNAMES} sidecar files containing the \code{\link[=colnames]{colnames()}} and \code{\link[=rownames]{rownames()}} of
the sparse matrix. Legacy support for manual loading of these sidecar files
is provided.
}

\section{General Feature Format (GFF/GTF)}{


The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.

Column names follow the \href{https://bit.ly/2K6EBla}{Ensembl conventions}.
Additional information:
\itemize{
\item \href{http://www.ensembl.org/info/website/upload/gff.html}{Ensembl}
\item \href{http://www.gencodegenes.org/gencodeformat.html}{Gencode}
}
}

\section{bcbio files}{


Also supports some additional extensions commonly used with the
\href{https://bcbio-nextgen.readthedocs.io}{bcbio} pipeline:
\itemize{
\item \code{COUNTS}: Counts table (e.g. RNA-seq aligned counts).
\item \code{COLNAMES}: Sidecar file containing column names.
\item \code{ROWNAMES}: Sidecar file containing row names.
}
}

\examples{
url <- basejumpCacheURL

## R Data
x <- import(file = file.path(url, "rnaseq_counts.rda"))
class(x)

## Comma Separated Values
x <- import(file = file.path(url, "example.csv"))
print(x)

## Tab Separated Values
x <- import(file = file.path(url, "example.tsv"))

## Microsoft Excel Worksheet
x <- import(file = file.path(url, "example.xlsx"))
print(x)

## GTF/GFF
x <- import(file = file.path(url, "example.gtf"))
summary(x)

## JSON
x <- import(file = file.path(url, "example.json"))
names(x)

## YAML
x <- import(file = file.path(url, "example.yml"))
names(x)

## Counts Table (i.e. aligned counts from bcbio)
x <- import(file = file.path(url, "example.counts"))
colSums(x)
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \href{https://cran.r-project.org/web/packages/Matrix/index.html}{Matrix}.
}
}
