% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSampleData.R
\name{importSampleData}
\alias{importSampleData}
\title{Import sample metadata}
\usage{
importSampleData(
  file,
  sheet = 1L,
  lanes = 0L,
  pipeline = c("none", "bcbio", "cellranger", "cpi"),
  autopadZeros = FALSE
)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{sheet}{\code{character(1)} or \code{integer(1)}.
Workbook sheet.}

\item{lanes}{\code{integer(1)}.
Number of lanes used to split the samples into technical replicates
suffix (i.e. \verb{_LXXX}).}

\item{pipeline}{\code{character(1)}.
Analysis pipeline:
\itemize{
\item \code{"none"}: Simple mode, requiring only "sampleId" column.
\item \code{"bcbio"}: bcbio mode. See section here in documentation for details.
\item \code{"cellranger"}: Cell Ranger mode. Currently requires "directory" column.
Used by Chromium R package.
}}

\item{autopadZeros}{\code{logical(1)}.
Autopad zeros in sample identifiers, for improved sorting.
Currently supported only for non-multiplexed samples.
For example: \code{sample_1}, \code{sample_2}, ... \code{sample_10} becomes
\code{sample_01}, \code{sample_02}, ... \code{sample10}.}
}
\value{
\code{DataFrame}.
}
\description{
This function imports user-defined sample metadata saved in a spreadsheet.
}
\note{
Works with local or remote files.

Updated 2020-07-24.
}
\section{bcbio pipeline}{


\strong{Required column names.} The \code{"description"} column is always required, and
must match the bcbio per sample directory names exactly. Inclusion of the
\code{"fileName"} column isn't required but is recommended for data provenance.
Note that some bcbio examples on readthedocs use \code{"samplename"} (note case)
instead of \code{"fileName"}. This function checks for that and will rename the
column to \code{"fileName"} automatically. We're using the \code{sampleName} column
(note case) to define unique sample names, in the event that bcbio has
processed multiplexed samples.

\strong{Demultiplexed samples.} The samples in the bcbio run must map to the
\code{"description"} column. The values provided in description for demultiplexed
samples must be unique. They must also be \emph{syntactically valid}, meaning that
they cannot contain illegal characters (e.g. spaces, non-alphanumerics,
\emph{dashes}) or \emph{begin with a number}. Consult the documentation in \code{help(topic = "make.names")} for more information on valid names in R.

\strong{Multiplexed samples.} This applies to some single-cell RNA-seq formats,
including inDrops. In this case, bcbio will output per-sample directories
with this this structure: \code{description-revcomp}. \code{\link[=readSampleData]{readSampleData()}} checks to
see if the \code{"description"} column is unique. If the values are duplicated,
the function assumes that bcbio processed multiplexed FASTQs, where multiple
samples of interest are barcoded inside a single FASTQ. This this case, you
must supply additional \code{"index"}, \code{"sequence"}, and \code{"sampleName"} columns.
Note that bcbio currently outputs the reverse complement index sequence in
the sample directory names (e.g. \code{"sample-ATAGAGAG"}). Define the forward
index barcode in the \code{sequence} column here, not the reverse complement. The
reverse complement will be calculated automatically and added as the
\code{revcomp} column in the sample metadata.
}

\examples{
## Demultiplexed ====
file <- file.path(basejumpTestsURL, "bcbio-metadata-demultiplexed.csv")
x <- importSampleData(file, pipeline = "bcbio")
print(x)

## Multiplexed ====
file <- file.path(basejumpTestsURL, "bcbio-metadata-multiplexed-indrops.csv")
x <- importSampleData(file, pipeline = "bcbio")
print(x)
}
\author{
Michael Steinbaugh
}
