% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeDirectory.R
\name{initializeDirectory}
\alias{initializeDirectory}
\title{Initialize Directory}
\usage{
initializeDirectory(dir)
}
\arguments{
\item{dir}{Directory path.}
}
\description{
Initialize Directory
}
\examples{
initializeDirectory(getwd())
}
