% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/integerCounts-methods.R
\name{integerCounts}
\alias{integerCounts}
\alias{integerCounts,matrix-method}
\alias{integerCounts,Matrix-method}
\alias{integerCounts,SummarizedExperiment-method}
\title{Integer counts}
\usage{
integerCounts(object, ...)

\S4method{integerCounts}{matrix}(object)

\S4method{integerCounts}{Matrix}(object)

\S4method{integerCounts}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Matrix.
Typically \code{matrix} or \code{Matrix} class.
}
\description{
Integer counts
}
\note{
Updated 2020-12-21.

For a \code{SummarizedExperiment} object, \code{"counts"} must be explicitly
defined in \code{assayNames}.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
x <- integerCounts(object)
class(x)
is.integer(x)
summary(x)
}
