% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-localOrRemoteFile.R
\name{.localOrRemoteFile}
\alias{.localOrRemoteFile}
\title{Dynamically Handle a Local or Remote File Path}
\usage{
.localOrRemoteFile(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Named character vector containing the original file name as the
name and local file path as the string.
}
\description{
Dynamically Handle a Local or Remote File Path
}
\examples{
\dontrun{
file.path(testDataURL, "mtcars.csv") \%>\%
    .localOrRemoteFile
}
}
