% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/intersectAll-methods.R
\name{intersectAll}
\alias{intersectAll}
\alias{intersectAll,list-method}
\title{Intersection of more than two vectors}
\usage{
intersectAll(object, ...)

\S4method{intersectAll}{list}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{vector}.
}
\description{
Intersection of more than two vectors
}
\note{
Updated 2020-08-18.
}
\examples{
object <- list(
    a = c("a", "b", "c", "d", "e", "f"),
    b = c("b", "c", "d", "e", "f", "g"),
    c = c("c", "d", "e", "f", "g", "h")
)
intersectAll(object)
}
