% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectionMatrix-methods.R
\name{intersectionMatrix}
\alias{intersectionMatrix}
\alias{intersectionMatrix,list-method}
\title{Intersection matrix}
\usage{
intersectionMatrix(object, ...)

\S4method{intersectionMatrix}{list}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Intersection matrix
}
\note{
Updated 2020-08-18.
}
\examples{
list <- list(
    a = c("a", "b", "c", "d", "e", "f"),
    b = c("b", "c", "d", "e", "f", "g"),
    c = c("c", "d", "e", "f", "g", "h")
)
mat <- intersectionMatrix(list)
rowSums(mat)
}
