% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intronicMappingRate.R
\name{intronicMappingRate}
\alias{intronicMappingRate}
\title{Intronic mapping rate plot}
\usage{
intronicMappingRate(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Intronic mapping rate plot
}
\examples{
\dontrun{
intronicMappingRate(summary)
}
}
\author{
Michael Steinbaugh
}
