% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-methods.R
\name{join}
\alias{join}
\alias{innerJoin}
\alias{leftJoin}
\alias{rightJoin}
\alias{fullJoin}
\alias{semiJoin}
\alias{antiJoin}
\alias{innerJoin,DataFrame,DataFrame,character-method}
\alias{leftJoin,DataFrame,DataFrame,character-method}
\alias{rightJoin,DataFrame,DataFrame,character-method}
\alias{fullJoin,DataFrame,DataFrame,character-method}
\alias{semiJoin,DataFrame,DataFrame,character-method}
\alias{antiJoin,DataFrame,DataFrame,character-method}
\title{Join operations}
\usage{
innerJoin(x, y, by, ...)

leftJoin(x, y, by, ...)

rightJoin(x, y, by, ...)

fullJoin(x, y, by, ...)

semiJoin(x, y, by, ...)

antiJoin(x, y, by, ...)

\S4method{innerJoin}{DataFrame,DataFrame,character}(x, y, by)

\S4method{leftJoin}{DataFrame,DataFrame,character}(x, y, by)

\S4method{rightJoin}{DataFrame,DataFrame,character}(x, y, by)

\S4method{fullJoin}{DataFrame,DataFrame,character}(x, y, by)

\S4method{semiJoin}{DataFrame,DataFrame,character}(x, y, by)

\S4method{antiJoin}{DataFrame,DataFrame,character}(x, y, by)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{by}{\code{character}.
Column names to use for join operation.}

\item{...}{Additional arguments.}
}
\value{
New object containing a merge of \code{x} and \code{y} objects.
}
\description{
Join operations
}
\note{
Updated 2019-09-03.
}
\section{Ordering}{


The join functions never rearrange rows. To accomplish this, we're currently
setting an internal \code{.idx} column that we can use to reorder the rows after
\code{\link[base:merge]{merge()}} operation.
}

\section{Row names}{


Unlike the S3 methods defined in dplyr, the join methods defined here for
\code{DataFrame} always preserve row names.
}

\examples{
data(band_members, band_instruments)

## DataFrame ====
x <- band_members
print(x)
y <- band_instruments
print(y)
by <- "name"
innerJoin(x = x, y = y, by = by)
leftJoin(x = x, y = y, by = by)
rightJoin(x = x, y = y, by = by)
fullJoin(x = x, y = y, by = by)
semiJoin(x = x, y = y, by = by)
antiJoin(x = x, y = y, by = by)
}
\seealso{
These functions are inspired by dplyr. However, they are designed to only
work on Bioconductor S4 class objects, and use base R code internally.
\itemize{
\item \code{help(topic = "join", package = "dplyr")}.
\item \code{help(topic = "merge", package = "S4Vectors")}.
\item https://support.bioconductor.org/p/120277/
}
}
