% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kable}
\alias{kable}
\title{Create tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
kable(x, ...)
}
\arguments{
\item{x}{An R object (typically a matrix or data frame)}

\item{...}{Other arguments}
}
\value{
Kable during knit (via \code{rmarkdown::render()}), otherwise the
  original data
}
\description{
Wrapper function for \code{knitr::kable()}
}
\examples{
kable(head(iris))
}
\author{
Michael Steinbaugh
}
\keyword{internal}
