% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{kables}
\alias{kables}
\title{Create Tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
kables(list, captions = NULL)
}
\arguments{
\item{list}{List of column data (e.g. \link{data.frame}, \link{matrix}).}

\item{captions}{Optional character vector of table captions.}
}
\value{
Knit tables, using \code{\link[knitr:kable]{knitr::kable()}}.
}
\description{
Handle multiple kables in a single RMarkdown chunk.
}
\examples{
list(starwars, head(mtcars)) \%>\% kables
}
