% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{lanePattern}
\alias{lanePattern}
\title{Sequencing lane grep pattern}
\format{An object of class \code{character} of length 1.}
\usage{
lanePattern
}
\description{
Sequencing lane grep pattern
}
\examples{
lanePattern
}
\keyword{datasets}
