% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latestGenomeVersions.R
\name{latestGenomeVersion}
\alias{latestGenomeVersion}
\alias{latestEnsemblVersion}
\alias{latestGencodeVersion}
\alias{latestRefseqVersion}
\alias{latestFlybaseVersion}
\alias{latestWormbaseVersion}
\title{Latest genome version}
\usage{
latestEnsemblVersion()

latestGencodeVersion(organism)

latestRefseqVersion()

latestFlybaseVersion(dmel = FALSE)

latestWormbaseVersion()
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{dmel}{\code{logical(1)}.
Return \emph{Drosophila melanogaster} genome release version.}
}
\value{
\code{character(1)}.
Release version.
}
\description{
Obtain the latest release version from various genome annotation sources.
}
\note{
Updated 2020-01-20.
}
\examples{
if (goalie::hasInternet(url = "ftp://ftp.ensembl.org/")) {
    currentEnsemblVersion()
}
}
\seealso{
Refer to the koopa package for shell variants:
\itemize{
\item \code{latest-ensembl-version}.
\item \code{latest-flybase-version}.
\item \code{latest-gencode-version}.
\item \code{latest-refseq-version}.
\item \code{latest-wormbase-version}.
}
}
