% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadData}
\alias{loadData}
\alias{loadRemote}
\title{Data loading utilities}
\usage{
loadData(data)

loadRemote(url)
}
\arguments{
\item{url}{URL.}

\item{...}{Data files as dot objects.}
}
\description{
Dynamically \code{\link[=load]{load()}} data from \code{data} directory or \code{\link[=source]{source()}}
the corresponding script frmo the \code{data-raw} directory.

Load a remote R binary file.
}
\examples{
\dontrun{
loadData(geneIDs, oligo)
}
\dontrun{
loadRemote("http://example.com/data.rda")
}
}
