% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Data}
\usage{
loadData(..., dir = getOption("basejump.load.dir", "."),
  envir = globalenv())
}
\arguments{
\item{...}{Object names. Note that these arguments are interpreted as symbols
using non-standard evaluation for convenience during interactive use, and
\emph{must not be quoted}.}

\item{dir}{\code{character(1)}. Directory path.}

\item{envir}{\code{environment} to use for assignment. Defaults to \code{parent.frame},
which will assign into the calling environment.}
}
\value{
Invisible \code{character}. File paths.
}
\description{
Load R data files from a directory using symbols rather than complete file
paths. Supports "\code{.rds}", "\code{.rda}", and "\code{.RData}" file extensions.
}
\details{
\code{loadData} is opinionated about the format of R data files it will accept.
\code{save} allows for the saving of multiple objects into a single R data file.
This can later result in unexpected accidental replacement of an existing
object in the current environment. Since an R data file internally stores the
name of an object, if the file is later renamed the object name will no
longer match.

To avoid any accidental replacements, \code{loadData} will only load R data
files that contain a single object, and the internal object name must match
the file name exactly. Additionally, \code{loadData} will intentionally error if
an object with the same name already exists in the destination \code{environment}.
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation.
}
\examples{
loadData(example, dir = system.file("extdata", package = "basejump"))
}
\seealso{
\code{load}, \code{readRDS}.
}
