% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadDataAsName}
\alias{loadDataAsName}
\title{Load data as name}
\usage{
loadDataAsName(..., dir = getOption("basejump.load.dir", "."),
  envir = globalenv())
}
\arguments{
\item{...}{Key value pairs, defining the name mappings. For example,
\code{newName1} = \code{oldName1}, \code{newName2} = \code{oldName2}. Note that these
arguments are interpreted using non-standard evaluation, and \emph{should not
be quoted}.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{envir}{\code{environment}.
Environment to use for assignment. Defaults to
\link[base:parent.frame]{parent.frame()}, which will assign into the calling
environment.}
}
\value{
Invisible named \code{character}. File paths.
}
\description{
Load data as name
}
\note{
This function is intended for interactive use and interprets object
names using non-standard evaluation.
}
\examples{
loadDataAsName(
    renamed = example,
    dir = system.file("extdata", package = "basejump")
)
class(renamed)
}
