% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataRaw.R
\name{loadDataRaw}
\alias{loadDataRaw}
\title{Load data-raw directory}
\usage{
loadDataRaw(data)
}
\arguments{
\item{data}{Data object}
}
\description{
Dynamically source or load raw data in the `data-raw` directory
}
\author{
Michael Steinbaugh
}
\keyword{internal}
