% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRemote.R
\name{loadRemote}
\alias{loadRemote}
\alias{load_remote}
\title{Load a remote R binary file.}
\usage{
loadRemote(url)

load_remote(url)
}
\arguments{
\item{url}{URL.}
}
\description{
Load a remote R binary file.
}
\examples{
\dontrun{
loadRemote("http://example.com/data.rda")
}
}
