% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRemoteData.R
\name{loadRemoteData}
\alias{loadRemoteData}
\title{Load Remote Data}
\usage{
loadRemoteData(object, envir = parent.frame())
}
\arguments{
\item{object}{Object.}
}
\value{
No value.
}
\description{
Load a remote R binary file.
}
\examples{
loadRemoteData(file.path(testDataURL, "mtcars.rda"))
}
\seealso{
Other Data Import and Project Utilities: \code{\link{readFileByExtension}},
  \code{\link{readGFF}},
  \code{\link{readSampleMetadataFile}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
\concept{Data Import and Project Utilities}
