% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically Handle a Local or Remote File Path}
\usage{
localOrRemoteFile(file)
}
\arguments{
\item{file}{\code{string}. Local file paths or remote URLs.}
}
\value{
\code{character}. Local file paths. Stops on a missing file.
}
\description{
Dynamically Handle a Local or Remote File Path
}
\note{
This function is vectorized and supports mixed local and remote paths. Remote
files are downloaded locally to a temporary directory.
}
\examples{
# Import a single file.
file <- file.path(basejumpCacheURL, "rnaseq_counts.csv.gz")
x <- localOrRemoteFile(file)
basename(x)

# Import multiple files (vectorized).
files <- file.path(
    basejumpCacheURL,
    c("rnaseq_counts.csv.gz", "single_cell_counts.mtx.gz")
)
x <- localOrRemoteFile(files)
basename(x)
}
\seealso{
\code{\link[base:tempdir]{base::tempdir()}}.
}
\author{
Michael Steinbaugh
}
