% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{fc2lr}
\alias{fc2lr}
\alias{lr2fc}
\title{Log ratio to fold change}
\usage{
fc2lr(x, base = 2L)

lr2fc(x, base = 2L)
}
\arguments{
\item{x}{Numeric vector of log ratio (\code{lr}) or fold change (\code{fc}) values.}

\item{base}{Logarithm base. Defaults to \code{2}, for compatibility with RNA-Seq
differential expression output.}
}
\value{
Numeric vector.
}
\description{
Convert log ratio normalized values to fold change.
}
\examples{
lr2fc(seq(-3, 3, 1))
fc2lr(c(-8, -4, -2, 1, 2, 4, 8))
}
\seealso{
\code{gtools::foldchange2logratio()}, \code{gtools::logratio2foldchange()}.
}
