% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2Symbol.R
\name{makeGene2Symbol}
\alias{makeGene2Symbol}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGene2SymbolFromGFF}
\alias{makeGene2SymbolFromGTF}
\title{Make a Gene2Symbol object}
\usage{
makeGene2SymbolFromEnsembl()

makeGene2SymbolFromEnsDb(object)

makeGene2SymbolFromGFF(file)
}
\arguments{
\item{object}{Object.}

\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Make a Gene2Symbol object
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
## makeGene2SymbolFromEnsembl ====
x <- makeGene2SymbolFromEnsembl(organism = "Homo sapiens")
print(x)

## makeTx2GeneFromEnsDb ====
x <- makeGene2SymbolFromEnsDb("EnsDb.Hsapiens.v75")
print(x)

## makeGene2SymbolFromGFF ====
## GTF
file <- file.path(basejumpCacheURL, "example.gtf")
x <- makeGene2SymbolFromGFF(file)
print(x)

## GFF3
file <- file.path(basejumpCacheURL, "example.gff3")
x <- makeGene2SymbolFromGFF(file)
print(x)
}
