% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2symbolFromEnsembl.R
\name{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromEnsembl}
\title{Gene-to-Symbol Mappings from Ensembl}
\usage{
makeGene2symbolFromEnsembl(organism, genomeBuild = NULL, release = NULL)
}
\arguments{
\item{organism}{Default recommended usage is to provide the full latin
organism name as a string.}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}
}
\value{
\code{data.frame}
}
\description{
Gene-to-Symbol Mappings from Ensembl
}
