% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRanges-methods.R
\name{makeGRanges}
\alias{makeGRanges}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromEnsDb}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\alias{annotable}
\title{Make a \code{GenomicRanges} object}
\usage{
makeGRangesFromEnsembl(organism, level = c("genes", "transcripts"),
  genomeBuild = NULL, release = NULL)

makeGRangesFromEnsDb(object, level = c("genes", "transcripts"))

makeGRangesFromGFF(file, level = c("genes", "transcripts"))

annotable(organism, level = c("genes", "transcripts"),
  genomeBuild = NULL, release = NULL)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{level}{\code{character(1)}.
Return ranges as "\code{genes}" or "\code{transcripts}".}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available. Note: don't pass in UCSC build
IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). If set \code{NULL}, defaults to the most
recent release available.}

\item{object}{Object.}

\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{GRanges}.
}
\description{
Make a \code{GenomicRanges} object
}
\section{Functions}{
\itemize{
\item \code{makeGRangesFromEnsembl}: Quickly obtain gene and transcript annotations from
\href{https://www.ensembl.org/}{Ensembl} using
\href{https://bioconductor.org/packages/AnnotationHub/}{AnnotationHub} and
\href{https://bioconductor.org/packages/ensembldb/}{ensembldb}.

Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.

Genome build: use "\code{GRCh38}" instead of "\code{hg38}" for the genome build, since
we're querying Ensembl and not UCSC.

\item \code{makeGRangesFromEnsDb}: Use specific \code{EnsDb} object as annotation source.
Alternatively, can pass in an EnsDb package name as a \code{character(1)}.

\item \code{makeGRangesFromGFF}: The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2. We recommend
using a GTF file instead of a GFF3 file, if possible.

The UCSC website has detailed conventions on the GFF3 format, including
the metadata columns.

Remote URLs and compressed files are supported.

\item \code{annotable}: \code{\link[=annotable]{annotable()}} is a legacy convenience function that calls
\code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} and returns a \code{tibble} instead of \code{GRanges}. Note
that \code{GRanges} can also be coercing using
\code{\link[BiocGenerics:as.data.frame]{as.data.frame()}}.
}}

\section{Broad class definitions}{


For gene and transcript annotations, a \code{broadClass} column is added, which
generalizes the gene types into a smaller number of semantically-meaningful
groups:
\itemize{
\item \code{coding}.
\item \code{noncoding}.
\item \code{pseudo}.
\item \code{small}.
\item \code{decaying}.
\item \code{ig} (immunoglobulin).
\item \code{tcr} (T cell receptor).
\item \code{other}.
}
}

\section{GRCh37 (hg19) legacy annotations}{


\code{makeGRangesFromEnsembl} supports the legacy \emph{Homo sapiens} GRCh37 (release
75) build by internally querying the \href{https://bioconductor.org/packages/EnsDb.Hsapiens.v75/}{EnsDb.Hsapiens.v75} package.
Alternatively, the corresponding GTF/GFF file can be loaded directly from
GENCODE or Ensembl.
}

\examples{
## makeGRangesFromEnsembl ====
## Genes
x <- makeGRangesFromEnsembl("Homo sapiens", level = "genes")
summary(x)

## Transcripts
x <- makeGRangesFromEnsembl("Homo sapiens", level = "transcripts")
summary(x)

## makeGRangesFromEnsDb ====
x <- makeGRangesFromEnsDb("EnsDb.Hsapiens.v75")

## makeGRangesFromGFF ====
file <- file.path(basejumpCacheURL, "example.gtf")

## Genes
x <- makeGRangesFromGFF(file = file, level = "genes")
summary(x)

## Transcripts
x <- makeGRangesFromGFF(file = file, level = "transcripts")
summary(x)
}
\seealso{
\itemize{
\item \href{https://bioconductor.org/packages/AnnotationHub/}{AnnotationHub}.
\item \href{https://bioconductor.org/packages/ensembldb/}{ensembldb}.
\item \code{rtracklayer::import()}.
\item \code{GenomicFeatures::makeTxDbFromGFF()}.
}
}
