% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromEnsembl.R
\name{makeGRangesFromEnsembl}
\alias{makeGRangesFromEnsembl}
\alias{annotable}
\title{Make GRanges from Ensembl}
\usage{
makeGRangesFromEnsembl(
  organism,
  level = c("genes", "transcripts"),
  genomeBuild = NULL,
  release = NULL,
  ignoreTxVersion = TRUE,
  synonyms = TRUE
)

annotable(...)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{ignoreTxVersion}{\code{logical(1)}. Don't the include the transcript version
in the identifier. Only applies when \code{level = "transcripts"}. This
simplifies identifier matching when generating a \code{tx2gene} file.}

\item{...}{Additional arguments.}
}
\value{
\code{GRanges}.
}
\description{
Quickly obtain gene and transcript annotations from
\href{https://www.ensembl.org/}{Ensembl} using
\href{https://bioconductor.org/packages/AnnotationHub/}{AnnotationHub} and
\href{https://bioconductor.org/packages/ensembldb/}{ensembldb}.
}
\details{
Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \verb{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.

Genome build: use \code{"GRCh38"} instead of \code{"hg38"} for the genome build, since
we're querying Ensembl and not UCSC.
}
\section{Functions}{
\itemize{
\item \code{annotable}: Legacy convenience function that calls \code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} and returns
a \code{tbl_df} (tibble) instead of \code{GRanges}. Note that \code{GRanges} can also be
coercing using \code{\link[BiocGenerics:as.data.frame]{as.data.frame()}}.
}}

\note{
Updated 2020-10-05.
}
\section{Broad class definitions}{


For gene and transcript annotations, a \code{broadClass} column is added, which
generalizes the gene types into a smaller number of semantically-meaningful
groups:
\itemize{
\item \code{coding}.
\item \code{noncoding}.
\item \code{pseudo}.
\item \code{small}.
\item \code{decaying}.
\item \code{ig} (immunoglobulin).
\item \code{tcr} (T cell receptor).
\item \code{other}.
}
}

\section{GRCh37 (hg19) legacy annotations}{


\code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} supports the legacy \emph{Homo sapiens} GRCh37 (release
75) build by internally querying the \href{https://bioconductor.org/packages/EnsDb.Hsapiens.v75/}{EnsDb.Hsapiens.v75} package.
Alternatively, the corresponding GTF/GFF file can be loaded directly from
GENCODE or Ensembl.
}

\section{AnnotationHub queries}{


Here's how to perform manual, customized AnnotationHub queries.\preformatted{library(AnnotationHub)
library(ensembldb)
ah <- AnnotationHub()

# Human ensembldb (EnsDb) records.
ahs <- query(
    x = ah,
    pattern = c(
        "Homo sapiens",
        "GRCh38",
        "Ensembl",
        "EnsDb"
    )
)
mcols(ahs)
print(ahs)
# EnsDb (Ensembl GRCh38 94; 2018-10-11)
ah[["AH64923"]]

# Human UCSC TxDb records.
ahs <- query(
    x = ah,
    pattern = c(
        "Homo sapiens",
        "UCSC",
        "TxDb",
        "knownGene"
    )
)
mcols(ahs)
print(ahs)
# TxDb (UCSC hg38 GENCODE 24; 2016-12-22)
ah[["AH52260"]]
}
}

\examples{
## Genes
x <- makeGRangesFromEnsembl("Homo sapiens", level = "genes")
summary(x)

## Transcripts
x <- makeGRangesFromEnsembl("Homo sapiens", level = "transcripts")
summary(x)
}
\seealso{
\itemize{
\item \href{https://bioconductor.org/packages/AnnotationHub/}{AnnotationHub}.
\item \href{https://bioconductor.org/packages/ensembldb/}{ensembldb}.
}
}
