% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromGFF.R
\name{makeGRangesFromGFF}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\title{Make \code{GRanges} from GFF/GTF file}
\usage{
makeGRangesFromGFF(file, level = c("genes", "transcripts"))

makeGRangesFromGTF(file, level = c("genes", "transcripts"))
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{level}{\code{character(1)}.
Return ranges as "\code{genes}" or "\code{transcripts}".}
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2. We recommend
using a GTF file instead of a GFF3 file, if possible.

The UCSC website has detailed conventions on the GFF3 format, including
the metadata columns.

Remote URLs and compressed files are supported.
}
\section{Functions}{
\itemize{
\item \code{makeGRangesFromGTF}: File format alias.
Runs the same internal code as \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}}.
}}

\examples{
file <- file.path(basejumpCacheURL, "example.gtf")

## Genes
x <- makeGRangesFromGFF(file = file, level = "genes")
summary(x)

## Transcripts
x <- makeGRangesFromGFF(file = file, level = "transcripts")
summary(x)
}
\seealso{
\itemize{
\item \code{rtracklayer::import()}.
\item \code{GenomicFeatures::makeTxDbFromGFF()}.
}
}
