% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNamesCamel.R
\name{makeNamesCamel}
\alias{makeNamesCamel}
\alias{camel}
\title{Make names camelCase}
\usage{
makeNamesCamel(character)

camel(character)
}
\arguments{
\item{character}{Character vector}
}
\value{
Character vector formatted as camelCase
}
\description{
Make names camelCase
}
\examples{
makeNamesCamel("RNAi clone")
}
\author{
Michael Steinbaugh
}
