% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNamesSnake.R
\name{makeNamesSnake}
\alias{makeNamesSnake}
\alias{snake}
\title{Make names snake_case}
\usage{
makeNamesSnake(character)

snake(character)
}
\arguments{
\item{character}{Character vector}
}
\value{
Character vector formatted as snake_case
}
\description{
Make names snake_case
}
\examples{
makeNamesSnake("RNAi clone")
}
\author{
Michael Steinbaugh
}
