% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeProtein2Gene.R
\name{makeProtein2Gene}
\alias{makeProtein2Gene}
\alias{makeProtein2GeneFromEnsembl}
\title{Map protein identifiers to genes}
\usage{
makeProtein2GeneFromEnsembl(
  ids,
  organism = NULL,
  genomeBuild = NULL,
  release = NULL
)
}
\arguments{
\item{ids}{\code{character}.
Ensembl protein identifiers.
Human proteins are prefixed with "ENSP", for example.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}
}
\value{
\code{Protein2Gene}.
}
\description{
Map protein identifiers to genes
}
\note{
Updated 2020-09-29.
}
\examples{
ids <- c("ENSP00000238714", "ENSP00000338157")
makeProtein2GeneFromEnsembl(ids)
}
