% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSampleData.R
\name{makeSampleData}
\alias{makeSampleData}
\title{Make sample data}
\usage{
makeSampleData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DataFrame}.
}
\description{
Utility function that prepares metadata to be slotted into \code{colData}.
}
\details{
This function adheres to the following conventions:
\itemize{
\item All column names will be converted to lower camel case
(see \code{camel} for details).
\item Required columns:
\itemize{
\item \code{sampleName}: Human readable sample names. Note that this column is
useful for plots and doesn't have to match the column names of a
\code{SummarizedExperiment} object, which should use valid names.
}
\item Blacklisted columns:
\itemize{
\item \code{filename} (use \code{fileName}).
\item \code{rowname}.
\item \code{sample}. Too vague. Does this represent an ID or human readable name?
\item \code{samplename} (use \code{sampleName}).
}
\item \code{sampleName} column is always placed first.
}
}
\examples{
object <- data.frame(
    genotype = rep(c("control", "wildtype"), times = 2L),
    treatment = rep(c("vector", "RNAi"), each = 2L),
    sampleName = paste("sample", seq_len(4L)),
    row.names = paste0("GSM000000", seq_len(4L))
)
}
\seealso{
\code{makeNames}.
}
