% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/makeSingleCellExperiment-methods.R
\docType{methods}
\name{makeSingleCellExperiment}
\alias{makeSingleCellExperiment}
\alias{makeSingleCellExperiment,SimpleList-method}
\alias{makeSingleCellExperiment,list-method}
\title{Make a SingleCellExperiment object}
\usage{
makeSingleCellExperiment(assays, ...)

\S4method{makeSingleCellExperiment}{SimpleList}(assays,
  rowRanges = GRangesList(), colData = DataFrame(),
  metadata = list(), reducedDims = SimpleList(),
  transgeneNames = NULL, spikeNames = NULL)

\S4method{makeSingleCellExperiment}{list}(assays,
  rowRanges = GRangesList(), colData = DataFrame(),
  metadata = list(), reducedDims = SimpleList(),
  transgeneNames = NULL, spikeNames = NULL)
}
\arguments{
\item{assays}{\code{SimpleList}.
Count matrices, which must have matching dimensions. Counts can be passed
in as either a dense matrix (\code{matrix}) or sparse matrix (\code{sparseMatrix}).}

\item{...}{Additional arguments.}

\item{rowRanges}{\code{GRanges} or \code{GRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{colData}{\code{DataFrame}.
Metadata describing the assay columns. For bulk RNA-seq, this data
describes the samples. For single-cell RNA-seq, this data describes the
cells.}

\item{metadata}{\code{list}.
Metadata.}

\item{reducedDims}{\code{SimpleList}.
List containing matrices of cell coordinates in reduced space.}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character}.
Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).}
}
\value{
\code{SingleCellExperiment}.
}
\description{
This function is a utility wrapper for \code{SummarizedExperiment} that provides
automatic subsetting for row and column data, as well as automatic handling
of transgenes and spike-ins.
}
\note{
Updated 2019-08-22.
}
\section{Session information}{


This function improves upon the standard constructor by slotting useful
session information into the \code{metadata} slot by default:
\itemize{
\item \code{date}: Today's date, returned from \code{Sys.Date}.
\item \code{wd}: Working directory, returned from \code{getwd}.
\item \code{sessionInfo}: \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} return.
}

This behavior can be disabled by setting \code{sessionInfo = FALSE}.
}

\examples{
data(SingleCellExperiment, package = "acidtest")

## SimpleList ====
object <- SingleCellExperiment
assays <- assays(object)
rowRanges <- rowRanges(object)
colData <- colData(object)
metadata <- metadata(object)
reducedDims <- reducedDims(object)

x <- makeSingleCellExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    metadata = metadata,
    reducedDims = reducedDims
)
print(x)
}
\seealso{
\itemize{
\item \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment()}}.
\item \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment()}}.
\item \code{help("RangedSummarizedExperiment-class", "SummarizedExperiment")}.
\item \code{help("SummarizedExperiment-class", "SummarizedExperiment")}.
\item \code{help("SingleCellExperiment-class", "SingleCellExperiment")}.
}
}
