% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2GeneFromFASTA.R
\name{makeTx2GeneFromFASTA}
\alias{makeTx2GeneFromFASTA}
\title{Make a Tx2Gene object from transcriptome FASTA}
\usage{
makeTx2GeneFromFASTA(file, source = c("ensembl", "gencode", "refseq",
  "flybase", "wormbase"))
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{source}{\code{character(1)}.
FASTA file source:
\itemize{
\item \code{"ensembl"}: Ensembl.
\item \code{"gencode"}: GENCODE.
\item \code{"refseq"}: RefSeq.
\item \code{"flybase"}: FlyBase.
\item \code{"wormbase"}: WormBase.
}

Assuming Ensembl transcriptome (i.e. cDNA) input by default.}
}
\description{
Make a Tx2Gene object from transcriptome FASTA
}
\note{
Updated 2019-10-24.
}
\examples{
file <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-98",
    "fasta",
    "homo_sapiens",
    "cdna",
    "Homo_sapiens.GRCh38.cdna.all.fa.gz",
    protocol = "ftp"
)
makeTx2GeneFromFASTA(file, source = "ensembl")
}
\seealso{
tx2gene importers defined in \link{koopa}.
\link{koopa}: https://koopa.acidgenomics.com/
}
