% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2geneFromGFF.R
\name{makeTx2geneFromGFF}
\alias{makeTx2geneFromGFF}
\alias{makeTx2geneFromGTF}
\title{Transcript-to-Gene Mappings from GFF File}
\usage{
makeTx2geneFromGFF(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{data.frame}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
# GTF
x <- makeTx2geneFromGFF("http://basejump.seq.cloud/example.gtf")
glimpse(x)

# GFFv3
x <- makeTx2geneFromGFF("http://basejump.seq.cloud/example.gff3")
glimpse(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{mgi2gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
