% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_pkg.R
\name{manage_pkg}
\alias{manage_pkg}
\title{Manage R packages.}
\usage{
manage_pkg(pkg, source = "cran")
}
\arguments{
\item{source}{}
}
\description{
This function allows for dynamic install and loading of packages from CRAN,
Bioconductor, and GitHub.
}

