% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCellsToSamples.R
\name{mapCellsToSamples}
\alias{mapCellsToSamples}
\title{Map cells to samples}
\usage{
mapCellsToSamples(cells, samples)
}
\arguments{
\item{cells}{\code{character}.
Cell identifiers.}

\item{samples}{\code{character}.
Sample identifiers.}
}
\value{
\code{factor}.
Cells as the names and samples as the levels.
}
\description{
This function extracts \code{sampleId} from the \code{cellId} column using grep
matching.
}
\note{
Updated 2021-01-14.
}
\examples{
samples <- paste0("sample", seq_len(2L))
print(samples)
cells <- paste(samples, c("AAAAAAAA", "CCCCCCCC"), sep = "_")
print(cells)
mapCellsToSamples(cells, samples)
}
