% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapEnsemblReleaseToURL.R
\name{mapEnsemblReleaseToURL}
\alias{mapEnsemblReleaseToURL}
\title{Map Ensembl release to archive URL.}
\usage{
mapEnsemblReleaseToURL(release)
}
\arguments{
\item{release}{\code{integer(1)} or \code{character(1)}.
Ensembl release (e.g. 99).}
}
\value{
\code{character(1)}.
URL.
}
\description{
Map Ensembl release to archive URL.
}
\note{
Updated 2020-10-05.
}
\examples{
mapEnsemblReleaseToURL(96L)
}
\seealso{
\itemize{
\item \code{biomaRt::listEnsemblArchives()}.
}
}
