% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mapGenes-methods.R
\docType{methods}
\name{mapGenesToIDs}
\alias{mapGenesToIDs}
\alias{mapGenesToRownames}
\alias{mapGenesToSymbols}
\alias{mapGenes}
\alias{mapGenesToRownames,Gene2Symbol-method}
\alias{mapGenesToRownames,SummarizedExperiment-method}
\alias{mapGenesToIDs,Gene2Symbol-method}
\alias{mapGenesToIDs,SummarizedExperiment-method}
\alias{mapGenesToSymbols,Gene2Symbol-method}
\alias{mapGenesToSymbols,SummarizedExperiment-method}
\title{Map genes}
\usage{
\S4method{mapGenesToRownames}{Gene2Symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToRownames}{SummarizedExperiment}(object, genes,
  strict = TRUE)

\S4method{mapGenesToIDs}{Gene2Symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToIDs}{SummarizedExperiment}(object, genes,
  strict = TRUE)

\S4method{mapGenesToSymbols}{Gene2Symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToSymbols}{SummarizedExperiment}(object, genes,
  strict = TRUE)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{strict}{\code{logical(1)}.
Require all genes to match. Recommended by default. If set \code{FALSE}, instead
will return a warning to the user, and subset the genes vector to only
include matches.}
}
\value{
\code{character}.
}
\description{
Take a user-defined gene vector and dynamically map the input to either the
object rownames or the gene names (symbols). These functions are useful for
writing code that needs to handle either gene identifier or gene name input
dynamically (e.g. for single-cell RNA-seq marker analysis).
}
\section{Ambiguous gene names}{


Some genomes (e.g. Homo sapiens, Mus musculus) contain duplicated gene names
for multiple gene identifiers. Normally we handle these ambiguous gene names
by sanitizing them with \code{make.names}. If a user requests a gene name that
is duplicated, these functions will return a warning.
}

\examples{
data(rse, package = "acidtest")
print(rse)

rownames <- head(rownames(rse))
print(rownames)

g2s <- Gene2Symbol(rse)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)

geneNames <- head(g2s[["geneName"]])
print(geneNames)

## Row names.
mapGenesToRownames(rse, genes = rownames)
mapGenesToRownames(rse, genes = geneIDs)
mapGenesToRownames(rse, genes = geneNames)

## Gene identifiers.
mapGenesToIDs(rse, genes = rownames)
mapGenesToIDs(rse, genes = geneIDs)
mapGenesToIDs(rse, genes = geneNames)

## Gene names (symbols).
mapGenesToSymbols(rse, genes = rownames)
mapGenesToSymbols(rse, genes = geneIDs)
mapGenesToSymbols(rse, genes = geneNames)
}
