% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappedReads.R
\name{mappedReads}
\alias{mappedReads}
\title{Mapped reads plot}
\usage{
mappedReads(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Mapped reads plot
}
\examples{
\dontrun{
totalReads(summary)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
