% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingRate.R
\name{mappingRate}
\alias{mappingRate}
\title{Mapping rate plot}
\usage{
mappingRate(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Mapping rate plot
}
\examples{
\dontrun{
mappingRate(summary)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
