% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-methods.R
\docType{methods}
\name{markdown}
\alias{markdown}
\alias{markdown,SummarizedExperiment-method}
\title{Markdown}
\usage{
markdown(object, ...)

\S4method{markdown}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Markdown output.
}
\description{
Markdown
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
markdown(rse)
}
