% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownHeader.R
\name{markdownHeader}
\alias{markdownHeader}
\alias{mdHeader}
\title{Markdown header}
\usage{
markdownHeader(text, level = 2L, tabset = FALSE, asis = FALSE)
}
\arguments{
\item{level}{\code{integer(1)}.
Markdown header level (1-7).}
}
\value{
\itemize{
\item \code{asis = TRUE}: \code{\link[knitr:asis_output]{knitr::asis_output()}} return.
\item \code{asis = FALSE}: \code{\link[base:writeLines]{writeLines()}} return.
}
}
\description{
Generate a Markdown header (levels 1-7) in any R Markdown code block. When
calling inside an \code{asis} chunk, set \code{asis = TRUE}.
}
\note{
Updated 2019-08-11.
}
\examples{
markdownHeader("Header", level = 2L)
markdownHeader("Header", tabset = TRUE)
markdownHeader("Header", asis = TRUE)
}
\seealso{
\href{https://daringfireball.net/projects/markdown/syntax}{Markdown Syntax}.
}
