% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdownList}
\alias{markdownList}
\alias{mdList}
\title{Markdown List}
\usage{
markdownList(text, ordered = FALSE, asis = FALSE)

mdList(text, ordered = FALSE, asis = FALSE)
}
\arguments{
\item{text}{\code{string}. Header text.}

\item{ordered}{\code{boolean}. Ordered ("\code{TRUE}") or unordered ("\code{FALSE}").}

\item{asis}{\code{boolean}. Set this to \code{TRUE} when using the function inside a
loop or inside an R Markdown chunk with '\code{results = "asis"}' enabled.}
}
\value{
\itemize{
\item "\code{asis = TRUE}": \code{knitr::asis_output()}.
\item "\code{asis = FALSE}": \code{writeLines()}.
}
}
\description{
Include a Markdown-formatted list, either ordered or unordered. This function
works in any R Markdown code block. When calling from inside an \code{asis} chunk,
set \code{asis = TRUE}.
}
\examples{
groceries <- c("milk", "eggs")
markdownList(groceries)
markdownList(groceries, ordered = TRUE)
markdownList(groceries, asis = TRUE)
}
\seealso{
\href{https://daringfireball.net/projects/markdown/syntax}{Markdown Syntax}.
}
