% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdownPlotlist}
\alias{markdownPlotlist}
\alias{mdPlotlist}
\title{Markdown Plotlist}
\usage{
markdownPlotlist(plotlist, headerLevel = 2L)
}
\arguments{
\item{plotlist}{\code{list} containing \code{ggplot} objects.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). Markdown header level.}
}
\value{
Graphical output of plots, separated by Markdown headers.
}
\description{
Supports using a named \code{list} containing multiple \code{ggplot} objects, which
can be used in an R Markdown report, separated by headers. Internally, the
headers are generated with the \code{markdownHeader()} function.
}
\examples{
loadRemoteData(url = file.path(basejumpCacheURL, "plotlist.rda"))
names(plotlist)
markdownPlotlist(plotlist)
}
\seealso{
\code{cowplot::plot_grid()}.
}
