% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownTables.R
\name{markdownTables}
\alias{markdownTables}
\alias{mdTables}
\title{Multiple Markdown tables}
\usage{
markdownTables(list, captions = NULL, force = FALSE)
}
\arguments{
\item{list}{Named \code{list}.
Column data that can be coerced to \code{data.frame}.}

\item{captions}{\code{character} or \code{NULL}.
Table captions. If \code{NULL}, the names of the list will be used automatically
as captions.}

\item{force}{\code{logical(1)}.
Force knit output using \code{\link[knitr:asis_output]{knitr::asis_output()}}.
\emph{Recommended for development and unit testing only.}}
}
\value{
\itemize{
\item knit call: \code{asis_output}
\item Otherwise: \code{list}.
}
}
\description{
Knit multiple tables in a single R Markdown chunk.
}
\note{
\code{\link[knitr:kable]{knitr::kable()}} now supports multiple tables as a \code{list} for the \code{x}
argument, but it still only supports a single caption. \code{markdownTables}
extends this functionality, but supporting captions for each table.

Updated 2019-07-28.
}
\examples{
list <- list(
    iris = head(datasets::iris),
    mtcars = head(datasets::mtcars)
)
captions <- c(
    iris = "Edgar Anderson's iris data",
    mtcars = "Motor Trend car road tests"
)
markdownTables(list = list, captions = captions)
}
\seealso{
\itemize{
\item \code{knitr::kable()}.
\item \href{https://stackoverflow.com/a/35149103/3911732}{Stack Overflow post}.
}
}
