% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat}
\alias{mat}
\title{Example matrix}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\usage{
mat
}
\description{
Example matrix
}
\examples{
data(mat)
mat
}
\keyword{datasets}
