% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCall.R
\name{matchCall}
\alias{matchCall}
\title{Functions to Access the Function Call Stack}
\usage{
matchCall(verbose = FALSE)
}
\arguments{
\item{verbose}{\code{boolean}. Run the function with verbose messages? It is only
recommended to enable this when debugging.}
}
\value{
\code{call}.
}
\description{
Functions to Access the Function Call Stack
}
\note{
\code{\link[base:match.call]{base::match.call()}} doesn't always work correctly inside S4 methods. Here we
are using a combination of \code{\link[base:sys.call]{base::sys.call()}} with \code{\link[pryr:standardise_call]{pryr::standardise_call()}}
to correctly capture named arguments inside an S4 method.
}
\examples{
x <- "XXX"

# Standard call
testing <- function(object, ...) {
    matchCall()
}
testing(x)

# S4 mode
setGeneric(
    name = "testing",
    def = function(object, ...) {
        standardGeneric("testing")
    }
)

setMethod(
    f = "testing",
    signature = signature("character"),
    definition = function(object, ...) {
        matchCall()
    }
)
testing(x)

setMethod(
    f = "testing",
    signature = signature("character"),
    definition = function(object, xxx, ...) {
        matchCall()
    }
)
testing(x)
}
\seealso{
\itemize{
\item \code{\link[base:sys.calls]{base::sys.calls()}}.
\item \code{\link[base:sys.call]{base::sys.call()}}.
\item \code{\link[base:match.call]{base::match.call()}}.
\item \code{\link[pryr:standardise_call]{pryr::standardise_call()}}.
}

Other Developer Functions: \code{\link{detectHPC}},
  \code{\link{dots}}, \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
