% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchEnsemblReleaseToURL.R
\name{matchEnsemblReleaseToURL}
\alias{matchEnsemblReleaseToURL}
\title{Match Ensembl release to archive URL.}
\usage{
matchEnsemblReleaseToURL(release)
}
\arguments{
\item{release}{\code{integer(1)} or \code{character(1)}.
Ensembl release (e.g. 96).}
}
\value{
\code{character(1)}.
URL.
}
\description{
Match Ensembl release to archive URL.
}
\note{
Updated 2019-07-28.
}
\examples{
matchEnsemblReleaseToURL(96L)
}
