% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchesInterestingGroups.R
\name{matchesInterestingGroups}
\alias{matchesInterestingGroups}
\title{Check Interesting Groups Value Matching}
\usage{
matchesInterestingGroups(x, interestingGroups)
}
\arguments{
\item{x}{S4 class.}

\item{interestingGroups}{\code{character}.
Interesting groups.}
}
\value{
No value.
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
data(rse)

## Checks that columns are defined in `colData`.
matchesInterestingGroups(rse, "condition")

## Currently allowing `NULL` to pass.
matchesInterestingGroups(rse, NULL)
}
