% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchHumanOrthologs.R
\name{matchHumanOrthologs}
\alias{matchHumanOrthologs}
\title{Match human gene orthologs}
\usage{
matchHumanOrthologs(genes, organism = NULL, ensemblRelease = NULL)
}
\value{
\code{DataFrame}.
Data frame containing mapping columns:
\itemize{
\item geneID
\item hgncID
\item geneName
\item hgncName
}
}
\description{
Match human gene orthologs
}
\note{
Updated 2019-08-11.
}
\examples{
genes <- c(
    "ENSMUSG00000000001", "ENSMUSG00000000003",
    "ENSMUSG00000000028", "ENSMUSG00000000031",
    "ENSMUSG00000000037", "ENSMUSG00000000049"
)
matchHumanOrthologs(genes, ensemblRelease = 87L)
}
