% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchInterestingGroups.R
\name{matchInterestingGroups}
\alias{matchInterestingGroups}
\title{Match interesting groups}
\usage{
matchInterestingGroups(object, interestingGroups = NULL)
}
\arguments{
\item{object}{\code{SummarizedExperiment}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{character}.
Interesting groups.
}
\description{
Match the user-defined interesting groups to the values slotted within a
\code{SummarizedExperiment} object, and check that they are valid. Otherwise
supports a missing \code{interestingGroups} argument, which will then use the
\code{interestingGroups} accessor function internally.
}
\examples{
data(rse, package = "acidData")
matchInterestingGroups(rse)
matchInterestingGroups(rse, interestingGroups = NULL)
matchInterestingGroups(rse, interestingGroups = substitute())
}
