% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchS4Call.R
\name{matchS4Call}
\alias{matchS4Call}
\alias{matchS4Args}
\title{Match S4 Call}
\usage{
matchS4Call(which = -2L)

matchS4Args(which = -2L)
}
\arguments{
\item{which}{the frame number if non-negative, the number of frames
    to go back if negative.}
}
\value{
\code{call}.
f <- function(x) {
matchCall()
}
f(x = "XXX")
}
\description{
Note that \code{\link[base:match.call]{base::match.call()}} doesn't always work correctly inside S4
methods. Here we are using a combination of \code{\link[base:sys.call]{base::sys.call()}} with
\code{\link[pryr:standardise_call]{pryr::standardise_call()}} to correctly capture named arguments inside an S4
method.
}
\seealso{
Other Developer Functions: \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
