% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/matchSampleColumn-methods.R
\name{matchSampleColumn}
\alias{matchSampleColumn}
\alias{matchSampleColumn,DataFrame-method}
\alias{matchSampleColumn,SummarizedExperiment-method}
\title{Match sample identifier column}
\usage{
matchSampleColumn(object, ...)

\S4method{matchSampleColumn}{DataFrame}(object)

\S4method{matchSampleColumn}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
}
\description{
Match sample identifier column
}
\note{
Updated 2021-01-16.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
id <- matchSampleColumn(rse)
print(id)
}
