% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mcolnames-methods.R
\name{mcolnames}
\alias{mcolnames}
\alias{mcolnames<-}
\alias{mcolnames,Vector-method}
\alias{mcolnames<-,Vector,character-method}
\title{Metadata column names}
\usage{
mcolnames(x, ...)

mcolnames(x, ...) <- value

\S4method{mcolnames}{Vector}(x)

\S4method{mcolnames}{Vector,character}(x) <- value
}
\arguments{
\item{x}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Metadata column names
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## Vector ====
object <- RangedSummarizedExperiment
mcolnames(object)
## Asignment method.
mcolnames(object) <- toupper(mcolnames(object))
mcolnames(object)
}
\seealso{
\code{S4Vectors::mcols()}.
}
