% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R, R/zzz.R
\name{mdHeader}
\alias{mdHeader}
\alias{md_header}
\title{Markdown Header}
\usage{
mdHeader(string, level = 2L)
}
\arguments{
\item{string}{String.}

\item{level}{Header level (1-7).}
}
\value{
\code{\link[=writeLines]{writeLines()}}.
}
\description{
Markdown Header
}
\examples{
mdHeader("Header")
mdHeader("Header", level = 4L)
}
\seealso{
Other Markdown Utilities: \code{\link{mdList}}
}
