% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-mdPlotlist.R
\docType{methods}
\name{mdPlotlist}
\alias{mdPlotlist}
\alias{mdPlotlist}
\alias{mdPlotlist,list-method}
\title{Markdown Plotlist}
\usage{
mdPlotlist(object, ...)

\S4method{mdPlotlist}{list}(object, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{headerLevel}{Header level.}
}
\description{
Markdown Plotlist
}
\examples{
loadRemoteData("http://basejump.seq.cloud/plotlist.rda")
mdPlotlist(plotlist)
}
\seealso{
\code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.
}
\author{
Michael Steinbaugh
}
