% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt-methods.R
\docType{methods}
\name{melt}
\alias{melt}
\alias{melt,matrix-method}
\alias{melt,table-method}
\alias{melt,Matrix-method}
\alias{melt,DataFrame-method}
\alias{melt,SummarizedExperiment-method}
\alias{melt,SingleCellExperiment-method}
\title{Melt columns into key-value pairs}
\usage{
melt(object, ...)

\S4method{melt}{matrix}(object, colnames = c("rowname", "colname",
  "value"), min = NULL, minMethod = c("perRow", "absolute"),
  trans = c("identity", "log2", "log10"))

\S4method{melt}{table}(object, colnames = c("rowname", "colname",
  "value"))

\S4method{melt}{Matrix}(object, colnames = c("rowname", "colname",
  "value"), min = NULL, minMethod = c("perRow", "absolute"),
  trans = c("identity", "log2", "log10"))

\S4method{melt}{DataFrame}(object, colnames = c("rowname", "colname",
  "value"))

\S4method{melt}{SummarizedExperiment}(object, assay = 1L, min = NULL,
  minMethod = c("perRow", "absolute"), trans = c("identity", "log2",
  "log10"))

\S4method{melt}{SingleCellExperiment}(object, assay = 1L, min = NULL,
  minMethod = c("perRow", "absolute"), trans = c("identity", "log2",
  "log10"))
}
\arguments{
\item{object}{Object.}

\item{colnames}{\code{character(3)}.
Column name mappings for melted data frame return.
Currently only applies to \code{matrix} and \code{DataFrame} methods.
Standardized for \code{SummarizedExperiment} and \code{SingleCellExperiment}.}

\item{min}{\code{integer(1)} or \code{NULL}.
Minimum count threshold to apply. Filters using "greater than or equal to"
logic internally. Note that this threshold gets applied prior to
logarithmic transformation, when \code{trans} argument applies.}

\item{minMethod}{\code{character(1)}.
Only applies when \code{min} argument is numeric.
Uses \code{\link[base:match.arg]{match.arg()}}.
\itemize{
\item \code{perRow}: \emph{Recommended}. Applies cutoff per row (i.e. gene).
Internally, \code{\link[base:rowSums]{rowSums()}} values are checked against this
cutoff threshold prior to the melt operation.
\item \code{absolute}: Applies hard cutoff to \code{counts} column after the melt
operation. This applies to all counts, not per feature.
}}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Unpivot column data from wide format to long format.
}
\note{
Updated 2019-08-27.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
dim(object)
x <- melt(object)
nrow(x)
print(x)

## SingleCellExperiment ====
object <- SingleCellExperiment
dim(object)
x <- melt(object)
nrow(x)
print(x)
}
\seealso{
tidyr:\preformatted{methods("gather")
methods("gather_")
getS3method("gather", "data.frame", envir = asNamespace("tidyr"))
getS3method("gather_", "data.frame", envir = asNamespace("tidyr"))
tidyr:::melt_dataframe
}

https://github.com/tidyverse/tidyr/blob/master/src/melt.cpp
https://github.com/tidyverse/tidyr/blob/master/src/RcppExports.cpp

reshape2 (deprecated):\preformatted{help(topic = "melt.array", package = "reshape2")
methods("melt")
getS3method("melt", "data.array", envir = asNamespace("tidyr"))
getS3method("melt", "data.frame", envir = asNamespace("tidyr"))
}
}
