% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltCounts-methods.R
\docType{methods}
\name{meltCounts}
\alias{meltCounts}
\alias{meltCounts,SummarizedExperiment-method}
\title{Melt count matrix into long format}
\usage{
\S4method{meltCounts}{SummarizedExperiment}(object, assay = 1L,
  minCountsPerFeature = -Inf, trans = c("identity", "log2", "log10"),
  ...)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{minCountsPerFeature}{\code{integer(1)}.
Minimum number of counts per row feature (i.e. gene). Internally,
\code{\link[base:rowSums]{rowSums()}} values are checked against this cutoff
threshold prior to the melt operation.}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}

\item{...}{Additional arguments.}
}
\value{
\code{grouped_df}.
Grouped by \code{colname} (e.g. sample ID) and \code{rowname} (e.g. gene ID).
}
\description{
Melt count matrix into long format
}
\examples{
data(rse, package = "acidtest")
x <- meltCounts(rse)
print(x)
}
\seealso{
\code{reshape2::melt()}.
}
