% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{metadataBlacklist}
\alias{metadataBlacklist}
\title{Sample metadata blacklist}
\format{An object of class \code{character} of length 12.}
\usage{
metadataBlacklist
}
\description{
Sample metadata blacklist
}
\examples{
metadataBlacklist
}
\keyword{datasets}
