% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodDefinition.R
\name{MethodDefinition}
\alias{MethodDefinition}
\alias{methodFormals}
\alias{methodFunction}
\title{Access to the Data Inside an S4 Method Definition}
\usage{
methodFormals(f, signature)

methodFunction(f, signature)
}
\arguments{
\item{f}{a generic function or the character-string name of one.}

\item{signature}{the signature of classes to match to the arguments
    of \code{f}.  See the details below.}
}
\value{
\itemize{
\item \code{\link[=methodFormals]{methodFormals()}}: \code{list}. Extract the sealed formal arguments.
\item \code{\link[=methodDefinition]{methodDefinition()}}: \code{function}. Extract the sealed function.
}
}
\description{
Access to the Data Inside an S4 Method Definition
}
\examples{
## Function
x <- methodFunction(f = "geometricMean", signature = "numeric")
class(x)

## Formals
x <- methodFormals(f = "geometricMean", signature = "numeric")
print(x)
}
\seealso{
Modified version of \href{https://goo.gl/ymX571}{John Chambers' code}.
}
