% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{metricsCols}
\alias{metricsCols}
\title{Quality control metric columns}
\format{
An object of class \code{character} of length 6.
}
\usage{
metricsCols
}
\description{
Column names returned by \code{calculateMetrics()}.
}
\note{
Previously: "nGene", "log10GenesPerUMI".

Updated 2019-08-21.
}
\examples{
metricsCols
}
\keyword{datasets}
