% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metricsPerSample-methods.R
\docType{methods}
\name{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample,SingleCellExperiment-method}
\title{Metrics per Sample}
\usage{
metricsPerSample(object, ...)

\S4method{metricsPerSample}{SingleCellExperiment}(object, f = c("mean",
  "median", "sum"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{f}{\code{string}. Mathematical function name to apply.}
}
\value{
\code{tbl_df}.
}
\description{
Calculate summary statistics per sample.
}
\examples{
x <- metricsPerSample(sce_small, f = "mean")
print(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCellsToSamples}},
  \code{\link{aggregateCols}}, \code{\link{cell2sample}},
  \code{\link{counts}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}},
  \code{\link{topCellsPerSample}}, \code{\link{tpm}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
