% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MGI2Ensembl-class}
\alias{MGI2Ensembl-class}
\title{MGI-to-Ensembl Gene Identifier Mappings}
\value{
\code{MGI2Ensembl}. Contains a \code{DataFrame} with \code{mgiID} and \code{geneID}
columns.
}
\description{
MGI-to-Ensembl Gene Identifier Mappings
}
\examples{
options(basejump.test = TRUE)
x <- MGI2Ensembl()
print(x)
}
\keyword{internal}
