% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGI2Ensembl.R
\name{MGI2Ensembl}
\alias{MGI2Ensembl}
\title{MGI-to-Ensembl Gene Identifier Mappings}
\usage{
MGI2Ensembl()
}
\value{
\code{MGI2Ensembl}.
Contains a \code{DataFrame} with \code{mgiID} and \code{geneID} columns.
}
\description{
MGI-to-Ensembl Gene Identifier Mappings
}
\examples{
options(basejump.test = TRUE)
x <- MGI2Ensembl()
print(x)
}
