% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi2gene.R
\name{mgi2gene}
\alias{mgi2gene}
\title{MGI to Ensembl Gene ID Mappings}
\usage{
mgi2gene()
}
\value{
\code{data.frame}.
}
\description{
MGI to Ensembl Gene ID Mappings
}
\examples{
x <- mgi2gene()
glimpse(x)
}
