% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalSampleData.R
\name{minimalSampleData}
\alias{minimalSampleData}
\title{Minimal Sample Data}
\usage{
minimalSampleData(sample)
}
\arguments{
\item{sample}{\code{character}.
Sample names (e.g. per-sample directory names).}
}
\value{
\code{DataFrame}.
}
\description{
Minimal Sample Data
}
\examples{
x <- minimalSampleData(sample = c("sample 1", "sample 2"))
print(x)
}
