% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mpg}
\alias{mpg}
\title{Miles per Gallon}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 234 rows and 11 columns.}
\usage{
mpg
}
\value{
\link{tibble}.
}
\description{
Miles per Gallon
}
\keyword{internal}
