% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassignAsEnvir.R
\name{multiassignAsEnvir}
\alias{multiassignAsEnvir}
\title{Assign multiple objects into a new environment}
\usage{
multiassignAsEnvir(..., envirName, parentFrame = parent.frame())
}
\arguments{
\item{...}{Objects as dots.}

\item{envirName}{\code{character(1)}.
Name of the new \code{environment} to create.}

\item{parentFrame}{\code{environment}.
Parent \code{environment} where to assign the new \code{environment}, specified by
\code{envirName} argument.}
}
\value{
\code{character}.
Object names defined in the new \code{environment}.
}
\description{
Assign multiple objects into a new environment
}
\note{
Updated 2020-01-20.
}
\examples{
data(data.frame, matrix, package = "acidtest")
multiassignAsEnvir(data.frame, matrix, envirName = "example")
class(example)
ls(example)
}
