% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassignAsNewEnvir.R
\name{multiassignAsNewEnvir}
\alias{multiassignAsNewEnvir}
\title{Assign Multiple Objects as a New Environment}
\usage{
multiassignAsNewEnvir(..., envirName, parentEnvir = parent.frame())
}
\arguments{
\item{...}{Objects as dots.}

\item{envirName}{New environment name (\code{string}).}

\item{parentEnvir}{Parent \code{environment} where to assign the new environment.}
}
\value{
\code{character} of object names in the new \code{environment}.
}
\description{
Assign Multiple Objects as a New Environment
}
\examples{
multiassignAsNewEnvir(rnaseqCounts, singleCellCounts, envirName = "example")
class(example)
ls(example)
}
\seealso{
Other Developer Functions: \code{\link{detectHPC}},
  \code{\link{dots}}, \code{\link{localOrRemoteFile}},
  \code{\link{methodFormals}}
}
\author{
Michael Steinbaugh
}
