% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\name{namespace}
\alias{namespace}
\title{Prepare the NAMESPACE.}
\usage{
namespace()
}
\description{
Prepare the NAMESPACE.
}
\seealso{
\itemize{
\item \link{reexports}.
\item \code{\link[=tidy_verbs]{tidy_verbs()}}.
}
}
\keyword{internal}
