% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonzeroRowsAndCols-methods.R
\docType{methods}
\name{nonzeroRowsAndCols}
\alias{nonzeroRowsAndCols}
\alias{nonzeroRowsAndCols,matrix-method}
\alias{nonzeroRowsAndCols,Matrix-method}
\alias{nonzeroRowsAndCols,DelayedArray-method}
\alias{nonzeroRowsAndCols,SummarizedExperiment-method}
\title{Subset object to keep only non-zero rows and columns}
\usage{
nonzeroRowsAndCols(object, ...)

\S4method{nonzeroRowsAndCols}{matrix}(object)

\S4method{nonzeroRowsAndCols}{Matrix}(object)

\S4method{nonzeroRowsAndCols}{DelayedArray}(object)

\S4method{nonzeroRowsAndCols}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Subset object to keep only non-zero rows and columns
}
\note{
Updated 2019-08-23.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
dim(object)
x <- nonzeroRowsAndCols(object)
dim(x)

## SingleCellExperiment ====
object <- SingleCellExperiment
dim(object)
x <- nonzeroRowsAndCols(object)
dim(x)
}
