% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/zzz.R
\name{objectSize}
\alias{objectSize}
\alias{object_size}
\title{Object size with automatic units}
\usage{
objectSize(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Object size in human readable units.
}
\description{
Object size with automatic units
}
