% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organism_mappings}
\alias{organism_mappings}
\title{Organism Mappings}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 6 columns.}
\usage{
organism_mappings
}
\description{
Organism Mappings
}
\examples{
data(organism_mappings)
organism_mappings
}
\keyword{datasets}
