% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageProj.R
\name{packageProj}
\alias{packageProj}
\title{Build package and website}
\usage{
packageProj(install = FALSE, test = FALSE)
}
\arguments{
\item{install}{Install package}

\item{test}{Run tests with \code{testthat}}
}
\description{
Build package and website
}
\author{
Michael Steinbaugh
}
