% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageProject.R
\name{packageProject}
\alias{packageProject}
\title{Build package and website}
\usage{
packageProject()
}
\description{
Build package and website
}
\author{
Michael Steinbaugh
}
\keyword{dev}
\keyword{package}
\keyword{website}
