% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R, R/zzz.R
\name{packageSE}
\alias{packageSE}
\alias{package_se}
\title{Package into a \code{SummarizedExperiment}}
\usage{
packageSE(assays, colData, rowData, metadata = NULL)
}
\arguments{
\item{assays}{assays.}

\item{colData}{Sample metadata.}

\item{rowData}{\href{http://www.ensembl.org/}{Ensembl} gene annotations.}

\item{metadata}{Custom metadata.}
}
\value{
\link{SummarizedExperiment}.
}
\description{
Package into a \code{SummarizedExperiment}
}
\seealso{
\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment}.
}
\author{
Michael Steinbaugh
}
