% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PANTHER-class}
\alias{PANTHER-class}
\title{PANTHER Database Annotations}
\value{
\code{PANTHER}. Contains a \code{DataFrame}.
}
\description{
\href{{http://www.pantherdb.org}{PANTHER}} gene ontology definitions. PANTHER stands
for \strong{P}rotein \strong{AN}alysis \strong{TH}rough \strong{E}volutionary \strong{R}elationships.
}
\examples{
options(basejump.test = TRUE)
x <- PANTHER("Homo sapiens", progress = FALSE)
summary(x)
}
\keyword{internal}
