% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANTHER.R
\name{PANTHER}
\alias{PANTHER}
\title{PANTHER Database Annotations}
\usage{
PANTHER(organism, release = NULL, progress = FALSE)
}
\arguments{
\item{organism}{\code{string}. Full Latin organism name. Currently supported
organisms: \emph{Homo sapiens}, \emph{Mus musculus}, \emph{Caenorhabditis elegans},
\emph{Drosophila melanogaster}.}

\item{release}{\code{string} or \code{NULL}. PANTHER release version. If set \code{NULL},
defaults to current release. Consult the PANTHER website for a list of
release versions available from the FTP server (e.g. \code{"13.0"}).}

\item{progress}{\code{boolean}. Call \code{pbapply::pblapply()} internally to show
progress.}
}
\value{
\code{PANTHER}. Contains a \code{DataFrame}.
}
\description{
\href{http://www.pantherdb.org}{PANTHER} gene ontology definitions. PANTHER stands
for \strong{P}rotein \strong{AN}alysis \strong{TH}rough \strong{E}volutionary \strong{R}elationships.
}
\examples{
options(basejump.test = TRUE)
x <- PANTHER("Homo sapiens", progress = FALSE)
summary(x)
}
