% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{counts}{\code{matrix}.
Count matrix. Normalized counts are recommended.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{direction}{\code{character(1)}.
Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{ensemblRelease, release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{envir}{\code{environment}.
Environment to use for assignment. Defaults to
\code{\link[base:parent.frame]{parent.frame()}}, which will assign into the calling
environment.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{file}{\code{character(1)}.
File path.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available. Note: don't pass in UCSC build
IDs (e.g. \code{"hg38"}).}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{gffFile}{\code{character(1)}.
GFF/GTF (General Feature Format) file. Generally, we recommend using a GTF
(GFFv2) instead of a GFFv3 file if possible.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{inherits}{\code{logical(1)}.
Should the enclosing frames of the \code{environment} be searched?}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per gene in the count matrix.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{ntop}{\code{integer(1)}.
Number of top genes to label.}

\item{object}{Object.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{progress}{\code{logical(1)}.
Show progress, using progress bars.}

\item{reducedDim}{\code{character(1)}.
Name of reduced dimension matrix slotted in
\code{\link[SingleCellExperiment:reducedDims]{reducedDims()}}.
Includes TNSE, UMAP, PCA, for example.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{samples}{\code{character}.
Specify a subset of samples to include.}

\item{sort}{\code{logical(1)}.
Resort using \code{sort}.}

\item{tx2gene}{\code{Tx2Gene}.
Transcript-to-gene mappings.}

\item{url}{\code{character(1)}.
Uniform Resource Locator (URL). HTTP or FTP address.}

\item{value}{Value to assign.}

\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
