% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGFFAttributes.R
\name{parseGFFAttributes}
\alias{parseGFFAttributes}
\title{Parse GFF Attributes Column}
\usage{
parseGFFAttributes(file, select = c("gene_", "transcript_"), unique = TRUE)
}
\arguments{
\item{file}{File path.}

\item{select}{Attribute to select. Supports partial matching. Defaults to
returning attributes matching both "\code{gene_}" and "\code{transcript_}".}

\item{unique}{Return unique attributes.}
}
\value{
\code{tbl_df}.
}
\description{
Parse GFF Attributes Column
}
\note{
Memory overhead and speed is greatly improved by requiring keys.
Deparsing only unique values is much faster and generally recommended.
}
\examples{
parseGFFAttributes("http://basejump.seq.cloud/mmusculus.gtf") \%>\%
    glimpse()
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
