% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pct.R
\docType{methods}
\name{pct,numeric-method}
\alias{pct,numeric-method}
\title{Convert Numeric to Percentage}
\usage{
\S4method{pct}{numeric}(object)
}
\arguments{
\item{number}{Number.}
}
\value{
Percentage.
}
\description{
Convert Numeric to Percentage
}
\examples{
pct(0.1)
}
