% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pct}
\alias{pct}
\title{Convert Numeric to Percentage}
\usage{
pct(number)
}
\arguments{
\item{number}{Number.}
}
\value{
Percentage.
}
\description{
Convert Numeric to Percentage
}
\examples{
pct(0.1)
}
\seealso{
Other Math Utilities: \code{\link{fc2lr}},
  \code{\link{geomean}}
}
