% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.R
\name{percentage}
\alias{percentage}
\title{Convert numeric to percentage}
\usage{
percentage(number)
}
\arguments{
\item{number}{Number}
}
\value{
Percentage
}
\description{
Convert numeric to percentage
}
\examples{
percentage(0.1)
}
\author{
Michael Steinbaugh
}
