% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlist.R
\name{\%>\%}
\alias{\%>\%}
\alias{mdlist}
\title{Markdown list}
\usage{
mdlist(character, ordered = FALSE)
}
\arguments{
\item{character}{\code{character} vector.}

\item{ordered}{Ordered \code{TRUE} or unordered \code{FALSE} list in Markdown format.}
}
\value{
Character vector.
}
\description{
Markdown list
}
\examples{
mdlist(c("item1", "item2"), ordered = TRUE)
}

