% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plot5Prime3PrimeBias}
\alias{plot5Prime3PrimeBias}
\title{Plot 5'->3' Bias}
\usage{
plot5Prime3PrimeBias(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
RNA-seq data can have specific biases at either the 5’ or 3’ end of sequenced
fragments.
}
\details{
It is common to see a small amount of bias, especially if polyA enrichment
was performed, or if there is any sample degradation. If a large amount of
bias is observed here, be sure to analyze the samples with a Bioanalyzer and
check the RIN scores.

5' (3') bias is generally calculated as the median of the following ratio:\preformatted{[mean expression of 5' (3')] / [mean expression of whole transcript]
}

For example:
\itemize{
\item Mean expression for 5' (3') is calculated as mean coverage of first (last)
100 bases.
\item Mean expression of transcript is the mean coverage of all bases in that
transcript.
\item Median is calculated for the representative set of 1000 transcripts.
}
}
