% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotBarcodeRanks}
\alias{plotBarcodeRanks}
\title{Plot Barcode Ranks}
\usage{
plotBarcodeRanks(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} grid.
}
\description{
Plot Barcode Ranks
}
