% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotCellCounts}
\alias{plotCellCounts}
\title{Plot Cell Counts}
\usage{
plotCellCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot Cell Counts
}
