% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R
\docType{methods}
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts,SummarizedExperiment-method}
\title{Plot counts}
\usage{
\S4method{plotCounts}{SummarizedExperiment}(object, genes, assay = 1L,
  interestingGroups = NULL, countsAxisLabel = "counts",
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE),
  style = c("facet", "wide"))
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{countsAxisLabel}{\code{character(1)}.
Label to use for the counts axis.}

\item{medianLine}{\code{logical(1)}.
Include median line for each group. Disabled if samples are colored by
sample name.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{legend}{\code{logical(1)}.
Show plot legend.}

\item{style}{\code{character(1)}.
Plot style.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot counts
}
\examples{
data(rse, package = "acidData")
object <- rse

rownames <- head(rownames(object))
print(rownames)
g2s <- Gene2Symbol(object)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

## Rownames, gene IDs, and gene names (symbols) are supported.
plotCounts(object, genes = geneIDs, style = "facet")
plotCounts(object, genes = geneNames, style = "wide")
}
\seealso{
\itemize{
\item \code{DESeq2::plotCounts()}.
}
}
