% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerBiotype-methods.R
\docType{methods}
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\alias{plotCountsPerBroadClass}
\alias{plotCountsPerBiotype,SummarizedExperiment-method}
\alias{plotCountsPerBiotype,SingleCellExperiment-method}
\alias{plotCountsPerBroadClass,SummarizedExperiment-method}
\alias{plotCountsPerBroadClass,SingleCellExperiment-method}
\title{Plot Counts per Biotype}
\usage{
plotCountsPerBiotype(object, ...)

plotCountsPerBroadClass(object, ...)

\S4method{plotCountsPerBiotype}{SummarizedExperiment}(object, assay = 1L,
  n = 9L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")

\S4method{plotCountsPerBiotype}{SingleCellExperiment}(object, assay = 1L,
  n = 9L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")

\S4method{plotCountsPerBroadClass}{SummarizedExperiment}(object,
  assay = 1L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")

\S4method{plotCountsPerBroadClass}{SingleCellExperiment}(object,
  assay = 1L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of count matrix slotted in
\code{SummarizedExperiment::assays()}. When passing in a \code{string}, the name must
be defined in \code{SummarizedExperiment::assayNames()}.}

\item{n}{\code{scalar integer}. Number to include.}

\item{interestingGroups}{\code{character}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{countsAxisLabel}{\code{string}. Counts axis label.}
}
\value{
\code{ggplot}.
}
\description{
Plot Counts per Biotype
}
\examples{
data(rse, sce)

## SummarizedExperiment ====
plotCountsPerBiotype(rse)
plotCountsPerBroadClass(rse)

## SingleCellExperiment ====
plotCountsPerBiotype(sce)
plotCountsPerBroadClass(sce)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
