% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerBroadClass-methods.R
\name{plotCountsPerBroadClass}
\alias{plotCountsPerBroadClass}
\title{Plot counts per broad class definition}
\description{
Plot counts per broad class definition
}
\examples{
data(rse, sce, package = "acidData")

## SummarizedExperiment ====
plotCountsPerBroadClass(rse)

## SingleCellExperiment ====
plotCountsPerBroadClass(sce)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
