% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerGene-methods.R
\docType{methods}
\name{plotCountsPerGene}
\alias{plotCountsPerGene}
\alias{plotCountsPerGene,SummarizedExperiment-method}
\alias{plotCountsPerGene,SingleCellExperiment-method}
\title{Plot Counts Per Gene}
\usage{
plotCountsPerGene(object, ...)

\S4method{plotCountsPerGene}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, geom = c("boxplot", "density", "violin"),
  trans = c("identity", "log2", "log10"),
  color = getOption("basejump.color.discrete", NULL),
  fill = getOption("basejump.fill.discrete", NULL),
  flip = getOption("basejump.flip", TRUE), countsAxisLabel = "counts",
  title = "counts per gene")

\S4method{plotCountsPerGene}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, geom = c("boxplot", "density", "violin"),
  trans = c("identity", "log2", "log10"),
  color = getOption("basejump.color.discrete", NULL),
  fill = getOption("basejump.fill.discrete", NULL),
  flip = getOption("basejump.flip", TRUE), countsAxisLabel = "counts",
  title = "counts per gene")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of count matrix slotted in
\code{SummarizedExperiment::assays()}. When passing in a \code{string}, the name must
be defined in \code{SummarizedExperiment::assayNames()}.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{geom}{\code{string}. Type of ggplot2 geometric object to use.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())}.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{ggplot2::scale_fill_manual()}. To set the discrete fill palette
globally, use \code{options(basejump.fill.discrete = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for plots containing
many samples.}

\item{countsAxisLabel}{\code{string}. Counts axis label.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\examples{
data(rse, sce)

## SummarizedExperiment ====
plotCountsPerGene(rse, geom = "boxplot")
plotCountsPerGene(rse, geom = "density")

## SingleCellExperiment ====
plotCountsPerGene(sce)
}
