% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotDEGPCA}
\alias{plotDEGPCA}
\title{Plot DEG PCA}
\usage{
plotDEGPCA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot DEG PCA
}
