% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\title{Plot exonic mapping rate}
\usage{
plotExonicMappingRate(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot exonic mapping rate
}
\examples{
isS4(plotExonicMappingRate)
}
\seealso{
bcbioRNASeq.
}
