% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGenderMarkers-methods.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,SummarizedExperiment-method}
\alias{plotGenderMarkers,SingleCellExperiment-method}
\title{Plot Sexually Dimorphic Gender Marker Genes}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, countsAxisLabel = "counts",
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE), style = "wide")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of count matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{countsAxisLabel}{\code{string}. Label to use for the counts axis.}

\item{medianLine}{\code{boolean}. Include median line for each group. Disabled if
samples are colored by sample name.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}. To set the discrete color
palette globally, use \code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{style}{\code{string}. Plot style.}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that wraps \code{\link[=plotGene]{plotGene()}} to quickly plot known
sexually dimorphic genes, and overall gene experssion for the X and Y
chromosomes. Currently only \emph{Homo sapiens} and \emph{Mus musculus}
genomes are supported.
}
\examples{
data(rse_small)
# Minimal example doesn't contain these genes.
x <- tryCatch(
    expr = plotGenderMarkers(rse_small),
    error = function(e) e
)
print(x)
}
\seealso{
\code{\link[=plotGene]{plotGene()}}.

Other Plots: \code{\link{plotCountsPerBiotype}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotHeatmap}}, \code{\link{plotQC}},
  \code{\link{plotTotalCounts}},
  \code{\link{plotZerosVsDepth}}
}
\concept{Plots}
