% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,SummarizedExperiment-method}
\alias{plotGene,SingleCellExperiment-method}
\title{Plot Gene Expression}
\usage{
plotGene(object, ...)

\S4method{plotGene}{SummarizedExperiment}(object, genes, assay = 1L,
  interestingGroups = NULL, countsAxisLabel = "counts",
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE),
  style = c("facet", "wide"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in \code{assays}. When passing in a
string, the name must be defined in \code{assayNames}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{countsAxisLabel}{\code{character(1)}.
Label to use for the counts axis.}

\item{medianLine}{\code{logical(1)}.
Include median line for each group. Disabled if samples are colored by
sample name.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{legend}{\code{logical(1)}.
Show plot legend.}

\item{style}{\code{character(1)}.
Plot style.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{ggplot2::facet_wrap} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot Gene Expression
}
\examples{
data(rse)
object <- rse

rownames <- head(rownames(object))
print(rownames)
g2s <- Gene2Symbol(object)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

## Rownames, gene IDs, and gene names (symbols) are supported.
plotGene(object, genes = geneIDs, style = "facet")
plotGene(object, genes = geneNames, style = "wide")
}
