% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGenesDetected-methods.R
\docType{methods}
\name{plotGenesDetected}
\alias{plotGenesDetected}
\alias{plotGenesDetected,SummarizedExperiment-method}
\title{Plot Genes Detected}
\usage{
plotGenesDetected(object, ...)

\S4method{plotGenesDetected}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, limit = 0L, minCounts = 1L,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), title = "genes detected")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{minCounts}{\code{scalar integer}. Minimum number of counts per gene in the
counts matrix.}

\item{title}{\code{string}. Plot title.}
}
\description{
Plot Genes Detected
}
\examples{
plotGenesDetected(rse_small)
}
\seealso{
Other Quality Control Functions: \code{\link{plotCountsPerGene}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotTotalCounts}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
