% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotMappedReads}
\alias{plotMappedReads}
\title{Plot Mapped Reads}
\usage{
plotMappedReads(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The number of mapped reads should correspond to the number of total reads.
}
\examples{
isS4(plotMappedReads)
}
\seealso{
bcbioRNASeq.
}
