% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotMeanSD}
\alias{plotMeanSD}
\title{Plot row standard deviations vs. row means}
\usage{
plotMeanSD(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot row standard deviations vs. row means
}
\examples{
isS4(plotMeanSD)
}
\seealso{
\href{https://steinbaugh.com/DESeqAnalysis/}{DESeqAnalysis}.
}
