% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA-methods.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,SummarizedExperiment-method}
\title{Principal component analysis plot}
\usage{
\S4method{plotPCA}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, ntop = 500L,
  label = getOption("basejump.label", default = FALSE),
  color = getOption("basejump.color.discrete", default = NULL),
  pointSize = getOption("basejump.point.size", default = 3L),
  title = "PCA", subtitle = NULL, return = c("ggplot", "DataFrame"),
  ...)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{ntop}{\code{integer(1)} or \code{Inf}.
Number of most variable genes to plot.
Use \code{Inf} to include all genes (\emph{not recommended}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.}

\item{title}{\code{character(1)}.
Plot title.}

\item{subtitle}{\code{character(1)}.
Plot subtitle.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} or \code{DataFrame}.
}
\description{
Principal component analysis plot
}
\section{Principal component analysis}{


PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.
}

\examples{
data(rse, package = "acidData")
plotPCA(rse, label = FALSE)
plotPCA(rse, label = TRUE)
}
\references{
Jolliffe, et al., 2002.
}
\seealso{
\code{DESeq2::plotPCA()}.

We're using a modified version of the \code{DESeqTransform} method here.\preformatted{methodFunction(
    f = "plotPCA",
    signature = "DESeqTransform",
    package = "DESeq2"
)
}
}
