% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SummarizedExperiment-method}
\title{Plot quality control metrics}
\usage{
plotQC(object, ...)

\S4method{plotQC}{SummarizedExperiment}(object, assay = 1L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\link[SummarizedExperiment:assays]{assays}. When passing in a string, the name
must be defined in \link[SummarizedExperiment:assayNames]{assayNames}.}
}
\value{
\code{ggplot}.
}
\description{
Plot quality control metrics
}
\examples{
data(rse)
plotQC(rse)
}
