% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotReadsPerCell}
\alias{plotReadsPerCell}
\title{Plot Read Counts per Cell}
\usage{
plotReadsPerCell(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot the distribution of read counts for all unfiltered cellular barcodes.
}
