% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTotalCounts-methods.R
\docType{methods}
\name{plotTotalCounts}
\alias{plotTotalCounts}
\alias{plotTotalCounts,SummarizedExperiment-method}
\alias{plotTotalCounts,SingleCellExperiment-method}
\title{Plot Total Counts}
\usage{
plotTotalCounts(object, ...)

\S4method{plotTotalCounts}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, perMillion = FALSE,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), title = "total counts")

\S4method{plotTotalCounts}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, perMillion = FALSE,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), title = "total counts")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of count matrix slotted in
\code{SummarizedExperiment::assays()}. When passing in a \code{string}, the name must
be defined in \code{SummarizedExperiment::assayNames()}.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{perMillion}{\code{boolean}. Display as counts per million.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{ggplot2::scale_fill_manual()}. To set the discrete fill palette
globally, use \code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for plots containing
many samples.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Total Counts
}
\examples{
data(rse, sce)
plotTotalCounts(rse)
plotTotalCounts(sce)
}
