% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotTotalReads}
\alias{plotTotalReads}
\title{Plot Total Reads}
\usage{
plotTotalReads(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
High quality RNA-seq samples ideally should have at least 10 million reads
per sample.
}
\examples{
isS4(plotUMIsVsGenes)
}
\seealso{
bcbioRNASeq.
}
