% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\title{Plot UMIs per cell}
\usage{
plotUMIsPerCell(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\examples{
isS4(plotUMIsVsGenes)
}
\seealso{
\itemize{
\item \href{http://bioinformatics.sph.harvard.edu/bcbioSingleCell/}{bcbioSingleCell}.
\item \href{https://github.com/steinbaugh/Chromium/}{Chromium}.
}
}
