% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\title{Plot UMI and Gene Correlation}
\usage{
plotUMIsVsGenes(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot UMI and Gene Correlation
}
\examples{
isS4(plotUMIsVsGenes)
}
\seealso{
\itemize{
\item bcbioSingleCell.
\item Chromium.
}
}
