% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot Volcano}
\usage{
plotVolcano(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot Volcano
}
\note{
Don't apply post hoc alpha level or LFC threshold cutoffs.
}
