% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTemplate.R
\name{prepareTemplate}
\alias{prepareTemplate}
\title{Prepare R Markdown template}
\usage{
prepareTemplate(package = NULL, overwrite = FALSE)
}
\arguments{
\item{package}{\code{character(1)} or \code{NULL}.
Name of package containing the R Markdown template. If \code{NULL}, only the
basic shared files defined in the basejump package will be copied.}

\item{overwrite}{\code{logical(1)}.
Should existing destination files be overwritten?}
}
\value{
Invisible \code{logical}.
Was the file copied?.
}
\description{
If the required template dependency files aren't present, copy them from the
requested package. Existing files are not overwritten by default. This
function will copy dependency files from a requested package inside the
\code{rmarkdown/shared/} directory. If a package doesn't contain this
subdirectory, the function will return an error.
}
\details{
This code is used internally by:
\itemize{
\item \code{bcbioRNASeq::prepareRNASeqTemplate()}.
\item \code{bcbioSingleCell::prepareSingleCellTemplate()}.
}
}
\note{
Updated 2019-10-10.
}
\examples{
## RNA-seq template.
## > prepareTemplate(package = "bcbioRNASeq")

## Single-cell RNA-seq template.
## > prepareTemplate(package = "bcbioSingleCell")
}
