% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printString.R
\name{printString}
\alias{printString}
\title{Print String}
\usage{
printString(x, max = 100L)
}
\arguments{
\item{x}{\code{atomic}.}

\item{max}{\code{scalar integer}. Maximum length of vector. Works like
\code{getOption("max.print")} without having to set globally.}
}
\value{
\code{character(1)}.
}
\description{
Capture \code{print} output of an \code{atomic} vector. Useful for returning
informative messages inside a function.
}
\examples{
printString(c("hello", "world"))
}
\seealso{
\code{cat}.
}
