% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTable.R
\name{printTable}
\alias{printTable}
\title{Smart data.frame printing}
\usage{
printTable(df, ...)
}
\arguments{
\item{df}{\code{data.frame}}

\item{...}{Passthrough to \code{kable()} during \code{RMarkdown} rendering
with \code{knitr}}
}
\value{
\code{data.frame} or \code{kable}, depending on the call
}
\description{
Display as `data.frame` in RMarkdown chunk output or `kable()` during a
`knitr` call
}
\keyword{report}
