% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{createProjDirs}
\alias{createProjDirs}
\alias{create_proj_dirs}
\alias{detectHPC}
\alias{detect_hpc}
\alias{packageProj}
\alias{package_proj}
\alias{renderProj}
\alias{render_proj}
\alias{clearWarnings}
\alias{clear_warnings}
\title{Project management utilities}
\usage{
createProjDirs()

create_proj_dirs()

detectHPC()

detect_hpc()

packageProj(install = FALSE, test = FALSE)

package_proj(install = FALSE, test = FALSE)

renderProj(outputDir = file.path("docs", Sys.Date()), recursive = FALSE)

render_proj(outputDir = file.path("docs", Sys.Date()), recursive = FALSE)

clearWarnings()

clear_warnings()
}
\arguments{
\item{install}{Install package.}

\item{test}{Run tests with \code{\link[devtools:test]{devtools::test()}}.}

\item{outputDir}{Output directory.}

\item{recursive}{Find files recursively.}
}
\description{
Create necessary directory structure for an
\href{http://rmarkdown.rstudio.com}{RMarkdown} report in
\href{https://www.rstudio.com}{RStudio}.

Detect if R is running on an HPC cluster.

Package project and build website.

Render all RMarkdown files in working directory.

Clear warnings.
}
\examples{
\dontrun{
renderProj()
}
\dontrun{
clearWarnings()
}
}
