% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Protein2Gene-class}
\alias{Protein2Gene-class}
\title{Protein-to-gene mappings}
\value{
\code{Protein2Gene}.
}
\description{
Protein-to-gene mappings
}
\details{
Contains a \code{DataFrame} with \code{proteinID}, \code{geneID}, and \code{geneName} columns.
}
\note{
Updated 2020-10-05.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:Annotated-class]{metadata()}}.
}

