% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{rdataLoadError}
\alias{rdataLoadError}
\title{R Data Load Error}
\format{An object of class \code{character} of length 1.}
\usage{
rdataLoadError
}
\description{
R Data Load Error
}
\examples{
message(rdataLoadError)
}
\keyword{datasets}
