% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{RdTags}
\alias{RdTags}
\title{R documentation tags}
\usage{
RdTags(object)
}
\arguments{
\item{object}{\code{Rd}.
R documentation, returned from \code{tools::Rd_db}.}
}
\description{
Modified version of the unexported \code{tools:::RdTags} function.
}
\examples{
db <- tools::Rd_db("base")
Rd <- db[["nrow.Rd"]]
RdTags(Rd)
}
