% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readDataRaw}
\alias{readDataRaw}
\alias{read_data_raw}
\alias{readYAML}
\alias{read_yaml}
\title{Read utilities}
\usage{
readDataRaw()

read_data_raw()

readYAML(yaml_file)

read_yaml(yaml_file)
}
\arguments{
\item{yaml_file}{YAML file.}
}
\description{
Read CSV files inside \code{data-raw} directory.

Read YAML file into a list.
}
