% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataRaw.R
\name{readDataRaw}
\alias{readDataRaw}
\title{Read CSV Files Inside \code{data-raw/} Directory}
\usage{
readDataRaw()
}
\value{
No value.
}
\description{
This function automatically loads raw CSVs using the readr package and
saves the data as R binary data into the \code{data/} directory.
}
