% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R, R/zzz.R
\name{readFileByExtension}
\alias{readFileByExtension}
\alias{read_file_by_extension}
\title{Read file by extension}
\usage{
readFileByExtension(file, ...)
}
\arguments{
\item{file}{File path.}

\item{...}{Additional parameters.}
}
\value{
\link{tibble} by default, or a sparse matrix for \code{.mtx} files.
}
\description{
Supports automatic loading of standard \code{.csv}, \code{.mtx}, \code{.tsv}, and \code{.xlsx}
files. Also supports bcbio-nextgen pipeline-specific \code{.counts}, \code{.colnames},
and \code{.rownames} files.
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \code{\link[Matrix:readMM]{Matrix::readMM()}}: Read a MatrixMarket file.
}
}
