% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJSON.R
\name{readJSON}
\alias{readJSON}
\title{Read JSON File}
\usage{
readJSON(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{list}.
}
\description{
Read JSON File
}
\examples{
x <- readJSON("http://basejump.seq.cloud/example.json")
names(x)
}
\seealso{
Other Read Functions: \code{\link{import}},
  \code{\link{loadDataAsName}}, \code{\link{loadData}},
  \code{\link{loadRemoteData}}, \code{\link{readGFF}},
  \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
