% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readLogFile.R
\docType{methods}
\name{readLogFile}
\alias{readLogFile}
\alias{readLogFile,character-method}
\title{Read Log File}
\usage{
readLogFile(object, ...)

\S4method{readLogFile}{character}(object)
}
\arguments{
\item{object}{Log file.}
}
\value{
Character vector.
}
\description{
Read Log File
}
\concept{bcbio Project Directory File Utilities}
\keyword{internal}
