% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readProgramVersions.R
\docType{methods}
\name{readProgramVersions}
\alias{readProgramVersions}
\alias{readProgramVersions,character-method}
\title{Read Program Versions}
\usage{
readProgramVersions(object, ...)

\S4method{readProgramVersions}{character}(object, quiet = FALSE)
}
\arguments{
\item{object}{Project directory path (character vector).}

\item{...}{Object names as symbols.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Read Program Versions
}
\examples{
readProgramVersions(file.path(testDataURL, "programs.txt"))
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readDataVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAMLMetrics}}, \code{\link{sampleYAML}}
}
\concept{bcbio Utilities}
\keyword{internal}
