% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-readSampleMetadataFile.R
\docType{methods}
\name{readSampleMetadataFile}
\alias{readSampleMetadataFile}
\alias{readSampleMetadataFile,character-method}
\title{Read Sample Metadata File}
\usage{
readSampleMetadataFile(object, ...)

\S4method{readSampleMetadataFile}{character}(object, lanes = 1,
  quiet = FALSE)
}
\arguments{
\item{object}{Metadata file. Supports CSV and XLSX file formats.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{lanes}{\emph{Optional}. Number of lanes used to split the samples into
technical replicates (\code{_LXXX}) suffix.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Read Sample Metadata File
}
\examples{
# Demultiplexed
demultiplexed <- file.path(
    "http://basejump.seq.cloud",
    "sample_metadata",
    "demultiplexed.xlsx")
readSampleMetadataFile(demultiplexed)

# Multiplexed (e.g. inDrop single-cell RNA-seq)
multiplexed <- file.path(
    "http://basejump.seq.cloud",
    "sample_metadata",
    "multiplexed.xlsx")
readSampleMetadataFile(multiplexed)
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
\concept{Data Import and Project Utilities}
