% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTx2Gene.R
\name{readTx2Gene}
\alias{readTx2Gene}
\title{Read transcript-to-gene annotations}
\usage{
readTx2Gene(file, organism = NULL, genomeBuild = NULL,
  ensemblRelease = NULL)
}
\value{
\code{Tx2Gene}.
}
\description{
Generates a \code{Tx2Gene} object containing \code{transcriptID} and \code{geneID} columns.
}
\note{
File should not contain column header names.

The functon doesn't attempt to strip transcript versions. Use
\code{\link[=stripTranscriptVersions]{stripTranscriptVersions()}} in a separate call, if necessary.

Updated 2019-08-13.
}
\examples{
file <- file.path(basejumpTestsURL, "tx2gene.csv")
x <- readTx2Gene(
    file = file,
    organism = "Mus musculus",
    genomeBuild = "GRCm38",
    ensemblRelease = 90L
)
print(x)
}
