% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-readYAML.R
\docType{methods}
\name{readYAML}
\alias{readYAML}
\alias{readYAML,character-method}
\title{Read YAML File}
\usage{
readYAML(object, ...)

\S4method{readYAML}{character}(object)
}
\arguments{
\item{object}{YAML file path.}

\item{...}{Additional arguments.}
}
\value{
\link{list}.
}
\description{
Read YAML File
}
