% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-classes.R
\name{reexports-classes}
\alias{reexports-classes}
\alias{Ensembl2Entrez-class}
\alias{Entrez2Ensembl-class}
\alias{Gene2Symbol-class}
\alias{HGNC-class}
\alias{HGNC2Ensembl-class}
\alias{MGI2Ensembl-class}
\alias{Protein2Gene-class}
\alias{Tx2Gene-class}
\title{Reexport S4 classes}
\description{
Reexport S4 classes
}
\note{
Updated 2020-10-09.
}
\keyword{internal}
