% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-transformer.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Coerce to \code{data.frame}}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{row.names, optional}{Refer to base \code{\link[base:as.data.frame]{as.data.frame()}} for details.}
}
\description{
Coerce to \code{data.frame}
}
\note{
Refer to \code{\link[transformer:as.data.frame]{transformer::as.data.frame()}} documentation for details.
}
\keyword{internal}
