% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R
\name{atomize}
\alias{atomize}
\title{Atomize}
\usage{
atomize(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Ensure all columns in the object are \code{atomic}.
}
\note{
Refer to \code{\link[brio:atomize]{brio::atomize()}} documentation for details.
}
\keyword{internal}
