% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-syntactic.R
\name{camel}
\alias{camel}
\title{Camel case}
\usage{
camel(object, ...)
}
\description{
Format character strings to use (lower) camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{thisIsCamelCase}).
}
\note{
Refer to \code{\link[syntactic:camel]{syntactic::camel()}} documentation for details.
}
\keyword{internal}
