% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-syntactic.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize}
\usage{
capitalize(object, ...)
}
\description{
Capitalize
}
\note{
Refer to \code{\link[syntactic:capitalize]{syntactic::capitalize()}} documentation for details.
}
\keyword{internal}
