% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R
\name{decode}
\alias{decode}
\title{Decode column data that uses run-length encoding}
\usage{
decode(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\description{
The Rle class is a general container for storing an atomic vector
  that is stored in a run-length encoding format. It is based on the
  \code{\link[base]{rle}} function from the base package.
}
\note{
Refer to \code{\link[brio:decode]{brio::decode()}} documentation for details.
}
\keyword{internal}
