% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-syntactic.R
\name{dotted}
\alias{dotted}
\title{Dotted case}
\usage{
dotted(object, ...)
}
\description{
Dotted case
}
\note{
Refer to \code{\link[syntactic:dotted]{syntactic::dotted()}} documentation for details.
}
\keyword{internal}
