% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R
\name{encode}
\alias{encode}
\title{Encode using run-length encoding}
\usage{
encode(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\description{
Encode using run-length encoding
}
\note{
Refer to \code{\link[brio:encode]{brio::encode()}} documentation for details.
}
\keyword{internal}
