% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R
\name{export}
\alias{export}
\title{Export}
\usage{
export(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\description{
Export data out of R and write to disk.
}
\note{
Refer to \code{\link[brio:export]{brio::export()}} documentation for details.
}
\keyword{internal}
