% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R
\name{factorize}
\alias{factorize}
\title{Factorize}
\usage{
factorize(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Coerce all columns in the object to \code{factor} (i.e. \code{stringsAsFactors}).
}
\note{
Refer to \code{\link[brio:factorize]{brio::factorize()}} documentation for details.
}
\keyword{internal}
