% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-transformer.R
\name{flatFiles}
\alias{flatFiles}
\title{Coerce any S4 object to a list}
\usage{
flatFiles(object, ...)
}
\description{
Coerce any S4 object to a list
}
\note{
Refer to \code{\link[transformer:flatFiles]{transformer::flatFiles()}} documentation for details.
}
\keyword{internal}
