% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-syntactic.R
\name{kebab}
\alias{kebab}
\title{Kebab case}
\usage{
kebab(object, ...)
}
\description{
Format character strings to use kebab-style formatting, where word boundaries
are defined by dashes/hyphens (e.g. \code{this-is-kebab-case}).
}
\note{
Refer to \code{\link[syntactic:kebab]{syntactic::kebab()}} documentation for details.
}
\keyword{internal}
