% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-bioverbs.R
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\title{Differentially expressed gene heatmap}
\usage{
plotDEGHeatmap(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
This function is an extension of \code{\link[=plotHeatmap]{plotHeatmap()}} that is optimized for
automatic handling differentially expressed genes, rather than requiring
manual input of a gene vector. All of the optional parameters for
\code{\link[=plotHeatmap]{plotHeatmap()}} are also supported by this function.
}
\note{
This generic is reexported here to maintain backward compatibility for
bcbioRNASeq v0.2 release series, and will be removed in a future update.
}
\keyword{internal}
