% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-bioverbs.R
\name{plotDEGPCA}
\alias{plotDEGPCA}
\title{Plot differentially expressed gene principal component analysis}
\usage{
plotDEGPCA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
This function is an extension of \code{\link[BiocGenerics:plotPCA]{plotPCA()}} that is
optimized for automatic handling of differentially expressed genes, rather
than requiring manual input of a gene vector or subset object.
}
\note{
This generic is reexported here to maintain backward compatibility for
bcbioRNASeq v0.2 release series, and will be removed in a future update.
}
\keyword{internal}
