% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R
\name{removeNA}
\alias{removeNA}
\title{Remove rows and columns containing only NA values}
\usage{
removeNA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Remove rows and columns containing only NA values
}
\note{
Refer to \code{\link[brio:removeNA]{brio::removeNA()}} documentation for details.
}
\keyword{internal}
