% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R
\name{sanitizeNA}
\alias{sanitizeNA}
\title{Sanitize NA values}
\usage{
sanitizeNA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Standardize empty strings (\code{""}), character NAs (\code{"NA"}), and \code{NULL} values
inside a character vector to \code{NA_character_}. Other \code{atomic} data types are
returned unmodified.
}
\note{
Refer to \code{\link[brio:sanitizeNA]{brio::sanitizeNA()}} documentation for details.
}
\keyword{internal}
