% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R
\name{sanitizePercent}
\alias{sanitizePercent}
\title{Sanitize percentage}
\usage{
sanitizePercent(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Take a \code{character} vector of percentages (e.g. \code{"50\%"}) and convert it to a
\code{numeric} vector (e.g. \code{0.5}).
}
\note{
Refer to \code{\link[brio:sanitizePercent]{brio::sanitizePercent()}} documentation for details.
}
\keyword{internal}
