% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-syntactic.R
\name{snake}
\alias{snake}
\title{Snake case}
\usage{
snake(object, ...)
}
\description{
Format character strings to use snake-style formatting, where word boundaries
are defined by underscores (e.g. \code{this_is_snake_case}).
}
\note{
Refer to \code{\link[syntactic:snake]{syntactic::snake()}} documentation for details.
}
\keyword{internal}
