% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-syntactic.R
\name{upperCamel}
\alias{upperCamel}
\title{Upper camel case}
\usage{
upperCamel(object, ...)
}
\description{
Format character strings to use upper camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{ThisIsCamelCase}).
}
\note{
Refer to \code{\link[syntactic:upperCamel]{syntactic::upperCamel()}} documentation for details.
}
\keyword{internal}
