% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plotGene}
\alias{GRanges}
\alias{GRangesList}
\alias{IRanges}
\alias{getURL}
\alias{DataFrame}
\alias{SimpleList}
\alias{mcols}
\alias{mcols<-}
\alias{metadata}
\alias{metadata<-}
\alias{reducedDims}
\alias{reducedDims<-}
\alias{assayNames}
\alias{assay}
\alias{assay<-}
\alias{assays}
\alias{assays<-}
\alias{colData}
\alias{colData<-}
\alias{rowData}
\alias{rowData<-}
\alias{rowRanges}
\alias{rowRanges<-}
\alias{basenameSansExt}
\alias{compress}
\alias{compressExtPattern}
\alias{decompress}
\alias{dots}
\alias{extPattern}
\alias{fileExt}
\alias{initDir}
\alias{matchArgsToDoCall}
\alias{methodFormals}
\alias{methodFunction}
\alias{pasteURL}
\alias{printString}
\alias{realpath}
\alias{standardizeCall}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{as.SummarizedExperiment}
\alias{as.data.frame}
\alias{as.data.table}
\alias{as_tibble}
\alias{assignAndSaveData}
\alias{atomize}
\alias{coerceToList}
\alias{decode}
\alias{encode}
\alias{export}
\alias{factorize}
\alias{import}
\alias{loadData}
\alias{loadDataAsName}
\alias{loadRemoteData}
\alias{localOrRemoteFile}
\alias{metadata2}
\alias{metadata2<-}
\alias{naStrings}
\alias{removeNA}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{saveData}
\alias{camel}
\alias{capitalize}
\alias{dotted}
\alias{kebab}
\alias{kebabCase}
\alias{makeDimnames}
\alias{makeLabel}
\alias{makeNames}
\alias{makeTitle}
\alias{makeWords}
\alias{snake}
\alias{upperCamel}
\alias{tibble}
\alias{tar}
\alias{untar}
\alias{unzip}
\alias{zip}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{acidbase}{\code{\link[acidbase]{basenameSansExt}}, \code{\link[acidbase]{compress}}, \code{\link[acidbase]{compressExtPattern}}, \code{\link[acidbase]{decompress}}, \code{\link[acidbase]{dots}}, \code{\link[acidbase]{extPattern}}, \code{\link[acidbase]{fileExt}}, \code{\link[acidbase]{initDir}}, \code{\link[acidbase]{matchArgsToDoCall}}, \code{\link[acidbase]{methodFormals}}, \code{\link[acidbase]{methodFunction}}, \code{\link[acidbase]{pasteURL}}, \code{\link[acidbase]{printString}}, \code{\link[acidbase]{realpath}}, \code{\link[acidbase]{standardizeCall}}}

  \item{acidgenerics}{\code{\link[acidgenerics]{plotGene}}}

  \item{GenomicRanges}{\code{\link[GenomicRanges]{GRanges}}, \code{\link[GenomicRanges]{GRangesList}}}

  \item{IRanges}{\code{\link[IRanges]{IRanges}}}

  \item{magrittr}{\code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%$\%}}, \code{\link[magrittr]{\%T>\%}}}

  \item{pipette}{\code{\link[pipette]{as_tibble}}, \code{\link[pipette]{as.data.frame}}, \code{\link[pipette]{as.data.table}}, \code{\link[pipette]{as.SummarizedExperiment}}, \code{\link[pipette]{assignAndSaveData}}, \code{\link[pipette]{atomize}}, \code{\link[pipette]{coerceToList}}, \code{\link[pipette]{decode}}, \code{\link[pipette]{encode}}, \code{\link[pipette]{export}}, \code{\link[pipette]{factorize}}, \code{\link[pipette]{import}}, \code{\link[pipette]{loadData}}, \code{\link[pipette]{loadDataAsName}}, \code{\link[pipette]{loadRemoteData}}, \code{\link[pipette]{localOrRemoteFile}}, \code{\link[pipette]{metadata2}}, \code{\link[pipette]{metadata2<-}}, \code{\link[pipette]{naStrings}}, \code{\link[pipette]{removeNA}}, \code{\link[pipette]{sanitizeNA}}, \code{\link[pipette]{sanitizePercent}}, \code{\link[pipette]{saveData}}}

  \item{RCurl}{\code{\link[RCurl]{getURL}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{mcols}}, \code{\link[S4Vectors]{mcols<-}}, \code{\link[S4Vectors]{metadata}}, \code{\link[S4Vectors]{metadata<-}}, \code{\link[S4Vectors]{SimpleList}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{reducedDims}}, \code{\link[SingleCellExperiment]{reducedDims<-}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment]{assay}}, \code{\link[SummarizedExperiment]{assay<-}}, \code{\link[SummarizedExperiment]{assayNames}}, \code{\link[SummarizedExperiment]{assays}}, \code{\link[SummarizedExperiment]{assays<-}}, \code{\link[SummarizedExperiment]{colData}}, \code{\link[SummarizedExperiment]{colData<-}}, \code{\link[SummarizedExperiment]{rowData}}, \code{\link[SummarizedExperiment]{rowData<-}}, \code{\link[SummarizedExperiment]{rowRanges}}, \code{\link[SummarizedExperiment]{rowRanges<-}}}

  \item{syntactic}{\code{\link[syntactic]{camel}}, \code{\link[syntactic]{capitalize}}, \code{\link[syntactic]{dotted}}, \code{\link[syntactic]{kebab}}, \code{\link[syntactic]{kebabCase}}, \code{\link[syntactic]{makeDimnames}}, \code{\link[syntactic]{makeLabel}}, \code{\link[syntactic]{makeNames}}, \code{\link[syntactic]{makeTitle}}, \code{\link[syntactic]{makeWords}}, \code{\link[syntactic]{snake}}, \code{\link[syntactic]{upperCamel}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{utils}{\code{\link[utils]{tar}}, \code{\link[utils]{untar}}, \code{\link[utils]{unzip}}, \code{\link[utils]{zip}}}
}}

