% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{relevel}
\alias{relevel}
\alias{relevelRowRanges}
\alias{relevelColData}
\title{Relevel row or column data}
\usage{
relevelRowRanges(object)

relevelColData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
Factor levels will be readjusted (i.e. superfluous levels are dropped).
}
\description{
Relevel row or column data
}
\examples{
data(rse, package = "acidData")

rowRanges <- SummarizedExperiment::rowRanges(rse)
x <- relevelRowRanges(rowRanges)
summary(x)

colData <- SummarizedExperiment::colData(rse)
x <- relevelColData(colData)
summary(x)
}
