% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeNA}
\alias{removeNA}
\alias{remove_na}
\title{Remove rows and columns containing only \code{NA} values}
\usage{
removeNA(x)
}
\arguments{
\item{x}{Object with column data (e.g. data frame, matrix).}
}
\value{
Sanitized data.
}
\description{
Remove rows and columns containing only \code{NA} values
}
\examples{
# Remove NA only rows and columns
matrix(c(1, NA, 3,
         NA, NA, NA,
         2, NA, 4),
       nrow = 3, ncol = 3) \%>\% removeNA

data.frame(a = c("A", NA, "C"),
           b = c(NA, NA, NA),
           c = c("B", NA, "D")) \%>\% removeNA

tibble(a = c("A", NA, "C"),
       b = c(NA, NA, NA),
       c = c("B", NA, "D")) \%>\% removeNA


# Return unmodified
list(a = c("A", NA, "C"),
     b = c(NA, NA, NA),
     c = c("B", NA, "D")) \%>\% removeNA
}
