% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{renamed}
\alias{renamed}
\alias{createNewProject}
\alias{dotNotation}
\alias{grepToString}
\alias{grep_to_string}
\alias{loadDataRaw}
\alias{toStringSummarize}
\title{Renamed functions}
\description{
Renamed functions
}
