% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderAll.R
\name{renderAll}
\alias{renderAll}
\title{Render all RMarkdown files in working directory}
\usage{
renderAll(dir = "docs", ...)
}
\arguments{
\item{dir}{Output directory.}

\item{...}{Optional \code{rmarkdown} parameters}
}
\description{
Render all RMarkdown files in working directory
}
