% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderProj.R
\name{renderProj}
\alias{renderProj}
\alias{render_proj}
\title{Render all RMarkdown files in working directory.}
\usage{
renderProj(outputDir = file.path("docs", Sys.Date()), recursive = FALSE)
}
\arguments{
\item{outputDir}{Output directory.}

\item{recursive}{Find files recursively.}
}
\description{
Render all RMarkdown files in working directory.
}
\examples{
\dontrun{
renderProj()
}
}
