% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaseq_counts}
\alias{rnaseq_counts}
\alias{rnaseqCounts}
\title{DEPRECATED Example RNA-Seq Counts}
\format{An object of class \code{matrix} with 1000 rows and 12 columns.}
\usage{
rnaseq_counts
}
\description{
Generated with \code{DESeq2::makeExampleDESeqDataSet()}.
}
\examples{
data(rnaseq_counts)
summary(rnaseq_counts)
}
\keyword{internal}
