% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rnaseqCounts}
\alias{rnaseqCounts}
\title{Example RNA-Seq Counts}
\format{An object of class \code{matrix} with 1000 rows and 12 columns.}
\usage{
rnaseqCounts
}
\description{
Generated with \code{DESeq2::makeExampleDESeqDataSet()}.
}
\examples{
summary(rnaseqCounts)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
