% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCollapse.R
\name{rowCollapse}
\alias{rowCollapse}
\title{Collapse rows in a data.frame.}
\usage{
rowCollapse(df)
}
\arguments{
\item{df}{long data.frame.}
}
\value{
collapsed data.frame.
}
\description{
Collapse rows in a data.frame.
}

