% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowRangesFromGFF.R
\name{rowRangesFromGFF}
\alias{rowRangesFromGFF}
\alias{rowRangesFromGTF}
\title{Define Row Annotations from GFF/GTF File}
\usage{
rowRangesFromGFF(file, level = c("genes", "transcripts"))

rowRangesFromGTF(file, level = c("genes", "transcripts"))
}
\value{
\code{data.frame}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
file <- "http://basejump.seq.cloud/mmusculus.gtf"

# Genes
x <- rowRangesFromGFF(file, level = "genes")
head(names(x))
head(mcols(x))

# Transcripts
x <- rowRangesFromGFF(file, level = "transcripts")
head(names(x))
head(mcols(x))
}
\seealso{
Other GFF Functions: \code{\link{gene2symbolFromGFF}},
  \code{\link{readGFF}}, \code{\link{tx2geneFromGFF}}
}
