% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrnaMappingRate.R
\name{rrnaMappingRate}
\alias{rrnaMappingRate}
\title{rRNA contamination mapping rate}
\usage{
rrnaMappingRate(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
rRNA contamination mapping rate
}
\examples{
\dontrun{
rrnaMappingRate(summary)
}
}
\author{
Michael Steinbaugh
}
