% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse}
\alias{rse}
\title{Example RNA-Seq Data Set}
\format{An object of class \code{RangedSummarizedExperiment} with 50 rows and 4 columns.}
\usage{
rse
}
\description{
Generated with \code{DESeq2::makeExampleDESeqDataSet()}. Contains actual gene
identifiers and Ensembl annotations.
}
\examples{
data(rse)
rse
}
\keyword{datasets}
