% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse_bcb}
\alias{rse_bcb}
\title{DEPRECATED Example RNA-Seq DataSet}
\format{An object of class \code{RangedSummarizedExperiment} with 500 rows and 6 columns.}
\usage{
rse_bcb
}
\description{
DEPRECATED Example RNA-Seq DataSet
}
\examples{
data(rse_bcb)
rse_bcb
}
\keyword{internal}
