% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse_dds}
\alias{rse_dds}
\title{DEPRECATED Example DESeq2 Data Set}
\format{An object of class \code{RangedSummarizedExperiment} with 1000 rows and 12 columns.}
\usage{
rse_dds
}
\description{
DEPRECATED Example DESeq2 Data Set
}
\examples{
data(rse_dds)
rse_dds
}
\keyword{internal}
