% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse_small}
\alias{rse_small}
\title{Example RNA-Seq Data Set}
\format{An object of class \code{RangedSummarizedExperiment} with 100 rows and 4 columns.}
\usage{
rse_small
}
\description{
Generated with \code{DESeq2::makeExampleDESeqDataSet()}. Contains actual gene
identifiers and Ensembl annotations.
}
\examples{
print(rse_small)
}
\seealso{
Other Minimal Example Data: \code{\link{tx_se_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
