% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDavid.R
\name{runDavid}
\alias{runDavid}
\title{DAVID enrichment analysis}
\usage{
runDavid(foreground, background = NULL, idType = "ENSEMBL_GENE_ID",
  writeFiles = TRUE, writePrefix = NULL, writeDir = "results/david",
  count = 3, fdr = 0.1)
}
\arguments{
\item{foreground}{Foreground identifiers}

\item{background}{Background identifiers}

\item{idType}{Identifier type (see DAVID website)}

\item{writeFiles}{Write files to disk (\code{TRUE/FALSE})}

\item{writePrefix}{Add an optional prefix to the file names}

\item{writeDir}{Directory to write the TSV files}

\item{count}{Minimum hit count}

\item{fdr}{False discovery rate cutoff (alpha)}
}
\value{
List of \code{RDAVIDWebService()} report objects
}
\description{
Wrapper function that performs gene set enrichment analysis (GSEA) with
RDAVIDWebService, using simplified input options.
}
\author{
Michael Steinbaugh
}
