% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStringdb.R
\name{runStringdb}
\alias{runStringdb}
\title{STRINGdb enrichment analysis}
\usage{
runStringdb(gene, background, species, cache = TRUE, csv = FALSE,
  png = TRUE, pdf = FALSE, version = "10")
}
\arguments{
\item{gene}{Gene identifier vector}

\item{background}{Background identifier vector}

\item{species}{Species (e.g. "Homo sapiens")}

\item{cache}{Cache files to disk (\code{TRUE/FALSE})}

\item{csv}{Export to CSV (\code{TRUE/FALSE})}

\item{png}{Export network as PNG file (\code{TRUE/FALSE})}

\item{pdf}{Export network as high resolution PDF file (\code{TRUE/FALSE})}

\item{version}{STRING version to use}
}
\value{
\code{STRINGdb} output list
}
\description{
Wrapper function that performs gene set enrichment analysis (GSEA) with
STRINGdb, using simplified input options.
}
\author{
Michael Steinbaugh
}
