% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_exec.R
\name{sam_exec}
\alias{sam_exec}
\title{SAM execution}
\usage{
sam_exec()
}
\value{
Opens SAM in a Shiny instance.
}
\description{
Run SAM (Significance Analysis of Microarrays) using Shiny.
}
\examples{
sam_exec()
}

