% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sampleData-methods.R,
%   R/sampleNames-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleData,SummarizedExperiment-method}
\alias{sampleData,SingleCellExperiment-method}
\alias{sampleData<-,SummarizedExperiment,DataFrame-method}
\alias{sampleData<-,SingleCellExperiment,DataFrame-method}
\alias{sampleNames,SummarizedExperiment-method}
\title{Sample data}
\usage{
sampleData(object, ...)

sampleData(object, ...) <- value

\S4method{sampleData}{SummarizedExperiment}(object, clean = FALSE)

\S4method{sampleData}{SingleCellExperiment}(object,
  blacklist = c("^G2M.Score$", "^Phase$", "^S.Score$", "^ident$",
  "^old.ident$", "^orig.ident$", "^res[.0-9]+$"))

\S4method{sampleData}{SummarizedExperiment,DataFrame}(object) <- value

\S4method{sampleData}{SingleCellExperiment,DataFrame}(object) <- value

\S4method{sampleNames}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object class that inherits \code{SummarizedExperiment}.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}

\item{clean}{\code{logical(1)}.
Only return \code{factor} columns. Useful when working with objects that contain
metrics in \code{colData}.}

\item{blacklist}{\code{character}.
Column names that should never be treated as sample-level metadata.
Applicable only to \code{SingleCellExperiment} objects.}
}
\value{
\code{DataFrame}.
}
\description{
Metadata that describes the samples.
}
\details{
This is a complement to the standard \code{SummarizedExperiment::colData}
function, but improves support for accessing sample metadata for datasets
where multiple items in the columns map to a single sample (e.g. cells for a
single-cell RNA-seq experiment).
}
\examples{
data(rse, sce)

## SummarizedExperiment ====
x <- rse
sampleData(x)

## Assignment support
sampleData(x)[["batch"]] <- 1L
## `batch` column should be now defined.
sampleData(x)

## SingleCellExperiment ====
x <- sce
sampleData(x)

## Assignment support.
sampleData(x)[["batch"]] <- 1L
## `batch` column should be now defined.
sampleData(x)
}
