% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAML.R
\docType{methods}
\name{sampleYAML}
\alias{sampleYAML}
\alias{sampleYAML,list,character-method}
\title{Sample YAML Metadata Utilities}
\usage{
sampleYAML(yaml, keys, ...)

\S4method{sampleYAML}{list,character}(yaml, keys)
}
\arguments{
\item{yaml}{Project summary YAML list.}

\item{keys}{Nested operator keys, supplied as a character vector.}
}
\value{
\link{tibble}.
}
\description{
Sample YAML Metadata Utilities
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\examples{
yaml <- readYAML("http://basejump.seq.cloud/project-summary.yaml")
sampleYAML(yaml, "metadata")
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAMLMetrics}}
}
\concept{bcbio Utilities}
\keyword{internal}
