% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAMLMetrics.R
\docType{methods}
\name{sampleYAMLMetrics}
\alias{sampleYAMLMetrics}
\alias{sampleYAMLMetrics,list-method}
\title{Sample Metrics from YAML File}
\usage{
sampleYAMLMetrics(yaml, ...)

\S4method{sampleYAMLMetrics}{list}(yaml, characterCols)
}
\arguments{
\item{yaml}{YAML \link{list}.}

\item{characterCols}{Columns to set as character.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metrics from YAML File
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAML}}
}
\concept{bcbio Utilities}
\keyword{internal}
