% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeAnnotable.R
\name{sanitizeAnnotable}
\alias{sanitizeAnnotable}
\title{Sanitize Ensembl Annotations}
\usage{
sanitizeAnnotable(object)
}
\arguments{
\item{object}{\link{data.frame} containing Ensembl gene annotations returned
from the \code{\link[=annotable]{annotable()}} function.}
}
\value{
Annotable without nested list items (e.g. Entrez IDs).
}
\description{
Sanitize Ensembl Annotations
}
\examples{
human <- annotable("Homo sapiens")
sanitizeAnnotable(human) \%>\% glimpse()
}
