% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeAnnotable.R
\name{sanitizeGeneAnnotations}
\alias{sanitizeGeneAnnotations}
\title{Sanitize Ensembl Gene Annotations}
\usage{
sanitizeGeneAnnotations(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Data frame without nested list columns.
}
\description{
Drops any nested list columns (e.g. \code{entrez}) that prevents an annotable
data frame from being written to disk as a CSV.
}
\examples{
human <- annotable("Homo sapiens")
sanitizeAnnotable(human) \%>\% glimpse()
}
\seealso{
Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{fixNA}}, \code{\link{removeNA}},
  \code{\link{sanitizeColData}}
}
