% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sanitizeNA-methods.R
\docType{methods}
\name{sanitizeNA}
\alias{sanitizeNA}
\alias{sanitizeNA,atomic-method}
\alias{sanitizeNA,character-method}
\alias{sanitizeNA,data.frame-method}
\alias{sanitizeNA,DataFrame-method}
\alias{sanitizeNA,tbl_df-method}
\title{Sanitize \code{NA} Values}
\usage{
sanitizeNA(object, ...)

\S4method{sanitizeNA}{atomic}(object)

\S4method{sanitizeNA}{character}(object)

\S4method{sanitizeNA}{data.frame}(object)

\S4method{sanitizeNA}{DataFrame}(object)

\S4method{sanitizeNA}{tbl_df}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Object containing proper \code{NA} values.
}
\description{
Standardize empty strings (\code{""}), character NAs (\code{"NA"}), and \code{NULL} values
inside a character vector to \code{NA_character_}. Other \code{atomic} data types are
returned unmodified.
}
\examples{
# character ====
sanitizeNA(c(1L, "x", "", "NA", "NULL"))

# DataFrame ====
object <- DataFrame(
    a = c("foo", ""),
    b = c(NA, "bar"),
    row.names = c("c", "d")
)
sanitizeNA(object)
}
\seealso{
Other Sanitization Functions: \code{\link{camel}},
  \code{\link{collapseToString}}, \code{\link{removeNA}},
  \code{\link{sanitizeRowData}},
  \code{\link{sanitizeSampleData}}
}
\author{
Michael Steinbaugh
}
\concept{Sanitization Functions}
