% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeNames.R
\name{sanitizeNames}
\alias{sanitizeNames}
\title{Sanitize names}
\usage{
sanitizeNames(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Sanitized data frame
}
\description{
This function prevents illegal characters from being set in variable names
and column or row names. Additionally, it will prepend an \code{X} character
if necessary
}
\author{
Michael Steinbaugh
}
\keyword{dev}
\keyword{text}
