% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSampleData.R
\name{sanitizeSampleData}
\alias{sanitizeSampleData}
\title{Sanitize sample data}
\usage{
sanitizeSampleData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DataFrame}.
Sanitized data frame containing only non-blacklisted columns and all
\code{character} columns coerced to \code{factor} (i.e. \code{stringsAsFactors}).
}
\description{
Sanitize sample data
}
\section{Blacklist}{


Here's the current column blacklist:
\itemize{
\item interestingGroups.
\item sampleID.
}
}

\examples{
data(rse)
from <- sampleData(rse)
print(from)
to <- sanitizeSampleData(from)
all(vapply(to, is.factor, logical(1L)))
print(to)
}
