% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeString.R
\name{sanitizeString}
\alias{sanitizeString}
\title{Sanitize a string}
\usage{
sanitizeString(string)
}
\arguments{
\item{string}{String}
}
\value{
Sanitized string
}
\description{
Sanitize a string
}
\author{
Michael Steinbaugh
}
\keyword{internal}
