% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/saveDataRaw-methods.R
\docType{methods}
\name{saveDataRaw}
\alias{saveDataRaw}
\alias{saveDataRaw,ANY-method}
\title{Save Raw Data}
\usage{
saveDataRaw(...)

\S4method{saveDataRaw}{ANY}(...)
}
\arguments{
\item{...}{Objects as dots.}
}
\value{
No value.
}
\description{
Wrapper for \code{\link[=saveData]{saveData()}} that enables quick saving of objects to the
\code{data-raw/} directory.
}
\seealso{
\code{\link[devtools:use_data_raw]{devtools::use_data_raw()}}.

Other Write Utilities: \code{\link{saveData}},
  \code{\link{writeCounts}}
}
