% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFunctionExamples.R
\name{saveFunctionExamples}
\alias{saveFunctionExamples}
\title{Save Functions Examples}
\usage{
saveFunctionExamples(f, package, dir = ".")
}
\arguments{
\item{f}{\code{character} or \code{NULL}. Function(s) from which to parse and save the
working examples. If \code{NULL}, all functions will be saved.}

\item{package}{\code{string}. Package name.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}
}
\value{
Invisible \code{character}. File path(s).
}
\description{
Parse the documentation for a function and save the working examples to an R
script. Note that the \code{f} argument is parameterized and can handle multiple
requests in a single call.
}
\examples{
saveFunctionExamples(
    fun = c("do.call", "droplevels"),
    package = "base",
    dir = "XXX"
)

# Clean up
unlink("XXX", recursive = TRUE)
}
\seealso{
Other Developer Functions: \code{\link{calls}},
  \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
