% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{saveRdExamples}
\alias{saveRdExamples}
\title{Save R documentation examples}
\usage{
saveRdExamples(Rd = NULL, package, dir = ".")
}
\arguments{
\item{Rd}{\code{character} or \code{NULL}.
R documentation name(s) from which to parse and save the working examples.
If \code{NULL}, all documentation files containing examples will be saved.}

\item{package}{\code{character(1)}.
Package name.}

\item{dir}{\code{character(1)}.
Directory path.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Parse the documentation for a function and save the working examples to an R
script. Note that the \code{f} argument is parameterized and can handle multiple
requests in a single call.
}
\examples{
saveRdExamples(
    Rd = c("do.call", "droplevels"),
    package = "base",
    dir = "example"
)

## Clean up.
unlink("example", recursive = TRUE)
}
