% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{Example Single-Cell RNA-Seq Data Set}
\format{An object of class \code{SingleCellExperiment} with 50 rows and 100 columns.}
\usage{
sce
}
\description{
Example Single-Cell RNA-Seq Data Set
}
\examples{
data(sce)
sce
}
\seealso{
\code{scater::sc_example_counts}.
}
\keyword{datasets}
