% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_small}
\alias{sce_small}
\title{Example Single-Cell RNA-Seq Data Set}
\format{An object of class \code{SingleCellExperiment} with 50 rows and 100 columns.}
\usage{
sce_small
}
\description{
Example Single-Cell RNA-Seq Data Set
}
\examples{
data(sce_small)
sce_small
}
\seealso{
\code{scater::sc_example_counts}.
}
\keyword{datasets}
