% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSamples-methods.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,SummarizedExperiment-method}
\alias{selectSamples,SingleCellExperiment-method}
\title{Select samples}
\usage{
\S4method{selectSamples}{SummarizedExperiment}(object, ...)

\S4method{selectSamples}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Selection arguments that map to the column names of
\code{sampleData()}. \code{atomic} values are supported. Avoid using \code{logical} or
\code{numeric} indices (e.g. \code{which()} calls) here, for improved code
readability.}
}
\value{
Modified object.
Subset (extraction) containing the desired samples.
}
\description{
Utility function that enables quick an easy sample selection in Bioconductor
contains that don't correspond to samples in the columns
(e.g. \code{SingleCellExperiment}).
}
\details{
Internally, pattern matching against sample and file names is applied using
logical grep matching.

Bracket based subsetting with \code{[} also works on \code{SingleCellExperiment}
objects but it's not intuitive. In this case, provide cellular barcode
identifiers for columns and gene identifiers for rows.
}
\examples{
data(rse, sce)

## SummarizedExperiment ====
object <- rse
sample <- sampleNames(object) \%>\% head(1L)
print(sample)
subset <- selectSamples(object, sampleName = sample)
print(subset)

## SingleCellExperiment ====
object <- sce
sample <- sampleNames(object) \%>\% head(1L)
print(sample)
subset <- selectSamples(object, sampleName = sample)
print(subset)
}
\seealso{
\itemize{
\item \code{sampleData()}.
\item \code{S4Vectors::split()}.
}
}
