% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sem-methods.R
\name{sem}
\alias{sem}
\alias{sem,AsIs-method}
\title{Return the first and last parts of an object}
\usage{
sem(x, ...)

\S4method{sem}{AsIs}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Applies to both rows and columns, enabling quick inspection during
interactive use.
}
\details{
Inspired by the \code{\link[base:print]{print()}} method for \code{DataFrame} class
objects.
}
\note{
\code{numeric} method is defined in AcidBase.

Updated 2020-12-04.
}
\examples{
## AsIs ====
## Useful for dplyr mutate calls.
x <- seq(from = 1L, to = 10L, by = 1L)
x <- I(x)
class(x)
x <- sem(x)
x
}
\seealso{
\itemize{
\item \code{utils::head()}, \code{utils::tail()}.
\item \code{cat()}.
}
}
