% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separator.R
\name{separator}
\alias{separator}
\title{Separator bar}
\usage{
separator(sep = c("─", "=", "-", "+"),
  times = min(c(getOption("width", 72L), 72L)))
}
\arguments{
\item{sep}{\code{character(1)}.
Separator character.}

\item{times}{\code{integer(1)}.
Number of times to repeat.}
}
\value{
\code{character(1)}.
}
\description{
Maximum of 72 characters wide.
}
\note{
Bioconductor HTML vignettes don't render correctly when printing > 76
characters, even though the default width is set at 80.
}
\examples{
cat(separator(sep = "=", times = 10L))
}
