% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{separatorBar}
\alias{separatorBar}
\title{Separator Bar}
\format{An object of class \code{character} of length 1.}
\usage{
separatorBar
}
\description{
Maximum of 72 characters wide.
}
\note{
Bioconductor HTML vignettes don't render correctly when printing > 76
characters, even though the default width is set at 80.
}
\examples{
cat(separatorBar)
}
\keyword{internal}
