% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNames}
\alias{setNames}
\alias{setNamesCamel}
\alias{set_names_camel}
\alias{setNamesDot}
\alias{set_names_dot}
\alias{setNamesSnake}
\alias{set_names_snake}
\alias{sanitizeNames}
\alias{sanitiseNames}
\alias{sanitize_names}
\alias{sanitise_names}
\alias{setColnames}
\alias{setRownames}
\title{Naming utility functions.}
\usage{
setNamesCamel(data)

setNamesDot(data)

setNamesSnake(data)

sanitizeNames(data)
}
\arguments{
\item{data}{Data type that supports name assignments.}
}
\value{
Unmodified data with reformatted names.
}
\description{
Naming utility functions.
}
\examples{
setNamesCamel(head(starwars))
setNamesDot(head(starwars))
setNamesSnake(head(starwars))
sanitizeNames(head(starwars))
}
