% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamesDot.R
\name{setNamesDot}
\alias{setNamesDot}
\title{Set names as dot notation}
\usage{
setNamesDot(data)
}
\arguments{
\item{data}{Data type that supports name assignments}
}
\value{
data Unmodified data with reformatted dot notation names
}
\description{
Set names as dot notation
}
\author{
Michael Steinbaugh
}
