% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamesSnake.R
\name{setNamesSnake}
\alias{setNamesSnake}
\title{Set names of tibble as snake_case}
\usage{
setNamesSnake(tibble)
}
\arguments{
\item{tibble}{}
}
\value{
tibble
}
\description{
Set names of tibble as snake_case
}

