% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRownames.R
\name{setRownames}
\alias{setRownames}
\title{Set row names}
\usage{
setRownames(df, column)
}
\arguments{
\item{df}{Data frame}

\item{column}{Column to use for \code{rownames}}
}
\description{
Wrapper function for \code{rownames} that works in a chain operation
}
\examples{
\dontrun{
setRownames(df, "description")
}
}
\author{
Michael Steinbaugh
}
