% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_pkg.R
\name{setup_pkg}
\alias{setup_pkg}
\title{Setup R packages}
\usage{
setup_pkg(pkg, source = "cran")
}
\arguments{
\item{pkg}{Character vector of package names}

\item{source}{Where to download packages (CRAN, Bioconductor, GitHub)}
}
\value{
Installs packages if necessary, then loads into current workspace.
}
\description{
This function allows for dynamic install and loading of packages from CRAN,
Bioconductor, and GitHub.
}
\examples{
setup_pkg(c("plyr", "readr"), source = "cran")
}

