% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showHeader.R
\name{showHeader}
\alias{showHeader}
\title{Header for object show method}
\usage{
showHeader(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output, via \code{\link[base:cat]{cat()}}.
}
\description{
Header for object show method
}
\note{
Updated 2019-08-11.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment
showHeader(rse)
}
\seealso{
\code{\link[=show]{show()}}.
}
