% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSlotInfo.R
\name{showSlotInfo}
\alias{showSlotInfo}
\title{Show slot information}
\usage{
showSlotInfo(list)
}
\arguments{
\item{list}{\code{list}.
Named list containing slot metadata.}
}
\value{
Invisible \code{NULL}.
}
\description{
Show slot information
}
\examples{
showSlotInfo(list(
    dir = "~",
    skip = NULL,
    filtered = TRUE
))
}
