% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_cell_counts}
\alias{single_cell_counts}
\alias{singleCellCounts}
\title{Example Single-Cell RNA-Seq Counts}
\format{An object of class \code{dgCMatrix} with 1000 rows and 40 columns.}
\usage{
single_cell_counts
}
\description{
Modified version of \code{scater::sc_example_counts}, with dimnames formatted
in camel case for package consistency.
}
\examples{
dim(single_cell_counts)
class(single_cell_counts)
is(single_cell_counts, "sparseMatrix")
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
