% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{singleCellCounts}
\alias{singleCellCounts}
\title{Example Single-Cell RNA-Seq Counts}
\format{An object of class \code{dgCMatrix} with 2000 rows and 40 columns.}
\usage{
singleCellCounts
}
\description{
Modified version of \code{scater::sc_example_counts}, with dimnames formatted
in camel case for package consistency.
}
\examples{
summary(singleCellCounts)
}
\keyword{datasets}
