% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFactors-methods.R
\docType{methods}
\name{sizeFactors}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{sizeFactors,SummarizedExperiment-method}
\alias{sizeFactors<-,SummarizedExperiment,numeric-method}
\alias{sizeFactors<-,SingleCellExperiment,numeric-method}
\title{Size factors}
\usage{
sizeFactors(object, ...)

sizeFactors(object, ...) <- value

\S4method{sizeFactors}{SummarizedExperiment}(object)

\S4method{sizeFactors}{SummarizedExperiment,numeric}(object) <- value

\S4method{sizeFactors}{SingleCellExperiment,numeric}(object, type = NULL,
  ...) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{type}{A string specifying the \emph{type} of size factor to get or set.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
Names correspond to object column names.
}
\description{
Size factors
}
\note{
Updated 2019-08-06.
}
\examples{
data(
    RangedSummarizedExperiemnt,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
object <- estimateSizeFactors(object)
head(sizeFactors(object))
mean(sizeFactors(object))

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- estimateSizeFactors(object)
head(sizeFactors(object))
mean(sizeFactors(object))
}
\seealso{
\itemize{
\item \code{DESeq2::sizeFactors()}.
\item \code{DESeq2::estimateSizeFactors()}.
\item \code{DESeq2::estimateSizeFactorsForMatrix()}.
\item \code{SingleCellExperiment::sizeFactors()}.
}
}
