% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/collapse.R, R/deprecated.R,
%   R/load.R, R/math.R, R/names.R, R/project.R, R/read.R, R/reexports.R,
%   R/rmarkdown.R, R/save.R, R/utils.R
\name{snake_aliases}
\alias{snake_aliases}
\alias{to_string_unique}
\alias{to_string_sort_unique}
\alias{summarize_rows}
\alias{create_new_project}
\alias{dot_notation}
\alias{load_data_raw}
\alias{to_string_summarize}
\alias{load_data}
\alias{load_remote}
\alias{log_ratio_to_fold_change}
\alias{sanitize_names}
\alias{title_case}
\alias{first_case}
\alias{create_project_dirs}
\alias{detect_hpc}
\alias{package_project}
\alias{render_project}
\alias{read_data_raw}
\alias{read_yaml}
\alias{tidy_select}
\alias{md_list}
\alias{save_data}
\alias{save_data_raw}
\alias{assign_as_new_env}
\alias{clear_warnings}
\alias{fix_na}
\alias{grep_string}
\alias{remove_na}
\alias{sort_unique}
\alias{update_packages}
\title{snake_case aliases}
\usage{
dot_notation(...)
}
\description{
Function name variants formatted in \code{snake_case}.
}
\keyword{internal}
