% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakeCase-methods.R
\docType{methods}
\name{snakeCase}
\alias{snakeCase}
\alias{snakeCase,atomic-method}
\alias{snakeCase,factor-method}
\alias{snakeCase,list-method}
\alias{snakeCase,matrix-method}
\alias{snakeCase,data.frame-method}
\alias{snakeCase,Vector-method}
\alias{snakeCase,DataTable-method}
\alias{snakeCase,Ranges-method}
\alias{snakeCase,Matrix-method}
\alias{snakeCase,SummarizedExperiment-method}
\title{Snake case}
\usage{
\S4method{snakeCase}{atomic}(object, names = TRUE)

\S4method{snakeCase}{factor}(object, names = TRUE)

\S4method{snakeCase}{list}(object, names = TRUE)

\S4method{snakeCase}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snakeCase}{data.frame}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{snakeCase}{Vector}(object, names = TRUE, mcols = TRUE,
  metadata = TRUE)

\S4method{snakeCase}{DataTable}(object, rownames = FALSE,
  colnames = TRUE, mcols = TRUE, metadata = TRUE)

\S4method{snakeCase}{Ranges}(object, names = FALSE, mcols = TRUE,
  metadata = TRUE)

\S4method{snakeCase}{Matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snakeCase}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE, assayNames = TRUE, rowData = TRUE,
  colData = TRUE, metadata = TRUE)
}
\arguments{
\item{object}{Object.}

\item{names}{\code{logical(1)}.
Sanitize names.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
}
\description{
Format character strings to use snake-style formatting, where word boundaries
are defined by underscores (e.g. \code{this_is_snake_case}).
}
\note{
Updated 2019-09-25.
}
\examples{
data(syntactic, package = "acidtest")
lapply(syntactic, snakeCase)
}
