% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortUnique.R
\name{sortUnique}
\alias{sortUnique}
\title{Sort unique.}
\usage{
sortUnique(vector)
}
\arguments{
\item{vector}{\code{vector} with duplicates, \code{NA} values.}
}
\value{
vector
}
\description{
Sort unique.
}

