% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sparse}
\alias{sparse}
\title{Example Sparse Matrix}
\format{An object of class \code{dgCMatrix} with 4 rows and 4 columns.}
\usage{
sparse
}
\description{
Example Sparse Matrix
}
\examples{
data(sparse)
sparse
}
\keyword{datasets}
