% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByLevel-methods.R
\name{splitByLevel}
\alias{splitByLevel}
\alias{splitByLevel,data.frame-method}
\alias{splitByLevel,DataFrame-method}
\title{Split by factor level}
\usage{
splitByLevel(x, f, ref = FALSE, ...)

\S4method{splitByLevel}{data.frame}(x, f, ref = FALSE)

\S4method{splitByLevel}{DataFrame}(x, f, ref = FALSE)
}
\arguments{
\item{x}{Object.}

\item{f}{\code{character(1)}.
Factor column name.}

\item{ref}{\code{logical(1)}.
Include the reference factor level (i.e. first defined) in each split.
Useful for pairwise contrasts.}

\item{...}{Additional arguments.}
}
\value{
\code{list}
List of data frames split by factors.
}
\description{
Split by factor level
}
\note{
Updated 2020-09-14.
}
\examples{
df <- DataFrame(
    compound = relevel(
        factor(
            c(
                rep(x = "dmso", each = 3L),
                rep(x = c("ts_0001", "ts_0002"), each = 9L)
            )
        ),
        ref = "dmso"
    ),
    concentration = factor(
        c(
            rep(x = 0, each = 3L),
            rep(x = c(0.1, 1, 10), each = 3L, times = 2L)
        )
    ),
    replicate = factor(
        rep(
            seq(from = 1L, to = 3L),
            times = 7L
        )
    )
)
lapply(df, levels)

splitByLevel(df, f = "compound", ref = FALSE)
splitByLevel(df, f = "compound", ref = TRUE)

splitByLevel(df, f = "concentration", ref = FALSE)
splitByLevel(df, f = "concentration", ref = TRUE)
}
\seealso{
\itemize{
\item \code{split()}
\item \code{S4Vectors::split()}
\item \code{dplyr::group_split()}
}
}
