% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{starwars}
\alias{starwars}
\title{Star Wars dataset.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 87 rows and 13 columns.}
\usage{
starwars
}
\description{
Star Wars dataset.
}
\keyword{internal}
