% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse.R
\name{str_collapse}
\alias{str_collapse}
\title{Stringr collapse call for dplyr summarize.}
\usage{
str_collapse(x)
}
\arguments{
\item{x}{Input.}
}
\value{
Output.
}
\description{
Stringr collapse call for dplyr summarize.
}

