% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/stripTranscriptVersions-methods.R
\docType{methods}
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions,character-method}
\alias{stripTranscriptVersions,matrix-method}
\alias{stripTranscriptVersions,sparseMatrix-method}
\alias{stripTranscriptVersions,SummarizedExperiment-method}
\title{Strip transcript versions}
\usage{
\S4method{stripTranscriptVersions}{character}(object)

\S4method{stripTranscriptVersions}{matrix}(object)

\S4method{stripTranscriptVersions}{sparseMatrix}(object)

\S4method{stripTranscriptVersions}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
Contains transcript identifiers without version numbers.
}
\description{
Strip transcript versions
}
\note{
This method is strict, and will only strip Ensembl transcript IDs
beginning with "ENS".
}
\examples{
## Ensembl (modify; contains versions)
stripTranscriptVersions(c(
    "ENSMUST00000000001.1",
    "ENSMUST00000000001-1",
    "ENSMUST00000000001_1"
))

## WormBase (keep; doesn't contain versions)
stripTranscriptVersions("cTel79B.1")
}
