% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,Gene2Symbol-method}
\title{Object Summaries}
\usage{
\S4method{summary}{Gene2Symbol}(object)
}
\arguments{
\item{object}{an object for which a summary is desired.}
}
\value{
Invisible \code{NULL}.
}
\description{
\code{summary} is a generic function used to produce result summaries
  of the results of various model fitting functions.  The function
  invokes particular \code{\link{methods}} which depend on the
  \code{\link{class}} of the first argument.
}
\details{
For \code{\link{factor}}s, the frequency of the first \code{maxsum - 1}
    most frequent levels is shown, and the less frequent levels are
    summarized in \code{"(Others)"} (resulting in at most \code{maxsum}
    frequencies).

    The functions \code{summary.lm} and \code{summary.glm} are examples
    of particular methods which summarize the results produced by
    \code{\link{lm}} and \code{\link{glm}}.
}
\examples{
data(rse, tx_se)

## Gene2Symbol ====
x <- Gene2Symbol(rse)
summary(x)

## Tx2Gene ====
x <- Tx2Gene(tx_se)
summary(x)
}
\references{
Chambers, J. M. and Hastie, T. J. (1992)
  \emph{Statistical Models in S}.
  Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{anova}},
  \code{\link{summary.glm}},
  \code{\link{summary.lm}}.
}
