% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synonyms}
\alias{synonyms}
\title{Gene Name Synonyms}
\format{An object of class \code{list} of length 4.}
\usage{
synonyms
}
\description{
A list of tibbles per organism containing gene name (a.k.a. symbol) synonyms.
}
\examples{
names(synonyms)
glimpse(synonyms[["homoSapiens"]])
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
