% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/syntactic-methods.R
\name{camelCase}
\alias{camelCase}
\alias{dottedCase}
\alias{snakeCase}
\alias{upperCamelCase}
\alias{syntactic}
\alias{camelCase,atomic-method}
\alias{camelCase,factor-method}
\alias{camelCase,list-method}
\alias{camelCase,matrix-method}
\alias{camelCase,data.frame-method}
\alias{camelCase,Vector-method}
\alias{camelCase,DataFrame-method}
\alias{camelCase,Ranges-method}
\alias{camelCase,Matrix-method}
\alias{camelCase,SummarizedExperiment-method}
\alias{dottedCase,atomic-method}
\alias{dottedCase,factor-method}
\alias{dottedCase,list-method}
\alias{dottedCase,matrix-method}
\alias{dottedCase,data.frame-method}
\alias{dottedCase,Vector-method}
\alias{dottedCase,DataFrame-method}
\alias{dottedCase,Ranges-method}
\alias{dottedCase,Matrix-method}
\alias{dottedCase,SummarizedExperiment-method}
\alias{snakeCase,atomic-method}
\alias{snakeCase,factor-method}
\alias{snakeCase,list-method}
\alias{snakeCase,matrix-method}
\alias{snakeCase,data.frame-method}
\alias{snakeCase,Vector-method}
\alias{snakeCase,DataFrame-method}
\alias{snakeCase,Ranges-method}
\alias{snakeCase,Matrix-method}
\alias{snakeCase,SummarizedExperiment-method}
\alias{upperCamelCase,atomic-method}
\alias{upperCamelCase,factor-method}
\alias{upperCamelCase,list-method}
\alias{upperCamelCase,matrix-method}
\alias{upperCamelCase,data.frame-method}
\alias{upperCamelCase,Vector-method}
\alias{upperCamelCase,DataFrame-method}
\alias{upperCamelCase,Ranges-method}
\alias{upperCamelCase,Matrix-method}
\alias{upperCamelCase,SummarizedExperiment-method}
\title{Syntactic naming functions}
\usage{
camelCase(object, ...)

dottedCase(object, ...)

snakeCase(object, ...)

upperCamelCase(object, ...)

\S4method{camelCase}{atomic}(object, names = TRUE, strict = TRUE)

\S4method{camelCase}{factor}(object, names = TRUE, strict = TRUE)

\S4method{camelCase}{list}(object, names = TRUE, strict = TRUE)

\S4method{camelCase}{matrix}(object, rownames = FALSE, colnames = TRUE, strict = TRUE)

\S4method{camelCase}{data.frame}(object, rownames = FALSE, colnames = TRUE, strict = TRUE)

\S4method{camelCase}{Vector}(object, names = TRUE, mcols = TRUE, metadata = TRUE, strict = TRUE)

\S4method{camelCase}{DataFrame}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  mcols = TRUE,
  metadata = TRUE,
  strict = TRUE
)

\S4method{camelCase}{Ranges}(object, names = FALSE, mcols = TRUE, metadata = TRUE, strict = TRUE)

\S4method{camelCase}{Matrix}(object, rownames = FALSE, colnames = TRUE, strict = TRUE)

\S4method{camelCase}{SummarizedExperiment}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  assayNames = TRUE,
  rowData = TRUE,
  colData = TRUE,
  metadata = TRUE,
  strict = TRUE
)

\S4method{dottedCase}{atomic}(object, names = TRUE)

\S4method{dottedCase}{factor}(object, names = TRUE)

\S4method{dottedCase}{list}(object, names = TRUE)

\S4method{dottedCase}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{dottedCase}{data.frame}(object, rownames = FALSE, colnames = TRUE)

\S4method{dottedCase}{Vector}(object, names = TRUE, mcols = TRUE, metadata = TRUE)

\S4method{dottedCase}{DataFrame}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  mcols = TRUE,
  metadata = TRUE
)

\S4method{dottedCase}{Ranges}(object, names = FALSE, mcols = TRUE, metadata = TRUE)

\S4method{dottedCase}{Matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{dottedCase}{SummarizedExperiment}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  assayNames = TRUE,
  rowData = TRUE,
  colData = TRUE,
  metadata = TRUE
)

\S4method{snakeCase}{atomic}(object, names = TRUE)

\S4method{snakeCase}{factor}(object, names = TRUE)

\S4method{snakeCase}{list}(object, names = TRUE)

\S4method{snakeCase}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snakeCase}{data.frame}(object, rownames = FALSE, colnames = TRUE)

\S4method{snakeCase}{Vector}(object, names = TRUE, mcols = TRUE, metadata = TRUE)

\S4method{snakeCase}{DataFrame}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  mcols = TRUE,
  metadata = TRUE
)

\S4method{snakeCase}{Ranges}(object, names = FALSE, mcols = TRUE, metadata = TRUE)

\S4method{snakeCase}{Matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snakeCase}{SummarizedExperiment}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  assayNames = TRUE,
  rowData = TRUE,
  colData = TRUE,
  metadata = TRUE
)

\S4method{upperCamelCase}{atomic}(object, names = TRUE, strict = TRUE)

\S4method{upperCamelCase}{factor}(object, names = TRUE, strict = TRUE)

\S4method{upperCamelCase}{list}(object, names = TRUE, strict = TRUE)

\S4method{upperCamelCase}{matrix}(object, rownames = FALSE, colnames = TRUE, strict = TRUE)

\S4method{upperCamelCase}{data.frame}(object, rownames = FALSE, colnames = TRUE, strict = TRUE)

\S4method{upperCamelCase}{Vector}(
  object,
  names = TRUE,
  mcols = TRUE,
  metadata = TRUE,
  strict = TRUE
)

\S4method{upperCamelCase}{DataFrame}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  mcols = TRUE,
  metadata = TRUE,
  strict = TRUE
)

\S4method{upperCamelCase}{Ranges}(
  object,
  names = FALSE,
  mcols = TRUE,
  metadata = TRUE,
  strict = TRUE
)

\S4method{upperCamelCase}{Matrix}(object, rownames = FALSE, colnames = TRUE, strict = TRUE)

\S4method{upperCamelCase}{SummarizedExperiment}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  assayNames = TRUE,
  rowData = TRUE,
  colData = TRUE,
  metadata = TRUE,
  strict = TRUE
)
}
\arguments{
\item{object}{Object.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{mcols}{\code{logical(1)}.
Sanitize names of metadata columns (i.e. \code{DataFrame}).}

\item{metadata}{\code{logical(1)}.
Sanitize metadata names.}

\item{assayNames}{\code{logical(1)}.
Sanitize assay names.}

\item{rowData}{\code{logical(1)}.
Sanitize the row data names.}

\item{colData}{\code{logical(1)}.
Sanitize column names of column data.}

\item{...}{Additional arguments.}
}
\description{
Syntactic naming functions
}
\note{
Updated 2021-02-03.
}
\examples{
data(syntactic, package = "AcidTest")
lapply(syntactic, camelCase)
}
