% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump-package.R
\docType{data}
\name{testDataURL}
\alias{testDataURL}
\title{Test Data URL}
\format{An object of class \code{character} of length 1.}
\usage{
testDataURL
}
\description{
Test Data URL
}
\keyword{internal}
