% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\alias{theme_paperwhite}
\alias{theme_midnight}
\title{Complete Themes}
\usage{
theme_paperwhite(base_size = 14L, base_family = "", face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("right", "bottom",
  "none"), grid = FALSE)

theme_midnight(base_size = 14L, base_family = "", face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("right", "bottom",
  "none"), grid = FALSE)
}
\arguments{
\item{base_size}{\code{scalar numeric}. Base font size.}

\item{base_family}{\code{string}. Base font family.}

\item{face}{\code{string}. Font face ("\code{bold}", "\code{plain}").}

\item{aspect_ratio}{\code{scalar numeric}. Aspect ratio, specifying the plot
proportions. Use \code{1} for a perfectly square plot (including the axis
labels).}

\item{legend_position}{\code{string}. Legend key position. We're being a little
more restrictive here, only allowing "\code{bottom}", "\code{right}", or "\code{none}".
Including the legend at the top or the left side of the plot rarely makes
sense and is discouraged.}

\item{grid}{\code{boolean}. Label the major panel grids. If \code{TRUE}, the axis lines
will also be removed.}
}
\value{
\code{theme}.
}
\description{
\href{http://ggplot2.tidyverse.org}{ggplot2} themes.
}
\section{Paperwhite}{

High contrast black and white theme optimized for print. Recommended for
scientific manuscripts and website tutorials.
}

\section{Midnight}{

Blackout theme that sets the plot background as black, with white text.
Inspired by \code{Seurat::DarkTheme()}, with some color modifications. Useful
for visualizing many points with a high dynamic color range, such as t-SNE
expression plots.
}

\examples{
library(ggplot2)
p <- ggplot(
    data = mpg,
    mapping = aes(
        x = manufacturer,
        y = displ,
        colour = manufacturer
    )
) +
    geom_point()
p + theme_paperwhite(aspect_ratio = 1L, legend_position = "none")
p + theme_midnight(aspect_ratio = 1L, legend_position = "none")
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.

Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
