% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R, R/reexports.R, R/tidy.R
\docType{methods}
\name{content,response-method}
\alias{content,response-method}
\alias{expand,data.frame-method}
\alias{first,data.frame-method}
\alias{intersect,data.frame,data.frame-method}
\alias{rename,data.frame-method}
\alias{select,data.frame-method}
\alias{setdiff,data.frame,data.frame-method}
\alias{setequal,data.frame,data.frame-method}
\alias{union,data.frame,data.frame-method}
\alias{tidy_select}
\alias{tidy}
\alias{tidyverse}
\title{\href{http://tidyverse.org/}{tidyverse} generics}
\usage{
\S4method{content}{response}(object)

\S4method{expand}{data.frame}(x, ...)

\S4method{first}{data.frame}(x, ...)

\S4method{intersect}{data.frame,data.frame}(x, y, ...)

\S4method{rename}{data.frame}(x, ...)

\S4method{setdiff}{data.frame,data.frame}(x, y, ...)

\S4method{setequal}{data.frame,data.frame}(x, y)

\S4method{union}{data.frame,data.frame}(x, y, ...)
}
\description{
Avoid NAMESPACE collisions with \href{https://www.bioconductor.org/}{Bioconductor}
generics by using \code{\link[=setMethod]{setMethod()}} to relevant \code{signature}.
}
\seealso{
\itemize{
\item \href{https://www.bioconductor.org/}{Bioconductor}:
\link{Biobase}, \link{BiocGenerics}, S4Vectors.
\item \href{http://tidyverse.org/}{tidyverse}:
\link{dplyr}, \link{httr}, \link{tidyr}.
}
}
