% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{content}
\alias{content}
\alias{intersect}
\alias{setdiff}
\alias{union}
\alias{tidy_select}
\alias{expand}
\alias{first}
\alias{rename}
\title{S4 Generics}
\usage{
content(object)

intersect(x, y, ...)

setdiff(x, y, ...)

union(x, y, ...)

expand(x, ...)

first(x, ...)

rename(x, ...)
}
\arguments{
\item{object}{Object.}

\item{x}{Primary object.}

\item{y}{Secondary object.}

\item{...}{Additional arguments.}
}
\description{
S4 Generics
}
