% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{tmm}
\alias{tmm}
\title{Trimmed Mean of M-Values}
\usage{
tmm(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
TMM normalization is recommended for RNA-seq data generally when the majority
of genes are not differentially expressed.
}
\note{
Only recommended for gene-level counts.
}
\references{
Robinson and Oshlack (2010).
}
\seealso{
\itemize{
\item \code{edgeR::calcNormFactors}.
\item \code{edgeR::cpm}.
\item bcbioRNASeq.
}
}
