% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString.R
\name{toStringUnique}
\alias{toStringUnique}
\alias{toStringSortUnique}
\alias{toStringSummarize}
\title{Utility functions for toString}
\usage{
toStringUnique(character)

toStringSortUnique(character)

toStringSummarize(data)
}
\arguments{
\item{character}{Character vector}

\item{data}{Data with rows and columns (e.g. data frame, matrix)}
}
\value{
Unique character vector

Sorted unique character vector

Summarized data frame that has been collapsed to a single
  \code{toString()}-formatted row, separated by commas
}
\description{
Utility functions for toString
}
\examples{
toStringUnique(c("milk", "eggs", "eggs", NA))
toStringSortUnique(c("milk", "eggs", "eggs", NA))
toStringSummarize(head(iris))
}
\author{
Michael Steinbaugh
}
