% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStringSortUnique.R
\name{toStringSortUnique}
\alias{toStringSortUnique}
\title{toString call that outputs sorted uniques}
\usage{
toStringSortUnique(vector)
}
\arguments{
\item{vector}{\code{vector}}
}
\value{
Sorted unique \code{string}
}
\description{
toString call that outputs sorted uniques
}
\keyword{general}

