% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString.R
\name{toStringSummarize}
\alias{toStringSummarize}
\title{Summarize columns with \code{\link[=toString]{toString()}}}
\usage{
toStringSummarize(x, sort = TRUE)
}
\arguments{
\item{x}{Data with rows and columns (e.g. data frame, matrix)}

\item{sort}{Sort the collapsed results.}
}
\value{
Summarized data frame collapsed to a single row.
}
\description{
Collapse column data into a single row summary columnwize using
\code{\link[=toString]{toString()}}.
}
\examples{
mtcars \%>\% head \%>\% toStringSummarize
}
