% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalReads.R
\name{totalReads}
\alias{totalReads}
\title{Total reads plot}
\usage{
totalReads(description, totalReads, group)
}
\arguments{
\item{description}{Sample description}

\item{totalReads}{Total number of reads}

\item{group}{Sample grouping, used for coloring}
}
\description{
Optimized for \code{bcbio-rnaseq} summary report.
}
\examples{
\dontrun{
totalReads(summary$description,
           summary$total_reads,
           summary$treatment)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
