% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tpm-methods.R
\docType{methods}
\name{tpm}
\alias{tpm}
\alias{tpm,SummarizedExperiment-method}
\title{Transcripts per Million}
\usage{
tpm(object, ...)

\S4method{tpm}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Transcripts per Million
}
\note{
Both gene- and transcript-level counts are supported, as long as they
were imported using a tximport caller (e.g. salmon, kallisto).

Both gene- and transcript-level counts are supported, as long as they
were imported using a tximport caller (e.g. salmon, kallisto).
}
\examples{
se <- SummarizedExperiment::SummarizedExperiment(
    assays = list(
        tpm = matrix(
            data = seq_len(4L),
            nrow = 2L,
            ncol = 2L,
            byrow = TRUE
        )
    )
)
x <- tpm(se)
class(x)
}
