% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{transmit}
\alias{transmit}
\title{Transmit files from a remote FTP server}
\usage{
transmit(remoteDir, localDir = ".", pattern, rename = NULL, compress = FALSE)
}
\arguments{
\item{remoteDir}{\code{character(1)}.
Remote FTP directory path.}

\item{localDir}{\code{character(1)}.
Directory where to save files locally.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{rename}{\code{character(1)} or \code{NULL}.
Rename the local files (including suffix), if desired.}

\item{compress}{\code{logical(1)}.
Compress the downloaded files.}
}
\value{
Invisible \code{character}.
Local file paths.
}
\description{
Utility function that supports easy file matching and download from a remote
FTP server. Also enables on-the-fly file renaming and compression.
}
\note{
Updated 2020-07-24.
}
\examples{
remoteDir <- "ftp://ftp.ncbi.nlm.nih.gov/genomes/"
if (
    !isTRUE(nzchar(Sys.getenv("CI"))) &&
    goalie::hasInternet(remoteDir)
) {
    readme <- transmit(
        remoteDir = remoteDir,
        pattern = "^README\\\\.txt$",
        rename = "ncbi-readme.txt",
        compress = FALSE
    )
    basename(readme)
    file.exists(readme)
    unlink(readme)
}
}
