% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Tx2Gene-class}
\alias{Tx2Gene-class}
\title{Transcript-to-Gene Identifier Mappings}
\value{
\code{Tx2Gene}. Contains a \code{DataFrame} with \code{transcriptID} and \code{geneID}
columns.
}
\description{
Transcript-to-Gene Identifier Mappings
}
\section{Genome metadata}{



We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:metadata]{S4Vectors::metadata()}}.

}

\examples{
## SummarizedExperiment ====
data(tx_se, package = "basejump.data")
x <- Tx2Gene(tx_se)
print(x)
}
\seealso{
\code{help(topic = "makeTx2Gene", package = "basejump.annotations")}.
}
\keyword{internal}
