% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Tx2Gene-methods.R
\docType{methods}
\name{Tx2Gene}
\alias{Tx2Gene}
\alias{Tx2Gene,data.frame-method}
\alias{Tx2Gene,DataFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,SummarizedExperiment-method}
\title{Transcript-to-Gene Identifier Mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{data.frame}(object)

\S4method{Tx2Gene}{DataFrame}(object)

\S4method{Tx2Gene}{GRanges}(object)

\S4method{Tx2Gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-Gene Identifier Mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.
}
\examples{
## SummarizedExperiment ====
data(tx_se_small)
x <- Tx2Gene(tx_se_small)
print(x)
}
\seealso{
\code{\link[=Tx2Gene]{Tx2Gene()}}, \link{makeTx2Gene}.

Other S4 classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}
}
