% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx2geneFromGFF.R
\name{tx2geneFromGFF}
\alias{tx2geneFromGFF}
\alias{tx2geneFromGTF}
\title{Define Transcript to Gene Mappings from GFF/GTF File}
\usage{
tx2geneFromGFF(file)

tx2geneFromGTF(file)
}
\arguments{
\item{file}{File path.}
}
\value{
\code{data.frame}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
tx2geneFromGFF("http://basejump.seq.cloud/mmusculus.gtf") \%>\% glimpse()
}
\seealso{
Other GFF Functions: \code{\link{gene2symbolFromGFF}},
  \code{\link{readGFF}}, \code{\link{rowRangesFromGFF}}
}
