% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tx_se}
\alias{tx_se}
\title{Example Transcript-Level Data Set}
\format{An object of class \code{SummarizedExperiment} with 6 rows and 4 columns.}
\usage{
tx_se
}
\description{
Example Transcript-Level Data Set
}
\examples{
data(tx_se)
tx_se
}
\keyword{datasets}
