% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tx_se_small}
\alias{tx_se_small}
\title{Example Transcript-Level Data Set}
\format{An object of class \code{SummarizedExperiment} with 6 rows and 4 columns.}
\usage{
tx_se_small
}
\description{
Example Transcript-Level Data Set
}
\examples{
print(tx_se_small)
}
\seealso{
Other Minimal Example Data: \code{\link{rse_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
