% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniteInterestingGroups-methods.R
\docType{methods}
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups,DataFrame-method}
\title{Unite interesting groups into a single column}
\usage{
uniteInterestingGroups(object, ...)

\S4method{uniteInterestingGroups}{DataFrame}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Now contains an \code{interestingGroups} column.
}
\description{
Create a single interesting groups column (\code{interestingGroups}) used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by a colon.
}
\note{
Updated 2019-07-28.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## DataFrame ====
object <- rse
x <- uniteInterestingGroups(
    object = sampleData(object),
    interestingGroups = interestingGroups(object)
)
print(x)
}
