% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsplit-methods.R
\docType{methods}
\name{unsplit}
\alias{unsplit}
\alias{unsplit,SplitDataFrameList,factor-method}
\title{Unsplit a split data frame}
\usage{
unsplit(value, f, drop = FALSE)

\S4method{unsplit}{SplitDataFrameList,factor}(value, f, drop = FALSE)
}
\arguments{
\item{value}{\code{SplitDataFrame}.}

\item{f, drop}{Intentionally not supported.}
}
\value{
\code{DataFrame}.
}
\description{
Unsplit a \code{SplitDataFrame} object.
}
\details{
Note that the grouping \code{factor} defined in \code{f} must be named with the row
names of the original \code{DataFrame}.
}
\note{
Updated 2019-08-23.
}
\examples{
df <- DataFrame(
    a = seq_len(4L),
    b = as.factor(rep(c("b", "a"), each = 2L)),
    row.names = LETTERS[seq_len(4L)]
)
print(df)

f <- df[["b"]]
names(f) <- rownames(df)
print(f)

split <- split(x = df, f = df[["b"]])
print(split)

unsplit <- unsplit(split)
print(unsplit)
}
\seealso{
\itemize{
\item \code{help(topic = "DataFrameList", package = "IRanges")}
\item https://support.bioconductor.org/p/124076/
\item \code{BiocGenerics::unsplit()}.
\item \code{dplyr::ungroup()}.
}\preformatted{getMethod(
    f = "unsplit",
    signature = "List",
    where = asNamespace("IRanges")
)
}\preformatted{getMethod(
    f = "stack",
    signature = "DataFrameList",
    where = asNamespace("IRanges")
)
}
}
