% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{updateMessage}
\alias{updateMessage}
\title{Update message}
\format{An object of class \code{character} of length 1.}
\usage{
updateMessage
}
\description{
Update message
}
\note{
Updated 2019-08-21.
}
\examples{
message(updateMessage)
}
\keyword{internal}
