% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/zzz.R
\name{updatePackages}
\alias{updatePackages}
\alias{update_packages}
\title{Update all installed packages}
\usage{
updatePackages()
}
\description{
Ensure that all GitHub, Bioconductor, and CRAN packages are up to date.
}
\details{
Internaly this is a wrapper for \code{\link[devtools:update_packages]{devtools::update_packages()}} that improves
support for Bioconductor.
}
\keyword{internal}
