% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperCamelCase-methods.R
\docType{methods}
\name{upperCamelCase}
\alias{upperCamelCase}
\alias{upperCamelCase,atomic-method}
\alias{upperCamelCase,factor-method}
\alias{upperCamelCase,list-method}
\alias{upperCamelCase,matrix-method}
\alias{upperCamelCase,data.frame-method}
\alias{upperCamelCase,Vector-method}
\alias{upperCamelCase,DataTable-method}
\alias{upperCamelCase,Ranges-method}
\alias{upperCamelCase,Matrix-method}
\alias{upperCamelCase,SummarizedExperiment-method}
\title{Upper camel case}
\usage{
\S4method{upperCamelCase}{atomic}(object, names = TRUE, strict = FALSE)

\S4method{upperCamelCase}{factor}(object, names = TRUE, strict = FALSE)

\S4method{upperCamelCase}{list}(object, names = TRUE, strict = FALSE)

\S4method{upperCamelCase}{matrix}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{data.frame}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{Vector}(object, names = TRUE, mcols = TRUE,
  metadata = TRUE, strict = FALSE)

\S4method{upperCamelCase}{DataTable}(object, rownames = FALSE,
  colnames = TRUE, mcols = TRUE, metadata = TRUE, strict = FALSE)

\S4method{upperCamelCase}{Ranges}(object, names = FALSE, mcols = TRUE,
  metadata = TRUE, strict = FALSE)

\S4method{upperCamelCase}{Matrix}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE, assayNames = TRUE, rowData = TRUE,
  colData = TRUE, metadata = TRUE, strict = FALSE)
}
\arguments{
\item{object}{Object.}

\item{names}{\code{logical(1)}.
Sanitize names.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
}
\description{
Format character strings to use upper camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{ThisIsCamelCase}).
}
\details{
Note that lower camel case is generally recommended in R over the use of
upper camel case. However, upper camel case is recommended by Bioconductor
for S4 class names and corresponding generators, but not variables or
functions.
}
\note{
Updated 2019-09-25.
}
\examples{
data(syntactic, package = "acidtest")
lapply(syntactic, upperCamelCase)
}
