% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{url}
\alias{url}
\title{Concatenate Strings to Form a URL}
\usage{
url(..., protocol = c("https", "http", "ftp", "none"))
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{protocol}{\code{character(1)}. Desired protocol to use. Defaults to HTTPS
but HTTP and FTP are also supported. Use \code{"none"} if you want to prepare
a URL that already contains a protocol in the first element of the dots.}
}
\description{
Concatenate Strings to Form a URL
}
\examples{
## HTTPS
x <- url(
    "steinbaugh.com",
    "basejump",
    "reference",
    protocol = "https"
)
print(x)

## FTP
x <- url(
    "ftp.ensembl.org",
    "pub",
    "release-94",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.94.gtf.gz",
    protocol = "ftp"
)
print(x)
}
