% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validateClasses}
\alias{validateClasses}
\title{Validate Classes}
\usage{
validateClasses(object, expected, subset = FALSE)
}
\arguments{
\item{object}{Object.}

\item{expected}{\code{list}. Named list of expected classes per slot.}

\item{subset}{\code{boolean}. Only check a subset of slots in the object.}
}
\value{
\code{boolean} (TRUE) on sucess or \code{string} containing informative
message on failure.
}
\description{
Validity check capable of validating multiple slots in a single call.
}
\details{
To be used inside S4 \code{methods::setValidity()} call or with
\code{assertthat::validate_that()}. Particularly useful for checking multiple
slotted objects inside \code{metadata()}.
}
\examples{
data(rse)
validateClasses(
    object = S4Vectors::metadata(rse),
    expected = list(
        version = c("package_version", "numeric_version"),
        date = "Date",
        interestingGroups = "character"
    )
)
}
\seealso{
\code{assertthat::validate_that()}.
}
