% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wash}
\alias{wash}
\title{Wash a data frame.}
\usage{
wash(df)
}
\arguments{
\item{df}{Dirty data frame.}
}
\value{
Clean data frame.
}
\description{
This function cleans poorly formed separators, leading and trailing commas or
spaces, empty cells, and sets \code{NA} values if necessary.
}
