% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCounts.R
\name{writeCounts}
\alias{writeCounts}
\title{Write Counts}
\usage{
writeCounts(..., dir = ".", compress = FALSE)
}
\arguments{
\item{...}{Symbols.
Unquoted object names containing count matrices.}

\item{dir}{\code{character(1)}.
Output directory.}

\item{compress}{\code{logical(1)}.
Compress the files using gzip.}
}
\value{
Invisible \code{list}.
File paths.
}
\description{
Supports both bulk and single-cell RNA-seq count matrices. Bulk RNA-seq
counts are written to disk as comma separated values ("\code{.CSV}"). Single-cell
RNA-seq counts are written to disk in MatrixMarket format ("\code{.MTX}"), along
with the sample barcodes ("\code{.COLNAMES}"), and gene identifiers
("\code{.ROWNAMES}").
}
\details{
Automatic gzip compression is offered as a user-defined option. This setting
is enabled by default to save disk space. Note that the \href{http://readr.tidyverse.org/}{readr} package,
built into \href{https://www.rstudio.com/}{RStudio}, now natively supports compressed files.
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation.
}
\examples{
library(SummarizedExperiment)
library(SingleCellExperiment)

data(rse, sce)

rnaseq_counts <- SummarizedExperiment::assay(rse)
single_cell_counts <- SummarizedExperiment::assay(sce)

writeCounts(rnaseq_counts, single_cell_counts, dir = "example")
list.files("example")

## Clean up.
unlink("example", recursive = TRUE)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
