% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCsv.R
\name{writeCsv}
\alias{writeCsv}
\title{Write CSV files}
\usage{
writeCsv(..., dir = "results")
}
\arguments{
\item{...}{Unquoted names of existing objects to export}

\item{dir}{Target directory}
}
\description{
Quickly write comma separated values (CSV) files. Uses non-standard
evaluation to obtain objects.
}
\examples{
\dontrun{
x <- data.frame(col1 = 1:10, col2 = 11:20)
y <- data.frame(col3 = 21:30, col4 = 31:40)
writeCsv(x, y)
}
}
