% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregate-methods.R
\docType{methods}
\name{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateReplicates}
\alias{aggregate}
\alias{aggregateReplicates,matrix-method}
\alias{aggregateReplicates,sparseMatrix-method}
\alias{aggregateReplicates,SummarizedExperiment-method}
\alias{aggregateFeatures,matrix-method}
\alias{aggregateFeatures,sparseMatrix-method}
\alias{aggregateFeatures,SummarizedExperiment-method}
\title{Aggregate Columns or Rows}
\usage{
aggregateFeatures(object, ...)

aggregateReplicates(object, ...)

\S4method{aggregateReplicates}{matrix}(object, groupings)

\S4method{aggregateReplicates}{sparseMatrix}(object, groupings)

\S4method{aggregateReplicates}{SummarizedExperiment}(object)

\S4method{aggregateFeatures}{matrix}(object, groupings)

\S4method{aggregateFeatures}{sparseMatrix}(object, groupings)

\S4method{aggregateFeatures}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{groupings}{\code{factor}. Defines the aggregation groupings.
The new aggregate names are defined as the \code{factor} \code{levels}, and the
original, unaggregated names are defined as the \code{names}.}
}
\value{
Modified object, with aggregated columns (samples) or rows
(features).
}
\description{
Aggregate sample replicates (columns) or gene/transcript features (rows).
}
\details{
\code{\link[=aggregateReplicates]{aggregateReplicates()}} works across the columns, and is designed to
aggregate sample replicates. \code{\link[=aggregateFeatures]{aggregateFeatures()}} works down the rows, and
is designed to aggregate features (e.g. genes or transcripts). Most commonly,
the \code{\link[=aggregateFeatures]{aggregateFeatures()}} function can be used to aggregate counts from
transcript-level to gene-level.
}
\examples{
counts <- DataFrame(
    sample1_replicate1 = as.integer(c(0, 1, 1, 1)),
    sample1_replicate2 = as.integer(c(1, 0, 1, 1)),
    sample2_replicate1 = as.integer(c(1, 1, 0, 1)),
    sample2_replicate2 = as.integer(c(1, 1, 1, 0)),
    row.names = paste0("transcript", seq_len(4))
)
print(counts)

# Dense matrix
matrix <- as(counts, "matrix")
class(sparse)
print(matrix)

# Sparse matrix
sparse <- as(matrix, "sparseMatrix")
class(sparse)
print(sparse)

samples <- factor(c("sample1", "sample1", "sample2", "sample2"))
names(samples) <- colnames(counts)
print(samples)

genes <- factor(c("gene1", "gene1", "gene2", "gene2"))
names(genes) <- rownames(counts)
print(genes)

# SummarizedExperiment
se <- SummarizedExperiment(
    assay = list(counts = sparse),
    colData = DataFrame(
        sampleName = names(samples),
        aggregate = samples
    ),
    rowData = DataFrame(aggregate = genes)
)

# aggregateReplicates ====
aggregateReplicates(matrix, groupings = samples)
aggregateReplicates(sparse, groupings = samples)
aggregateReplicates(se)

# aggregateFeatures ====
aggregateFeatures(matrix, groupings = genes)
aggregateFeatures(sparse, groupings = genes)
aggregateReplicates(se)
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
