% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregate-methods.R
\docType{methods}
\name{aggregateCols}
\alias{aggregateCols}
\alias{aggregateRows}
\alias{aggregate}
\alias{aggregateCols,matrix-method}
\alias{aggregateCols,sparseMatrix-method}
\alias{aggregateCols,SummarizedExperiment-method}
\alias{aggregateCols,SingleCellExperiment-method}
\alias{aggregateRows,matrix-method}
\alias{aggregateRows,sparseMatrix-method}
\alias{aggregateRows,SummarizedExperiment-method}
\alias{aggregateReplicates}
\alias{aggregateFeatures}
\alias{aggregateSamples}
\title{Aggregate Columns or Rows}
\usage{
aggregateCols(object, ...)

aggregateRows(object, ...)

\S4method{aggregateCols}{matrix}(object, groupings)

\S4method{aggregateCols}{sparseMatrix}(object, groupings)

\S4method{aggregateCols}{SummarizedExperiment}(object)

\S4method{aggregateCols}{SingleCellExperiment}(object)

\S4method{aggregateRows}{matrix}(object, groupings)

\S4method{aggregateRows}{sparseMatrix}(object, groupings)

\S4method{aggregateRows}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{groupings}{\code{factor}. Defines the aggregation groupings.
The new aggregate names are defined as the \code{factor} \code{levels}, and the
original, unaggregated names are defined as the \code{names}.}
}
\value{
Modified object, with aggregated columns (samples) or rows
(features).
}
\description{
Aggregate sample replicates (columns) or gene/transcript features (rows).
}
\details{
\code{\link[=aggregateCols]{aggregateCols()}} works across the columns, and is designed to aggregate
sample replicates. \code{\link[=aggregateRows]{aggregateRows()}} works down the rows, and is designed to
aggregate features (e.g. genes or transcripts). Most commonly, the
\code{\link[=aggregateRows]{aggregateRows()}} function can be used to aggregate counts from
transcript-level to gene-level.
}
\section{SummarizedExperiment}{


Slot an \code{aggregate} column into \code{\link[SummarizedExperiment:colData]{SummarizedExperiment::colData()}} for
\code{\link[=aggregateCols]{aggregateCols()}}, or into \code{\link[SummarizedExperiment:rowData]{SummarizedExperiment::rowData()}} for
\code{aggregateRows()}. The S4 method will define the \code{groupings} automatically
and perform the aggregation.
}

\examples{
# Example data ====
genes <- factor(paste0("gene", rep(seq_len(2L), each = 2L)))
names(genes) <- rownames(counts)
print(genes)

samples <- factor(paste0("sample", rep(seq_len(2L), each = 2L)))
names(samples) <- colnames(counts)
print(samples)

cells <- factor(paste0("cell", rep(seq_len(2L), each = 2L)))
names(samples) <- colnames(counts)
print(samples)

counts <- data.frame(
    sample1_replicate1 = as.integer(c(0, 1, 1, 1)),
    sample1_replicate2 = as.integer(c(1, 0, 1, 1)),
    sample2_replicate1 = as.integer(c(1, 1, 0, 1)),
    sample2_replicate2 = as.integer(c(1, 1, 1, 0)),
    row.names = paste0("transcript", seq_len(4))
)
print(counts)

# matrix
matrix <- as(counts, "matrix")
class(matrix)
print(matrix)

# sparseMatrix
sparse <- as(matrix, "sparseMatrix")
class(sparse)
print(sparse)

# SummarizedExperiment
se <- SummarizedExperiment::SummarizedExperiment(
    assay = list(counts = sparse),
    colData = S4Vectors::DataFrame(
        sampleName = names(samples),
        aggregate = samples
    ),
    rowData = S4Vectors::DataFrame(aggregate = genes)
)
print(se)

# SingleCellExperiment
sce <- SingleCellExperiment::SingleCellExperiment(se)
colnames(sce) <- gsub("sample", "cell", colnames(sce))
print(sce)

# aggregateCols ====
aggregateCols(matrix, groupings = samples)
aggregateCols(sparse, groupings = samples)
aggregateCols(se)
aggregateCols(sce)  # FIXME

# aggregateRows ====
aggregateRows(matrix, groupings = genes)
aggregateRows(sparse, groupings = genes)
aggregateRows(se)
aggregateRows(sce)  # FIXME
}
\seealso{
Other Data Functions: \code{\link{cell2sample}},
  \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}, \code{\link{tpm}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
