% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/aggregateCellsToSamples-methods.R
\docType{methods}
\name{aggregateCellsToSamples}
\alias{aggregateCellsToSamples}
\alias{aggregateCellsToSamples,SingleCellExperiment-method}
\title{Aggregate Cells to Samples}
\usage{
aggregateCellsToSamples(object, ...)

\S4method{aggregateCellsToSamples}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{RangedSummarizedExperiment}.
}
\description{
Aggregate Cells to Samples
}
\examples{
x <- aggregateCellsToSamples(sce_small)
print(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}},
  \code{\link{topCellsPerSample}}, \code{\link{tpm}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
