% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/aggregateCellsToSamples-methods.R
\docType{methods}
\name{aggregateCellsToSamples}
\alias{aggregateCellsToSamples}
\alias{aggregateCellsToSamples,SingleCellExperiment-method}
\title{Aggregate Cells to Samples}
\usage{
aggregateCellsToSamples(object, ...)

\S4method{aggregateCellsToSamples}{SingleCellExperiment}(object,
  fun = c("sum", "mean"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{fun}{\code{string}. Name of the aggregation function. Uses \code{\link[=match.arg]{match.arg()}}.}
}
\value{
\code{SummarizedExperiment}. Object with cell-level counts aggregated
to sample-level.
}
\description{
Utilty function that factilites cell-to-sample aggregation. By default, this
function will sum the counts across cells to sample level.
}
\details{
This function is intended primarily for quality control analysis.

Internally it automatically obtains the cell-to-sample groupings and then
performs aggregation with the \code{\link[=aggregateCols]{aggregateCols()}} function.
}
\examples{
data(sce_small)
x <- aggregateCellsToSamples(sce_small)
print(x)
}
\seealso{
Other SingleCellExperiment Functions: \code{\link{cell2sample}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{subsetPerSample}},
  \code{\link{topCellsPerSample}}
}
\concept{SingleCellExperiment Functions}
