% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregateCols-methods.R
\name{aggregateCols}
\alias{aggregateCols}
\alias{aggregateCols,matrix-method}
\alias{aggregateCols,Matrix-method}
\alias{aggregateCols,SummarizedExperiment-method}
\alias{aggregateCols,SingleCellExperiment-method}
\title{Aggregate columns}
\usage{
aggregateCols(x, ...)

\S4method{aggregateCols}{matrix}(x, by, fun = c("sum", "mean", "median", "geometricMean"))

\S4method{aggregateCols}{Matrix}(x, by, fun = c("sum", "mean"))

\S4method{aggregateCols}{SummarizedExperiment}(x, col = "aggregate", fun = "sum")

\S4method{aggregateCols}{SingleCellExperiment}(x, fun)
}
\arguments{
\item{x}{Object.}

\item{by}{\code{factor}.
Aggregation groupings. The new aggregate names are defined as the \code{factor}
\link[base:levels]{levels}, and the original, unaggregated names are defined as
the \link[base:names]{names}.}

\item{fun}{\code{character(1)}.
Name of the aggregation function to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally.}

\item{col}{\code{character(1)}.
Name of column in either \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} that defines the desired
aggregation groupings.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with columns aggregated.
}
\description{
Aggregate columns
}
\section{Methods (by class)}{

\itemize{
\item \code{matrix}, \code{Matrix}:
Aggregate using a grouping \code{factor}.
\item \code{SummarizedExperiment}:
Aggregate data slotted in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays()}} using an automatically
generated grouping \code{factor}, which is obtained from a user-defined column
(\code{col} argument) in either the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} of the object. Slot an
\code{aggregate} column into \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData()}}
for \code{\link[=aggregateRows]{aggregateRows()}}, or into
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} for \code{\link[=aggregateCols]{aggregateCols()}}. This
method will define the \code{groupings} automatically, and perform the
aggregation.
\item \code{SingleCellExperiment}:
Aggregate \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays()}} across cell-level
groupings, defined by a column in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}}. Inherits from
\code{SummarizedExperiment}, and still relies upon slotting an \code{aggregate}
column into \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}}. Note that these
groupings will map to cells, so care must be taken to properly aggregate
samples.
}
}

\examples{
counts <- matrix(
    data = c(
        0L, 2L, 2L, 2L,
        2L, 0L, 2L, 2L,
        2L, 2L, 0L, 2L,
        2L, 2L, 2L, 0L
    ),
    nrow = 4L,
    ncol = 4L,
    byrow = TRUE,
    dimnames = list(
        paste0("transcript", seq_len(4L)),
        paste(
            paste0("sample", rep(seq_len(2L), each = 2L)),
            paste0("replicate", rep(seq_len(2L), times = 2L)),
            sep = "_"
        )
    )
)

samples <- factor(paste0("sample", rep(seq_len(2L), each = 2L)))
names(samples) <- colnames(counts)
print(samples)

## matrix ====
print(counts)
aggregateCols(counts, by = samples)

## Matrix ====
sparse <- as(counts, "sparseMatrix")
print(sparse)
aggregateCols(sparse, by = samples)

## SummarizedExperiment ====
se <- SummarizedExperiment::SummarizedExperiment(
    assays = SimpleList(counts = counts),
    colData = DataFrame(
        sampleName = as.factor(names(samples)),
        aggregate = samples
    )
)
print(se)
aggregateCols(se)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
