% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregateFeatures-methods.R
\docType{methods}
\name{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateFeatures,dgCMatrix-method}
\alias{aggregateFeatures,matrix-method}
\title{Aggregate Features}
\usage{
aggregateFeatures(object, ...)

\S4method{aggregateFeatures}{dgCMatrix}(object, groupings)

\S4method{aggregateFeatures}{matrix}(object, groupings)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{groupings}{\code{character}. Feature groupings (e.g. gene or transcript
IDs), defined as a named factor. The pooled features must be defined as the
factor levels, and the original features as the names of the factor.}
}
\value{
Object of same class with aggregated features (rows).
}
\description{
Aggregate Features
}
\examples{
counts <- data.frame(
    "sample_1" = as.integer(c(0, 1, 2, 3)),
    "sample_2" = as.integer(c(1, 2, 3, 4)),
    row.names = c("gene_1.1", "gene_1.2", "gene_2.1", "gene_2.2")
)

groupings <- factor(c("gene_1", "gene_1", "gene_2", "gene_2"))
names(groupings) <- rownames(counts)

# matrix ====
mat <- as(counts, "matrix")
aggregateFeatures(mat, groupings = groupings)

# dgCMatrix ====
dgc <- as(mat, "dgCMatrix")
aggregateFeatures(dgc, groupings = groupings)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{counts}}, \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
