% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregateReplicates-methods.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,dgCMatrix-method}
\alias{aggregateReplicates,matrix-method}
\title{Aggregate Replicates}
\usage{
aggregateReplicates(object, ...)

\S4method{aggregateReplicates}{dgCMatrix}(object, groupings)

\S4method{aggregateReplicates}{matrix}(object, groupings)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{groupings}{\code{character}. Factor that defines the aggregation groupings.
The new aggregation names are defined as the factor levels, and the
original replicates are defined as the names of the factor.}
}
\value{
Object with aggregated counts per pooled sample (columns).
}
\description{
Aggregate Replicates
}
\examples{
counts <- DataFrame(
    "sample1_rep1" = as.integer(c(0, 0, 0, 1, 2)),
    "sample1_rep2" = as.integer(c(0, 0, 0, 3, 4)),
    "sample2_rep1" = as.integer(c(1, 2, 0, 0, 0)),
    "sample2_rep2" = as.integer(c(3, 4, 0, 0, 0))
)

groupings <- factor(c("sample1", "sample1", "sample2", "sample2"))
names(groupings) <- colnames(counts)
print(groupings)

# matrix ====
mat <- as(counts, "matrix")
print(mat)
aggregateReplicates(mat, groupings = groupings)

# dgCMatrix ====
dgc <- as(mat, "dgCMatrix")
print(dgc)
aggregateReplicates(dgc, groupings = groupings)
}
\seealso{
Other Data Functions: \code{\link{aggregateFeatures}},
  \code{\link{counts}}, \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
