% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,dgCMatrix-method}
\alias{aggregateReplicates,matrix-method}
\title{Aggregate Replicates}
\usage{
aggregateReplicates(object, ...)

\S4method{aggregateReplicates}{dgCMatrix}(object, groupings)

\S4method{aggregateReplicates}{matrix}(object, groupings)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{groupings}{Factor that defines the aggregation groupings. The new
aggregation names are defined as the factor levels, and the original
replicates are defined as the names of the factor.}
}
\value{
Object with aggregated counts per pooled sample (columns).
}
\description{
Aggregate Replicates
}
\examples{
counts <- data.frame(
    sample1_rep1 = c(0, 0, 0, 1, 2),
    sample1_rep2 = c(0, 0, 0, 3, 4),
    sample2_rep1 = c(1, 2, 0, 0, 0),
    sample2_rep2 = c(3, 4, 0, 0, 0)
)

groupings <- factor(c("sample1", "sample1", "sample2", "sample2"))
names(groupings) <- colnames(counts)

# matrix
mat <- as(counts, "matrix")
aggregateReplicates(mat, groupings = groupings)

# dgCMatrix
dgc <- as(mat, "dgCMatrix")
aggregateReplicates(dgc, groupings = groupings)
}
\seealso{
Other Aggregation Utilities: \code{\link{aggregateFeatures}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
