% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{annotable}
\alias{annotable}
\alias{arrange}
\alias{as_tibble}
\alias{filter}
\alias{gene2entrez}
\alias{gene2symbol}
\alias{mutate}
\alias{tx2gene}
\title{S4 generics}
\usage{
annotable(object)

arrange(object, ...)

as_tibble(from)

filter(object, ...)

gene2entrez(object)

gene2symbol(object)

mutate(object, ...)

tx2gene(object)
}
\arguments{
\item{object}{Primary object.}

\item{...}{Additional arguments.}

\item{from}{Object to coerce.}
}
\description{
S4 generics
}
