% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/reexports.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{aggregateReplicates}
\alias{bcbio}
\alias{bcbio<-}
\alias{interestingGroups}
\alias{metadataTable}
\alias{metrics}
\alias{sampleDirs}
\alias{sampleMetadata}
\alias{selectSamples}
\alias{tidy_filter}
\alias{tidy_select}
\alias{tidy_rename}
\title{S4 Generics}
\usage{
aggregateReplicates(object, ...)

bcbio(object, ...)

bcbio(object, ...) <- value

interestingGroups(object)

metadataTable(object, ...)

metrics(object)

sampleDirs(object)

sampleMetadata(object, ...)

selectSamples(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{x}{Primary object.}

\item{y}{Secondary object.}
}
\description{
S4 Generics
}
\keyword{internal}
