% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{plotGene}
\alias{plotQC}
\alias{tx2gene}
\title{S4 Generics}
\usage{
plotGene(object, ...)

plotQC(object, ...)

tx2gene(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending upon the method.
}
\description{
S4 Generics
}
\keyword{internal}
