% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{aggregateReplicates}
\alias{bcbio}
\alias{bcbio<-}
\alias{interestingGroups}
\alias{metrics}
\alias{plotGene}
\alias{sampleDirs}
\alias{sampleMetadata}
\alias{selectSamples}
\title{S4 Generics}
\usage{
aggregateReplicates(object, ...)

bcbio(object, ...)

bcbio(object, ...) <- value

interestingGroups(object, ...)

metrics(object, ...)

plotGene(object, ...)

sampleDirs(object, ...)

sampleMetadata(object, ...)

selectSamples(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{x}{Primary object.}

\item{y}{Secondary object.}
}
\description{
S4 Generics
}
\keyword{internal}
