% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{markdown}
\alias{plotQC}
\alias{tx2gene}
\title{S4 Generics}
\usage{
markdown(object, ...)

plotQC(object, ...)

tx2gene(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending upon the method.
}
\description{
S4 Generics
}
\seealso{
\itemize{
\item \code{\link[methods:isGeneric]{methods::isGeneric()}}
\item \code{\link[methods:showMethods]{methods::showMethods()}}
\item \code{\link[methods:getMethod]{methods::getMethod()}}
}
}
\keyword{internal}
