% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{alphaSummary}
\alias{alphaSummary}
\title{Alpha Level Cutoff Summary Statistics}
\usage{
alphaSummary(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{integer matrix}.
}
\description{
Quickly generate a summary table of various alpha level cutoffs.
}
\examples{
isS4(alphaSummary)
}
\seealso{
DESeqAnalysis.
}
