% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaThreshold-methods.R
\docType{methods}
\name{alphaThreshold}
\alias{alphaThreshold}
\alias{alphaThreshold<-}
\alias{alphaThreshold,Annotated-method}
\alias{alphaThreshold<-,Annotated,numeric-method}
\title{Alpha threshold}
\usage{
alphaThreshold(object, ...)

alphaThreshold(object, ...) <- value

\S4method{alphaThreshold}{Annotated}(object)

\S4method{alphaThreshold}{Annotated,numeric}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)}.
Alpha cutoff level.
}
\description{
Alpha threshold
}
\examples{
data(rse, package = "acidtest")

## Annotated
alphaThreshold(rse) <- 0.05
alphaThreshold(rse)
}
