% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-annotable.R
\docType{methods}
\name{annotable}
\alias{annotable}
\alias{annotable,character-method}
\title{\href{http://www.ensembl.org/}{Ensembl} Annotations}
\usage{
annotable(object, ...)

\S4method{annotable}{character}(object, format = "gene", release = 90L)
}
\arguments{
\item{object}{Object. Default usage is to provide Ensembl genome build as a
string.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{format}{Desired table format, either \code{gene}, \code{tx2gene}, or
\code{gene2symbol}.}

\item{release}{Ensembl release version. This function defaults to using the
most recent annotations available on AnnotationHub.}
}
\value{
\link{data.frame} with unique rows per gene or transcript.
}
\description{
Quickly access gene annotations and transcript-to-gene (tx2gene) mappings
pre-compiled from \href{http://www.ensembl.org/}{Ensembl} with the
\href{http://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
package. For gene annotables, the Entrez identifier is removed, to allow
for unique Ensembl gene identifiers.
}
\examples{
annotable("Mus musculus") \%>\% glimpse
}
\seealso{
Other Gene Annotation Utilities: \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGTF}},
  \code{\link{gene2symbol}}, \code{\link{symbol2gene}},
  \code{\link{tx2geneFromGTF}}, \code{\link{tx2gene}}
}
\author{
Broad class definitions by Rory Kirchner
}
\concept{Gene Annotation Utilities}
