% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotable.R
\name{annotable}
\alias{annotable}
\title{Ensembl Annotations}
\usage{
annotable(organism, format = c("genes", "transcripts"),
  genomeBuild = NULL, release = NULL)
}
\arguments{
\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{format}{\code{string}. Fetch ranges as "\code{genes}" or "\code{transcripts}".}

\item{genomeBuild}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}
}
\value{
\code{data.frame}.
}
\description{
Quickly obtain gene and transcript annotations from
\href{http://www.ensembl.org}{Ensembl} using
\href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub} and
\href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.
}
\details{
Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.
}
\section{Broad class definitions}{

For gene and transcript annotations, a \code{broadClass} column is added, which
generalizes the gene types into a smaller number of semantically-meaningful
groups:
\itemize{
\item \code{coding}
\item \code{noncoding}
\item \code{pseudo}
\item \code{small}
\item \code{decaying}
\item \code{ig} (immunoglobulin)
\item \code{tcr} (T cell receptor)
\item \code{other}
}
}

\section{GRCh37 legacy annotations}{

The \emph{Homo sapiens} GRCh37 (release 75) build is supported by internally
querying the
\href{https://doi.org/doi:10.18129/B9.bioc.EnsDb.Hsapiens.v75}{EnsDb.Hsapiens.v75}
package.
}

\examples{
# Genes ====
x <- annotable("Homo sapiens")
glimpse(x)

# Transcripts ====
x <- annotable("Homo sapiens", format = "transcripts")
glimpse(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Gene Annotation Functions}
