% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-annotable.R
\docType{methods}
\name{annotable}
\alias{annotable}
\alias{annotable,character-method}
\title{\href{http://www.ensembl.org/}{Ensembl} Annotations}
\usage{
annotable(object, ...)

\S4method{annotable}{character}(object, format = "gene",
  release = "current")
}
\arguments{
\item{object}{Object. Default usage is to provide Ensembl genome build as a
string.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{format}{Desired table format, either \code{gene}, \code{tx2gene}, or
\code{gene2symbol}.}

\item{release}{Ensembl release version. This function defaults to using the
most current release available on AnnotationHub (\code{current}).}
}
\value{
\link{data.frame} with unique rows per gene or transcript.
}
\description{
Fetch annotations from AnnotationHub using the ensembldb package. This
function defaults to obtaining the latest annotations, unless the \code{release}
argument is set to a numeric version (e.g. 88).
}
\examples{
annotable("Mus musculus") \%>\%
    str()
}
\seealso{
\itemize{
\item \href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
\item \href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.
}

Other Gene Annotation Utilities: \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{symbol2gene}},
  \code{\link{tx2geneFromGFF}}, \code{\link{tx2gene}}
}
\author{
Broad class definitions by Rory Kirchner
}
\concept{Gene Annotation Utilities}
