% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-annotable.R
\docType{methods}
\name{annotable}
\alias{annotable}
\alias{annotable}
\alias{annotable,character-method}
\title{\href{http://www.ensembl.org/}{Ensembl} Annotations}
\usage{
annotable(object, ...)

\S4method{annotable}{character}(object, format = "gene")
}
\arguments{
\item{object}{Object. Default usage is to provide Ensembl genome build as a
string.}

\item{...}{Additional arguments.}

\item{format}{Desired table format, either \strong{\code{gene}}, \code{tx2gene},
\code{gene2symbol}, or \code{gene2entrez}.}
}
\value{
\link{data.frame} with unique rows per gene or transcript.
}
\description{
Quickly access gene annotations and transcript-to-gene (tx2gene) mappings
pre-compiled from \href{http://www.ensembl.org/}{Ensembl} with the
\href{http://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
package. For gene annotables, the Entrez identifier is removed, to allow
for unique Ensembl gene identifiers.
}
\note{
If the \code{format} argument is set to \code{gene2entrez}, \code{\link[=annotable]{annotable()}} returns
a \link{tibble} with non-unique rows grouped by \code{ensgene}, instead of a
\link{data.frame}.
}
\examples{
annotable("hg38") \%>\% glimpse
}
\seealso{
Consult the annotables package documentation (\code{help("annotables")})
for a list of currently supported genomes.

Other Gene Annotation Utilities: \code{\link{detectOrganism}},
  \code{\link{gene2entrez}}, \code{\link{gene2symbol}},
  \code{\link{tx2geneFromGTF}}, \code{\link{tx2gene}}
}
\author{
Broad class definitions by Rory Kirchner
}
