% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-annotable.R
\docType{methods}
\name{annotable}
\alias{annotable}
\alias{annotable}
\alias{annotable,character-method}
\alias{annotable,data.frame-method}
\title{Ensembl Annotations (Legacy Method)}
\usage{
annotable(object, ...)

\S4method{annotable}{character}(object, format = "genes",
  genomeBuild = NULL, release = NULL, uniqueSymbol = FALSE)

\S4method{annotable}{data.frame}(object)
}
\arguments{
\item{format}{Desired annotation data format, either "\code{genes}",
"\code{transcripts}", "\code{gene2symbol}"., or "\code{tx2gene}".}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{uniqueSymbol}{Make gene symbols unique. Only applies to \code{genes} and
\code{gene2symbol} output.}
}
\value{
\link{data.frame}.
}
\description{
Ensembl Annotations (Legacy Method)
}
\details{
Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.
}
\section{Broad Class Definitions}{

For gene and transcript tables, when \code{broadClass = TRUE}, a \code{broadClass}
column is added, which generalizes the gene types into a smaller number of
semantically-meaningful groups:
\itemize{
\item \code{coding}
\item \code{noncoding}
\item \code{pseudo}
\item \code{small}
\item \code{decaying}
\item \code{ig} (immunoglobulin)
\item \code{tcr} (T cell receptor)
\item \code{other}
}
}

\examples{
annotable("Homo sapiens") \%>\% glimpse()

# Legacy GRCh37/hg19 genome build support
annotable("Homo sapiens", genomeBuild = "GRCh37") \%>\% glimpse()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{ensembl}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{genes}},
  \code{\link{sanitizeAnnotable}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{transcripts}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
