% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antilog2.R
\name{antilog2}
\alias{antilog2}
\title{Log base 2 antilog conversion}
\usage{
antilog2(log2)
}
\arguments{
\item{log2}{\code{numeric} vector of log base 2 normalized values.}
}
\value{
Antilog \code{numeric} vector with proper sign.
}
\description{
Convert a string of log2 normalized values (e.g. fold changes) for better
readability.
}
\examples{
antilog2(c(-2, -1, 0, 1, 2))
}
\keyword{math}

