% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-methods.R
\docType{methods}
\name{as}
\alias{as}
\alias{coerce}
\alias{coerce,sparseMatrix,data.frame-method}
\alias{coerce,sparseMatrix,DataFrame-method}
\alias{coerceS4ToList}
\alias{coerce,SummarizedExperiment,list-method}
\alias{as.list,SummarizedExperiment-method}
\title{Force an Object to Belong to a Class}
\usage{
coerceS4ToList(from)

\S4method{as.list}{SummarizedExperiment}(x)
}
\arguments{
\item{from}{The classes between which the coerce methods
    \code{def} and \code{replace} perform coercion.
  }

\item{x}{Object.}
}
\value{
Object of new class.
}
\description{
Coerce an object to a given class.
}
\section{list}{


It is often useful to coerce an S4 object to a flat \code{list} for archival
storage. Here we are providing the \code{\link[=coerceS4ToList]{coerceS4ToList()}} function, which
consistently coerces the slots in any S4 to a standard \code{list}. Additionally,
here we have improved support for \code{SummarizedExperiment} to \code{list} coercion,
returning the slots as a \code{list}.
}

\section{Description}{

    \code{as(object)}
      returns the version of this object coerced to be the given
      \code{Class}.  When used in the replacement form on the left of
      an assignment, the portion of the object corresponding to
      \code{Class} is replaced by \code{value}.

      The operation of \code{as()} in either form depends on the
      definition of coerce methods.  Methods are defined automatically
      when the two classes are related by inheritance; that is, when
      one of the classes is a subclass of the other.

      Coerce methods are also predefined for basic classes (including all
      the types of vectors, functions and a few others).

      Beyond these two sources of methods, further methods are defined
      by calls to the \code{\link{setAs}} function.  See that
      documentation also for details of how coerce methods work. Use
      \code{showMethods(coerce)} for a list of all currently defined methods, as in the
      example below.

 
  
}

\section{Basic Coercion Methods}{

  Methods are pre-defined for coercing any object to one of the basic
  datatypes.  For example, \code{as(x, "numeric")} uses the existing
  \code{as.numeric} function.  These and all other existing methods
  can be listed as shown in the example.

}

\examples{
data(rse_small)

## SummarizedExperiment to list ====
x <- as(rse_small, Class = "list")
class(x)
names(x)
}
\references{
Chambers, John M. (2016)
 \emph{Extending R},
  Chapman & Hall.
(Chapters 9 and 10.)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:canCoerce]{methods::canCoerce()}}.
\item \code{\link[methods:showMethods]{methods::showMethods()}}.
\item \code{\link[utils:methods]{utils::methods()}}.
\item \code{\link[=as.list]{as.list()}}.
}
}
