% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\alias{assert_formal_annotation_col}
\alias{assert_formal_color_function}
\alias{assert_formal_compress}
\alias{assert_formal_header_level}
\alias{assert_has_rownames}
\alias{assert_is_a_string_or_null}
\alias{assert_is_character_or_null}
\alias{assert_is_implicit_integer}
\alias{assert_is_numeric_scalar_or_null}
\alias{has_rownames}
\alias{is_implicit_integer}
\title{Assert Check Functions}
\usage{
assert_formal_annotation_col(object, colData)

assert_formal_color_function(x)

assert_formal_compress(x)

assert_formal_header_level(x)

assert_has_rownames(x)

assert_is_a_string_or_null(x)

assert_is_character_or_null(x)

assert_is_implicit_integer(x)

assert_is_numeric_scalar_or_null(x)

has_rownames(x)

is_implicit_integer(x)
}
\arguments{
\item{object}{Object.}

\item{colData}{Column data.}

\item{x}{Primary object.}
}
\description{
These functions are generally structured following the syntax in the
assertive package.
}
\note{
\code{\link[=assert_has_rownames]{assert_has_rownames()}} is a stricter alternative to the assertive
version that works properply with data frames.

\code{tibble::has_rownames()} works better than \code{assertive::has_rownames()}
for data frames and tibbles.
}
\seealso{
\code{\link[base:all.equal]{base::all.equal()}} \code{tolerance} parameter.
}
\keyword{internal}
