% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_is_annotable.R
\name{assert_is_annotable}
\alias{assert_is_annotable}
\title{Ensembl Annotations Assert Check}
\usage{
assert_is_annotable(x)
}
\arguments{
\item{x}{\link{data.frame} containing Ensembl gene annotations returned
from the \code{\link[=annotable]{annotable()}} function.}
}
\value{
Abort on mismatch.
}
\description{
Ensembl Annotations Assert Check
}
\examples{
annotable <- annotable("Homo sapiens")
assert_is_annotable(annotable)
}
\seealso{
Other Assert Checks: \code{\link{assert_formal_annotation_col}},
  \code{\link{assert_formal_color_function}},
  \code{\link{assert_formal_compress}},
  \code{\link{assert_formal_header_level}},
  \code{\link{assert_has_rownames}},
  \code{\link{assert_is_a_string_or_null}},
  \code{\link{assert_is_character_or_null}},
  \code{\link{assert_is_gene2symbol}},
  \code{\link{assert_is_implicit_integer}},
  \code{\link{assert_is_numeric_scalar_or_null}},
  \code{\link{assert_is_tx2gene}}, \code{\link{assert}}
}
