% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_is_gene2symbol.R
\name{assert_is_gene2symbol}
\alias{assert_is_gene2symbol}
\title{Check Gene to Symbol Mapping Data}
\usage{
assert_is_gene2symbol(object)
}
\arguments{
\item{object}{\link{data.frame} containing Ensembl gene identifier to gene symbol
mappings. Must be structured as a two column \link{data.frame} with "ensgene"
and "symbol" columns.}
}
\value{
Invisible \code{TRUE} on pass, \code{\link[=stop]{stop()}} on error.
}
\description{
Check Gene to Symbol Mapping Data
}
\examples{
gene2symbol <- annotable("Homo sapiens", format = "gene2symbol")
assert_is_gene2symbol(gene2symbol)
}
