% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_is_tx2gene.R
\name{assert_is_tx2gene}
\alias{assert_is_tx2gene}
\title{Check Transcript to Gene Mapping Data}
\usage{
assert_is_tx2gene(object)
}
\arguments{
\item{object}{\link{data.frame} containing Ensembl transcript to gene identifier
mappings. Must be structured as a two column \link{data.frame} with "enstxp" and
"ensgene" columns.}
}
\value{
Abort on mismatch.
}
\description{
Check Transcript to Gene Mapping Data
}
\examples{
tx2gene <- annotable("Homo sapiens", format = "tx2gene")
assert_is_tx2gene(tx2gene)
}
\seealso{
Other Assert Checks: \code{\link{assert_formal_annotation_col}},
  \code{\link{assert_formal_color_function}},
  \code{\link{assert_formal_compress}},
  \code{\link{assert_formal_gene2symbol}},
  \code{\link{assert_formal_header_level}},
  \code{\link{assert_has_rownames}},
  \code{\link{assert_is_a_number_or_null}},
  \code{\link{assert_is_a_string_or_null}},
  \code{\link{assert_is_annotable}},
  \code{\link{assert_is_character_or_null}},
  \code{\link{assert_is_gene2symbol}},
  \code{\link{assert_is_implicit integer}},
  \code{\link{assert}}
}
