% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertAllAreNonExisting.R
\name{assertAllAreNonExisting}
\alias{assertAllAreNonExisting}
\alias{assert_all_are_non_existing}
\title{All Variables Do Not Exist Assert Check}
\usage{
assertAllAreNonExisting(x, envir = parent.frame(), inherits = FALSE)

assert_all_are_non_existing(x, envir = parent.frame(), inherits = FALSE)
}
\arguments{
\item{x}{Character vector of variable names to check in environment.}

\item{envir}{Environment.}

\item{inherits}{Should the enclosing frames of the environment be searched?}
}
\value{
Abort on mismatch.
}
\description{
All Variables Do Not Exist Assert Check
}
\examples{
a <- 1L
b <- 2L
tryCatch(
    assertAllAreNonExisting(c("a", "b", "c")),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
