% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertAllAreNonExisting}
\alias{assertAllAreNonExisting}
\title{Assert All Variables Are Non-Existing}
\usage{
assertAllAreNonExisting(object, envir = parent.frame(),
  inherits = FALSE)
}
\arguments{
\item{object}{\code{character}. Variable names to check in \code{environment}.}

\item{envir}{\code{environment}.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}
}
\value{
Always stop on error.
}
\description{
Assert All Variables Are Non-Existing
}
\examples{
assertAllAreNonExisting(c("XXX", "YYY"))
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
