% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertAllAreNonExisting}
\alias{assertAllAreNonExisting}
\title{Assert All Variables Are Non-Existing}
\usage{
assertAllAreNonExisting(x, envir = parent.frame(), inherits = FALSE)
}
\arguments{
\item{x}{Character vector of variable names to check in environment.}

\item{envir}{Environment.}

\item{inherits}{Should the enclosing frames of the environment be searched?}
}
\value{
Stop on mismatch.
}
\description{
Assert All Variables Are Non-Existing
}
\examples{
assertAllAreNonExisting(c("XXX", "YYY"))
}
\seealso{
Other Assert Check Functions: \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGFF}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assertIsURL}}
}
\author{
Michael Steinbaugh
}
