% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertAllAreNonExisting}
\alias{assertAllAreNonExisting}
\title{Assert All Variables Are Non-Existing}
\usage{
assertAllAreNonExisting(object, envir = parent.frame(),
  inherits = FALSE)
}
\arguments{
\item{object}{\code{character}. Variable names to check in \code{environment}.}

\item{envir}{\code{environment}.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}
}
\value{
Always stop on error.
}
\description{
Assert All Variables Are Non-Existing
}
\examples{
assertAllAreNonExisting(c("XXX", "YYY"))
}
