% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertAllAreUniqueGeneNames}
\alias{assertAllAreUniqueGeneNames}
\title{Assert All Are Unique Gene Names}
\usage{
assertAllAreUniqueGeneNames(object, genes)
}
\arguments{
\item{object}{\code{SummarizedExperiment}.}

\item{genes}{\code{character}. Input vector to check against definitions in the
corresponding \code{SummarizedExperiment}.}
}
\value{
Always stop on error.
}
\description{
This assert check determines if a user-defined gene name query is using only
unique (non-amgibuous) symbols. It is designed to be used for gene plotting
particularly when performing single-cell RNA-seq marker analysis.
}
\examples{
object <- rse_small
print(object)
genes <- head(as.character(rowData(rse_small)[["geneName"]]))
print(genes)
assertAllAreUniqueGeneNames(object = object, genes = genes)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
