% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertAllAreURL}
\alias{assertAllAreURL}
\alias{isURL}
\title{Assert All Are URL}
\usage{
assertAllAreURL(object)

isURL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Stop on mismatch.
}
\description{
Assert All Are URL
}
\examples{
assertAllAreURL(c(
    "https://www.r-project.org",
    "ftp://r-project.org"
))
isURL(c(
    "http://www.r-project.org",
    "https://www.r-project.org",
    "ftp://r-project.org",
    "r-project.org"
))
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
