% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertAllAreURL}
\alias{assertAllAreURL}
\alias{isURL}
\title{Assert All Are URL}
\usage{
assertAllAreURL(object)

isURL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert All Are URL
}
\examples{
assertAllAreURL(c(
    "https://www.r-project.org",
    "ftp://r-project.org"
))
isURL(c(
    "http://www.r-project.org",
    "https://www.r-project.org",
    "ftp://r-project.org",
    "r-project.org"
))
}
