% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertAllAreValidNames}
\alias{assertAllAreValidNames}
\alias{assertHasValidDimnames}
\alias{assertHasValidNames}
\alias{validNames}
\title{Assert All Are Valid Names}
\usage{
assertAllAreValidNames(object)

assertHasValidDimnames(object)

assertHasValidNames(object)

validNames(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Stop on mismatch.
}
\description{
Assert All Are Valid Names
}
\examples{
assertAllAreValidNames(c("sample_1", "sample_2"))
validNames(c(
    "sample_1",
    "gene_1",
    "293cells",
    "cell-AAAAAAAA",
    "GFP+ sort"
))
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
