% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertAllAreValidNames}
\alias{assertAllAreValidNames}
\alias{assertHasValidDimnames}
\alias{assertHasValidNames}
\alias{validDimnames}
\alias{validNames}
\title{Assert All Are Valid Names}
\usage{
assertAllAreValidNames(object)

assertHasValidDimnames(object)

assertHasValidNames(object)

validDimnames(object)

validNames(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert All Are Valid Names
}
\examples{
assertAllAreValidNames(c("sample1", "sample2"))
validDimnames(basejump::rse_small)
validDimnames(datasets::mtcars)
validNames(c(
    "sample_1",
    "gene_1",
    "293cells",
    "cell-AAAAAAAA",
    "GFP+ sort"
))
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2Symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDir}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
