% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertAllAreValidNames}
\alias{assertAllAreValidNames}
\alias{assertHasValidDimnames}
\alias{assertHasValidNames}
\alias{validDimnames}
\alias{validNames}
\title{Assert All Are Valid Names}
\usage{
assertAllAreValidNames(object)

assertHasValidDimnames(object)

assertHasValidNames(object)

validDimnames(object)

validNames(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert All Are Valid Names
}
\examples{
assertAllAreValidNames(c("sample1", "sample2"))
validDimnames(datasets::mtcars)
validNames(c(
    "sample_1",
    "gene_1",
    "293cells",
    "cell-AAAAAAAA",
    "GFP+ sort"
))
}
