% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertAreGeneAnnotations.R
\name{assertAreGeneAnnotations}
\alias{assertAreGeneAnnotations}
\title{Assert Are Ensembl Gene Annotations}
\usage{
assertAreGeneAnnotations(x, severity = "stop")
}
\arguments{
\item{x}{\link{data.frame} containing Ensembl gene annotations returned
by the \code{\link[=genes]{genes()}} function.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Assert Are Ensembl Gene Annotations
}
\examples{
# Success
genes <- genes("Homo sapiens")
assertAreGeneAnnotations(genes)

# Failure
transcripts <- transcripts("Homo sapiens")
tryCatch(
    assertAreGeneAnnotations(transcripts),
    error = function(e) e
)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
