% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertAreGeneAnnotations}
\alias{assertAreGeneAnnotations}
\title{Assert Are Gene Annotations}
\usage{
assertAreGeneAnnotations(object)
}
\arguments{
\item{object}{Object that can be coerced to \code{data.frame}.}
}
\value{
Always stop on error.
}
\description{
Must contain \code{geneID} and \code{geneName} columns. Does not need to contain
rownames, so \code{tibble} class is supported.
}
\examples{
object <- tibble::tibble(
    geneID = "ENSG00000000003",
    geneName = "TSPAN6"
)
assertAreGeneAnnotations(object)
}
