% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertAreTranscriptAnnotations.R
\name{assertAreTranscriptAnnotations}
\alias{assertAreTranscriptAnnotations}
\title{Assert Are Ensembl Transcript Annotations}
\usage{
assertAreTranscriptAnnotations(x, severity = "stop")
}
\arguments{
\item{x}{\link{data.frame} containing Ensembl transcript annotations returned
by the \code{\link[=transcripts]{transcripts()}} function.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Assert Are Ensembl Transcript Annotations
}
\examples{
# Success
transcripts <- transcripts("Homo sapiens")
assertAreTranscriptAnnotations(transcripts)

# Failure
genes <- genes("Homo sapiens")
tryCatch(
    assertAreTranscriptAnnotations(genes),
    error = function(e) e)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
