% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertAreTranscriptAnnotations}
\alias{assertAreTranscriptAnnotations}
\title{Assert Are Ensembl Transcript Annotations}
\usage{
assertAreTranscriptAnnotations(object)
}
\arguments{
\item{object}{Object that can be coerced to \code{data.frame}.}
}
\value{
Always stop on error.
}
\description{
Must contain \code{transcriptID} and \code{transcriptName} columns. Does not need to
contain rownames, so \code{tibble} class is supported.
}
\examples{
object <- tibble(
    transcriptID = "ENST00000000233",
    geneID = "ENSG00000004059"
)
assertAreTranscriptAnnotations(object)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
