% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertAreTranscriptAnnotations}
\alias{assertAreTranscriptAnnotations}
\title{Assert Are Ensembl Transcript Annotations}
\usage{
assertAreTranscriptAnnotations(object)
}
\arguments{
\item{object}{Object that can be coerced to \code{data.frame}.}
}
\value{
Always stop on error.
}
\description{
Must contain \code{transcriptID} and \code{transcriptName} columns. Does not need to
contain rownames, so \code{tibble} class is supported.
}
\examples{
object <- tibble::tibble(
    transcriptID = "ENST00000000233",
    geneID = "ENSG00000004059"
)
assertAreTranscriptAnnotations(object)
}
