% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFormalAnnotationCol.R
\name{assertFormalAnnotationCol}
\alias{assertFormalAnnotationCol}
\title{Assert Formal Annotation Column}
\usage{
assertFormalAnnotationCol(x, colData,
  severity = getOption("assertive.severity", "stop"))
}
\arguments{
\item{x}{Object.}

\item{colData}{Column data.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Assert Formal Annotation Column
}
\examples{
x <- data.frame(
    "sample_1" = c(1L, 2L),
    "sample_2" = c(3L, 4L),
    row.names = c("gene_1", "gene_2"),
    stringsAsFactors = FALSE
)
colData <- data.frame(
    "genotype" = c("wt", "ko"),
    row.names = c("sample_1", "sample_2"),
    stringsAsFactors = TRUE
)
assertFormalAnnotationCol(x, colData)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGFF}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
