% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFormalColorFunction.R
\name{assertFormalColorFunction}
\alias{assertFormalColorFunction}
\alias{assert_formal_color_function}
\title{Color Function Formal Assert Check}
\usage{
assertFormalColorFunction(x, severity = "stop")

assert_formal_color_function(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Color Function Formal Assert Check
}
\examples{
assertFormalColorFunction(viridis)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
