% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFormalCompress.R
\name{assertFormalCompress}
\alias{assertFormalCompress}
\alias{assert_formal_compress}
\title{Compression Formal Assert Check}
\usage{
assertFormalCompress(x, severity = "stop")

assert_formal_compress(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Compression Formal Assert Check
}
\examples{
assertFormalCompress("xz")
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
