% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertFormalCompress}
\alias{assertFormalCompress}
\title{Assert Formal Compression}
\usage{
assertFormalCompress(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Formal Compression
}
\examples{
assertFormalCompress("xz")
}
