% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertFormalGene2symbol}
\alias{assertFormalGene2symbol}
\title{Assert Formal Gene to Symbol Mappings}
\usage{
assertFormalGene2symbol(object, genes, gene2symbol)
}
\arguments{
\item{object}{Object containing gene identifiers in \code{rownames}. Cannot
contain gene names, or the check will fail.}

\item{genes}{\code{character}. Gene identifiers.}

\item{gene2symbol}{\code{DataFrame}, \code{tbl_df}, or \code{NULL}. Gene-to-symbol mappings.}
}
\value{
Stop on mismatch.
}
\description{
Assert Formal Gene to Symbol Mappings
}
\examples{
gene2symbol <- tibble(
    geneID = c("ENSG00000000003", "ENSG00000000005"),
    geneName = c("TSPAN6", "TNMD")
)
genes <- pull(gene2symbol, "geneID")
object <- DataFrame(
    "sample_1" = c(1L, 2L),
    "sample_2" = c(3L, 4L),
    row.names = genes
)
assertFormalGene2symbol(object, genes, gene2symbol)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
