% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFormalGene2symbol.R
\name{assertFormalGene2symbol}
\alias{assertFormalGene2symbol}
\alias{assert_formal_gene2symbol}
\title{Gene to Symbol Mappings Formal Assert Check}
\usage{
assertFormalGene2symbol(x, genes, gene2symbol, severity = "stop")

assert_formal_gene2symbol(x, genes, gene2symbol, severity = "stop")
}
\arguments{
\item{x}{Object containing gene identifiers in the rownames.}

\item{genes}{Character vector of gene identifiers.}

\item{gene2symbol}{Gene-to-symbol mappings data frame.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Gene to Symbol Mappings Formal Assert Check
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
