% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFormalGene2symbol.R
\name{assertFormalGene2symbol}
\alias{assertFormalGene2symbol}
\title{Assert Formal Gene to Symbol Mappings}
\usage{
assertFormalGene2symbol(x, genes, gene2symbol, severity = "stop")
}
\arguments{
\item{x}{Object containing gene identifiers in the rownames.}

\item{genes}{Character vector of gene identifiers.}

\item{gene2symbol}{Gene-to-symbol mappings data frame.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Assert Formal Gene to Symbol Mappings
}
\examples{
gene2symbol <- gene2symbol("Homo sapiens")
genes <- head(rownames(gene2symbol), 2L)
x <- data.frame(
    sample1 = c(1L, 2L),
    saple2 = c(3L, 4L),
    row.names = genes,
    stringsAsFactors = FALSE)
assertFormalGene2symbol(x, genes, gene2symbol)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
