% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertFormalGene2symbol}
\alias{assertFormalGene2symbol}
\title{Assert Formal Gene to Symbol Mappings}
\usage{
assertFormalGene2symbol(x, genes, gene2symbol,
  severity = getOption("assertive.severity", "stop"))
}
\arguments{
\item{x}{Object containing gene identifiers in the row names.}

\item{genes}{\code{character} vector of gene identifiers.}

\item{gene2symbol}{Gene-to-symbol mappings. Must contain a \code{data.frame} or
\code{NULL}.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Stop on mismatch.
}
\description{
Assert Formal Gene to Symbol Mappings
}
\examples{
gene2symbol <- data.frame(
    "geneID" = c("ENSG00000000003", "ENSG00000000005"),
    "geneName" = c("TSPAN6", "TNMD"),
    row.names = c("ENSG00000000003", "ENSG00000000005")
)
genes <- head(rownames(gene2symbol), 2L)
x <- data.frame(
    "sample_1" = c(1L, 2L),
    "sample_2" = c(3L, 4L),
    row.names = genes,
    stringsAsFactors = FALSE
)
assertFormalGene2symbol(x, genes, gene2symbol)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGFF}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assertIsURL}}
}
\author{
Michael Steinbaugh
}
