% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalie.R
\name{assertFormalInterestingGroups}
\alias{assertFormalInterestingGroups}
\title{Check Interesting Groups}
\usage{
assertFormalInterestingGroups(x, interestingGroups)
}
\arguments{
\item{x}{S4 class x.}

\item{interestingGroups}{\code{character}. Interesting groups.}
}
\value{
No value.
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
data(rse, package = "basejump")
assertFormalInterestingGroups(rse, "treatment")
assertFormalInterestingGroups(rse, NULL)
}
