% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertFormalInterestingGroups}
\alias{assertFormalInterestingGroups}
\title{Interesting Groups Formal Assert Check}
\usage{
assertFormalInterestingGroups(object, interestingGroups)
}
\arguments{
\item{object}{\code{SummarizedExperiment}.}

\item{interestingGroups}{\code{character}. Groups of interest.}
}
\value{
Always stop on error.
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
data(rse_small)
assertFormalInterestingGroups(rse_small, "treatment")
assertFormalInterestingGroups(rse_small, NULL)
}
