% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertHasRownames}
\alias{assertHasRownames}
\alias{hasRownames}
\title{Assert Has Rownames}
\usage{
assertHasRownames(object)

hasRownames(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
A stricter alternative to the assertive version that works properly with
data frames.
}
\examples{
object <- DataFrame(
    "sample1" = c(1L, 2L),
    "sample2" = c(3L, 4L),
    row.names = c("gene1", "gene2")
)
print(object)
assertHasRownames(object)
}
