% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertHasRownames.R
\name{assertHasRownames}
\alias{assertHasRownames}
\alias{hasRownames}
\title{Assert Has Rownames}
\usage{
assertHasRownames(x, severity = getOption("assertive.severity", "stop"))

hasRownames(x)
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
A stricter alternative to the assertive version that works properply with
data frames.
}
\note{
\code{tibble::has_rownames()} works better than \code{assertive::has_rownames()}
for data frames and tibbles.
}
\examples{
data <- data.frame(
    "sample_1" = c(1L, 2L),
    "sample_2" = c(3L, 4L),
    row.names = c("gene_1", "gene_2"),
    stringsAsFactors = FALSE
)
assertHasRownames(data)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGFF}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
\author{
Michael Steinbaugh
}
