% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertHasRownames.R
\name{assertHasRownames}
\alias{assertHasRownames}
\alias{hasRownames}
\title{Has Rownames Assert Check}
\usage{
assertHasRownames(x, severity = "stop")

hasRownames(x)
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
A stricter alternative to the assertive version that works properply with
data frames.
}
\note{
\code{tibble::has_rownames()} works better than \code{assertive::has_rownames()}
for data frames and tibbles.
}
\examples{
data <- data.frame(
    sample1 = c(1L, 2L),
    sample2 = c(3L, 4L),
    row.names = c("gene1", "gene2"),
    stringsAsFactors = FALSE)
assertHasRownames(data)

# Now set the rownames as NULL
rownames(data) <- NULL
tryCatch(
    assertHasRownames(data),
    error = function(e) e)

tibble <- tibble(
    sample1 = c(1L, 2L),
    sample2 = c(3L, 4L)
)
tryCatch(
    assertHasRownames(tibble),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
