% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertions}
\alias{assertions}
\title{Assertions}
\arguments{
\item{object}{Object.}

\item{envir}{\code{environment}.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}
}
\value{
Always stop on error.
}
\description{
Assertions
}
\keyword{internal}
