% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsAlpha}
\alias{assertIsAlpha}
\title{Assert Is Alpha Level}
\usage{
assertIsAlpha(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
An alpha level must be between 0 and 1, but not equal either 0 or 1.
}
\examples{
assertIsAlpha(0.05)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
