% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsAnIntegerOrNULL.R
\name{assertIsAnIntegerOrNULL}
\alias{assertIsAnIntegerOrNULL}
\alias{assert_is_an_integer_or_null}
\title{Is an Integer or NULL Assert Check}
\usage{
assertIsAnIntegerOrNULL(x, severity = "stop")

assert_is_an_integer_or_null(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Is an Integer or NULL Assert Check
}
\examples{
# Success
assertIsAnIntegerOrNULL(1L)

# Failure (not scalar)
tryCatch(
    assertIsAnIntegerOrNULL(c(1L, 2L)),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
