% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsAnIntegerOrNULL}
\alias{assertIsAnIntegerOrNULL}
\title{Assert Is an Integer or NULL}
\usage{
assertIsAnIntegerOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is an Integer or NULL
}
\examples{
assertIsAnIntegerOrNULL(1L)
assertIsAnIntegerOrNULL(NULL)
}
