% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsANumberOrNULL.R
\name{assertIsANumberOrNULL}
\alias{assertIsANumberOrNULL}
\alias{assert_is_a_number_or_null}
\title{Is a Number or NULL Assert Check}
\usage{
assertIsANumberOrNULL(x, severity = "stop")

assert_is_a_number_or_null(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Is a Number or NULL Assert Check
}
\examples{
# Success
assertIsANumberOrNULL(1.1)
assertIsANumberOrNULL(NULL)

# Failure
tryCatch(
    assertIsANumberOrNULL(c(1.1, 1.2)),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
