% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsANumberOrNULL}
\alias{assertIsANumberOrNULL}
\title{Assert Is a Number or NULL}
\usage{
assertIsANumberOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is a Number or NULL
}
\examples{
assertIsANumberOrNULL(1.1)
assertIsANumberOrNULL(NULL)
}
