% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsAStringOrNULL.R
\name{assertIsAStringOrNULL}
\alias{assertIsAStringOrNULL}
\alias{assert_is_a_string_or_null}
\title{Is a String or NULL Assert Check}
\usage{
assertIsAStringOrNULL(x, severity = "stop")

assert_is_a_string_or_null(x, severity = "stop")
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Is a String or NULL Assert Check
}
\examples{
# Success
assertIsAStringOrNULL("hello world")
assertIsAStringOrNULL(NULL)

# Failure
tryCatch(
    assertIsAStringOrNULL(c("hello", "world")),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
