% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsAStringOrNULL}
\alias{assertIsAStringOrNULL}
\title{Assert Is a String or NULL}
\usage{
assertIsAStringOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is a String or NULL
}
\examples{
assertIsAStringOrNULL("hello world")
assertIsAStringOrNULL(NULL)
}
