% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsCharacterOrNULL}
\alias{assertIsCharacterOrNULL}
\title{Assert Is Character Vector or NULL}
\usage{
assertIsCharacterOrNULL(x, severity = getOption("assertive.severity",
  "stop"))
}
\arguments{
\item{x}{Object.}

\item{severity}{\code{string}. How severe should the consequences of the assertion
be? Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Stop on mismatch.
}
\description{
Assert Is Character Vector or NULL
}
\examples{
assertIsCharacterOrNULL(c("hello", "world"))
assertIsCharacterOrNULL(NULL)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
