% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsColorScaleContinuousOrNULL}
\alias{assertIsColorScaleContinuousOrNULL}
\title{Assert Is Color Palette Scale Continuous or NULL}
\usage{
assertIsColorScaleContinuousOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Color Palette Scale Continuous or NULL
}
\examples{
color <- ggplot2::scale_color_gradient(low = "red", high = "blue")
class(color)
assertIsColorScaleContinuousOrNULL(color)
assertIsColorScaleContinuousOrNULL(NULL)
}
