% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsDir}
\alias{assertIsDir}
\title{Assert Is Directory}
\usage{
assertIsDir(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Directory
}
\examples{
assertIsDir("~")
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
