% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsFillScaleContinuousOrNULL}
\alias{assertIsFillScaleContinuousOrNULL}
\title{Assert Is Fill Palette Scale Continuous or NULL}
\usage{
assertIsFillScaleContinuousOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Fill Palette Scale Continuous or NULL
}
\examples{
fill <- ggplot2::scale_fill_gradient(low = "red", high = "blue")
class(fill)
assertIsFillScaleContinuousOrNULL(fill)
assertIsFillScaleContinuousOrNULL(NULL)
}
