% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsFillScaleDiscreteOrNULL.R
\name{assertIsFillScaleDiscreteOrNULL}
\alias{assertIsFillScaleDiscreteOrNULL}
\title{Assert Is Fill Palette Scale Discrete or NULL}
\usage{
assertIsFillScaleDiscreteOrNULL(x, severity = getOption("assertive.severity",
  "stop"))
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Assert Is Fill Palette Scale Discrete or NULL
}
\examples{
fill <- ggplot2::scale_fill_manual(values = c("red", "blue"))
class(fill)
assertIsFillScaleDiscreteOrNULL(fill)
assertIsFillScaleDiscreteOrNULL(NULL)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsGFF}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
