% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsFillScaleDiscreteOrNULL}
\alias{assertIsFillScaleDiscreteOrNULL}
\title{Assert Is Fill Palette Scale Discrete or NULL}
\usage{
assertIsFillScaleDiscreteOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Fill Palette Scale Discrete or NULL
}
\examples{
fill <- ggplot2::scale_fill_manual(values = c("red", "blue"))
class(fill)
assertIsFillScaleDiscreteOrNULL(fill)
assertIsFillScaleDiscreteOrNULL(NULL)
}
