% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsGene2symbol.R
\name{assertIsGene2symbol}
\alias{assertIsGene2symbol}
\alias{assert_is_gene2symbol}
\title{Check Gene to Symbol Mapping Data}
\usage{
assertIsGene2symbol(x, severity = "stop")

assert_is_gene2symbol(x, severity = "stop")
}
\arguments{
\item{x}{\link{data.frame} containing Ensembl gene identifier to gene symbol
mappings. Must be structured as a two column \link{data.frame} with "ensgene"
and "symbol" columns.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Check Gene to Symbol Mapping Data
}
\examples{
# Success
gene2symbol <- annotable("Homo sapiens", format = "gene2symbol")
assertIsGene2symbol(gene2symbol)

# Failure
tryCatch(
    assertIsGene2symbol(mtcars),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
\keyword{internal}
