% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsGene2symbol}
\alias{assertIsGene2symbol}
\title{Assert Is Gene to Symbol Mapping Data Frame}
\usage{
assertIsGene2symbol(object)
}
\arguments{
\item{object}{\code{DataFrame} or \code{tbl_df}. Must contain "\code{geneID}" and
"\code{geneName}" columns. If \code{DataFrame}, must also have \code{rownames} set.}
}
\value{
Always stop on error.
}
\description{
Assert Is Gene to Symbol Mapping Data Frame
}
\note{
Standard \code{data.frame} class is not supported. Use either \code{DataFrame}
or \code{tbl_df} class.
}
\examples{
# DataFrame ====
object <- DataFrame(
    geneID = "ENSG00000000003",
    geneName = "TSPAN6",
    row.names = "ENSG00000000003"
)
assertIsGene2symbol(object)

# tibble ====
object <- tibble(
    geneID = "ENSG00000000003",
    geneName = "TSPAN6"
)
assertIsGene2symbol(object)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDir}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
