% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsHeaderLevel}
\alias{assertIsHeaderLevel}
\title{Assert Is Markdown Header Level}
\usage{
assertIsHeaderLevel(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Markdown supports header levels 1-7 (\code{<H1>}-\code{<H7>}).
}
\examples{
assertIsHeaderLevel(1L)
}
