% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsHexColorFunctionOrNULL}
\alias{assertIsHexColorFunctionOrNULL}
\title{Assert Is Hex Color Function or NULL}
\usage{
assertIsHexColorFunctionOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Hex Color Function or NULL
}
\examples{
hex <- function(n) {
    colors <- c("#FFFFFF", "#000000")
    colors[seq_len(n)]
}
assertIsHexColorFunctionOrNULL(hex)
assertIsHexColorFunctionOrNULL(NULL)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2Symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDir}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsImplicitInteger}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
