% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsHexColorFunctionOrNULL.R
\name{assertIsHexColorFunctionOrNULL}
\alias{assertIsHexColorFunctionOrNULL}
\title{Assert Is Hex Color Function or NULL}
\usage{
assertIsHexColorFunctionOrNULL(x, severity = getOption("assertive.severity",
  "stop"))
}
\arguments{
\item{x}{Object.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Assert Is Hex Color Function or NULL
}
\examples{
hex <- function(n) {
    colors <- c("#FFFFFF", "#000000")
    colors[seq_len(n)]
}
assertIsHexColorFunctionOrNULL(hex)
assertIsHexColorFunctionOrNULL(NULL)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGFF}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsTx2gene}}, \code{\link{assertIsURL}}
}
\author{
Michael Steinbaugh
}
