% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertIsHexColorFunctionOrNULL}
\alias{assertIsHexColorFunctionOrNULL}
\title{Assert Is Hex Color Function or NULL}
\usage{
assertIsHexColorFunctionOrNULL(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Hex Color Function or NULL
}
\examples{
hex <- function(n) {
    colors <- c("#FFFFFF", "#000000")
    colors[seq_len(n)]
}
assertIsHexColorFunctionOrNULL(hex)
assertIsHexColorFunctionOrNULL(NULL)
}
