% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsImplicitInteger}
\alias{assertIsImplicitInteger}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsImplicitIntegerOrNULL}
\alias{isAnImplicitInteger}
\alias{isImplicitInteger}
\title{Assert Is Implicit Integer}
\usage{
assertIsAnImplicitInteger(object)

assertIsAnImplicitIntegerOrNULL(object)

assertIsImplicitInteger(object)

assertIsImplicitIntegerOrNULL(object)

isAnImplicitInteger(object)

isImplicitInteger(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Always stop on error.
}
\description{
Assert Is Implicit Integer
}
\examples{
assertIsAnImplicitInteger(1)
assertIsAnImplicitIntegerOrNULL(1)
assertIsAnImplicitIntegerOrNULL(NULL)
assertIsImplicitInteger(c(1, 2))
assertIsImplicitIntegerOrNULL(c(1, 2))
assertIsImplicitIntegerOrNULL(NULL)
isAnImplicitInteger(1)
isImplicitInteger(list(1, 1L, 1.1, "XXX"))
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreURL}},
  \code{\link{assertAllAreUniqueGeneNames}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAlpha}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHeaderLevel}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsTx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
