% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsImplicitInteger.R
\name{assertIsImplicitInteger}
\alias{assertIsImplicitInteger}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsImplicitInteger}
\alias{assertIsImplicitIntegerOrNULL}
\alias{isImplicitInteger}
\alias{assert_is_an_implicit_integer}
\alias{assert_is_an_implicit_integer_or_null}
\alias{assert_is_implicit_integer}
\alias{assert_is_implicit_integer_or_null}
\alias{is_implicit_integer}
\title{Implicit Integer Assert Check}
\usage{
assertIsAnImplicitInteger(x)

assertIsAnImplicitIntegerOrNULL(x)

assertIsImplicitInteger(x)

assertIsImplicitIntegerOrNULL(x)

isImplicitInteger(x)

assert_is_an_implicit_integer(x)

assert_is_an_implicit_integer_or_null(x)

assert_is_implicit_integer(x)

assert_is_implicit_integer_or_null(x)

is_implicit_integer(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Abort on mismatch.
}
\description{
Implicit Integer Assert Check
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
