% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsImplicitInteger}
\alias{assertIsImplicitInteger}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsImplicitIntegerOrNULL}
\alias{isAnImplicitInteger}
\alias{isImplicitInteger}
\title{Assert Is Implicit Integer}
\usage{
assertIsAnImplicitInteger(x)

assertIsAnImplicitIntegerOrNULL(x)

assertIsImplicitInteger(x)

assertIsImplicitIntegerOrNULL(x)

isAnImplicitInteger(x)

isImplicitInteger(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Stop on mismatch.
}
\description{
Assert Is Implicit Integer
}
\examples{
assertIsAnImplicitInteger(1)
assertIsAnImplicitIntegerOrNULL(1)
assertIsAnImplicitIntegerOrNULL(NULL)
assertIsImplicitInteger(c(1, 2))
assertIsImplicitIntegerOrNULL(c(1, 2))
assertIsImplicitIntegerOrNULL(NULL)
isAnImplicitInteger(1)
isImplicitInteger(list(1, 1L, 1.1, "XXX"))
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAllAreValidNames}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsTx2gene}}, \code{\link{assertIsURL}}
}
\author{
Michael Steinbaugh
}
\concept{Assert Check Functions}
