% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsImplicitInteger.R
\name{assertIsImplicitInteger}
\alias{assertIsImplicitInteger}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsImplicitInteger}
\alias{assertIsImplicitIntegerOrNULL}
\alias{isImplicitInteger}
\title{Assert Is Implicit Integer}
\usage{
assertIsAnImplicitInteger(x)

assertIsAnImplicitIntegerOrNULL(x)

assertIsImplicitInteger(x)

assertIsImplicitIntegerOrNULL(x)

isImplicitInteger(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Abort on mismatch.
}
\description{
Assert Is Implicit Integer
}
\examples{
assertIsAnImplicitInteger(1)
assertIsAnImplicitIntegerOrNULL(1)
assertIsAnImplicitIntegerOrNULL(NULL)
assertIsImplicitInteger(c(1, 2))
assertIsImplicitIntegerOrNULL(c(1, 2))
assertIsImplicitIntegerOrNULL(NULL)
isImplicitInteger(1)
# Also returns `TRUE` for explicit integers
isImplicitInteger(1L)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsTx2gene}}, \code{\link{assert}}
}
