% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsTx2gene.R
\name{assertIsTx2gene}
\alias{assertIsTx2gene}
\title{Check Transcript to Gene Mapping Data}
\usage{
assertIsTx2gene(x, severity = "stop")
}
\arguments{
\item{x}{\code{data.frame} containing Ensembl transcript to gene identifier
mappings. Must be structured as a two column \code{data.frame} with "txID" and
"geneID" columns.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Check Transcript to Gene Mapping Data
}
\examples{
tx2gene <- tx2gene("Homo sapiens", genomeBuild = "GRCh37")
assertIsTx2gene(tx2gene)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assert}}
}
