% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertIsTx2gene}
\alias{assertIsTx2gene}
\title{Assert Is Transcript-to-Gene Mapping Data}
\usage{
assertIsTx2gene(x, severity = getOption("assertive.severity", "stop"))
}
\arguments{
\item{x}{\code{data.frame} containing Ensembl transcript to gene identifier
mappings. Must be structured as a two column \code{data.frame} with
"transcriptID" and "geneID" columns.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Stop on mismatch.
}
\description{
Assert Is Transcript-to-Gene Mapping Data
}
\examples{
x <- data.frame(
    transcriptID = "ENST00000000233",
    geneID = "ENSG00000004059"
)
assertIsTx2gene(x)
}
\seealso{
Other Assert Check Functions: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertAreGeneAnnotations}},
  \code{\link{assertAreTranscriptAnnotations}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalInterestingGroups}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsAHeaderLevel}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsColorScaleContinuousOrNULL}},
  \code{\link{assertIsColorScaleDiscreteOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsFillScaleContinuousOrNULL}},
  \code{\link{assertIsFillScaleDiscreteOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsHexColorFunctionOrNULL}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assertIsURL}}
}
\author{
Michael Steinbaugh
}
