% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsTx2gene.R
\name{assertIsTx2gene}
\alias{assertIsTx2gene}
\alias{assert_is_tx2gene}
\title{Check Transcript to Gene Mapping Data}
\usage{
assertIsTx2gene(x, severity = "stop")

assert_is_tx2gene(x, severity = "stop")
}
\arguments{
\item{x}{\link{data.frame} containing Ensembl transcript to gene identifier
mappings. Must be structured as a two column \link{data.frame} with "enstxp" and
"ensgene" columns.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Abort on mismatch.
}
\description{
Check Transcript to Gene Mapping Data
}
\examples{
# Success
tx2gene <- tx2gene("Homo sapiens")
assertIsTx2gene(tx2gene)

# Failure
tryCatch(
    assertIsTx2gene(mtcars),
    error = function(e) e)
}
\seealso{
Other Assert Checks: \code{\link{assertAllAreNonExisting}},
  \code{\link{assertFormalAnnotationCol}},
  \code{\link{assertFormalColorFunction}},
  \code{\link{assertFormalCompress}},
  \code{\link{assertFormalGene2symbol}},
  \code{\link{assertFormalHeaderLevel}},
  \code{\link{assertHasRownames}},
  \code{\link{assertIsANumberOrNULL}},
  \code{\link{assertIsAStringOrNULL}},
  \code{\link{assertIsAnIntegerOrNULL}},
  \code{\link{assertIsAnnotable}},
  \code{\link{assertIsCharacterOrNULL}},
  \code{\link{assertIsDataFrameOrNULL}},
  \code{\link{assertIsGene2symbol}},
  \code{\link{assertIsImplicitInteger}},
  \code{\link{assert}}
}
