% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assignMultiple}
\alias{assignMultiple}
\title{Assign multiple variables}
\usage{
assignMultiple(..., envir = parent.frame(), prefix = NULL)
}
\arguments{
\item{...}{List of variables to assign.}

\item{envir}{Desired environment (optional).}

\item{prefix}{Name prefix (optional).}
}
\description{
Optionally, you can specify a name prefix and the desired environment.
}
