% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assignAsNewEnv}
\alias{assignAsNewEnv}
\title{Assign multiple objects to a new environment}
\usage{
assignAsNewEnv(..., envir_name)
}
\arguments{
\item{...}{List of objects to assign.}

\item{envir_name}{Environment name.}
}
\value{
Object names.
}
\description{
Assign multiple objects to a new environment
}
\examples{
assignAsNewEnv(mtcars, starwars, envir_name = "testenv")
}
