% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignAndSaveData.R
\name{assignAndSaveData}
\alias{assignAndSaveData}
\title{Assign and Save Data}
\usage{
assignAndSaveData(name, object, dir = "data", compress = TRUE)
}
\arguments{
\item{name}{Desired variable name.}

\item{object}{Object.}

\item{dir}{Output directory. Defaults to \strong{data}.}

\item{compress}{Logical or character string specifying compression format.
\code{TRUE} corresponds to gzip compression (\strong{default}), and character strings
\code{gzip}, \code{bzip2} or \code{xz} specify the type of compression.}
}
\value{
Assigned object name as a string.
}
\description{
Assigns a new object by name to the current working environment then saves
the newly assigned object, specified by \code{dir}.
}
\examples{
assignAndSaveData("test", mtcars)
}
\seealso{
Other Object Assignment Utilities: \code{\link{multiassignAsNewEnv}}
}
\concept{Object Assignment Utilities}
