% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignAndSaveData.R
\name{assignAndSaveData}
\alias{assignAndSaveData}
\title{Assign and Save Data}
\usage{
assignAndSaveData(name, object, dir = "data", compress = TRUE,
  envir = parent.frame(), quiet = TRUE)
}
\arguments{
\item{name}{Desired variable name.}

\item{object}{Object.}

\item{dir}{Output directory. Defaults to \strong{data}.}

\item{compress}{logical or character string specifying whether saving
    to a named file is to use compression.  \code{TRUE} corresponds to
    \command{gzip} compression, and character strings \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} specify the type of
    compression.  Ignored when \code{file} is a connection and
    for workspace format version 1.}

\item{envir}{Environment to use for assignment. Defaults to \code{\link[=parent.frame]{parent.frame()}},
the calling environment.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Silent named character vector of file path.
}
\description{
Assigns a new object by name to the current working environment then saves
the newly assigned object, specified by \code{dir}.
}
\note{
This function attempts to follow the same order as \code{\link[base:assign]{base::assign()}}.
}
\examples{
assignAndSaveData("test", mtcars)
}
\seealso{
Other Object Assignment Utilities: \code{\link{multiassignAsNewEnvir}}
}
\concept{Object Assignment Utilities}
