% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{assignAndSaveData}
\alias{assignAndSaveData}
\title{Assign and Save Data}
\usage{
assignAndSaveData(name, object, envir = parent.frame(),
  dir = getOption("basejump.save.dir", "."),
  ext = getOption("basejump.save.ext", "rds"),
  overwrite = getOption("basejump.save.overwrite", TRUE),
  compress = getOption("basejump.save.compress", TRUE))
}
\arguments{
\item{name}{\code{string}. Desired variable name.}

\item{object}{Object.}

\item{envir}{\code{environment}. Environment to use for assignment. Defaults to
\code{\link[=parent.frame]{parent.frame()}}, the calling environment.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}

\item{ext}{\code{string}. R data serialized (RDS; "\code{rds}") or R data ("\code{rda}",
"\code{RData}"). RDS is preferred when saving single objects per file, which is
always the convention of \code{\link[=saveData]{saveData()}}, regardless of the extension used.}

\item{overwrite}{\code{boolean}. Overwrite existing file.}

\item{compress}{logical or character string specifying whether saving
    to a named file is to use compression.  \code{TRUE} corresponds to
    \command{gzip} compression, and character strings \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} specify the type of
    compression.  Ignored when \code{file} is a connection and
    for workspace format version 1.}
}
\value{
Invisible named \code{string}. File path.
}
\description{
Assigns a new object by name to the current working environment then saves
the newly assigned object, specified by the "\code{dir}" argument.
}
\note{
This function attempts to follow the same order as \code{\link[base:assign]{base::assign()}}.
}
\examples{
data(rse_small)
assignAndSaveData(name = "example", object = rse_small, ext = "rds")
exists("example", inherits = FALSE)
file.exists("example.rds")

## Clean up.
rm(example)
unlink("example.rds")
}
\seealso{
Other Import/Export Functions: \code{\link{export}},
  \code{\link{import}}, \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{saveData}}, \code{\link{transmit}},
  \code{\link{writeCounts}}
}
\concept{Import/Export Functions}
