% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignAndSaveData-S3.R
\name{assignAndSaveData}
\alias{assignAndSaveData}
\title{Assign and Save Data}
\usage{
assignAndSaveData(name, object, dir = "data", compress = TRUE,
  env = parent.frame())
}
\arguments{
\item{name}{Desired variable name.}

\item{object}{Object.}

\item{dir}{Output directory. Defaults to \strong{data}.}

\item{compress}{Compression method, supporting \code{xz} (\strong{preferred}), \code{bzip2},
or \code{gzip}. Compression can be disabled by setting as \code{FALSE}, although this
is not generally recommended.}

\item{env}{Destination environment.}
}
\value{
No value.
}
\description{
Assigns a new object by name to the current working environment then saves
the newly assigned object, specified by \code{dir}.
}
\examples{
assignAndSaveData("test", mtcars)
}
\seealso{
Other Object Assignment Utilities: \code{\link{assignIntoNewEnv}}
}
