% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignAndSaveData.R
\name{assignAndSaveData}
\alias{assignAndSaveData}
\title{Assign and Save Data}
\usage{
assignAndSaveData(name, object, dir = ".", compress = TRUE,
  envir = parent.frame())
}
\arguments{
\item{name}{\code{string}. Desired variable name.}

\item{object}{Object.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}

\item{compress}{logical or character string specifying whether saving
    to a named file is to use compression.  \code{TRUE} corresponds to
    \command{gzip} compression, and character strings \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} specify the type of
    compression.  Ignored when \code{file} is a connection and
    for workspace format version 1.}

\item{envir}{\code{environment}. Environment to use for assignment. Defaults to
\code{\link[=parent.frame]{parent.frame()}}, the calling environment.}
}
\value{
Invisible named \code{character} containing file paths.
}
\description{
Assigns a new object by name to the current working environment then saves
the newly assigned object, specified by the "\code{dir}" argument.
}
\note{
This function attempts to follow the same order as \code{\link[base:assign]{base::assign()}}.
}
\examples{
assignAndSaveData(name = "example", object = rnaseq_counts)
exists("example", inherits = FALSE)
file.exists("example.rda")

# Clean up
rm(example)
unlink("example.rda")
}
\seealso{
Other Write Functions: \code{\link{initializeDirectory}},
  \code{\link{saveData}}, \code{\link{transmit}},
  \code{\link{writeCounts}}
}
\author{
Michael Steinbaugh
}
\concept{Write Functions}
