% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/zzz.R
\name{assignAsNewEnv}
\alias{assignAsNewEnv}
\alias{assign_as_new_env}
\alias{assign_and_save_data}
\title{Assign Multiple Objects to a New Environment}
\usage{
assignAsNewEnv(..., envirName)
}
\arguments{
\item{...}{Dot objects to assign.}

\item{envirName}{Environment name.}
}
\value{
Object names.
}
\description{
Assign Multiple Objects to a New Environment
}
\examples{
assignAsNewEnv(mtcars, starwars, envirName = "testenv")
}
