% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassignAsNewEnv.R
\name{multiassignAsNewEnv}
\alias{multiassignAsNewEnv}
\title{Assign Multiple Objects as a New Environment}
\usage{
multiassignAsNewEnv(..., newEnv, parentEnv = parent.frame())
}
\arguments{
\item{...}{Objects as dots.}

\item{newEnv}{New environment name (string).}

\item{parentEnv}{Parent environment of new environment.}
}
\value{
Character vector of assigned objects in the new environment.
}
\description{
Assign Multiple Objects as a New Environment
}
\examples{
multiassignAsNewEnv(mtcars, starwars, newEnv = "test")
}
\seealso{
\code{\link[Biobase:multiassign]{Biobase::multiassign()}}.

Other Object Assignment Utilities: \code{\link{assignAndSaveData}}
}
\concept{Object Assignment Utilities}
