% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignIntoNewEnv.R
\name{assignIntoNewEnv}
\alias{assignIntoNewEnv}
\title{Assign Multiple Objects to a New Environment}
\usage{
assignIntoNewEnv(..., envName, parentEnv = parent.frame())
}
\arguments{
\item{...}{Objects as dots.}

\item{envName}{New environment name.}

\item{parentEnv}{Parent environment of new environment.}
}
\value{
No value.
}
\description{
Assign Multiple Objects to a New Environment
}
\examples{
assignIntoNewEnv(mtcars, starwars, envName = "test")
}
\seealso{
\code{\link[Biobase:multiassign]{Biobase::multiassign()}}.

Other Object Assignment Utilities: \code{\link{assignAndSaveData}}
}
