% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/autopadZeros-methods.R
\docType{methods}
\name{autopadZeros}
\alias{autopadZeros}
\alias{autopadZeros,character-method}
\alias{autopadZeros,matrix-method}
\alias{autopadZeros,SummarizedExperiment-method}
\title{Autopad Zeros in Vector}
\usage{
autopadZeros(object, ...)

\S4method{autopadZeros}{character}(object)

\S4method{autopadZeros}{matrix}(object, rownames = FALSE,
  colnames = TRUE, sort = TRUE)

\S4method{autopadZeros}{SummarizedExperiment}(object, rownames = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{rownames}{\code{boolean}. Apply sanitization on row names. This is not
generally recommended by default, since rownames commonly contain gene
identifiers that should not be modified.}

\item{colnames}{\code{boolean}. Apply sanitization on column names. This is
generally recommended by default.}

\item{sort}{\code{boolean}. Resort using \code{sort()}.}
}
\value{
\code{character}.
}
\description{
Autopad Zeros in Vector
}
\note{
For methods on objects supporting \code{dim()} (e.g. \code{matrix} or
\code{SummarizedExperiment}), the object will be returned with the rows and/or
columns resorted by default. This does not apply to the \code{character} method.
}
\examples{
data(rse)

## character ====
autopadZeros(c("A1", "B10"))
autopadZeros(c("A1", "B10", "C100"))

## SummarizedExperiment ====
autopadZeros(rse)
}
