% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopadZeros-methods.R
\name{autopadZeros}
\alias{autopadZeros}
\alias{autopadZeros,matrix-method}
\alias{autopadZeros,SummarizedExperiment-method}
\title{Autopad zeros}
\usage{
autopadZeros(object, ...)

\S4method{autopadZeros}{matrix}(object, rownames = FALSE, colnames = TRUE, sort = TRUE)

\S4method{autopadZeros}{SummarizedExperiment}(object, rownames = FALSE, colnames = TRUE, sort = TRUE)
}
\arguments{
\item{object}{Object.}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{sort}{\code{logical(1)}.
Resort using \code{sort}.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Autopad zeros
}
\note{
For methods on objects supporting \code{\link[base:dim]{dim()}} (e.g. \code{matrix}),
the object will be returned with the rows and/or columns resorted by default.
This does not apply to the \code{character} method defined in syntactic.

Updated 2020-06-15.
}
\section{SummarizedExperiment sample names}{


If \code{sampleName} column is defined in
\code{\link[SummarizedExperiment:colData]{colData()}}, these values will also get
padded, if necessary. This improves # downstream handling in functions that
rely on this feature.
}

\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
autopadZeros(object, rownames = TRUE, colnames = TRUE)

## SingleCellExperiment ====
object <- SingleCellExperiment
autopadZeros(object, rownames = TRUE, colnames = TRUE)
}
