% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopadZeros.R
\name{autopadZeros}
\alias{autopadZeros}
\title{Autopad Zeros in Vector}
\usage{
autopadZeros(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character}.
}
\description{
Autopad Zeros in Vector
}
\examples{
autopadZeros(c("A1", "B10"))
autopadZeros(c("A1", "B10", "C100"))
}
