% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopadZeros-methods.R
\docType{methods}
\name{autopadZeros}
\alias{autopadZeros}
\alias{autopadZeros,character-method}
\alias{autopadZeros,matrix-method}
\alias{autopadZeros,SummarizedExperiment-method}
\title{Autopad zeros}
\usage{
autopadZeros(object, ...)

\S4method{autopadZeros}{character}(object)

\S4method{autopadZeros}{matrix}(object, rownames = FALSE,
  colnames = TRUE, sort = TRUE)

\S4method{autopadZeros}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE, sort = TRUE)
}
\arguments{
\item{object}{Object.}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{sort}{\code{logical(1)}.
Resort using \code{sort}.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Autopad zeros
}
\note{
For methods on objects supporting \code{\link[base:dim]{dim()}} (e.g. \code{matrix}),
the object will be returned with the rows and/or columns resorted by default.
This does not apply to the \code{character} method.

Updated 2019-07-28.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)
rse <- RangedSummarizedExperiment
sce <- SingleCellExperiment

## character ====
autopadZeros(c("A1", "B10"))
autopadZeros(c("A1", "B10", "C100"))

## SummarizedExperiment ====
autopadZeros(rse, rownames = TRUE, colnames = TRUE)

## SingleCellExperiment ====
autopadZeros(sce, rownames = TRUE, colnames = TRUE)
}
