% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{barcodePattern}
\alias{barcodePattern}
\title{Single-Cell Barcode Pattern
Trailing number is to match cellranger output.}
\format{An object of class \code{character} of length 1.}
\usage{
barcodePattern
}
\description{
Single-Cell Barcode Pattern
Trailing number is to match cellranger output.
}
\examples{
barcodePattern
}
\keyword{datasets}
