% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{barcodePattern}
\alias{barcodePattern}
\title{Single-sell barcode pattern}
\format{
An object of class \code{character} of length 1.
}
\usage{
barcodePattern
}
\description{
Trailing number is to match cellranger output.
}
\note{
Updated 2019-08-21.
}
\examples{
barcodePattern
}
\keyword{datasets}
