% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\title{Barcode Ranks per Sample}
\usage{
barcodeRanksPerSample(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
}
\description{
Barcode Ranks per Sample
}
\examples{
isS4(barcodeRanksPerSample)
}
\seealso{
\itemize{
\item DropletUtils.
\item SingleCellExperiment.
\item bcbioSingleCell.
\item Chromium.
}
}
