% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump-package.R
\docType{package}
\name{basejump-package}
\alias{basejump}
\alias{basejump-package}
\title{basejump}
\description{
Base functions for bioinformatics and R package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://steinbaugh.com/basejump}
  \item \url{https://github.com/steinbaugh/basejump}
  \item Report bugs at \url{https://github.com/steinbaugh/basejump/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Rory Kirchner \email{roryk@alum.mit.edu} (0000-0003-4814-5885) [contributor]
  \item Mary Piper \email{mary.piper@gmail.com} (0000-0003-2699-3840) [contributor]
}

}
