% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\docType{package}
\name{basejump-package}
\alias{basejump}
\alias{basejump-package}
\title{basejump}
\description{
Base functions for bioinformatics and R package development.
}
\details{
The package exports \code{\link[=camel]{camel()}} and \code{\link[=snake]{snake()}} variants of all functions,
for seamless integration into \href{https://www.bioconductor.org}{Bioconductor}
and \href{http://tidyverse.org}{tidyverse} workflows. Function parameters are
always passed in \code{\link[=camel]{camel()}} case. We also provide British spelling variants
(e.g. colour) of the function names wherever applicable.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://steinbaugh.com/basejump}
  \item \url{https://github.com/steinbaugh/basejump}
  \item Report bugs at \url{https://github.com/steinbaugh/basejump/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com}

}
