% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{basejumpCacheURL}
\alias{basejumpCacheURL}
\title{Cache URL}
\format{An object of class \code{character} of length 1.}
\usage{
basejumpCacheURL
}
\description{
Cache URL
}
\examples{
basejumpCacheURL
}
\keyword{internal}
