% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{basejumpTestsURL}
\alias{basejumpTestsURL}
\title{basejump test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
basejumpTestsURL
}
\description{
basejump test data URL
}
\note{
Updated 2020-06-09.
}
\examples{
basejumpTestsURL
}
\keyword{internal}
