% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioProject.R
\name{bcbioProject}
\alias{bcbioProject}
\title{Load a remote bcbio project}
\usage{
bcbioProject(workflow, mountDir = file.path("~", "Orchestra", "bcbio", "PIs",
  basename(dirname(getwd())), basename(getwd())), subdirCreate = TRUE)
}
\arguments{
\item{workflow}{bcbio workflow template}

\item{mountDir}{SSH mount directory. Defaults to \code{bcbio} group share.}

\item{subdirCreate}{Create subdirectories}
}
\value{
List with directory paths
}
\description{
We recommend loading the \code{mountDir} as a connection over \code{sshfs}.
}
\examples{
\dontrun{
bcbioProject(workflow = "illumina_rnaseq",
             subdirCreate = FALSE)
}
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{hpc}
\keyword{sshfs}
