% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioProject.R
\name{bcbioProject}
\alias{bcbioProject}
\title{Load a remote bcbio project}
\usage{
bcbioProject(researcher, project, workflow, mountDir = file.path("~",
  "Orchestra", "bcbio", "PIs"), subdirCreate = TRUE)
}
\arguments{
\item{researcher}{Researcher (primary investigator)}

\item{project}{Project name}

\item{workflow}{bcbio workflow template}

\item{mountDir}{SSH mount directory. Defaults to \code{bcbio} group share.}

\item{subdirCreate}{Create subdirectories}
}
\value{
List with directory paths
}
\description{
We recommend loading the \code{mountDir} as a connection over \code{sshfs}.
}
\examples{
\dontrun{
bcbioProject(researcher = "joe_smith",
             project = "gene_ko_rnaseq",
             workflow = "illumina_rnaseq",
             subdirCreate = FALSE)
}
}
