% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSailfish.R
\name{bcbioSailfish}
\alias{bcbioSailfish}
\title{Sailfish data}
\usage{
bcbioSailfish(project, summary)
}
\arguments{
\item{project}{bcbio project}

\item{summary}{Summary data frame}
}
\value{
txi \code{tximport} count data generated by \code{sailfish}
}
\description{
Import sailfish data from \code{bcbio-rnaseq} run
}
\examples{
\dontrun{
bcbioSailfish(project, summary)
}
}
\author{
Michael Steinbaugh
}
