% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSummary.R
\name{bcbioSummary}
\alias{bcbioSummary}
\title{Import a bcbio project summary}
\usage{
bcbioSummary(project, metadata)
}
\arguments{
\item{project}{bcbio project}

\item{metadata}{bcbio project metadata}
}
\description{
Import a bcbio project summary
}
\examples{
\dontrun{
bcbioSummary(project)
}
}
\author{
Michael Steinbaugh
}
