% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{bioc}
\alias{bioc}
\title{Install or update Bioconductor packages}
\usage{
bioc(pkgs = NULL, ...)
}
\arguments{
\item{pkgs}{Character vector of package names to install or update. Supports
Bioconductor, CRAN, and GitHub repositories (e.g. \code{steinbaugh/basejump}
format).}

\item{...}{Passthrough parameters.}
}
\description{
This function is a wrapper for \code{\link[=biocLite]{biocLite()}}. It first sources the code from
the Bioconductor website then calls \code{\link[=biocLite]{biocLite()}}.
}
\examples{
\dontrun{
# Load Bioconductor and update packages
bioc()

# Install `limma`
bioc("limma")
}
}
\seealso{
This function is a modified version of \code{rafalib::install_bioc()}.
It calls \code{\link[BiocInstaller:biocLite]{BiocInstaller::biocLite()}} internally.
}
