% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{biocLite}
\alias{biocLite}
\title{Install or Update Bioconductor Packages}
\usage{
biocLite(pkgs = NULL, ...)
}
\arguments{
\item{pkgs}{Character vector of package names.}

\item{...}{Additional arguments, passed to \code{\link[BiocInstaller:biocLite]{BiocInstaller::biocLite()}}.}
}
\description{
This function is a convenience wrapper for \code{\link[BiocInstaller:biocLite]{BiocInstaller::biocLite()}}. It
sources the required \href{https://bioconductor.org/biocLite.R}{biocLite.R} file
from the Bioconductor website automatically then calls \code{\link[=biocLite]{biocLite()}}.
}
\details{
Supports \href{https://bioconductor.org}{Bioconductor},
\href{https://cran.r-project.org}{CRAN}, and \href{https://github.com}{GitHub}
repositories.
}
\examples{
\dontrun{
# Update all packages
biocLite()

# Install packages
biocLite("limma")            # Bioconductor
biocLite("dplyr")            # CRAN
biocLite("tidyverse/dplyr")  # GitHub
}
}
\seealso{
Modified variant of \code{rafalib::install_bioc()}.
}
\keyword{internal}
