% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{biocLite}
\alias{biocLite}
\title{Install or update Bioconductor packages}
\usage{
biocLite(pkgs = NULL, ...)
}
\arguments{
\item{pkgs}{Character vector of package names to install or update. Supports
Bioconductor, CRAN, and GitHub repositories.}

\item{...}{Passthrough parameters.}
}
\description{
This function is a convenience wrapper for \code{\link[BiocInstaller:biocLite]{BiocInstaller::biocLite()}}. It
sources the required \code{biocLite.R} file from the Bioconductor website
automatically then calls \code{\link[=biocLite]{biocLite()}}.
}
\examples{
\dontrun{
# Load Bioconductor and update packages
biocLite()

# Install packages
biocLite("limma")            # Bioconductor
biocLite("dplyr")            # CRAN
biocLite("tidyverse/dplyr")  # GitHub
}
}
\seealso{
This function is a modified version of \code{rafalib::install_bioc()}.
}
