% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-broadClass.R
\docType{methods}
\name{broadClass}
\alias{broadClass}
\alias{broadClass}
\alias{broadClass,DataFrame-method}
\alias{broadClass,data.frame-method}
\alias{broadClass,GRanges-method}
\title{Broad Class Definitions}
\usage{
broadClass(object, ...)

\S4method{broadClass}{DataFrame}(object)

\S4method{broadClass}{data.frame}(object)

\S4method{broadClass}{GRanges}(object)
}
\arguments{
\item{object}{Object that can be coerced to \code{data.frame}, containing gene or
transcript annotations.}
}
\value{
Named \code{factor} containing broad class definitions.
}
\description{
Broad Class Definitions
}
\examples{
# GRanges ====
# Ensembl
x <- makeGRangesFromEnsembl("Homo sapiens")
broadClass(x)

# data.frame ====
x <- annotable("Homo sapiens")
broadClass(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}},
  \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Rory Kirchner and Michael Steinbaugh
}
