% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/broadClass-methods.R
\docType{methods}
\name{broadClass}
\alias{broadClass}
\alias{broadClass,GRanges-method}
\alias{broadClass,SummarizedExperiment-method}
\title{Broad Class Definitions}
\usage{
broadClass(object, ...)

\S4method{broadClass}{GRanges}(object)

\S4method{broadClass}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object that can be coerced to \code{DataFrame}, containing gene or
transcript annotations.}

\item{...}{Additional arguments.}
}
\value{
Named \code{factor} containing broad class definitions.
}
\description{
Broad Class Definitions
}
\examples{
# GRanges ====
x <- broadClass(makeGRangesFromEnsembl("Homo sapiens"))
table(x)

# SummarizedExperiment ====
x <- broadClass(rse_small)
table(x)
}
\seealso{
Other Annotation Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2gene}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Rory Kirchner and Michael Steinbaugh
}
\concept{Annotation Functions}
