% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetrics-methods.R
\docType{methods}
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,matrix-method}
\alias{calculateMetrics,Matrix-method}
\alias{calculateMetrics,DelayedArray-method}
\alias{calculateMetrics,RangedSummarizedExperiment-method}
\alias{calculateMetrics,SingleCellExperiment-method}
\title{Calculate quality control metrics}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{matrix}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{calculateMetrics}{Matrix}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{calculateMetrics}{DelayedArray}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{calculateMetrics}{RangedSummarizedExperiment}(object,
  prefilter = FALSE)

\S4method{calculateMetrics}{SingleCellExperiment}(object,
  prefilter = FALSE)
}
\arguments{
\item{object}{Object.}

\item{rowRanges}{\code{GRanges} or \code{GRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{prefilter}{\code{logical(1)}.
Drop very low quality samples/cells from the object.
This can resize the number of columns but the rows (i.e. features) do not
change with this operation.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{matrix} / \code{Matrix} / \code{DelayedArray}: \code{DataFrame} containing metrics.
\item \code{SingleCellExperiment} / \code{SummarizedExperiment}: Modified object, with
metrics in \code{\link[SummarizedExperiment:colData]{colData()}}.
}
}
\description{
Calculate quality control metrics
}
\note{
Input a raw count matrix. Do not use size factor adjusted or log
normalized counts here.

Updated 2019-08-27.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
x <- calculateMetrics(object)
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
