% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calculateMetrics-methods.R
\docType{methods}
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,DelayedArray-method}
\alias{calculateMetrics,Matrix-method}
\alias{calculateMetrics,SingleCellExperiment-method}
\title{Calculate cellular barcode quality metrics}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{DelayedArray}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{calculateMetrics}{Matrix}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{calculateMetrics}{SingleCellExperiment}(object,
  prefilter = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{rowRanges}{\code{GRanges} or \code{GRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{prefilter}{\code{logical(1)}.
Apply prefiltering. Remove zero count genes.}
}
\value{
\code{DataFrame}.
}
\description{
Calculate cellular barcode quality metrics
}
\note{
Input a raw count matrix. Do not use size factor adjusted or log
normalized counts here.

Updated 2019-08-07.
}
\examples{
data(SingleCellExperiment, package = "acidtest")
counts <- counts(indrops)
rowRanges <- rowRanges(indrops)

# Metrics using genome annotations (recommended).
x <- calculateMetrics(counts, rowRanges = rowRanges)
head(x)

# Minimal metrics (supported, but not recommended).
x <- calculateMetrics(counts, rowRanges = NULL)
head(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
