% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel.R
\name{camel}
\alias{camel}
\title{camelCase}
\usage{
camel(string)
}
\arguments{
\item{string}{\code{string}}
}
\value{
\code{string} with camelCase formatting
}
\description{
camelCase
}
\examples{
camel("RNAi clone")
}
\keyword{general}

