% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelCase-methods.R
\name{camelCase}
\alias{camelCase}
\alias{camelCase,atomic-method}
\alias{camelCase,factor-method}
\alias{camelCase,list-method}
\alias{camelCase,matrix-method}
\alias{camelCase,data.frame-method}
\alias{camelCase,Vector-method}
\alias{camelCase,DataTable-method}
\alias{camelCase,Ranges-method}
\alias{camelCase,Matrix-method}
\alias{camelCase,SummarizedExperiment-method}
\title{Camel case}
\usage{
camelCase(object, ...)

\S4method{camelCase}{atomic}(object, names = TRUE, strict = FALSE)

\S4method{camelCase}{factor}(object, names = TRUE, strict = FALSE)

\S4method{camelCase}{list}(object, names = TRUE, strict = FALSE)

\S4method{camelCase}{matrix}(object, rownames = FALSE, colnames = TRUE, strict = FALSE)

\S4method{camelCase}{data.frame}(object, rownames = FALSE, colnames = TRUE, strict = FALSE)

\S4method{camelCase}{Vector}(object, names = TRUE, mcols = TRUE, metadata = TRUE, strict = FALSE)

\S4method{camelCase}{DataTable}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  mcols = TRUE,
  metadata = TRUE,
  strict = FALSE
)

\S4method{camelCase}{Ranges}(object, names = FALSE, mcols = TRUE, metadata = TRUE, strict = FALSE)

\S4method{camelCase}{Matrix}(object, rownames = FALSE, colnames = TRUE, strict = FALSE)

\S4method{camelCase}{SummarizedExperiment}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  assayNames = TRUE,
  rowData = TRUE,
  colData = TRUE,
  metadata = TRUE,
  strict = FALSE
)
}
\arguments{
\item{object}{Object.}

\item{names}{\code{logical(1)}.
Sanitize names.

Only applies to string mode (\code{rename = FALSE}).}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{mcols}{\code{logical(1)}.
Sanitize names of metadata columns (i.e. \code{DataFrame}).}

\item{metadata}{\code{logical(1)}.
Sanitize metadata names.}

\item{assayNames}{\code{logical(1)}.
Sanitize assay names.}

\item{rowData}{\code{logical(1)}.
Sanitize the row data names.}

\item{colData}{\code{logical(1)}.
Sanitize column names of column data.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
Returns invisible modified file path when \code{rename = TRUE}.
Returns invisible \code{NULL} when \code{recursive = TRUE}.
}
\description{
Format character strings to use (lower) camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{thisIsCamelCase}).
}
\details{
Camel case is recommended by Bioconductor for variable and function names.
}
\note{
Updated 2019-09-25.
}
\examples{
data(syntactic, package = "acidtest")
lapply(syntactic, camelCase)
}
