% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstCase.R, R/titleCase.R
\name{firstCase}
\alias{firstCase}
\alias{titleCase}
\title{Change case of a character vector}
\usage{
firstCase(character)

titleCase(character)
}
\arguments{
\item{character}{Character vector}
}
\value{
Reformatted character vector
}
\description{
Change case of a character vector
}
\examples{
firstCase("RNA sequencing")
titleCase("RNA sequencing")
}
\author{
Michael Steinbaugh
}
