% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{titleCase}
\alias{titleCase}
\alias{firstCase}
\title{Change case of a character vector}
\usage{
titleCase(character)

firstCase(character)
}
\arguments{
\item{character}{Character vector}
}
\value{
Reformatted character vector
}
\description{
Title case
}
\examples{
titleCase("RNA sequencing")
firstCase("RNA sequencing")
}
\author{
Michael Steinbaugh
}
