% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{titleCase}
\alias{titleCase}
\alias{title_case}
\alias{firstCase}
\alias{first_case}
\title{Change case of a character vector.}
\usage{
titleCase(character)

firstCase(character)
}
\arguments{
\item{character}{Character vector.}
}
\value{
Reformatted character vector.
}
\description{
Title case.

First case.
}
\examples{
titleCase("RNA sequencing")
firstCase("RNA sequencing")
}
