% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cell2sample-methods.R
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample,SingleCellExperiment-method}
\title{Cell-to-sample mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{SingleCellExperiment}(object, return = c("factor", "DataFrame", "tbl_df"))
}
\arguments{
\item{object}{Object.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{"factor"}: Named \code{factor} containing as the \link[base:levels]{levels} and
cell IDs as the \link[base:names]{names}.
\item \code{DataFrame}: Data frame containing \code{sampleId} column and cell identifiers
as the row names.
\item \code{"tbl_df"}: Tibble containing \code{cellId} and \code{sampleId} columns.
}
}
\description{
Cell-to-sample mappings
}
\note{
\code{sampleId} column must be defined in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}}.

Updated 2021-01-17.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")
sce <- SingleCellExperiment

## SingleCellExperiment ====
x <- cell2sample(sce)
table(x)
}
