% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cell2sample-methods.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample,SingleCellExperiment-method}
\title{Cell-to-Sample Mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{SingleCellExperiment}(object, return = c("factor",
  "tibble"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{return}{\code{string}. Return type. Uses \code{base::match.arg()}
internally and defaults to the first argument in the \code{character} vector.}
}
\value{
Named \code{factor}. Sample IDs as the \code{levels} and cell IDs as the
\code{names}.
\itemize{
\item \code{"factor"}: Named \code{factor} containing sample IDs as the levels and cell IDs
as the names.
\item \code{"tibble"}: \code{tbl_df} containing \code{cellID} and \code{sampleID} columns.
}
}
\description{
Cell-to-Sample Mappings
}
\note{
\code{sampleID} column must be defined in \code{colData()}.
}
\examples{
data(sce)
x <- cell2sample(sce)
table(x)
}
