% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2sample-methods.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample,SingleCellExperiment-method}
\title{Cell-to-sample mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{SingleCellExperiment}(object, return = c("factor",
  "tbl_df"))
}
\arguments{
\item{object}{Object.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{"factor"}: Named \code{factor} containing as the \link[base:levels]{levels} and
cell IDs as the \link[base:names]{names}.
\item \code{"tbl_df"}: Tibble containing \code{cellID} and \code{sampleID} columns.
}
}
\description{
Cell-to-sample mappings
}
\note{
\code{sampleID} column must be defined in
\code{\link[SummarizedExperiment:colData]{colData()}}.

Updated 2019-08-18.
}
\examples{
data(SingleCellExperiment, package = "acidtest")
sce <- SingleCellExperiment

## SingleCellExperiment ====
x <- cell2sample(sce)
table(x)
}
