% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2sample-methods.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample,SingleCellExperiment-method}
\title{Cell-to-sample mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{SingleCellExperiment}(object, return = c("factor",
  "tibble"))
}
\arguments{
\item{object}{Object.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{"factor"}: Named \code{factor} containing as the \link[base:levels]{levels} and
cell IDs as the \link[base:names]{names}.
\item \code{"tibble"}: \code{tbl_df} containing \code{cellID} and \code{sampleID} columns.
}
}
\description{
Cell-to-sample mappings
}
\note{
\code{sampleID} column must be defined in
\code{\link[SummarizedExperiment:colData]{colData()}}.
}
\examples{
data(sce, package = "acidtest")
x <- cell2sample(sce)
table(x)
}
