% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cell2sample-methods.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample,SingleCellExperiment-method}
\title{Cell to Sample Mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Named \code{factor} containing sample IDs as the levels and cell IDs as
the names.
}
\description{
Cell to Sample Mappings
}
\note{
\code{sampleID} column must be defined in \code{\link[=colData]{colData()}}.
}
\examples{
x <- cell2sample(sce_small)
table(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}, \code{\link{tpm}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
