% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cell2sample-methods.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample,SingleCellExperiment-method}
\title{Cell to Sample Mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{SingleCellExperiment}(object, return = c("factor",
  "tibble"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\value{
\itemize{
\item \code{"factor"}: Named \code{factor} containing sample IDs as the levels and cell IDs
as the names.
\item \code{"tibble"}: \code{tbl_df} containing \code{cellID} and \code{sampleID} columns.
}
}
\description{
Cell to Sample Mappings
}
\note{
\code{sampleID} column must be defined in \code{\link[=colData]{colData()}}.
}
\examples{
x <- cell2sample(sce_small)
table(x)
}
\seealso{
Other Single-Cell Functions: \code{\link{makeSingleCellExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{topCellsPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{Single-Cell Functions}
