% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClasses.R
\name{checkClasses}
\alias{checkClasses}
\title{Check Classes}
\usage{
checkClasses(object, expected)
}
\arguments{
\item{object}{Object.}

\item{expected}{\code{list}. Named list of expected classes per slot.}
}
\description{
Check Classes
}
\examples{
checkClasses(
    object = metadata(rse_small),
    expected = list(
        version = c("package_version" "numeric_version"),
        date = "Date",
        interestingGroups = "character"
    )
)
}
