% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{import}
\name{classes}
\alias{classes}
\alias{EggNOG-class}
\alias{Ensembl2Entrez-class}
\alias{Gene2Symbol-class}
\alias{HGNC2Ensembl-class}
\alias{MGI2Ensembl-class}
\alias{PANTHER-class}
\alias{Tx2Gene-class}
\title{Classes exported from other packages}
\description{
These classes are imported from other packages.
Follow the links below to see their documentation.
}
\details{
\itemize{
\item \code{help(topic = "EggNOG-class", package = "basejump.classes")}.
\item \code{help(topic = "Ensembl2Entrez-class", package = "basejump.classes")}.
\item \code{help(topic = "Gene2Symbol-class", package = "basejump.classes")}.
\item \code{help(topic = "HGNC2Ensembl-class", package = "basejump.classes")}.
\item \code{help(topic = "MGI2Ensembl-class", package = "basejump.classes")}.
\item \code{help(topic = "PANTHER-class", package = "basejump.classes")}.
\item \code{help(topic = "Tx2Gene-class", package = "basejump.classes")}.
}
}
\keyword{internal}
