% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSystemLibrary.R
\name{cleanSystemLibrary}
\alias{cleanSystemLibrary}
\title{Check for Clean System Library}
\usage{
cleanSystemLibrary()
}
\value{
\code{logical}. Is system library clean?
}
\description{
Determine whether a user has installed packages into the system library.
}
\examples{
cleanSystemLibrary()
}
\author{
Michael Steinbaugh
}
