% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSystemLibrary.R
\name{cleanSystemLibrary}
\alias{cleanSystemLibrary}
\title{Check for Clean System Library}
\usage{
cleanSystemLibrary()
}
\value{
\code{logical}. Is system library clean?
}
\description{
Determine whether a user has installed packages into the system library.
}
\examples{
cleanSystemLibrary()
}
\seealso{
Other Developer Functions: \code{\link{calls}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
