% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce,SummarizedExperiment,list-method}
\alias{coerce,matrix,tbl_df-method}
\alias{coerce,data.frame,tbl_df-method}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GRanges,tbl_df-method}
\title{Methods for Coercing an Object to a Class}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\section{tibble}{

Coerce an object to a \code{tibble} (\code{tbl_df}). Tibbles don't support rowname
assignemnt, so here we are ensuring they are kept by moving them to a column
named \code{rowname} upon coercion. This helps avoid downstream unexpected data
loss when using the dplyr chain of single table verbs, such as
\code{\link[dplyr:arrange]{dplyr::arrange()}}, \code{\link[dplyr:filter]{dplyr::filter()}}, or \code{\link[dplyr:mutate]{dplyr::mutate()}}.
}

\examples{
# data.frame ====
# Automatically move rownames to `rowname` column
as(datasets::mtcars, "tbl_df") \%>\% glimpse()

# tibble ====
# Return unmodified
as(ggplot2::mpg, "tbl_df") \%>\% glimpse()
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:canCoerce]{methods::canCoerce()}}.
\item \code{\link[tibble:tibble]{tibble::tibble()}}.
}
}
\author{
Michael Steinbaugh
}
