% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce,SummarizedExperiment,list-method}
\alias{coerce,matrix,tbl_df-method}
\alias{coerce,data.frame,tbl_df-method}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GRanges,tbl_df-method}
\alias{coerce,tbl_df,DataFrame-method}
\title{Methods for Coercing an Object to a Class}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\section{tibble}{

Coerce an object to a \code{tibble} using \code{as(object, Class = "tbl_df")}. Tibbles
don't support rowname assignment, so here we are ensuring they are kept by
moving them to a column named \code{rowname} upon coercion. This helps avoid
downstream unexpected data loss when using the dplyr chain of single table
verbs, such as \code{\link[dplyr:arrange]{dplyr::arrange()}}, \code{\link[dplyr:filter]{dplyr::filter()}}, or \code{\link[dplyr:mutate]{dplyr::mutate()}}.

Conversely, when coercing a \code{tibble} back to an S4 \code{DataFrame}, our
\code{as(tbl_df, Class = "DataFrame")} method looks for the "rowname" column and
will attempt to move it back to \code{\link[=rownames]{rownames()}} automatically, unless there are
duplicates present.
}

\examples{
# DataFrame to tbl_df ====
# Automatically move rownames to `rowname` column
data <- colData(rse_bcb)
class(data)
hasRownames(data)

tbl_df <- as(data, Class = "tbl_df")
hasRownames(tbl_df)
print(tbl_df)

# tbl_df back to DataFrame ====
data <- as(tbl_df, "DataFrame")
hasRownames(data)
class(data)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:canCoerce]{methods::canCoerce()}}.
\item \code{\link[tibble:tibble]{tibble::tibble()}}.
}
}
\author{
Michael Steinbaugh
}
