% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce-methods.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce,matrix,tbl_df-method}
\alias{coerce,data.frame,tbl_df-method}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GRanges,tbl_df-method}
\alias{coerce,tbl_df,DataFrame-method}
\alias{coerceS4ToList,ANY-method}
\alias{coerce,SummarizedExperiment,list-method}
\alias{as.list,SummarizedExperiment-method}
\title{Methods for Coercing an Object to a Class}
\usage{
coerceS4ToList(from)

\S4method{coerceS4ToList}{ANY}(from)

\S4method{as.list}{SummarizedExperiment}(x)
}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\section{\code{tbl_df} (tibble)}{


Coerce an object to a \code{tibble} using \code{as(object, Class = "tbl_df")}. Tibbles
don't support rowname assignment, so here we are ensuring they are kept by
moving them to a column named \code{rowname} upon coercion. This helps avoid
downstream unexpected data loss when using the dplyr chain of single table
verbs, such as \code{\link[dplyr:arrange]{dplyr::arrange()}}, \code{\link[dplyr:filter]{dplyr::filter()}}, or \code{\link[dplyr:mutate]{dplyr::mutate()}}.

Conversely, when coercing a \code{tibble} back to an S4 \code{DataFrame}, our
\code{as(tbl_df, Class = "DataFrame")} method looks for the "rowname" column and
will attempt to move it back to \code{\link[=rownames]{rownames()}} automatically, unless there are
duplicates present.
}

\section{\code{list}}{


It is often useful to coerce an S4 object to a flat \code{list} for archival
storage. Here we are providing the \code{\link[=coerceS4ToList]{coerceS4ToList()}} generic, which
consistently coerces the slots in any S4 to a standard \code{list}. Additionally,
here we have improved support for \code{SummarizedExperiment} to \code{list} coercion,
returning the slots as a \code{list}.
}

\examples{
# DataFrame to tbl_df ====
# Automatically move row names to `rowname` column.
data <- colData(rse_small)
class(data)
hasRownames(data)

tbl_df <- as(data, Class = "tbl_df")
hasRownames(tbl_df)
print(tbl_df)

# tbl_df to DataFrame ====
data <- as(tbl_df, "DataFrame")
hasRownames(data)
class(data)

# SummarizedExperiment to list ====
x <- as(rse_small, "list")
class(x)
names(x)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:canCoerce]{methods::canCoerce()}}.
\item \code{\link[S4Vectors:as.list]{S4Vectors::as.list()}}.
\item \code{\link[tibble:tibble]{tibble::tibble()}}, \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.
}
}
\author{
Michael Steinbaugh
}
