% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GRanges,tbl_df-method}
\alias{coerce,sparseMatrix,data.frame-method}
\alias{coerce,tbl_df,DataFrame-method}
\alias{coerceS4ToList}
\alias{coerce,SummarizedExperiment,list-method}
\alias{as.list,SummarizedExperiment-method}
\title{Methods for Coercing an Object to a Class}
\usage{
coerceS4ToList(from)

\S4method{as.list}{SummarizedExperiment}(x)
}
\arguments{
\item{from}{The classes between which the coerce methods
    \code{def} and \code{replace} perform coercion.
  }

\item{x}{Object.}
}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\section{tibble}{


Coerce an object to a \code{tibble} (\code{tbl_df}) data frame using either S3 or S4
methods:
\itemize{
\item S3: \code{as_tibble(x)} (or \code{as.tibble()} alias).
\item S4: \code{as(object, Class = "tbl_df")}.
}

Tibbles don't support row name assignment, so here we are ensuring they are
kept by moving them to a column named \code{rowname} by default. This helps avoid
downstream unexpected data loss when using the dplyr chain of single table
verbs, such as \code{\link[dplyr:arrange]{dplyr::arrange()}}, \code{\link[dplyr:filter]{dplyr::filter()}}, or \code{\link[dplyr:mutate]{dplyr::mutate()}}.

This behavior can be overriden in the S3 method by setting \code{rowname = NULL}
instead, which is the current default in the tibble package. The S4 coercion
method doesn't support arguments, and therefore always attempts to move
rownames automatically, if defined.

Conversely, when coercing a \code{tibble} back to an S4 \code{DataFrame}, our
\code{as(tbl_df, Class = "DataFrame")} method looks for the \code{rowname} column and
will attempt to move it back to \code{\link[=rownames]{rownames()}} automatically, unless there are
duplicates present.
}

\section{list}{


It is often useful to coerce an S4 object to a flat \code{list} for archival
storage. Here we are providing the \code{\link[=coerceS4ToList]{coerceS4ToList()}} function, which
consistently coerces the slots in any S4 to a standard \code{list}. Additionally,
here we have improved support for \code{SummarizedExperiment} to \code{list} coercion,
returning the slots as a \code{list}.
}

\examples{
data(rse_small)

## DataFrame to tbl_df ====
data <- colData(rse_small)
print(data)

## Automatically move row names to `rowname` column by default.
tbl_df <- as(data, Class = "tbl_df")

## tbl_df to DataFrame ====
data <- as(tbl_df, Class = "DataFrame")
print(data)

## SummarizedExperiment to list ====
x <- as(rse_small, Class = "list")
class(x)
names(x)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:canCoerce]{methods::canCoerce()}}.
\item \code{\link[methods:showMethods]{methods::showMethods()}}.
\item \code{\link[utils:methods]{utils::methods()}}.
\item \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.
\item \code{\link[=as.list]{as.list()}}.
}
}
