% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerceS4ToList.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{coerceS4ToList}
\alias{flatFiles,SummarizedExperiment-method}
\title{Coerce Any S4 Object to List}
\usage{
flatFiles(object, ...)

coerceS4ToList(from)

\S4method{flatFiles}{SummarizedExperiment}(object)
}
\arguments{
\item{from}{The classes between which the coerce methods
    \code{def} and \code{replace} perform coercion.
  }
}
\description{
Coerce Any S4 Object to List
}
\examples{
data(rse)

x <- coerceS4ToList(rse)
class(x)
names(x)

x <- flatFiles(rse)
class(x)
names(x)
}
\seealso{
\code{methods::coerce}.
}
