% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{coerceS4ToList}
\alias{coerceS4ToList}
\alias{flatFiles}
\title{Coerce Any S4 Object to List}
\usage{
coerceS4ToList(from)

flatFiles(from)
}
\arguments{
\item{from}{The classes between which the coerce methods
    \code{def} and \code{replace} perform coercion.
  }
}
\description{
Coerce Any S4 Object to List
}
\examples{
data(rse)
list <- coerceS4ToList(rse)
class(list)
names(list)
}
\seealso{
\code{methods::coerce}.
}
