% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{toStringUnique}
\alias{toStringUnique}
\alias{toStringSortUnique}
\alias{summarizeRows}
\title{Collapse utilities}
\usage{
toStringUnique(x, sep = ", ")

toStringSortUnique(x, sep = ", ")

summarizeRows(x, sep = ", ")
}
\arguments{
\item{x}{The object to be converted.}

\item{sep}{Separator used for collapse.}
}
\value{
A character vector of length 1 is returned.
}
\description{
Helper functions for to produce collapsed output for vectors (\code{\link[=toString]{toString()}})
and data frames (summarizeRows()]).
}
\examples{
toStringUnique(c("milk", "eggs", "eggs", NA))
toStringSortUnique(c("milk", "eggs", "eggs", NA))

mtcars \%>\% head \%>\% summarizeRows
}
