% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-collapse.R
\docType{methods}
\name{collapse}
\alias{collapse}
\alias{collapse,character-method}
\alias{collapse,integer-method}
\alias{collapse,data.frame-method}
\alias{collapse,DataFrame-method}
\alias{collapse,matrix-method}
\alias{collapse,tbl_df-method}
\title{Collapse Utilities}
\usage{
collapse(object, ...)

\S4method{collapse}{character}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapse}{integer}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapse}{data.frame}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapse}{DataFrame}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapse}{matrix}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapse}{tbl_df}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[glue:collapse]{glue::collapse()}}.}

\item{sep}{Separator. Defaults to comma.}

\item{unique}{Unique values.}

\item{sort}{Sort values.}

\item{keepNA}{Keep \code{NA} values.}
}
\value{
\itemize{
\item For vector: String.
\item For column data: Single row.
}
}
\description{
Collapse Utilities
}
\examples{
# Designed primarily for `dims` objects
mtcars \%>\% head \%>\% collapse \%>\% t

# Vectors are also supported
groceries <- c("milk", "eggs", "eggs", "veggies", NA)
collapse(groceries)
collapse(groceries, unique = FALSE, sort = FALSE)
}
\seealso{
\itemize{
\item \code{\link[base:toString]{base::toString()}}.
\item \code{\link[glue:collapse]{glue::collapse()}}.
}
}
\concept{Data Manipulation Utilities}
