% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse rows in a data.frame.}
\usage{
collapse(tibble)
}
\arguments{
\item{tibble}{long tibble.}
}
\value{
collapsed tibble.
}
\description{
Collapse rows in a data.frame.
}

