% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{toStringUnique}
\alias{toStringUnique}
\alias{toStringSortUnique}
\alias{summarizeRows}
\alias{summariseRows}
\title{Collapse Utilities}
\usage{
toStringUnique(x, sep = ", ")

toStringSortUnique(x, sep = ", ")

summarizeRows(x, sep = ", ")
}
\arguments{
\item{x}{The object to be converted.}

\item{sep}{Separator used for collapse.}
}
\value{
A character vector of length 1 is returned.
}
\description{
Helper functions for to produce collapsed output for vectors
(\code{\link[=toString]{toString()}}) and data frames (\code{\link[=summarizeRows]{summarizeRows()}}).
}
\examples{
groceries <- c("milk", "eggs", "eggs", NA)

toStringUnique(groceries)
toStringSortUnique(groceries)

mtcars \%>\% head \%>\% summarizeRows
}
