% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-collapse.R
\docType{methods}
\name{collapse}
\alias{collapse}
\alias{summarizeRows}
\alias{.collapseVec}
\alias{.collapseCols}
\alias{collapse,character-method}
\alias{collapse,integer-method}
\alias{collapse,data.frame-method}
\alias{collapse,DataFrame-method}
\alias{collapse,matrix-method}
\alias{collapse,tbl_df-method}
\title{Collapse Utilities}
\usage{
collapse(object, ...)

summarizeRows(object, ...)

.collapseVec(object, sep = ", ", unique = TRUE, sort = TRUE,
  keepNA = FALSE, ...)

.collapseCols(object, ...)

\S4method{collapse}{character}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapse}{integer}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapse}{data.frame}(object)

\S4method{collapse}{DataFrame}(object)

\S4method{collapse}{matrix}(object)

\S4method{collapse}{tbl_df}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{sep}{Separator. Defaults to comma, like \code{\link[base:toString]{base::toString()}}.}

\item{unique}{Unique values.}

\item{sort}{Sort values.}

\item{keepNA}{Keep \code{NA} values.}
}
\value{
\itemize{
\item Vector: String.
\item Column data: Single row.
}
}
\description{
Collapse Utilities
}
\examples{
groceries <- c("milk", "eggs", "eggs", "veggies", NA)

collapse(groceries)
collapse(groceries, sort = )

mtcars \%>\% head \%>\% collapse \%>\% t
}
\seealso{
\itemize{
\item \code{\link[base:toString]{base::toString()}}.
\item \code{\link[glue:collapse]{glue::collapse()}}.
}
}
