% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse rows in a tibble}
\usage{
collapse(tibble)
}
\arguments{
\item{tibble}{Long \code{tibble}}
}
\value{
Collapsed \code{tibble}
}
\description{
Collapse rows in a tibble
}
\keyword{general}

