% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/collapseToString-methods.R
\docType{methods}
\name{collapseToString}
\alias{collapseToString}
\alias{collapseToString,atomic-method}
\alias{collapseToString,data.frame-method}
\alias{collapseToString,DataFrame-method}
\alias{collapseToString,matrix-method}
\title{Collapse to String}
\usage{
collapseToString(object, ...)

\S4method{collapseToString}{atomic}(object, sep = ", ", sort = FALSE,
  removeNA = FALSE, unique = FALSE)

\S4method{collapseToString}{data.frame}(object, sep = ", ",
  sort = FALSE, removeNA = FALSE, unique = FALSE)

\S4method{collapseToString}{DataFrame}(object, sep = ", ",
  sort = FALSE, removeNA = FALSE, unique = FALSE)

\S4method{collapseToString}{matrix}(object, sep = ", ", sort = FALSE,
  removeNA = FALSE, unique = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{sep}{\code{string}. Separator. Defaults to comma.}

\item{sort}{\code{boolean}. Sort values.}

\item{removeNA}{\code{boolean}. Remove NA values.}

\item{unique}{\code{boolean}. Unique values.}
}
\value{
\itemize{
\item For \code{atomic}: \code{string}.
\item For column data: Object collapsed to a single row.
}
}
\description{
Collapse to String
}
\examples{
# character ====
groceries <- c(NA, NA, "milk", "eggs", "eggs", "veggies")
collapseToString(
    groceries,
    unique = TRUE,
    sort = TRUE,
    removeNA = TRUE
)
collapseToString(
    groceries,
    unique = FALSE,
    sort = FALSE,
    removeNA = FALSE
)

# numeric ====
collapseToString(seq(1:5))

# logical ====
collapseToString(c(TRUE, FALSE))
collapseToString(c(NA, NaN))

# data.frame ====
# Objects supporting `dim` function similarly
ggplot2::mpg \%>\%
    head() \%>\%
    collapseToString(sort = TRUE, unique = TRUE) \%>\%
    t()
}
\seealso{
\itemize{
\item \code{\link[base:toString]{base::toString()}}.
}

Other Sanitization Functions: \code{\link{fixNA}},
  \code{\link{removeNA}}, \code{\link{sanitizeRowData}},
  \code{\link{sanitizeSampleData}}
}
\author{
Michael Steinbaugh
}
\concept{Sanitization Functions}
