% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-collapseToString.R
\docType{methods}
\name{collapseToString}
\alias{collapseToString}
\alias{collapseToString,character-method}
\alias{collapseToString,integer-method}
\alias{collapseToString,data.frame-method}
\alias{collapseToString,DataFrame-method}
\alias{collapseToString,matrix-method}
\alias{collapseToString,tbl_df-method}
\title{Collapse to String}
\usage{
collapseToString(object, ...)

\S4method{collapseToString}{character}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapseToString}{integer}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapseToString}{data.frame}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapseToString}{DataFrame}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapseToString}{matrix}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

\S4method{collapseToString}{tbl_df}(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[glue:collapse]{glue::collapse()}}.}

\item{sep}{Separator. Defaults to comma.}

\item{unique}{Unique values.}

\item{sort}{Sort values.}

\item{keepNA}{Keep \code{NA} values.}
}
\value{
\itemize{
\item For vector: String.
\item For column data: Single row.
}
}
\description{
Collapse to String
}
\examples{
# Designed primarily for `dims` objects
mtcars \%>\%
    head() \%>\%
    collapseToString() \%>\%
    t()

# Vectors are also supported
groceries <- c("milk", "eggs", "eggs", "veggies", NA)
collapseToString(groceries)
collapseToString(groceries, unique = FALSE, sort = FALSE)
}
\seealso{
\itemize{
\item \code{\link[base:toString]{base::toString()}}.
\item \code{\link[glue:collapse]{glue::collapse()}}.
}
}
\concept{Data Manipulation Utilities}
