% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-collapseToString.R
\docType{methods}
\name{collapseToString}
\alias{collapseToString}
\alias{collapseToString,character-method}
\alias{collapseToString,data.frame-method}
\alias{collapseToString,DataFrame-method}
\alias{collapseToString,integer-method}
\alias{collapseToString,logical-method}
\alias{collapseToString,matrix-method}
\alias{collapseToString,numeric-method}
\alias{collapseToString,tbl_df-method}
\title{Collapse to String}
\usage{
collapseToString(object, ...)

\S4method{collapseToString}{character}(object, sep = ", ", unique = TRUE,
  sort = TRUE, ...)

\S4method{collapseToString}{data.frame}(object, sep = ", ", unique = TRUE,
  sort = TRUE, ...)

\S4method{collapseToString}{DataFrame}(object, sep = ", ", unique = TRUE,
  sort = TRUE, ...)

\S4method{collapseToString}{integer}(object, sep = ", ", unique = TRUE,
  sort = TRUE, ...)

\S4method{collapseToString}{logical}(object, sep = ", ", unique = TRUE,
  sort = TRUE, ...)

\S4method{collapseToString}{matrix}(object, sep = ", ", unique = TRUE,
  sort = TRUE, ...)

\S4method{collapseToString}{numeric}(object, sep = ", ", unique = TRUE,
  sort = TRUE, ...)

\S4method{collapseToString}{tbl_df}(object, sep = ", ", unique = TRUE,
  sort = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{sep}{Separator. Defaults to comma.}

\item{unique}{Unique values.}

\item{sort}{Sort values.}
}
\value{
\itemize{
\item For vector: String.
\item For column data: Single row.
}
}
\description{
Collapse to String
}
\examples{
# character
groceries <- c("milk", "eggs", "eggs", "veggies", NA)
collapseToString(groceries, unique = TRUE, sort = FALSE)
collapseToString(groceries, unique = FALSE, sort = FALSE)

# numeric
collapseToString(seq(1:5))

# logical
collapseToString(c(TRUE, FALSE))
collapseToString(c(NA, NaN))

# data.frame
# Objects supporting `dim` function similarly
mtcars \%>\%
    head() \%>\%
    collapseToString() \%>\%
    t()
}
\seealso{
\itemize{
\item \code{\link[base:toString]{base::toString()}}.
}
}
\concept{Data Manipulation Utilities}
