% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,SummarizedExperiment,SummarizedExperiment-method}
\alias{combine,SingleCellExperiment,SingleCellExperiment-method}
\title{Combine Multiple Objects}
\usage{
combine(x, y, ...)

\S4method{combine}{SummarizedExperiment,SummarizedExperiment}(x, y)

\S4method{combine}{SingleCellExperiment,SingleCellExperiment}(x, y)
}
\arguments{
\item{x}{One of the values.}

\item{y}{A second value.}

\item{...}{Any other objects of the same class as \code{x} and \code{y}.}
}
\value{
\code{SummarizedExperiment}.
}
\description{
Combine Multiple Objects
}
\note{
We're attempting to make this as strict as possible, requiring:
\itemize{
\item Rows (genes) across objects must be identical.
\item \code{\link[SummarizedExperiment:rowRanges]{SummarizedExperiment::rowRanges()}} and/or
\code{\link[SummarizedExperiment:rowData]{SummarizedExperiment::rowData()}} metadata must be identical.
\item \code{\link[SummarizedExperiment:colData]{SummarizedExperiment::colData()}} must contain the same columns.
\item Specific metadata must be identical (see \code{metadata} parameter).
}
}
\examples{
## SummarizedExperiment ====
x <- basejump::rse_small
colnames(x)
SummarizedExperiment::colData(x)

## Create a copy of our minimal example.
y <- x
colnames(y) <- paste0("sample", seq(from = 5L, to = 8L))
colnames(y)
SummarizedExperiment::colData(y)

## Combine two SummarizedExperiment objects.
c <- combine(x, y)
print(c)
colnames(c)
SummarizedExperiment::colData(c)

## SingleCellExperiment ====
x <- basejump::sce_small
head(colnames(x))
sampleData(x)

## Here we're faking a distinct replicate, just as an example.
y <- x
## Increase the cell ID numbers.
cells <- colnames(y) \%>\%
    sub("cell", "", .) \%>\%
    as.integer() \%>\%
    `+`(ncol(y)) \%>\%
    paste0("cell", .)
colnames(y) <- cells
head(colnames(y))
## Increase the sample ID numbers.
y$sampleID <- gsub("1$", "3", y$sampleID)
y$sampleID <- gsub("2$", "4", y$sampleID)
sampleData(y)

## Combine two SingleCellExperiment objects.
c <- combine(x, y)
print(c)
sampleData(c)
}
\seealso{
\itemize{
\item \code{\link[BiocGenerics:combine]{BiocGenerics::combine()}}.
\item \code{help("merge.Matrix", "Matrix.utils")}.
}
}
