% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,SummarizedExperiment,SummarizedExperiment-method}
\title{Combine Multiple Objects}
\usage{
\S4method{combine}{SummarizedExperiment,SummarizedExperiment}(x, y,
  metadata = c("version", "interestingGroups", "organism", "genomeBuild",
  "ensemblRelease", "rowRangesMetadata", "gffFile"))
}
\arguments{
\item{x}{One of the values.}

\item{y}{A second value.}

\item{metadata}{\code{character}. Metadata slot names that must be identical
between the datasets.}
}
\value{
\code{SummarizedExperiment}.
}
\description{
Combine Multiple Objects
}
\note{
We're attempting to make this as strict as possible, requiring:
\itemize{
\item Rows (genes) across objects must be identical.
\item \code{\link[=rowRanges]{rowRanges()}} and/or \code{\link[=rowData]{rowData()}} metadata must be identical.
\item \code{\link[=colData]{colData()}} must contain the same columns.
\item Specific metadata must be identical (see \code{metadata} parameter).
}
}
\examples{
# For this simple working example, let's duplicate our minimal dataset.

x <- rse_small
colnames(x)
colData(x)

y <- x
colnames(y) <- paste0("sample", seq(from = 5L, to = 8L))
colnames(y)
colData(y)[["sampleName"]] <- colnames(y)
colData(y)

# Combine two SummarizedExperiment objects.
c <- combine(x, y)
print(c)
colnames(c)
colData(c)
}
\seealso{
\code{\link[BiocGenerics:combine]{BiocGenerics::combine()}}.
}
\author{
Michael Steinbaugh
}
