% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,SummarizedExperiment,SummarizedExperiment-method}
\alias{combine,SingleCellExperiment,SingleCellExperiment-method}
\title{Combine Multiple Objects}
\usage{
\S4method{combine}{SummarizedExperiment,SummarizedExperiment}(x, y)

\S4method{combine}{SingleCellExperiment,SingleCellExperiment}(x, y)
}
\arguments{
\item{x}{One of the values.}

\item{y}{A second value.}
}
\value{
\code{SummarizedExperiment}.
}
\description{
Combine Multiple Objects
}
\note{
We're attempting to make this as strict as possible, requiring:
\itemize{
\item Rows (genes) across objects must be identical.
\item \code{\link[=rowRanges]{rowRanges()}} and/or \code{\link[=rowData]{rowData()}} metadata must be identical.
\item \code{\link[=colData]{colData()}} must contain the same columns.
\item Specific metadata must be identical (see \code{metadata} parameter).
}
}
\examples{
data(rse_small, sce_small)

# SummarizedExperiment ====
x <- rse_small
colnames(x)
colData(x)

# Create a copy of our minimal example.
y <- x
colnames(y) <- paste0("sample", seq(from = 5L, to = 8L))
colnames(y)
colData(y)

# Combine two SummarizedExperiment objects.
c <- combine(x, y)
print(c)
colnames(c)
colData(c)

# SingleCellExperiment ====
x <- sce_small
head(colnames(x))
sampleData(x)

# Here we're faking a distinct replicate, just as an example.
y <- x
# Increase the cell ID numbers.
cells <- colnames(y) \%>\%
    sub("cell", "", .) \%>\%
    as.integer() \%>\%
    `+`(ncol(y)) \%>\%
    paste0("cell", .)
colnames(y) <- cells
head(colnames(y))
# Increase the sample ID numbers.
y$sampleID <- gsub("1$", "3", y$sampleID)
y$sampleID <- gsub("2$", "4", y$sampleID)
sampleData(y)

# Combine two SingleCellExperiment objects.
c <- combine(x, y)
print(c)
sampleData(c)
}
\seealso{
\itemize{
\item \code{\link[BiocGenerics:combine]{BiocGenerics::combine()}}.
\item \code{help("merge.Matrix", "Matrix.utils")}.
}

Other Data Functions: \code{\link{aggregate}},
  \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
