% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,SummarizedExperiment,SummarizedExperiment-method}
\alias{combine,SingleCellExperiment,SingleCellExperiment-method}
\title{Combine Multiple Objects}
\usage{
\S4method{combine}{SummarizedExperiment,SummarizedExperiment}(x, y,
  metadata = c("version", "interestingGroups", "organism", "genomeBuild",
  "ensemblRelease", "rowRangesMetadata", "gffFile"))

\S4method{combine}{SingleCellExperiment,SingleCellExperiment}(x, y,
  metadata = c("version", "pipeline", "level", "interestingGroups",
  "organism", "genomeBuild", "ensemblRelease", "umiType", "gffFile",
  "dataVersions", "programVersions"))
}
\arguments{
\item{x}{One of the values.}

\item{y}{A second value.}

\item{metadata}{\code{character}. Metadata slot names that must be identical
between the datasets.}
}
\value{
\code{SummarizedExperiment}.
}
\description{
Combine Multiple Objects
}
\note{
We're attempting to make this as strict as possible, requiring:
\itemize{
\item Rows (genes) across objects must be identical.
\item \code{\link[=rowRanges]{rowRanges()}} and/or \code{\link[=rowData]{rowData()}} metadata must be identical.
\item \code{\link[=colData]{colData()}} must contain the same columns.
\item Specific metadata must be identical (see \code{metadata} parameter).
}
}
\examples{
# SummarizedExperiment ====
# For this simple working example, let's duplicate our minimal dataset.

x <- rse_small
colnames(x)
colData(x)

y <- x
colnames(y) <- paste0("sample", seq(from = 5L, to = 8L))
colnames(y)
colData(y)[["sampleName"]] <- colnames(y)
colData(y)

# Combine two SummarizedExperiment objects.
c <- combine(x, y)
print(c)
colnames(c)
colData(c)

# SingleCellExperiment ====
# For this simple working example, let's duplicate our minimal dataset,
# modify the cell barcodes, and combine.

x <- indrops_small
validObject(x)
sampleData(x)
sampleNames(x)

# Here we're faking a distinct replicate, just as an example.
y <- indrops_small
# Modify the cellular barcodes prefix.
colnames(y) <- gsub(
    pattern = "^multiplexed_AAAAAAAA",
    replacement = "multiplexed_CCCCCCCC",
    x = colnames(y)
)
# Modify the column data.
colData(y)[["description"]] <- factor("multiplexed-CCCCCCCC")
colData(y)[["sampleID"]] <- factor("multiplexed_CCCCCCCC")
colData(y)[["sampleName"]] <- factor("rep_2")
colData(y)[["sequence"]] <- factor("GGGGGGGG")
colData(y)[["revcomp"]] <- factor("CCCCCCCC")
# Modify the sample data.
sampleData(y)[["sampleName"]] <- factor("rep_2")
sampleData(y)[["description"]] <- factor("multiplexed-CCCCCCCC")
sampleData(y)[["sequence"]] <- factor("GGGGGGGG")
sampleData(y)[["revcomp"]] <- factor("CCCCCCCC")
rownames(sampleData(y)) <- "multiplexed_CCCCCCCC"
validObject(y)
sampleData(y)
sampleNames(y)

# Combine two SingleCellExperiment objects.
c <- combine(x, y)
print(c)
sampleNames(c)
}
\seealso{
\itemize{
\item \code{\link[BiocGenerics:combine]{BiocGenerics::combine()}}.
\item \code{help("merge.Matrix", "Matrix.utils")}.
}
}
\author{
Michael Steinbaugh
}
