% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-convertGenesToSymbols.R
\docType{methods}
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertGenesToSymbols,character-method}
\alias{convertGenesToSymbols,matrix-method}
\alias{convertGenesToSymbols,data.frame-method}
\alias{convertGenesToSymbols,DataFrame-method}
\alias{convertGenesToSymbols,dgCMatrix-method}
\alias{convertGenesToSymbols,dgTMatrix-method}
\title{Convert Ensembl Identifiers to Gene Symbols}
\usage{
convertGenesToSymbols(object, ...)

\S4method{convertGenesToSymbols}{character}(object, gene2symbol, organism,
  genomeBuild = NULL, release = NULL)

\S4method{convertGenesToSymbols}{matrix}(object, ...)

\S4method{convertGenesToSymbols}{data.frame}(object, ...)

\S4method{convertGenesToSymbols}{DataFrame}(object, ...)

\S4method{convertGenesToSymbols}{dgCMatrix}(object, ...)

\S4method{convertGenesToSymbols}{dgTMatrix}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{gene2symbol}{\emph{Optional.} Gene-to-symbol mappings. If \code{NULL}, will
attempt to download from Ensembl using the desired \code{organism},
\code{genomeBuild}, and \code{release} arguments.}

\item{organism}{\emph{Optional.} Organism name. Normally this argument is
unnecessary and can be left unset. If a count matrix starts with a
FASTA spike-in (e.g. "EGFP"), then automatic genome detection based on the
first gene identifier will fail. In this case, the desired organism must be
manually declared.}

\item{genomeBuild}{\emph{Optional.} Genome build assembly name (e.g. "\code{GRCh38}").
If \code{NULL}, defaults to the most recent build available.}

\item{release}{\emph{Optional.} Release version (e.g. \code{90}). If \code{NULL}, defaults
to the most recent release available.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Identifiers to Gene Symbols
}
\examples{
# character ====
x <- c("ENSMUSG00000000001", "ENSMUSG00000000003")
convertGenesToSymbols(x)

# matrix ====
mat <- matrix(
    data = seq(1L:4L),
    byrow = TRUE,
    nrow = 2L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUSG00000000001", "ENSMUSG00000000003"),
        c("sample_1", "sample_2")
    )
)
print(mat)
mat <- convertGenesToSymbols(mat)
print(mat)
rownames(mat)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}}, \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
