% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertGenesToSymbols-methods.R, R/convertSymbolsToGenes-methods.R
\docType{methods}
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertSymbolsToGenes}
\alias{convertGenesToSymbols,character-method}
\alias{convertGenesToSymbols,matrix-method}
\alias{convertGenesToSymbols,data.frame-method}
\alias{convertGenesToSymbols,DataFrame-method}
\alias{convertGenesToSymbols,dgCMatrix-method}
\alias{convertGenesToSymbols,dgTMatrix-method}
\alias{convertGenesToSymbols,SummarizedExperiment-method}
\alias{convertSymbolsToGenes,SummarizedExperiment-method}
\title{Convert Ensembl Identifiers to Gene Symbols}
\usage{
convertGenesToSymbols(object, ...)

convertSymbolsToGenes(object, ...)

\S4method{convertGenesToSymbols}{character}(object, gene2symbol = NULL,
  ...)

\S4method{convertGenesToSymbols}{matrix}(object, ...)

\S4method{convertGenesToSymbols}{data.frame}(object, ...)

\S4method{convertGenesToSymbols}{DataFrame}(object, ...)

\S4method{convertGenesToSymbols}{dgCMatrix}(object, ...)

\S4method{convertGenesToSymbols}{dgTMatrix}(object, ...)

\S4method{convertGenesToSymbols}{SummarizedExperiment}(object)

\S4method{convertSymbolsToGenes}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough to \code{\link[=makeGene2symbolFromEnsembl]{makeGene2symbolFromEnsembl()}}.}

\item{gene2symbol}{\code{data.frame} or \code{NULL}. Gene-to-symbol mappings. If set
\code{NULL}, the function will attempt to download the mappings from Ensembl
automatically.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Identifiers to Gene Symbols
}
\examples{
# character ====
x <- c("ENSMUSG00000000001", "ENSMUSG00000000003")
convertGenesToSymbols(x)

# matrix ====
mat <- matrix(
    data = seq(1L:4L),
    byrow = TRUE,
    nrow = 2L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUSG00000000001", "ENSMUSG00000000003"),
        c("sample_1", "sample_2")
    )
)
print(mat)
mat <- convertGenesToSymbols(mat)
print(mat)
rownames(mat)

# SummarizedExperiment ====
x <- convertGenesToSymbols(rse_bcb)
print(x)

y <- convertSymbolsToGenes(x)
print(y)
}
\seealso{
Other Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
