% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-convertGenesToSymbols.R
\docType{methods}
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertGenesToSymbols,character-method}
\alias{convertGenesToSymbols,data.frame-method}
\alias{convertGenesToSymbols,DataFrame-method}
\alias{convertGenesToSymbols,dgCMatrix-method}
\alias{convertGenesToSymbols,dgTMatrix-method}
\alias{convertGenesToSymbols,matrix-method}
\title{Convert Ensembl Identifier to Gene Symbol}
\usage{
convertGenesToSymbols(object, ...)

\S4method{convertGenesToSymbols}{character}(object, organism, release = NULL,
  quiet = FALSE)

\S4method{convertGenesToSymbols}{data.frame}(object, organism, release = NULL,
  quiet = FALSE)

\S4method{convertGenesToSymbols}{DataFrame}(object, organism, release = NULL,
  quiet = FALSE)

\S4method{convertGenesToSymbols}{dgCMatrix}(object, organism, release = NULL,
  quiet = FALSE)

\S4method{convertGenesToSymbols}{dgTMatrix}(object, organism, release = NULL,
  quiet = FALSE)

\S4method{convertGenesToSymbols}{matrix}(object, organism, release = NULL,
  quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{organism}{\emph{Optional}. Organism name. Normally this argument is
unnecessary and can be left unset. If a count matrix starts with a
FASTA spike-in (e.g. "EGFP"), then automatic genome detection based on the
first gene identifier will fail. In this case, the desired organism must be
manually declared.}

\item{release}{\emph{Optional}. Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Identifier to Gene Symbol
}
\examples{
# character
vec <- c("ENSMUSG00000000001", "ENSMUSG00000000003")
convertGenesToSymbols(vec)

# matrix
mat <- matrix(
    data = seq(1L:4L),
    byrow = TRUE,
    nrow = 2L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUSG00000000001", "ENSMUSG00000000003"),
        c("sample1", "sample2")
    )
)
convertGenesToSymbols(mat)
}
\seealso{
\code{\link[=detectOrganism]{detectOrganism()}}.

Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{symbol2gene}},
  \code{\link{tx2geneFromGFF}}, \code{\link{tx2gene}}
}
