% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-convertGenesToSymbols.R
\docType{methods}
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertGenesToSymbols,character-method}
\alias{convertGenesToSymbols,data.frame-method}
\alias{convertGenesToSymbols,DataFrame-method}
\alias{convertGenesToSymbols,dgCMatrix-method}
\alias{convertGenesToSymbols,dgTMatrix-method}
\alias{convertGenesToSymbols,matrix-method}
\title{Convert Ensembl Identifiers to Gene Symbols}
\usage{
convertGenesToSymbols(object, ...)

\S4method{convertGenesToSymbols}{character}(object, gene2symbol = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL)

\S4method{convertGenesToSymbols}{data.frame}(object, gene2symbol = NULL,
  genomeBuild = NULL, organism = NULL, release = NULL)

\S4method{convertGenesToSymbols}{DataFrame}(object, gene2symbol = NULL,
  genomeBuild = NULL, organism = NULL, release = NULL)

\S4method{convertGenesToSymbols}{dgCMatrix}(object, gene2symbol = NULL,
  genomeBuild = NULL, organism = NULL, release = NULL)

\S4method{convertGenesToSymbols}{dgTMatrix}(object, gene2symbol = NULL,
  genomeBuild = NULL, organism = NULL, release = NULL)

\S4method{convertGenesToSymbols}{matrix}(object, gene2symbol = NULL,
  genomeBuild = NULL, organism = NULL, release = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{gene2symbol}{\emph{Optional.} Gene-to-symbol mappings. If \code{NULL}, will
attempt to download from Ensembl using the desired \code{organism},
\code{genomeBuild}, and \code{release} arguments.}

\item{organism}{\emph{Optional.} Organism name. Normally this argument is
unnecessary and can be left unset. If a count matrix starts with a
FASTA spike-in (e.g. "EGFP"), then automatic genome detection based on the
first gene identifier will fail. In this case, the desired organism must be
manually declared.}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Identifiers to Gene Symbols
}
\examples{
# character ====
genes <- c("ENSMUSG00000000001", "ENSMUSG00000000003")
convertGenesToSymbols(genes)

# matrix ====
mat <- matrix(
    data = seq(1L:4L),
    byrow = TRUE,
    nrow = 2L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUSG00000000001", "ENSMUSG00000000003"),
        c("sample_1", "sample_2")
    )
)
print(mat)
mat <- convertGenesToSymbols(mat)
print(mat)
rownames(mat)
}
\seealso{
Other Gene Functions: \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{tx2geneFromGFF}}
}
