% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertGenesToSymbols-methods.R
\docType{methods}
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertSymbolsToGenes}
\alias{convertGenesToSymbols,character-method}
\alias{convertGenesToSymbols,matrix-method}
\alias{convertGenesToSymbols,sparseMatrix-method}
\alias{convertGenesToSymbols,SummarizedExperiment-method}
\alias{convertSymbolsToGenes,SummarizedExperiment-method}
\title{Convert Ensembl Identifiers to Gene Symbols}
\usage{
convertGenesToSymbols(object, ...)

convertSymbolsToGenes(object, ...)

\S4method{convertGenesToSymbols}{character}(object, gene2symbol = NULL,
  organism = NULL, build = NULL, release = NULL)

\S4method{convertGenesToSymbols}{matrix}(object, gene2symbol = NULL,
  organism = NULL, build = NULL, release = NULL)

\S4method{convertGenesToSymbols}{sparseMatrix}(object,
  gene2symbol = NULL, organism = NULL, build = NULL,
  release = NULL)

\S4method{convertGenesToSymbols}{SummarizedExperiment}(object)

\S4method{convertSymbolsToGenes}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{gene2symbol}{\code{Gene2Symbol} or \code{NULL}. Gene-to-symbol mappings. If set
\code{NULL}, the function will attempt to download the mappings from Ensembl
automatically.}

\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{build}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}
}
\value{
Object with gene IDs converted to names (symbols).
}
\description{
Convert Ensembl Identifiers to Gene Symbols
}
\examples{
gene2symbol <- makeGene2symbolFromEnsembl("Mus musculus")
print(gene2symbol)
geneIDs <- c("ENSMUSG00000000001", "ENSMUSG00000000003")
sampleIDs = paste0("sample", seq_len(2L))

# character ====
x <- convertGenesToSymbols(geneIDs, gene2symbol = gene2symbol)
print(x)

# matrix ====
counts <- matrix(
    data = seq_len(length(geneIDs) * length(sampleIDs)),
    byrow = TRUE,
    nrow = length(geneIDs),
    ncol = length(sampleIDs),
    dimnames = list(geneIDs, sampleIDs)
)
print(counts)
x <- convertGenesToSymbols(counts, gene2symbol = gene2symbol)
print(x)

# SummarizedExperiment ====
x <- convertGenesToSymbols(rse_small)
print(x)
# Interconvert back to gene IDs.
y <- convertSymbolsToGenes(x)
print(y)
}
\seealso{
Other Annotation Functions: \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{makeGRanges}}, \code{\link{makeGene2Symbol}},
  \code{\link{makeTx2Gene}}, \code{\link{organism}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
