% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertGenesToSymbols-methods.R
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertSymbolsToGenes}
\alias{convertGenesToSymbols,character-method}
\alias{convertGenesToSymbols,matrix-method}
\alias{convertGenesToSymbols,Matrix-method}
\alias{convertGenesToSymbols,GRanges-method}
\alias{convertGenesToSymbols,SummarizedExperiment-method}
\alias{convertSymbolsToGenes,SummarizedExperiment-method}
\title{Convert genes to symbols}
\usage{
convertGenesToSymbols(object, ...)

convertSymbolsToGenes(object, ...)

\S4method{convertGenesToSymbols}{character}(object, gene2symbol)

\S4method{convertGenesToSymbols}{matrix}(object, gene2symbol)

\S4method{convertGenesToSymbols}{Matrix}(object, gene2symbol)

\S4method{convertGenesToSymbols}{GRanges}(object)

\S4method{convertGenesToSymbols}{SummarizedExperiment}(object)

\S4method{convertSymbolsToGenes}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneId} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class.
}
\description{
Convert genes to symbols
}
\note{
Updated 2021-02-02.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment
object <- rse

g2s <- Gene2Symbol(object)
print(g2s)
genes <- head(g2s[["geneId"]])
print(genes)

## character ====
x <- convertGenesToSymbols(genes, gene2symbol = g2s)
print(x)

## matrix ====
samples <- head(colnames(object))
counts <- matrix(
    data = seq_len(length(genes) * length(samples)),
    byrow = TRUE,
    nrow = length(genes),
    ncol = length(samples),
    dimnames = list(genes, samples)
)
print(counts)
x <- convertGenesToSymbols(counts, gene2symbol = g2s)
print(x)

## SummarizedExperiment ====
x <- convertGenesToSymbols(rse)
print(x)
## Interconvert back to gene IDs.
y <- convertSymbolsToGenes(x)
print(y)
}
