% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSampleIDsToNames-methods.R
\docType{methods}
\name{convertSampleIDsToNames}
\alias{convertSampleIDsToNames}
\alias{convertSampleIDsToNames,SummarizedExperiment-method}
\alias{convertSampleIDsToNames,SingleCellExperiment-method}
\title{Convert sample identifiers to names}
\usage{
convertSampleIDsToNames(object, ...)

\S4method{convertSampleIDsToNames}{SummarizedExperiment}(object)

\S4method{convertSampleIDsToNames}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending on the method.
}
\description{
Convert sample identifiers to names
}
\note{
Updated 2019-07-28.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment
convertSampleIDsToNames(rse)
}
