% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertSampleIDsToNames-methods.R
\docType{methods}
\name{convertSampleIDsToNames}
\alias{convertSampleIDsToNames}
\alias{convertSampleIDsToNames,SummarizedExperiment-method}
\alias{convertSampleIDsToNames,SingleCellExperiment-method}
\title{Convert Sample Identifiers to Names}
\usage{
convertSampleIDsToNames(object, ...)

\S4method{convertSampleIDsToNames}{SummarizedExperiment}(object)

\S4method{convertSampleIDsToNames}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Convert Sample Identifiers to Names
}
\examples{
data(rse_small)
convertSampleIDsToNames(rse_small)
}
