% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertTranscriptsToGenes-methods.R
\docType{methods}
\name{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes,character-method}
\alias{convertTranscriptsToGenes,matrix-method}
\alias{convertTranscriptsToGenes,data.frame-method}
\alias{convertTranscriptsToGenes,DataFrame-method}
\alias{convertTranscriptsToGenes,dgCMatrix-method}
\alias{convertTranscriptsToGenes,dgTMatrix-method}
\title{Convert Ensembl Transcripts to Genes}
\usage{
convertTranscriptsToGenes(object, ...)

\S4method{convertTranscriptsToGenes}{character}(object, tx2gene = NULL,
  organism = NULL, build = NULL, release = NULL)

\S4method{convertTranscriptsToGenes}{matrix}(object, tx2gene = NULL,
  organism = NULL, build = NULL, release = NULL)

\S4method{convertTranscriptsToGenes}{data.frame}(object, tx2gene = NULL,
  organism = NULL, build = NULL, release = NULL)

\S4method{convertTranscriptsToGenes}{DataFrame}(object, tx2gene = NULL,
  organism = NULL, build = NULL, release = NULL)

\S4method{convertTranscriptsToGenes}{dgCMatrix}(object, tx2gene = NULL,
  organism = NULL, build = NULL, release = NULL)

\S4method{convertTranscriptsToGenes}{dgTMatrix}(object, tx2gene = NULL,
  organism = NULL, build = NULL, release = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough to \code{\link[=makeTx2geneFromEnsembl]{makeTx2geneFromEnsembl()}}.}

\item{tx2gene}{\code{data.frame} or \code{NULL}. Transcript-to-gene mappings. If set
\code{NULL}, the function will attempt to download the mappings from Ensembl
automatically.}

\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{build}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Transcripts to Genes
}
\examples{
# character ====
x <- c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041")
y <- convertTranscriptsToGenes(x)
print(y)

# matrix ====
mat <- matrix(
    data = seq(1L:6L),
    byrow = TRUE,
    nrow = 3L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041"),
        c("sample_1", "sample_2")
    )
)
print(mat)
mat <- convertTranscriptsToGenes(mat)
print(mat)
rownames(mat)
}
\seealso{
Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2gene}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
