% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-convertTranscriptsToGenes.R
\docType{methods}
\name{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes,character-method}
\alias{convertTranscriptsToGenes,matrix-method}
\alias{convertTranscriptsToGenes,data.frame-method}
\alias{convertTranscriptsToGenes,DataFrame-method}
\alias{convertTranscriptsToGenes,dgCMatrix-method}
\alias{convertTranscriptsToGenes,dgTMatrix-method}
\title{Convert Ensembl Transcripts to Genes}
\usage{
convertTranscriptsToGenes(object, ...)

\S4method{convertTranscriptsToGenes}{character}(object, tx2gene, organism,
  genomeBuild = NULL, release = NULL)

\S4method{convertTranscriptsToGenes}{matrix}(object, ...)

\S4method{convertTranscriptsToGenes}{data.frame}(object, ...)

\S4method{convertTranscriptsToGenes}{DataFrame}(object, ...)

\S4method{convertTranscriptsToGenes}{dgCMatrix}(object, ...)

\S4method{convertTranscriptsToGenes}{dgTMatrix}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{tx2gene}{\emph{Optional.} Transcript-to-gene mappings. If missing, the
function will attempt to download mappings from Ensembl according to the
\code{organism}, \code{genomeBuild}, and \code{release} parameters.}

\item{organism}{\emph{Optional.} Organism name. Normally this argument is
unnecessary and can be left unset. If a count matrix starts with a
FASTA spike-in (e.g. "EGFP"), then automatic genome detection based on the
first gene identifier will fail. In this case, the desired organism must be
manually declared.}

\item{genomeBuild}{\emph{Optional.} Genome build assembly name (e.g. "\code{GRCh38}").
If \code{NULL}, defaults to the most recent build available.}

\item{release}{\emph{Optional.} Release version (e.g. \code{90}). If \code{NULL}, defaults
to the most recent release available.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Transcripts to Genes
}
\examples{
# character ====
x <- c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041")
convertTranscriptsToGenes(x)

# matrix ====
mat <- matrix(
    data = seq(1L:6L),
    byrow = TRUE,
    nrow = 3L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041"),
        c("sample_1", "sample_2")
    )
)
print(mat)
mat <- convertTranscriptsToGenes(mat)
print(mat)
rownames(mat)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
