% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-convertTranscriptsToGenes.R
\docType{methods}
\name{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes}
\alias{convertTranscriptsToGenes,character-method}
\alias{convertTranscriptsToGenes,data.frame-method}
\alias{convertTranscriptsToGenes,DataFrame-method}
\alias{convertTranscriptsToGenes,dgCMatrix-method}
\alias{convertTranscriptsToGenes,dgTMatrix-method}
\alias{convertTranscriptsToGenes,matrix-method}
\title{Convert Ensembl Transcripts to Genes}
\usage{
convertTranscriptsToGenes(object, ...)

\S4method{convertTranscriptsToGenes}{character}(object, tx2gene = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL, quiet = FALSE)

\S4method{convertTranscriptsToGenes}{data.frame}(object, tx2gene = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL, quiet = FALSE)

\S4method{convertTranscriptsToGenes}{DataFrame}(object, tx2gene = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL, quiet = FALSE)

\S4method{convertTranscriptsToGenes}{dgCMatrix}(object, tx2gene = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL, quiet = FALSE)

\S4method{convertTranscriptsToGenes}{dgTMatrix}(object, tx2gene = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL, quiet = FALSE)

\S4method{convertTranscriptsToGenes}{matrix}(object, tx2gene = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL, quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genomeBuild}{\emph{Optional}. Check to see if a specific genome build is
supported.}

\item{release}{\emph{Optional}. Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Same class as object.
}
\description{
Convert Ensembl Transcripts to Genes
}
\examples{
# character
c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041") \%>\%
    convertTranscriptsToGenes()

# matrix
matrix(
    data = seq(1L:6L),
    byrow = TRUE,
    nrow = 3L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041"),
        c("sample1", "sample2")
    )
) \%>\%
    convertTranscriptsToGenes()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
