% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUCSCBuildToEnsembl.R
\name{convertUCSCBuildToEnsembl}
\alias{convertUCSCBuildToEnsembl}
\title{Convert UCSC Build to Ensembl}
\usage{
convertUCSCBuildToEnsembl(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character}. Ensembl genome build as the value, UCSC build as the
name. Stops on match failure.
}
\description{
Convert UCSC Build to Ensembl
}
\examples{
from <- c("hg19", "hg38")
to <- convertUCSCBuildToEnsembl(from)
print(to)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{Ensembl2Entrez}},
  \code{\link{Gene2Symbol}}, \code{\link{HGNC2Ensembl}},
  \code{\link{MGI2Ensembl}}, \code{\link{Tx2Gene}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{mapGenes}}
}
\concept{Identifier Mapping Functions}
