% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUCSCBuildToEnsembl.R
\name{convertUCSCBuildToEnsembl}
\alias{convertUCSCBuildToEnsembl}
\title{Convert UCSC Build to Ensembl}
\usage{
convertUCSCBuildToEnsembl(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Character vector containing Ensembl genome build names.
}
\description{
Convert UCSC Build to Ensembl
}
\examples{
convertUCSCBuildToEnsembl(c("hg19", "hg38"))
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}},
  \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
