% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUCSCBuildToEnsembl.R
\name{convertUCSCBuildToEnsembl}
\alias{convertUCSCBuildToEnsembl}
\title{Convert UCSC Build to Ensembl}
\usage{
convertUCSCBuildToEnsembl(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character}. Ensembl genome build as the value, UCSC build as the
name. Stops on match failure.
}
\description{
Convert UCSC Build to Ensembl
}
\examples{
convertUCSCBuildToEnsembl(c("hg19", "hg38"))
}
\seealso{
Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{eggnog}}, \code{\link{emptyRanges}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{geneSynonyms}}, \code{\link{hgnc2ensembl}},
  \code{\link{makeGRanges}}, \code{\link{makeGene2symbol}},
  \code{\link{makeTx2gene}}, \code{\link{mgi2ensembl}},
  \code{\link{organism}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
