% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUCSCBuildToEnsembl.R
\name{convertUCSCBuildToEnsembl}
\alias{convertUCSCBuildToEnsembl}
\title{Convert UCSC Build to Ensembl}
\usage{
convertUCSCBuildToEnsembl(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Character vector containing Ensembl genome build names.
}
\description{
Convert UCSC Build to Ensembl
}
\examples{
convertUCSCBuildToEnsembl(c("hg19", "hg38"))
}
