% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corHeatmap.R
\name{corHeatmap}
\alias{corHeatmap}
\title{Correlation matrix heatmap}
\usage{
corHeatmap(counts, metadata, factor = "group", method = "pearson",
  format = "DESeqTransform", ...)
}
\arguments{
\item{counts}{RNA-Seq counts}

\item{metadata}{Metadata data frame}

\item{factor}{Factor groups to use for clustering color headers}

\item{method}{Correlation coefficient (or covariance) to be computed}

\item{format}{Format of input count data}

\item{...}{Passthrough to \code{pheatmap()}}
}
\value{
Heatmap
}
\description{
Correlation matrix heatmap
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
