% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/correlation-methods.R
\name{correlation}
\alias{correlation}
\alias{correlation,numeric,numeric-method}
\alias{correlation,matrix,missingOrNULL-method}
\alias{correlation,matrix,matrix-method}
\alias{correlation,Matrix,missingOrNULL-method}
\alias{correlation,Matrix,Matrix-method}
\alias{correlation,SummarizedExperiment,missingOrNULL-method}
\alias{correlation,SummarizedExperiment,SummarizedExperiment-method}
\title{Correlation}
\usage{
correlation(x, y, ...)

\S4method{correlation}{numeric,numeric}(x, y, method = c("pearson", "kendall", "spearman"))

\S4method{correlation}{matrix,missingOrNULL}(x, y = NULL, method = c("pearson", "kendall", "spearman"))

\S4method{correlation}{matrix,matrix}(x, y, method = c("pearson", "kendall", "spearman"))

\S4method{correlation}{Matrix,missingOrNULL}(x, y = NULL, method = c("pearson", "kendall", "spearman"))

\S4method{correlation}{Matrix,Matrix}(x, y, method = c("pearson", "kendall", "spearman"))

\S4method{correlation}{SummarizedExperiment,missingOrNULL}(
  x,
  y = NULL,
  i = 1L,
  j = NULL,
  method = c("pearson", "kendall", "spearman")
)

\S4method{correlation}{SummarizedExperiment,SummarizedExperiment}(x, y, i = 1L, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{i}{\code{integer(1)} or \code{character(1)}.
For \code{SummarizedExperiment}, primary assay.}

\item{j}{\code{integer(1)}, \code{character(1)}, or \code{NULL}.
For \code{SummarizedExperiment}, optional secondary assay.
If \code{NULL}, calculates correlation matrix only on the primary assay.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)} or \code{matrix}.
}
\description{
Correlation
}
\note{
Updated 2020-01-20.
}
\examples{
data(correlation, package = "AcidTest")
list <- correlation

## vector ====
x <- list[["vector_x"]]
y <- list[["vector_y"]]

head(x)
head(y)

correlation(x = x, y = y)

## matrix ====
x <- list[["matrix_x"]]
y <- list[["matrix_y"]]

head(x)
head(y)

stats::cor(x)
correlation(x)

stats::cor(x = c(x), y = c(y))
correlation(x = x, y = y)

## SummarizedExperiment ====
x <- list[["SummarizedExperiment_x"]]
y <- list[["SummarizedExperiment_y"]]

correlation(x = x, i = 1L)
correlation(x = x, i = 1L, j = 2L)
correlation(x = x, y = y)
}
\seealso{
\itemize{
\item \code{stats::cor()}.
\item \code{stats::cor.test()}.
\item \code{stats::sd()}.
\item \code{S4Vectors::cor()}.
\item https://stats.stackexchange.com/questions/24980
}
}
