% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDensity.R
\name{countDensity}
\alias{countDensity}
\title{Count density plot}
\usage{
countDensity(counts, metadata)
}
\arguments{
\item{counts}{Counts matrix}

\item{metadata}{Metadata data frame}
}
\value{
Density plot
}
\description{
Count density plot
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{plot}
\keyword{rnaseq}
