% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts,SummarizedExperiment-method}
\alias{counts<-,SummarizedExperiment,matrix-method}
\alias{counts<-,SummarizedExperiment,Matrix-method}
\title{Counts}
\usage{
counts(object, ...)

counts(object, ...) <- value

\S4method{counts}{SummarizedExperiment}(object)

\S4method{counts}{SummarizedExperiment,matrix}(object) <- value

\S4method{counts}{SummarizedExperiment,Matrix}(object) <- value
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
Matrix.
Typically \code{matrix} or \code{sparseMatrix} class.
}
\description{
Count matrix.
}
\note{
For a \code{SummarizedExperiment} object, \code{"counts"} must be explicitly
defined in \code{assayNames}.

Updated 2019-08-06.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
x <- counts(rse)
summary(x)
}
