% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts,SummarizedExperiment-method}
\alias{counts<-,SummarizedExperiment-method}
\title{Counts}
\usage{
counts(object, ...)

counts(object, ...) <- value

\S4method{counts}{SummarizedExperiment}(object)

\S4method{counts}{SummarizedExperiment}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
Matrix.
Typically \code{matrix} or \code{sparseMatrix} class.
}
\description{
Count matrix.
}
\note{
For a \code{SummarizedExperiment} object, \code{"counts"} must be explicitly
defined in \code{assayNames}.
}
\examples{
data(rse, package = "acidtest")
x <- counts(rse)
summary(x)
}
