% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts,SummarizedExperiment-method}
\alias{counts<-,SummarizedExperiment,ANY-method}
\alias{counts<-,SummarizedExperiment,matrix-method}
\title{Counts}
\usage{
\S4method{counts}{SummarizedExperiment}(object)

\S4method{counts}{SummarizedExperiment,ANY}(object) <- value

\S4method{counts}{SummarizedExperiment,matrix}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Matrix. Typically \code{matrix} or \code{sparseMatrix} class.
}
\description{
Count matrix.
}
\note{
For a \code{SummarizedExperiment} object, \code{"counts"} must be explicitly
defined in \code{\link[=assayNames]{assayNames()}}.
}
\examples{
# SummarizedExperiment ====
x <- counts(rse_small)
summary(x)
}
\seealso{
Other Data Functions: \code{\link{aggregate}},
  \code{\link{combine}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{tpm}}
}
\concept{Data Functions}
