% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,SummarizedExperiment-method}
\title{Counts}
\usage{
\S4method{counts}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
}
\description{
Counts
}
\examples{
# SummarizedExperiment ====
x <- counts(rse_small)
summary(x)
}
\seealso{
Other Data Functions: \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{sampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}}
}
\concept{Data Functions}
