% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts,SummarizedExperiment-method}
\alias{counts<-,SummarizedExperiment,ANY-method}
\alias{counts<-,SummarizedExperiment,matrix-method}
\alias{counts<-,SummarizedExperiment,sparseMatrix-method}
\title{Counts}
\usage{
\S4method{counts}{SummarizedExperiment}(object)

\S4method{counts}{SummarizedExperiment,ANY}(object) <- value

\S4method{counts}{SummarizedExperiment,matrix}(object) <- value

\S4method{counts}{SummarizedExperiment,sparseMatrix}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Counts matrix. Typically \code{matrix} or \code{sparseMatrix} class.
}
\description{
Counts
}
\note{
For a \code{SummarizedExperiment} object, \code{"counts"} must be explicitly
defined in \code{\link[=assayNames]{assayNames()}}.
}
\examples{
# SummarizedExperiment ====
x <- counts(rse_small)
summary(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topCells}},
  \code{\link{tpm}}, \code{\link{tx2gene}}
}
\concept{Data Functions}
