% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsPerGene.R
\name{countsPerGene}
\alias{countsPerGene}
\title{Counts per gene plot}
\usage{
countsPerGene(counts, metadata, flip = TRUE, print = TRUE)
}
\arguments{
\item{counts}{Counts matrix}

\item{metadata}{Metadata data frame}

\item{flip}{Flip x and y axes}

\item{print}{Print plot}
}
\value{
Boxplot
}
\description{
Counts per gene plot
}
\author{
Michael Steinbaugh
}
