% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectDirs.R
\name{createProjectDirs}
\alias{createProjectDirs}
\title{Create Project Directory Structure}
\usage{
createProjectDirs()
}
\value{
No value.
}
\description{
Create necessary directory structure for an
\href{http://rmarkdown.rstudio.com}{RMarkdown} report in
\href{https://www.rstudio.com}{RStudio}.
}
