% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruft.R
\name{cruft}
\alias{cruft}
\title{Cruft removal for tibbles.}
\usage{
cruft(crufty)
}
\arguments{
\item{crufty}{Crufty \code{tibble} with leading/trailing spaces, dashes.}
}
\value{
A reformatted, clean \code{tibble}
}
\description{
Cruft removal for tibbles.
}

