% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruft.R
\name{cruft}
\alias{cruft}
\title{Cruft removal for data frame cells}
\usage{
cruft(data)
}
\arguments{
\item{data}{Data with leading/trailing spaces/commas, empty cells.}
}
\value{
A reformatted, clean \code{tibble}
}
\description{
Cruft removal for data frame cells
}

