% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruft.R
\name{cruft}
\alias{cruft}
\title{Cruft removal for data frame cells}
\usage{
cruft(df)
}
\arguments{
\item{df}{Messy data frame -- leading/trailing spaces, dashes}
}
\value{
A reformatted, clean data frame
}
\description{
Cruft removal for data frame cells
}

