% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/davidQuery.R
\name{davidQuery}
\alias{davidQuery}
\title{RDAVIDWebService query}
\usage{
davidQuery(email, foregroundGenes, backgroundGenes, idType)
}
\arguments{
\item{email}{Registered email address}

\item{foregroundGenes}{Gene ID vector}

\item{backgroundGenes}{Gene ID vector}

\item{idType}{Gene identifier type}
}
\description{
RDAVIDWebService query
}

