% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{aggregateFeatures}
\alias{aggregateReplicates}
\alias{aggregateSamples}
\alias{broadClass}
\alias{checkClasses}
\alias{eggnog}
\alias{ensembl2entrez}
\alias{gene2symbol}
\alias{grepString}
\alias{hgnc2gene}
\alias{hgnc2ensembl}
\alias{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGTF}
\alias{makeTx2geneFromEnsembl}
\alias{makeTx2geneFromGFF}
\alias{makeTx2geneFromGTF}
\alias{mgi2gene}
\alias{mgi2ensembl}
\alias{panther}
\alias{setArgsToDoCall}
\alias{tx2gene}
\alias{assertHasRownames}
\alias{theme_midnight}
\alias{theme_paperwhite}
\alias{tx2geneFromGFF}
\alias{separatorBar}
\alias{assertFormalInterestingGroups}
\alias{assertIsAHeaderLevel}
\alias{assertIsAStringOrNULL}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsColorScaleDiscreteOrNULL}
\alias{assertIsFillScaleDiscreteOrNULL}
\alias{assertIsGene2symbol}
\alias{assertIsHexColorFunctionOrNULL}
\alias{assertIsImplicitInteger}
\alias{assertIsTx2gene}
\alias{assertAreGeneAnnotations}
\alias{assertFormalGene2symbol}
\title{Defunct functions}
\usage{
aggregateFeatures(...)

aggregateReplicates(...)

aggregateSamples(...)

broadClass(...)

checkClasses(...)

eggnog(...)

ensembl2entrez(...)

gene2symbol(...)

grepString(...)

hgnc2gene(...)

hgnc2ensembl(...)

makeGene2symbolFromEnsembl(...)

makeGene2symbolFromGFF(...)

makeGene2symbolFromGTF(...)

makeTx2geneFromEnsembl(...)

makeTx2geneFromGFF(...)

makeTx2geneFromGTF(...)

mgi2gene(...)

mgi2ensembl(...)

panther(...)

setArgsToDoCall(...)

tx2gene(...)

assertHasRownames(...)

theme_midnight(...)

theme_paperwhite(...)

tx2geneFromGFF(...)

separatorBar(...)

assertFormalInterestingGroups(...)

assertIsAHeaderLevel(...)

assertIsAStringOrNULL(...)

assertIsAnImplicitInteger(...)

assertIsAnImplicitIntegerOrNULL(...)

assertIsColorScaleDiscreteOrNULL(...)

assertIsFillScaleDiscreteOrNULL(...)

assertIsGene2symbol(...)

assertIsHexColorFunctionOrNULL(...)

assertIsImplicitInteger(...)

assertIsTx2gene(...)

assertAreGeneAnnotations(...)

assertFormalGene2symbol(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\examples{
a <- function(...) {
    .Defunct("b")
}

withCallingHandlers(
    expr = tryCatch(
        expr = a(1L),
        error = function(e) {
            print(e)
            invisible()
        }
    )
)
}
\seealso{
\link[base:Defunct]{Defunct}.
}
\keyword{internal}
