% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{assertIsGFF}
\alias{parseGFFAttributes}
\alias{dynamicPlotlist}
\alias{synonyms}
\alias{assertIsCharacterOrNULL}
\alias{assertIsGene2symbol}
\alias{assertIsTx2gene}
\alias{assertIsDataFrameOrNULL}
\alias{broadClass}
\alias{separatorBar}
\alias{setArgsToDoCall}
\title{Defunct Functions}
\usage{
assertIsGFF(...)

parseGFFAttributes(...)

dynamicPlotlist(...)

synonyms(...)

assertIsCharacterOrNULL(object, ...)

assertIsGene2symbol(...)

assertIsTx2gene(...)

assertIsDataFrameOrNULL(object, ...)

broadClass(object, ...)

separatorBar(...)

setArgsToDoCall(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\keyword{internal}
