% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{parseGFFAttributes}
\alias{dynamicPlotlist}
\alias{synonyms}
\alias{broadClass}
\alias{separatorBar}
\alias{setArgsToDoCall}
\title{Defunct Functions}
\usage{
parseGFFAttributes(...)

dynamicPlotlist(...)

synonyms(...)

broadClass(object, ...)

separatorBar(...)

setArgsToDoCall(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{.Defunct()}.
}
\description{
Defunct Functions
}
\keyword{internal}
