% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{theme_midnight}
\alias{theme_paperwhite}
\alias{tx2geneFromGFF}
\alias{separatorBar}
\alias{relevelColData}
\alias{relevelRowData}
\alias{relevelRowRanges}
\alias{flatFiles}
\alias{flatFiles,SummarizedExperiment-method}
\title{Defunct functions}
\usage{
theme_midnight(...)

theme_paperwhite(...)

tx2geneFromGFF(...)

separatorBar(...)

relevelColData(...)

relevelRowData(...)

relevelRowRanges(...)

\S4method{flatFiles}{SummarizedExperiment}(object)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\examples{
a <- function(...) {
    .Defunct("b")
}

withCallingHandlers(
    expr = tryCatch(
        expr = a(1L),
        error = function(e) {
            print(e)
            invisible()
        }
    )
)
}
\seealso{
\link[base:Defunct]{Defunct}.
}
\keyword{internal}
