% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{summarizeRows}
\alias{wash}
\alias{packageSE}
\alias{prepareSE}
\alias{metadataTable}
\alias{pct}
\alias{fc2lr}
\alias{lr2fc}
\alias{comp}
\alias{revcomp}
\alias{symbol2gene}
\alias{tx2gene}
\alias{assertIsGFF}
\alias{parseGFFAttributes}
\alias{dynamicPlotlist}
\title{Defunct Functions}
\usage{
summarizeRows(...)

wash(...)

packageSE(...)

prepareSE(...)

metadataTable(...)

pct(...)

fc2lr(...)

lr2fc(...)

comp()

revcomp()

symbol2gene()

tx2gene(...)

assertIsGFF(...)

parseGFFAttributes(...)

dynamicPlotlist(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
