% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{fc2lr}
\alias{lr2fc}
\alias{comp}
\alias{revcomp}
\alias{symbol2gene}
\alias{assertIsGFF}
\alias{parseGFFAttributes}
\alias{dynamicPlotlist}
\alias{synonyms}
\alias{assertIsGene2symbol}
\alias{assertIsTx2gene}
\alias{broadClass}
\alias{separatorBar}
\alias{setArgsToDoCall}
\title{Defunct Functions}
\usage{
fc2lr(...)

lr2fc(...)

comp(...)

revcomp(...)

symbol2gene(...)

assertIsGFF(...)

parseGFFAttributes(...)

dynamicPlotlist(...)

synonyms(...)

assertIsGene2symbol(...)

assertIsTx2gene(...)

broadClass(object, ...)

separatorBar(...)

setArgsToDoCall(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\keyword{internal}
