% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{defunct}
\alias{defunct}
\alias{fc2lr}
\alias{lr2fc}
\alias{comp}
\alias{revcomp}
\alias{symbol2gene}
\alias{tx2gene}
\alias{assertIsGFF}
\alias{parseGFFAttributes}
\alias{dynamicPlotlist}
\alias{broadClass,data.frame-method}
\alias{broadClass,DataFrame-method}
\alias{synonyms}
\title{Defunct Functions}
\usage{
fc2lr(...)

lr2fc(...)

comp(...)

revcomp(...)

symbol2gene(...)

tx2gene(...)

assertIsGFF(...)

parseGFFAttributes(...)

dynamicPlotlist(...)

\S4method{broadClass}{data.frame}(object)

\S4method{broadClass}{DataFrame}(object)

synonyms(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
