% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{aggregateFeatures}
\alias{aggregateFeatures,ANY-method}
\alias{aggregateReplicates}
\alias{aggregateReplicates,ANY-method}
\alias{aggregateSamples}
\alias{aggregateSamples,ANY-method}
\alias{checkClasses}
\alias{eggnog}
\alias{ensembl2entrez}
\alias{fixNA}
\alias{gene2symbol}
\alias{grepString}
\alias{hgnc2gene}
\alias{hgnc2ensembl}
\alias{initializeDirectory}
\alias{kables}
\alias{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGTF}
\alias{makeTx2geneFromEnsembl}
\alias{makeTx2geneFromGFF}
\alias{makeTx2geneFromGTF}
\alias{mgi2gene}
\alias{mgi2ensembl}
\alias{panther}
\alias{readFileByExtension}
\alias{readGFF}
\alias{readGTF}
\alias{readJSON}
\alias{readYAML}
\alias{tx2gene}
\alias{markdownPlotlist}
\alias{mdPlotlist}
\alias{assertAreGeneAnnotations}
\alias{assertFormalGene2symbol}
\alias{assertFormalInterestingGroups}
\alias{assertIsAHeaderLevel}
\alias{assertIsAStringOrNULL}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsColorScaleDiscreteOrNULL}
\alias{assertIsFillScaleDiscreteOrNULL}
\alias{assertIsGene2symbol}
\alias{assertIsHexColorFunctionOrNULL}
\alias{assertIsImplicitInteger}
\alias{assertIsTx2gene}
\title{Deprecated functions}
\usage{
aggregateFeatures(object, ...)

\S4method{aggregateFeatures}{ANY}(object, ...)

aggregateReplicates(object, ...)

\S4method{aggregateReplicates}{ANY}(object, ...)

aggregateSamples(object, ...)

\S4method{aggregateSamples}{ANY}(object, ...)

checkClasses(...)

eggnog(...)

ensembl2entrez(...)

fixNA(...)

gene2symbol(...)

grepString(...)

hgnc2gene(...)

hgnc2ensembl(...)

initializeDirectory(...)

kables(...)

makeGene2symbolFromEnsembl(...)

makeGene2symbolFromGFF(...)

makeGene2symbolFromGTF(...)

makeTx2geneFromEnsembl(...)

makeTx2geneFromGFF(...)

makeTx2geneFromGTF(...)

mgi2gene(...)

mgi2ensembl(...)

panther(...)

readFileByExtension(...)

readGFF(...)

readGTF(...)

readJSON(...)

readYAML(...)

tx2gene(...)

markdownPlotlist(plotlist, ...)

mdPlotlist(plotlist, ...)

assertAreGeneAnnotations(x)

assertFormalGene2symbol(x, genes, gene2symbol)

assertFormalInterestingGroups(x, interestingGroups)

assertIsAHeaderLevel(x)

assertIsAStringOrNULL(x)

assertIsAnImplicitInteger(x)

assertIsAnImplicitIntegerOrNULL(x)

assertIsColorScaleDiscreteOrNULL(x)

assertIsFillScaleDiscreteOrNULL(x)

assertIsGene2symbol(x)

assertIsHexColorFunctionOrNULL(x)

assertIsImplicitInteger(x)

assertIsTx2gene(x)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{plotlist}{\code{list}.
List containing \code{ggplot} objects.}

\item{x}{Object.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
