% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R,
%   R/uniteInterestingGroups-methods.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{fixNA}
\alias{initializeDirectory}
\alias{kables}
\alias{readFileByExtension}
\alias{readGFF}
\alias{readGTF}
\alias{readJSON}
\alias{readYAML}
\alias{markdownPlotlist}
\alias{mdPlotlist}
\alias{assertAreGeneAnnotations}
\alias{assertFormalGene2symbol}
\alias{assertFormalInterestingGroups}
\alias{assertIsAHeaderLevel}
\alias{assertIsAStringOrNULL}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsColorScaleDiscreteOrNULL}
\alias{assertIsFillScaleDiscreteOrNULL}
\alias{assertIsGene2symbol}
\alias{assertIsHexColorFunctionOrNULL}
\alias{assertIsImplicitInteger}
\alias{assertIsTx2gene}
\alias{relevelColData}
\alias{relevelRowData}
\alias{relevelRowRanges}
\alias{uniteInterestingGroups,data.frame-method}
\title{Deprecated functions}
\usage{
fixNA(...)

initializeDirectory(...)

kables(...)

readFileByExtension(...)

readGFF(...)

readGTF(...)

readJSON(...)

readYAML(...)

markdownPlotlist(plotlist, ...)

mdPlotlist(plotlist, ...)

assertAreGeneAnnotations(x)

assertFormalGene2symbol(x, genes, gene2symbol)

assertFormalInterestingGroups(x, interestingGroups)

assertIsAHeaderLevel(x)

assertIsAStringOrNULL(x)

assertIsAnImplicitInteger(x)

assertIsAnImplicitIntegerOrNULL(x)

assertIsColorScaleDiscreteOrNULL(x)

assertIsFillScaleDiscreteOrNULL(x)

assertIsGene2symbol(x)

assertIsHexColorFunctionOrNULL(x)

assertIsImplicitInteger(x)

assertIsTx2gene(x)

relevelColData(object)

relevelRowData(object)

relevelRowRanges(object)
}
\arguments{
\item{...}{Additional arguments.}

\item{plotlist}{\code{list}.
List containing \code{ggplot} objects.}

\item{x}{Object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{object}{Object.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\examples{
a <- function(...) {
    .Deprecated("b")
    b(...)
}

b <- function(x) {
    x + 1L
}

withCallingHandlers(
    expr = a(1L),
    warning = function(w) {
        print(w)
        invokeRestart("muffleWarning")
    }
)
}
\seealso{
\link[base:Deprecated]{Deprecated}.
}
\keyword{internal}
