% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/separator.R,
%   R/uniteInterestingGroups-methods.R
\docType{data}
\name{deprecated}
\alias{deprecated}
\alias{fixNA}
\alias{initializeDirectory}
\alias{kables}
\alias{readFileByExtension}
\alias{readGFF}
\alias{readGTF}
\alias{readJSON}
\alias{readYAML}
\alias{markdownPlotlist}
\alias{mdPlotlist}
\alias{separatorBar}
\alias{uniteInterestingGroups,data.frame-method}
\title{Deprecated functions}
\format{An object of class \code{character} of length 1.}
\usage{
fixNA(...)

initializeDirectory(...)

kables(...)

readFileByExtension(...)

readGFF(...)

readGTF(...)

readJSON(...)

readYAML(...)

markdownPlotlist(plotlist, ...)

mdPlotlist(plotlist, ...)

separatorBar
}
\arguments{
\item{...}{Additional arguments.}

\item{plotlist}{\code{list}.
List containing \code{ggplot} objects.}

\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\examples{
a <- function(...) {
    .Deprecated("b")
    b(...)
}

b <- function(x) {
    x + 1L
}

withCallingHandlers(
    expr = a(1L),
    warning = function(w) {
        print(w)
        invokeRestart("muffleWarning")
    }
)
}
\seealso{
\link[base:Deprecated]{Deprecated}.
}
\keyword{datasets}
\keyword{internal}
