% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{ensembl}
\alias{gene2symbolFromGFF}
\alias{gene2symbolFromGTF}
\alias{sanitizeAnnotable}
\alias{tx2geneFromGFF}
\alias{tx2geneFromGTF}
\alias{multiassignAsNewEnvir}
\alias{geomean}
\alias{midnightTheme}
\alias{paperwhiteTheme}
\alias{aggregateFeatures}
\alias{aggregateFeatures,ANY-method}
\alias{aggregateReplicates}
\alias{aggregateReplicates,ANY-method}
\alias{aggregateSamples}
\alias{aggregateSamples,ANY-method}
\alias{assertFormalGene2symbol}
\alias{checkClasses}
\alias{eggnog}
\alias{ensembl2entrez}
\alias{fixNA}
\alias{gene2symbol}
\alias{grepString}
\alias{hgnc2gene}
\alias{hgnc2ensembl}
\alias{initializeDirectory}
\alias{kables}
\alias{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGTF}
\alias{makeTx2geneFromEnsembl}
\alias{makeTx2geneFromGFF}
\alias{makeTx2geneFromGTF}
\alias{mgi2gene}
\alias{mgi2ensembl}
\alias{panther}
\alias{readFileByExtension}
\alias{readGFF}
\alias{readGTF}
\alias{readJSON}
\alias{readYAML}
\alias{tx2gene}
\alias{markdownPlotlist}
\alias{mdPlotlist}
\title{Deprecated Functions}
\usage{
ensembl(...)

gene2symbolFromGFF(...)

gene2symbolFromGTF(...)

sanitizeAnnotable(...)

tx2geneFromGFF(...)

tx2geneFromGTF(...)

multiassignAsNewEnvir(...)

geomean(...)

midnightTheme(...)

paperwhiteTheme(...)

aggregateFeatures(object, ...)

\S4method{aggregateFeatures}{ANY}(object, ...)

aggregateReplicates(object, ...)

\S4method{aggregateReplicates}{ANY}(object, ...)

aggregateSamples(object, ...)

\S4method{aggregateSamples}{ANY}(object, ...)

assertFormalGene2symbol(...)

checkClasses(...)

eggnog(...)

ensembl2entrez(...)

fixNA(...)

gene2symbol(...)

grepString(...)

hgnc2gene(...)

hgnc2ensembl(...)

initializeDirectory(...)

kables(...)

makeGene2symbolFromEnsembl(...)

makeGene2symbolFromGFF(...)

makeGene2symbolFromGTF(...)

makeTx2geneFromEnsembl(...)

makeTx2geneFromGFF(...)

makeTx2geneFromGTF(...)

mgi2gene(...)

mgi2ensembl(...)

panther(...)

readFileByExtension(...)

readGFF(...)

readGTF(...)

readJSON(...)

readYAML(...)

tx2gene(...)

markdownPlotlist(plotlist, ...)

mdPlotlist(plotlist, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}

\item{plotlist}{\code{list} containing \code{ggplot} objects.}
}
\value{
\code{.Deprecated()}.
}
\description{
Deprecated Functions
}
\keyword{internal}
