% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R,
%   R/uniteInterestingGroups-methods.R
\name{deprecated}
\alias{deprecated}
\alias{readFileByExtension}
\alias{readGFF}
\alias{readGTF}
\alias{readJSON}
\alias{readYAML}
\alias{readSampleData}
\alias{readTx2Gene}
\alias{coerceS4ToList}
\alias{matchEnsemblReleaseToURL}
\alias{uniteInterestingGroups,data.frame-method}
\title{Deprecated functions}
\usage{
readFileByExtension(...)

readGFF(...)

readGTF(...)

readJSON(...)

readYAML(...)

readSampleData(...)

readTx2Gene(...)

coerceS4ToList(...)

matchEnsemblReleaseToURL(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\examples{
a <- function(...) {
    .Deprecated("b")
    b(...)
}

b <- function(x) {
    x + 1L
}

withCallingHandlers(
    expr = a(1L),
    warning = function(w) {
        print(w)
        invokeRestart("muffleWarning")
    }
)
}
\seealso{
\link[base:Deprecated]{Deprecated}.
}
\keyword{internal}
