% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-annotable.R, R/deprecated.R
\docType{methods}
\name{annotable}
\alias{annotable}
\alias{annotable,character-method}
\alias{annotable,data.frame-method}
\alias{deprecated}
\alias{summarizeRows}
\alias{wash}
\alias{packageSE}
\alias{prepareSE}
\alias{metadataTable}
\alias{pct}
\alias{fc2lr}
\alias{lr2fc}
\alias{comp}
\alias{revcomp}
\alias{symbol2gene}
\alias{checkAnnotable}
\alias{checkGene2symbol}
\alias{checkTx2gene}
\alias{assertFormalHeaderLevel}
\alias{assertFormalColorFunction}
\alias{initializeDir}
\alias{assertIsAnnotable}
\alias{sanitizeAnnotable}
\title{Defunct or Deprecated Functions}
\usage{
annotable(object, ...)

\S4method{annotable}{character}(object, ...)

\S4method{annotable}{data.frame}(object)

summarizeRows(...)

wash(...)

packageSE(...)

prepareSE(...)

metadataTable(...)

pct(...)

fc2lr(...)

lr2fc(...)

comp()

revcomp()

symbol2gene()

checkAnnotable(...)

checkGene2symbol(...)

checkTx2gene(...)

assertFormalHeaderLevel(...)

assertFormalColorFunction(...)

initializeDir(...)

assertIsAnnotable(...)

sanitizeAnnotable(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}} or \code{\link[=.Deprecated]{.Deprecated()}} calls.
}
\description{
Defunct or Deprecated Functions

#' @rdname deprecated
#' @export
sampleDirs <- function(...) {
.Defunct()
}
}
\keyword{internal}
