% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{createNewProject}
\alias{createNewProject}
\alias{dotNotation}
\alias{grepToString}
\alias{loadDataRaw}
\alias{toStringSummarize}
\title{Deprecated Functions}
\usage{
createNewProject(...)

dotNotation(...)

grepToString()

loadDataRaw(...)

toStringSummarize(...)
}
\arguments{
\item{...}{Additional parameters.}
}
\description{
Deprecated Functions
}
\keyword{internal}
