% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{save_data}
\alias{write_counts}
\alias{dotNotation}
\alias{grepToString}
\alias{loadDataRaw}
\alias{toStringSummarize}
\alias{assignAsNewEnv}
\alias{getObjsFromDots}
\alias{readDataRaw}
\alias{sanitizeNames}
\alias{saveDataRaw}
\alias{wash}
\title{Deprecated Functions}
\usage{
save_data(...)

write_counts(...)

dotNotation(...)

grepToString(...)

loadDataRaw(...)

toStringSummarize(...)

assignAsNewEnv(...)

getObjsFromDots(...)

readDataRaw(...)

sanitizeNames(...)

saveDataRaw(...)

wash(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
Varies by function.
}
\description{
Deprecated Functions
}
\examples{
\dontrun{
# snake_case variants used in RMarkdown
save_data()
write_counts()

# 0.0.19
dotNotation()
grepToString()
loadDataRaw()
toStringSummarize()

# 0.0.20
assignAsNewEnv()
getObjsFromDots()
readDataRaw()
sanitizeNames()
saveDataRaw()
wash()
}
}
\keyword{internal}
