% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{packageSE}
\alias{prepareSE}
\title{Deprecated Functions}
\usage{
packageSE(...)

prepareSE(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
Deprecation warning.
}
\description{
Deprecated Functions
}
\keyword{internal}
