% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{summarizeRows}
\alias{wash}
\alias{packageSE}
\alias{prepareSE}
\alias{metadataTable}
\alias{pct}
\alias{sampleDirs}
\alias{fc2lr}
\alias{lr2fc}
\title{Deprecated Functions}
\usage{
summarizeRows()

wash()

packageSE(...)

prepareSE(...)

metadataTable()

pct(...)

sampleDirs()

fc2lr(...)

lr2fc(...)
}
\arguments{
\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Soft deprecation to new functions.
}
\description{
Deprecated Functions
}
\keyword{internal}
