% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{summarizeRows}
\alias{wash}
\alias{packageSE}
\alias{prepareSE}
\alias{metadataTable}
\alias{pct}
\alias{fc2lr}
\alias{lr2fc}
\alias{comp}
\alias{revcomp}
\alias{symbol2gene}
\alias{checkAnnotable}
\alias{checkGene2symbol}
\alias{checkTx2gene}
\alias{assertFormalHeaderLevel}
\alias{assertFormalColorFunction}
\alias{initializeDir}
\alias{annotable}
\alias{assertIsAnnotable}
\alias{ensembl}
\alias{gene2symbolFromGFF}
\alias{gene2symbolFromGTF}
\alias{sanitizeAnnotable}
\alias{tx2geneFromGFF}
\alias{tx2geneFromGTF}
\title{Defunct or Deprecated Functions}
\usage{
summarizeRows(...)

wash(...)

packageSE(...)

prepareSE(...)

metadataTable(...)

pct(...)

fc2lr(...)

lr2fc(...)

comp()

revcomp()

symbol2gene()

checkAnnotable(...)

checkGene2symbol(...)

checkTx2gene(...)

assertFormalHeaderLevel(...)

assertFormalColorFunction(...)

initializeDir(...)

annotable(...)

assertIsAnnotable(...)

ensembl(...)

gene2symbolFromGFF(...)

gene2symbolFromGTF(...)

sanitizeAnnotable(...)

tx2geneFromGFF(...)

tx2geneFromGTF(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}} or \code{\link[=.Deprecated]{.Deprecated()}} calls.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
