% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{checkAnnotable}
\alias{checkGene2symbol}
\alias{checkTx2gene}
\alias{assertFormalHeaderLevel}
\alias{assertFormalColorFunction}
\alias{initializeDir}
\alias{assertIsAnnotable}
\alias{ensembl}
\alias{gene2symbolFromGFF}
\alias{gene2symbolFromGTF}
\alias{sanitizeAnnotable}
\alias{tx2geneFromGFF}
\alias{tx2geneFromGTF}
\alias{multiassignAsNewEnvir}
\alias{geomean}
\alias{midnightTheme}
\alias{paperwhiteTheme}
\alias{assertIsURL}
\alias{assertIsCharacterOrNULL}
\alias{assertIsDataFrameOrNULL}
\alias{fixNA}
\alias{setArgsToDoCall}
\title{Deprecated Functions}
\usage{
checkAnnotable(...)

checkGene2symbol(...)

checkTx2gene(...)

assertFormalHeaderLevel(...)

assertFormalColorFunction(...)

initializeDir(...)

assertIsAnnotable(...)

ensembl(...)

gene2symbolFromGFF(...)

gene2symbolFromGTF(...)

sanitizeAnnotable(...)

tx2geneFromGFF(...)

tx2geneFromGTF(...)

multiassignAsNewEnvir(...)

geomean(...)

midnightTheme(...)

paperwhiteTheme(...)

assertIsURL(...)

assertIsCharacterOrNULL(object, ...)

assertIsDataFrameOrNULL(object, ...)

fixNA(...)

setArgsToDoCall(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
