% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{packageSE}
\alias{prepareSE}
\alias{metadataTable}
\alias{.dataVersions}
\alias{.logFile}
\alias{.programs}
\alias{.sampleYAML}
\alias{.sampleYAMLMetadata}
\alias{.sampleYAMLMetrics}
\alias{pct}
\alias{sampleDirs}
\title{Deprecated Functions}
\usage{
packageSE(...)

prepareSE(...)

metadataTable()

.dataVersions()

.logFile()

.programs()

.sampleYAML(yaml, keys)

.sampleYAMLMetadata()

.sampleYAMLMetrics(yaml)

pct(...)

sampleDirs()
}
\arguments{
\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Soft deprecation to new functions.
}
\description{
Deprecated Functions
}
\keyword{internal}
