% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{summarizeRows}
\alias{wash}
\alias{packageSE}
\alias{prepareSE}
\alias{metadataTable}
\alias{pct}
\alias{sampleDirs}
\alias{fc2lr}
\alias{lr2fc}
\alias{comp}
\alias{revcomp}
\alias{symbol2gene}
\alias{checkAnnotable}
\alias{checkGene2symbol}
\alias{checkTx2gene}
\alias{assertFormalHeaderLevel}
\alias{assert_formal_header_level}
\alias{assertFormalColorFunction}
\alias{assert_formal_color_function}
\title{Deprecated Functions}
\usage{
summarizeRows()

wash()

packageSE()

prepareSE()

metadataTable()

pct(...)

sampleDirs()

fc2lr(...)

lr2fc(...)

comp()

revcomp()

symbol2gene()

checkAnnotable(...)

checkGene2symbol(...)

checkTx2gene(...)

assertFormalHeaderLevel(x, severity = "stop")

assert_formal_header_level(x, severity = "stop")

assertFormalColorFunction(x, severity = "stop")

assert_formal_color_function(x, severity = "stop")
}
\arguments{
\item{...}{Additional arguments.}

\item{x}{Primary object.}
}
\value{
Soft deprecation to new functions.
}
\description{
Deprecated Functions
}
\keyword{internal}
