% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{summarizeRows}
\alias{wash}
\alias{packageSE}
\alias{prepareSE}
\alias{metadataTable}
\alias{pct}
\alias{sampleDirs}
\alias{fc2lr}
\alias{lr2fc}
\alias{comp}
\alias{revcomp}
\alias{symbol2gene}
\alias{checkAnnotable}
\alias{checkGene2symbol}
\alias{checkTx2gene}
\alias{assertFormalHeaderLevel}
\alias{assertFormalColorFunction}
\alias{initializeDir}
\alias{annotable}
\title{Deprecated Functions}
\usage{
summarizeRows()

wash()

packageSE()

prepareSE()

metadataTable()

pct(...)

sampleDirs()

fc2lr(...)

lr2fc(...)

comp()

revcomp()

symbol2gene()

checkAnnotable(...)

checkGene2symbol(...)

checkTx2gene(...)

assertFormalHeaderLevel(...)

assertFormalColorFunction(...)

initializeDir(...)

annotable()
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
Soft deprecation to new functions.
}
\description{
Deprecated Functions
}
\keyword{internal}
