% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{createNewProject}
\alias{createNewProject}
\alias{dotNotation}
\alias{get_objs_from_dots}
\alias{grepToString}
\alias{loadDataRaw}
\alias{save_data}
\alias{toStringSummarize}
\alias{write_counts}
\title{Deprecated Functions}
\usage{
createNewProject(...)

dotNotation(...)

get_objs_from_dots(...)

grepToString()

loadDataRaw(...)

save_data(...)

toStringSummarize(...)

write_counts(...)
}
\arguments{
\item{...}{Additional parameters.}
}
\description{
Deprecated Functions
}
\keyword{internal}
