% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{dotNotation}
\alias{grepToString}
\alias{loadDataRaw}
\alias{toStringSummarize}
\alias{assignAsNewEnv}
\alias{getObjsFromDots}
\alias{sanitizeNames}
\alias{save_data}
\alias{write_counts}
\title{Deprecated Functions}
\usage{
dotNotation()

grepToString()

loadDataRaw()

toStringSummarize()

assignAsNewEnv()

getObjsFromDots()

sanitizeNames()

save_data(...)

write_counts(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
Varies by function.
}
\description{
Deprecated Functions
}
\keyword{internal}
