% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{checkAnnotable}
\alias{checkGene2symbol}
\alias{checkTx2gene}
\alias{assertFormalHeaderLevel}
\alias{assertFormalColorFunction}
\alias{assertIsAnnotable}
\alias{ensembl}
\alias{gene2symbolFromGFF}
\alias{gene2symbolFromGTF}
\alias{sanitizeAnnotable}
\alias{tx2geneFromGFF}
\alias{tx2geneFromGTF}
\alias{multiassignAsNewEnvir}
\alias{geomean}
\alias{midnightTheme}
\alias{paperwhiteTheme}
\alias{assertIsURL}
\alias{aggregateFeatures}
\alias{aggregateReplicates}
\alias{aggregateSamples}
\alias{assertIsAHeaderLevel}
\alias{assertIsCharacterOrNULL}
\alias{assertIsDataFrameOrNULL}
\alias{detectOrganism}
\alias{fixNA}
\alias{flatFiles}
\alias{hgnc2gene}
\alias{kables}
\alias{mgi2gene}
\alias{readFileByExtension}
\alias{readGFF}
\alias{readGTF}
\alias{readJSON}
\alias{readYAML}
\title{Deprecated Functions}
\usage{
checkAnnotable(...)

checkGene2symbol(...)

checkTx2gene(...)

assertFormalHeaderLevel(...)

assertFormalColorFunction(...)

assertIsAnnotable(...)

ensembl(...)

gene2symbolFromGFF(...)

gene2symbolFromGTF(...)

sanitizeAnnotable(...)

tx2geneFromGFF(...)

tx2geneFromGTF(...)

multiassignAsNewEnvir(...)

geomean(...)

midnightTheme(...)

paperwhiteTheme(...)

assertIsURL(...)

aggregateFeatures(...)

aggregateReplicates(...)

aggregateSamples(...)

assertIsAHeaderLevel(...)

assertIsCharacterOrNULL(object, ...)

assertIsDataFrameOrNULL(object, ...)

detectOrganism(object, ...)

fixNA(...)

flatFiles(object)

hgnc2gene(...)

kables(...)

mgi2gene(...)

readFileByExtension(...)

readGFF(...)

readGTF(...)

readJSON(...)

readYAML(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
