% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-detectHPC.R
\docType{methods}
\name{detectHPC}
\alias{detectHPC}
\alias{detectHPC,missing-method}
\title{Detect HPC Environment}
\usage{
detectHPC(object, ...)

\S4method{detectHPC}{missing}()
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Scheduler name if HPC is detected (e.g. LSF, SLURM), otherwise
\code{FALSE}.
}
\description{
Detect if R is running on a high-performance computing (HPC) cluster.
}
\note{
Currently supports detection of
\href{https://rc.hms.harvard.edu/#orchestra}{HMS Orchestra}.
}
\examples{
detectHPC()
}
\seealso{
\itemize{
\item \code{Sys.getenv()}.
\item \code{Sys.info()}.
\item \code{R.version}.
\item \code{.Platform}.
}

Other Infrastructure Utilities: \code{\link{prepareTemplate}}
}
\concept{Infrastructure Utilities}
