% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectHPC.R
\name{detectHPC}
\alias{detectHPC}
\title{Detect HPC Environment}
\usage{
detectHPC()
}
\value{
\code{string} or \code{boolean}. Scheduler name if HPC is detected (e.g. LSF,
SLURM), otherwise \code{FALSE}.
}
\description{
Detect if R is running on a high-performance computing (HPC) cluster.
}
\note{
Currently supports detection of
\href{https://rc.hms.harvard.edu/#orchestra}{HMS Orchestra}.
}
\examples{
detectHPC()
}
\seealso{
\itemize{
\item \code{Sys.getenv()}.
\item \code{Sys.info()}.
\item \code{R.version}.
\item \code{.Platform}.
}

Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{dots}}, \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}, \code{\link{separator}},
  \code{\link{standardizeCall}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
