% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectHPC.R
\name{detectHPC}
\alias{detectHPC}
\title{Detect HPC environment}
\usage{
detectHPC()
}
\value{
\code{character(1)} or \code{logical(1)}.
Workload manager (scheduler) name if HPC is detected (e.g. \code{"SLURM"} or
\code{"LSF"}), otherwise \code{FALSE}.
}
\description{
Detect if R is running on a high-performance computing (HPC) cluster.
}
\note{
Currently supports detection of Slurm or LSF.

Updated 2019-07-28.
}
\examples{
detectHPC()
}
\seealso{
\itemize{
\item \code{Sys.getenv}.
\item \code{Sys.info}.
\item \code{R.version}.
\item \code{.Platform}.
}
}
