% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectLanes}
\alias{detectLanes}
\title{Detect Sequencing Lanes}
\usage{
detectLanes(object, pattern = lanePattern)
}
\arguments{
\item{object}{\code{character}. Sequencing file or directory paths.}

\item{pattern}{\code{string}. Grep pattern. Default pattern checks for \code{"_LXXX"}.}
}
\value{
\code{integer}. Lane number.
}
\description{
Detect Sequencing Lanes
}
\examples{
files <- c("sample1_L001.fastq.gz", "sample1_L002.fastq.gz")
detectLanes(files)
}
\seealso{
Other Atomic Vector Functions: \code{\link{sortUnique}},
  \code{\link{toStringUnique}}
}
\author{
Michael Steinbaugh
}
\concept{Atomic Vector Functions}
