% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectLanes.R
\name{detectLanes}
\alias{detectLanes}
\title{Detect sequencing lanes}
\usage{
detectLanes(path, pattern = "_L([[:digit:]]{3})")
}
\arguments{
\item{path}{\code{character}.
Sequencing file paths, e.g. FASTQs.}

\item{pattern}{\code{character(1)}.
Lane grep pattern. Defaults to \link{lanePattern} global variable.}
}
\value{
\code{integer}.
Lane numbers.
}
\description{
Detect sequencing lanes
}
\note{
Updated 2019-08-07.
}
\examples{
## Matching lanes 1-4 for paired-end FASTQ files.
files <- paste0(
    "sample1",
    paste0("_R", seq_len(2L)),
    paste0("_L00", seq_len(4L)),
    ".fastq.gz"
)
detectLanes(files)
}
