% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R, R/zzz.R
\name{detectOrganism}
\alias{detectOrganism}
\alias{detect_organism}
\title{Detect the organism from the genome build name}
\usage{
detectOrganism(genomeBuild)
}
\arguments{
\item{genomeBuild}{Genome build.}
}
\value{
Organism string.
}
\description{
Detect the organism from the genome build name
}
