% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOrganism.R
\name{detectOrganism}
\alias{detectOrganism}
\title{Detect Organism}
\usage{
detectOrganism(object)
}
\arguments{
\item{object}{\code{character}.}
}
\value{
\code{string}. Full latin organism name. Stops on match failure.
}
\description{
Supports organism detection from Ensembl identifier or genome build.
}
\details{
Only the first match is returned. We're using a while loop approach here so
we can skip transgenes or spike-ins. The function fails after a maximum of 50
unknowns, for speed.
}
\note{
\code{\link[=organism]{organism()}} \code{character} method conflicts with annotate package, which
gets loaded into the namespace when DESeq2 is attached. Instead, we're
exporting the character method here as a separate function named
\code{\link[=detectOrganism]{detectOrganism()}}.
}
\section{Supported organisms}{

\itemize{
\item \emph{Caenorhabditis elegans} (roundworm)
\item \emph{Danio rerio} (zebrafish)
\item \emph{Drosophila melanogaster} (fruitfly)
\item \emph{Gallus gallus} (chicken)
\item \emph{Homo sapiens} (human)
\item \emph{Mus musculus} (mouse)
\item \emph{Ovis aries} (sheep)
\item \emph{Rattus norvegicus} (rat)
\item \emph{Saccharomyces cerevisiae} (yeast)
}
}

\examples{
## Match by gene identifier.
detectOrganism("ENSG00000000003")

## Match by genome build.
detectOrganism("GRCh38")  # Ensembl
detectOrganism("hg38")    # UCSC

## Match by alternate organism name.
detectOrganism("H. sapiens")
detectOrganism("hsapiens")

## The function will skip transgenes/spike-ins until we find a match.
detectOrganism(c("EGFP", "TDTOMATO", "ENSG00000000003"))

## But it only returns the first match, if there are multiple genomes.
detectOrganism(c("ENSG00000000003", "ENSMUSG00000000001"))
}
\seealso{
\code{\link[=organism]{organism()}}.

Other Metadata Functions: \code{\link{detectHPC}},
  \code{\link{detectLanes}},
  \code{\link{interestingGroups}},
  \code{\link{microplate}},
  \code{\link{minimalSampleData}}, \code{\link{organism}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{uniteInterestingGroups}}
}
\concept{Metadata Functions}
