% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-detectOrganism.R
\docType{methods}
\name{detectOrganism}
\alias{detectOrganism}
\alias{detectOrganism,character-method}
\title{Detect Organism}
\usage{
detectOrganism(object, ...)

\S4method{detectOrganism}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Full latin organism name.
}
\description{
Supports organism detection from genome build or Ensembl identifier.
}
\details{
Currently supported organisms:
\itemize{
\item \emph{Homo sapiens} (human)
\item \emph{Mus musculus} (mouse)
\item \emph{Rattus norvegicus} (rat)
\item \emph{Drosophila melanogaster} (fruitfly)
\item \emph{Caenorhabditis elegans} (roundworm)
\item \emph{Saccharomyces cerevisiae} (yeast)
\item \emph{Gallus gallus} (chicken)
\item \emph{Ovis aries} (sheep)
}
}
\examples{
# H. sapiens
detectOrganism("GRCh38")
detectOrganism("hg38")
detectOrganism("ENSG00000000003")
detectOrganism("Homo sapiens")
detectOrganism("Hsapiens")
detectOrganism("human")
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{symbol2gene}},
  \code{\link{tx2geneFromGFF}}, \code{\link{tx2gene}}
}
\concept{Gene Annotation Utilities}
