% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/detectOrganism-methods.R
\docType{methods}
\name{detectOrganism}
\alias{detectOrganism}
\alias{detectOrganism,character-method}
\alias{detectOrganism,factor-method}
\alias{detectOrganism,matrix-method}
\alias{detectOrganism,data.frame-method}
\alias{detectOrganism,DataFrame-method}
\alias{detectOrganism,dgCMatrix-method}
\alias{detectOrganism,tbl_df-method}
\title{Detect Organism}
\usage{
detectOrganism(object, ...)

\S4method{detectOrganism}{character}(object, unique = FALSE)

\S4method{detectOrganism}{factor}(object, unique = FALSE)

\S4method{detectOrganism}{matrix}(object)

\S4method{detectOrganism}{data.frame}(object)

\S4method{detectOrganism}{DataFrame}(object)

\S4method{detectOrganism}{dgCMatrix}(object)

\S4method{detectOrganism}{tbl_df}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{unique}{\code{boolean}. Only return unique matching organisms. Applies to
\code{character} input.}
}
\value{
Full latin organism name. Stops on detection failure.
\itemize{
\item \code{atomic}: Named \code{character} vector containing organism name or \code{NA} for
individual match failures (e.g. spike-ins like EGFP, TDTOMATO).
\item \code{dim}: Unique \code{character} vector of the organism(s) detected. Warns if
multiple organisms are detected.
}
}
\description{
Supports organism detection from genome build or Ensembl identifier.
}
\details{
Currently supported organisms:
\itemize{
\item \emph{Homo sapiens} (human)
\item \emph{Mus musculus} (mouse)
\item \emph{Rattus norvegicus} (rat)
\item \emph{Danio rerio} (zebrafish)
\item \emph{Drosophila melanogaster} (fruitfly)
\item \emph{Caenorhabditis elegans} (roundworm)
\item \emph{Saccharomyces cerevisiae} (yeast)
\item \emph{Gallus gallus} (chicken)
\item \emph{Ovis aries} (sheep)
}
}
\examples{
# By gene identifier
detectOrganism("ENSG00000000003")
detectOrganism(c("ENSG00000000003", "EGFP", "TDTOMATO"))

# By genome build
detectOrganism("GRCh38")  # Ensembl
detectOrganism("hg38")    # UCSC

# By alternate organism name
detectOrganism("H. sapiens")
detectOrganism("hsapiens")
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{emptyRanges}}, \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Gene Annotation Functions}
