% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-detectOrganism.R
\docType{methods}
\name{detectOrganism}
\alias{detectOrganism}
\alias{detectOrganism,character-method}
\title{Detect Organism}
\usage{
detectOrganism(object, ...)

\S4method{detectOrganism}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Full latin scientific organism name.
}
\description{
Supports organism detection from genome build or Ensembl identifier.
}
\examples{
# H. sapiens
detectOrganism("GRCh38")
detectOrganism("hg38")
detectOrganism("ENSG00000000003")
detectOrganism("Homo sapiens")
detectOrganism("Hsapiens")
detectOrganism("human")
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{gene2symbolFromGTF}},
  \code{\link{gene2symbol}}, \code{\link{symbol2gene}},
  \code{\link{tx2geneFromGTF}}, \code{\link{tx2gene}}
}
\concept{Gene Annotation Utilities}
