% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-detectOrganism.R
\docType{methods}
\name{detectOrganism}
\alias{detectOrganism}
\alias{detectOrganism}
\alias{detectOrganism,character-method}
\alias{detectOrganism,data.frame-method}
\alias{detectOrganism,DataFrame-method}
\alias{detectOrganism,dgCMatrix-method}
\alias{detectOrganism,matrix-method}
\alias{detectOrganism,tbl_df-method}
\title{Detect Organism}
\usage{
detectOrganism(object, ...)

\S4method{detectOrganism}{character}(object, unique = FALSE)

\S4method{detectOrganism}{data.frame}(object)

\S4method{detectOrganism}{DataFrame}(object)

\S4method{detectOrganism}{dgCMatrix}(object)

\S4method{detectOrganism}{matrix}(object)

\S4method{detectOrganism}{tbl_df}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{unique}{Only return unique matching organisms. Applies to character
vector input.}
}
\value{
Full latin organism name. Always aborts on detection failure.
\itemize{
\item Vector: Named character vector containing organism name or \code{NA} for
individual match failures (e.g. spike-ins like EGFP, TDTOMATO).
\item Dim: Unique character vector of the organism(s) detected.  Warns if
multiple organisms are detected.
}
}
\description{
Supports organism detection from genome build or Ensembl identifier.
}
\details{
Currently supported organisms:
\itemize{
\item \emph{Homo sapiens} (human)
\item \emph{Mus musculus} (mouse)
\item \emph{Rattus norvegicus} (rat)
\item \emph{Danio rerio} (zebrafish)
\item \emph{Drosophila melanogaster} (fruitfly)
\item \emph{Caenorhabditis elegans} (roundworm)
\item \emph{Saccharomyces cerevisiae} (yeast)
\item \emph{Gallus gallus} (chicken)
\item \emph{Ovis aries} (sheep)
}
}
\examples{
loadRemoteData("http://basejump.seq.cloud/counts.rda")

# String
detectOrganism("ENSG00000000003")
detectOrganism("GRCh38")
detectOrganism("hg38")
detectOrganism("H. sapiens")

# Character vector with spike-ins
detectOrganism(c("ENSG00000000003", "EGFP", "TDTOMATO"))

# Matrix
detectOrganism(counts)

# Tibble
tibble <- as(counts, "tibble")
detectOrganism(tibble)
}
\seealso{
Other Gene Annotation Utilities: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{ensemblAnnotations}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
