% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-detectOrganism.R
\docType{methods}
\name{detectOrganism}
\alias{detectOrganism}
\alias{detectOrganism,character-method}
\title{Detect Organism}
\usage{
detectOrganism(object)

\S4method{detectOrganism}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Full latin scientific organism name.
}
\description{
Supports organism detection from genome build or Ensembl identifier.
}
\examples{
# H. sapiens
detectOrganism("hg38")
detectOrganism("ENSG00000000003")

# M. musculus
detectOrganism("mm10")
detectOrganism("ENSMUSG00000002459")

# C. elegans
detectOrganism("WBcel235")
detectOrganism("WBGene00000001")

# D. melanogaster
detectOrganism("BDGP6")
detectOrganism("FBgn0000003")
}
