% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Example DataFrame}
\format{An object of class \code{DataFrame} with 4 rows and 4 columns.}
\usage{
df
}
\description{
Example DataFrame
}
\examples{
data(df)
df
}
\keyword{datasets}
