% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biology.R
\name{comp}
\alias{comp}
\alias{revcomp}
\title{DNA sequence utilites}
\usage{
comp(dna)

revcomp(dna)
}
\arguments{
\item{dna}{DNA sequence (\code{ATGC} nucleotides).}
}
\value{
Matching DNA sequence.
}
\description{
Generate complement or reverse complement sequences.
}
\examples{
dna <- "ATGCATGC"
comp(dna)
revcomp(dna)
}
