% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aliases.R, R/methods-dna.R
\docType{methods}
\name{complement}
\alias{complement}
\alias{reverseComplement}
\alias{comp}
\alias{revcomp}
\alias{dna}
\alias{complement,character-method}
\alias{complement,integer-method}
\alias{complement,numeric-method}
\alias{reverseComplement,character-method}
\alias{reverseComplement,integer-method}
\alias{reverseComplement,numeric-method}
\title{DNA Sequence Utilites}
\usage{
complement(object, ...)

reverseComplement(object, ...)

comp(...)

revcomp(...)

\S4method{complement}{character}(object)

\S4method{complement}{integer}(object)

\S4method{complement}{numeric}(object)

\S4method{reverseComplement}{character}(object)

\S4method{reverseComplement}{integer}(object)

\S4method{reverseComplement}{numeric}(object)
}
\arguments{
\item{object}{DNA sequence (\code{ATGC} nucleotides).}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Matching DNA sequence.
}
\description{
Generate complement or reverse complement sequences.
}
\examples{
dna <- "ATGCATGC"
complement(dna)
reverseComplement(dna)
}
