% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-interestingColData.R
\name{.interestingColData}
\alias{.interestingColData}
\title{Interesting Column Data}
\usage{
.interestingColData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}.
}
\description{
Interesting Column Data
}
\seealso{
Other Exported Constructor Functions: \code{\link{.SummarizedExperiment}},
  \code{\link{.interestingGroups}},
  \code{\link{.sampleYAMLMetadata}},
  \code{\link{.sampleYAMLMetrics}},
  \code{\link{.sampleYAML}},
  \code{\link{.setMetadataFactors}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
