% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-interestingGroup.R
\name{.interestingGroup}
\alias{.interestingGroup}
\title{Primary Interesting Group}
\usage{
.interestingGroup(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{string}.
}
\description{
Constructor function.
}
\concept{Internal Constructors}
\keyword{internal}
