% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-interestingGroups.R
\name{.interestingGroups}
\alias{.interestingGroups}
\title{Interesting Groups}
\usage{
.interestingGroups(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\seealso{
Other Exported Constructor Functions: \code{\link{.SummarizedExperiment}},
  \code{\link{.arrangeMetadataByPriorityCols}},
  \code{\link{.interestingColData}},
  \code{\link{.interestingGroup}},
  \code{\link{.sampleYAMLMetadata}},
  \code{\link{.sampleYAMLMetrics}},
  \code{\link{.sampleYAML}},
  \code{\link{.setMetadataFactors}}
}
\concept{Exported Constructor Functions}
\keyword{internal}
