% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor-sampleMetadata.R
\name{.sampleMetadata}
\alias{.sampleMetadata}
\title{Sample Metadata Constructor}
\usage{
.sampleMetadata(meta)
}
\arguments{
\item{meta}{Metadata \link{data.frame}.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata Constructor
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAMLMetrics}}, \code{\link{sampleYAML}}
}
\concept{bcbio Utilities}
\keyword{internal}
