% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots}
\alias{dots}
\title{Extract Dots from Function}
\usage{
dots(..., character = FALSE)
}
\arguments{
\item{...}{Objects as dots.}

\item{character}{Return dots (\code{...}) as \code{character} vector.}
}
\value{
\itemize{
\item For "\code{character = FALSE}": \code{list} containing objects as names. Can return
object from the stored \code{name} class with \code{\link[=eval]{eval()}}.
\item For "\code{character = TRUE}": \code{character} containing the dot names.
}
}
\description{
Extract Dots from Function
}
\examples{
# names
dots(a, b, c)

# character
dots(a, b, c, character = TRUE)
}
\seealso{
\itemize{
\item \code{help("dotsMethods", "methods")}.
\item \href{http://tidyverse.org}{tidyverse} code:
\href{http://rlang.tidyverse.org}{rlang}
\href{https://goo.gl/fhAuak}{dplyr utils},
\href{https://goo.gl/bM5TrP}{devtools infrastructure}.
}

Other Developer Functions: \code{\link{detectHPC}},
  \code{\link{localOrRemoteFile}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}
}
\author{
Michael Steinbaugh
}
