% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots}
\alias{dots}
\title{Extract Dots from Function}
\usage{
dots(..., character = FALSE)
}
\arguments{
\item{...}{Objects as dots.}

\item{character}{\code{boolean}. Return dots (\code{...}) as \code{character}.}
}
\value{
\itemize{
\item "\code{character = FALSE}": \code{list}. Objects as \code{name} class. Can return the
object from the \code{name} with \code{\link[=eval]{eval()}}.
\item "\code{character = TRUE}": \code{character}. Names of the dots.
}
}
\description{
Extract Dots from Function
}
\examples{
dots(a, b, c, character = FALSE)
dots(a, b, c, character = TRUE)
}
\seealso{
\itemize{
\item \code{help("dotsMethods", "methods")}.
\item \href{http://tidyverse.org}{tidyverse} documentation:
\itemize{
\item \href{http://rlang.tidyverse.org}{rlang}.
\item \href{https://goo.gl/fhAuak}{dplyr utils}.
\item \href{https://goo.gl/bM5TrP}{devtools infrastructure}.
}
}
}
