% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots}
\alias{dots}
\title{Extract Dots from Function}
\usage{
dots(..., character = FALSE)
}
\arguments{
\item{...}{Objects as dots.}

\item{character}{\code{boolean}. Return dots (\code{...}) as \code{character}.}
}
\value{
\itemize{
\item "\code{character = FALSE}": \code{list} containing objects as names. Can return
object from the stored \code{name} class with \code{\link[=eval]{eval()}}.
\item "\code{character = TRUE}": \code{character} containing the dot names.
}
}
\description{
Extract Dots from Function
}
\examples{
dots(a, b, c, character = FALSE)
dots(a, b, c, character = TRUE)
}
\seealso{
\itemize{
\item \code{help("dotsMethods", "methods")}.
\item \href{http://tidyverse.org}{tidyverse} code:
\href{http://rlang.tidyverse.org}{rlang}
\href{https://goo.gl/fhAuak}{dplyr utils},
\href{https://goo.gl/bM5TrP}{devtools infrastructure}.
}

Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}}, \code{\link{matchCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
