% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dots-methods.R
\docType{methods}
\name{dots}
\alias{dots}
\alias{dots,ANY-method}
\title{Extract Dots from Function}
\usage{
dots(..., character = FALSE)

\S4method{dots}{ANY}(..., character = FALSE)
}
\arguments{
\item{...}{Objects as dots.}

\item{character}{Return \code{...} as character.}
}
\value{
\itemize{
\item For \code{character = FALSE}: \link{list} containing objects as names. Can return
object from the stored \link{name} class with \code{\link[=eval]{eval()}}.
\item For \code{character = TRUE}: Character vector.
}
}
\description{
Extract Dots from Function
}
\examples{
dots(mtcars, starwars)
dots(mtcars, starwars, character = TRUE)
}
\seealso{
\itemize{
\item \code{help("dotsMethods", "methods")}
\item \code{browseVignettes("rlang")}.
\item \href{http://rlang.tidyverse.org}{rlang} functions:
\code{\link[rlang:quos]{rlang::quos()}},
\code{\link[rlang:dots_definitions]{rlang::dots_definitions()}},
\code{\link[rlang:dots_values]{rlang::dots_values()}},
\code{\link[rlang:eval_tidy]{rlang::eval_tidy()}},
\code{\link[rlang:eval_bare]{rlang::eval_bare()}}.
\item \href{http://tidyverse.org}{tidyverse} code:
\href{https://goo.gl/fhAuak}{dplyr utils},
\href{https://goo.gl/bM5TrP}{devtools infrastructure}
}
}
