% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots}
\alias{dots}
\alias{get_objs_from_dots}
\title{Evaluate dots}
\usage{
dots(...)

get_objs_from_dots(.dots)
}
\arguments{
\item{...}{Ellipsis.}

\item{.dots}{Dots values.}
}
\description{
Capture function ellipses as dots using rlang syntax.
}
\seealso{
\itemize{
\item \code{browseVignettes("rlang")}.
\item \code{\link[rlang:dots_values]{rlang::dots_values()}}, \code{\link[rlang:eval_tidy]{rlang::eval_tidy()}}.
\item https://github.com/tidyverse/dplyr/blob/master/R/utils.r
\item https://github.com/hadley/devtools/blob/master/R/infrastructure.R
}
}
