% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-dotted.R
\docType{methods}
\name{dotted}
\alias{dotted}
\alias{dotted}
\alias{dotted,ANY-method}
\alias{dotted,character-method}
\alias{dotted,data.frame-method}
\alias{dotted,DataFrame-method}
\alias{dotted,factor-method}
\alias{dotted,GRanges-method}
\alias{dotted,list-method}
\alias{dotted,List-method}
\alias{dotted,matrix-method}
\alias{dotted,SimpleList-method}
\alias{dotted,tbl_df-method}
\title{Dotted Case}
\usage{
dotted(object, ...)

\S4method{dotted}{ANY}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{character}(object)

\S4method{dotted}{data.frame}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{DataFrame}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{factor}(object)

\S4method{dotted}{GRanges}(object)

\S4method{dotted}{list}(object)

\S4method{dotted}{List}(object)

\S4method{dotted}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{SimpleList}(object)

\S4method{dotted}{tbl_df}(object)
}
\arguments{
\item{object}{Character vector or an object for which \code{\link[=names]{names()}} assignment
will be meaningful.}

\item{...}{Additional arguments.}

\item{rownames}{Apply sanitization on row names. This is not recommended
by default, since rownames commonly contain gene identifiers that should
not be modified.}

\item{colnames}{Apply sanitization on column names. This is generally
recommended by default.}
}
\value{
Object with syntatically valid names. For objects supporting
\code{\link[base:names]{base::names()}}, the underlying data returns unchanged.
}
\description{
For unnamed character vectors, this function will sanitize the underlying
values. Otherwise, it will set \code{\link[=names]{names()}} and/or \code{\link[=rownames]{rownames()}} on objects
supporting name assignments. They return the object without modification of
the underlying data.
}
\note{
\code{dotted.case} support is provided for matching against base R
parameters, but we strongly advise against using it for object and/or
argument name assignments.
}
\examples{
load(system.file("extdata/mn.rda", package = "basejump"))

# character ====
character <- mn$character
print(character)
dotted(character)

namedCharacter <- mn$namedCharacter
dotted(namedCharacter)

# factor ====
factor <- mn$factor
print(factor)
dotted(factor)

# data.frame ====
dataFrame <- mn$dataFrame
print(dataFrame)
dotted(dataFrame, rownames = TRUE)
dotted(dataFrame, rownames = FALSE)

# list ====
list <- mn$list
print(list)
dotted(list)
}
\seealso{
Other Name Functions: \code{\link{camel}},
  \code{\link{snake}}
}
