% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-camel.R,
%   R/methods-dotted.R, R/methods-upperCamel.R
\docType{methods}
\name{dotted}
\alias{dotted}
\alias{camel,List-method}
\alias{camel,SimpleList-method}
\alias{dotted}
\alias{dotted,ANY-method}
\alias{dotted,character-method}
\alias{dotted,data.frame-method}
\alias{dotted,DataFrame-method}
\alias{dotted,factor-method}
\alias{dotted,list-method}
\alias{dotted,List-method}
\alias{dotted,matrix-method}
\alias{dotted,SimpleList-method}
\alias{dotted,tbl_df-method}
\alias{upperCamel,List-method}
\alias{upperCamel,SimpleList-method}
\title{Dotted Case}
\usage{
dotted(object, ...)

\S4method{camel}{List}(object, strict = FALSE)

\S4method{camel}{SimpleList}(object, strict = FALSE)

\S4method{dotted}{ANY}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{character}(object)

\S4method{dotted}{data.frame}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{DataFrame}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{factor}(object)

\S4method{dotted}{list}(object)

\S4method{dotted}{List}(object)

\S4method{dotted}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{SimpleList}(object)

\S4method{dotted}{tbl_df}(object)

\S4method{upperCamel}{List}(object, strict = FALSE)

\S4method{upperCamel}{SimpleList}(object, strict = FALSE)
}
\arguments{
\item{object}{Character vector or an object for which \code{\link[=names]{names()}} assignment
will be meaningful.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{rownames}{Apply sanitization on row names. This is not recommended
by default, since rownames commonly contain gene identifiers that should
not be modified.}
}
\value{
Object with syntatically valid names. For objects supporting
\code{\link[base:names]{base::names()}}, the underlying data returns unchanged.
}
\description{
Dotted Case
}
\details{
For unnamed character vectors, this function will sanitize the
underlying values. Otherwise, it will set \code{\link[=names]{names()}} and/or \code{\link[=rownames]{rownames()}} on
objects supporting name assignments. They return the object without
modification of the underlying data.
}
\note{
\code{dotted.case} support is provided for matching against base R
parameters, but we strongly advise against using it for object and/or
argument name assignments.
}
\examples{
load(system.file(
    file.path("extdata", "makeNames.rda"),
    package = "basejump"
))

# Character vector
character <- makeNames$character
print(character)
dotted(character)

# Named character vector
namedCharacter <- makeNames$namedCharacter
dotted(namedCharacter)

# Factor
factor <- makeNames$factor
print(factor)
dotted(factor)

# data.frame
dataFrame <- makeNames$dataFrame
print(dataFrame)
dotted(dataFrame, rownames = TRUE)
dotted(dataFrame, rownames = FALSE)

# Named list
list <- makeNames$list
print(list)
dotted(list)
}
