% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottedCase-methods.R
\docType{methods}
\name{dottedCase}
\alias{dottedCase}
\alias{dottedCase,atomic-method}
\alias{dottedCase,factor-method}
\alias{dottedCase,list-method}
\alias{dottedCase,matrix-method}
\alias{dottedCase,data.frame-method}
\alias{dottedCase,Vector-method}
\alias{dottedCase,DataTable-method}
\alias{dottedCase,Ranges-method}
\alias{dottedCase,Matrix-method}
\alias{dottedCase,SummarizedExperiment-method}
\title{Dotted case}
\usage{
\S4method{dottedCase}{atomic}(object, names = TRUE)

\S4method{dottedCase}{factor}(object, names = TRUE)

\S4method{dottedCase}{list}(object, names = TRUE)

\S4method{dottedCase}{matrix}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{dottedCase}{data.frame}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{dottedCase}{Vector}(object, names = TRUE, mcols = TRUE,
  metadata = TRUE)

\S4method{dottedCase}{DataTable}(object, rownames = FALSE,
  colnames = TRUE, mcols = TRUE, metadata = TRUE)

\S4method{dottedCase}{Ranges}(object, names = FALSE, mcols = TRUE,
  metadata = TRUE)

\S4method{dottedCase}{Matrix}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{dottedCase}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE, assayNames = TRUE, rowData = TRUE,
  colData = TRUE, metadata = TRUE)
}
\arguments{
\item{object}{Object.}

\item{names}{\code{logical(1)}.
Sanitize names.}
}
\value{
Modified object, with names formatted in dotted case.
}
\description{
Dotted case
}
\note{
Updated 2019-09-25.
}
\examples{
data(syntactic, package = "acidtest")
lapply(syntactic, dottedCase)
}
