% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottedCase-methods.R
\name{dottedCase}
\alias{dottedCase}
\alias{dottedCase,atomic-method}
\alias{dottedCase,factor-method}
\alias{dottedCase,list-method}
\alias{dottedCase,matrix-method}
\alias{dottedCase,data.frame-method}
\alias{dottedCase,Vector-method}
\alias{dottedCase,DataFrame-method}
\alias{dottedCase,Ranges-method}
\alias{dottedCase,Matrix-method}
\alias{dottedCase,SummarizedExperiment-method}
\title{Dotted case}
\usage{
dottedCase(object, ...)

\S4method{dottedCase}{atomic}(object, names = TRUE)

\S4method{dottedCase}{factor}(object, names = TRUE)

\S4method{dottedCase}{list}(object, names = TRUE)

\S4method{dottedCase}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{dottedCase}{data.frame}(object, rownames = FALSE, colnames = TRUE)

\S4method{dottedCase}{Vector}(object, names = TRUE, mcols = TRUE, metadata = TRUE)

\S4method{dottedCase}{DataFrame}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  mcols = TRUE,
  metadata = TRUE
)

\S4method{dottedCase}{Ranges}(object, names = FALSE, mcols = TRUE, metadata = TRUE)

\S4method{dottedCase}{Matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{dottedCase}{SummarizedExperiment}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  assayNames = TRUE,
  rowData = TRUE,
  colData = TRUE,
  metadata = TRUE
)
}
\arguments{
\item{object}{Object.}

\item{names}{\code{logical(1)}.
Sanitize names.

Only applies to string mode (\code{rename = FALSE}).}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{mcols}{\code{logical(1)}.
Sanitize names of metadata columns (i.e. \code{DataFrame}).}

\item{metadata}{\code{logical(1)}.
Sanitize metadata names.}

\item{assayNames}{\code{logical(1)}.
Sanitize assay names.}

\item{rowData}{\code{logical(1)}.
Sanitize the row data names.}

\item{colData}{\code{logical(1)}.
Sanitize column names of column data.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names.
For objects with \code{\link[base:names]{names()}} defined, the underlying data
returns unchanged, except for \code{character} or \code{vector} class.
}
\description{
Support is provided for matching against base R parameters. However, it is
recommended to avoid using it for variable assignments into an \code{environment},
as that can introduce conflicts with base functions.
}
\note{
Updated 2020-05-11.
}
\examples{
data(syntactic, package = "AcidTest")
lapply(syntactic, dottedCase)
}
