% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFile.R
\name{downloadFile}
\alias{downloadFile}
\title{Match and download a file from a remote directory}
\usage{
downloadFile(remoteDir, string, rename = NULL, compress = FALSE,
  localDir = "data-raw")
}
\arguments{
\item{remoteDir}{Remote directory URL}

\item{string}{String to match against remote file names}

\item{rename}{Rename the local file, if desired}

\item{compress}{Whether to compress the file (\code{gzip}) after download}

\item{localDir}{Directory where to save file locally}
}
\value{
Local file name
}
\description{
Match and download a file from a remote directory
}
\author{
Michael Steinbaugh
}
