% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicPlotlist.R
\name{dynamicPlotlist}
\alias{dynamicPlotlist}
\title{Dynamic Plotlist}
\usage{
dynamicPlotlist(plotlist, headerLevel = 2L, return = c("grid", "list",
  "markdown"))
}
\arguments{
\item{plotlist}{\code{list} containing \code{ggplot} objects.}

\item{headerLevel}{Markdown header level.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\value{
\itemize{
\item \code{grid}: \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.
\item \code{list}: \code{list}.
\item \code{markdown}: Markdown output, with headers for each plot.
}
}
\description{
Dynamic Plotlist
}
\examples{
loadRemoteData("http://basejump.seq.cloud/plotlist.rda")

# Grid
dynamicPlotlist(plotlist, return = "grid")

# List
list <- dynamicPlotlist(plotlist, return = "list")
names(list)

# Markdown
dynamicPlotlist(plotlist, return = "markdown")
}
\seealso{
\code{\link[=markdownPlotlist]{markdownPlotlist()}}.

Other Plot Functions: \code{\link{themes}}
}
\author{
Michael Steinbaugh
}
