% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-dynamicPlotlist.R
\docType{methods}
\name{dynamicPlotlist}
\alias{dynamicPlotlist}
\alias{dynamicPlotlist}
\alias{dynamicPlotlist,list-method}
\title{Dynamic Plotlist}
\usage{
dynamicPlotlist(object, ...)

\S4method{dynamicPlotlist}{list}(object, return = "grid", headerLevel = 2L)
}
\arguments{
\item{return}{Return type. "grid", "list", and "markdown" are supported.}

\item{headerLevel}{Markdown header level. Only applicable when
\code{return = "markdown"}.}
}
\value{
\itemize{
\item \code{grid}: \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.
\item \code{list}: \link{list}.
\item \code{markdown}: Markdown output, with headers for each plot.
}
}
\description{
Dynamic Plotlist
}
\author{
Michael Steinbaugh
}
