% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{EggNOG}
\alias{EggNOG}
\title{EggNOG database annotations}
\usage{
EggNOG()
}
\value{
\code{EggNOG}.
}
\description{
\href{http://eggnogdb.embl.de}{EggNOG} is a database of biological information
hosted by the EMBL. It is based on the original idea of COGs (\strong{c}lusters of
\strong{o}rthologous \strong{g}roups) and expands that idea to non-supervised
orthologous groups constructed from numerous organisms. eggNOG stands for
\strong{e}volutionary \strong{g}enealogy of \strong{g}enes: \strong{N}on-supervised
\strong{O}rthologous \strong{G}roups.
}
\note{
Updated 2019-08-15.
}
\examples{
options(acid.test = TRUE)
x <- EggNOG()
print(x)
}
