% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog.R
\name{eggnog}
\alias{eggnog}
\alias{EggNOG}
\title{EggNOG Database Annotations}
\usage{
eggnog(.test = FALSE)
}
\arguments{
\item{.test}{\code{boolean}. \emph{For unit testing only}.}
}
\value{
\code{EggNOG}.
}
\description{
\href{http://eggnogdb.embl.de}{EggNOG} is a database of biological information
hosted by the EMBL. It is based on the original idea of COGs (\strong{c}lusters of
\strong{o}rthologous \strong{g}roups) and expands that idea to non-supervised
orthologous groups constructed from numerous organisms. eggNOG stands for
\strong{e}volutionary \strong{g}enealogy of \strong{g}enes: \strong{N}on-supervised
\strong{O}rthologous \strong{G}roups.
}
\examples{
x <- eggnog(.test = TRUE)
print(x)
}
\seealso{
Other Annotation Functions: \code{\link{emptyRanges}},
  \code{\link{geneSynonyms}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2Symbol}}, \code{\link{makeTx2Gene}},
  \code{\link{panther}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
