% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog.R
\name{eggnog}
\alias{eggnog}
\title{EggNOG Annotations}
\usage{
eggnog()
}
\value{
\code{list} containing (1; "\code{cogFunctionalCategories}") \strong{C}luster of
\strong{O}rthologous \strong{G}roups (COG) functional category information and (2;
"\code{annotations}") up-to-date functional descriptions and categories for
\strong{Eu}karyotes \strong{N}on-supervised \strong{O}rthologous \strong{G}roups (euNOG) and
\strong{N}on-supervised \strong{O}rthologous \strong{G}roups (NOG) protein identifiers.
}
\description{
EggNOG is a hierarchical orthology framework with functional annotations for
eukaryotic, prokaryotic, and viral genomes.
}
\examples{
x <- eggnog()
glimpse(x)
}
\seealso{
\href{http://eggnogdb.embl.de/download/latest/README.txt}{EggNOG README}

Other Database Functions: \code{\link{panther}}
}
\author{
Michael Steinbaugh
}
