% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog.R
\name{eggnog}
\alias{eggnog}
\alias{EggNOG}
\title{\code{EggNOG} Generator}
\usage{
eggnog(.test = FALSE)
}
\arguments{
\item{.test}{\code{boolean}. \emph{For unit testing only}.}
}
\value{
\code{EggNOG}.
}
\description{
\code{EggNOG} Generator
}
\examples{
x <- eggnog(.test = TRUE)
print(x)
}
\seealso{
Other S4 Generators: \code{\link{ensembl2entrez}},
  \code{\link{gene2symbol}}, \code{\link{hgnc2ensembl}},
  \code{\link{mgi2ensembl}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
