% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog.R
\name{eggnog}
\alias{eggnog}
\title{EggNOG Annotations}
\usage{
eggnog()
}
\value{
\code{list} containing:
\enumerate{
\item "\code{cogFunctionalCategories}": \strong{C}luster of \strong{O}rthologous \strong{G}roups
(COG) functional category information.
\item "\code{annotations}": up-to-date functional descriptions and categories
for \strong{Eu}karyotes \strong{N}on-supervised \strong{O}rthologous \strong{G}roups (euNOG)
and \strong{N}on-supervised \strong{O}rthologous \strong{G}roups (NOG) protein
identifiers.
}
}
\description{
EggNOG is a hierarchical orthology framework with functional annotations for
eukaryotic, prokaryotic, and viral genomes.
}
\examples{
x <- eggnog()
print(x)
}
\seealso{
\href{http://eggnogdb.embl.de/download/latest/README.txt}{EggNOG README}.

Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{emptyRanges}}, \code{\link{ensembl2entrez}},
  \code{\link{gene2symbol}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2ensembl}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2ensembl}}, \code{\link{organism}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
