% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyRanges.R
\name{emptyRanges}
\alias{emptyRanges}
\title{Generate Empty Genomic Ranges}
\usage{
emptyRanges(names, seqname = "transgene", mcolsNames = NULL)
}
\value{
\code{GRanges}.
}
\description{
Utility function that provides support for creating internal \code{GRanges} for
transgene and FASTA spike-in sequences.
}
\examples{
emptyRanges("EGFP", seqname = "transgene")
}
\seealso{
\code{help("seqinfo", "GenomeInfoDb")}.

Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}}, \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
