% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyRanges.R
\name{emptyRanges}
\alias{emptyRanges}
\title{Generate Empty Genomic Ranges}
\usage{
emptyRanges(names, seqname = c("unknown", "transgene", "spike"),
  mcolsNames = NULL)
}
\arguments{
\item{names}{\code{character}. Gene or transcript names.}

\item{seqname}{\code{string} Name of the alternative chromosome to be defined in
\code{\link[GenomeInfoDb:seqnames]{GenomeInfoDb::seqnames()}} where these ranges will be grouped. Defaults to
"\code{unknown}" but "\code{transgene}" (transgenes) and "\code{spike}" (spike-ins) are
also supported.}

\item{mcolsNames}{\code{character} or \code{NULL}. Metadata column names to be defined
in the \code{\link[S4Vectors:mcols]{S4Vectors::mcols()}} of the \code{GRanges} return. Normally this does not
need to be defined; useful when combining with another \code{GRanges} that
contains metadata.}
}
\value{
\code{GRanges}.
}
\description{
Utility function that provides support for creating internal \code{GRanges} for
transgene and FASTA spike-in sequences.
}
\examples{
# Unknown/dead genes.
emptyRanges("ENSG00000000000", seqname = "unknown")

# Transgenes.
emptyRanges(c("EGFP", "TDTOMATO", "GAL4"), seqname = "transgene")

# Spike-ins.
emptyRanges("ERCC", seqname = "spike")
}
\seealso{
\code{help("seqinfo", "GenomeInfoDb")}.

Other Annotation Functions: \code{\link{eggnog}},
  \code{\link{geneSynonyms}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2Symbol}}, \code{\link{makeTx2Gene}},
  \code{\link{panther}}
}
\concept{Annotation Functions}
