% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/encode-methods.R
\docType{methods}
\name{encode}
\alias{encode}
\alias{encode,DataFrame-method}
\title{Encode Column Data Using Run-Length Encoding}
\usage{
encode(x, ...)

\S4method{encode}{DataFrame}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with \code{atomic} columns converted to \code{Rle}.
}
\description{
Encode Column Data Using Run-Length Encoding
}
\examples{
binary <- seq(from = 0L, to = 1L)
df <- S4Vectors::DataFrame(
    a = rep(x = binary, times = 50L),
    b = rep(x = binary, each = 50L)
)
lapply(df, class)
x <- encode(df)
lapply(x, class)
}
\seealso{
\code{S4Vectors::Rle}.
}
