% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Rle-methods.R
\docType{methods}
\name{encode}
\alias{encode}
\alias{encode,DataFrame-method}
\title{Encode Column Data Using Run-Length Encoding}
\usage{
encode(x, ...)

\S4method{encode}{DataFrame}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with \code{atomic} columns converted to \code{Rle}.
}
\description{
Encode Column Data Using Run-Length Encoding
}
\seealso{
\code{\link[S4Vectors:Rle]{S4Vectors::Rle()}}.
}
