% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensembl}
\alias{ensembl}
\title{Ensembl Annotations}
\usage{
ensembl(organism, format = "genes", genomeBuild = NULL, release = NULL,
  uniqueSymbol = FALSE, broadClass = TRUE, sanitizeColnames = TRUE,
  return = "GRanges")
}
\arguments{
\item{organism}{Default recommended usage is to provide the full latin
organism name as a string.}

\item{format}{Desired annotation data format, either "\code{genes}",
"\code{transcripts}", "\code{gene2symbol}"., or "\code{tx2gene}".}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{uniqueSymbol}{Make gene symbols unique. Only applies to \code{genes} and
\code{gene2symbol} output.}

\item{broadClass}{Include broad class definitions, based on \code{biotype} and/or
gene \code{symbol} annotations. Only applies to \code{genes} and \code{transcripts}
output.}

\item{sanitizeColnames}{Improve column names and drop duplicated columns:
\itemize{
\item Rename \code{geneID} to \code{ensgene}.
\item Rename \code{txID} to \code{enstxp}.
\item Rename \code{geneBiotype} or \code{txBiotype} to simply \code{biotype}.
\item Rename \code{entrezid} to \code{entrez}.
}}

\item{return}{Class of the returned object. Can be "\code{data.frame}",
"\code{DataFrame}" or "\code{GRanges}. See \code{help("genes", "ensembldb")} for
additional information.}
}
\value{
\code{GRanges}, \code{data.frame}, or \code{DataFrame}.
}
\description{
Quickly obtain gene and transcript annotations from
\href{http://www.ensembl.org/}{Ensembl} using
\href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
}
\details{
Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.
}
\note{
Use GRCh38 instead of hg38 for the genome build, since we're
querying Ensembl and not UCSC. Unfortunately, GRCh37 is not currently
suported on AnnotationHub.
}
\section{Broad Class Definitions}{

For gene and transcript tables, when \code{broadClass = TRUE}, a \code{broadClass}
column is added, which generalizes the gene types into a smaller number of
semantically-meaningful groups:
\itemize{
\item \code{coding}
\item \code{noncoding}
\item \code{pseudo}
\item \code{small}
\item \code{decaying}
\item \code{ig} (immunoglobulin)
\item \code{tcr} (T cell receptor)
\item \code{other}
}
}

\examples{
genes <- ensembl("Homo sapiens", format = "genes")
summary(genes)

transcripts <- ensembl("Homo sapiens", format = "transcripts")
summary(transcripts)
}
\seealso{
\itemize{
\item \href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
\item \href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.
}

Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{genes}},
  \code{\link{sanitizeAnnotable}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{transcripts}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
