% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensembl}
\alias{ensembl}
\title{Ensembl Annotations}
\usage{
ensembl(organism, format = c("genes", "gene2symbol", "transcripts",
  "tx2gene"), genomeBuild = NULL, release = NULL, metadata = FALSE,
  return = c("GRanges", "DataFrame", "data.frame"))
}
\arguments{
\item{organism}{Default recommended usage is to provide the full latin
organism name as a string.}

\item{format}{Desired annotation data format, either "\code{genes}",
"\code{transcripts}", "\code{gene2symbol}"., or "\code{tx2gene}".}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{metadata}{Include the AnnotationHub metadata inside a list. Useful
for documenting where the annotations were sourced from inside
AnnotationHub.}

\item{return}{Class of the returned object. Only applies when \code{format} equals
"\code{genes}" or "\code{transcripts}". Can be "\code{GRanges}" (default), "\code{DataFrame}",
or "\code{data.frame}". See \code{help("genes", "ensembldb")} for additional
information. "\code{gene2symbol}" and "\code{tx2gene}" are always returned as
\code{data.frame}.}
}
\value{
\code{GRanges}, \code{DataFrame}, or \code{data.frame} containing gene or
transcript annotations.
}
\description{
Quickly obtain gene and transcript annotations from
\href{http://www.ensembl.org/}{Ensembl} using
\href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
}
\details{
Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.
}
\note{
Use GRCh38 instead of hg38 for the genome build, since we're
querying Ensembl and not UCSC. Unfortunately, GRCh37 is not currently
suported on AnnotationHub.
}
\section{Broad Class Definitions}{

For gene and transcript tables, when \code{broadClass = TRUE}, a \code{broadClass}
column is added, which generalizes the gene types into a smaller number of
semantically-meaningful groups:
\itemize{
\item \code{coding}
\item \code{noncoding}
\item \code{pseudo}
\item \code{small}
\item \code{decaying}
\item \code{ig} (immunoglobulin)
\item \code{tcr} (T cell receptor)
\item \code{other}
}
}

\examples{
# Genes
genes <- ensembl("Homo sapiens", format = "genes")
summary(genes)

# Transcripts
transcripts <- ensembl("Homo sapiens", format = "transcripts")
summary(transcripts)

# Include AnnotationHub metadata
x <- ensembl("Homo sapiens", metadata = TRUE)
class(x)
names(x)
glimpse(x[["metadata"]])
}
\seealso{
\itemize{
\item \href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
\item \href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.
}

Other Gene Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{genes-method}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{transcripts-method}},
  \code{\link{tx2geneFromGFF}}, \code{\link{tx2gene}}
}
