% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/AllGenerators.R
\docType{class}
\name{Ensembl2Entrez-class}
\alias{Ensembl2Entrez-class}
\alias{Ensembl2Entrez}
\alias{Ensembl2Entrez,DataFrame-method}
\alias{Ensembl2Entrez,GRanges-method}
\alias{Ensembl2Entrez,SummarizedExperiment-method}
\title{Ensembl-to-Entrez gene identifier mappings}
\usage{
Ensembl2Entrez(object, ...)

\S4method{Ensembl2Entrez}{DataFrame}(object, format = c("1:1", "long"))

\S4method{Ensembl2Entrez}{GRanges}(object, format = c("1:1", "long"))

\S4method{Ensembl2Entrez}{SummarizedExperiment}(object, format = c("1:1",
  "long"))
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Defines 1:1 mappings from Ensembl gene IDs to Entrez IDs. Uses the oldest
Entrez ID if there are multiple identifiers that map to an Ensembl gene ID.
}
\details{
Contains a \code{DataFrame} with \code{geneID} and \code{entrezID} columns.
}
\note{
Updated 2019-08-08.
}
