% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Ensembl2Entrez-class}
\alias{Ensembl2Entrez-class}
\title{Ensembl-to-Entrez Gene Identifier Mappings}
\value{
\code{Ensembl2Entrez}.
Contains a \code{DataFrame} with \code{geneID} and \code{entrezID} columns.
}
\description{
Defines 1:1 mappings from Ensembl gene IDs to Entrez IDs. Uses the oldest
Entrez ID if there are multiple identifiers that map to an Ensembl gene ID.
}
\author{
Michael Steinbaugh
}
