% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ensembl2entrez-class}
\alias{ensembl2entrez-class}
\title{ensembl2entrez Class}
\value{
\code{ensembl2entrez}.
}
\description{
Ensembl gene ID to Entrez ID mappings.
}
\details{
Contains a \code{DataFrame} with \code{geneID} and \code{entrezID} columns. Requires 1:1
mappings.
}
\examples{
x <- new(
    Class = "ensembl2entrez",
    DataFrame(
        geneID = "ENSG00000000003",
        entrezID = 7105L,
        row.names = "ENSG00000000003"
    )
)
print(x)
}
\seealso{
\code{\link[=ensembl2entrez]{ensembl2entrez()}}.

Other S4 Classes: \code{\link{PANTHER-class}},
  \code{\link{gene2symbol-class}},
  \code{\link{hgnc2ensembl-class}},
  \code{\link{mgi2ensembl-class}},
  \code{\link{tx2gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
