% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Ensembl2Entrez-class}
\alias{Ensembl2Entrez-class}
\title{\code{Ensembl2Entrez} Class}
\description{
Ensembl gene ID to Entrez ID mappings.
}
\details{
Contains a \code{DataFrame} with \code{geneID} and \code{entrezID} columns. Requires 1:1
mappings.
}
\seealso{
\code{\link[=ensembl2entrez]{ensembl2entrez()}}.

Other S4 Classes: \code{\link{EggNOG-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
