% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Ensembl2Entrez-class}
\alias{Ensembl2Entrez-class}
\title{Ensembl-to-Entrez gene identifier mappings}
\value{
\code{Ensembl2Entrez}.
}
\description{
Ensembl-to-Entrez gene identifier mappings
}
\details{
Contains a \code{DataFrame} with \code{ensembl} and \code{entrez} columns.
}
\note{
Updated 2020-10-05.
}
