% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ensembl2Entrez-methods.R
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{Entrez2Ensembl}
\alias{Ensembl2Entrez,character-method}
\alias{Entrez2Ensembl,integer-method}
\alias{Ensembl2Entrez,GRanges-method}
\alias{Ensembl2Entrez,RangedSummarizedExperiment-method}
\title{Ensembl-to-Entrez gene identifier mappings}
\usage{
Ensembl2Entrez(object, ...)

Entrez2Ensembl(object, ...)

\S4method{Ensembl2Entrez}{character}(object, organism = NULL, format = c("1:1", "long"))

\S4method{Entrez2Ensembl}{integer}(object, organism = "Homo sapiens", format = c("1:1", "long"))

\S4method{Ensembl2Entrez}{GRanges}(object, format = c("1:1", "long"))

\S4method{Ensembl2Entrez}{RangedSummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.}
Return with 1:1 mappings. For Ensembl genes that don't map 1:1 with
Entrez, pick the oldest Entrez ID. Genes that don't map to Entrez will
contain \code{NA} in \code{entrezID} column.
\item \code{"long"}:
Return 1:many in long format.
}}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Ensembl-to-Entrez gene identifier mappings
}
\note{
Updated 2020-10-01.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## character ====
## Ensembl-to-Entrez.
genes <- c("ENSG00000000003", "ENSG00000000005")
x <- Ensembl2Entrez(genes)
print(x)

## integer ====
## Entrez-to-Ensembl.
genes <- c(1L, 2L)
x <- Entrez2Ensembl(genes)
print(x)

## SummarizedExperiment ====
x <- Ensembl2Entrez(rse)
print(x)
}
