% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ensembl2Entrez.R
\docType{methods}
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{Ensembl2Entrez,SummarizedExperiment-method}
\title{Ensembl-to-Entrez Gene Identifier Mappings}
\usage{
Ensembl2Entrez(object, ...)

\S4method{Ensembl2Entrez}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Defines 1:1 mappings from Ensembl gene IDs to Entrez IDs. Uses the oldest
Entrez ID if there are multiple identifiers that map to an Ensembl gene ID.
}
\examples{
data(rse_small)
x <- Ensembl2Entrez(rse_small)
print(x)
}
\seealso{
\code{\link[=Ensembl2Entrez]{Ensembl2Entrez()}}.

Other S4 Classes: \code{\link{EggNOG-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
