% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllGenerators.R
\docType{methods}
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{Ensembl2Entrez,DataFrame-method}
\alias{Ensembl2Entrez,GRanges-method}
\alias{Ensembl2Entrez,SummarizedExperiment-method}
\title{Ensembl-to-Entrez gene identifier mappings}
\usage{
Ensembl2Entrez(object, ...)

\S4method{Ensembl2Entrez}{DataFrame}(object, format = c("1:1", "long"))

\S4method{Ensembl2Entrez}{GRanges}(object, format = c("1:1", "long"))

\S4method{Ensembl2Entrez}{SummarizedExperiment}(object, format = c("1:1",
  "long"))
}
\arguments{
\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.}
Return with 1:1 mappings. For Ensembl genes that don't map 1:1 with
Entrez, pick the oldest Entrez ID. Genes that don't map to Entrez will
contain \code{NA} in \code{entrezID} column.
\item \code{"long"}:
Return 1:many in long format.
}}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Defines 1:1 mappings from Ensembl gene IDs to Entrez IDs. Uses the oldest
Entrez ID if there are multiple identifiers that map to an Ensembl gene ID.
}
\note{
Updated 2019-08-08.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
x <- Ensembl2Entrez(rse)
print(x)
}
