% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ensembl2entrez.R
\docType{methods}
\name{ensembl2entrez}
\alias{ensembl2entrez}
\alias{Ensembl2Entrez}
\alias{ensembl2entrez,SummarizedExperiment-method}
\title{\code{Ensembl2Entrez} Generator}
\usage{
ensembl2entrez(object, ...)

\S4method{ensembl2entrez}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Defines 1:1 mappings from Ensembl gene IDs to Entrez IDs. Uses the oldest
Entrez ID if there are multiple identifiers that map to an Ensembl gene ID.
}
\examples{
data(rse_small)
x <- ensembl2entrez(rse_small)
print(x)
}
\seealso{
Other S4 Generators: \code{\link{eggnog}},
  \code{\link{gene2symbol}}, \code{\link{hgnc2ensembl}},
  \code{\link{mgi2ensembl}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
