% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ensembl2entrez.R
\docType{methods}
\name{ensembl2entrez}
\alias{ensembl2entrez}
\alias{ensembl2entrez,SummarizedExperiment-method}
\title{Ensembl Gene ID to Entrez ID Mappings}
\usage{
ensembl2entrez(object, ...)

\S4method{ensembl2entrez}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ensembl2entrez}.
}
\description{
Defines 1:1 mappings from Ensembl gene IDs to Entrez IDs. Uses the oldest
Entrez ID if there are multiple identifiers that map to an Ensembl gene ID.
}
\examples{
x <- ensembl2entrez(rse_small)
print(x)
}
\seealso{
Other Annotation Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{eggnog}}, \code{\link{emptyRanges}},
  \code{\link{gene2symbol}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2ensembl}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2ensembl}}, \code{\link{organism}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
