% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecated.R,
%   R/ensembl2entrez.R
\docType{methods}
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{ensembl2entrez}
\alias{Ensembl2Entrez,SummarizedExperiment-method}
\title{Ensembl-to-Entrez Gene Identifier Mappings}
\usage{
Ensembl2Entrez(object, ...)

\S4method{Ensembl2Entrez}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Defines 1:1 mappings from Ensembl gene IDs to Entrez IDs. Uses the oldest
Entrez ID if there are multiple identifiers that map to an Ensembl gene ID.
}
\examples{
data(rse_small)
x <- Ensembl2Entrez(rse_small)
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{Gene2Symbol}},
  \code{\link{HGNC2Ensembl}}, \code{\link{MGI2Ensembl}},
  \code{\link{Tx2Gene}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{mapGenes}}
}
\concept{Identifier Mapping Functions}
