% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblVersion.R
\name{ensemblVersion}
\alias{ensemblVersion}
\title{Ensembl Build Version}
\usage{
ensemblVersion()
}
\value{
String containing Ensembl build version of annotables.
}
\description{
Ensembl Build Version
}
\keyword{internal}
