% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,ANY-method}
\title{Export}
\usage{
\S4method{export}{ANY}(object, format = "csv", dir = ".")
}
\description{
Export data out of R and write to disk.
}
\examples{
export(mtcars, format = "csv", dir = ".")
}
\seealso{
\code{\link[rio:export]{rio::export()}}.
}
\author{
Michael Steinbaugh
}
