% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,ANY-method}
\alias{export,sparseMatrix-method}
\title{Export}
\usage{
\S4method{export}{ANY}(x, file, format, ...)

\S4method{export}{sparseMatrix}(x, file, format)
}
\arguments{
\item{x}{Object.}

\item{file}{A character string naming a file. Must specify \code{file} and/or \code{format}.}

\item{format}{An optional character string containing the file format, which can be used to override the format inferred from \code{file} or, in lieu of specifying \code{file}, a file with the symbol name of \code{x} and the specified file extension will be created. Must specify \code{file} and/or \code{format}. Shortcuts include: \dQuote{,} (for comma-separated values), \dQuote{;} (for semicolon-separated values), \dQuote{|} (for pipe-separated values), and \dQuote{dump} for \code{\link[base]{dump}}.}

\item{...}{Additional arguments for the underlying export functions. See examples.}
}
\description{
Export data out of R and write to disk.
}
\details{
This is a wrapper for \code{\link[rio:export]{rio::export()}} that adds support for additional S4
classes in Bioconductor.
}
\examples{
export(mtcars, format = "csv")
}
\seealso{
\code{\link[rio:export]{rio::export()}}.
}
\author{
Michael Steinbaugh
}
