% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,ANY-method}
\alias{export,sparseMatrix-method}
\alias{export,SummarizedExperiment-method}
\title{Export}
\usage{
export(x, ...)

\S4method{export}{ANY}(x, file, format, ...)

\S4method{export}{sparseMatrix}(x, file, format)

\S4method{export}{SummarizedExperiment}(x, dir = ".")
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments for the underlying export functions. See examples.}

\item{file}{A character string naming a file. Must specify \code{file} and/or \code{format}.}

\item{format}{An optional character string containing the file format, which can be used to override the format inferred from \code{file} or, in lieu of specifying \code{file}, a file with the symbol name of \code{x} and the specified file extension will be created. Must specify \code{file} and/or \code{format}. Shortcuts include: \dQuote{,} (for comma-separated values), \dQuote{;} (for semicolon-separated values), \dQuote{|} (for pipe-separated values), and \dQuote{dump} for \code{\link[base]{dump}}.}
}
\value{
Invisible \code{character}. File path(s).
}
\description{
Export data out of R and write to disk.
}
\details{
This is a wrapper for \code{\link[rio:export]{rio::export()}} that adds support for additional S4
classes in Bioconductor.
}
\note{
The standard \code{\link[rio:export]{rio::export()}} call will drop rownames when exporting to
CSV. We're performing any internal tibble coercion step to ensure rownames
are always moved to a "rowname" column in the CSV output.
}
\examples{
export(rnaseq_counts, format = "csv")
export(single_cell_counts, format = "mtx")

# Clean up
file.remove(c(
    "rnaseq_counts.csv",
    "single_cell_counts.mtx",
    "single_cell_counts.mtx.colnames",
    "single_cell_counts.mtx.rownames"
))
}
\seealso{
\code{\link[rio:export]{rio::export()}}.
}
\author{
Michael Steinbaugh
}
