% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{filterCells}
\alias{filterCells}
\title{Filter cells}
\usage{
filterCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Filter cells
}
\examples{
isS4(filterCells)
}
\seealso{
\itemize{
\item bcbioSingleCell.
\item Chromium.
}
}
