% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells-methods.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells,SingleCellExperiment-method}
\title{Filter cells}
\usage{
filterCells(object, ...)

\S4method{filterCells}{SingleCellExperiment}(object, nCells = Inf,
  minCounts = 1L, maxCounts = Inf, minFeatures = 1L,
  maxFeatures = Inf, minNovelty = 0L, maxMitoRatio = 1L,
  minCellsPerFeature = 1L)
}
\arguments{
\item{object}{Object.}

\item{nCells}{\code{integer(1)}.
Expected number of cells per sample.
Don't set this by default, unless you're confident of your capture.}

\item{minCounts, maxCounts}{\code{integer(1)}.
Minimum/maximum number of counts per cell.
Applies to UMI disambiguated counts for droplet scRNA-seq.
Matches \code{nUMI} then \code{nCount} column in
\code{\link[=SummarizedExperiment]{colData()}} internally.
Previously named \code{minUMIs}/\code{maxUMIs} in bcbioSingleCell.}

\item{minFeatures, maxFeatures}{\code{integer(1)}.
Minimum/maximum number of features (i.e. genes) detected.
Matches \code{nFeature}in \code{\link[=SummarizedExperiment]{colData()}} internally.
Previously named \code{minGenes}/\code{maxGenes} in bcbioSingleCell.}

\item{minNovelty}{\code{integer(1)} (\code{0}-\code{1}).
Minimum novelty score (log10 features per UMI).
Matches \code{log10FeaturesPerCount} then \code{log10FeaturesPerUMI} (legacy)
\code{\link[=SummarizedExperiment]{colData()}} internally.}

\item{maxMitoRatio}{\code{integer(1)} (\code{0}-\code{1}).
Maximum relative mitochondrial abundance.}

\item{minCellsPerFeature}{\code{integer(1)}.
Include genes with non-zero expression in at least this many cells.
Previously named \code{minCellsPerGene} in bcbioSingleCell.}

\item{...}{Additional arguments.}
}
\value{
\code{SingleCellExperiment}.
Filtering information gets slotted into \code{\link[S4Vectors:metadata]{metadata()}}
as \code{filterCells} metadata.
}
\description{
Filter cells
}
\details{
Apply gene detection, novelty score, and mitochondrial abundance cutoffs to
cellular barcodes. By default we recommend applying the same filtering cutoff
to all samples. The filtering parameters now support per-sample cutoffs,
defined using a named \code{numeric} vector. When matching per sample, be sure to
use the \code{\link[=sampleNames]{sampleNames()}} return values (i.e. the \code{sampleName} column in
\code{\link[=sampleData]{sampleData()}}).
}
\note{
Updated 2019-08-08.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- calculateMetrics(object)

show(object)
x <- filterCells(object)
show(x)

## Per sample cutoffs.
sampleNames(object)
x <- filterCells(
    object = object,
    minCounts = c(sample1 = 100L)
)
show(x)
}
\author{
Michael Steinbaugh
}
