% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fixNA.R
\docType{methods}
\name{fixNA}
\alias{fixNA}
\alias{fixNA,ANY-method}
\alias{fixNA,character-method}
\alias{fixNA,data.frame-method}
\alias{fixNA,DataFrame-method}
\alias{fixNA,tbl_df-method}
\title{Fix Character Strings Missing \code{NA}}
\usage{
fixNA(object, ...)

\S4method{fixNA}{ANY}(object)

\S4method{fixNA}{character}(object)

\S4method{fixNA}{data.frame}(object)

\S4method{fixNA}{DataFrame}(object)

\S4method{fixNA}{tbl_df}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Object containing proper \code{NA} values.
}
\description{
Fix Character Strings Missing \code{NA}
}
\examples{
fixNA(c(1L, "x", "", "NA", "NULL"))

data.frame(a = c("foo", ""),
           b = c(NA, "bar")) \%>\%
    fixNA()
}
\seealso{
Other Cleanup Utilities: \code{\link{removeNA}},
  \code{\link{sortUnique}}
}
\concept{Cleanup Utilities}
