% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/zzz.R
\name{fixNA}
\alias{fixNA}
\alias{fix_na}
\title{Fix character strings missing \code{NA}}
\usage{
fixNA(string)
}
\arguments{
\item{string}{String missing \code{NA}.}
}
\value{
String containing \code{NA}.
}
\description{
Fix character strings missing \code{NA}
}
\examples{
fixNA(c(1, "x", "", "NA"))
}
