% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNA.R
\name{fixNA}
\alias{fixNA}
\title{Fix empty and "NA" character strings.}
\usage{
fixNA(a)
}
\arguments{
\item{a}{Values missing \code{NA}.}
}
\value{
Values containing \code{NA}.
}
\description{
Fix empty and "NA" character strings.
}
\examples{
fixNA(c(1, "x", "", "NA"))
}

