% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fixNA.R
\docType{methods}
\name{fixNA}
\alias{fixNA}
\alias{fixNA}
\alias{fixNA,ANY-method}
\alias{fixNA,character-method}
\alias{fixNA,data.frame-method}
\alias{fixNA,DataFrame-method}
\alias{fixNA,tbl_df-method}
\title{Fix Character Strings Missing \code{NA}}
\usage{
fixNA(object, ...)

\S4method{fixNA}{ANY}(object)

\S4method{fixNA}{character}(object)

\S4method{fixNA}{data.frame}(object)

\S4method{fixNA}{DataFrame}(object)

\S4method{fixNA}{tbl_df}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Object containing proper \code{NA} values.
}
\description{
Fix Character Strings Missing \code{NA}
}
\examples{
# character ====
fixNA(c(1L, "x", "", "NA", "NULL"))

# data.frame ====
x <- data.frame(
    a = c("foo", ""),
    b = c(NA, "bar"),
    row.names = c("a", "b"),
    stringsAsFactors = FALSE
)
fixNA(x)

# DataFrame ====

}
\seealso{
Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{removeNA}}, \code{\link{sanitizeRowData}},
  \code{\link{sanitizeSampleData}}
}
\author{
Michael Steinbaugh
}
