% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/flatFiles-methods.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{flatFiles,SummarizedExperiment-method}
\title{Flat Files from S4 Object}
\usage{
flatFiles(object, ...)

\S4method{flatFiles}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list} containing the S4 object slots.
}
\description{
Extract the slots inside an S4 object for archival storage.
}
\examples{
# SummarizedExperiment ====
x <- flatFiles(rse_small)
class(x)
names(x)
}
\seealso{
Other Developer Functions: \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{matchS4Call}}, \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
