% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{formalsList}
\alias{formalsList}
\title{Formals List
Formals that the user can set globally with \code{base::getOption}.}
\format{An object of class \code{list} of length 14.}
\usage{
formalsList
}
\description{
Formals List
Formals that the user can set globally with \code{base::getOption}.
}
\examples{
formalsList
}
\keyword{datasets}
