% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Gene2Symbol-class}
\alias{Gene2Symbol-class}
\title{Gene-to-Symbol Mappings}
\value{
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\description{
Gene-to-Symbol Mappings
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:metadata]{S4Vectors::metadata()}}.
}

\examples{
data(rse_small)
x <- Gene2Symbol(rse_small)
print(x)
}
\seealso{
\code{\link[=Gene2Symbol]{Gene2Symbol()}}, \link{makeGene2Symbol}.

Other S4 classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 classes}
