% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Gene2Symbol-class}
\alias{Gene2Symbol-class}
\title{Gene-to-Symbol Mappings}
\value{
\code{Gene2Symbol}. Contains a \code{DataFrame} with \code{geneID} and \code{geneName}
columns.
}
\description{
Gene-to-Symbol Mappings
}
\section{Genome metadata}{



We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:metadata]{S4Vectors::metadata()}}.

}

\examples{
data(rse, package = "basejump.data")
x <- Gene2Symbol(rse)
print(x)
}
\seealso{
\code{help(topic = "makeGene2Symbol", package = "basejump.annotations")}.
}
\keyword{internal}
