% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/AllGenerators.R
\docType{class}
\name{Gene2Symbol-class}
\alias{Gene2Symbol-class}
\alias{Gene2Symbol}
\alias{Gene2Symbol,DataFrame-method}
\alias{Gene2Symbol,GRanges-method}
\alias{Gene2Symbol,SummarizedExperiment-method}
\title{Gene-to-symbol mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{DataFrame}(object, format = c("makeUnique",
  "unmodified", "1:1"))

\S4method{Gene2Symbol}{GRanges}(object, format = c("makeUnique",
  "unmodified", "1:1"))

\S4method{Gene2Symbol}{SummarizedExperiment}(object,
  format = c("makeUnique", "unmodified", "1:1"))
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-symbol mappings
}
\details{
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\note{
Updated 2019-08-08.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:metadata]{metadata()}}.
}

