% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Gene2Symbol-class}
\alias{Gene2Symbol-class}
\title{Gene-to-symbol mappings}
\value{
\code{Gene2Symbol}.
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\description{
Gene-to-symbol mappings
}
\note{
For some organisms, gene names and gene symbols do not map 1:1
(e.g. \emph{Homo sapiens} and \emph{Mus musculus}). Refer to the \code{format} argument here
in the documentation for approaches that deal with this issue.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\link[S4Vectors:metadata]{metadata}.
}

\seealso{
\code{makeGene2Symbol}.
}
\author{
Michael Steinbaugh
}
