% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/Gene2Symbol-methods.R
\docType{class}
\name{Gene2Symbol-class}
\alias{Gene2Symbol-class}
\alias{Gene2Symbol}
\alias{Gene2Symbol,DataFrame-method}
\alias{Gene2Symbol,GRanges-method}
\alias{Gene2Symbol,SummarizedExperiment-method}
\title{Gene-to-symbol mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{DataFrame}(object, format = c("makeUnique",
  "unmodified", "1:1"))

\S4method{Gene2Symbol}{GRanges}(object, format = c("makeUnique",
  "unmodified", "1:1"))

\S4method{Gene2Symbol}{SummarizedExperiment}(object,
  format = c("makeUnique", "unmodified", "1:1"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"makeUnique"}: \emph{Recommended.} Apply \code{\link[base:make.unique]{make.unique()}}
to the \code{geneName} column. Gene symbols are made unique, while the gene
IDs remain unmodified.
\item \code{"unmodified"}: Return \code{geneID} and \code{geneName} columns unmodified, in
long format.
\item \code{"1:1"}: For gene symbols that map to multiple gene IDs, select only the
first annotated gene ID.
}}
}
\value{
\code{Gene2Symbol}.
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\description{
Gene-to-symbol mappings
}
\note{
For some organisms, gene names and gene symbols do not map 1:1
(e.g. \emph{Homo sapiens} and \emph{Mus musculus}). Refer to the \code{format} argument here
in the documentation for approaches that deal with this issue.

For the \code{format} argument, note that "long" was used instead of
"unmodified" prior to v0.10.10.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:metadata]{metadata()}}.
}

\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
x <- Gene2Symbol(rse)
print(x)
}
\seealso{
\code{makeGene2Symbol()}.
}
\author{
Michael Steinbaugh
}
