% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Gene2Symbol-class}
\alias{Gene2Symbol-class}
\title{Gene-to-Symbol Mappings}
\value{
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\description{
Gene-to-Symbol Mappings
}
\examples{
# SummarizedExperiment ====
data(rse_small)
x <- gene2symbol(rse_small)
print(x)
}
\seealso{
\itemize{
\item \code{\link[=gene2symbol]{gene2symbol()}}.
\item \link{makeGene2Symbol}.
}

Other S4 Classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
