% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{gene2symbol-class}
\alias{gene2symbol-class}
\title{\code{gene2symbol} Class}
\value{
\code{gene2symbol}.
}
\description{
Gene-to-symbol mappings.
}
\details{
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\examples{
x <- new(
    "gene2symbol",
    DataFrame(
        geneID = "ENSG00000000003",
        geneName = "TSPAN6",
        row.names = "ENSG00000000003"
    )
)
print(x)
}
\seealso{
Other S4 Classes: \code{\link{PANTHER-class}},
  \code{\link{ensembl2entrez-class}},
  \code{\link{hgnc2ensembl-class}},
  \code{\link{mgi2ensembl-class}},
  \code{\link{tx2gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
