% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Gene2Symbol-class}
\alias{Gene2Symbol-class}
\title{Gene-to-symbol mappings}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-symbol mappings
}
\details{
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\note{
Updated 2019-08-08.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:Annotated-class]{metadata()}}.
}

