% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Gene2Symbol-methods.R
\docType{methods}
\name{Gene2Symbol}
\alias{Gene2Symbol}
\alias{Gene2Symbol,DataFrame-method}
\alias{Gene2Symbol,GRanges-method}
\alias{Gene2Symbol,SummarizedExperiment-method}
\title{Gene-to-symbol mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{DataFrame}(object, format = c("makeUnique", "1:1",
  "long"))

\S4method{Gene2Symbol}{GRanges}(object, format = c("makeUnique", "1:1",
  "long"))

\S4method{Gene2Symbol}{SummarizedExperiment}(object,
  format = c("makeUnique", "1:1", "long"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"makeUnique"}: \emph{Recommended.} Apply \code{make.unique} to the \code{geneName}
column. Gene symbols are made unique, while the gene IDs remain
unmodified.
\item \code{"1:1"}: For gene symbols that map to multiple gene IDs, select only the
first annotated gene ID.
\item \code{"long"}: Return \code{geneID} and \code{geneName} columns unmodified in long
format.
}}
}
\value{
\code{Gene2Symbol}.
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\description{
Gene-to-symbol mappings
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\link[S4Vectors:metadata]{metadata}.
}

\examples{
data(rse)
x <- Gene2Symbol(rse)
print(x)
}
\seealso{
\code{makeGene2Symbol}.
}
