% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Gene2Symbol-methods.R
\docType{methods}
\name{Gene2Symbol}
\alias{Gene2Symbol}
\alias{Gene2Symbol,DataFrame-method}
\alias{Gene2Symbol,GRanges-method}
\alias{Gene2Symbol,SummarizedExperiment-method}
\title{Gene-to-Symbol Mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{DataFrame}(object)

\S4method{Gene2Symbol}{GRanges}(object)

\S4method{Gene2Symbol}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-Symbol Mappings
}
\note{
\itemize{
\item This function will make any duplicated symbols unique by applying
\code{\link[base:make.unique]{base::make.unique()}}, which will add ".1" to the end of the gene name.
\item No attempt is made to arrange the rows by gene identifier.
}
}
\examples{
data(rse_small)
x <- Gene2Symbol(rse_small)
print(x)
}
\seealso{
\code{\link[=Gene2Symbol]{Gene2Symbol()}}, \link{makeGene2Symbol}.

Other S4 classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
