% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gene2symbol-methods.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,GRanges-method}
\alias{gene2symbol,SummarizedExperiment-method}
\title{Gene-to-Symbol Mappings}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{GRanges}(object)

\S4method{gene2symbol}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-Symbol Mappings
}
\note{
This function will make any duplicated symbols unique by applying
\code{\link[base:make.unique]{base::make.unique()}}, which will add ".1" to the end of the gene name.
}
\examples{
# SummarizedExperiment ====
x <- gene2symbol(rse_small)
print(x)
}
\seealso{
Other S4 Generators: \code{\link{eggnog}},
  \code{\link{ensembl2entrez}}, \code{\link{hgnc2ensembl}},
  \code{\link{mgi2ensembl}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
