% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,character-method}
\title{Gene to Symbol Mappings}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{character}(object, genomeBuild = NULL,
  release = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}
}
\value{
\code{data.frame}.
}
\description{
Gene to Symbol Mappings
}
\examples{
gene2symbol("Homo sapiens") \%>\% glimpse()
}
\seealso{
Other Gene Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{ensembl}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{tx2gene}}
}
