% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,character-method}
\alias{gene2symbol,data.frame-method}
\alias{gene2symbol,DataFrame-method}
\alias{gene2symbol,dgCMatrix-method}
\alias{gene2symbol,dgTMatrix-method}
\alias{gene2symbol,matrix-method}
\title{Convert Ensembl Identifier to Gene Symbol}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{character}(object, organism = NULL)

\S4method{gene2symbol}{data.frame}(object, organism = NULL)

\S4method{gene2symbol}{DataFrame}(object, organism = NULL)

\S4method{gene2symbol}{dgCMatrix}(object, organism = NULL)

\S4method{gene2symbol}{dgTMatrix}(object, organism = NULL)

\S4method{gene2symbol}{matrix}(object, organism = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{organism}{\emph{Optional}. Organism name. Normally this argument is
unnecessary and can be left \code{NULL}. If a count matrix starts with a
FASTA spike-in (e.g. "EGFP"), then automatic genome detection based on the
first gene identifier will fail. In this case, the desired organism must be
manually declared.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Identifier to Gene Symbol
}
\examples{
# character
c("ENSMUSG00000000001",
  "ENSMUSG00000000003") \%>\% gene2symbol

# matrix
matrix(
    data = seq(1L:4L),
    byrow = TRUE,
    nrow = 2L,
    ncol = 2L,
    dimnames = list(c("ENSMUSG00000000001",
                      "ENSMUSG00000000003"),
                    c("sample1",
                      "sample2"))) \%>\% gene2symbol
}
\seealso{
\code{\link[=detectOrganism]{detectOrganism()}}.

Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGTF}},
  \code{\link{symbol2gene}}, \code{\link{tx2geneFromGTF}},
  \code{\link{tx2gene}}
}
\concept{Gene Annotation Utilities}
