% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gene2symbol-methods.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,SummarizedExperiment-method}
\title{Gene to Symbol Mappings}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Gene to Symbol Mappings
}
\note{
This function will make any duplicated symbols unique by applying
\code{\link[base:make.unique]{base::make.unique()}}, which will add ".1" to the end of the gene name.
}
\examples{
# SummarizedExperiment ====
gene2symbol(rse_small) \%>\% head()
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
