% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gene2symbol-methods.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{Gene2Symbol}
\alias{gene2symbol,data.frame-method}
\alias{gene2symbol,DataFrame-method}
\alias{gene2symbol,GRanges-method}
\alias{gene2symbol,SummarizedExperiment-method}
\title{Gene-to-Symbol Mappings}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{data.frame}(object)

\S4method{gene2symbol}{DataFrame}(object)

\S4method{gene2symbol}{GRanges}(object)

\S4method{gene2symbol}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-Symbol Mappings
}
\note{
\itemize{
\item This function will make any duplicated symbols unique by applying
\code{\link[base:make.unique]{base::make.unique()}}, which will add ".1" to the end of the gene name.
\item No attempt is made to arrange the rows by gene identifier.
}
}
\examples{
# SummarizedExperiment ====
data(rse_small)
x <- gene2symbol(rse_small)
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{ensembl2entrez}}, \code{\link{hgnc2ensembl}},
  \code{\link{mapGenes}}, \code{\link{mgi2ensembl}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Identifier Mapping Functions}
