% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gene2symbol-methods.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,SummarizedExperiment-method}
\title{Gene-to-Symbol Mappings}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{gene2symbol}.
}
\description{
Gene-to-Symbol Mappings
}
\note{
This function will make any duplicated symbols unique by applying
\code{\link[base:make.unique]{base::make.unique()}}, which will add ".1" to the end of the gene name.
}
\examples{
# SummarizedExperiment ====
x <- gene2symbol(rse_small)
print(x)
}
\seealso{
Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{eggnog}}, \code{\link{emptyRanges}},
  \code{\link{ensembl2entrez}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2ensembl}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2ensembl}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
