% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,character-method}
\title{Convert Ensembl Identifier to Gene Symbol}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Convert Ensembl Identifier to Gene Symbol
}
\examples{
# character
gene2symbol(c("ENSMUSG00000000001", "ENSMUSG00000000003"))
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}}, \code{\link{gene2entrez}},
  \code{\link{tx2gene}}
}
