% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,character-method}
\title{Ensembl Gene to Symbol Mappings}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{character}(object, genomeBuild = NULL,
  release = NULL, uniqueSymbol = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{data.frame}.
}
\description{
Ensembl Gene to Symbol Mappings
}
\examples{
gene2symbol("Homo sapiens") \%>\% glimpse()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}},
  \code{\link{ensemblAnnotations}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{tx2geneFromGFF}}, \code{\link{tx2gene}}
}
