% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,character-method}
\title{Ensembl Gene to Symbol Mappings}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{character}(object, genomeBuild = NULL,
  release = NULL, uniqueSymbol = FALSE, quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{genomeBuild}{\emph{Optional}. Check to see if a specific genome build is
supported.}

\item{release}{\emph{Optional}. Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{uniqueSymbol}{Make gene symbols unique.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Ensembl Gene to Symbol Mappings
}
\examples{
gene2symbol("Homo sapiens") \%>\% glimpse()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{symbol2gene}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
