% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,character-method}
\alias{gene2symbol,data.frame-method}
\alias{gene2symbol,DataFrame-method}
\alias{gene2symbol,dgCMatrix-method}
\alias{gene2symbol,dgTMatrix-method}
\alias{gene2symbol,matrix-method}
\title{Convert Ensembl Identifier to Gene Symbol}
\usage{
gene2symbol(object, ...)

\S4method{gene2symbol}{character}(object, genomeBuild = NULL)

\S4method{gene2symbol}{data.frame}(object, genomeBuild = NULL)

\S4method{gene2symbol}{DataFrame}(object, genomeBuild = NULL)

\S4method{gene2symbol}{dgCMatrix}(object, genomeBuild = NULL)

\S4method{gene2symbol}{dgTMatrix}(object, genomeBuild = NULL)

\S4method{gene2symbol}{matrix}(object, genomeBuild = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genomeBuild}{Genome build.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Identifier to Gene Symbol
}
\examples{
# character
c("ENSMUSG00000000001",
  "ENSMUSG00000000003") \%>\% gene2symbol

# matrix
matrix(
    data = seq(1L:4L),
    byrow = TRUE,
    nrow = 2L,
    ncol = 2L,
    dimnames = list(c("ENSMUSG00000000001",
                      "ENSMUSG00000000003"),
                    c("sample1",
                      "sample2"))) \%>\% gene2symbol
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGTF}},
  \code{\link{tx2geneFromGTF}}, \code{\link{tx2gene}}
}
