% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2symbolFromGFF.R
\name{gene2symbolFromGFF}
\alias{gene2symbolFromGFF}
\alias{gene2symbolFromGTF}
\title{Define Gene to Symbol Mappings from GFF/GTF File}
\usage{
gene2symbolFromGFF(file)

gene2symbolFromGTF(file)
}
\arguments{
\item{file}{File path.}
}
\value{
\code{data.frame}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
gene2symbolFromGFF("http://basejump.seq.cloud/mmusculus.gtf") \%>\% glimpse()
}
\seealso{
Other GFF Functions: \code{\link{readGFF}},
  \code{\link{rowRangesFromGFF}},
  \code{\link{tx2geneFromGFF}}
}
