% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbolFromGFF.R
\docType{methods}
\name{gene2symbolFromGFF}
\alias{gene2symbolFromGFF}
\alias{gene2symbolFromGFF,character-method}
\alias{gene2symbolFromGFF,data.frame-method}
\alias{gene2symbolFromGTF}
\title{Generate gene2symbol from GFF/GTF File}
\usage{
gene2symbolFromGFF(object, ...)

\S4method{gene2symbolFromGFF}{character}(object)

\S4method{gene2symbolFromGFF}{data.frame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame}.
}
\description{
Generate gene2symbol from GFF/GTF File
}
\examples{
# From URL (recommended)
url <- file.path(testDataURL, "mmusculus.gtf")
gene2symbolFromGFF(url) \%>\%
    str()

# GFF data.frame
gff <- readGFF(url)
gene2symbolFromGFF(gff) \%>\%
    str()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}}, \code{\link{gene2symbol}},
  \code{\link{symbol2gene}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
\concept{Gene Annotation Utilities}
