% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbolFromGTF.R
\docType{methods}
\name{gene2symbolFromGTF}
\alias{gene2symbolFromGTF}
\alias{gene2symbolFromGTF}
\alias{gene2symbolFromGTF,character-method}
\title{Generate gene2symbol from GTF File}
\usage{
gene2symbolFromGTF(object, ...)

\S4method{gene2symbolFromGTF}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{data.frame}.
}
\description{
Generate gene2symbol from GTF File

Generate gene2symbol from GTF File
}
\examples{
# Mouse (Ensembl)
gene2symbolFromGTF("http://steinbaugh.com/basejump/tests/mmusculus.gtf")

# Fruitfly (FlyBase)
gene2symbolFromGTF("http://steinbaugh.com/basejump/tests/dmelanogaster.gtf")
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}}, \code{\link{gene2symbol}},
  \code{\link{tx2geneFromGTF}}, \code{\link{tx2gene}}
}
