% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbolFromGTF.R
\docType{methods}
\name{gene2symbolFromGTF}
\alias{gene2symbolFromGTF}
\alias{gene2symbolFromGTF,character-method}
\alias{gene2symbolFromGTF,data.frame-method}
\title{Generate gene2symbol from GTF File}
\usage{
gene2symbolFromGTF(object, ...)

\S4method{gene2symbolFromGTF}{character}(object)

\S4method{gene2symbolFromGTF}{data.frame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame}.
}
\description{
Generate gene2symbol from GTF File
}
\examples{
# GTF URL
url <- file.path(testDataURL, "mmusculus.gtf")
gene2symbolFromGTF(url) \%>\%
    str

# GTF data.frame
gtf <- readGTF(url)
gene2symbolFromGTF(gtf) \%>\%
    str
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}}, \code{\link{gene2symbol}},
  \code{\link{symbol2gene}}, \code{\link{tx2geneFromGTF}},
  \code{\link{tx2gene}}
}
\concept{Gene Annotation Utilities}
