% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{atomic}{\code{atomic} vector.}

\item{object}{Object.}

\item{value}{Value to assign.}

\item{x}{Object.}

\item{y}{Secondary object.}

\item{...}{Additional arguments.}

\item{dir}{Output directory. Defaults to the current working directory.}

\item{envir}{Environment to use for assignment. Defaults to \code{parent.frame()},
which will assign into the calling environment.}

\item{file}{File path.}

\item{headerLevel}{Markdown header level.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{organism}{Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{plotlist}{\code{list} containing \code{ggplot} objects.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\description{
General Arguments
}
\author{
Michael Steinbaugh
}
\keyword{internal}
