% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{file}{File path.}

\item{headerLevel}{Markdown header level.}

\item{object}{Object.}

\item{plotlist}{List containing \code{ggplot} objects.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}

\item{value}{Value to assign.}

\item{x}{Primary object.}

\item{y}{Secondary object.}

\item{...}{Additional arguments.}
}
\description{
General Arguments
}
\author{
Michael Steinbaugh
}
\keyword{internal}
