% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{arrange}
\alias{arrange}
\alias{filter}
\alias{mutate}
\title{All generics}
\usage{
arrange(object, ...)

filter(object, ...)

mutate(object, ...)
}
\description{
All generics
}
\keyword{internal}
