% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-genes.R
\docType{methods}
\name{genes}
\alias{genes}
\alias{genes,character-method}
\alias{annotable}
\title{Gene Annotations}
\usage{
\S4method{genes}{character}(x, genomeBuild = NULL, release = NULL,
  uniqueSymbol = FALSE, return = "GRanges")

annotable(object, ..., return = "data.frame")
}
\arguments{
\item{uniqueSymbol}{Make gene symbols unique.}
}
\value{
\code{GRanges}, \code{data.frame}, or \code{DataFrame}.
}
\description{
Gene Annotations
}
\section{Broad Class Definitions}{

A \code{broadClass} column is added, which generalizes the gene types into a
smaller number of semantically-meaningful groups:
\itemize{
\item \code{coding}
\item \code{noncoding}
\item \code{pseudo}
\item \code{small}
\item \code{decaying}
\item \code{ig} (immunoglobulin)
\item \code{tcr} (T cell receptor)
\item \code{other}
}
}

\examples{
# Legacy GRCh37/hg19 genome build support
genes("Homo sapiens", genomeBuild = "GRCh37") \%>\% glimpse()
}
