% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-genes.R
\docType{methods}
\name{genes}
\alias{genes}
\alias{genes,character-method}
\title{Gene Annotations}
\usage{
\S4method{genes}{character}(x, genomeBuild = NULL, release = NULL,
  uniqueSymbol = FALSE, return = "GRanges")
}
\arguments{
\item{x}{Primary object.}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{uniqueSymbol}{Make gene symbols unique. Only applies to \code{genes} and
\code{gene2symbol} output.}

\item{return}{Class of the returned object. Can be "\code{data.frame}",
"\code{DataFrame}" or "\code{GRanges}. See \code{help("genes", "ensembldb")} for
additional information.}
}
\value{
\code{GRanges}, \code{data.frame}, or \code{DataFrame}.
}
\description{
Gene Annotations
}
\details{
Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.
}
\section{Broad Class Definitions}{

For gene and transcript tables, when \code{broadClass = TRUE}, a \code{broadClass}
column is added, which generalizes the gene types into a smaller number of
semantically-meaningful groups:
\itemize{
\item \code{coding}
\item \code{noncoding}
\item \code{pseudo}
\item \code{small}
\item \code{decaying}
\item \code{ig} (immunoglobulin)
\item \code{tcr} (T cell receptor)
\item \code{other}
}
}

\examples{
# GRanges return
data <- genes("Homo sapiens")
summary(data)
colnames(mcols(data))

# Legacy GRCh37/hg19 genome build support
genes("Homo sapiens", genomeBuild = "GRCh37") \%>\% glimpse()
}
\seealso{
Other Gene Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{ensembl}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{transcripts}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
