% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSynonyms.R
\name{geneSynonyms}
\alias{geneSynonyms}
\title{Gene synonyms}
\usage{
geneSynonyms(organism = c("Homo sapiens", "Mus musculus",
  "Drosophila melanogaster"))
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}
}
\value{
\code{grouped_df}.
Grouped by \code{geneID} column.
}
\description{
Look up gene synonyms from NCBI.
}
\note{
Synonym support for \emph{Caenorhabditis elegans} is poor on NCBI.
Use the \href{https://steinbaugh.com/wormbase/}{wormbase} package instead.

Updated 2019-07-28.
}
\examples{
options(acid.test = TRUE)
x <- geneSynonyms(organism = "Homo sapiens")
print(x)
}
