% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSynonyms.R
\name{geneSynonyms}
\alias{geneSynonyms}
\title{Gene synonyms}
\usage{
geneSynonyms(organism)
}
\arguments{
\item{organism}{\code{character(1)}.
Supported organisms:
\itemize{
\item \emph{Homo sapiens}.
\item \emph{Mus musculus}.
\item \emph{Drosophila melanogaster}.
}}
}
\value{
\code{grouped_df}.
Grouped by \code{geneID} column.
}
\description{
Look up gene synonyms from NCBI.
}
\note{
Synonym support for \emph{Caenorhabditis elegans} is poor on NCBI.
Use the \href{https://steinbaugh.com/wormbase/}{wormbase} package instead.
}
\examples{
options(acid.test = TRUE)
x <- geneSynonyms(organism = "Homo sapiens")
print(x)
}
