% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSynonyms.R
\name{geneSynonyms}
\alias{geneSynonyms}
\title{Gene Synonyms}
\usage{
geneSynonyms(organism)
}
\arguments{
\item{organism}{\code{string}. Supported organisms: \emph{Homo sapiens},
\emph{Mus musculus}, \emph{Drosophila melanogaster}.}
}
\value{
\code{grouped_df}. Grouped by \code{geneID} column.
}
\description{
Look up gene synonyms from NCBI.
}
\note{
Synonym support for \emph{Caenorhabditis elegans} is poor on NCBI.
Use the \href{https://steinbaugh.com/wormbase/}{wormbase} package instead.
}
\examples{
options(basejump.test = TRUE)
x <- geneSynonyms(organism = "Homo sapiens")
print(x)
}
