% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSynonyms.R
\name{geneSynonyms}
\alias{geneSynonyms}
\title{Gene Synonyms}
\usage{
geneSynonyms(organism, .test = FALSE)
}
\arguments{
\item{organism}{\code{string}. Supported organisms: \emph{Homo sapiens},
\emph{Mus musculus}, \emph{Drosophila melanogaster}.}

\item{.test}{\code{boolean}. \emph{For unit testing only}.}
}
\value{
\code{grouped_df}. Grouped by \code{geneID} column.
}
\description{
Look up gene synonyms from NCBI.
}
\note{
Synonym support for \emph{Caenorhabditis elegans} is poor on NCBI.
Use the wormbase package instead.
}
\examples{
x <- geneSynonyms(organism = "Homo sapiens", .test = TRUE)
print(x)
}
\seealso{
Other Annotation Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{emptyRanges}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2Symbol}}, \code{\link{makeTx2Gene}},
  \code{\link{organism}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
