% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSynonyms.R
\name{geneSynonyms}
\alias{geneSynonyms}
\title{Gene synonyms}
\usage{
geneSynonyms(
  organism = c("Homo sapiens", "Mus musculus", "Drosophila melanogaster"),
  return = c("DataFrame", "SplitDataFrameList")
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}
}
\value{
\itemize{
\item \code{DataFrame}:
Returns unique row for each \code{geneID}.
Returns only \code{geneID} and \code{geneSynonyms} columns.
The \code{geneSynonyms} column returns as character vector, with synonyms
arranged alphabetically and delimited by \code{", "}.
\item \code{SplitDataFrameList}:
Split by \code{geneID} column.
Keeps duplicate rows mapping to same \code{geneID} (e.g. ENSG00000004866).
Returns \code{geneID}, \code{geneName}, and \code{geneSynonyms} columns in the split.
}
}
\description{
Look up gene synonyms from NCBI.
}
\note{
Synonym support for \emph{Caenorhabditis elegans} is poor on NCBI.
Use the \href{https://wormbase.acidgenomics.com/}{wormbase} package instead.

Updated 2020-10-05.
}
\examples{
object <- geneSynonyms(organism = "Homo sapiens")
print(object)
}
