% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{genomeMetadataNames}
\alias{genomeMetadataNames}
\title{Slot Names in Metadata Containing Genome Information}
\format{An object of class \code{character} of length 3.}
\usage{
genomeMetadataNames
}
\description{
Slot Names in Metadata Containing Genome Information
}
\examples{
genomeMetadataNames
}
\keyword{datasets}
