% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{genomeMetadataNames}
\alias{genomeMetadataNames}
\title{Slot names in metadata containing genome information}
\format{
An object of class \code{character} of length 3.
}
\usage{
genomeMetadataNames
}
\description{
Slot names in metadata containing genome information
}
\note{
Updated 2019-08-21.
}
\examples{
genomeMetadataNames
}
\keyword{datasets}
