% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{geomean}
\alias{geomean}
\title{Geometric mean}
\usage{
geomean(object, na_rm = TRUE)
}
\arguments{
\item{object}{Vector or column data (data frame, matrix).}

\item{na_rm}{Remove \code{NA} values before processing.}
}
\value{
Geometric means.
}
\description{
The geometric mean is the nth root of n products or e to the mean log of \code{x}.
Useful for describing non-normal (i.e. geometric) distributions.
}
\details{
Modified version of \code{\link[psych:geometric.mean]{psych::geometric.mean()}}.
}
\note{
Not particularly useful if there are elements that are <= 0.
}
\examples{
# Vector
vec <- seq(1,5)
vec2 <- vec^2
vec2[2] <- NA
geomean(vec)
geomean(vec2)

# Data frame
df <- data.frame(vec, vec2)
geomean(df)
geomean(df, na_rm = FALSE)
}
