% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-geomean.R
\docType{methods}
\name{geomean}
\alias{geomean}
\alias{geomean}
\alias{geomean,data.frame-method}
\alias{geomean,integer-method}
\alias{geomean,matrix-method}
\alias{geomean,numeric-method}
\title{Geometric Mean}
\usage{
geomean(object, ...)

\S4method{geomean}{data.frame}(object)

\S4method{geomean}{integer}(object, removeNA = TRUE, zeroPropagate = FALSE)

\S4method{geomean}{matrix}(object)

\S4method{geomean}{numeric}(object, removeNA = TRUE, zeroPropagate = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{removeNA}{Remove \code{NA} values from calculations.}

\item{zeroPropagate}{Allow propagation of zeroes.}
}
\value{
Numeric containing geometric means.
}
\description{
The geometric mean is the nth root of n products or e to the mean log of \code{x}.
Useful for describing non-normal (i.e. geometric) distributions.
}
\details{
Modified version of \code{psych::geometric.mean()} and/or Paul McMurdie's
code.
}
\note{
This function should be fully zero- and \code{NA}-tolerant. This calculation
is not particularly useful if there are elements that are <= 0 and will
return \code{NaN}.
}
\examples{
# integer
vec <- seq(1L, 5L, 1L)
geomean(vec)
vec2 <- vec ^ 2L
geomean(vec2)

# data.frame
df <- data.frame(vec, vec2)
geomean(df)

# matrix
mat <- as.matrix(df)
geomean(mat)
}
\seealso{
\href{https://stackoverflow.com/a/25555105}{Paul McMurdie's post} on
Stack Overflow

Other Math Utilities: \code{\link{fc2lr}}
}
