% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-geomean.R
\docType{methods}
\name{geomean}
\alias{geomean}
\alias{geomean}
\alias{geomean,ANY-method}
\alias{geomean,data.frame-method}
\alias{geomean,matrix-method}
\alias{geomean,numeric-method}
\title{Geometric Mean}
\usage{
geomean(object, ...)

\S4method{geomean}{ANY}(object)

\S4method{geomean}{data.frame}(object)

\S4method{geomean}{matrix}(object)

\S4method{geomean}{numeric}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Numeric containing geometric means.
}
\description{
The geometric mean is the nth root of n products or e to the mean log of \code{x}.
Useful for describing non-normal (i.e. geometric) distributions.
}
\details{
Modified version of \code{psych::geometric.mean()}.
}
\note{
Not particularly useful if there are elements that are <= 0.
}
\examples{
# Vector
vec <- seq(1L, 5L, 1L)
geomean(vec)

vec2 <- vec ^ 2L
geomean(vec2)


# Data frame
df <- data.frame(vec, vec2)
geomean(df)


# Matrix
mat <- as.matrix(df)
geomean(mat)
}
\seealso{
Other Math Utilities: \code{\link{fc2lr}},
  \code{\link{pct}}
}
