% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geometricMean-methods.R
\name{geometricMean}
\alias{geometricMean}
\alias{geometricMean,Matrix-method}
\title{Geometric mean}
\usage{
geometricMean(object, ...)

\S4method{geometricMean}{Matrix}(x, MARGIN = 2L)
}
\arguments{
\item{x}{Object.}

\item{MARGIN}{a vector giving the subscripts which the function will
    be applied over.  E.g., for a matrix \code{1} indicates rows,
    \code{2} indicates columns, \code{c(1, 2)} indicates rows and
    columns. Where \code{X} has named dimnames, it can be a character
    vector selecting dimension names.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
}
\description{
The geometric mean is the nth root of n products or e to the mean log of \code{x}.
}
\details{
Useful for describing non-normal (i.e. geometric) distributions.
}
\note{
Updated 2021-02-02.
}
\examples{
## Matrix ====
sparse <- as(matrix, "sparseMatrix")
print(sparse)
geometricMean(sparse)
}
