% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsDb.R
\name{getEnsDb}
\alias{getEnsDb}
\title{Get EnsDb from Bioconductor AnnotationHub}
\usage{
getEnsDb(organism, genomeBuild = NULL, release = NULL)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}
}
\value{
\code{EnsDb}.
}
\description{
Get EnsDb from Bioconductor AnnotationHub
}
\details{
Remaps UCSC genome build to Ensembl automatically, if necessary.
Provides legacy support for GRCh37 (hg19).
}
\note{
Updated 2020-09-25.
}
\examples{
getEnsDb(organism = "Homo sapiens", release = 100L)
}
