% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\docType{data}
\name{globals}
\alias{globals}
\title{Global variables}
\format{An object of class \code{character} of length 2.}
\usage{
globals
}
\description{
Global variables
}
\seealso{
\itemize{
\item \code{\link[=globalVariables]{globalVariables()}}.
\item \link{reexports}.
}
}
\keyword{internal}
