% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goseq_exec.R
\name{goseq_exec}
\alias{goseq_exec}
\title{Execute goseq}
\usage{
goseq_exec(deg, bg, envir = .GlobalEnv, save_dir = ".")
}
\arguments{
\item{save_dir}{}
}
\description{
Allows for input of differentially expressed genes (DEG),
background genes (bg), output environemnt, and data save folder.
}

