% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepString.R
\name{grepString}
\alias{grepString}
\title{Generate a \code{grep} String}
\usage{
grepString(object, sep = ", ")
}
\arguments{
\item{object}{\code{string}. Term to use for grep matching.}

\item{sep}{\code{string}. Separator. Defaults to comma.}
}
\value{
\code{string}. Regular expression pattern.
}
\description{
A \code{grep} string is useful for pattern matching against a delimited
\code{character} string returned by \code{\link[base:toString]{base::toString()}} or \code{\link[base:paste]{base::paste()}} using
the \code{collapse} argument.
}
\examples{
grepString("gene")
}
\seealso{
\itemize{
\item \code{\link[base:grep]{base::grep()}}.
\item \code{\link[base:toString]{base::toString()}}.
}

Other Atomic Vector Functions: \code{\link{detectLanes}},
  \code{\link{sortUnique}}, \code{\link{toStringUnique}}
}
\author{
Michael Steinbaugh
}
\concept{Atomic Vector Functions}
