% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepString.R
\name{grepString}
\alias{grepString}
\title{\code{grep} String}
\usage{
grepString(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character} string for \code{\link[=grep]{grep()}} matching.
}
\description{
Generate a \code{\link[=grep]{grep()}} string for pattern matching against comma separated
\code{\link[=toString]{toString()}} output.
}
\examples{
grepString("gene")
}
\seealso{
Other Atomic Vector Functions: \code{\link{sortUnique}},
  \code{\link{toStringUnique}}
}
\author{
Michael Steinbaugh
}
