% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-grepString.R
\docType{methods}
\name{grepString}
\alias{grepString}
\alias{grepString}
\alias{grepString,character-method}
\title{\code{grep} String Generator}
\usage{
grepString(object)

\S4method{grepString}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Comma separated grep string.
}
\description{
Generate a grep string for pattern matching against comma separated
\code{\link[base:toString]{base::toString()}} output.
}
\examples{
grepString("gene")
}
