% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grepString}
\alias{grepString}
\title{\code{\link[=grep]{grep()}} pattern string generator}
\usage{
grepString(identifier)
}
\arguments{
\item{identifier}{Identifier.}
}
\value{
Comma separated string for matching against \code{\link[=toString]{toString()}} return.
}
\description{
Generate a \code{\link[=grep]{grep()}} string for pattern matching against \code{\link[=toString]{toString()}} return
(comma separated).
}
\examples{
grepString("gene")
}
