% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepString.R
\name{grepString}
\alias{grepString}
\title{\code{grep} String Generator}
\usage{
grepString(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Comma delimited grep string.
}
\description{
Generate a grep string for pattern matching against comma separated
\code{\link[base:toString]{base::toString()}} output.
}
\examples{
grepString("gene")
}
\seealso{
Other Atomic Vector Functions: \code{\link{sortUnique}},
  \code{\link{toStringUnique}}
}
\author{
Michael Steinbaugh
}
