% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/zzz.R
\name{grepString}
\alias{grepString}
\alias{grep_string}
\title{grep string generator}
\usage{
grepString(identifier)
}
\arguments{
\item{identifier}{Identifier.}
}
\value{
Comma separated grep string.
}
\description{
Generate a grep string for pattern matching against comma separated
\code{\link[base:toString]{base::toString()}} output.
}
\examples{
grepString("gene")
}
