% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepToString.R
\name{grepToString}
\alias{grepToString}
\title{Grep pattern for toString}
\usage{
grepToString(identifier)
}
\arguments{
\item{identifier}{Identifier}
}
\value{
Comma separated string for matching against \code{toString} output
}
\description{
Grep pattern for toString
}
\author{
Michael Steinbaugh
}
\keyword{internal}
