% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grepToString}
\alias{grepToString}
\alias{grep_to_string}
\title{\code{\link[=grep]{grep()}} against \code{\link[=toString]{toString()}} formatted data.}
\usage{
grepToString(identifier)
}
\arguments{
\item{identifier}{Identifier.}
}
\value{
Comma separated string for matching against \code{\link[=toString]{toString()}} return.
}
\description{
Generates a grep pattern that will match an identifier in a comma separated
string.
}
\examples{
grepToString("gene")
}
