% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepToString.R
\name{grepToString}
\alias{grepToString}
\title{Generate a grep for matching against toString}
\usage{
grepToString(identifier)
}
\arguments{
\item{identifier}{Identifier to match in toString}
}
\description{
Generate a grep for matching against toString
}
