% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDavid.R
\name{gseaDavid}
\alias{gseaDavid}
\title{Gene set enrichment analysis (GSEA) with DAVID}
\usage{
gseaDavid(foreground, background, format = "ENSEMBL_GENE_ID",
  saveFiles = FALSE)
}
\arguments{
\item{foreground}{Foreground identifier vector}

\item{background}{Background identifier vector}

\item{format}{Identifier format (see DAVID website)}

\item{saveFiles}{Save files to disk (\code{TRUE} or \code{FALSE})}
}
\value{
List of DAVID chart and cluster
}
\description{
Gene set enrichment analysis (GSEA) with DAVID
}
