% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/headtail-methods.R
\docType{methods}
\name{headtail}
\alias{headtail}
\alias{headtail,atomic-method}
\alias{headtail,matrix-method}
\alias{headtail,sparseMatrix-method}
\alias{headtail,data.frame-method}
\alias{headtail,DataFrame-method}
\alias{headtail,GRanges-method}
\alias{headtail,SummarizedExperiment-method}
\title{Return the first and last parts of an object}
\usage{
headtail(x, ...)

\S4method{headtail}{atomic}(x, n = 2L)

\S4method{headtail}{matrix}(x, n = 2L, ascii = FALSE)

\S4method{headtail}{sparseMatrix}(x, n = 2L, ascii = FALSE)

\S4method{headtail}{data.frame}(x, n = 2L, ascii = FALSE)

\S4method{headtail}{DataFrame}(x, n = 2L, ascii = FALSE)

\S4method{headtail}{GRanges}(x, n = 2L, ascii = FALSE)

\S4method{headtail}{SummarizedExperiment}(x, n = 2L, ascii = FALSE)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{n}{\code{integer(1)}.
Positive integer denoting the number of first and last items to include.}

\item{ascii}{\code{logical(1)}.
Require separators to use ASCII instead of Unicode.}
}
\value{
Invisible \code{NULL}.
}
\description{
Inspired by the \code{print} method for \code{DataFrame} class objects. Applies
to both rows and columns, enabling quick inspection during interactive use.
}
\section{Methods (by class)}{
\itemize{
\item \code{atomic}: Paste collapse to a \code{character(1)}.

\item \code{matrix}: Show first and last rows.

\item \code{sparseMatrix}: Same method as \code{matrix}.

\item \code{data.frame}: Same method as \code{matrix}.

\item \code{DataFrame}: Same method as \code{data.frame}.

\item \code{GRanges}: Summarize the ranges.

\item \code{SummarizedExperiment}: Summarize the primary \code{assay}.
}}

\examples{
data(rse)

## data.frame ====
headtail(datasets::mtcars, ascii = TRUE)
headtail(dplyr::starwars)

## SummarizedExperiment ====
headtail(rse, ascii = TRUE)
}
\seealso{
\itemize{
\item \code{utils::head}, \code{utils::tail}, \code{base::cat}.
\item \code{getMethod("show", "DataTable")}.
}
}
