% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/headtail-methods.R
\docType{methods}
\name{headtail}
\alias{headtail}
\alias{headtail,atomic-method}
\title{Return the First and Last Part of an Object}
\usage{
headtail(x, ...)

\S4method{headtail}{atomic}(x, n = 2L)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{n}{\code{scalar integer}. Positive integer denoting the number of first and
last items to include.}
}
\value{
\code{character}.
}
\description{
Return the First and Last Part of an Object
}
\section{Methods (by class)}{
\itemize{
\item \code{atomic}: Paste collapse to a \code{string}.
}}

\seealso{
\link{head}, \link{tail}.
}
