% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{HGNC2Ensembl-class}
\alias{HGNC2Ensembl-class}
\title{HGNC-to-Ensembl Gene Identifier Mappings}
\value{
\code{HGNC2Ensembl}. Contains a \code{DataFrame} with \code{hgncID} and \code{geneID}
columns.
}
\description{
HGNC-to-Ensembl Gene Identifier Mappings
}
\examples{
options(basejump.test = TRUE)
x <- HGNC2Ensembl()
print(x)
}
\keyword{internal}
