% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{hgnc2ensembl-class}
\alias{hgnc2ensembl-class}
\title{\code{hgnc2ensembl} Class}
\description{
HGNC to Ensembl gene ID mappings.
}
\details{
Contains a \code{DataFrame} with \code{hgncID} and \code{geneID} columns.
}
\examples{
x <- new(
    Class = "hgnc2ensembl",
    DataFrame(
        hgncID = 5L,
        geneID = "ENSG00000121410",
        row.names = 5L
    )
)
}
\seealso{
Other S4 Classes: \code{\link{PANTHER-class}},
  \code{\link{ensembl2entrez-class}},
  \code{\link{gene2symbol-class}},
  \code{\link{mgi2ensembl-class}},
  \code{\link{tx2gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
