% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HGNC2Ensembl-class}
\alias{HGNC2Ensembl-class}
\title{\code{HGNC2Ensembl} Class}
\description{
HGNC-to-Ensembl gene ID mappings.
}
\details{
Contains a \code{DataFrame} with \code{hgncID} and \code{geneID} columns.
}
\seealso{
\code{\link[=hgnc2ensembl]{hgnc2ensembl()}}.

Other S4 Classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
