% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HGNC2Ensembl-class}
\alias{HGNC2Ensembl-class}
\title{HGNC-to-Ensembl gene identifier mappings}
\value{
\code{HGNC2Ensembl}.
}
\description{
HGNC-to-Ensembl gene identifier mappings
}
\details{
Contains a \code{DataFrame} with \code{hgnc} and \code{ensembl} columns.
}
\note{
Updated 2020-10-05.
}
