% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC2Ensembl.R
\name{HGNC2Ensembl}
\alias{HGNC2Ensembl}
\title{HGNC-to-Ensembl Gene Identifier Mappings}
\usage{
HGNC2Ensembl()
}
\value{
\code{HGNC2Ensembl}.
}
\description{
HGNC-to-Ensembl Gene Identifier Mappings
}
\examples{
options(basejump.test = TRUE)
x <- HGNC2Ensembl()
print(x)
}
\seealso{
\code{\link[=HGNC2Ensembl]{HGNC2Ensembl()}}.

Other S4 classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
