% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC2Ensembl.R
\name{HGNC2Ensembl}
\alias{HGNC2Ensembl}
\title{HGNC-to-Ensembl Gene Identifier Mappings}
\usage{
HGNC2Ensembl()
}
\value{
\code{HGNC2Ensembl}.
}
\description{
HGNC-to-Ensembl Gene Identifier Mappings
}
\examples{
options(basejump.test = TRUE)
x <- HGNC2Ensembl()
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{Ensembl2Entrez}},
  \code{\link{Gene2Symbol}}, \code{\link{MGI2Ensembl}},
  \code{\link{Tx2Gene}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{mapGenes}}
}
\concept{Identifier Mapping Functions}
