% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc2ensembl.R
\name{hgnc2ensembl}
\alias{hgnc2ensembl}
\alias{HGNC2Ensembl}
\title{HGNC-to-Ensembl Gene Identifier Mappings}
\usage{
hgnc2ensembl(.test = FALSE)
}
\arguments{
\item{.test}{\code{boolean}. \emph{For unit testing only}.}
}
\value{
\code{HGNC2Ensembl}.
}
\description{
HGNC-to-Ensembl Gene Identifier Mappings
}
\examples{
x <- hgnc2ensembl(.test = TRUE)
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{mapGenes}}, \code{\link{mgi2ensembl}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Identifier Mapping Functions}
