% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc2ensembl.R
\name{hgnc2ensembl}
\alias{hgnc2ensembl}
\title{HGNC to Ensembl Gene ID Mappings.}
\usage{
hgnc2ensembl(url = paste("ftp://ftp.ebi.ac.uk", "pub", "databases",
  "genenames", "new", "tsv", "hgnc_complete_set.txt", sep = "/"))
}
\arguments{
\item{url}{\code{string}. Uniform Resource Locator (URL). HTTP or FTP address.}
}
\value{
\code{hgnc2ensembl}.
}
\description{
HGNC to Ensembl Gene ID Mappings.
}
\examples{
x <- hgnc2ensembl()
print(x)
}
\seealso{
Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{eggnog}}, \code{\link{emptyRanges}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{geneSynonyms}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2ensembl}}, \code{\link{organism}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
