% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc2ensembl.R
\name{hgnc2ensembl}
\alias{hgnc2ensembl}
\title{HGNC to Ensembl Gene ID Mappings.}
\usage{
hgnc2ensembl(.test = FALSE)
}
\arguments{
\item{.test}{\code{boolean}. \emph{For unit testing only}.}
}
\value{
\code{HGNC2Ensembl}.
}
\description{
HGNC to Ensembl Gene ID Mappings.
}
\examples{
x <- hgnc2ensembl(.test = TRUE)
print(x)
}
\seealso{
Other S4 Generators: \code{\link{eggnog}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{mgi2ensembl}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
