% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc2ensembl.R
\name{hgnc2ensembl}
\alias{hgnc2ensembl}
\title{\code{hgnc2ensembl} Generator}
\usage{
hgnc2ensembl()
}
\value{
\code{hgnc2ensembl}.
}
\description{
\code{hgnc2ensembl} Generator
}
\examples{
x <- hgnc2ensembl()
print(x)
}
\seealso{
\code{hgnc2ensembl-class}.

Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{ensembl2entrez}},
  \code{\link{geneSynonyms}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2ensembl}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
