% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC2Ensembl.R, R/deprecated.R
\name{HGNC2Ensembl}
\alias{HGNC2Ensembl}
\alias{hgnc2ensembl}
\title{HGNC-to-Ensembl Gene Identifier Mappings}
\usage{
HGNC2Ensembl(.test = FALSE)
}
\arguments{
\item{.test}{\code{boolean}. \emph{For unit testing only}.}

\item{...}{Additional arguments.}
}
\value{
\code{HGNC2Ensembl}.
}
\description{
HGNC-to-Ensembl Gene Identifier Mappings
}
\examples{
x <- HGNC2Ensembl(.test = TRUE)
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{Ensembl2Entrez}},
  \code{\link{Gene2Symbol}}, \code{\link{MGI2Ensembl}},
  \code{\link{Tx2Gene}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{mapGenes}}
}
\concept{Identifier Mapping Functions}
