% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC2Ensembl.R
\name{HGNC2Ensembl}
\alias{HGNC2Ensembl}
\title{HGNC-to-Ensembl gene identifier mappings}
\usage{
HGNC2Ensembl()
}
\value{
\code{HGNC2Ensembl}.
}
\description{
HGNC-to-Ensembl gene identifier mappings
}
\note{
Updated 2020-10-03.
}
\examples{
object <- HGNC2Ensembl()
print(object)
}
