% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc2gene.R
\name{hgnc2gene}
\alias{hgnc2gene}
\title{HGNC to Ensembl Gene ID Mappings}
\usage{
hgnc2gene()
}
\value{
\code{data.frame}
}
\description{
HGNC to Ensembl Gene ID Mappings
}
\examples{
x <- hgnc2gene()
glimpse(x)
}
\seealso{
Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{makeGRanges}}, \code{\link{makeGene2symbol}},
  \code{\link{makeTx2gene}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
