% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc2gene.R
\name{hgnc2gene}
\alias{hgnc2gene}
\title{HGNC to Ensembl Gene ID Mappings}
\usage{
hgnc2gene()
}
\value{
\code{data.frame}
}
\description{
HGNC to Ensembl Gene ID Mappings
}
\examples{
x <- hgnc2gene()
glimpse(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Gene Annotation Functions}
