% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/humanize-methods.R
\docType{methods}
\name{humanize}
\alias{humanize}
\alias{humanize,SummarizedExperiment-method}
\title{Humanize an R object}
\usage{
\S4method{humanize}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains human-friendly rownames (e.g. gene symbols instead of stable gene
IDs) and colnames (e.g. sample names instead of sample IDs).
}
\description{
Humanize an R object
}
\details{
This can make dimnames invalid (see \code{\link[base:make.names]{make.names()}}) and
should only be called prior to writing files to disk.
}
\examples{
data(rse, sce, package = "acidtest")

## SummarizedExperiment ====
lapply(dimnames(rse), head)
x <- humanize(rse)
lapply(dimnames(x), head)


## SingleCellExperiment ====
lapply(dimnames(sce), head)
x <- humanize(sce)
lapply(dimnames(x), head)
}
