% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanize-methods.R
\docType{methods}
\name{humanize}
\alias{humanize}
\alias{humanize,SummarizedExperiment-method}
\title{Humanize an R object}
\usage{
humanize(object, ...)

\S4method{humanize}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains human-friendly rownames (e.g. gene symbols instead of stable gene
IDs) and colnames (e.g. sample names instead of sample IDs).
}
\description{
Humanize an R object
}
\details{
This can make dimnames invalid (see \code{\link[base:make.names]{make.names()}}) and
should only be called prior to writing files to disk.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)
rse <- RangedSummarizedExperiment
sce <- SingleCellExperiment

## SummarizedExperiment ====
lapply(dimnames(rse), head)
x <- humanize(rse)
lapply(dimnames(x), head)


## SingleCellExperiment ====
lapply(dimnames(sce), head)
x <- humanize(sce)
lapply(dimnames(x), head)
}
