% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanize-methods.R
\name{humanize}
\alias{humanize}
\alias{humanize,SummarizedExperiment-method}
\title{Humanize an R object}
\usage{
humanize(object, ...)

\S4method{humanize}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains human-friendly rownames (e.g. gene symbols instead of stable gene
IDs) and colnames (e.g. sample names instead of sample IDs).
}
\description{
Humanize an R object
}
\details{
This can make dimnames invalid (see \code{\link[base:make.names]{make.names()}}) and
should only be called prior to writing files to disk.
}
\note{
Updated 2020-01-20.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
lapply(dimnames(object), head)
x <- humanize(object)
lapply(dimnames(x), head)

## SingleCellExperiment ====
object <- SingleCellExperiment
lapply(dimnames(object), head)
x <- humanize(object)
lapply(dimnames(x), head)
}
