% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import}
\usage{
import(file, ..., dataFrame = getOption("basejump.data.frame",
  "data.frame"))
}
\arguments{
\item{file}{\code{string}. File path.}

\item{...}{Additional arguments.}

\item{dataFrame}{\code{string}. Data frame class to return. Can be set globally
using the \code{basejump.data.frame} option.

Current recommendations (by priority):
\itemize{
\item \code{data.frame}: Base R default. Generally recommended.
\itemize{
\item S3 class.
\item Allows rownames, but they're required and can't be set \code{NULL}.
\item See \code{help(topic = "data.frame", package = "base")} for details.
}
\item \code{DataFrame}: Recommended when working with Bioconductor packages.
\itemize{
\item S4 class; inherits \code{DataTable} virtual class.
\item Allows rownames, but they're optional and can be set \code{NULL}.
\item See \code{help(topic = "DataFrame", package = "S4Vectors")} for details.
}
\item \code{tibble}: Recommended when working with tidyverse packages.
\itemize{
\item S3 class (\code{tbl_df}); inherits \code{data.frame}.
\item Does not allow rownames.
\item See \code{help(topic = "tibble", package = "tibble")} for details.
}
\item \code{data.table}: Recommended when working with the data.table package.
\itemize{
\item S3 class; inherits \code{data.frame}.
\item Does not allow rownames.
\item See \code{help(topic = "data.table", package = "data.table")} for details.
}
}}
}
\value{
Varies, depending on the file extension:
\itemize{
\item Default: \code{DataFrame}.
\item \code{MTX}: \code{sparseMatrix}.
\item \code{GTF}/\code{GFF}: \code{GRanges}.
\item \code{JSON}/\code{YAML}: \code{list}.
\item Source code or log: \code{character}.
}
}
\description{
Read file by extension into R.
}
\details{
Remote URLs and compressed files are supported. All extensions are case
insensitive.

This function supports automatic loading of common file types:
\itemize{
\item \code{CSV}: Comma Separated Values.\cr
Imported by \code{\link[data.table:fread]{data.table::fread()}}.
\item \code{TSV} Tab Separated Values.\cr
Imported by \code{\link[data.table:fread]{data.table::fread()}}.
\item \code{TXT}: Text file. \emph{Ambiguous, not recommended}.\cr
Imported by \code{\link[data.table:fread]{data.table::fread()}}.
\item \code{XLSX}/\code{XLS}: Excel workbook.\cr
Imported by \code{\link[readxl:read_excel]{readxl::read_excel()}}.
\item \code{MTX}: MatrixMarket sparse matrix.\cr
Imported by \code{\link[Matrix:readMM]{Matrix::readMM()}}.
\item \code{GTF}/\code{GFF}/\code{GFF3}: General Feature Format.\cr
Imported by \code{\link[rtracklayer:import]{rtracklayer::import()}}.
\item \code{JSON}: JSON.
Imported by \code{\link[jsonlite:read_json]{jsonlite::read_json()}}.
\item \code{YAML}/\code{YML}: YAML.
Imported by \code{\link[yaml:yaml.load_file]{yaml::yaml.load_file()}}
\item \code{RDA}/\code{RDATA}: R Data.
\itemize{
\item Imported by \code{\link[base:load]{base::load()}}.
\item Must contain a single object.
\item Doesn't require internal object name to match, unlike \code{\link[=loadData]{loadData()}}.
}
\item \code{RDS}: R Data Serialized.\cr
Imported by \code{\link[base:readRDS]{base::readRDS()}}.
}

These file formats will be imported as source code lines by
\code{\link[readr:read_lines]{readr::read_lines()}}: \code{LOG}, \code{MD}, \code{PY}, \code{R}, \code{RMD}, \code{SH}.

These file formats are blacklisted, and intentionally not supported:
\code{DOC}, \code{DOCX}, \code{PDF}, \code{PPT}, \code{PPTX}.

If a file format isn't supported natively (or blacklisted), the
\href{https://cran.r-project.org/web/packages/rio/index.html}{rio} package will
be used as a fallback attempt. See \code{\link[rio:import]{rio::import()}} for details.
}
\section{Delimited Files (CSV/TSV)}{


\code{\link[=import]{import()}} uses the \code{fread()} function of the  \href{https://cran.r-project.org/package=data.table}{data.table} package to
import standard CSV and TSV files. This should work automatically for most
files without issue.

Here are some notable exceptions:
\itemize{
\item Columns headers should be \code{character}. In the event that a column name
is \code{numeric}, set \code{header = TRUE} here to force the column name.
}

See \code{help(topic = "fread", package = "data.table")} for details.

The \code{read_csv()} and \code{read_tsv()} functions of the \code{\link[=readr]{readr()}} package
are good alternatives, which return \code{tibble} data frames.
}

\section{Matrix Market Exchange (MTX/MEX)}{


Reading a Matrix Market Exchange (\code{MTX}) file now requires \code{COLNAMES} and
\code{ROWNAMES} sidecar files containing the \code{\link[=colnames]{colnames()}} and \code{\link[=rownames]{rownames()}} of
the sparse matrix. Legacy support for manual loading of these sidecar files
is provided.
}

\section{General Feature Format (GFF/GTF)}{


The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.

Column names follow the \href{https://bit.ly/2K6EBla}{Ensembl conventions}.
Additional information:
\itemize{
\item \href{http://www.ensembl.org/info/website/upload/gff.html}{Ensembl}
\item \href{http://www.gencodegenes.org/gencodeformat.html}{Gencode}
}
}

\section{bcbio files}{


Also supports some additional extensions commonly used with the
\href{https://bcbio-nextgen.readthedocs.io}{bcbio} pipeline:
\itemize{
\item \code{COUNTS}: Counts table (e.g. RNA-seq aligned counts).
\item \code{COLNAMES}: Sidecar file containing column names.
\item \code{ROWNAMES}: Sidecar file containing row names.
}
}

\examples{
url <- basejumpCacheURL

## R Data
x <- import(file = file.path(url, "rnaseq_counts.rda"))
class(x)

## Comma Separated Values
x <- import(file = file.path(url, "example.csv"))
print(x)

## Tab Separated Values
x <- import(file = file.path(url, "example.tsv"))

## Microsoft Excel Worksheet
x <- import(file = file.path(url, "example.xlsx"))
print(x)

## GTF/GFF
x <- import(file = file.path(url, "example.gtf"))
summary(x)

## JSON
x <- import(file = file.path(url, "example.json"))
names(x)

## YAML
x <- import(file = file.path(url, "example.yml"))
names(x)

## Counts Table (i.e. aligned counts from bcbio)
x <- import(file = file.path(url, "example.counts"))
colSums(x)
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \href{https://cran.r-project.org/web/packages/Matrix/index.html}{Matrix}.
}
}
