% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import}
\usage{
import(file, ...)
}
\arguments{
\item{file}{\code{string}. File path.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending on the file extension:
\itemize{
\item Default: \code{DataFrame}.
\item \code{MTX}: \code{sparseMatrix}.
\item \code{GTF}/\code{GFF}: \code{GRanges}.
\item \code{JSON}/\code{YAML}: \code{list}.
\item Source code or log: \code{character}.
}

Note that data frames are returned as S4 \code{DataFrame} class instead of
\code{data.frame}, when applicable.
}
\description{
Read file by extension into R.
}
\details{
This is a wrapper for \code{\link[rio:import]{rio::import()}} that adds support for additional common
genomic data files. Remote URLs and compressed files are supported. All
extensions are case insensitive.

This function supports automatic loading of common file types:
\itemize{
\item \code{CSV}: Comma Separated Values.
\item \code{TSV} Tab Separated Values.
\item \code{XLSX}: Excel workbook.
\item \code{MTV}: MatrixMarket sparse matrix.
\item \code{GTF}/\code{GFF}/\code{GFF3}: General Feature Format.
\item \code{RDA}/\code{RDATA}: R Data.
\itemize{
\item Must contain a single object.
\item Doesn't require internal object name to match, unlike \code{\link[=loadData]{loadData()}}.
}
\item \code{RDS}: R Data Serialized.
\item \code{JSON}: JSON.
\item \code{YAML}/\code{YML}: YAML.
}

These file formats will be read as (source code) lines:
\code{LOG}, \code{MD}, \code{PY}, \code{R}, \code{RMD}, \code{SH}.

These file formats are blacklisted, and intentionally not supported:
\code{DOC}, \code{DOCX}, \code{PDF}, \code{PPT}, \code{PPTX}, \code{TXT}.

If the file format isn't supported natively (or blacklisted), the
\href{https://cran.r-project.org/web/packages/rio/index.html}{rio} package will
be used as a fallback attempt.
}
\section{Matrix Market Exchange (MTX/MEX)}{


Reading a Matrix Market Exchange (\code{MTX}) file now requires \code{COLNAMES} and
\code{ROWNAMES} sidecar files containing the \code{\link[=colnames]{colnames()}} and \code{\link[=rownames]{rownames()}} of
the sparse matrix. Legacy support for manual loading of these sidecar files
is provided.
}

\section{General Feature Format (GFF/GTF)}{


The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.

Column names follow the \href{https://bit.ly/2K6EBla}{Ensembl conventions}.
Additional information:
\itemize{
\item \href{http://www.ensembl.org/info/website/upload/gff.html}{Ensembl}
\item \href{http://www.gencodegenes.org/gencodeformat.html}{Gencode}
}
}

\section{bcbio}{


Also supports some additional extensions commonly used with the
\href{https://bcbio-nextgen.readthedocs.io}{bcbio} pipeline:
\itemize{
\item \code{COUNTS}: Counts table (e.g. RNA-seq aligned counts).
\item \code{COLNAMES}: Sidecar file containing column names.
\item \code{ROWNAMES}: Sidecar file containing row names.
}
}

\examples{
## R Data
x <- import(file = file.path(basejumpCacheURL, "rnaseq_counts.rda"))
class(x)

## Comma Separated Values
x <- import(file = file.path(basejumpCacheURL, "example.csv"))
print(x)

## Microsoft Excel Worksheet
x <- import(file = file.path(basejumpCacheURL, "example.xlsx"))
print(x)

## GTF/GFF
x <- import(file = file.path(basejumpCacheURL, "example.gtf"))
summary(x)

## JSON
x <- import(file = file.path(basejumpCacheURL, "example.json"))
names(x)

## YAML
x <- import(file = file.path(basejumpCacheURL, "example.yml"))
names(x)

## Counts Table (i.e. aligned counts from bcbio)
x <- import(file = file.path(basejumpCacheURL, "example.counts"))
colSums(x)
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \href{https://cran.r-project.org/web/packages/Matrix/index.html}{Matrix}.
}

Other Import/Export Functions: \code{\link{assignAndSaveData}},
  \code{\link{export}}, \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{saveData}}, \code{\link{transmit}},
  \code{\link{writeCounts}}
}
\concept{Import/Export Functions}
