% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{importTidyVerbs}
\alias{importTidyVerbs}
\title{Import \href{https://www.bioconductor.org/}{Bioconductor}-masked generic
\href{http://tidyverse.org/}{tidyverse} verbs to environment.}
\usage{
importTidyVerbs(envir = parent.frame())
}
\arguments{
\item{envir}{Environment where to assign the verbs.}
}
\description{
Avoid NAMESPACE collisions with \href{https://www.bioconductor.org/}{Bioconductor}
BiocGenerics and S4Vectors.
}
\seealso{
\itemize{
\item \href{http://bioconductor.org/packages/release/bioc/html/BiocGenerics.html}{BiocGenerics}.
\item \href{http://bioconductor.org/packages/release/bioc/html/S4Vectors.html}{S4Vectors}.
\item \href{http://style.tidyverse.org/}{tidyverse style guide}.
\item \href{http://dplyr.tidyverse.org/}{dplyr}.
\item \href{http://tidyr.tidyverse.org/}{tidyr}.
}
}
\keyword{internal}
