% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTx2Gene.R
\name{importTx2Gene}
\alias{importTx2Gene}
\title{Import transcript-to-gene annotations}
\usage{
importTx2Gene(
  file,
  organism = NULL,
  genomeBuild = NULL,
  ensemblRelease = NULL,
  ignoreTxVersion = FALSE,
  ignoreGeneVersion = TRUE
)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{ensemblRelease}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{ignoreTxVersion}{\code{logical(1)}.
Don't include the transcript version in the identifier.
\strong{Not recommended} by default when handing off salmon or kallisto output
to tximport-DESeq2 workflow.}

\item{ignoreGeneVersion}{\code{logical(1)}.
Don't include the gene version in the identifier.
\strong{Recommended} by default when handing off salmon or kallisto output to
tximport-DESeq2 workflow.}
}
\value{
\code{Tx2Gene}.
}
\description{
Generates a \code{Tx2Gene} object containing \code{transcriptID} and \code{geneID} columns.
}
\note{
File should not contain column header names.

Updated 2020-05-10.
}
\examples{
file <- file.path(basejumpTestsURL, "tx2gene.csv")
x <- importTx2Gene(
    file = file,
    organism = "Homo sapiens",
    genomeBuild = "GRCh38",
    ensemblRelease = 90L
)
print(x)
}
\seealso{
\itemize{
\item \code{stripTranscriptVersions}, \code{stripGeneVersions}
\item \code{download-ensembl-genome} script (which generates \code{tx2gene.csv} output)
defined in koopa shell bootloader.
}
}
