% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/initializeDirectory.R
\name{initializeDir}
\alias{initializeDir}
\alias{initializeDirectory}
\title{Initialize Directory}
\usage{
initializeDir(...)

initializeDirectory(dir)
}
\arguments{
\item{...}{Additional arguments.}

\item{dir}{Directory path.}
}
\value{
Directory path string.
}
\description{
Checks and creates a directory recursively automatically. Useful for setting
up functions that require writes to a user-specified directory.
}
\examples{
initializeDirectory("testdir")
dir.exists("testdir")
unlink("testdir")
}
