% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeDirectory.R
\name{initializeDirectory}
\alias{initializeDirectory}
\title{Initialize Directory}
\usage{
initializeDirectory(dir)
}
\arguments{
\item{dir}{Directory path.}
}
\value{
Directory path string.
}
\description{
Checks and creates a directory recursively automatically. Useful for setting
up functions that require writes to a user-specified directory.
}
\examples{
initializeDirectory("testdir")
dir.exists("testdir")

# Clean up
unlink("testdir", recursive = TRUE)
}
\seealso{
Other Write Functions: \code{\link{assignAndSaveData}},
  \code{\link{saveData}}, \code{\link{transmit}},
  \code{\link{writeCounts}}
}
