% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeDirectory.R
\name{initializeDirectory}
\alias{initializeDirectory}
\title{Initialize Directory}
\usage{
initializeDirectory(dir)
}
\arguments{
\item{dir}{\code{string}. Directory path.}
}
\value{
\code{string}. Directory path.
}
\description{
Checks and creates a directory recursively automatically. Useful for setting
up functions that require writes to a user-specified directory.
}
\examples{
initializeDirectory("testdir")
dir.exists("testdir")

# Clean up
unlink("testdir", recursive = TRUE)
}
\seealso{
Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}}, \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}, \code{\link{standardizeCall}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
