% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeDirectory.R
\name{initializeDirectory}
\alias{initializeDirectory}
\alias{initializeDir}
\title{Initialize Directory}
\usage{
initializeDirectory(dir)

initializeDir(...)
}
\arguments{
\item{dir}{Directory path.}

\item{...}{Additional arguments.}
}
\value{
Directory path string.
}
\description{
Checks and creates a directory recursively automatically. Useful for setting
up functions that require writes to a user-specified directory.
}
\examples{
initializeDirectory("testdir")
dir.exists("testdir")
unlink("testdir")
}
