% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeDirectory.R
\name{initializeDirectory}
\alias{initializeDirectory}
\alias{initDir}
\title{Initialize Directory}
\usage{
initializeDirectory(dir)

initDir(dir)
}
\arguments{
\item{dir}{\code{string}. Directory path.}
}
\value{
\code{string}. Directory path.
}
\description{
Checks and creates a directory recursively automatically. Useful for setting
up functions that require writes to a user-specified directory.
}
\examples{
initializeDirectory("testdir")
dir.exists("testdir")

## Clean up.
unlink("testdir", recursive = TRUE)
}
