% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/initializeDirectory.R
\name{initializeDir}
\alias{initializeDir}
\alias{initializeDirectory}
\title{Initialize Directory}
\usage{
initializeDir(...)

initializeDirectory(dir)
}
\arguments{
\item{dir}{Directory path.}
}
\description{
Initialize Directory
}
\examples{
initializeDirectory(getwd())
}
