% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestingGroups-methods.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,SummarizedExperiment-method}
\alias{interestingGroups<-,SummarizedExperiment,character-method}
\alias{interestingGroups<-,SummarizedExperiment,NULL-method}
\title{Interesting groups}
\usage{
\S4method{interestingGroups}{SummarizedExperiment}(object, check = FALSE)


  \S4method{interestingGroups}{SummarizedExperiment,character}(object) <- value


  \S4method{interestingGroups}{SummarizedExperiment,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{check}{\code{logical(1)}.
Perform assert checks.}

\item{value}{Value.}
}
\value{
\code{character}.
}
\description{
Interesting groups
}
\examples{
data(rse)
object <- rse

intgroup <- interestingGroups(object)
print(intgroup)

## Assignment support.
interestingGroups(object) <- intgroup[[1L]]
interestingGroups(object)
}
