% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestingGroups-methods.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{interestingGroups,Annotated-method}
\alias{interestingGroups<-,Annotated,character-method}
\alias{interestingGroups<-,SummarizedExperiment,character-method}
\alias{interestingGroups<-,Annotated,NULL-method}
\title{Interesting groups}
\usage{
interestingGroups(object, ...)

interestingGroups(object, ...)  <- value

\S4method{interestingGroups}{Annotated}(object)

\S4method{interestingGroups}{Annotated,character}(object) <- value


  \S4method{interestingGroups}{SummarizedExperiment,character}(object) <- value

\S4method{interestingGroups}{Annotated,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Interesting groups
}
\examples{
data(rse, package = "acidtest")
object <- rse

intgroup <- interestingGroups(object)
print(intgroup)

## Assignment support.
interestingGroups(object) <- intgroup[[1L]]
interestingGroups(object)
}
