% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interestingGroups-methods.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{interestingGroups,SummarizedExperiment-method}
\alias{interestingGroups<-,SummarizedExperiment,character-method}
\alias{interestingGroups<-,SummarizedExperiment,NULL-method}
\title{Interesting Groups}
\usage{
interestingGroups(object, ...)

interestingGroups(object, ...) <- value

\S4method{interestingGroups}{SummarizedExperiment}(object, check = TRUE)


  \S4method{interestingGroups}{SummarizedExperiment,character}(object) <- value


  \S4method{interestingGroups}{SummarizedExperiment,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}

\item{check}{\code{boolean}. Perform assert checks.}
}
\value{
\code{character}.
}
\description{
Interesting Groups
}
\examples{
data(rse_small)
object <- rse_small

intgroup <- interestingGroups(object)
print(intgroup)

# Assignment support.
interestingGroups(object) <- intgroup[[1L]]
interestingGroups(object)
}
\seealso{
Other Metadata Functions: \code{\link{detectHPC}},
  \code{\link{detectLanes}}, \code{\link{detectOrganism}},
  \code{\link{microplate}},
  \code{\link{minimalSampleData}}, \code{\link{organism}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{uniteInterestingGroups}}

Other SummarizedExperiment Functions: \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{sampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}},
  \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
\concept{Metadata Functions}
\concept{SummarizedExperiment Functions}
