% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interestingGroups-methods.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{interestingGroups,SummarizedExperiment-method}
\alias{interestingGroups<-,SummarizedExperiment,character-method}
\alias{interestingGroups<-,SummarizedExperiment,NULL-method}
\title{Interesting Groups}
\usage{
interestingGroups(object, ...)

interestingGroups(object, ...) <- value

\S4method{interestingGroups}{SummarizedExperiment}(object)


  \S4method{interestingGroups}{SummarizedExperiment,character}(object) <- value


  \S4method{interestingGroups}{SummarizedExperiment,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{character}.
}
\description{
Interesting Groups
}
\examples{
# SummarizedExperiment ====
object <- rse_small
intgroup <- interestingGroups(object)
print(intgroup)

# Assignment support.
interestingGroups(object) <- intgroup[[1L]]
interestingGroups(object)
}
\author{
Michael Steinbaugh
}
