% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interestingGroups-methods.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{interestingGroups}
\alias{interestingGroups,SummarizedExperiment-method}
\alias{interestingGroups<-,SummarizedExperiment,character-method}
\title{Interesting Groups}
\usage{
interestingGroups(object, ...)

interestingGroups(object, ...) <- value

\S4method{interestingGroups}{SummarizedExperiment}(object)

\S4method{interestingGroups}{SummarizedExperiment,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{character}.
}
\description{
Interesting Groups
}
\examples{
# SummarizedExperiment ====
interestingGroups(rse_dds)
colnames(colData(rse_dds))
interestingGroups(rse_dds) <- colnames(colData(rse_dds))[[1L]]
interestingGroups(rse_dds)
}
\author{
Michael Steinbaugh
}
