% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interestingGroups-methods.R
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{interestingGroups,Annotated-method}
\alias{interestingGroups<-,Annotated,character-method}
\alias{interestingGroups<-,SummarizedExperiment,character-method}
\alias{interestingGroups<-,Annotated,NULL-method}
\title{Interesting groups}
\usage{
interestingGroups(object, ...)

interestingGroups(object, ...)  <- value

\S4method{interestingGroups}{Annotated}(object)

\S4method{interestingGroups}{Annotated,character}(object) <- value

\S4method{interestingGroups}{SummarizedExperiment,character}(object) <- value

\S4method{interestingGroups}{Annotated,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Interesting groups
}
\note{
Updated 2019-08-11.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
intgroup <- interestingGroups(rse)
print(intgroup)
## Assignment support.
interestingGroups(rse) <- intgroup[[1L]]
interestingGroups(rse)
}
