% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kable}
\alias{kable}
\title{kable rework with custom defaults}
\usage{
kable(x, ...)
}
\arguments{
\item{x}{An R object (typically a matrix or data frame)}

\item{...}{Pass through to \code{knitr::kable()}}
}
\description{
Display as data frame in RMarkdown chunk or \code{kable()} during knit
}
\examples{
kable(head(iris))
}
\author{
Michael Steinbaugh
}
