% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kable}
\alias{kable}
\title{Create tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
kable(x, ...)
}
\arguments{
\item{x}{An R object (typically a matrix or data frame)}

\item{...}{Other arguments supported by \code{knitr::kable}}
}
\value{
Kable during knit, otherwise the original data
}
\description{
Wrapper for \code{knitr::kable}. Improves appearance in RStudio RMarkdown
chunks by returning a data frame instead of a kable outside of a knit.
Customizes knit output defaults based on format matches. The function applies
grep matching to both default and custom document templates by format (e.g.
\code{BiocStyle::pdf_document2}).
}
\examples{
kable(head(iris))
}
\author{
Michael Steinbaugh
}
