% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kable}
\alias{kable}
\title{Create tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
kable(x, digits = 3, ...)
}
\arguments{
\item{x}{An R object (typically a matrix or data frame)}

\item{digits}{Maximum number of digits for numeric columns}

\item{...}{Other arguments supported by \code{knitr::kable}}
}
\value{
Kable during knit, otherwise the original data
}
\description{
Wrapper for \code{knitr::kable}. Improves appearance in RStudio RMarkdown
chunks by returning a data frame instead of a kable outside of a knit.
Customizes knit output defaults based on format.
}
\examples{
kable(head(iris))
}
\author{
Michael Steinbaugh
}
