% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kables.R
\name{kables}
\alias{kables}
\title{Handle multiple kables in a single RMarkdown chunk}
\usage{
kables(list, captions = NULL)
}
\arguments{
\item{list}{List of tables (e.g. data frame, matrix)}

\item{captions}{Caption character vector}
}
\value{
Knit tables, using \code{kable}
}
\description{
Utility function for improving knit reports
}
\examples{
kables(list(iris, mtcars))
}
\author{
Michael Steinbaugh
}
