% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-kables.R
\docType{methods}
\name{kables}
\alias{kables}
\alias{kables}
\alias{kables,list-method}
\title{Create Tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
kables(object, ...)

\S4method{kables}{list}(object, captions = NULL, force = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{captions}{Optional character vector of table captions.}

\item{force}{Force kable output.}

\item{list}{List of column data (e.g. \link{data.frame}, \link{matrix}).}
}
\value{
Knit tables, using \code{\link[knitr:kable]{knitr::kable()}}.
}
\description{
Handle multiple kables in a single RMarkdown chunk.
}
\examples{
list(head(starwars), head(mtcars)) \%>\% kables()
}
\seealso{
\href{https://stackoverflow.com/a/35149103/3911732}{Stack Overflow}.

Other Report Utilities: \code{\link{markdownHeader}},
  \code{\link{markdownList}}
}
