% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kables.R
\name{kables}
\alias{kables}
\title{Handle multiple kables in a single RMarkdown chunk}
\usage{
kables(list)
}
\arguments{
\item{list}{List of tables (e.g. data frame, matrix)}
}
\value{
Knit tables, using \code{kable}
}
\description{
Utility function for improving knit reports
}
\examples{
kables(list(example1 = iris, example2 = mtcars))
}
\author{
Michael Steinbaugh
}
