% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/kables-methods.R
\docType{methods}
\name{kables}
\alias{kables}
\alias{kables,list-method}
\title{Create Tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
kables(object, ...)

\S4method{kables}{list}(object, captions = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{captions}{Optional character vector of table captions.}

\item{list}{List of column data (e.g. \link{data.frame}, \link{matrix}).}
}
\value{
Knit tables, using \code{\link[knitr:kable]{knitr::kable()}}.
}
\description{
Handle multiple kables in a single RMarkdown chunk.
}
\examples{
list(starwars, head(mtcars)) \%>\% kables
}
\seealso{
Other Markdown Utilities: \code{\link{mdHeader}},
  \code{\link{mdList}}
}
