% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kables.R
\name{kables}
\alias{kables}
\title{Create Tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
kables(list, captions, force = FALSE)
}
\arguments{
\item{list}{List of column data (e.g. \code{data.frame}, \code{matrix}).}

\item{captions}{Optional character vector of table captions.}

\item{force}{Force kable output. Recommended for development and unit testing
only.}
}
\value{
Knit tables, using \code{\link[knitr:kable]{knitr::kable()}}.
}
\description{
Handle multiple kables in a single R Markdown chunk.
}
\examples{
kables(
    list = list(head(starwars), head(mtcars)),
    captions = c("starwars", "mtcars")
)
}
\seealso{
\href{https://stackoverflow.com/a/35149103/3911732}{Stack Overflow post}.

Other R Markdown Functions: \code{\link{markdownHeader}},
  \code{\link{markdownList}},
  \code{\link{markdownPlotlist}}
}
\author{
Michael Steinbaugh
}
