% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kables.R
\name{kables}
\alias{kables}
\title{Create Multiple Tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
kables(list, captions, force = FALSE)
}
\arguments{
\item{list}{\code{list} containing column data (e.g. \code{data.frame}, \code{matrix}).}

\item{captions}{\code{character}. Table captions.}

\item{force}{\code{boolean}. Force knit output. Recommended for development and
unit testing only.}
}
\value{
Knit tables.
}
\description{
Knit multiple tables in a single R Markdown chunk.
}
\examples{
kables(
    list = list(head(ggplot2::mpg), head(datasets::mtcars)),
    captions = c("mpg", "mtcars")
)
}
\seealso{
\itemize{
\item \code{\link[knitr:kable]{knitr::kable()}}.
\item \href{https://stackoverflow.com/a/35149103/3911732}{Stack Overflow post}.
}

Other R Markdown Functions: \code{\link{markdownHeader}},
  \code{\link{markdownLink}}, \code{\link{markdownList}},
  \code{\link{markdownPlotlist}}
}
\author{
Michael Steinbaugh
}
\concept{R Markdown Functions}
