% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{lanePattern}
\alias{lanePattern}
\title{Sequencing Lane Grep Pattern}
\format{An object of class \code{character} of length 1.}
\usage{
lanePattern
}
\description{
Sequencing Lane Grep Pattern
}
\examples{
lanePattern
}
\keyword{datasets}
