% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDataRaw}
\alias{loadDataRaw}
\alias{loadRemote}
\title{Data loading utilities}
\usage{
loadDataRaw(data)

loadRemote(url)
}
\arguments{
\item{data}{Data object.}

\item{url}{URL.}
}
\description{
Dynamically \code{\link[=source]{source()}} or \code{\link[=load]{load()}} raw data in the \code{data-raw}
directory.

Load a remote R binary file.
}
\examples{
\dontrun{
loadDataRaw("counts")
}
\dontrun{
loadRemote("http://example.com/data.rda")
}
}
