% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-loadRemote.R
\docType{methods}
\name{loadRemote}
\alias{loadRemote}
\alias{loadRemote,character-method}
\title{Load Remote Data}
\usage{
loadRemote(object)

\S4method{loadRemote}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Load a remote R binary file.
}
\examples{
\dontrun{
loadRemoteData("http://example.com/data.rda")
}
}
