% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R, R/zzz.R
\name{loadData}
\alias{loadData}
\alias{loadRemote}
\alias{load_data}
\alias{load_remote}
\title{Data Loading Utilities}
\usage{
loadData(...)

loadRemote(url)
}
\arguments{
\item{...}{Data files as dot objects.}

\item{url}{URL.}
}
\description{
Dynamically \code{\link[=load]{load()}} data from \code{data} directory or \code{\link[=source]{source()}}
the corresponding script frmo the \code{data-raw} directory.

Load a remote R binary file.
}
\examples{
\dontrun{
loadData(geneIDs, oligo)
loadRemote("http://example.com/data.rda")
}

}
