% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Local Data}
\usage{
loadData(..., dir = "data")
}
\arguments{
\item{...}{Object names as symbols.}

\item{dir}{Output directory. Defaults to \strong{data}.}
}
\value{
No value.
}
\description{
Load RData (\code{.rda}) files from a directory using symbols rather than complete
file paths.
}
\examples{
\dontrun{
loadData(geneIDs, oligo)
}
}
