% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Local Data}
\usage{
loadData(..., dir = "data")
}
\arguments{
\item{dir}{Directory where the RData (\code{.rda}) files are located.}
}
\value{
No value.
}
\description{
Load RData (\code{.rda}) files from a directory using symbols rather than complete
file paths.
}
\examples{
\dontrun{
loadData(geneIDs, oligo)
}
}
