% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Local Data}
\usage{
loadData(..., dir = "data", envir = parent.frame())
}
\arguments{
\item{dir}{Directory where the RData (\code{.rda}) files are located.}

\item{envir}{Environment where the objects will be loaded. Defaults to
\code{parent.frame()}, also referred to as the calling environment. This
argument supports new environment creation when a character string is used.}
}
\value{
No value.
}
\description{
Dynamically \code{\link[=load]{load()}} data from \code{data/} directory or \code{\link[=source]{source()}} the
corresponding script from the \code{data-raw/} directory.
}
\examples{
\dontrun{
loadData(geneIDs, oligo)
loadData(geneIDs, envir = "newenv")
}
}
