% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Local Data}
\usage{
loadData(...)
}
\arguments{
\item{...}{Data files as symbols (dot objects).}
}
\value{
No value.
}
\description{
Dynamically \code{\link[=load]{load()}} data from \code{data} directory or \code{\link[=source]{source()}} the
corresponding script frmo the \code{data-raw} directory.
}
\examples{
\dontrun{
loadData(geneIDs, oligo)
}
}
