% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Local Data}
\usage{
loadData(..., dir = "data", envir = parent.frame(), quiet = FALSE)
}
\arguments{
\item{...}{Object names as symbols.}

\item{dir}{Output directory. Defaults to \strong{data}.}

\item{envir}{Environment to use for assignment. Defaults to \code{parent.frame()},
which will assign into the calling environment.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Silent named character vector of file paths.
}
\description{
Load RData (\code{.rda}) files from a directory using symbols rather than complete
file paths.
}
\examples{
\dontrun{
loadData(geneIDs, oligo)
}
}
