% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Data}
\usage{
loadData(..., dir = ".", envir = parent.frame())
}
\arguments{
\item{...}{Object names. Note that these arguments are interpreted as symbols
using non-standard evaluation for convenience during interactive use, and
\emph{should not be quoted}.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}

\item{envir}{\code{environment} to use for assignment. Defaults to
\code{parent.frame()}, which will assign into the calling environment.}
}
\value{
Invisible \code{character} containing file paths.
}
\description{
Load R data files from a directory using symbols rather than complete file
paths. Supports "\code{.RData}", "\code{.rda}", and "\code{.rds}" file extensions.
}
\details{
\code{\link[=loadData]{loadData()}} is opinionated about the format of R data files it will accept.
\code{\link[base:save]{base::save()}} allows for the saving of multiple objects into a single R data
file. This can later result in unexpected accidental replacement of an
existing object in the current environment. Additionally, since an R data
file internally stores the name of an object, if the file is later renamed
the object name will no longer match.

To avoid any accidental replacements, \code{\link[=loadData]{loadData()}} will only load R data
files that contain a single object, and the internal object name must match
the file name exactly.
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation.
}
\examples{
loadData(
    example,
    dir = system.file("extdata", package = "basejump")
)
}
\seealso{
Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{readJSON}}, \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
