% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Local Data}
\usage{
loadData(..., dir = getwd(), ext = "rda", envir = parent.frame(),
  replace = TRUE, quiet = FALSE)
}
\arguments{
\item{...}{Object names as symbols.}

\item{dir}{Output directory. Defaults to the current working directory.}

\item{ext}{R data file extension. Defaults to \code{rda} and typically should not
be changed.}

\item{envir}{Environment to use for assignment. Defaults to \code{parent.frame()},
which will assign into the calling environment.}

\item{replace}{Replace existing object in destination environment.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Silent named character vector of file paths.
}
\description{
Load RData (\code{.rda}) files from a directory using symbols rather than complete
file paths.
}
\details{
\code{\link[=loadData]{loadData()}} is opinionated about the format of R data files it will accept.
\code{\link[base:save]{base::save()}} allows for the saving of multiple objects into a single R data
file. This can later result in unexpected accidental replacement of an
existing object in the current environment. Additionally, since an R data
file internally stores the name of an object, if the file is later renamed
the object name will no longer match.

To avoid any accidental replacements, \code{\link[=loadData]{loadData()}} will only load R data
files that contain a single object, and the internal object name must match
the file name exactly. These conventions match the recommendations of the
RStudio team, which recommends saving single objects per file.
}
\examples{
# Load the internal GRCh37 gene annotations data
loadData(grch37, dir = system.file("extdata", package = "basejump"))
glimpse(grch37)
}
