% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-loadData.R
\docType{methods}
\name{loadData}
\alias{loadData}
\alias{loadData,ANY-method}
\title{Load Local Data}
\usage{
loadData(...)

\S4method{loadData}{ANY}(...)
}
\arguments{
\item{...}{Objects as dots.}
}
\value{
No value.
}
\description{
Dynamically \code{\link[=load]{load()}} data from \code{data/} directory or \code{\link[=source]{source()}} the
corresponding script from the \code{data-raw/} directory.
}
\examples{
\dontrun{
loadData(geneIDs, oligo)
}
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{packageSE}},
  \code{\link{readFileByExtension}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
