% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataAsName.R
\name{loadDataAsName}
\alias{loadDataAsName}
\title{Load Data File as Name}
\usage{
loadDataAsName(..., dir = ".", envir = parent.frame())
}
\arguments{
\item{...}{Key value pairs, defining the name mappings. The argument name
defines the new name of the object in the environment, whereas the value
(string) denotes the original object name. For example, \code{newName1 = "oldName1", newName2 = "oldName2"}.}

\item{dir}{Output directory. Defaults to the current working directory.}

\item{envir}{Environment to use for assignment. Defaults to \code{parent.frame()},
which will assign into the calling environment.}
}
\value{
Silently return named character vector of file paths.
}
\description{
Load Data File as Name
}
\examples{
loadDataAsName(
    annotable = "grch37",
    tx2gene = "grch37Tx2gene",
    dir = system.file("extdata", package = "basejump")
)
glimpse(annotable)
}
