% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataAsName.R
\name{loadDataAsName}
\alias{loadDataAsName}
\title{Load Data as Name}
\usage{
loadDataAsName(..., dir = ".", envir = parent.frame())
}
\arguments{
\item{...}{Key value pairs, defining the name mappings. For example,
\code{newName1} = \code{oldName1}, \code{newName2} = \code{oldName2}. Note that these
arguments are interpreted using non-standard evaluation, and \emph{should not
be quoted}.}

\item{dir}{Output directory. Defaults to the current working directory.}

\item{envir}{Environment to use for assignment. Defaults to \code{parent.frame()},
which will assign into the calling environment.}
}
\value{
Silently return named character vector of file paths.
}
\description{
Load Data as Name
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation.
}
\examples{
loadDataAsName(
    hg19 = grch37,
    hg19_tx2gene = grch37Tx2gene,
    dir = system.file("extdata", package = "basejump")
)
glimpse(hg19)
}
\seealso{
Other Read Functions: \code{\link{loadData}},
  \code{\link{loadRemoteData}},
  \code{\link{localOrRemoteFile}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{readYAML}}
}
