% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataAsName.R
\name{loadDataAsName}
\alias{loadDataAsName}
\title{Load Data File as Name}
\usage{
loadDataAsName(mappings, dir = "data")
}
\arguments{
\item{mappings}{Named character vector to define mappings. The name defines
the new name of the object in the environment, whereas the value in the
vector denotes the original object name. This is designed to function
like a key value pair.}

\item{dir}{Output directory. Defaults to \strong{data}.}
}
\value{
No value.
}
\description{
Load Data File as Name
}
\examples{
\dontrun{
loadDataAsName(c(foo = "mtcars", bar = "starwars"))
}
}
