% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataAsName.R
\name{loadDataAsName}
\alias{loadDataAsName}
\title{Load Data File as Name}
\usage{
loadDataAsName(..., dir = "data", envir = parent.frame())
}
\arguments{
\item{...}{Key value pairs, defining the name mappings. The argument name
defines the new name of the object in the environment, whereas the value
(string) denotes the original object name. For example, \code{newName1 = "oldName1", newName2 = "oldName2"}.}

\item{dir}{Output directory. Defaults to \strong{data}.}

\item{envir}{The environment where the data should be loaded.}
}
\value{
Silently return named character vector of file paths.
}
\description{
Load Data File as Name
}
\examples{
# New dots method (>= 0.1.1)
\dontrun{
loadDataAsName(foo = "mtcars", bar = "starwars")
}

# Legacy named character vector mappings method
\dontrun{
loadDataAsName(c(foo = "mtcars", bar = "starwars"))
}
}
