% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadDataAsName}
\alias{loadDataAsName}
\title{Load Data as Name}
\usage{
loadDataAsName(..., dir = ".", envir = parent.frame())
}
\arguments{
\item{...}{Key value pairs, defining the name mappings. For example,
\code{newName1} = \code{oldName1}, \code{newName2} = \code{oldName2}. Note that these
arguments are interpreted using non-standard evaluation, and \emph{should not
be quoted}.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}

\item{envir}{\code{environment} to use for assignment. Defaults to
\code{parent.frame()}, which will assign into the calling environment.}
}
\value{
Invisible named \code{character}. File paths.
}
\description{
Load Data as Name
}
\note{
This function is intended for interactive use and interprets object
names using non-standard evaluation.
}
\examples{
loadDataAsName(
    renamed = example,
    dir = system.file("extdata", package = "basejump")
)
class(renamed)
}
\seealso{
Other Read Functions: \code{\link{import}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
