% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataRaw.R
\name{loadDataRaw}
\alias{loadDataRaw}
\alias{load_data_raw}
\title{Load data-raw directory.}
\usage{
loadDataRaw(data)

load_data_raw(data)
}
\arguments{
\item{data}{Data object.}
}
\description{
Dynamically source or load raw data in the \code{data-raw} directory.
}
\examples{
\dontrun{
loadDataRaw("counts")
}
}
