% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRemote.R
\name{loadRemote}
\alias{loadRemote}
\title{Load a remote R binary data file}
\usage{
loadRemote(url, ...)
}
\arguments{
\item{url}{URL}

\item{...}{Passthrough to `load()` function.}
}
\value{
Loaded R data file
}
\description{
Load a remote R binary data file
}

