% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRemoteData.R
\name{loadRemoteData}
\alias{loadRemoteData}
\title{Load Remote Data}
\usage{
loadRemoteData(url, envir = parent.frame())
}
\arguments{
\item{url}{Remote URL file path to R data. Supports multiple URLs passed in
as a character vector.}

\item{envir}{Environment to use for assignment. Defaults to \code{parent.frame()},
which will assign into the calling environment.}
}
\value{
Invisible named \code{character} containing the local object name as the
name, and the remote URL as the value.
}
\description{
Load a remote R binary file. This function is vectorized and supports
multiple URLs in a single call.
}
\examples{
loadRemoteData(c(
    "http://basejump.seq.cloud/rnaseqCounts.rda",
    "http://basejump.seq.cloud/singleCellCounts.rda"
))
class(rnaseqCounts)
class(singleCellCounts)
}
\seealso{
Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
