% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRemoteData.R
\name{loadRemoteData}
\alias{loadRemoteData}
\title{Load Remote Data}
\usage{
loadRemoteData(object, envir = parent.frame(), quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{envir}{Environment to use for assignment. Defaults to \code{parent.frame()},
which will assign into the calling environment.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Silently return loaded object name.
}
\description{
Load a remote R binary file.
}
\examples{
loadRemoteData("http://basejump.seq.cloud/mtcars.rda")
}
\seealso{
Other Data Import and Project Utilities: \code{\link{readFileByExtension}},
  \code{\link{readGFF}},
  \code{\link{readSampleMetadataFile}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
