% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadRemoteData}
\alias{loadRemoteData}
\title{Load Remote Data}
\usage{
loadRemoteData(url, envir = parent.frame())
}
\arguments{
\item{url}{\code{character}. Remote URL file path(s) to R data.}

\item{envir}{\code{environment} to use for assignment. Defaults to
\code{parent.frame()}, which will assign into the calling environment.}
}
\value{
Invisible named \code{character}. Local object name as the name, and the
remote URL as the value.
}
\description{
Load a remote R binary file. This function is vectorized and supports
multiple URLs in a single call.
}
\examples{
url <- file.path(
    basejumpCacheURL,
    paste0(c("rnaseq_counts", "single_cell_counts"), ".rds")
)
print(url)
x <- loadRemoteData(url)
print(x)
}
\seealso{
Other Import/Export Functions: \code{\link{assignAndSaveData}},
  \code{\link{export}}, \code{\link{import}},
  \code{\link{loadDataAsName}}, \code{\link{loadData}},
  \code{\link{saveData}}, \code{\link{transmit}},
  \code{\link{writeCounts}}
}
\concept{Import/Export Functions}
