% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-loadRemoteData.R
\docType{methods}
\name{loadRemoteData}
\alias{loadRemoteData}
\alias{loadRemoteData}
\alias{loadRemoteData,character-method}
\title{Load Remote Data}
\usage{
loadRemoteData(object, ...)

\S4method{loadRemoteData}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Load a remote R binary file.
}
\examples{
loadRemoteData(file.path(testDataURL, "mtcars.rda"))
}
\seealso{
Other Data Import and Project Utilities: \code{\link{prepareSE}},
  \code{\link{readFileByExtension}}, \code{\link{readGTF}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
