% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRemoteData.R
\name{loadRemoteData}
\alias{loadRemoteData}
\title{Load Remote Data}
\usage{
loadRemoteData(url, envir = parent.frame())
}
\arguments{
\item{url}{\code{character}. Remote URL file path(s) to R data.}

\item{envir}{\code{environment} to use for assignment. Defaults to
\code{parent.frame()}, which will assign into the calling environment.}
}
\value{
Invisible named \code{character}. Local object name as the name, and the
remote URL as the value.
}
\description{
Load a remote R binary file. This function is vectorized and supports
multiple URLs in a single call.
}
\examples{
loadRemoteData(c(
    "http://basejump.seq.cloud/rnaseq_counts.rda",
    "http://basejump.seq.cloud/single_cell_counts.rda"
))
class(rnaseq_counts)
class(single_cell_counts)
}
\seealso{
Other Read Functions: \code{\link{import}},
  \code{\link{loadDataAsName}}, \code{\link{loadData}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
