% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRemoteData.R
\name{loadRemoteData}
\alias{loadRemoteData}
\title{Load Remote Data}
\usage{
loadRemoteData(url, envir = parent.frame())
}
\arguments{
\item{url}{Remote URL file path to R Data file. Supports multiple URLs
passed in as a character vector.}

\item{envir}{Environment to use for assignment. Defaults to \code{parent.frame()},
which will assign into the calling environment.}
}
\value{
Silently return a character matrix containing URL and tempfile paths.
}
\description{
Load a remote R binary file. This function is vectorized and supports
multiple URLs in a single call.
}
\examples{
objects <- loadRemoteData(c(
    "http://basejump.seq.cloud/mtcars.rda",
    "http://basejump.seq.cloud/starwars.rda"
))
objects
}
\seealso{
Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{localOrRemoteFile}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{readYAML}}
}
