% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically Handle a Local or Remote File Path}
\usage{
localOrRemoteFile(object, severity = "stop", quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{severity}{Return \code{stop} (default), \code{warning}, or \code{message} if file
doesn't exist.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Named character vector containing the original file name as the
name and local file path as the string. Aborts on a missing file.
}
\description{
Dynamically Handle a Local or Remote File Path
}
\examples{
localOrRemoteFile("http://basejump.seq.cloud/mtcars.rda")
}
