% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically Handle a Local or Remote File Path}
\usage{
localOrRemoteFile(file)
}
\arguments{
\item{file}{File path.}
}
\value{
Named \code{character} containing the original basename as the name and
local file path (i.e. tempfile) as the string. Aborts on a missing file.
}
\description{
This function is vectorized and supports mixed local and remote paths.
}
\examples{
# Single file
x <- localOrRemoteFile("http://basejump.seq.cloud/rnaseqCounts.csv.gz")
names(x)

# Vectorized
x <- localOrRemoteFile(c(
    "http://basejump.seq.cloud/rnaseqCounts.csv.gz",
    "http://basejump.seq.cloud/singleCellCounts.mtx.gz"
))
names(x)
}
\seealso{
Other Developer Functions: \code{\link{detectHPC}},
  \code{\link{dots}}, \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}
}
