% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically Handle a Local or Remote File Path}
\usage{
localOrRemoteFile(file)
}
\arguments{
\item{file}{File path.}
}
\value{
Named character vector containing the original basename as the name
and local file path (i.e. tempfile) as the string. Aborts on a missing
file.
}
\description{
This function is vectorized and supports mixed local and remote paths.
}
\examples{
# Single file
file <- localOrRemoteFile("http://basejump.seq.cloud/mtcars.csv")
names(file)

# Vectorized
files <- localOrRemoteFile(c(
    "http://basejump.seq.cloud/mtcars.csv",
    "http://basejump.seq.cloud/mtcars.rda"
))
names(files)
}
\seealso{
Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}}, \code{\link{readYAML}}
}
