% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically Handle a Local or Remote File Path}
\usage{
localOrRemoteFile(file)
}
\arguments{
\item{file}{\code{string}. Local file paths or remote URLs.}
}
\value{
\code{character}. Local file paths.
}
\description{
Dynamically Handle a Local or Remote File Path
}
\note{
This function is vectorized and supports mixed local and remote paths. Remote
files are downloaded locally to a temporary directory.

Stops on a missing file.
}
\examples{
# Single file
x <- localOrRemoteFile("http://basejump.seq.cloud/rnaseq_counts.csv.gz")
basename(x)

# Vectorized
x <- localOrRemoteFile(c(
    "http://basejump.seq.cloud/rnaseq_counts.csv.gz",
    "http://basejump.seq.cloud/single_cell_counts.mtx.gz"
))
basename(x)
}
\seealso{
\code{\link[base:tempdir]{base::tempdir()}}.

Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{matchArgsToDoCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
