% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically Handle a Local or Remote File Path}
\usage{
localOrRemoteFile(object, severity = "stop", quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{severity}{Return \code{stop} (default), \code{warning}, or \code{message} if file
doesn't exist.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Named list containing the original basename as the name and local
file path (i.e. tempfile) as the string. Aborts on a missing file.
}
\description{
Dynamically Handle a Local or Remote File Path
}
\examples{
files <- localOrRemoteFile(c(
    "http://basejump.seq.cloud/mtcars.csv",
    "http://basejump.seq.cloud/mtcars.rda"
))
glimpse(files)
}
