% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically handle a local or remote file path}
\usage{
localOrRemoteFile(file)
}
\arguments{
\item{file}{\code{character(1)}.
Local file paths or remote URLs.}
}
\value{
\code{character}.
Local file paths. Stops on a missing file.
}
\description{
Dynamically handle a local or remote file path
}
\note{
This function is vectorized and supports mixed local and remote paths. Remote
files are downloaded locally to a temporary directory.
}
\examples{
## Import a single file.
file <- file.path(basejumpCacheURL, "rnaseq_counts.csv.gz")
x <- localOrRemoteFile(file)
basename(x)

## Import multiple files (vectorized).
files <- file.path(
    basejumpCacheURL,
    c("rnaseq_counts.csv.gz", "single_cell_counts.mtx.gz")
)
x <- localOrRemoteFile(files)
basename(x)
}
\seealso{
\code{tempdir}.
}
