% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatio.R
\name{fc2lr}
\alias{fc2lr}
\alias{lr2fc}
\title{Interconvert Log Ratio and Fold Change Values}
\usage{
fc2lr(x, base = 2L)

lr2fc(x, base = 2L)
}
\arguments{
\item{x}{Numeric vector of log ratio (\code{lr}) or fold change (\code{fc}) values.}

\item{base}{Logarithm base. Defaults to \code{2}, for compatibility with RNA-Seq
differential expression output.}
}
\value{
Numeric vector.
}
\description{
Interconvert Log Ratio and Fold Change Values
}
\examples{
lr2fc(seq(-3, 3, 1))
fc2lr(c(-8, -4, -2, 1, 2, 4, 8))
}
\seealso{
Modified variants of \code{gtools::foldchange2logratio()} and
\code{gtools::logratio2foldchange()}.

Other Math Utilities: \code{\link{geomean}},
  \code{\link{pct}}
}
