% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logRatio-methods.R
\docType{methods}
\name{logRatio}
\alias{logRatio}
\alias{foldChangeToLogRatio}
\alias{logRatioToFoldChange}
\alias{foldChangeToLogRatio,numeric-method}
\alias{logRatioToFoldChange,numeric-method}
\title{Interconvert Log Ratio and Fold Change Values}
\usage{
foldChangeToLogRatio(object, ...)

logRatioToFoldChange(object, ...)

\S4method{foldChangeToLogRatio}{numeric}(object, base = 2L)

\S4method{logRatioToFoldChange}{numeric}(object, base = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{base}{\code{scalar integer}. Logarithm base. Defaults to \code{2}, for
compatibility with RNA-Seq differential expression output.}
}
\value{
\code{numeric}.
}
\description{
Interconvert Log Ratio and Fold Change Values
}
\examples{
# Convert fold change to log ratio.
foldChangeToLogRatio(c(-8, -4, -2, 1, 2, 4, 8))

# Convert log ratio to fold change.
logRatioToFoldChange(seq(-3, 3, 1))
}
\seealso{
Modified variants of:
\itemize{
\item \code{gtools::foldchange2logratio()}.
\item \code{gtools::logratio2foldchange()}.
}

Other Calculations: \code{\link{aggregate}},
  \code{\link{geometricMean}}
}
\concept{Calculations}
