% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fc2lr.R,
%   R/methods-lr2fc.R
\docType{methods}
\name{fc2lr}
\alias{fc2lr}
\alias{lr2fc}
\alias{logRatio}
\alias{fc2lr,numeric-method}
\alias{lr2fc,numeric-method}
\title{Interconvert Log Ratio and Fold Change Values}
\usage{
fc2lr(object, ...)

lr2fc(object, ...)

\S4method{fc2lr}{numeric}(object, base = 2L)

\S4method{lr2fc}{numeric}(object, base = 2L)
}
\arguments{
\item{object}{Numeric vector of log ratio (\code{lr}) or fold change (\code{fc})
values.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{base}{Logarithm base. Defaults to \code{2}, for compatibility with RNA-Seq
differential expression output.}
}
\value{
Numeric vector.
}
\description{
Interconvert Log Ratio and Fold Change Values
}
\examples{
# Convert log ratio to fold change
lr2fc(seq(-3, 3, 1))

# Convert fold change to log ratio
fc2lr(c(-8, -4, -2, 1, 2, 4, 8))
}
\seealso{
Modified variants of \code{gtools::foldchange2logratio()`` and}gtools::logratio2foldchange()`.

Other Math Utilities: \code{\link{geomean}}
}
\concept{Math Utilities}
