% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-foldChangeToLogRatio.R, R/methods-logRatioToFoldChange.R
\docType{methods}
\name{foldChangeToLogRatio}
\alias{foldChangeToLogRatio}
\alias{logRatioToFoldChange}
\alias{logRatio}
\alias{foldChangeToLogRatio,integer-method}
\alias{foldChangeToLogRatio,numeric-method}
\alias{logRatioToFoldChange,integer-method}
\alias{logRatioToFoldChange,numeric-method}
\title{Interconvert Log Ratio and Fold Change Values}
\usage{
foldChangeToLogRatio(object, ...)

logRatioToFoldChange(object, ...)

\S4method{foldChangeToLogRatio}{integer}(object, base = 2L)

\S4method{foldChangeToLogRatio}{numeric}(object, base = 2L)

\S4method{logRatioToFoldChange}{integer}(object, base = 2L)

\S4method{logRatioToFoldChange}{numeric}(object, base = 2L)
}
\arguments{
\item{object}{Numeric of log ratio or fold change values.}

\item{...}{Additional arguments.}

\item{base}{Logarithm base. Defaults to \code{2}, for compatibility with RNA-Seq
differential expression output.}
}
\value{
\link{numeric}.
}
\description{
Interconvert Log Ratio and Fold Change Values
}
\examples{
# Convert log ratio to fold change
logRatioToFoldChange(seq(-3, 3, 1))

# Convert fold change to log ratio
foldChangeToLogRatio(c(-8, -4, -2, 1, 2, 4, 8))
}
\seealso{
Modified variants of \code{gtools::foldchange2logratio()} and
\code{gtools::logratio2foldchange()}.

Other Math Functions: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{geometricMean}}, \code{\link{microplate}}
}
