% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logRatio-methods.R
\name{foldChangeToLogRatio}
\alias{foldChangeToLogRatio}
\alias{logRatioToFoldChange}
\alias{logRatio}
\alias{foldChangeToLogRatio,numeric-method}
\alias{logRatioToFoldChange,numeric-method}
\title{Interconvert log ratio and fold change values}
\usage{
foldChangeToLogRatio(object, ...)

logRatioToFoldChange(object, ...)

\S4method{foldChangeToLogRatio}{numeric}(object, base = 2L)

\S4method{logRatioToFoldChange}{numeric}(object, base = 2L)
}
\arguments{
\item{object}{Object.}

\item{base}{\code{integer(1)}.
Logarithm base.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
}
\description{
Interconvert log ratio and fold change values
}
\note{
Updated 2019-07-28.
}
\examples{
## Convert fold change to log ratio.
foldChangeToLogRatio(c(-8, -4, -2, 1, 2, 4, 8))

## Convert log ratio to fold change.
logRatioToFoldChange(seq(-3, 3, 1))
}
\seealso{
\itemize{
\item \code{gtools::foldchange2logratio()}.
\item \code{gtools::logratio2foldchange()}.
}
}
