% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{logRatioToFoldChange}
\alias{logRatioToFoldChange}
\title{log ratio to fold change}
\usage{
logRatioToFoldChange(x, base = 2)
}
\arguments{
\item{x}{Numeric vector of log ratio values.}

\item{base}{Logarithm base. Defaults to \code{2}, for compatibility with RNA-Seq
differential expression output.}
}
\value{
Fold change values.
}
\description{
Convert log ratio normalized values to fold change. Based on the approach
used in \code{gtools::logratio2foldchange()}.
}
\examples{
logRatioToFoldChange(seq(-3, 3, 0.5))
}
