% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioToFoldChange.R
\name{logRatioToFoldChange}
\alias{logRatioToFoldChange}
\title{log ratio to fold change}
\usage{
logRatioToFoldChange(logRatio, base = 2)
}
\arguments{
\item{logRatio}{\code{numeric} vector of log ratio values}

\item{base}{Logarithm base. Defaults to 2, for compatibility with RNA-Seq
differential expression output.}
}
\value{
Fold change values
}
\description{
Convert log ratio normalized values to fold change
}
\examples{
logRatioToFoldChange(seq(-3, 3, 0.5))
}
\keyword{antilog}
\keyword{log}
\keyword{math}
\keyword{rnaseq}
