% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2Symbol.R
\name{makeGene2Symbol}
\alias{makeGene2Symbol}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGene2SymbolFromGFF}
\alias{makeGene2SymbolFromGTF}
\title{Make a Gene2Symbol object}
\usage{
makeGene2SymbolFromEnsembl(organism, genomeBuild = NULL,
  release = NULL, format = c("makeUnique", "unmodified", "1:1"))

makeGene2SymbolFromEnsDb(object, format = c("makeUnique", "unmodified",
  "1:1"))

makeGene2SymbolFromGFF(file, format = c("makeUnique", "unmodified",
  "1:1"))
}
\arguments{
\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"makeUnique"}: \emph{Recommended.} Apply \code{\link[base:make.unique]{make.unique()}}
to the \code{geneName} column. Gene symbols are made unique, while the gene
IDs remain unmodified.
\item \code{"unmodified"}: Return \code{geneID} and \code{geneName} columns unmodified, in
long format.
\item \code{"1:1"}: For gene symbols that map to multiple gene IDs, select only the
first annotated gene ID.
}}
}
\value{
\code{Gene2Symbol}.
}
\description{
Make a Gene2Symbol object
}
\note{
Updated 2019-07-28.
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
## makeGene2SymbolFromEnsembl ====
x <- makeGene2SymbolFromEnsembl(organism = "Homo sapiens")
print(x)

## makeTx2GeneFromEnsDb ====
x <- makeGene2SymbolFromEnsDb("EnsDb.Hsapiens.v75")
print(x)

## makeGene2SymbolFromGFF ====
## GTF
file <- file.path(basejumpTestsURL, "example.gtf")
x <- makeGene2SymbolFromGFF(file)
print(x)

## GFF3
file <- file.path(basejumpTestsURL, "example.gff3")
x <- makeGene2SymbolFromGFF(file)
print(x)
}
