% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2Symbol-methods.R
\name{makeGene2Symbol}
\alias{makeGene2Symbol}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGene2SymbolFromGFF}
\alias{makeGene2SymbolFromGTF}
\title{Make Gene-to-Symbol Mappings}
\usage{
makeGene2SymbolFromEnsembl(organism, genomeBuild = NULL,
  release = NULL)

makeGene2SymbolFromEnsDb(object)

makeGene2SymbolFromGFF(file)
}
\arguments{
\item{organism}{\code{character(1)}. Full Latin organism name
(e.g. "\code{Homo sapiens}").}

\item{genomeBuild}{\code{character(1)}. Ensembl genome build assembly name (e.g.
\code{"GRCh38"}). If set \code{NULL}, defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{scalar integer}. Ensembl release version (e.g. \code{90}). If set
\code{NULL}, defaults to the most recent release available.}

\item{object}{Object.}

\item{file}{\code{character(1)}. File path.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Make Gene-to-Symbol Mappings
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
## makeGene2SymbolFromEnsembl ====
x <- makeGene2SymbolFromEnsembl(organism = "Homo sapiens")
print(x)

## makeTx2GeneFromEnsDb ====
x <- makeGene2SymbolFromEnsDb("EnsDb.Hsapiens.v75")
print(x)

## makeGene2SymbolFromGFF ====
## GTF
file <- file.path(basejumpCacheURL, "example.gtf")
x <- makeGene2SymbolFromGFF(file)
print(x)

## GFF3
file <- file.path(basejumpCacheURL, "example.gff3")
x <- makeGene2SymbolFromGFF(file)
print(x)
}
\seealso{
\code{makeGRanges}.
}
