% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2Symbol.R
\name{makeGene2Symbol}
\alias{makeGene2Symbol}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGene2SymbolFromGFF}
\alias{makeGene2SymbolFromGTF}
\title{Make a Gene2Symbol object}
\usage{
makeGene2SymbolFromEnsembl(organism, genomeBuild = NULL,
  release = NULL, format = c("makeUnique", "unmodified", "1:1"))

makeGene2SymbolFromEnsDb(object, format = c("makeUnique", "unmodified",
  "1:1"))

makeGene2SymbolFromGFF(file, format = c("makeUnique", "unmodified",
  "1:1"))
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{object}{Object.}

\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Make a Gene2Symbol object
}
\note{
Updated 2019-07-28.
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
## makeGene2SymbolFromEnsembl ====
x <- makeGene2SymbolFromEnsembl(organism = "Homo sapiens")
print(x)

## makeTx2GeneFromEnsDb ====
x <- makeGene2SymbolFromEnsDb("EnsDb.Hsapiens.v75")
print(x)

## makeGene2SymbolFromGFF ====
## GTF
file <- file.path(basejumpTestsURL, "example.gtf")
x <- makeGene2SymbolFromGFF(file)
print(x)

## GFF3
file <- file.path(basejumpTestsURL, "example.gff3")
x <- makeGene2SymbolFromGFF(file)
print(x)
}
