% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2Symbol.R
\name{makeGene2Symbol}
\alias{makeGene2Symbol}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromGFF}
\alias{makeGene2SymbolFromGTF}
\title{Make Gene-to-Symbol Mappings}
\usage{
makeGene2SymbolFromEnsembl(organism, build = NULL, release = NULL)

makeGene2SymbolFromGFF(file)
}
\arguments{
\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{build}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}

\item{file}{\code{string}. File path.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Make Gene-to-Symbol Mappings
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
# makeGene2SymbolFromEnsembl ====
x <- makeGene2SymbolFromEnsembl(organism = "Homo sapiens")
print(x)

# makeGene2SymbolFromGFF ====
# GTF
file <- file.path(basejumpCacheURL, "example.gtf")
x <- makeGene2SymbolFromGFF(file)
print(x)

# GFF3
file <- file.path(basejumpCacheURL, "example.gff3")
x <- makeGene2SymbolFromGFF(file)
print(x)
}
\seealso{
\link{makeGRanges}.

Other Annotation Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{makeGRanges}}, \code{\link{makeTx2Gene}},
  \code{\link{organism}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
