% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2symbol.R
\name{makeGene2symbol}
\alias{makeGene2symbol}
\alias{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGTF}
\title{Make Gene-to-Symbol Mappings}
\usage{
makeGene2symbolFromEnsembl(organism, build = NULL, release = NULL)

makeGene2symbolFromGFF(file)

makeGene2symbolFromGTF(file)
}
\arguments{
\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{build}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}

\item{file}{\code{string}. File path.}
}
\value{
\code{gene2symbol}.
}
\description{
Make Gene-to-Symbol Mappings
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
# makeGene2symbolFromEnsembl ====
x <- makeGene2symbolFromEnsembl(organism = "Homo sapiens")
print(x)

# makeGene2symbolFromGFF ====
# GTF
file <- file.path(basejumpCacheURL, "example.gtf")
x <- makeGene2symbolFromGFF(file)
print(x)

# GFF3
file <- file.path(basejumpCacheURL, "example.gff3")
x <- makeGene2symbolFromGFF(file)
print(x)
}
\seealso{
\link{makeGRanges}.

Other Annotation Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{eggnog}}, \code{\link{emptyRanges}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{geneSynonyms}}, \code{\link{hgnc2ensembl}},
  \code{\link{makeGRanges}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2ensembl}}, \code{\link{organism}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
