% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2symbol.R
\name{makeGene2symbol}
\alias{makeGene2symbol}
\alias{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGTF}
\title{Make Gene-to-Symbol Mappings}
\usage{
makeGene2symbolFromEnsembl(...)

makeGene2symbolFromGFF(file)
}
\arguments{
\item{...}{Additional arguments.}

\item{file}{\code{string}. File path.}
}
\value{
\code{DataFrame}.
}
\description{
Make Gene-to-Symbol Mappings
}
\examples{
# makeGene2symbolFromEnsembl ====
x <- makeGene2symbolFromEnsembl("Homo sapiens")
glimpse(x)

# makeGene2symbolFromGFF ====
# GTF
x <- makeGene2symbolFromGFF("http://basejump.seq.cloud/example.gtf")
glimpse(x)

# GFF3
x <- makeGene2symbolFromGFF("http://basejump.seq.cloud/example.gff3")
glimpse(x)
}
\seealso{
\link{makeGRanges}.

Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}}, \code{\link{makeGRanges}},
  \code{\link{makeTx2gene}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
