% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2symbol.R
\name{makeGene2symbol}
\alias{makeGene2symbol}
\alias{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGTF}
\title{Make Gene-to-Symbol Mappings}
\usage{
makeGene2symbolFromEnsembl(organism, build = NULL, release = NULL, ...)

makeGene2symbolFromGFF(file)
}
\arguments{
\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{build}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}

\item{...}{Additional arguments.}

\item{file}{\code{string}. File path.}
}
\value{
\code{DataFrame}.
}
\description{
Make Gene-to-Symbol Mappings
}
\examples{
# makeGene2symbolFromEnsembl ====
x <- makeGene2symbolFromEnsembl("Homo sapiens")
print(x)

# makeGene2symbolFromGFF ====
# GTF
x <- makeGene2symbolFromGFF("http://basejump.seq.cloud/example.gtf")
print(x)

# GFF3
x <- makeGene2symbolFromGFF("http://basejump.seq.cloud/example.gff3")
print(x)
}
\seealso{
\link{makeGRanges}.

Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2ensembl}}, \code{\link{makeGRanges}},
  \code{\link{makeTx2gene}}, \code{\link{mgi2ensembl}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
