% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2symbolFromEnsembl.R
\name{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromEnsembl}
\title{Gene-to-Symbol Mappings from Ensembl}
\usage{
makeGene2symbolFromEnsembl(organism, genomeBuild = NULL,
  release = NULL)
}
\arguments{
\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{genomeBuild}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}
}
\value{
\code{data.frame}.
}
\description{
Gene-to-Symbol Mappings from Ensembl
}
\examples{
x <- makeGene2symbolFromEnsembl("Homo sapiens")
glimpse(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Gene Annotation Functions}
