% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2symbolFromEnsembl.R
\name{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromEnsembl}
\title{Gene-to-Symbol Mappings from Ensembl}
\usage{
makeGene2symbolFromEnsembl(..., unique = TRUE)
}
\arguments{
\item{...}{Passthrough to \code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}}.}

\item{unique}{\code{boolean}. Ensure that all gene symbols are unique. Uses
\code{\link[base:make.unique]{base::make.unique()}} internally.}
}
\value{
\code{data.frame}.
}
\description{
Gene-to-Symbol Mappings from Ensembl
}
\examples{
x <- makeGene2symbolFromEnsembl("Homo sapiens")
glimpse(x)
}
\seealso{
Other Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
