% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2symbolFromEnsembl.R
\name{makeGene2symbolFromEnsembl}
\alias{makeGene2symbolFromEnsembl}
\title{Gene-to-Symbol Mappings from Ensembl}
\usage{
makeGene2symbolFromEnsembl(organism, genomeBuild = NULL, release = NULL)
}
\arguments{
\item{organism}{Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{genomeBuild}{\emph{Optional.} Genome build assembly name (e.g. "\code{GRCh38}").
If \code{NULL}, defaults to the most recent build available.}

\item{release}{\emph{Optional.} Release version (e.g. \code{90}). If \code{NULL}, defaults
to the most recent release available.}
}
\value{
\code{data.frame}.
}
\description{
Gene-to-Symbol Mappings from Ensembl
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}},
  \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
