% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGene2symbolFromGFF.R
\name{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGFF}
\alias{makeGene2symbolFromGTF}
\title{Gene-to-Symbol Mappings from GFF File}
\usage{
makeGene2symbolFromGFF(file)
}
\arguments{
\item{file}{File path.}
}
\value{
\code{data.frame}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
invisible(capture.output(
    x <- makeGene2symbolFromGFF("http://basejump.seq.cloud/mmusculus.gtf")
))
glimpse(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
