% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRanges.R
\name{makeGRanges}
\alias{makeGRanges}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\alias{annotable}
\title{Make Genomic Ranges}
\usage{
makeGRangesFromEnsembl(organism, level = c("genes", "transcripts"),
  build = NULL, release = NULL)

makeGRangesFromGFF(file, level = c("genes", "transcripts"))

annotable(organism, level = c("genes", "transcripts"), build = NULL,
  release = NULL)
}
\arguments{
\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{level}{\code{string}. Return ranges as "\code{genes}" or "\code{transcripts}".}

\item{build}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}

\item{file}{\code{string}. File path.}
}
\value{
\itemize{
\item \code{makeGRangesFromEnsembl()}, \code{makeGRangesFromGFF()}: \code{GRanges}.
\item \code{annotable()}: \code{tbl_df}.
}
}
\description{
Make Genomic Ranges
}
\section{Ensembl annotations}{


Quickly obtain gene and transcript annotations from
\href{http://www.ensembl.org}{Ensembl} using
\href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub} and
\href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.

Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.

Genome build: use "\code{GRCh38}" instead of "\code{hg38}" for the genome build, since
we're querying Ensembl and not UCSC.
}

\section{GTF/GFF file}{


The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2. We recommend
using a GTF file instead of a GFF3 file, if possible.

The UCSC website has detailed conventions on the GFF3 format, including
the metadata columns.

Remote URLs and compressed files are supported.
}

\section{GRCh37 legacy annotations}{


\code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} supports the legacy \emph{Homo sapiens} GRCh37 (release
75) build by internally querying the
\href{https://doi.org/doi:10.18129/B9.bioc.EnsDb.Hsapiens.v75}{EnsDb.Hsapiens.v75}
package. Alternatively, the corresponding GTF/GFF file can be loaded directly
from GENCODE or Ensembl.
}

\section{Broad class definitions}{


For gene and transcript annotations, a \code{broadClass} column is added, which
generalizes the gene types into a smaller number of semantically-meaningful
groups:
\itemize{
\item \code{coding}.
\item \code{noncoding}.
\item \code{pseudo}.
\item \code{small}.
\item \code{decaying}.
\item \code{ig} (immunoglobulin).
\item \code{tcr} (T cell receptor).
\item \code{other}.
}
}

\section{annotable}{


\code{\link[=annotable]{annotable()}} is a legacy convenience function that calls
\code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} and returns a \code{tibble} instead of \code{GRanges}. Note
that \code{GRanges} also support coercion to a basic \code{data.frame} using
\code{\link[=as.data.frame]{as.data.frame()}}.
}

\examples{
# makeGRangesFromEnsembl ====
# Genes
x <- makeGRangesFromEnsembl("Homo sapiens", level = "genes")
summary(x)

# Transcripts
x <- makeGRangesFromEnsembl("Homo sapiens", level = "transcripts")
summary(x)

# makeGRangesFromGFF ====
file <- file.path(basejumpCacheURL, "example.gtf")

# Genes
x <- makeGRangesFromGFF(file = file, level = "genes")
summary(x)

# Transcripts
x <- makeGRangesFromGFF(file = file, level = "transcripts")
summary(x)

# annotable ====
# Genes
x <- annotable("Homo sapiens")
glimpse(x)

# Transcripts
x <- annotable("Homo sapiens", level = "transcripts")
glimpse(x)
}
\seealso{
\itemize{
\item \href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
\item \href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.
\item \code{\link[rtracklayer:import]{rtracklayer::import()}}.
\item \code{\link[GenomicFeatures:makeTxDbFromGFF]{GenomicFeatures::makeTxDbFromGFF()}}.
}

Other Annotation Functions: \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{makeGene2Symbol}}, \code{\link{makeTx2Gene}},
  \code{\link{panther}}

Other Transcript-Level Functions: \code{\link{convertTranscriptsToGenes}},
  \code{\link{makeTx2Gene}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{tx2gene}}
}
\concept{Annotation Functions}
\concept{Transcript-Level Functions}
