% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromEnsDb.R
\name{makeGRangesFromEnsDb}
\alias{makeGRangesFromEnsDb}
\title{Make GRanges from EnsDb object}
\usage{
makeGRangesFromEnsDb(object, level = c("genes", "transcripts"))
}
\arguments{
\item{object}{Object.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}
}
\description{
Make GRanges from EnsDb object
}
\details{
Use specific \code{EnsDb} object as annotation source.
Alternatively, can pass in an EnsDb package name as a \code{character(1)}.
}
\examples{
if ("EnsDb.Hsapiens.v75" \%in\% rownames(installed.packages())) {
    x <- makeGRangesFromEnsDb("EnsDb.Hsapiens.v75")
}
}
