% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromEnsembl.R
\name{makeGRangesFromEnsembl}
\alias{makeGRangesFromEnsembl}
\title{Genomic Ranges from Ensembl}
\usage{
makeGRangesFromEnsembl(organism, format = c("genes", "transcripts"),
  genomeBuild = NULL, release = NULL, metadata = FALSE)
}
\arguments{
\item{organism}{Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{format}{Fetch ranges as "\code{genes}" or "\code{transcripts}".}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. "\code{GRCh38}"). If
\code{NULL}, defaults to the most recent build available.}

\item{release}{\emph{Optional.} Release version (e.g. \code{90}). If \code{NULL}, defaults
to the most recent release available.}

\item{metadata}{Include the AnnotationHub metadata inside a \code{list}.}
}
\value{
\code{GRanges}.
}
\description{
Quickly obtain gene and transcript annotations from
\href{http://www.ensembl.org}{Ensembl} using
\href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub} and
\href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.
}
\details{
Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.
}
\note{
Use "\code{GRCh38}" instead of "\code{hg38}" for the genome build, since we're
querying Ensembl and not UCSC.
}
\section{Broad Class Definitions}{

For gene and transcript tables, when \code{broadClass = TRUE}, a \code{broadClass}
column is added, which generalizes the gene types into a smaller number of
semantically-meaningful groups:
\itemize{
\item \code{coding}
\item \code{noncoding}
\item \code{pseudo}
\item \code{small}
\item \code{decaying}
\item \code{ig} (immunoglobulin)
\item \code{tcr} (T cell receptor)
\item \code{other}
}
}

\section{GRCh37 Legacy Annotations}{

The \emph{Homo sapiens} GRCh37 (release 75) build is supported by internally
querying the
\href{https://doi.org/doi:10.18129/B9.bioc.EnsDb.Hsapiens.v75}{EnsDb.Hsapiens.v75}
package.
}

\examples{
# Genes ====
x <- makeGRangesFromEnsembl("Homo sapiens", format = "genes")
summary(x)

# Transcripts ====
x <- makeGRangesFromEnsembl("Homo sapiens", format = "transcripts")
summary(x)
}
\seealso{
\itemize{
\item \href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
\item \href{https://doi.org/doi:10.18129/B9.bioc.ensembldb}{ensembldb}.
}

Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}},
  \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
