% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromGFF.R
\name{makeGRangesFromGFF}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\title{Make \code{GRanges} from a GFF/GTF file}
\usage{
makeGRangesFromGFF(file, level = c("genes", "transcripts"),
  strict = FALSE)

makeGRangesFromGTF(file, level = c("genes", "transcripts"),
  strict = FALSE)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{level}{\code{character(1)}.
Return ranges as "\code{genes}" or "\code{transcripts}".}

\item{strict}{\code{logical(1)}.
Strict mode. Generate an internal \code{TxDb} using
\code{\link[GenomicFeatures:makeTxDbFromGRanges]{GenomicFeatures::makeTxDbFromGRanges()}} and check that the
\code{\link[IRanges:ranges]{ranges()}}, \code{\link[GenomeInfoDb:seqnames]{seqnames()}}, and
identifiers defined in \code{\link[base:names]{names()}} are identical. Doesn't work
for all GFF/GTF files due to some current limitations in the
GenomicFeatures package, so this is disabled by default. Generally,
GenomicFeatures parses GTF files better than GFF files. However, it's a
useful sanity check and should be enabled if possible.}
}
\description{
Make \code{GRanges} from a GFF/GTF file
}
\details{
Remote URLs and compressed files are supported.
}
\section{Functions}{
\itemize{
\item \code{makeGRangesFromGTF}: GTF file extension alias.
Runs the same internal code as \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}}.
}}

\section{Recommendations}{


\strong{GTF over GFF3.} We recommend using a GTF file instead of a GFF3 file,
when possible. The file format is more compact and easier to parse.

\strong{Ensembl over RefSeq.} We generally recommend using Ensembl over RefSeq,
if possible. It's better supported in R and generally used by most NGS
vendors.
}

\section{GFF/GTF specification}{


The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines.

The GTF (General Transfer Format) format is identical to GFF version 2.

The UCSC website has detailed conventions on the GFF3 format, including
the metadata columns.
}

\section{Supported sources}{


Currently \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}} supports genomes from these sources:
\itemize{
\item Ensembl (GTF, GFF3).
\item GENCODE, which inherits from Ensembl (GTF, GFF3).
\item RefSeq (GFF3).
\item FlyBase (GTF).
\item WormBase (GTF).
}

UCSC support will be added in a future release.
}

\section{Ensembl}{


Note that \code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} offers native support for Ensembl genome
builds and returns additional useful metadata that isn't defined inside a
GFF/GTF file.

If you must load a GFF/GTF file directly, then use \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}}.

GTF key-value pairs:
\itemize{
\item source
\item type
\item score
\item phase
\item gene_id
\item gene_version
\item gene_name
\item gene_source
\item gene_biotype
\item transcript_id
\item transcript_version
\item transcript_name
\item transcript_source
\item transcript_biotype
\item tag
\item transcript_support_level
\item exon_number
\item exon_id
\item exon_version
\item protein_id
\item protein_version
\item ccds_id
}

GFF key-value pairs:
\itemize{
\item "source"
\item "type"
\item "score"
\item "phase"
\item "ID"
\item "Alias"
\item "external_name"
\item "logic_name"
\item "Name"
\item "biotype"
\item "description"
\item "gene_id"
\item "version"
\item "Parent"
\item "tag"
\item "transcript_id"
\item "transcript_support_level"
\item "constitutive"
\item "ensembl_end_phase"
\item "ensembl_phase"
\item "exon_id"
\item "rank"
\item "protein_id"
\item "ccdsid"
}
}

\section{GENCODE vs. Ensembl}{


Annotations available from Ensembl and GENCODE are very similar.

The GENCODE annotation is made by merging the manual gene annotation produced
by the Ensembl-Havana team and the Ensembl-genebuild automated gene
annotation. The GENCODE annotation is the default gene annotation displayed
in the Ensembl browser. The GENCODE releases coincide with the Ensembl
releases, although GENCODE can skip an Ensembl release if there is no update
to the annotation with respect to the previous release. In practical terms,
the GENCODE annotation is essentially identical to the Ensembl annotation.

However, GENCODE handles pseudoautosomal regions (PAR) differently than
Ensembl. The Ensembl GTF file only includes this annotation once, for
chromosome X. However, GENCODE GTF/GFF3 files include the annotation in the
PAR regions of both chromosomes.

The \href{https://www.gencodegenes.org/pages/faq.html}{GENCODE FAQ} has additional
details.
}

\section{RefSeq}{


Refer to the
\href{ftp://ftp.ncbi.nlm.nih.gov/genomes/README_GFF3.txt}{Current RefSeq GFF3 spec}
for additional details.

GFF key-value pairs:
\itemize{
\item source
\item type
\item score
\item phase
\item ID
\item Dbxref
\item Name
\item chromosome
\item gbkey
\item genome
\item mol_type
\item description
\item gene
\item gene_biotype
\item pseudo
\item Parent
\item product
\item transcript_id
\item gene_synonym
\item model_evidence
\item protein_id
}

See also:
\itemize{
\item \href{https://www.ncbi.nlm.nih.gov/books/NBK50679/}{RefSeq FAQ}
\item ftp://ftp.ncbi.nih.gov/gene/DATA/gene2refseq.gz
}
}

\section{UCSC (not supported)}{


Loading UCSC genome annotations from a GFF/GTF file are
\emph{intentionally not supported} by this function.

Use a pre-built TxDb package instead
(e.g. TxDb.Hsapiens.UCSC.hg38.knownGene).

Note that UCSC doesn't provide direct GFF/GTF file downloads.
Use of the \href{https://genome.ucsc.edu/cgi-bin/hgTables}{hgTables} table
browser is required in a web browser.

Select the following options to download hg38:
\itemize{
\item clade: Mammal
\item genome: Human
\item assembly: Dec. 2013 (GRCh38/hg38)
\item group: Genes and Gene Predictions
\item track: GENCODE v29
\item table: knownGene
\item region: genome.
\item output format: GTF - gene transfer format
\item output file: enter a file name
}

See also:
\itemize{
\item http://genome.ucsc.edu/cgi-bin/hgTables
\item http://hgdownload.soe.ucsc.edu/downloads.html
\item ftp://hgdownload.soe.ucsc.edu/goldenPath/hg38/
}
}

\section{Homo sapiens URLs}{

\itemize{
\item Ensembl GTF:\cr
ftp://ftp.ensembl.org/pub/release-95/gtf/homo_sapiens/Homo_sapiens.GRCh38.95.gtf.gz
\item Ensembl GFF:\cr
ftp://ftp.ensembl.org/pub/release-95/gff3/homo_sapiens/Homo_sapiens.GRCh38.95.gff3.gz
\item GENCODE GTF:\cr
ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_29/gencode.v29.annotation.gtf.gz
\item GENCODE GFF:\cr
ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_29/gencode.v29.annotation.gff3.gz
\item RefSeq GFF:\cr
ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/vertebrate_mammalian/Homo_sapiens/reference/GCF_000001405.38_GRCh38.p12/GCF_000001405.38_GRCh38.p12_genomic.gff.gz
}
}

\examples{
file <- system.file("extdata/ensembl.gtf.gz", package = "basejump")

## Genes
x <- makeGRangesFromGFF(file = file, level = "genes")
summary(x)

## Transcripts
x <- makeGRangesFromGFF(file = file, level = "transcripts")
summary(x)
}
\seealso{
\itemize{
\item \code{rtracklayer::import()}.
\item \code{GenomicFeatures::makeTxDbFromGRanges()}.
\item \code{GenomicFeatures::makeTxDbFromGFF()}.
}
}
