% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromGFF.R
\name{makeGRangesFromGFF}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\title{Make \code{GRanges} from GFF/GTF file}
\usage{
makeGRangesFromGFF(file, level = c("genes", "transcripts"))

makeGRangesFromGTF(file, level = c("genes", "transcripts"))
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{level}{\code{character(1)}.
Return ranges as "\code{genes}" or "\code{transcripts}".}
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2. We recommend
using a GTF file instead of a GFF3 file, if possible.

The UCSC website has detailed conventions on the GFF3 format, including
the metadata columns.

Remote URLs and compressed files are supported.
}
\section{Functions}{
\itemize{
\item \code{makeGRangesFromGTF}: GTF file extension alias.
Runs the same internal code as \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}}.
}}

\section{Commonly used GFF/GTF files}{

\itemize{
\item Ensembl GTF:\cr
ftp://ftp.ensembl.org/pub/release-95/gtf/homo_sapiens/Homo_sapiens.GRCh38.95.gtf.gz
\item Ensembl GFF3:\cr
ftp://ftp.ensembl.org/pub/release-95/gff3/homo_sapiens/Homo_sapiens.GRCh38.95.gff3.gz
\item RefSeq:\cr
ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/vertebrate_mammalian/Homo_sapiens/reference/GCF_000001405.38_GRCh38.p12/GCF_000001405.38_GRCh38.p12_genomic.gff.gz
}
}

\examples{
file <- file.path(basejumpCacheURL, "example.gtf")

## Genes
x <- makeGRangesFromGFF(file = file, level = "genes")
summary(x)

## Transcripts
x <- makeGRangesFromGFF(file = file, level = "transcripts")
summary(x)
}
\seealso{
\itemize{
\item \code{rtracklayer::import()}.
\item \code{GenomicFeatures::makeTxDbFromGFF()}.
}
}
