% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromGFF.R
\name{makeGRangesFromGFF}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\title{Make \code{GRanges} from a GFF/GTF file}
\usage{
makeGRangesFromGFF(file, level = c("genes", "transcripts"),
  .checkAgainstTxDb = FALSE)

makeGRangesFromGTF(file, level = c("genes", "transcripts"),
  .checkAgainstTxDb = FALSE)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{.checkAgainstTxDb}{\code{logical(1)}.
Enable strict mode, intended for development and unit testing only.
Generate an internal \code{TxDb} using \code{\link[GenomicFeatures:makeTxDbFromGRanges]{GenomicFeatures::makeTxDbFromGRanges()}}
and check that the \code{\link[IRanges:ranges]{ranges()}},
\code{\link[GenomeInfoDb:seqnames]{seqnames()}}, and identifiers defined in
\code{\link[base:names]{names()}} are identical. Doesn't work for all GFF/GTF files
due to some current limitations in the GenomicFeatures package, so this is
disabled by default. Generally, GenomicFeatures parses GTF files better
than GFF files. However, it's a useful sanity check and should be enabled
if possible.}
}
\description{
Make \code{GRanges} from a GFF/GTF file
}
\details{
Remote URLs and compressed files are supported.
}
\section{Functions}{
\itemize{
\item \code{makeGRangesFromGTF}: GTF file extension alias.
Runs the same internal code as \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}}.
}}

\section{Recommendations}{

\itemize{
\item \strong{Use GTF over GFF3.} We recommend using a GTF file instead of a GFF3
file, when possible. The file format is more compact and easier to parse.
\item \strong{Use Ensembl over RefSeq.} We generally recommend using Ensembl over
RefSeq, if possible. It's better supported in R and generally used by most
NGS vendors.
}
}

\section{GFF/GTF specification}{


The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines.

The GTF (General Transfer Format) format is identical to GFF version 2.

The UCSC website has detailed conventions on the GFF3 format, including
the metadata columns.
}

\section{Supported sources}{


Currently \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}} supports genomes from these sources:

\tabular{lrrl}{
\tab   GTF \tab  GFF3 \tab                                     \cr
Ensembl  \tab  TRUE \tab  TRUE \tab Also available via ensembldb.       \cr
GENCODE  \tab  TRUE \tab  TRUE \tab Similar to Ensembl; duplicate PAR_Y.\cr
RefSeq   \tab    NA \tab  TRUE \tab Experimental support.               \cr
UCSC     \tab FALSE \tab FALSE \tab Use pre-built Bioconductor packages.\cr
FlyBase  \tab  TRUE \tab FALSE \tab Newer than Ensembl.                 \cr
WormBase \tab  TRUE \tab FALSE \tab Very similar to Ensembl.            \cr
}
}

\section{Ensembl}{


Note that \code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} offers native support for Ensembl genome
builds and returns additional useful metadata that isn't defined inside a
GFF/GTF file.

If you must load a GFF/GTF file directly, then use \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}}.
}

\section{GENCODE vs. Ensembl}{


Annotations available from Ensembl and GENCODE are very similar.

The GENCODE annotation is made by merging the manual gene annotation produced
by the Ensembl-Havana team and the Ensembl-genebuild automated gene
annotation. The GENCODE annotation is the default gene annotation displayed
in the Ensembl browser. The GENCODE releases coincide with the Ensembl
releases, although GENCODE can skip an Ensembl release if there is no update
to the annotation with respect to the previous release. In practical terms,
the GENCODE annotation is essentially identical to the Ensembl annotation.

However, GENCODE handles pseudoautosomal regions (PAR) differently than
Ensembl. The Ensembl GTF file only includes this annotation once, for
chromosome X. However, GENCODE GTF/GFF3 files include the annotation in the
PAR regions of both chromosomes. You'll see these genes contain a "_PAR_Y"
suffix.

Additionally, GENCODE GFF/GTF files import with a gene identifier containing
a suffix, which differs slightly from the Ensembl GFF/GTF spec
(e.g. GENCODE: ENSG00000000003.14; Ensembl: ENSG00000000003).

The \href{https://www.gencodegenes.org/pages/faq.html}{GENCODE FAQ} has additional
details.
}

\section{RefSeq}{


Refer to the
\href{ftp://ftp.ncbi.nlm.nih.gov/genomes/README_GFF3.txt}{current RefSeq spec}
for details.

See also:
\itemize{
\item \href{https://www.ncbi.nlm.nih.gov/books/NBK50679/}{RefSeq FAQ}
\item ftp://ftp.ncbi.nih.gov/gene/DATA/gene2refseq.gz
}
}

\section{UCSC}{


Loading UCSC genome annotations from a GFF/GTF file are
\emph{intentionally not supported} by this function.

We recommend using a pre-built \code{TxDb} package from Bioconductor instead.
For example, load \code{TxDb.Hsapiens.UCSC.hg38.knownGene} for hg38.

For reference, note that UCSC doesn't provide direct GFF/GTF file downloads.
Use of the \href{https://genome.ucsc.edu/cgi-bin/hgTables}{hgTables} table
browser is required in a web browser.

Select the following options to download hg38:
\itemize{
\item clade: \code{Mammal}
\item genome: \code{Human}
\item assembly: \code{Dec. 2013 (GRCh38/hg38)}
\item group: \code{Genes and Gene Predictions}
\item track: \code{GENCODE v29}
\item table: \code{knownGene}
\item region: \code{genome}
\item output format: \code{GTF - gene transfer format}
\item output file: \code{<Enter a file name>}
}

See also:
\itemize{
\item http://genome.ucsc.edu/cgi-bin/hgTables
\item http://hgdownload.soe.ucsc.edu/downloads.html
\item ftp://hgdownload.soe.ucsc.edu/goldenPath/hg38/
}
}

\section{Example URLs}{

\itemize{
\item Ensembl GTF:\cr
ftp://ftp.ensembl.org/pub/release-95/gtf/homo_sapiens/Homo_sapiens.GRCh38.95.gtf.gz
\item Ensembl GFF3:\cr
ftp://ftp.ensembl.org/pub/release-95/gff3/homo_sapiens/Homo_sapiens.GRCh38.95.gff3.gz
\item GENCODE GTF:\cr
ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_29/gencode.v29.annotation.gtf.gz
\item GENCODE GFF3:\cr
ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_29/gencode.v29.annotation.gff3.gz
\item RefSeq GFF3:\cr
ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/vertebrate_mammalian/Homo_sapiens/reference/GCF_000001405.38_GRCh38.p12/GCF_000001405.38_GRCh38.p12_genomic.gff.gz
\item FlyBase GTF:\cr
ftp://ftp.flybase.net/releases/FB2018_05/dmel_r6.24/gtf/dmel-all-r6.24.gtf.gz
\item WormBase GTF:\cr
ftp://ftp.wormbase.org/pub/wormbase/releases/WS267/species/c_elegans/PRJNA13758/c_elegans.PRJNA13758.WS267.canonical_geneset.gtf.gz
}
}

\examples{
file <- system.file("extdata/ensembl.gtf.gz", package = "basejump")

## Genes
x <- makeGRangesFromGFF(file = file, level = "genes")
summary(x)

## Transcripts
x <- makeGRangesFromGFF(file = file, level = "transcripts")
summary(x)
}
\seealso{
\itemize{
\item \code{\link[rtracklayer:import]{rtracklayer::import()}}.
\item \code{\link[GenomicFeatures:makeTxDbFromGRanges]{GenomicFeatures::makeTxDbFromGRanges()}}.
\item \code{\link[GenomicFeatures:makeTxDbFromGFF]{GenomicFeatures::makeTxDbFromGFF()}}.
}
}
