% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromGFF.R
\name{makeGRangesFromGFF}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\title{Genomic Ranges from GFF File}
\usage{
makeGRangesFromGFF(file, level = c("genes", "transcripts"))

makeGRangesFromGTF(file, level = c("genes", "transcripts"))
}
\arguments{
\item{file}{File path.}

\item{level}{Fetch annotations as "\code{genes}" or "\code{transcripts}".}
}
\value{
\code{data.frame}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
file <- "http://basejump.seq.cloud/mmusculus.gtf"

# Genes
x <- makeGRangesFromGFF(file, level = "genes")
head(names(x))
head(mcols(x))

# Transcripts
x <- makeGRangesFromGFF(file, level = "transcripts")
head(names(x))
head(mcols(x))
}
\seealso{
Other GFF Functions: \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{readGFF}}
}
