% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromGFF.R
\name{makeGRangesFromGFF}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\title{Genomic Ranges from GFF File}
\usage{
makeGRangesFromGFF(file, format = c("genes", "transcripts"))
}
\arguments{
\item{file}{\code{string}. File path.}

\item{format}{Output as genes or transcripts.}
}
\value{
\code{GRanges}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\details{
The UCSC website has detailed conventions on the GFF3 format, including
the metadata columns.
}
\examples{
file <- "http://basejump.seq.cloud/example.gtf"

# Genes
x <- makeGRangesFromGFF(file = file, format = "genes")
summary(x)
as.data.frame(x) \%>\% glimpse()

# Transcripts
x <- makeGRangesFromGFF(file = file, format = "transcripts")
summary(x)
as.data.frame(x) \%>\% glimpse()
}
\seealso{
\itemize{
\item \code{\link[rtracklayer:import]{rtracklayer::import()}}.
\item \code{\link[GenomicFeatures:makeTxDbFromGFF]{GenomicFeatures::makeTxDbFromGFF()}}.
}

Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
