% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromGFF.R
\name{makeGRangesFromGFF}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\title{Genomic Ranges from GFF File}
\usage{
makeGRangesFromGFF(file, format = c("genes", "transcripts"))
}
\arguments{
\item{file}{File path.}

\item{format}{Output as genes or transcripts.}
}
\value{
\code{GRanges}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
x <- makeGRangesFromGFF("http://basejump.seq.cloud/mmusculus.gtf")
summary(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
