% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Make Syntactically Valid Names}
\usage{
makeNames(names, unique = FALSE)
}
\arguments{
\item{names}{character vector to be coerced to syntactically valid
   names.  This is coerced to character if necessary.}

\item{unique}{logical; if \code{TRUE}, the resulting elements are
     unique.  This may be desired for, e.g., column names.}
}
\value{
Character vector.
}
\description{
Make Syntactically Valid Names
}
\note{
Sanitizes names using "\code{_}" instead of "\code{.}".
}
\seealso{
\code{\link[base:make.names]{base::make.names()}}.
}
\author{
Michael Steinbaugh
}
