% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R, R/makeNamesCamel.R,
%   R/makeNamesSnake.R
\name{makeNames}
\alias{makeNames}
\alias{makeNamesCamel}
\alias{camel}
\alias{makeNamesSnake}
\alias{snake}
\title{Make syntactically valid names out of character vectors}
\usage{
makeNames(character)

makeNamesCamel(character)

camel(character)

makeNamesSnake(character)

snake(character)
}
\arguments{
\item{character}{Character vector}
}
\value{
Character vector formatted with valid names
}
\description{
Make syntactically valid names out of character vectors
}
\examples{
makeNames("RNAi clone")
makeNamesCamel("RNAi clone")
makeNamesSnake("RNAi clone")
}
\author{
Michael Steinbaugh
}
