% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Make syntactically valid names out of character vectors}
\usage{
makeNames(character)
}
\arguments{
\item{character}{Character vector}
}
\value{
Character vector formatted in dot notation
}
\description{
Make syntactically valid names out of character vectors
}
\examples{
makeNames("RNAi clone")
}
\author{
Michael Steinbaugh
}
