% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Make Syntactically Valid Names}
\usage{
makeNames(names, unique = FALSE)
}
\arguments{
\item{names}{character vector to be coerced to syntactically valid
   names.  This is coerced to character if necessary.}

\item{unique}{logical; if \code{TRUE}, the resulting elements are
     unique.  This may be desired for, e.g., column names.}
}
\value{
Character vector.
}
\description{
Make Syntactically Valid Names
}
\details{
A syntactically valid name consists of letters, numbers and the dot or
  underline characters and starts with a letter or the dot not followed
  by a number.  Names such as \code{".2way"} are not valid, and neither
  are the \link{reserved} words.

  The definition of a \emph{letter} depends on the current locale, but
  only ASCII digits are considered to be digits.

  The character \code{"X"} is prepended if necessary.
  All invalid characters are translated to \code{"."}.  A missing value
  is translated to \code{"NA"}.  Names which match \R keywords have a dot
  appended to them.  Duplicated values are altered by
  \code{\link{make.unique}}.
}
\note{
Sanitizes names using "\code{_}" instead of "\code{.}".
}
\section{Warning}{

  Some OSes, notably FreeBSD, report extremely incorrect information about
  which characters are alphabetic in some locales (typically, all
  multi-byte locales including UTF-8 locales).  However, \R provides
  substitutes on Windows, macOS and AIX.

}

\seealso{
\code{\link[=make.names]{make.names()}}.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
