% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Make Syntactically Valid Names}
\usage{
makeNames(names, unique = FALSE)
}
\arguments{
\item{names}{\code{atomic} to be coerced to syntactically valid names. Will be
coerced to \code{character} if necessary.}

\item{unique}{\code{logical}; if \code{TRUE}, the resulting elements are unique.}
}
\value{
Character vector.
}
\description{
Make Syntactically Valid Names
}
\note{
Sanitizes names using underscore instead of dot.
}
\examples{
makeNames(c(123L, "hello world"))
}
\seealso{
\code{\link[base:make.names]{base::make.names()}}.

Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{fixNA}}, \code{\link{removeNA}},
  \code{\link{sanitizeRowData}},
  \code{\link{sanitizeSampleData}}
}
\author{
Michael Steinbaugh
}
