% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\alias{make_names}
\alias{makeNamesCamel}
\alias{make_names_camel}
\alias{camel}
\alias{makeNamesSnake}
\alias{make_names_snake}
\alias{snake}
\title{Make syntactically valid names out of character vectors.}
\usage{
makeNames(character)

makeNamesCamel(character)

makeNamesSnake(character)
}
\arguments{
\item{character}{Character vector.}
}
\value{
Character vector formatted with valid names.
}
\description{
Make syntactically valid names out of character vectors.
}
\examples{
makeNames("RNAi clone")
makeNamesCamel("RNAi clone")
makeNamesSnake("RNAi clone")
}
