% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/makeSingleCellExperiment-methods.R
\docType{methods}
\name{makeSingleCellExperiment}
\alias{makeSingleCellExperiment}
\alias{makeSingleCellExperiment,SimpleList-method}
\alias{makeSingleCellExperiment,list-method}
\title{Make a SingleCellExperiment object}
\usage{
makeSingleCellExperiment(assays, ...)

\S4method{makeSingleCellExperiment}{SimpleList}(assays,
  rowRanges = GRangesList(), colData = DataFrame(),
  metadata = list(), reducedDims = SimpleList(),
  transgeneNames = NULL, spikeNames = NULL)

\S4method{makeSingleCellExperiment}{list}(assays,
  rowRanges = GRangesList(), colData = DataFrame(),
  metadata = list(), reducedDims = SimpleList(),
  transgeneNames = NULL, spikeNames = NULL)
}
\value{
\code{SingleCellExperiment}.
}
\description{
This function is a utility wrapper for \code{SummarizedExperiment} that provides
automatic subsetting for row and column data, as well as automatic handling
of transgenes and spike-ins.
}
\note{
Updated 2019-08-05.
}
\section{Session information}{


This function improves upon the standard constructor by slotting useful
session information into the \code{metadata} slot by default:
\itemize{
\item \code{date}: Today's date, returned from \code{Sys.Date}.
\item \code{wd}: Working directory, returned from \code{getwd}.
\item \code{sessionInfo}: \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} return.
}

This behavior can be disabled by setting \code{sessionInfo = FALSE}.
}

\examples{
data(SingleCellExperiment, package = "acidtest")

## SimpleList ====
object <- SingleCellExperiment
assays <- assays(object)
rowRanges <- rowRanges(object)
colData <- colData(object)
metadata <- metadata(object)
reducedDims <- reducedDims(object)

x <- makeSingleCellExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    metadata = metadata,
    reducedDims = reducedDims
)
print(x)
}
\seealso{
\itemize{
\item \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment()}}.
\item \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment()}}.
\item \code{help("RangedSummarizedExperiment-class", "SummarizedExperiment")}.
\item \code{help("SummarizedExperiment-class", "SummarizedExperiment")}.
\item \code{help("SingleCellExperiment-class", "SingleCellExperiment")}.
}
}
