% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSingleCellExperiment.R
\name{makeSingleCellExperiment}
\alias{makeSingleCellExperiment}
\title{Make a \code{SingleCellExperiment} object}
\usage{
makeSingleCellExperiment(assays, rowRanges, colData, metadata,
  transgeneNames = NULL, spikeNames = NULL)
}
\arguments{
\item{assays}{\code{list}.
Count matrices, which must have matching dimensions. Counts can be passed
in as either a dense matrix (\code{matrix}) or sparse matrix (\code{sparseMatrix}).}

\item{rowRanges}{\code{GRanges}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{colData}{\code{DataFrame}.
Metadata describing the assay columns. For bulk RNA-seq, this data
describes the samples. For single-cell RNA-seq, this data describes the
cells.}

\item{metadata}{\code{list}.
Metadata.}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character}.
Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).}
}
\value{
\code{SingleCellExperiment}.
}
\description{
This function is a utility wrapper for \code{SummarizedExperiment} that provides
automatic subsetting for row and column data, as well as automatic handling
of transgenes and spike-ins. Additionally, it improves upon the standard
constructor by slotting useful session information into the \code{metadata} slot:
}
\details{
\itemize{
\item \code{date}: Today's date, returned from \code{Sys.Date}.
\item \code{wd}: Working directory, returned from \code{getwd}.
\item \code{sessionInfo}: \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} return.
}
}
\examples{
data(sce)
object <- sce
x <- makeSingleCellExperiment(
    assays = assays(object),
    rowRanges = rowRanges(object),
    colData = colData(object),
    metadata = metadata(object)
)
print(x)
}
\seealso{
\itemize{
\item \code{SummarizedExperiment::SummarizedExperiment()}.
\item \code{SingleCellExperiment::SingleCellExperiment()}.
\item \code{help("RangedSummarizedExperiment-class", "SummarizedExperiment")}.
\item \code{help("SummarizedExperiment-class", "SummarizedExperiment")}.
\item \code{help("SingleCellExperiment-class", "SingleCellExperiment")}.
}
}
