% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSingleCellExperiment.R
\name{makeSingleCellExperiment}
\alias{makeSingleCellExperiment}
\title{Make Single-Cell Experiment}
\usage{
makeSingleCellExperiment(assays, rowRanges, colData, metadata,
  transgeneNames = NULL, spikeNames = NULL)
}
\arguments{
\item{assays}{\code{list}. RNA-seq count matrices, which must have matching
dimensions. Counts can be passed in either a dense matrix (\code{matrix}) or
sparse matrix (\code{sparseMatrix}).}

\item{rowRanges}{\code{GRanges} or \code{NULL}. \emph{Recommended}. Metadata describing the
assay rows. If defined, must contain genomic ranges. Can be left \code{NULL} if
the genome is poorly annotated and/or ranges aren't available from
AnnotationHub.}

\item{colData}{\code{DataFrame} \code{data.frame}, or \code{matrix}. Metadata
describing the assay columns. For bulk RNA-seq, this data describes the
samples. For single-cell RNA-seq, this data describes the cells.}

\item{metadata}{\code{list} or \code{NULL}. Metadata.}

\item{transgeneNames}{\code{character} or \code{NULL}. Vector indicating which
\code{\link[=assay]{assay()}} rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character} or \code{NULL}. Vector indicating which \code{\link[=assay]{assay()}}
rows denote spike-in sequences (e.g. ERCCs).}
}
\value{
\code{SingleCellExperiment}.
}
\description{
This function is a utility wrapper for
\code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}} that provides automatic
subsetting for row and column data, as well as automatic handling of
transgenes and spike-ins. Additionally, it improves upon the standard
constructor by slotting useful session information into the \code{\link[=metadata]{metadata()}}
slot:
}
\details{
\itemize{
\item \code{date}: Today's date, returned from \code{\link[base:Sys.Date]{base::Sys.Date()}}.
\item \code{wd}: Working directory, returned from \code{\link[base:getwd]{base::getwd()}}.
\item \code{sessionInfo}: \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} return.
}
}
\examples{
object <- sce_small
assays <- assays(object)
rowRanges <- rowRanges(object)
colData <- colData(object)
metadata <- metadata(object)

x <- makeSingleCellExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    metadata = metadata
)
print(x)
}
\seealso{
Other SingleCellExperiment Functions: \code{\link{aggregateCellsToSamples}},
  \code{\link{cell2sample}},
  \code{\link{mapCellsToSamples}},
  \code{\link{subsetPerSample}},
  \code{\link{topCellsPerSample}}
}
\concept{SingleCellExperiment Functions}
