% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummarizedExperiment.R
\name{makeSummarizedExperiment}
\alias{makeSummarizedExperiment}
\title{Make Summarized Experiment}
\usage{
makeSummarizedExperiment(assays, rowRanges = NULL, colData = NULL,
  metadata = NULL, transgeneNames = NULL, spikeNames = NULL)
}
\arguments{
\item{assays}{\code{list}. RNA-seq count matrices, which must have matching
dimensions. Counts can be passed in either dense (\code{matrix}) or sparse
(\code{dgCMatrix}, \code{dgTMatrix}) format.}

\item{rowRanges}{\code{GRanges} or \code{NULL}. Metadata describing the
assay rows. If defined, must contain genomic ranges. Can be left \code{NULL} if
the genome is poorly annotated and/or ranges aren't available from
AnnotationHub.}

\item{colData}{\code{DataFrame} \code{data.frame}, or \code{matrix}. Metadata
describing the assay columns. For bulk RNA-seq, this data describes the
samples. For single-cell RNA-seq, this data describes the cells.}

\item{metadata}{\code{list} or \code{NULL}. Metadata.}

\item{transgeneNames}{\code{character} or \code{NULL}. Vector indicating which
\code{\link[=assay]{assay()}} rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character} or \code{NULL}. Vector indicating which \code{\link[=assay]{assay()}}
rows denote spike-in sequences (e.g. ERCCs).}
}
\value{
\code{RangedSummarizedExperiment}.
}
\description{
This function is a utility wrapper for
\code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}} that provides automatic
subsetting for row and column data, as well as automatic handling of
transgenes and spike-ins. Additionally, it improves upon the standard
constructor by slotting useful session information into the \code{\link[=metadata]{metadata()}}
slot:
}
\details{
\itemize{
\item \code{date}: Today's date, returned from \code{\link[base:Sys.Date]{base::Sys.Date()}}.
\item \code{wd}: Working directory, returned from \code{\link[base:getwd]{base::getwd()}}.
\item \code{utilsSessionInfo}: \code{\link[utils:sessionInfo]{utils::sessionInfo()}}.
\item \code{devtoolsSessionInfo}: \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}}.
}
}
\examples{
genes <- c(
    "EGFP",  # transgene
    "gene1",
    "gene2",
    "gene3"
)
samples <- c(
    "sample1",
    "sample2",
    "sample3",
    "sample4"
)

# Example matrix
mat <- matrix(
    seq(1L:16L),
    nrow = 4L,
    ncol = 4L,
    dimnames = list(genes, samples)
)

# Primary assay must be named "counts"
assays = list(counts = mat)

# rowRanges won't contain transgenes or spike-ins
rowRanges <- GRanges(
    seqnames = c("1", "1", "1"),
    ranges = IRanges(
        start = c(1L, 101L, 201L),
        end = c(100L, 200L, 300L)
    )
)
names(rowRanges) <- c("gene1", "gene2", "gene3")

colData <- data.frame(
    genotype = c(
        "wildtype",
        "wildtype",
        "knockout",
        "knockout"
    ),
    age = c(3L, 6L, 3L, 6L),
    row.names = samples
)

makeSummarizedExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    transgeneNames = "EGFP"
)
}
\seealso{
\itemize{
\item \code{help("RangedSummarizedExperiment-class", "SummarizedExperiment")}.
\item \code{help("SummarizedExperiment-class", "SummarizedExperiment")}.
\item \code{help("SingleCellExperiment-class", "SingleCellExperiment")}.
\item \code{SummarizedExperiment::SummarizedExperiment()}.
\item \code{SingleCellExperiment::SingleCellExperiment()}.
}

Other Data Functions: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}}, \code{\link{counts}},
  \code{\link{flatFiles}}, \code{\link{gene2symbol}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
