% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2Gene.R
\name{makeTx2Gene}
\alias{makeTx2Gene}
\alias{makeTx2GeneFromEnsembl}
\alias{makeTx2GeneFromEnsDb}
\alias{makeTx2GeneFromGFF}
\alias{makeTx2GeneFromGTF}
\title{Make a Tx2Gene object}
\usage{
makeTx2GeneFromEnsembl()

makeTx2GeneFromEnsDb(object)

makeTx2GeneFromGFF(file)
}
\arguments{
\item{object}{Object.}

\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{Tx2Gene}.
}
\description{
Make a Tx2Gene object
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
## makeTx2GeneFromEnsembl ====
x <- makeTx2GeneFromEnsembl(organism = "Homo sapiens")
print(x)

## makeTx2GeneFromEnsDb ====
x <- makeTx2GeneFromEnsDb("EnsDb.Hsapiens.v75")
print(x)

## makeTx2GeneFromGFF ====
## GTF
file <- file.path(basejumpCacheURL, "example.gtf")
x <- makeTx2GeneFromGFF(file)
print(x)

## GFF3
file <- file.path(basejumpCacheURL, "example.gff3")
x <- makeTx2GeneFromGFF(file)
print(x)
}
