% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2gene.R
\name{makeTx2gene}
\alias{makeTx2gene}
\alias{makeTx2geneFromEnsembl}
\alias{makeTx2geneFromGFF}
\alias{makeTx2geneFromGTF}
\title{Make Transcript-to-Gene Mappings}
\usage{
makeTx2geneFromEnsembl(organism, build = NULL, release = NULL)

makeTx2geneFromGFF(file)

makeTx2geneFromGTF(file)
}
\arguments{
\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{build}{\code{string} or \code{NULL}. Genome build assembly name (e.g.
"\code{GRCh38}"). If set \code{NULL}, defaults to the most recent build available.}

\item{release}{\code{scalar integer} or \code{NULL}. Release version (e.g. \code{90}). If
set \code{NULL}, defaults to the most recent release available.}

\item{file}{\code{string}. File path.}
}
\value{
\code{tx2gene}.
}
\description{
Make Transcript-to-Gene Mappings
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
# makeTx2geneFromEnsembl ====
x <- makeTx2geneFromEnsembl("Homo sapiens")
print(x)

# makeTx2geneFromGFF ====
# GTF
x <- makeTx2geneFromGFF("http://basejump.seq.cloud/example.gtf")
print(x)

# GFF3
x <- makeTx2geneFromGFF("http://basejump.seq.cloud/example.gff3")
print(x)
}
\seealso{
\link{makeGRanges}.

Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{eggnog}}, \code{\link{emptyRanges}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{geneSynonyms}}, \code{\link{hgnc2ensembl}},
  \code{\link{makeGRanges}}, \code{\link{makeGene2symbol}},
  \code{\link{mgi2ensembl}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
