% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2gene.R
\name{makeTx2gene}
\alias{makeTx2gene}
\alias{makeTx2geneFromEnsembl}
\alias{makeTx2geneFromGFF}
\alias{makeTx2geneFromGTF}
\title{Make Transcript-to-Gene Mappings}
\usage{
makeTx2geneFromEnsembl(...)

makeTx2geneFromGFF(file)
}
\arguments{
\item{...}{Passthrough to \code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}}.}

\item{file}{\code{string}. File path.}
}
\value{
\code{data.frame}.
}
\description{
Make Transcript-to-Gene Mappings
}
\examples{
# makeTx2geneFromEnsembl ====
x <- makeTx2geneFromEnsembl("Homo sapiens")
glimpse(x)

# makeTx2geneFromGFF ====
# GTF
x <- makeTx2geneFromGFF("http://basejump.seq.cloud/example.gtf")
glimpse(x)

# GFF3
x <- makeTx2geneFromGFF("http://basejump.seq.cloud/example.gff3")
glimpse(x)
}
\seealso{
\code{help(topic = "makeGRanges", package = "basejump")}.

Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
