% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2geneFromEnsembl.R
\name{makeTx2geneFromEnsembl}
\alias{makeTx2geneFromEnsembl}
\title{Transcript-to-Gene Mappings from Ensembl}
\usage{
makeTx2geneFromEnsembl(organism, genomeBuild = NULL, release = NULL)
}
\arguments{
\item{organism}{Default recommended usage is to provide the full latin
organism name as a string.}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}
}
\value{
\code{data.frame}
}
\description{
Transcript-to-Gene Mappings from Ensembl
}
