% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2geneFromEnsembl.R
\name{makeTx2geneFromEnsembl}
\alias{makeTx2geneFromEnsembl}
\title{Transcript-to-Gene Mappings from Ensembl}
\usage{
makeTx2geneFromEnsembl(...)
}
\arguments{
\item{...}{Passthrough to \code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}}.}
}
\value{
\code{data.frame}.
}
\description{
Transcript-to-Gene Mappings from Ensembl
}
\examples{
x <- makeTx2geneFromEnsembl("Homo sapiens")
glimpse(x)
}
\seealso{
Other Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
