% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2GeneFromFASTA.R
\name{makeTx2GeneFromFASTA}
\alias{makeTx2GeneFromFASTA}
\title{Make a Tx2Gene object from transcriptome FASTA}
\usage{
makeTx2GeneFromFASTA(
  file,
  source = c("ensembl", "gencode", "flybase", "wormbase")
)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{source}{\code{character(1)}.
FASTA file source:
\itemize{
\item \code{"ensembl"}: Ensembl.
\item \code{"gencode"}: GENCODE.
\item \code{"flybase"}: FlyBase.
\item \code{"wormbase"}: WormBase.
}

Assuming Ensembl transcriptome (i.e. cDNA) input by default.}
}
\value{
\code{Tx2Gene}.
}
\description{
Make a Tx2Gene object from transcriptome FASTA
}
\note{
RefSeq transcript FASTA (e.g. "GRCh38_latest_rna.fna.gz") doesn't
contain gene identifiers, and is not supported.

Updated 2019-11-06.
}
\examples{
## Ensembl ====
file <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-98",
    "fasta",
    "homo_sapiens",
    "cdna",
    "Homo_sapiens.GRCh38.cdna.all.fa.gz",
    protocol = "ftp"
)
## > makeTx2GeneFromFASTA(file, source = "ensembl")

## GENCODE ====
file <- pasteURL(
    "ftp.ebi.ac.uk",
    "pub",
    "databases",
    "gencode",
    "Gencode_human",
    "release_32",
    "gencode.v32.transcripts.fa.gz",
    protocol = "ftp"
)
## > makeTx2GeneFromFASTA(file, source = "gencode")

## FlyBase ====
file <- pasteURL(
    "ftp.flybase.net",
    "releases",
    "FB2019_05",
    "dmel_r6.30",
    "fasta",
    "dmel-all-transcript-r6.30.fasta.gz",
    protocol = "ftp"
)
## > makeTx2GeneFromFASTA(file, source = "flybase")

## WormBase ====
file <- pasteURL(
    "ftp.wormbase.org",
    "pub",
    "wormbase",
    "releases",
    "WS272",
    "species",
    "c_elegans",
    "PRJNA13758",
    "c_elegans.PRJNA13758.WS272.mRNA_transcripts.fa.gz",
    protocol = "ftp"
)
## > makeTx2GeneFromFASTA(file, source = "wormbase")
}
\seealso{
tx2gene importers defined in \href{https://koopa.acidgenomics.com/}{koopa}.
}
