% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCellsToSamples.R
\name{mapCellsToSamples}
\alias{mapCellsToSamples}
\title{Map Cells to Samples}
\usage{
mapCellsToSamples(cells, samples)
}
\arguments{
\item{cells}{\code{character}. Cell identifiers.}

\item{samples}{\code{character}. Sample identifiers.}
}
\value{
\code{factor}. Cells as the names and samples as the levels.
}
\description{
This function extracts \code{sampleID} from the \code{cellID} column using grep
matching.
}
\seealso{
Other SingleCellExperiment Functions: \code{\link{aggregateCellsToSamples}},
  \code{\link{cell2sample}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{subsetPerSample}},
  \code{\link{topCellsPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{SingleCellExperiment Functions}
