% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCellsToSamples.R
\name{mapCellsToSamples}
\alias{mapCellsToSamples}
\title{Map Cells to Samples}
\usage{
mapCellsToSamples(cells, samples)
}
\arguments{
\item{cells}{\code{character}. Cell identifiers.}

\item{samples}{\code{character}. Sample identifiers.}
}
\value{
\code{factor}. Cells as the names and samples as the levels.
}
\description{
This function extracts \code{sampleID} from the \code{cellID} column using grep
matching.
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{tpm}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
