% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mapGenes-methods.R
\docType{methods}
\name{mapGenesToIDs}
\alias{mapGenesToIDs}
\alias{mapGenesToRownames}
\alias{mapGenesToSymbols}
\alias{mapGenes}
\alias{mapGenesToRownames,SummarizedExperiment-method}
\alias{mapGenesToIDs,SummarizedExperiment-method}
\alias{mapGenesToSymbols,SummarizedExperiment-method}
\title{Map Genes}
\usage{
mapGenesToIDs(object, ...)

mapGenesToRownames(object, ...)

mapGenesToSymbols(object, ...)

\S4method{mapGenesToRownames}{SummarizedExperiment}(object, genes)

\S4method{mapGenesToIDs}{SummarizedExperiment}(object, genes)

\S4method{mapGenesToSymbols}{SummarizedExperiment}(object, genes)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Take a user-defined gene vector and dynamically map the input to either the
object rownames or the gene names (symbols). These functions are useful for
writing code that needs to handle either gene identifier or gene name input
dynamically (e.g. for single-cell RNA-seq marker analysis).
}
\section{Ambiguous gene names}{


Some genomes (e.g. Homo sapiens, Mus musculus) contain duplicated gene names
for multiple gene identifiers. Normally we handle these ambiguous gene names
by sanitizing them with \code{\link[base:make.names]{base::make.names()}}. If a user requests a gene name
that is duplicated, these functions will return a warning.
}

\examples{
object <- rse_small
head(rownames(object))

object_symbols <- convertGenesToSymbols(object)
head(rownames(object_symbols))

geneIDs <- head(rownames(object))
geneNames <- head(as.character(rowRanges(object)$geneName))

# Row names
mapGenesToRownames(object, genes = geneIDs)
mapGenesToRownames(object, genes = geneNames)
mapGenesToRownames(object_symbols, genes = geneIDs)
mapGenesToRownames(object_symbols, genes = geneNames)

# Gene identifiers
mapGenesToIDs(object, genes = geneIDs)
mapGenesToIDs(object, genes = geneNames)
mapGenesToIDs(object_symbols, genes = geneIDs)
mapGenesToIDs(object_symbols, genes = geneNames)

# Gene names (symbols)
mapGenesToSymbols(object, genes = geneIDs)
mapGenesToSymbols(object, genes = geneNames)
mapGenesToSymbols(object_symbols, genes = geneIDs)
mapGenesToSymbols(object_symbols, genes = geneNames)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{counts}}, \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{metrics}}, \code{\link{sampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
