% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mapGenes-methods.R
\docType{methods}
\name{mapGenesToIDs}
\alias{mapGenesToIDs}
\alias{mapGenesToRownames}
\alias{mapGenesToSymbols}
\alias{mapGenes}
\alias{mapGenesToRownames,gene2symbol-method}
\alias{mapGenesToRownames,SummarizedExperiment-method}
\alias{mapGenesToIDs,gene2symbol-method}
\alias{mapGenesToIDs,SummarizedExperiment-method}
\alias{mapGenesToSymbols,gene2symbol-method}
\alias{mapGenesToSymbols,SummarizedExperiment-method}
\title{Map Genes}
\usage{
mapGenesToIDs(object, ...)

mapGenesToRownames(object, ...)

mapGenesToSymbols(object, ...)

\S4method{mapGenesToRownames}{gene2symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToRownames}{SummarizedExperiment}(object, genes,
  strict = TRUE)

\S4method{mapGenesToIDs}{gene2symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToIDs}{SummarizedExperiment}(object, genes,
  strict = TRUE)

\S4method{mapGenesToSymbols}{gene2symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToSymbols}{SummarizedExperiment}(object, genes,
  strict = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers.}

\item{strict}{\code{boolean}. Require all genes to match. Recommended by default.
If set \code{FALSE}, instead will return a warning to the user, and subset the
genes vector to only include matches.}
}
\value{
\code{character}.
}
\description{
Take a user-defined gene vector and dynamically map the input to either the
object rownames or the gene names (symbols). These functions are useful for
writing code that needs to handle either gene identifier or gene name input
dynamically (e.g. for single-cell RNA-seq marker analysis).
}
\section{Ambiguous gene names}{


Some genomes (e.g. Homo sapiens, Mus musculus) contain duplicated gene names
for multiple gene identifiers. Normally we handle these ambiguous gene names
by sanitizing them with \code{\link[base:make.names]{base::make.names()}}. If a user requests a gene name
that is duplicated, these functions will return a warning.
}

\examples{
object <- rse_small
print(object)

rownames <- head(rownames(object))
print(rownames)

geneIDs <- head(rowData(object)[["geneID"]])
print(geneIDs)

geneNames <- head(as.character(rowData(object)[["geneName"]]))
print(geneNames)

# Row names
mapGenesToRownames(object, genes = rownames)
mapGenesToRownames(object, genes = geneIDs)
mapGenesToRownames(object, genes = geneNames)

# Gene identifiers
mapGenesToIDs(object, genes = rownames)
mapGenesToIDs(object, genes = geneIDs)
mapGenesToIDs(object, genes = geneNames)

# Gene names (symbols)
mapGenesToSymbols(object, genes = rownames)
mapGenesToSymbols(object, genes = geneIDs)
mapGenesToSymbols(object, genes = geneNames)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{meltCounts}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
