% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mapGenes-methods.R
\name{mapGenesToRownames}
\alias{mapGenesToRownames}
\alias{mapGenesToIDs}
\alias{mapGenesToSymbols}
\alias{mapGenes}
\alias{mapGenesToRownames,Gene2Symbol-method}
\alias{mapGenesToRownames,SummarizedExperiment-method}
\alias{mapGenesToIDs,Gene2Symbol-method}
\alias{mapGenesToIDs,SummarizedExperiment-method}
\alias{mapGenesToSymbols,Gene2Symbol-method}
\alias{mapGenesToSymbols,SummarizedExperiment-method}
\title{Map genes}
\usage{
mapGenesToRownames(object, ...)

mapGenesToIDs(object, ...)

mapGenesToSymbols(object, ...)

\S4method{mapGenesToRownames}{Gene2Symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToRownames}{SummarizedExperiment}(object, genes, strict = TRUE)

\S4method{mapGenesToIDs}{Gene2Symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToIDs}{SummarizedExperiment}(object, genes, strict = TRUE)

\S4method{mapGenesToSymbols}{Gene2Symbol}(object, genes, strict = TRUE)

\S4method{mapGenesToSymbols}{SummarizedExperiment}(object, genes, strict = TRUE)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{strict}{\code{logical(1)}.
Require all genes to match. Recommended by default. If set \code{FALSE}, instead
will return a warning to the user, and subset the genes vector to only
include matches.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Take a user-defined gene vector and dynamically map the input to either the
object rownames or the gene names (symbols). These functions are useful for
writing code that needs to handle either gene identifier or gene name input
dynamically (e.g. for single-cell RNA-seq marker analysis).
}
\note{
Updated 2021-02-02.
}
\section{Ambiguous gene names}{


Some genomes (e.g. Homo sapiens, Mus musculus) contain duplicated gene names
for multiple gene identifiers. Normally we handle these ambiguous gene names
by sanitizing them with \code{make.names}. If a user requests a gene name that
is duplicated, these functions will return a warning.
}

\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment
print(rse)

rownames <- head(rownames(rse))
print(rownames)

g2s <- Gene2Symbol(rse)
geneIds <- head(g2s[["geneId"]])
print(geneIds)

geneNames <- head(g2s[["geneName"]])
print(geneNames)

## Row names.
mapGenesToRownames(rse, genes = rownames)
mapGenesToRownames(rse, genes = geneIds)
mapGenesToRownames(rse, genes = geneNames)

## Gene identifiers.
mapGenesToIDs(rse, genes = rownames)
mapGenesToIDs(rse, genes = geneIds)
mapGenesToIDs(rse, genes = geneNames)

## Gene names (symbols).
mapGenesToSymbols(rse, genes = rownames)
mapGenesToSymbols(rse, genes = geneIds)
mapGenesToSymbols(rse, genes = geneNames)
}
