% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-methods.R
\docType{methods}
\name{markdown}
\alias{markdown}
\alias{markdown,SummarizedExperiment-method}
\title{Markdown}
\usage{
\S4method{markdown}{SummarizedExperiment}(object, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{headerLevel}{\code{scalar integer}. Markdown header level (1-7).}
}
\value{
Markdown output.
}
\description{
Markdown
}
\examples{
markdown(rse_small)
}
\author{
Michael Steinbaugh
}
\concept{Markdown Functions}
