% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/markdown-methods.R
\docType{methods}
\name{markdown}
\alias{markdown}
\alias{markdown,SummarizedExperiment-method}
\title{Markdown}
\usage{
markdown(object, ...)

\S4method{markdown}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Markdown output.
}
\description{
Markdown
}
\examples{
data(rse_small)
markdown(rse_small)
}
\author{
Michael Steinbaugh
}
\concept{Markdown Functions}
