% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R, R/zzz.R
\name{mdList}
\alias{mdList}
\alias{md_list}
\title{Markdown utilities}
\usage{
mdList(character, ordered = FALSE)
}
\arguments{
\item{character}{Character vector.}

\item{ordered}{Ordered (\code{TRUE}; \code{1.}) or unordered (\code{FALSE}; \code{-}) list in
Markdown format.}
}
\value{
Printed Markdown syntax.
}
\description{
Markdown list.
}
\examples{
mdList(c("milk", "eggs"))
mdList(c("milk", "eggs"), ordered = TRUE)
}
