% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/markdown-methods.R
\docType{methods}
\name{markdown}
\alias{markdown}
\alias{markdown,SummarizedExperiment-method}
\title{Markdown}
\usage{
markdown(object, ...)

\S4method{markdown}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Markdown output.
}
\description{
Markdown
}
\examples{
data(rse_small)
markdown(rse_small)
}
\seealso{
Other Markdown Functions: \code{\link{markdownHeader}},
  \code{\link{markdownLink}}, \code{\link{markdownList}},
  \code{\link{markdownPlotlist}},
  \code{\link{markdownTables}},
  \code{\link{prepareTemplate}}
}
\concept{Markdown Functions}
