% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aliases.R,
%   R/methods-markdownHeader.R
\docType{methods}
\name{markdownHeader}
\alias{markdownHeader}
\alias{mdHeader}
\alias{markdownHeader}
\alias{markdownHeader,character-method}
\title{Markdown Header}
\usage{
markdownHeader(object, ...)

mdHeader(...)

\S4method{markdownHeader}{character}(object, level = 2L, tabset = FALSE,
  asis = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{level}{Header level (1-7).}

\item{tabset}{Include tabset marker.}

\item{asis}{Set this to \code{TRUE} when using the function inside a loop or
inside an RMarkdown chunk with \code{results="asis"} enabled.}
}
\value{
Markdown formatted code as \link[knitr:asis_output]{knitr::asis_output} or
\code{\link[base:writeLines]{base::writeLines()}}.
}
\description{
Markdown Header
}
\examples{
markdownHeader("Header")
markdownHeader("Header", level = 4L)
markdownHeader("Header", tabset = TRUE)
markdownHeader("Header", asis = TRUE)
}
\seealso{
Other Report Utilities: \code{\link{kables}},
  \code{\link{markdownList}}
}
