% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownHeader.R
\name{markdownHeader}
\alias{markdownHeader}
\alias{mdHeader}
\title{Markdown Header}
\usage{
markdownHeader(string, level = 2L, tabset = FALSE, asis = FALSE)

mdHeader(string, level = 2L, tabset = FALSE, asis = FALSE)
}
\arguments{
\item{string}{Character string.}

\item{level}{Header level (1-7).}

\item{tabset}{Include tabset marker.}

\item{asis}{Set this to \code{TRUE} when using the function inside a loop or
inside an RMarkdown chunk with \code{results="asis"} enabled.}
}
\value{
Markdown formatted code as \link[knitr:asis_output]{knitr::asis_output} or
\code{\link[base:writeLines]{base::writeLines()}}.
}
\description{
Markdown Header
}
\examples{
markdownHeader("Header")
markdownHeader("Header", level = 4L)
markdownHeader("Header", tabset = TRUE)
markdownHeader("Header", asis = TRUE)
}
\seealso{
Other R Markdown Functions: \code{\link{kables}},
  \code{\link{markdownList}},
  \code{\link{markdownPlotlist}}
}
