% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownHeader.R
\name{markdownHeader}
\alias{markdownHeader}
\alias{mdHeader}
\title{Markdown Header}
\usage{
markdownHeader(object, level = 2L, tabset = FALSE, asis = FALSE)
}
\arguments{
\item{object}{\code{scalar}, which will be coerced to \code{character}.}

\item{level}{Header level (1-7).}

\item{tabset}{Include tabset marker.}

\item{asis}{Set this to \code{TRUE} when using the function inside a loop or
inside an RMarkdown chunk with '\code{results = "asis"}' enabled.}
}
\value{
\itemize{
\item "\code{asis = TRUE}": \code{\link[knitr:asis_output]{knitr::asis_output()}}.
\item "\code{asis = FALSE}": \code{\link[=writeLines]{writeLines()}}.
}
}
\description{
Markdown Header
}
\examples{
markdownHeader("Header")
markdownHeader("Header", level = 4L)
markdownHeader("Header", tabset = TRUE)
markdownHeader("Header", asis = TRUE)
}
\seealso{
Other R Markdown Functions: \code{\link{kables}},
  \code{\link{markdownList}},
  \code{\link{markdownPlotlist}}
}
\author{
Michael Steinbaugh
}
