% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownHeader.R, R/markdownLink.R
\name{markdownHeader}
\alias{markdownHeader}
\alias{mdHeader}
\alias{mdLink}
\title{Markdown Header}
\usage{
markdownHeader(text, level = 2L, tabset = FALSE, asis = FALSE)
}
\arguments{
\item{text}{\code{string}. Header text.}

\item{level}{\code{scalar integer}. Header level (1-7).}

\item{tabset}{\code{boolean}. Include tabset marker.}

\item{asis}{\code{boolean}. Set this to \code{TRUE} when using the function inside a
loop or inside an R Markdown chunk with '\code{results = "asis"}' enabled.}
}
\value{
\itemize{
\item "\code{asis = TRUE}": \code{\link[knitr:asis_output]{knitr::asis_output()}}.
\item "\code{asis = FALSE}": \code{\link[=writeLines]{writeLines()}}.
}
}
\description{
Markdown Header
}
\examples{
markdownHeader("Header")
markdownHeader("Header", level = 4L)
markdownHeader("Header", tabset = TRUE)
markdownHeader("Header", asis = TRUE)
}
\seealso{
\itemize{
\item \href{https://daringfireball.net/projects/markdown/syntax}{Markdown Syntax}.
}

Other R Markdown Functions: \code{\link{kables}},
  \code{\link{markdownLink}}, \code{\link{markdownList}},
  \code{\link{markdownPlotlist}},
  \code{\link{prepareTemplate}}
}
\author{
Michael Steinbaugh
}
\concept{R Markdown Functions}
