% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownHeader.R
\name{markdownHeader}
\alias{markdownHeader}
\alias{mdHeader}
\title{Markdown Header}
\usage{
markdownHeader(text, level = 2L, tabset = FALSE, asis = FALSE)

mdHeader(text, level = 2L, tabset = FALSE, asis = FALSE)
}
\arguments{
\item{text}{\code{string}. Header text.}

\item{level}{\code{scalar integer}. Header level (1-7).}

\item{tabset}{\code{boolean}. Include tabset marker.}

\item{asis}{\code{boolean}. Set this to \code{TRUE} when using the function inside a
loop or inside an R Markdown chunk with '\code{results = "asis"}' enabled.}
}
\value{
\itemize{
\item "\code{asis = TRUE}": \code{\link[knitr:asis_output]{knitr::asis_output()}}.
\item "\code{asis = FALSE}": \code{\link[=writeLines]{writeLines()}}.
}
}
\description{
Generate a Markdown header (levels 1-7) in any R Markdown code block. When
calling inside an \code{asis} chunk, set \code{asis = TRUE}.
}
\examples{
markdownHeader("Header", level = 2L)
markdownHeader("Header", tabset = TRUE)
markdownHeader("Header", asis = TRUE)
}
\seealso{
\href{https://daringfireball.net/projects/markdown/syntax}{Markdown Syntax}.

Other Markdown Functions: \code{\link{markdownLink}},
  \code{\link{markdownList}},
  \code{\link{markdownPlotlist}},
  \code{\link{markdownTables}}, \code{\link{markdown}},
  \code{\link{prepareTemplate}}
}
\author{
Michael Steinbaugh
}
\concept{Markdown Functions}
