% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownLink.R
\name{markdownLink}
\alias{markdownLink}
\alias{mdLink}
\title{Markdown hyperlink}
\usage{
markdownLink(text, url, title = NULL)
}
\value{
Markdown-formatted link.
}
\description{
For use in \code{asis} blocks only.
}
\note{
Updated 2019-08-05.
}
\examples{
markdownLink(
    text = "R",
    url = "https://www.r-project.org",
    title = "The R Project for Statistical Computing"
)
}
\seealso{
\href{https://daringfireball.net/projects/markdown/syntax}{Markdown Syntax}.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
