% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownLink.R
\name{markdownLink}
\alias{markdownLink}
\alias{mdLink}
\title{Markdown hyperlink}
\usage{
markdownLink(text, url, title = NULL)
}
\arguments{
\item{text}{\code{character(1)}.
Header text.}

\item{url}{\code{character(1)}.
URL.}

\item{title}{\code{character(1)} or \code{NULL}.
Link title attribute. This will appear in a mouse-over pop-up box.}
}
\value{
Markdown-formatted link.
}
\description{
For use in \code{asis} blocks only.
}
\examples{
markdownLink(
    text = "R",
    url = "https://www.r-project.org",
    title = "The R Project for Statistical Computing"
)
}
\seealso{
\href{https://daringfireball.net/projects/markdown/syntax}{Markdown Syntax}.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
