% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownLink.R
\name{markdownLink}
\alias{markdownLink}
\alias{mdLink}
\title{Markdown Hyperlink}
\usage{
markdownLink(text, url, title = NULL)

mdLink(text, url, title = NULL)
}
\arguments{
\item{text}{\code{string}. Header text.}

\item{url}{\code{string}. URL.}

\item{title}{\code{string} or \code{NULL}. Link title attribute. This will appear in a
mouse-over pop-up box.}
}
\value{
Markdown-formatted link.
}
\description{
For use in \code{asis} blocks only.
}
\examples{
markdownLink(
    text = "R",
    url = "https://www.r-project.org",
    title = "The R Project for Statistical Computing"
)
}
\seealso{
Other Markdown Functions: \code{\link{markdownHeader}},
  \code{\link{markdownList}},
  \code{\link{markdownPlotlist}},
  \code{\link{markdownTables}}, \code{\link{markdown}},
  \code{\link{prepareTemplate}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Markdown Functions}
