% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownLink.R
\name{markdownLink}
\alias{markdownLink}
\title{Markdown Link}
\usage{
markdownLink(text, url, title = NULL)
}
\arguments{
\item{text}{Link text.}

\item{url}{Link URL.}

\item{title}{\emph{Optional.} Link title attribute. This will appear in mouse-over
popup.}
}
\value{
Markdown-formatted link.
}
\description{
For use in 'asis' blocks.
}
\examples{
markdownLink(
    text = "R",
    url = "https://www.r-project.org",
    title = "The R Project for Statistical Computing"
)
}
\seealso{
\itemize{
\item \href{https://daringfireball.net/projects/markdown/syntax}{Markdown Syntax}.
}

Other R Markdown Functions: \code{\link{kables}},
  \code{\link{markdownHeader}}, \code{\link{markdownList}},
  \code{\link{markdownPlotlist}}
}
\author{
Rory Kirchner
}
