% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-markdownList.R
\docType{methods}
\name{markdownList}
\alias{markdownList}
\alias{markdownList}
\alias{markdownList,character-method}
\alias{mdList}
\title{Markdown List}
\usage{
markdownList(object, ...)

\S4method{markdownList}{character}(object, ordered = FALSE, asis = FALSE)

mdList(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{ordered}{Ordered (\code{TRUE}; \code{1.}) or unordered (\code{FALSE}; \code{-}) list in
Markdown format.}

\item{asis}{Set this to \code{TRUE} when using the function inside a loop or
inside an RMarkdown chunk with \code{results="asis"} enabled.}
}
\value{
Markdown formatted code as \link[knitr:asis_output]{knitr::asis_output} or
\code{\link[base:writeLines]{base::writeLines()}}.
}
\description{
Markdown List
}
\examples{
groceries <- c("milk", "eggs")
markdownList(groceries)
markdownList(groceries, ordered = TRUE)
markdownList(groceries, asis = TRUE)
}
\seealso{
Other R Markdown Functions: \code{\link{kables}},
  \code{\link{markdownHeader}},
  \code{\link{markdownPlotlist}}
}
