% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownList.R
\name{markdownList}
\alias{markdownList}
\alias{mdList}
\title{Markdown List}
\usage{
markdownList(object, ordered = FALSE, asis = FALSE)

mdList(object, ordered = FALSE, asis = FALSE)
}
\arguments{
\item{object}{\code{atomic}, which will be coerced to \code{character}.}

\item{ordered}{Ordered ("\code{TRUE}") or unordered ("\code{FALSE}") list.}

\item{asis}{Set this to \code{TRUE} when using the function inside a loop or
inside an RMarkdown chunk with \code{results="asis"} enabled.}
}
\value{
Markdown formatted code as \link[knitr:asis_output]{knitr::asis_output} or
\code{\link[base:writeLines]{base::writeLines()}}.
}
\description{
Markdown List
}
\examples{
groceries <- c("milk", "eggs")
markdownList(groceries)
markdownList(groceries, ordered = TRUE)
markdownList(groceries, asis = TRUE)
}
\seealso{
Other R Markdown Functions: \code{\link{kables}},
  \code{\link{markdownHeader}},
  \code{\link{markdownPlotlist}}
}
