% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlist.R
\name{markdownList}
\alias{markdownList}
\title{Markdown list}
\usage{
markdownList(character, ordered = FALSE)
}
\arguments{
\item{character}{\code{character} vector.}

\item{ordered}{Ordered \code{TRUE} or unordered \code{FALSE} list in Markdown format.}
}
\value{
\code{character} vector.
}
\description{
Markdown list
}
\examples{
markdownList(c("item1", "item2"), ordered = TRUE)
}

