% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-markdownPlotlist.R
\docType{methods}
\name{markdownPlotlist}
\alias{markdownPlotlist}
\alias{markdownPlotlist}
\alias{markdownPlotlist,list-method}
\alias{mdPlotlist}
\title{Markdown Plotlist}
\usage{
markdownPlotlist(object, ...)

\S4method{markdownPlotlist}{list}(object, headerLevel = 2L)

mdPlotlist(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{headerLevel}{Header level.}
}
\value{
Graphical output of plots, separated by Markdown headers.
}
\description{
Markdown Plotlist
}
\examples{
loadRemoteData("http://basejump.seq.cloud/plotlist.rda")
markdownPlotlist(plotlist)
}
\seealso{
\code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.
}
\author{
Michael Steinbaugh
}
