% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownPlotlist.R
\name{markdownPlotlist}
\alias{markdownPlotlist}
\alias{mdPlotlist}
\title{Markdown Plotlist}
\usage{
markdownPlotlist(plotlist, headerLevel = 2L)

mdPlotlist(plotlist, headerLevel = 2L)
}
\arguments{
\item{plotlist}{List containing \code{ggplot} objects.}

\item{headerLevel}{Header level.}
}
\value{
Graphical output of plots, separated by Markdown headers.
}
\description{
Markdown Plotlist
}
\examples{
loadRemoteData("http://basejump.seq.cloud/plotlist.rda")
markdownPlotlist(plotlist)
}
\seealso{
\code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.

Other R Markdown Functions: \code{\link{kables}},
  \code{\link{markdownHeader}}, \code{\link{markdownList}}
}
