% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aliases.R,
%   R/methods-markdownPlotlist.R
\docType{methods}
\name{markdownPlotlist}
\alias{markdownPlotlist}
\alias{mdPlotlist}
\alias{markdownPlotlist}
\alias{markdownPlotlist,list-method}
\title{Markdown Plotlist}
\usage{
markdownPlotlist(object, ...)

mdPlotlist(...)

\S4method{markdownPlotlist}{list}(object, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{headerLevel}{Header level.}
}
\description{
Markdown Plotlist
}
\examples{
loadRemoteData("http://basejump.seq.cloud/plotlist.rda")
markdownPlotlist(plotlist)
}
\seealso{
\code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.
}
\author{
Michael Steinbaugh
}
