% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownPlots.R
\name{markdownPlots}
\alias{markdownPlots}
\alias{mdPlots}
\title{Multiple Markdown plots}
\usage{
markdownPlots(list, headerLevel = 2L)
}
\arguments{
\item{list}{\code{list}.
Named list containing \code{ggplot} objects.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}
}
\value{
Graphical output of plots.
Separated by Markdown headers.
}
\description{
Supports using a named \code{list} containing multiple \code{ggplot} objects, which
can be used in an R Markdown report, separated by headers. Internally, the
headers are generated with the \code{markdownHeader} function.
}
\note{
Updated 2019-07-28.
}
\examples{
loadRemoteData(url = file.path(basejumpTestsURL, "plotlist.rds"))
names(plotlist)
markdownPlots(list = plotlist)
}
\seealso{
\code{cowplot::plot_grid()}.
}
