% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdownTables}
\alias{markdownTables}
\title{Create Multiple Tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
markdownTables(list, captions = NULL, force = FALSE)
}
\arguments{
\item{list}{Named \code{list}. Column data that can be coerced to \code{data.frame}.}

\item{captions}{\code{character} or \code{NULL}. Table captions. If \code{NULL}, the names
of the list will be used automatically as captions.}

\item{force}{\code{boolean}. Force knit output using \code{\link[knitr:asis_output]{knitr::asis_output()}}.
Recommended for development and unit testing only.}
}
\value{
\code{asis_output} if in a knit call or \code{list}.
}
\description{
Knit multiple tables in a single R Markdown chunk.
}
\examples{
list <- list(mpg = head(ggplot2::mpg), mtcars = head(datasets::mtcars))
captions <- c(
    mpg = "Miles per gallon",
    mtcars = "Motor Trend car road tests"
)
markdownTables(list = list, captions = captions)
}
\seealso{
\itemize{
\item \code{\link[knitr:kable]{knitr::kable()}}.
\item \href{https://stackoverflow.com/a/35149103/3911732}{Stack Overflow post}.
}
}
