% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownTables.R
\name{markdownTables}
\alias{markdownTables}
\title{Create Multiple Tables in LaTeX, HTML, Markdown and reStructuredText}
\usage{
markdownTables(list, captions, force = FALSE)
}
\arguments{
\item{list}{\code{list} containing column data (e.g. \code{data.frame}, \code{matrix}).}

\item{captions}{\code{character}. Table captions.}

\item{force}{\code{boolean}. Force knit output. Recommended for development and
unit testing only.}
}
\value{
\code{asis_output} if in a knit call or \code{list}.
}
\description{
Knit multiple tables in a single R Markdown chunk.
}
\examples{
markdownTables(
    list = list(head(ggplot2::mpg), head(datasets::mtcars)),
    captions = c("mpg", "mtcars")
)
}
\seealso{
\itemize{
\item \code{\link[knitr:kable]{knitr::kable()}}.
\item \href{https://stackoverflow.com/a/35149103/3911732}{Stack Overflow post}.
}

Other Markdown Functions: \code{\link{markdownHeader}},
  \code{\link{markdownLink}}, \code{\link{markdownList}},
  \code{\link{markdownPlotlist}}, \code{\link{markdown}},
  \code{\link{prepareTemplate}}
}
\concept{Markdown Functions}
