% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchArgsToDoCall.R
\name{matchArgsToDoCall}
\alias{matchArgsToDoCall}
\title{Match Arguments to Do Call}
\usage{
matchArgsToDoCall(args, removeArgs = NULL, which = sys.parent(),
  verbose = FALSE)
}
\arguments{
\item{args}{
    The default method expects a \emph{list} of arguments to the function
    call (the \code{names} attribute of \code{args} gives the argument names).
    See \code{?base::\link[base]{do.call}} for the details.
    
    Specific methods can support other objects.
    Please refer to the documentation of a particular method for the details.
  }

\item{removeArgs}{\code{character}. Names of arguments to remove from \code{call}
and/or \code{fun} returns before passing to \code{do.call()}.}

\item{which}{the frame number if non-negative, the number of frames
    to go back if negative.}

\item{verbose}{\code{boolean}. Run the function with verbose messages? It is only
recommended to enable this when debugging.}
}
\value{
\code{list}. Arguments to pass to \code{\link[=do.call]{do.call()}}.
}
\description{
Match Arguments to Do Call
}
\examples{
example <- function(object, xxx, ...) {
    args <- matchArgsToDoCall(
        args = list(
            object = object,
            collapse = " "
        ),
        removeArgs = "xxx"
    )
    print(args)
    do.call(what = paste, args = args)
}
example(c("hello", "world"))
}
\seealso{
Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}}, \code{\link{matchCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
