% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchArgsToDoCall.R
\name{matchArgsToDoCall}
\alias{matchArgsToDoCall}
\title{Match Arguments to Do Call}
\usage{
matchArgsToDoCall(args, removeFormals = NULL, n = 1L,
  verbose = FALSE)
}
\arguments{
\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{removeFormals}{\code{character}. Names of formal arguments to remove from
\code{args} list before passing to \code{do.call()}.}

\item{n}{the number of generations to go back.  (See the
    \sQuote{Details} section.)}

\item{verbose}{\code{boolean}. Run the function with verbose messages? It is only
recommended to enable this when debugging.}
}
\value{
\code{list}. Arguments to pass to \code{\link[=do.call]{do.call()}}.
}
\description{
Match Arguments to Do Call
}
\examples{
example <- function(object, xxx, ...) {
    args <- matchArgsToDoCall(
        args = list(
            object = object,
            collapse = " "
        ),
        removeFormals = "xxx"
    )
    print(args)
    do.call(what = paste, args = args)
}
example(c("hello", "world"))
}
\seealso{
\itemize{
\item \code{\link[base:do.call]{base::do.call()}}.
\item \code{\link[base:sys.call]{base::sys.call()}}.
\item \code{\link[base:sys.parent]{base::sys.parent()}}.
}

Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}, \code{\link{standardizeCall}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
