% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchArgsToDoCall.R
\name{matchArgsToDoCall}
\alias{matchArgsToDoCall}
\title{Match Arguments to Do Call}
\usage{
matchArgsToDoCall(args = NULL, removeFormals = NULL,
  which = sys.parent(n = 1L), verbose = FALSE)
}
\arguments{
\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{removeFormals}{\code{character}. Names of formal arguments to remove from
\code{args} list before passing to \code{do.call()}.}

\item{which}{the frame number if non-negative, the number of frames
    to go back if negative.}

\item{verbose}{\code{boolean}. Run the function with verbose messages? It is only
recommended to enable this when debugging.}
}
\value{
\code{list}. Arguments to pass to \code{\link[=do.call]{do.call()}}.
}
\description{
Match Arguments to Do Call
}
\examples{
example <- function(object, xxx, ...) {
    args <- matchArgsToDoCall(
        args = list(
            object = object,
            collapse = " "
        ),
        removeFormals = "xxx"
    )
    print(args)
    do.call(what = paste, args = args)
}
example(c("hello", "world"))

example <- function(object, xxx, ...) {
    do.call(
        what = paste,
        args = matchArgsToDoCall(
            args = list(collapse = " "),
            removeFormals = "xxx"
        )
    )
}
example(c("hello", "world"))
}
\seealso{
\itemize{
\item \link{standardizeCall}.
\item \link[base:do.call]{base::do.call}.
}
}
\author{
Michael Steinbaugh
}
