% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchesGene2Symbol.R
\name{matchesGene2Symbol}
\alias{matchesGene2Symbol}
\title{Check Gene-to-Symbol Value Matching}
\usage{
matchesGene2Symbol(x, genes, gene2symbol)
}
\arguments{
\item{x}{Object class supporting \code{rownames}. All rownames in this object
must intersect with the rownames defined in the \code{gene2symbol} argument.}

\item{genes}{\code{character}. Gene identifiers. It is considered better practice
to input the stable gene identifiers from Ensembl (e.g. "ENSG00000000003")
and not the (HGNC) gene symbols (e.g. "TSPN6"), if possible.}

\item{gene2symbol}{\code{Gene2Symbol}. Gene-to-symbol mappings. Must contain
\code{geneID} and \code{geneName} columns. See \code{Gene2Symbol()} for more information.}
}
\value{
No value.
}
\description{
Check Gene-to-Symbol Value Matching
}
\examples{
x <- DataFrame(
    "sample1" = c(1L, 2L),
    "sample2" = c(3L, 4L),
    row.names = c("gene1", "gene2")
)
print(x)

g2s <- Gene2Symbol(
    object = DataFrame(
        geneID = c("ENSG00000000003", "ENSG00000000005"),
        geneName = c("TSPAN6", "TNMD"),
        row.names = rownames(x)
    )
)
print(g2s)

geneIDs <- g2s[["geneID"]]
print(geneIDs)

geneNames <- g2s[["geneName"]]
print(geneNames)

matchesGene2Symbol(x = x, genes = geneIDs, gene2symbol = g2s)
matchesGene2Symbol(x = x, genes = geneNames, gene2symbol = g2s)
}
