% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchesGene2Symbol.R
\name{matchesGene2Symbol}
\alias{matchesGene2Symbol}
\title{Check that user-defined gene input matches expected values}
\usage{
matchesGene2Symbol(x, genes, gene2symbol, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneId} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Check that user-defined gene input matches expected values
}
\note{
Updated 2020-10-05.
}
\examples{
x <- S4Vectors::DataFrame(
    "sample1" = c(1L, 2L),
    "sample2" = c(3L, 4L),
    row.names = c("gene1", "gene2")
)
print(x)

g2s <- Gene2Symbol(
    object = S4Vectors::DataFrame(
        "geneId" = c("ENSG00000000003", "ENSG00000000005"),
        "geneName" = c("TSPAN6", "TNMD"),
        row.names = rownames(x)
    )
)
print(g2s)

geneIds <- g2s[["geneId"]]
print(geneIds)

geneNames <- g2s[["geneName"]]
print(geneNames)

matchesGene2Symbol(x = x, genes = geneIds, gene2symbol = g2s)
matchesGene2Symbol(x = x, genes = geneNames, gene2symbol = g2s)
}
