% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchesInterestingGroups.R
\name{matchesInterestingGroups}
\alias{matchesInterestingGroups}
\title{Check that interesting groups match a defined value}
\usage{
matchesInterestingGroups(x, interestingGroups,
  .xname = getNameInParent(x))
}
\arguments{
\item{x}{S4 class.}

\item{interestingGroups}{\code{character}.
Interesting groups.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
No value.
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
data(rse, package = "acidData")

## Checks that columns are defined in `colData()`.
matchesInterestingGroups(rse, "condition")

## Currently allowing `NULL` to pass.
matchesInterestingGroups(rse, NULL)
}
