% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchesInterestingGroups.R
\name{matchesInterestingGroups}
\alias{matchesInterestingGroups}
\title{Check that interesting groups match a defined value}
\usage{
matchesInterestingGroups(x, interestingGroups,
  .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\note{
Updated 2019-07-28.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## Checks that columns are defined in `sampleData()`.
matchesInterestingGroups(rse, "condition")

## Currently allowing `NULL` to pass.
matchesInterestingGroups(rse, NULL)
}
