% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchHumanOrthologs.R
\name{matchHumanOrthologs}
\alias{matchHumanOrthologs}
\title{Match human gene orthologs}
\usage{
matchHumanOrthologs(genes, organism = NULL, ensemblRelease = NULL)
}
\arguments{
\item{genes}{Ensembl gene identifiers.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{ensemblRelease}{Ensembl release (e.g. 96).}
}
\value{
\code{DataFrame}.
Data frame containing mapping columns:
\itemize{
\item geneID
\item hgncID
\item geneName
\item hgncName
}
}
\description{
Match human gene orthologs
}
\examples{
genes <- c(
    "ENSMUSG00000000001", "ENSMUSG00000000003",
    "ENSMUSG00000000028", "ENSMUSG00000000031",
    "ENSMUSG00000000037", "ENSMUSG00000000049"
)
matchHumanOrthologs(genes, ensemblRelease = 87L)
}
