% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchHumanOrthologs.R
\name{matchHumanOrthologs}
\alias{matchHumanOrthologs}
\title{Match human gene orthologs}
\usage{
matchHumanOrthologs(genes, organism = NULL, ensemblRelease = NULL)
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{ensemblRelease}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}
}
\value{
\code{DataFrame}.
Data frame containing mapping columns:
\itemize{
\item geneID
\item hgncID
\item geneName
\item hgncName
}
}
\description{
Match human gene orthologs
}
\note{
Updated 2019-08-18.
}
\examples{
genes <- c(
    "ENSMUSG00000000001", "ENSMUSG00000000003",
    "ENSMUSG00000000028", "ENSMUSG00000000031",
    "ENSMUSG00000000037", "ENSMUSG00000000049"
)
## Protect against Ensembl timeouts causing build checks to fail.
goalie::hasInternet("https://useast.ensembl.org") {
    matchHumanOrthologs(genes = genes, ensemblRelease = 87L)
}
}
