% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchInterestingGroups.R
\name{matchInterestingGroups}
\alias{matchInterestingGroups}
\title{Match Interesting Groups}
\usage{
matchInterestingGroups(object, ...)
}
\value{
\code{character}.
}
\description{
Match Interesting Groups
}
\examples{
matchInterestingGroups(rse_bcb)
matchInterestingGroups(rse_bcb, interestingGroups = "sampleName")
}
\author{
Michael Steinbaugh
}
\keyword{internal}
