% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchInterestingGroups.R
\name{matchInterestingGroups}
\alias{matchInterestingGroups}
\title{Match Interesting Groups}
\usage{
matchInterestingGroups(object, interestingGroups, ...)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Match Interesting Groups
}
\examples{
matchInterestingGroups(rse_small)
matchInterestingGroups(rse_small, interestingGroups = NULL)
matchInterestingGroups(rse_small, interestingGroups = "sampleName")
}
\seealso{
Other Developer Functions: \code{\link{calls}},
  \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
