% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchInterestingGroups.R
\name{matchInterestingGroups}
\alias{matchInterestingGroups}
\title{Match Interesting Groups}
\usage{
matchInterestingGroups(object, interestingGroups)
}
\arguments{
\item{object}{\code{SummarizedExperiment}.}

\item{interestingGroups}{\code{character}. Groups of interest.}
}
\value{
\code{character}. Interesting groups.
}
\description{
Match the user-defined interesting groups to the values slotted within a
\code{SummarizedExperiment} object, and check that they are valid. Otherwise
supports a missing \code{interestingGroups} argument, which will then use the
\code{\link[=interestingGroups]{interestingGroups()}} accessor function internally.
}
\examples{
matchInterestingGroups(rse_small)
matchInterestingGroups(rse_small, interestingGroups = NULL)
matchInterestingGroups(rse_small, interestingGroups = "sampleName")
}
\seealso{
Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}}, \code{\link{matchCall}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
