% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchInterestingGroups.R
\name{matchInterestingGroups}
\alias{matchInterestingGroups}
\title{Match Interesting Groups}
\usage{
matchInterestingGroups(object, interestingGroups, ...)
}
\value{
\code{character}.
}
\description{
Match Interesting Groups
}
\examples{
matchInterestingGroups(rse_bcb)
matchInterestingGroups(rse_bcb, interestingGroups = NULL)
matchInterestingGroups(rse_bcb, interestingGroups = "sampleName")
}
\seealso{
Other Developer Functions: \code{\link{detectHPC}},
  \code{\link{dots}}, \code{\link{localOrRemoteFile}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
