% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchInterestingGroups.R
\name{matchInterestingGroups}
\alias{matchInterestingGroups}
\title{Match interesting groups}
\usage{
matchInterestingGroups(object, interestingGroups = NULL)
}
\arguments{
\item{object}{\code{SummarizedExperiment}.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}
}
\value{
\code{character}.
Interesting groups.
}
\description{
Match the user-defined interesting groups to the values slotted within a
\code{SummarizedExperiment} object, and check that they are valid. Otherwise
supports a missing \code{interestingGroups} argument, which will then use the
\code{interestingGroups} accessor function internally.
}
\note{
Updated 2019-07-28.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment
matchInterestingGroups(rse)
matchInterestingGroups(rse, interestingGroups = NULL)
matchInterestingGroups(rse, interestingGroups = substitute())
}
