% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchS4Call.R
\name{matchS4Call}
\alias{matchS4Call}
\title{Functions to Access the Function Call Stack}
\usage{
matchS4Call(verbose = FALSE)
}
\arguments{
\item{verbose}{\code{boolean}. Run the function with verbose messages? It is only
recommended to enable this when debugging.}
}
\value{
\code{call}.
}
\description{
Note that \code{\link[base:match.call]{base::match.call()}} doesn't always work correctly inside S4
methods. Here we are using a combination of \code{\link[base:sys.call]{base::sys.call()}} with
\code{\link[pryr:standardise_call]{pryr::standardise_call()}} to correctly capture named arguments inside an S4
method. The function raverses up the call stack and find the first
\code{.local()}. The desired S4 call to match is one level above.
}
\examples{
setGeneric(
    name = "testing",
    def = function(object, ...) {
        standardGeneric("testing")
    }
)
setMethod(
    f = "testing",
    signature = signature("character"),
    definition = function(object, ...) {
        call <- matchS4Call()
        call
    }
)
setMethod(
    f = "testing",
    signature = signature("numeric"),
    definition = function(object, xxx, ...) {
        call <- matchS4Call()
        call
    }
)
testing("abc")
testing(123L)
}
\seealso{
\itemize{
\item \code{\link[base:sys.calls]{base::sys.calls()}}.
\item \code{\link[base:sys.call]{base::sys.call()}}.
\item \code{\link[base:match.call]{base::match.call()}}.
\item \code{\link[pryr:standardise_call]{pryr::standardise_call()}}.
}

Other Developer Functions: \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}, \code{\link{setArgsToDoCall}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
