% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mcolnames-methods.R
\docType{methods}
\name{mcolnames}
\alias{mcolnames}
\alias{mcolnames<-}
\alias{mcolnames,Vector-method}
\alias{mcolnames<-,Vector,character-method}
\title{Metadata column names}
\usage{
\S4method{mcolnames}{Vector}(x)

\S4method{mcolnames}{Vector,character}(x) <- value
}
\arguments{
\item{x}{Object.}

\item{value}{Value.}
}
\description{
Metadata column names
}
\examples{
data(rse, package = "acidtest")
mcolnames(rse)

## Asignment method.
mcolnames(rse) <- toupper(mcolnames(rse))
mcolnames(rse)
}
\seealso{
\code{S4Vectors::mcols()}.
}
