% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcolnames.R
\name{mcolnames}
\alias{mcolnames}
\alias{mcolnames<-}
\title{Metadata column names}
\usage{
mcolnames(object)

mcolnames(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\description{
Metadata column names
}
\examples{
data(rse, package = "acidData")
mcolnames(rse)

## Asignment method.
mcolnames(rse) <- snake(mcolnames(rse))
mcolnames(rse)
}
\seealso{
\itemize{
\item \code{\link[S4Vectors:mcols]{S4Vectors::mcols()}}.
}
}
