% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-mdHeader.R
\docType{methods}
\name{mdHeader}
\alias{mdHeader}
\alias{mdHeader,character-method}
\title{Markdown Header}
\usage{
mdHeader(object, ...)

\S4method{mdHeader}{character}(object, level = 2L, tabset = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{level}{Header level (1-7).}

\item{tabset}{Include tabset marker.}
}
\value{
\code{\link[=writeLines]{writeLines()}}.
}
\description{
Markdown Header
}
\examples{
mdHeader("Header")
mdHeader("Header", level = 4L)
mdHeader("Header", tabset = TRUE)
}
\seealso{
Other Report Utilities: \code{\link{kables}},
  \code{\link{mdList}}
}
\concept{Report Utilities}
