% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{mdList}
\alias{mdList}
\title{Markdown List}
\usage{
mdList(vec, ordered = FALSE)
}
\arguments{
\item{vec}{Character vector.}

\item{ordered}{Ordered (\code{TRUE}; \code{1.}) or unordered (\code{FALSE}; \code{-}) list in
Markdown format.}
}
\value{
\code{\link[=writeLines]{writeLines()}}.
}
\description{
Markdown List
}
\examples{
mdList(c("milk", "eggs"))
mdList(c("milk", "eggs"), ordered = TRUE)
}
\seealso{
Other Markdown Utilities: \code{\link{mdHeader}}
}
