% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-mdList.R
\docType{methods}
\name{mdList}
\alias{mdList}
\alias{mdList,character-method}
\title{Markdown List}
\usage{
mdList(object, ...)

\S4method{mdList}{character}(object, ordered = FALSE, asis = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{ordered}{Ordered (\code{TRUE}; \code{1.}) or unordered (\code{FALSE}; \code{-}) list in
Markdown format.}

\item{asis}{Set this to \code{TRUE} when using the function inside a loop or
inside an RMarkdown chunk with \code{results="asis"} enabled.}
}
\value{
Markdown formatted code as \link[knitr:asis_output]{knitr::asis_output} or
\code{\link[base:writeLines]{base::writeLines()}}.
}
\description{
Markdown List
}
\examples{
groceries <- c("milk", "eggs")
mdList(groceries)
mdList(groceries, ordered = TRUE)
mdList(groceries, asis = TRUE)
}
\seealso{
Other Report Utilities: \code{\link{kables}},
  \code{\link{mdHeader}}
}
\concept{Report Utilities}
