% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R, R/zzz.R
\name{mdList}
\alias{mdList}
\alias{md_list}
\title{Markdown list}
\usage{
mdList(vec, ordered = FALSE)
}
\arguments{
\item{vec}{Character vector.}

\item{ordered}{Ordered (\code{TRUE}; \code{1.}) or unordered (\code{FALSE}; \code{-}) list in
Markdown format.}
}
\description{
Markdown list
}
\examples{
mdList(c("milk", "eggs"))
mdList(c("milk", "eggs"), ordered = TRUE)
}
