% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdList.R
\name{mdList}
\alias{mdList}
\title{Markdown list}
\usage{
mdList(character, ordered = FALSE)
}
\arguments{
\item{character}{\code{character} vector}

\item{ordered}{Ordered \code{TRUE} or unordered \code{FALSE} list in Markdown format}
}
\value{
\code{character} vector
}
\description{
Markdown list
}
\examples{
mdList(c("milk", "eggs"))
mdList(c("milk", "eggs"), ordered = TRUE)
}
\author{
Michael Steinbaugh
}
