% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-mdList.R
\docType{methods}
\name{mdList}
\alias{mdList}
\alias{mdList,character-method}
\title{Markdown List}
\usage{
mdList(object, ...)

\S4method{mdList}{character}(object, ordered = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{ordered}{Ordered (\code{TRUE}; \code{1.}) or unordered (\code{FALSE}; \code{-}) list in
Markdown format.}
}
\value{
\code{\link[=writeLines]{writeLines()}}.
}
\description{
Markdown List
}
\examples{
mdList(c("milk", "eggs"))
mdList(c("milk", "eggs"), ordered = TRUE)
}
\seealso{
Other Report Utilities: \code{\link{kables}},
  \code{\link{mdHeader}}
}
