% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-mdPlotlist.R
\docType{methods}
\name{mdPlotlist}
\alias{mdPlotlist}
\alias{mdPlotlist}
\alias{mdPlotlist,list-method}
\title{Markdown Plotlist}
\usage{
mdPlotlist(object, ...)

\S4method{mdPlotlist}{list}(object, headerLevel)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{headerLevel}{Header level.}
}
\description{
Markdown Plotlist
}
\author{
Michael Steinbaugh
}
