% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/meltCounts-methods.R
\docType{methods}
\name{meltCounts}
\alias{meltCounts}
\alias{meltCounts,SummarizedExperiment-method}
\title{Melt Count Matrix to Long Format}
\usage{
meltCounts(object, ...)

\S4method{meltCounts}{SummarizedExperiment}(object, assay = 1L,
  nonzeroGenes = FALSE, trans = c("identity", "log2", "log10"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in \code{assays}. When passing in a
string, the name must be defined in \code{assayNames}.}

\item{nonzeroGenes}{\code{logical(1)}.
Return only non-zero genes.}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}
}
\value{
\code{grouped_df}.
Grouped by \code{colname} (e.g. sample ID) and \code{rowname} (e.g. gene ID).
}
\description{
Melt Count Matrix to Long Format
}
\examples{
data(rse)
x <- meltCounts(rse)
print(x)
}
\seealso{
\code{reshape2::melt}.
}
