% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/meltCounts-methods.R
\docType{methods}
\name{meltCounts}
\alias{meltCounts}
\alias{meltCounts,SummarizedExperiment-method}
\title{Melt Counts Matrix to Long Format}
\usage{
meltCounts(object, ...)

\S4method{meltCounts}{SummarizedExperiment}(object,
  log2Transform = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{log2Transform}{\code{boolean}. Apply \code{log2(x + 1L)} transformation to the
counts matrix before melt?}
}
\value{
\code{grouped_df}, grouped by \code{colname} (e.g. sample ID) and \code{rowname}
(e.g. gene ID).
}
\description{
Melt Counts Matrix to Long Format
}
\examples{
x <- meltCounts(rse_small)
print(x)
}
\seealso{
\code{\link[reshape2:melt]{reshape2::melt()}}.

Other Data Functions: \code{\link{counts}},
  \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
