% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/meltCounts-methods.R
\docType{methods}
\name{meltCounts}
\alias{meltCounts}
\alias{meltCounts,SummarizedExperiment-method}
\title{Melt Counts Matrix to Long Format}
\usage{
meltCounts(object, ...)

\S4method{meltCounts}{SummarizedExperiment}(object, assay = 1L,
  nonzeroGenes = FALSE, trans = c("identity", "log2", "log10"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{nonzeroGenes}{\code{boolean}. Return only non-zero genes.}

\item{trans}{\code{string}. Apply a log transformation (e.g. \code{log2(x + 1L)}) to
the counts matrix prior to melting, if desired. Use \code{"identity"} to return
unmodified (default).}
}
\value{
\code{grouped_df}, grouped by \code{colname} (e.g. sample ID) and \code{rowname}
(e.g. gene ID).
}
\description{
Melt Counts Matrix to Long Format
}
\examples{
x <- meltCounts(rse_small)
print(x)
}
\seealso{
\code{\link[reshape2:melt]{reshape2::melt()}}.

Other Data Functions: \code{\link{aggregate}},
  \code{\link{combine}}, \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{metrics}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
