% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltLog10.R
\name{meltLog10}
\alias{meltLog10}
\title{Melt RNA-Seq count data to long format and log10 transform}
\usage{
meltLog10(rawCounts, metadata = NULL)
}
\arguments{
\item{rawCounts}{Raw counts `matrix`}

\item{metadata}{Metadata `data.frame`}
}
\value{
log10 melted `data.frame`
}
\description{
Melt RNA-Seq count data to long format and log10 transform
}
