% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{metadataBlacklist}
\alias{metadataBlacklist}
\title{Sample metadata blacklist}
\format{
An object of class \code{character} of length 17.
}
\usage{
metadataBlacklist
}
\description{
Sample metadata blacklist
}
\note{
\code{sampleID} is set automatically for multiplexed/cell-level data.

Updated 2019-09-05.
}
\examples{
metadataBlacklist
}
\keyword{datasets}
