% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodFormals.R
\name{methodFormals}
\alias{methodFormals}
\title{Access to the Formal Arguments of an S4 Method}
\usage{
methodFormals(f, signature)
}
\arguments{
\item{f}{Generic \code{function} or \code{string} referencing a generic.}

\item{signature}{The signature of classes to match to the \code{f} argument.}
}
\value{
\code{list} of formal arguments.
}
\description{
Access to the Formal Arguments of an S4 Method
}
\examples{
x <- methodFormals("geometricMean", "numeric")
class(x)
names(x)
}
\seealso{
Modified version of \href{https://goo.gl/ymX571}{John Chambers' code}.

Other Developer Functions: \code{\link{calls}},
  \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{saveFunctionExamples}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
