% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodFormals.R
\name{methodFormals}
\alias{methodFormals}
\title{Access to the Formal Arguments of an S4 Method}
\usage{
methodFormals(f, signature)
}
\arguments{
\item{f}{a generic function or the character-string name of one.}

\item{signature}{the signature of classes to match to the arguments
    of \code{f}.  See the details below.}
}
\value{
\code{list}. Formal arguments.
}
\description{
Access to the Formal Arguments of an S4 Method
}
\examples{
x <- methodFormals("geometricMean", "numeric")
class(x)
names(x)
}
\seealso{
Modified version of \href{https://goo.gl/ymX571}{John Chambers' code}.

Other Developer Functions: \code{\link{detectHPC}},
  \code{\link{dots}}, \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}}, \code{\link{matchCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
