% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metricsPerSample}
\alias{metrics,SummarizedExperiment-method}
\alias{metrics,SingleCellExperiment-method}
\alias{metricsPerSample,SingleCellExperiment-method}
\title{Metrics}
\usage{
metrics(object, ...)

metricsPerSample(object, ...)

\S4method{metrics}{SummarizedExperiment}(object, return = c("tibble",
  "DataFrame"))

\S4method{metrics}{SingleCellExperiment}(object, return = c("tibble",
  "DataFrame"))

\S4method{metricsPerSample}{SingleCellExperiment}(object, fun = c("mean",
  "median", "sum"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{return}{\code{character(1)}. Return type. Uses \code{match.arg} internally and
defaults to the first argument in the \code{character} vector.}

\item{fun}{\code{character(1)}. Mathematical function name to apply. Uses \code{match.arg}.}
}
\value{
\itemize{
\item \code{"tibble"}: \code{grouped_df}. Grouped by \code{sampleID} column.
\item \code{"DataFrame"}: \code{DataFrame}. Rownames are identical to the \code{colnames}, of
the object, like \code{colData}.
}
}
\description{
This function takes data stored in \code{colData} and consistently returns a
tibble grouped by sample by default (\code{sampleID}).
}
\details{
\code{metrics} always returns \code{sampleName} and \code{interestingGroups} columns, even
when these columns are not defined in \code{colData}. This is designed to
integrate with plotting functions that use ggplot2 internally.
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: Metrics are sample level. \code{sampleID} column corresponds
to \code{colnames}.

\item \code{SingleCellExperiment}: Metrics are cell level. \code{cellID} column corresponds to
\code{colnames}. Tibble is returned grouped by sample (\code{sampleID} column).
}}

\examples{
data(rse, sce)

## SummarizedExperiment ====
x <- metrics(rse)
print(x)

## SingleCellExperiment ====
x <- metrics(sce)
print(x)

x <- metricsPerSample(sce, fun = "mean")
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
