% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metricsPerSample}
\alias{metrics,SummarizedExperiment-method}
\alias{metrics,SingleCellExperiment-method}
\alias{metricsPerSample,SingleCellExperiment-method}
\title{Metrics}
\usage{
metrics(object, ...)

metricsPerSample(object, ...)

\S4method{metrics}{SummarizedExperiment}(object, return = c("tibble",
  "DataFrame"))

\S4method{metrics}{SingleCellExperiment}(object, return = c("tibble",
  "DataFrame"))

\S4method{metricsPerSample}{SingleCellExperiment}(object, fun = c("mean",
  "median", "sum"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{return}{\code{string}. Return type. Uses \code{base::match.arg()}
internally and defaults to the first argument in the \code{character} vector.}

\item{fun}{\code{string}. Mathematical function name to apply.
Uses \code{\link[base:match.arg]{base::match.arg()}}.}
}
\value{
\code{DataFrame}.

\itemize{
\item \code{"tibble"}: \code{grouped_df}. Grouped by \code{sampleID} column.
\item \code{"DataFrame"}: \code{DataFrame}. Rownames are identical to \code{\link[base:colnames]{base::colnames()}},
of the object, like \code{\link[SummarizedExperiment:colData]{SummarizedExperiment::colData()}}.
}
}
\description{
This function takes data stored in \code{SummarizedExperiment::colData()} and
consistently returns a tibble grouped by sample by default (\code{sampleID}).
}
\details{
\code{\link[=metrics]{metrics()}} always returns \code{sampleName} and \code{interestingGroups} columns, even
when these columns are not defined in \code{colData}. This is designed to
integrate with plotting functions that use ggplot2 internally.
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: Metrics are sample level. \code{sampleID} column corresponds
to \code{\link[base:colnames]{base::colnames()}}.

\item \code{SingleCellExperiment}: Metrics are cell level. \code{cellID} column corresponds to
\code{\link[base:colnames]{base::colnames()}}. Tibble is returned grouped by sample
(\code{sampleID} column).
}}

\examples{
data(rse, sce)

## SummarizedExperiment ====
x <- metrics(rse)
print(x)

## SingleCellExperiment ====
x <- metrics(sce)
print(x)

x <- metricsPerSample(sce, fun = "mean")
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
