% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metricsPerSample}
\alias{metrics,SummarizedExperiment-method}
\alias{metrics,SingleCellExperiment-method}
\alias{metricsPerSample,SingleCellExperiment-method}
\title{Quality control metrics}
\usage{
metrics(object, ...)

metricsPerSample(object, ...)

\S4method{metrics}{SummarizedExperiment}(object, return = c("tbl_df",
  "DataFrame"))

\S4method{metrics}{SingleCellExperiment}(object, return = c("tbl_df",
  "DataFrame"))

\S4method{metricsPerSample}{SingleCellExperiment}(object, fun = c("mean",
  "median", "sum"))
}
\arguments{
\item{object}{Object.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{fun}{\code{character(1)}.
Mathematical function name to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{"tbl_df"}: \code{grouped_df}.
Tibble grouped by \code{sampleID} column.
\item \code{"DataFrame"}: \code{DataFrame}.
Row names are identical to the column names of the object, like
\code{\link[SummarizedExperiment:colData]{colData()}}.
}
}
\description{
Quality control metrics
}
\details{
\code{\link[=metrics]{metrics()}} takes data stored in \code{\link[SummarizedExperiment:colData]{colData()}}
and consistently returns a tibble grouped by sample by default (\code{sampleID}).
It always returns \code{sampleName} and \code{interestingGroups} columns, even when
these columns are not defined in \code{colData}. This is designed to integrate
with plotting functions that use ggplot2 internally.
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: Metrics are sample level. \code{sampleID} column corresponds
to \code{colnames}.

\item \code{SingleCellExperiment}: Metrics are cell level. \code{cellID} column corresponds to
\code{colnames}. Tibble is returned grouped by sample (\code{sampleID} column).
}}

\note{
Updated 2019-08-06.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)
rse <- RangedSummarizedExperiment
sce <- SingleCellExperiment

## SummarizedExperiment ====
x <- metrics(rse)
print(x)

## SingleCellExperiment ====
x <- metrics(sce)
print(x)

x <- metricsPerSample(sce, fun = "mean")
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
