% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metricsPerSample}
\alias{metrics,SummarizedExperiment-method}
\alias{metrics,SingleCellExperiment-method}
\alias{metricsPerSample,SingleCellExperiment-method}
\title{Metrics}
\usage{
metrics(object, ...)

metricsPerSample(object, ...)

\S4method{metrics}{SummarizedExperiment}(object)

\S4method{metrics}{SingleCellExperiment}(object)

\S4method{metricsPerSample}{SingleCellExperiment}(object, fun = c("mean",
  "median", "sum"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{fun}{\code{string}. Mathematical function name to apply. Uses \code{\link[=match.arg]{match.arg()}}.}
}
\value{
\code{grouped_df}. Grouped by \code{sampleID} column.
}
\description{
This function takes data stored in \code{\link[=colData]{colData()}} and consistently returns a
tibble grouped by sample, to be used for plotting with \code{ggplot2}.
}
\section{SingleCellExperiment}{


Note that metrics are cell level but grouped by sample.
}

\examples{
data(rse_small, sce_small)

# SummarizedExperiment ====
x <- metrics(rse_small)
print(x)

# SingleCellExperiment ====
x <- metrics(sce_small)
print(x)

x <- metricsPerSample(sce_small, fun = "mean")
print(x)
}
\seealso{
Other SummarizedExperiment Functions: \code{\link{counts}},
  \code{\link{interestingGroups}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{meltCounts}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{SummarizedExperiment Functions}
