% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,SummarizedExperiment-method}
\title{Metrics}
\usage{
\S4method{metrics}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DataFrame}.
}
\description{
By default, \code{\link[=metrics]{metrics()}} returns \code{\link[=colData]{colData()}} with an \code{interestingGroups}
column automatically added. This is useful for \code{ggplot2} scripting. Column
names are always sorted alphabetically.
}
\note{
\code{\link[=metrics]{metrics()}} Requires that the \code{sampleName} column is defined, and will
error otherwise.
}
\examples{
x <- metrics(rse_bcb)
glimpse(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}}, \code{\link{counts}},
  \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{sampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
