% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,SummarizedExperiment-method}
\title{Metrics}
\usage{
metrics(object, ...)

\S4method{metrics}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
By default, \code{\link[=metrics]{metrics()}} returns \code{\link[=colData]{colData()}} with an \code{interestingGroups}
column automatically added. This is useful for \code{ggplot2} scripting. Column
names are always sorted alphabetically.
}
\note{
\code{\link[=metrics]{metrics()}} Requires that the \code{sampleName} column is defined, and will
error otherwise.
}
\examples{
x <- metrics(rse_small)
glimpse(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
