% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,SummarizedExperiment-method}
\alias{metrics,SingleCellExperiment-method}
\title{Metrics}
\usage{
metrics(object, ...)

\S4method{metrics}{SummarizedExperiment}(object)

\S4method{metrics}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{grouped_df}. Grouped by \code{sampleID} column.
}
\description{
This function takes data stored in \code{\link[=colData]{colData()}} and consistently returns a
tibble grouped by sample, to be used for plotting with \code{ggplot2}.
}
\section{SingleCellExperiment}{


Note that metrics are cell level but grouped by sample.
}

\examples{
# SummarizedExperiment ====
x <- metrics(rse_small)
print(x)

# SingleCellExperiment ====
x <- metrics(sce_small)
print(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{tpm}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
