% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\name{metrics}
\alias{metrics}
\alias{metrics,SummarizedExperiment-method}
\title{Quality control metrics}
\usage{
metrics(object, ...)

\S4method{metrics}{SummarizedExperiment}(object, return = c("tbl_df", "DataFrame"))
}
\arguments{
\item{object}{Object.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}

\item{fun}{\code{character(1)}.
Mathematical function name to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally.}
}
\value{
Object of class determined by \code{return} argument.
}
\description{
Quality control metrics
}
\details{
\code{\link[=metrics]{metrics()}} takes data stored in \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}}
and consistently returns a \code{tbl_df} or \code{DataFrame} with \code{sampleName} and
\code{interestingGroups} columns, even when these columns are not defined in
\code{colData()}. This is designed to integrate with plotting functions that use
ggplot2 internally.
}
\section{Functions}{
\itemize{
\item \code{metrics,SummarizedExperiment-method}: Sample-level metrics.
}}

\note{
These functions will error intentionally if no numeric columns are
defined in \code{colData()}.

Updated 2021-02-02.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
object <- calculateMetrics(object)
x <- metrics(object)
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
