% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MGI2Ensembl-class}
\alias{MGI2Ensembl-class}
\title{MGI-to-Ensembl gene identifier mappings}
\value{
\code{MGI2Ensembl}.
Contains a \code{DataFrame} with \code{mgiID} and \code{geneID} columns.
}
\description{
MGI-to-Ensembl gene identifier mappings
}
\author{
Michael Steinbaugh
}
