% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/MGI2Ensembl.R
\docType{class}
\name{MGI2Ensembl-class}
\alias{MGI2Ensembl-class}
\alias{MGI2Ensembl}
\title{MGI-to-Ensembl gene identifier mappings}
\usage{
MGI2Ensembl()
}
\value{
\code{MGI2Ensembl}.
Contains a \code{DataFrame} with \code{mgiID} and \code{geneID} columns.
}
\description{
MGI-to-Ensembl gene identifier mappings
}
\note{
Updated 2019-07-28.
}
\examples{
options(acid.test = TRUE)
x <- MGI2Ensembl()
print(x)
}
\author{
Michael Steinbaugh
}
