% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{mgi2ensembl-class}
\alias{mgi2ensembl-class}
\title{mgi2ensembl Class}
\value{
\code{mgi2ensembl}.
}
\description{
MGI to Ensembl gene ID mappings.
}
\details{
Contains a \code{DataFrame} with \code{mgiID} and \code{geneID} columns.
}
\examples{
x <- new(
    Class = "mgi2ensembl",
    DataFrame(
        mgiID = 87853L,
        geneID = "ENSMUSG00000027596",
        row.names = 87853L
    )
)
print(x)
}
\seealso{
\code{\link[=mgi2ensembl]{mgi2ensembl()}}.

Other S4 Classes: \code{\link{PANTHER-class}},
  \code{\link{ensembl2entrez-class}},
  \code{\link{gene2symbol-class}},
  \code{\link{hgnc2ensembl-class}},
  \code{\link{tx2gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
