% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MGI2Ensembl-class}
\alias{MGI2Ensembl-class}
\title{MGI-to-Ensembl Identifier Mappings.}
\value{
Contains a \code{DataFrame} with \code{mgiID} and \code{geneID} columns.
}
\description{
MGI-to-Ensembl Identifier Mappings.
}
\seealso{
\code{\link[=mgi2ensembl]{mgi2ensembl()}}.

Other S4 Classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
