% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{MGI2Ensembl}
\alias{MGI2Ensembl}
\title{MGI-to-Ensembl gene identifier mappings}
\usage{
MGI2Ensembl()
}
\value{
\code{MGI2Ensembl}.
}
\description{
MGI-to-Ensembl gene identifier mappings
}
\note{
Updated 2019-08-08.
}
\examples{
options(acid.test = TRUE)
x <- MGI2Ensembl()
print(x)
}
