% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGI2Ensembl.R
\name{MGI2Ensembl}
\alias{MGI2Ensembl}
\title{MGI-to-Ensembl Gene Identifier Mappings}
\usage{
MGI2Ensembl()
}
\value{
\code{MGI2Ensembl}.
}
\description{
MGI-to-Ensembl Gene Identifier Mappings
}
\examples{
options(basejump.test = TRUE)
x <- MGI2Ensembl()
print(x)
}
\seealso{
\code{\link[=MGI2Ensembl]{MGI2Ensembl()}}.

Other S4 classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{PANTHER-class}}, \code{\link{Tx2Gene-class}}
}
