% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGI2Ensembl.R
\name{MGI2Ensembl}
\alias{MGI2Ensembl}
\title{MGI-to-Ensembl Gene Identifier Mappings}
\usage{
MGI2Ensembl()
}
\value{
\code{MGI2Ensembl}.
}
\description{
MGI-to-Ensembl Gene Identifier Mappings
}
\examples{
options(basejump.test = TRUE)
x <- MGI2Ensembl()
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{Ensembl2Entrez}},
  \code{\link{Gene2Symbol}}, \code{\link{HGNC2Ensembl}},
  \code{\link{Tx2Gene}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{mapGenes}}
}
\concept{Identifier Mapping Functions}
