% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGI2Ensembl.R, R/deprecated.R
\name{MGI2Ensembl}
\alias{MGI2Ensembl}
\alias{mgi2ensembl}
\title{MGI-to-Ensembl Gene Identifier Mappings}
\usage{
MGI2Ensembl(.test = FALSE)
}
\arguments{
\item{.test}{\code{boolean}. \emph{For unit testing only}.}

\item{...}{Additional arguments.}
}
\value{
\code{MGI2Ensembl}.
}
\description{
MGI-to-Ensembl Gene Identifier Mappings
}
\examples{
x <- MGI2Ensembl(.test = TRUE)
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{Ensembl2Entrez}},
  \code{\link{Gene2Symbol}}, \code{\link{HGNC2Ensembl}},
  \code{\link{Tx2Gene}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{mapGenes}}
}
\concept{Identifier Mapping Functions}
