% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi2ensembl.R
\name{mgi2ensembl}
\alias{mgi2ensembl}
\title{MGI to Ensembl Gene ID Mappings}
\usage{
mgi2ensembl(.test = FALSE)
}
\arguments{
\item{.test}{\code{boolean}. \emph{For unit testing only}.}
}
\value{
\code{mgi2ensembl}.
}
\description{
MGI to Ensembl Gene ID Mappings
}
\examples{
x <- mgi2ensembl(.test = TRUE)
print(x)
}
\seealso{
\code{mgi2ensembl-class}.

Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{eggnog}}, \code{\link{emptyRanges}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{geneSynonyms}}, \code{\link{hgnc2ensembl}},
  \code{\link{makeGRanges}}, \code{\link{makeGene2symbol}},
  \code{\link{makeTx2gene}}, \code{\link{organism}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
