% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi2ensembl.R
\name{mgi2ensembl}
\alias{mgi2ensembl}
\alias{MGI2Ensembl}
\title{\code{MGI2Ensembl} Generator}
\usage{
mgi2ensembl(.test = FALSE)
}
\arguments{
\item{.test}{\code{boolean}. \emph{For unit testing only}.}
}
\value{
\code{MGI2Ensembl}.
}
\description{
\code{MGI2Ensembl} Generator
}
\examples{
x <- mgi2ensembl(.test = TRUE)
print(x)
}
\seealso{
Other S4 Generators: \code{\link{eggnog}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{hgnc2ensembl}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
