% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGI2Ensembl.R
\name{MGI2Ensembl}
\alias{MGI2Ensembl}
\title{MGI-to-Ensembl gene identifier mappings}
\usage{
MGI2Ensembl()
}
\value{
\code{MGI2Ensembl}.
}
\description{
MGI-to-Ensembl gene identifier mappings
}
\note{
Updated 2020-10-05.
}
\examples{
object <- MGI2Ensembl()
print(object)
}
