% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi2gene.R
\name{mgi2gene}
\alias{mgi2gene}
\title{MGI to Ensembl Gene ID Mappings}
\usage{
mgi2gene()
}
\value{
\code{DataFrame}.
}
\description{
MGI to Ensembl Gene ID Mappings
}
\examples{
x <- mgi2gene()
print(x)
}
\seealso{
Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
