% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi2gene.R
\name{mgi2gene}
\alias{mgi2gene}
\title{MGI to Ensembl Gene ID Mappings}
\usage{
mgi2gene()
}
\value{
\code{data.frame}.
}
\description{
MGI to Ensembl Gene ID Mappings
}
\examples{
x <- mgi2gene()
glimpse(x)
}
\seealso{
Other Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\concept{Annotation Functions}
