% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplate.R
\name{microplate}
\alias{microplate}
\title{Microtiter Plate Well Identifiers}
\usage{
microplate(plates = 1L, wells = 96L, controls = 0L, prefix = NULL)
}
\arguments{
\item{plates}{Number of plates.}

\item{wells}{Number of wells (\code{96}, \code{384}).}

\item{controls}{Number of control wells.}

\item{prefix}{\emph{Optional}. Plate name prefix.}
}
\value{
Character vector containing well identifiers.
}
\description{
Microtiter Plate Well Identifiers
}
\examples{
# Single 96-well plate
microplate(wells = 96L)

# 2 96-well plates
microplate(plates = 2L, wells = 96L)

# Single 384-well plate
microplate(wells = 384L)

# 2 96-well plates with 6 control wells per plate
microplate(plates = 2L, wells = 96L, controls = 6L)
}
