% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midnightTheme.R
\name{midnightTheme}
\alias{midnightTheme}
\title{Midnight Theme}
\usage{
midnightTheme(...)
}
\arguments{
\item{...}{Passthrough arguments to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
ggplot theme.
}
\description{
Blackout palette that sets the plot background as black, with white text.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars) +
    geom_point(aes(x = wt, y = mpg, colour = factor(gear)))
p + midnightTheme()
}
\seealso{
\itemize{
\item \code{\link[ggplot2:theme]{ggplot2::theme()}}.
\item Inspired by \code{Seurat::DarkTheme()}, with some modifications.
}
}
