% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midnightTheme.R
\name{midnightTheme}
\alias{midnightTheme}
\title{Midnight Theme for \code{ggplot} Objects}
\usage{
midnightTheme(...)
}
\arguments{
\item{...}{Passthrough arguments to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
ggplot2 theme.
}
\description{
Blackout palette that sets the plot background as black, with white text.
}
\seealso{
\itemize{
\item \code{\link[ggplot2:theme]{ggplot2::theme()}}.
\item \code{Seurat::DarkTheme()}.
}
}
