% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Motor Trend Car Road Tests}
\format{An object of class \code{data.frame} with 32 rows and 11 columns.}
\usage{
mtcars
}
\description{
The data was extracted from the 1974 \emph{Motor Trend} US magazine,
  and comprises fuel consumption and 10 aspects of
  automobile design and performance for 32 automobiles (1973--74
  models).
}
\keyword{internal}
