% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Motor Trend car road tests.}
\format{An object of class \code{data.frame} with 32 rows and 11 columns.}
\usage{
mtcars
}
\description{
Motor Trend car road tests.
}
\keyword{internal}
