% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Motor Trend Car Road Tests}
\format{An object of class \code{data.frame} with 32 rows and 11 columns.}
\usage{
mtcars
}
\value{
\code{data.frame}.
}
\description{
Motor Trend Car Road Tests
}
\examples{
glimpse(mtcars)
}
\keyword{internal}
