% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAssign.R
\name{multiAssign}
\alias{multiAssign}
\title{Assign multiple variables}
\usage{
multiAssign(..., envir = .GlobalEnv, prefix = NULL)
}
\arguments{
\item{...}{\code{list} of variables to assign}

\item{envir}{Desired \code{environment} (optional)}

\item{prefix}{Name prefix (optional)}
}
\description{
Optionally, you can specify a name prefix and the desired environment
}
\author{
Michael Steinbaugh
}
\keyword{dev}
\keyword{package}
