% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassign.R
\name{multiassign}
\alias{multiassign}
\title{Assign multiple variables}
\usage{
multiassign(..., envir = .GlobalEnv, prefix = NULL)
}
\arguments{
\item{...}{List of variables to assign}

\item{envir}{Desired environment (optional)}

\item{prefix}{Name prefix (optional)}
}
\value{
Assigns variables to new environment with a name prefix, if desired.
}
\description{
Optionally, you can specify a name prefix and the desired environment.
}

