% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassignAsEnvir.R
\name{multiassignAsEnvir}
\alias{multiassignAsEnvir}
\title{Assign Multiple Objects as an Environment}
\usage{
multiassignAsEnvir(..., envirName, parentEnvir = parent.frame())
}
\arguments{
\item{...}{Objects as dots.}

\item{envirName}{New environment name (\code{string}).}

\item{parentEnvir}{Parent \code{environment} where to assign the new environment.}
}
\value{
\code{character} of object names in the new \code{environment}.
}
\description{
Assign Multiple Objects as an Environment
}
\examples{
multiassignAsEnvir(rnaseqCounts, singleCellCounts, envirName = "example")
class(example)
ls(example)
}
\seealso{
Other Developer Functions: \code{\link{detectHPC}},
  \code{\link{dots}}, \code{\link{localOrRemoteFile}},
  \code{\link{methodFormals}}, \code{\link{printString}}
}
\author{
Michael Steinbaugh
}
