% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassignAsEnvir.R
\name{multiassignAsEnvir}
\alias{multiassignAsEnvir}
\title{Assign Multiple Objects as an Environment}
\usage{
multiassignAsEnvir(..., envirName, parentEnvir = parent.frame())
}
\arguments{
\item{...}{Objects as dots.}

\item{envirName}{\code{string}. Name of the new \code{environment} to create.}

\item{parentEnvir}{\code{environment}. Parent \code{environment} where to assign the
new \code{environment}, specified by \code{envirName} argument.}
}
\value{
\code{character} vector of object names in the new \code{environment}.
}
\description{
Assign Multiple Objects as an Environment
}
\examples{
multiassignAsEnvir(rnaseq_counts, single_cell_counts, envirName = "example")
class(example)
ls(example)
}
\seealso{
Other Developer Functions: \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{matchS4Call}}, \code{\link{methodFormals}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
