% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassignAsEnvir.R
\name{multiassignAsEnvir}
\alias{multiassignAsEnvir}
\title{Assign Multiple Objects as an Environment}
\usage{
multiassignAsEnvir(..., envirName, parentFrame = parent.frame())
}
\arguments{
\item{...}{Objects as dots.}

\item{envirName}{\code{string}. Name of the new \code{environment} to create.}

\item{parentFrame}{\code{environment}. Parent \code{environment} where to assign the
new \code{environment}, specified by \code{envirName} argument.}
}
\value{
\code{character}. Object names defined in the new \code{environment}.
}
\description{
Assign Multiple Objects as an Environment
}
\examples{
multiassignAsEnvir(rse_small, sce_small, envirName = "example")
class(example)
ls(example)
}
