% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiassignAsNewEnvir.R
\name{multiassignAsNewEnvir}
\alias{multiassignAsNewEnvir}
\title{Assign Multiple Objects as a New Environment}
\usage{
multiassignAsNewEnvir(..., envirName, parentEnvir = parent.frame())
}
\arguments{
\item{...}{Objects as dots.}

\item{envirName}{New environment name (string).}

\item{parentEnvir}{Parent environment of new environment.}
}
\value{
Character vector of assigned objects in the new environment.
}
\description{
Assign Multiple Objects as a New Environment
}
\examples{
multiassignAsNewEnvir(mtcars, starwars, envirName = "test")
}
\seealso{
\code{\link[Biobase:multiassign]{Biobase::multiassign()}}.

Other Object Assignment Utilities: \code{\link{assignAndSaveData}}
}
\concept{Object Assignment Utilities}
