% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{titleCase}
\alias{titleCase}
\alias{title_case}
\alias{firstCase}
\alias{first_case}
\alias{makeNames}
\alias{make_names}
\alias{makeNamesCamel}
\alias{make_names_camel}
\alias{camel}
\alias{makeNamesSnake}
\alias{make_names_snake}
\alias{snake}
\alias{setNamesCamel}
\alias{set_names_camel}
\alias{setNamesDot}
\alias{set_names_dot}
\alias{setNamesSnake}
\alias{set_names_snake}
\alias{setColnames}
\alias{setRownames}
\alias{sanitizeNames}
\alias{sanitize_names}
\alias{sanitiseNames}
\alias{sanitise_names}
\title{Generate syntactically valid names.}
\usage{
titleCase(character)

title_case(character)

firstCase(character)

first_case(character)

makeNames(character)

make_names(character)

makeNamesCamel(character)

make_names_camel(character)

camel(character)

makeNamesSnake(character)

make_names_snake(character)

snake(character)

setNamesCamel(object)

set_names_camel(object)

setNamesDot(object)

set_names_dot(object)

setNamesSnake(object)

set_names_snake(object)

sanitizeNames(object)

sanitize_names(object)

sanitiseNames(object)

sanitise_names(object)
}
\arguments{
\item{character}{Character vector.}

\item{object}{Object for which a \code{\link[=names]{names()}} attribute will be meaningful.}
}
\value{
Character vector or object with valid names.
}
\description{
These are convenience function that set \code{\link[=names]{names()}} on an object and return the
object, without modifying the underlying data.
}
\details{
\itemize{
\item The \code{\link[=makeNames]{makeNames()}} and \code{*Case()} family of functions manipulate a character
vector, supporting either \code{camelCase} or \code{snake_case}.
\item The \code{\link[=setNames]{setNames()}} family of functions work similarly to \code{\link[=makeNames]{makeNames()}}
overall, but are to be used on objects supporting name assignments (e.g.
\code{data.frame}, \code{matrix}, \code{list}).
\item \code{\link[=sanitizeNames]{sanitizeNames()}} modifies both \code{\link[=colnames]{colnames()}} and \code{\link[=rownames]{rownames()}} to
\code{snake_case} format, and is generally recommended for clean-up of RNA-seq
count matrices containing gene names.
}
}
\examples{
# Create character vectors of valid names
camel("RNAi clone")
snake("RNAi clone")
titleCase("RNA sequencing")
firstCase("RNA sequencing")

# Set names on objects
setNamesCamel(head(starwars))
setNamesSnake(head(starwars))
setNamesDot(head(starwars))

# Sanitize both colnames and rownames
sanitizeNames(head(mtcars))
}
\seealso{
\itemize{
\item \code{\link[magrittr:set_names]{magrittr::set_names()}}.
\item \code{\link[stats:setNames]{stats::setNames()}}.
}
}
