% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R, R/zzz.R
\name{.makeTitleCase}
\alias{.makeTitleCase}
\alias{.makeFirstCase}
\alias{.makeNamesDotted}
\alias{.makeNamesCamel}
\alias{.makeNamesSnake}
\alias{.checkNames}
\alias{.checkRownames}
\alias{.setNamesDotted}
\alias{.setNamesCamel}
\alias{.setNamesSnake}
\alias{camel}
\alias{dotted}
\alias{snake}
\alias{sanitizeNames}
\alias{titleCase}
\alias{firstCase}
\alias{first_case}
\alias{sanitiseNames}
\alias{sanitize_names}
\alias{sanitise_names}
\alias{title_case}
\title{Make syntactically valid names}
\usage{
camel(object, rownames = TRUE)

dotted(object, rownames = TRUE)

snake(object, rownames = TRUE)

titleCase(object)

firstCase(object)
}
\arguments{
\item{object}{Character vector or an object for which \code{\link[=names]{names()}} assignment
will be meaningful.}

\item{rownames}{Include row names.}
}
\value{
Object with syntatically valid names. For objects supporting
\code{\link[base:names]{base::names()}}, the underlying data returns unchanged.
}
\description{
These are convenience functions that sanitize names into \code{camelCase}
(preferred), \code{snake_case}, or \code{dot.notation}. Dot assignment is provided for
use with base R operations, but we advise against using \code{dot.notation} name
assignments for values and/or function.
}
\details{
For unnamed character vectors, these functions will sanitize the underlying
values. Otherwise, the functions will set \code{\link[=names]{names()}} and/or \code{\link[=rownames]{rownames()}} on
objects supporting name assignments. They return the object without
modification of the underlying data.
}
\examples{
# Unnamed character vector, with no names assigned
unnamed_vec <- c("hello world",
                 "HELLO WORLD",
                 "RNAi clones",
                 "worfdbHTMLRemap",
                 123,
                 NA)

# Named character vector
named_vec <- c(Item.A = "hello world", Item.B = "HELLO WORLD")

# Data frame with colnames and rownames
df <- head(mtcars)

# Tibble, with colnames. Rownames aren't supported on these!
tbl <- head(starwars)

# List, with assigned namess
lst <- list(Item.A = c(1, 2),
            Item.B = c(3, 4))


# camelCase
camel(unnamed_vec)
camel(named_vec)
camel(df)
camel(df, rownames = FALSE)
camel(tbl)
camel(lst)


# snake_case
snake(unnamed_vec)
snake(named_vec)
snake(df)
snake(df, rownames = FALSE)
snake(tbl)
snake(lst)


# dotted.case
dotted(unnamed_vec)
dotted(named_vec)
dotted(df)
dotted(df, rownames = FALSE)
dotted(tbl)
dotted(lst)
}
