% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{naStrings}
\alias{naStrings}
\title{\code{NA} strings}
\format{An object of class \code{character} of length 5.}
\usage{
naStrings
}
\description{
\code{NA} strings
}
\examples{
naStrings
}
\keyword{datasets}
