% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nonzeroRowsAndCols-methods.R
\name{nonzeroRowsAndCols}
\alias{nonzeroRowsAndCols}
\alias{nonzeroRowsAndCols,matrix-method}
\alias{nonzeroRowsAndCols,Matrix-method}
\alias{nonzeroRowsAndCols,SummarizedExperiment-method}
\title{Subset object to keep only non-zero rows and columns}
\usage{
nonzeroRowsAndCols(object, ...)

\S4method{nonzeroRowsAndCols}{matrix}(object)

\S4method{nonzeroRowsAndCols}{Matrix}(object)

\S4method{nonzeroRowsAndCols}{SummarizedExperiment}(object, assay = 1L)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Subset object to keep only non-zero rows and columns
}
\note{
Updated 2020-01-20.
}
\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "AcidTest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
dim(object)
x <- nonzeroRowsAndCols(object)
dim(x)

## SingleCellExperiment ====
object <- SingleCellExperiment
dim(object)
x <- nonzeroRowsAndCols(object)
dim(x)
}
