% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism-methods.R
\docType{methods}
\name{organism}
\alias{organism}
\alias{organism<-}
\alias{organism,matrix-method}
\alias{organism,sparseMatrix-method}
\alias{organism,data.frame-method}
\alias{organism,Annotated-method}
\alias{organism,DataTable-method}
\alias{organism,GRanges-method}
\alias{organism,SummarizedExperiment-method}
\title{Organism}
\usage{
organism(object)

organism(object) <- value

\S4method{organism}{matrix}(object)

\S4method{organism}{sparseMatrix}(object)

\S4method{organism}{data.frame}(object)

\S4method{organism}{Annotated}(object)

\S4method{organism}{DataTable}(object)

\S4method{organism}{GRanges}(object)

\S4method{organism}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
Latin organism name (e.g. \emph{Homo sapiens}).
}
\description{
Organism
}
\note{
Updated 2019-07-28.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
organism(rse)
}
\seealso{
[detectOrganism()][.
}
