% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism-methods.R
\docType{methods}
\name{organism}
\alias{organism}
\alias{organism,matrix-method}
\alias{organism,sparseMatrix-method}
\alias{organism,data.frame-method}
\alias{organism,DataFrame-method}
\alias{organism,GRanges-method}
\alias{organism,SummarizedExperiment-method}
\title{Organism}
\usage{
\S4method{organism}{matrix}(object)

\S4method{organism}{sparseMatrix}(object)

\S4method{organism}{data.frame}(object)

\S4method{organism}{DataFrame}(object)

\S4method{organism}{GRanges}(object)

\S4method{organism}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Organism
}
\examples{
data(rse)
organism(rse)
}
\seealso{
\code{detectOrganism()}.
}
