% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism-methods.R
\docType{methods}
\name{organism}
\alias{organism}
\alias{organism,character-method}
\alias{organism,factor-method}
\alias{organism,matrix-method}
\alias{organism,data.frame-method}
\alias{organism,DataFrame-method}
\alias{organism,sparseMatrix-method}
\alias{organism,GRanges-method}
\alias{organism,SummarizedExperiment-method}
\title{Organism Accessor}
\usage{
\S4method{organism}{character}(object)

\S4method{organism}{factor}(object)

\S4method{organism}{matrix}(object)

\S4method{organism}{data.frame}(object)

\S4method{organism}{DataFrame}(object)

\S4method{organism}{sparseMatrix}(object)

\S4method{organism}{GRanges}(object)

\S4method{organism}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{string}. Full latin organism name. Stops on match failure.
}
\description{
Supports organism detection from Ensembl identifier or genome build.
}
\section{Supported organisms}{

\itemize{
\item \emph{Caenorhabditis elegans} (roundworm)
\item \emph{Danio rerio} (zebrafish)
\item \emph{Drosophila melanogaster} (fruitfly)
\item \emph{Gallus gallus} (chicken)
\item \emph{Homo sapiens} (human)
\item \emph{Mus musculus} (mouse)
\item \emph{Ovis aries} (sheep)
\item \emph{Rattus norvegicus} (rat)
\item \emph{Saccharomyces cerevisiae} (yeast)
}
}

\examples{
# Match by gene identifier.
organism("ENSG00000000003")

# Match by genome build.
organism("GRCh38")  # Ensembl
organism("hg38")    # UCSC

# Match by alternate organism name.
organism("H. sapiens")
organism("hsapiens")

# The function will skip transgenes/spike-ins until we find a match.
organism(c("EGFP", "TDTOMATO", "ENSG00000000003"))

# But it only returns the first match, if there are multiple genomes.
organism(c("ENSG00000000003", "ENSMUSG00000000001"))

# SummarizedExperiment support.
data(rse_small)
organism(rse_small)
}
\seealso{
Other Annotation Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{makeGRanges}}, \code{\link{makeGene2Symbol}},
  \code{\link{makeTx2Gene}}, \code{\link{panther}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
