% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism-methods.R
\docType{methods}
\name{organism}
\alias{organism}
\alias{organism,matrix-method}
\alias{organism,sparseMatrix-method}
\alias{organism,data.frame-method}
\alias{organism,DataFrame-method}
\alias{organism,GRanges-method}
\alias{organism,SummarizedExperiment-method}
\title{Organism Accessor}
\usage{
\S4method{organism}{matrix}(object)

\S4method{organism}{sparseMatrix}(object)

\S4method{organism}{data.frame}(object)

\S4method{organism}{DataFrame}(object)

\S4method{organism}{GRanges}(object)

\S4method{organism}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Organism Accessor
}
\examples{
data(rse_small)
organism(rse_small)
}
\seealso{
\itemize{
\item \code{\link[BiocGenerics:organism]{BiocGenerics::organism()}}.
\item \code{\link[=detectOrganism]{detectOrganism()}}.
}

Other Metadata Functions: \code{\link{detectHPC}},
  \code{\link{detectLanes}}, \code{\link{detectOrganism}},
  \code{\link{interestingGroups}},
  \code{\link{microplate}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Metadata Functions}
