% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/PANTHER.R
\docType{class}
\name{PANTHER-class}
\alias{PANTHER-class}
\alias{PANTHER}
\title{PANTHER database annotations}
\usage{
PANTHER(organism, release = NULL,
  progress = getOption("basejump.progress", FALSE))
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name.

Supported organisms:
\itemize{
\item \emph{Caenorhabditis elegans}
\item \emph{Drosophila melanogaster}
\item \emph{Homo sapiens}
\item \emph{Mus musculus}
}}

\item{release}{\code{character(1)} or \code{NULL}.
PANTHER release version. If \code{NULL}, defaults to current release. Consult
the PANTHER website for a list of release versions available from the FTP
server (e.g. \code{"14.0"}).}

\item{progress}{\code{logical(1)}.
Use \code{\link[pbapply:pblapply]{pbapply::pblapply()}} to show progress.}
}
\value{
\code{PANTHER}. Contains a \code{DataFrame}.
}
\description{
\href{http://www.pantherdb.org}{PANTHER} gene ontology definitions. PANTHER stands
for \strong{P}rotein \strong{AN}alysis \strong{TH}rough \strong{E}volutionary \strong{R}elationships.
}
\examples{
options(basejump.test = TRUE)
x <- PANTHER("Homo sapiens", progress = FALSE)
summary(x)
}
\author{
Michael Steinbaugh
}
