% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PANTHER-class}
\alias{PANTHER-class}
\title{PANTHER Class}
\value{
\code{PANTHER}.
}
\description{
PANTHER Gene Ontology definitions.
}
\seealso{
\code{\link[=panther]{panther()}}.

Other S4 Classes: \code{\link{ensembl2entrez-class}},
  \code{\link{gene2symbol-class}},
  \code{\link{hgnc2ensembl-class}},
  \code{\link{mgi2ensembl-class}},
  \code{\link{tx2gene-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
