% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PANTHER-class}
\alias{PANTHER-class}
\title{PANTHER database annotations}
\value{
\code{PANTHER}.
}
\description{
\href{http://www.pantherdb.org}{PANTHER} gene ontology definitions. PANTHER stands
for \strong{P}rotein \strong{AN}alysis \strong{TH}rough \strong{E}volutionary \strong{R}elationships.
}
\note{
Updated 2019-08-08.
}
