% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{panther}
\alias{panther}
\title{PANTHER Gene Ontology Annotations}
\usage{
panther(organism = c(human = "Homo sapiens", mouse = "Mus musculus", nematode
  = "Caenorhabditis elegans", fruit_fly = "Drosophila melanogaster"),
  release = NULL)
}
\description{
Only the current PANTHER release is currently supported.
}
\examples{
x <- panther("Homo sapiens")
glimpse(hs)
}
\author{
Michael Steinbaugh
}
