% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANTHER.R, R/deprecated.R
\name{PANTHER}
\alias{PANTHER}
\alias{panther}
\title{PANTHER Database Annotations}
\usage{
PANTHER(organism, release = NULL, progress = TRUE, .test = FALSE)
}
\arguments{
\item{organism}{\code{string}. Full Latin organism name. Currently supported
organisms: \emph{Homo sapiens}, \emph{Mus musculus}, \emph{Caenorhabditis elegans},
\emph{Drosophila melanogaster}.}

\item{release}{\code{string} or \code{NULL}. PANTHER release version. If set \code{NULL},
defaults to current release. Consult the PANTHER website for a list of
release versions available from the FTP server (e.g. \code{"13.0"}).}

\item{progress}{\code{boolean}. Use \code{\link[=pblapply]{pblapply()}} internally to show progress.}

\item{.test}{\code{boolean}. \emph{For unit testing only}.}

\item{...}{Additional arguments.}
}
\value{
\code{PANTHER}.
}
\description{
\href{http://www.pantherdb.org}{PANTHER} gene ontology definitions. PANTHER stands
for \strong{P}rotein \strong{AN}alysis \strong{TH}rough \strong{E}volutionary \strong{R}elationships.
}
\examples{
x <- PANTHER("Homo sapiens", progress = FALSE, .test = TRUE)
summary(x)
}
\seealso{
Other Annotation Functions: \code{\link{EggNOG}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{makeGRanges}}, \code{\link{makeGene2Symbol}},
  \code{\link{makeTx2Gene}}
}
\concept{Annotation Functions}
