% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{panther}
\alias{panther}
\title{PANTHER Gene Ontology Annotations}
\usage{
panther(organism = c(human = "Homo sapiens", mouse = "Mus musculus",
  nematode_worm = "Caenorhabditis elegans", fruit_fly =
  "Drosophila melanogaster"), release = NULL)
}
\description{
PANTHER Gene Ontology Annotations
}
\examples{
x <- panther("Homo sapiens")
glimpse(hs)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
