% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{panther}
\alias{panther}
\title{PANTHER Gene Ontology Annotations}
\usage{
panther(organism = c(human = "Homo sapiens", mouse = "Mus musculus",
  nematode_worm = "Caenorhabditis elegans", fruit_fly =
  "Drosophila melanogaster"), release = NULL)
}
\description{
PANTHER Gene Ontology Annotations
}
\examples{
x <- panther("Homo sapiens")
glimpse(hs)
}
\author{
Michael Steinbaugh
}
