% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{panther}
\alias{panther}
\title{PANTHER Gene Ontology Annotations}
\usage{
panther(organism, release = NULL)
}
\arguments{
\item{organism}{\code{string}. Supported organisms:
\itemize{
\item \emph{Homo sapiens}
\item \emph{Mus musculus}
\item \emph{Caenorhabditis elegans}
\item \emph{Drosophila melanogaster}
}}

\item{release}{\code{string} or \code{NULL}. PANTHER release version. If set \code{NULL},
defaults to current release. Consult the PANTHER website for a list of
release versions available from the FTP server (e.g. \code{"13.0"}).}
}
\value{
\code{DataFrame}.
}
\description{
PANTHER Gene Ontology Annotations
}
\examples{
invisible(capture.output(
    x <- panther("Homo sapiens")
))
glimpse(x)
}
\seealso{
Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
