% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{panther}
\alias{panther}
\title{PANTHER Gene Ontology Annotations}
\usage{
panther(organism = c("Homo sapiens", "Mus musculus", "Caenorhabditis elegans",
  "Drosophila melanogaster"), release = NULL)
}
\arguments{
\item{organism}{\code{string}. Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{release}{\code{character} string denoting PANTHER release version. If
\code{NULL}, defaults to current release. Consult the PANTHER website for a list
of release versions available from the FTP server (e.g. \code{"13.0"}).}
}
\value{
\code{data.frame}.
}
\description{
PANTHER Gene Ontology Annotations
}
\examples{
invisible(capture.output(
    x <- panther("Homo sapiens")
))
glimpse(x)

# Clean up
unlink("PTHR*_human.gz")
}
\seealso{
Other Database Functions: \code{\link{eggnog}}
}
\author{
Michael Steinbaugh
}
