% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{panther}
\alias{panther}
\title{PANTHER Gene Ontology Annotations}
\usage{
panther(organism = c(human = "Homo sapiens", mouse = "Mus musculus",
  nematode_worm = "Caenorhabditis elegans", fruit_fly =
  "Drosophila melanogaster"), release = NULL)
}
\arguments{
\item{organism}{Full latin organism name (e.g. "\code{Homo sapiens}").}

\item{release}{\code{character} string denoting PANTHER release version. If
\code{NULL}, defaults to current release. Consult the PANTHER website for a list
of release versions available from the FTP server (e.g. \code{"13.0"}).}
}
\value{
\code{data.frame}.
}
\description{
PANTHER Gene Ontology Annotations
}
\examples{
x <- panther("Homo sapiens")
glimpse(x)
}
\seealso{
Other Gene Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{emptyRanges}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{parseGFFAttributes}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
