% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGFFAttributes.R
\name{parseGFFAttributes}
\alias{parseGFFAttributes}
\title{Parse GFF Attributes Column}
\usage{
parseGFFAttributes(file, select = c("gene_", "transcript_"), unique = TRUE,
  return = c("data.frame", "list"))
}
\arguments{
\item{select}{Attribute keys to select. Supports partial matching. Defaults
to returning attributes matching both "\code{gene_}" and "\code{transcript_}".}

\item{unique}{Return unique attributes.}
}
\value{
\code{data.frame}.
}
\description{
Parse GFF Attributes Column
}
\note{
Memory overhead and speed is greatly improved by requiring keys.
Deparsing only unique values is much faster and generally recommended.
}
\examples{
parseGFFAttributes("mmusculus.gtf") \%>\% glimpse()
}
