% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRd.R
\name{parseRd}
\alias{parseRd}
\title{Parse R Documentation}
\usage{
parseRd(object, kind)
}
\description{
Modified version of \code{tools:::.Rd_get_metadata()} that keeps whitespace and
returns \code{character} instead of \code{matrix}.
}
\examples{
db <- tools::Rd_db("base")
head(names(db))
Rd <- db[["nrow.Rd"]]
class(Rd)
summary(Rd)
examples <- parseRd(Rd, kind = "examples")
print(examples)
}
\seealso{
Other Developer Functions: \code{\link{calls}},
  \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{saveFunctionExamples}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
