% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRd.R
\name{parseRd}
\alias{parseRd}
\title{Parse R Documentation}
\usage{
parseRd(object, tag)
}
\arguments{
\item{object}{\code{Rd}. R documentation, returned from \code{\link[tools:Rd_db]{tools::Rd_db()}}.}

\item{tag}{\code{string}. Desired metadata type. For example, these are supported:
\itemize{
\item "\code{title}".
\item "\code{description}".
\item "\code{usage}".
\item "\code{arguments}".
\item "\code{value}".
\item "\code{references}".
\item "\code{seealso}".
\item "\code{examples}".
}}
}
\description{
Modified version of \code{tools:::.Rd_get_metadata()} that keeps whitespace and
returns \code{character} instead of \code{matrix}.
}
\examples{
db <- tools::Rd_db("base")
head(names(db))
Rd <- db[["nrow.Rd"]]
class(Rd)
summary(Rd)
RdTags(Rd)
examples <- parseRd(Rd, tag = "examples")
print(examples)
}
\seealso{
\code{\link[tools:Rd_db]{tools::Rd_db()}}.

Other Documentation Functions: \code{\link{RdTags}},
  \code{\link{saveRdExamples}}
}
\concept{Documentation Functions}
