% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-pct.R
\docType{methods}
\name{pct}
\alias{pct}
\alias{pct,numeric-method}
\title{Convert Numeric to Percentage}
\usage{
pct(object, ...)

\S4method{pct}{numeric}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Percentage.
}
\description{
Convert Numeric to Percentage
}
\examples{
pct(0.1)
}
\seealso{
Other Math Utilities: \code{\link{fc2lr}},
  \code{\link{geomean}}
}
\concept{Math Utilities}
