% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{Convert numeric to percentage}
\usage{
pct(number)
}
\arguments{
\item{number}{Number}
}
\value{
Percentage
}
\description{
Convert numeric to percentage
}
\examples{
pct(0.1)
}
\author{
Michael Steinbaugh
}
