% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotCellCounts}
\alias{plotCellCounts}
\title{Plot cell counts}
\usage{
plotCellCounts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot cell counts
}
\examples{
isS4(plotCellCounts)
}
\seealso{
\href{https://bioinformatics.sph.harvard.edu/bcbioRNASeq/}{bcbioRNASeq}.
}
