% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCorrelationHeatmap-methods.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,SummarizedExperiment-method}
\title{Plot Correlation Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{SummarizedExperiment}(object,
  interestingGroups = NULL, method = c("pearson", "spearman"),
  clusteringMethod = "ward.D2", showRownames = TRUE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = NULL, borderColor = NULL,
  title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{method}{\code{string}. Correlation coefficient (or covariance) method to be
computed. Defaults to "\code{pearson}" but "\code{spearman}" can also be used.
Consult the \code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{clusteringMethod}{\code{string}. Clustering method. Accepts the same values
as \code{\link[stats:hclust]{stats::hclust()}}.}

\item{showRownames}{\code{boolean}. Show row or column names.}

\item{showColnames}{\code{boolean}. Show row or column names.}

\item{treeheightRow}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{treeheightCol}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{color}{\code{function}, \code{character}, or \code{NULL}. Hexadecimal color function
or values to use for plot. We generally recommend these hexadecimal
functions from the viridis package:
\itemize{
\item \code{\link[viridis:viridis]{viridis::viridis()}} (\emph{default}).
\item \code{\link[viridis:inferno]{viridis::inferno()}}.
\item \code{\link[viridis:magma]{viridis::magma()}}.
\item \code{\link[viridis:plasma]{viridis::plasma()}}.
Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.
}}

\item{legendColor}{\code{function} or \code{NULL}. Hexadecimal color function to use
for legend labels. Note that hexadecimal values are not supported. If set
\code{NULL}, will use the default pheatmap colors.}

\item{borderColor}{\code{string} or \code{NULL}. Border color. Disabled by default for
improved aesthetics.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
Construct a correlation heatmap comparing the columns of the matrix.
}
\examples{
# SummarizedExperiment ====
plotCorrelationHeatmap(rse_small)

# Using viridis color palettes.
plotCorrelationHeatmap(
    object = rse_small,
    color = viridis::plasma,
    legendColor = viridis::viridis
)

# Using hexadecimal color input.
library("RColorBrewer")
purple_orange <- colorRampPalette(brewer.pal(n = 11L, name = "PuOr"))(256L)
plotCorrelationHeatmap(rse_small, color = purple_orange)

# Using default pheatmap colors.
plotCorrelationHeatmap(rse_small, color = NULL)
}
\seealso{
\itemize{
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}

Other Plot Functions: \code{\link{ggplot2_geoms}},
  \code{\link{ggplot2_themes}}, \code{\link{plotHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
