% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCorrelationHeatmap.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,matrix-method}
\title{Plot Correlation Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{matrix}(object, method = c("pearson",
  "spearman"), clusteringMethod = "ward.D2", annotationCol = NULL,
  color = viridis, legendColor = viridis, borderColor = NULL,
  title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}

\item{method}{Correlation coefficient (or covariance) method to be computed.
Defaults to "\code{pearson}" but "\code{spearman}" can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{clusteringMethod}{Clustering method. Accepts the same values as
\code{\link[stats:hclust]{stats::hclust()}}.}

\item{annotationCol}{\emph{Optional.} \code{data.frame} that defines annotation
mappings for the columns.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[=viridis]{viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[=viridis]{viridis()}} palette.}

\item{borderColor}{Border color.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and return \code{list}, containing \code{gtable}.
}
\description{
Construct a correlation heatmap comparing the columns of the matrix.
}
\examples{
mat <- as.matrix(mtcars)
plotCorrelationHeatmap(mat)
}
\seealso{
\itemize{
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}

Other Plot Functions: \code{\link{dynamicPlotlist}},
  \code{\link{midnightTheme}}, \code{\link{plotHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
\author{
Michael Steinbaugh
}
