% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerBiotype-methods.R
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\title{Plot Counts per Biotype}
\usage{
plotCountsPerBiotype(object, ...)
}
\description{
Plot Counts per Biotype
}
\examples{
plotCountsPerBiotype(rse_small)
}
\seealso{
Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotGene}}, \code{\link{plotHeatmap}},
  \code{\link{plotQC}}, \code{\link{plotQuantileHeatmap}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Plot Functions}
