% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerBiotype-methods.R
\docType{methods}
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\alias{plotCountsPerBroadClass}
\alias{plotCountsPerBiotype,SummarizedExperiment-method}
\alias{plotCountsPerBiotype,SingleCellExperiment-method}
\alias{plotCountsPerBroadClass,SummarizedExperiment-method}
\alias{plotCountsPerBroadClass,SingleCellExperiment-method}
\title{Plot Counts per Biotype}
\usage{
plotCountsPerBiotype(object, ...)

plotCountsPerBroadClass(object, ...)

\S4method{plotCountsPerBiotype}{SummarizedExperiment}(object, assay = 1L,
  n = 9L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")

\S4method{plotCountsPerBiotype}{SingleCellExperiment}(object, assay = 1L,
  n = 9L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")

\S4method{plotCountsPerBroadClass}{SummarizedExperiment}(object,
  assay = 1L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")

\S4method{plotCountsPerBroadClass}{SingleCellExperiment}(object,
  assay = 1L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{countsAxisLabel}{\code{string}. Counts axis label.}
}
\description{
Plot Counts per Biotype
}
\examples{
plotCountsPerBiotype(rse_small)
plotCountsPerBroadClass(rse_small)
}
\seealso{
\code{\link[=broadClass]{broadClass()}}.

Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotGene}}, \code{\link{plotHeatmap}},
  \code{\link{plotQC}}, \code{\link{plotQuantileHeatmap}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Plot Functions}
