% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerBiotype-methods.R,
%   R/plotCountsPerBroadClass-methods.R
\docType{methods}
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\alias{plotCountsPerBiotype,SummarizedExperiment-method}
\alias{plotCountsPerBroadClass,SummarizedExperiment-method}
\title{Plot counts per biotype}
\usage{
\S4method{plotCountsPerBiotype}{SummarizedExperiment}(object, assay = 1L,
  n = 9L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")

\S4method{plotCountsPerBroadClass}{SummarizedExperiment}(object,
  assay = 1L, interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{countsAxisLabel}{\code{character(1)}.
Counts axis label.}
}
\value{
\code{ggplot}.
}
\description{
Plot counts per biotype
}
\examples{
data(rse, sce, package = "acidData")

## SummarizedExperiment ====
plotCountsPerBiotype(rse)

## SingleCellExperiment ====
plotCountsPerBiotype(sce)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
