% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerGene-methods.R
\docType{methods}
\name{plotCountsPerGene}
\alias{plotCountsPerGene}
\alias{plotCountsPerGene,SummarizedExperiment-method}
\title{Plot Counts Per Gene}
\usage{
plotCountsPerGene(object, ...)

\S4method{plotCountsPerGene}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, geom = c("density", "boxplot", "violin"),
  trans = "identity", color = getOption("basejump.discrete.color",
  NULL), fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), countsAxisLabel = "counts",
  title = "counts per gene")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{geom}{\code{string}. Type of ggplot2 geometric object to use.}

\item{trans}{Either the name of a transformation object, or the
object itself. Built-in transformations include "asn", "atanh",
"boxcox", "exp", "identity", "log", "log10", "log1p", "log2",
"logit", "probability", "probit", "reciprocal", "reverse" and "sqrt".

A transformation object bundles together a transform, it's inverse,
and methods for generating breaks and labels. Transformation objects
are defined in the scales package, and are called \code{name_trans}, e.g.
\code{\link[scales:boxcox_trans]{scales::boxcox_trans()}}. You can create your own
transformation with \code{\link[scales:trans_new]{scales::trans_new()}}.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{title}{\code{string}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\examples{
plotCountsPerGene(rse_small, geom = "density")
plotCountsPerGene(rse_small, geom = "boxplot")
plotCountsPerGene(rse_small, geom = "violin")
}
\seealso{
Other Quality Control Functions: \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotTotalCounts}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
