% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCountsPerGene-methods.R
\docType{methods}
\name{plotCountsPerGene}
\alias{plotCountsPerGene}
\alias{plotCountsPerGene,SummarizedExperiment-method}
\alias{plotCountsPerGene,SingleCellExperiment-method}
\title{Plot Counts Per Gene}
\usage{
plotCountsPerGene(object, ...)

\S4method{plotCountsPerGene}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, geom = c("density", "boxplot", "violin"),
  trans = c("identity", "log2", "log10"),
  color = getOption("basejump.discrete.color", NULL),
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), countsAxisLabel = "counts",
  title = "counts per gene")

\S4method{plotCountsPerGene}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, geom = c("density", "boxplot", "violin"),
  trans = c("identity", "log2", "log10"),
  color = getOption("basejump.discrete.color", NULL),
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), countsAxisLabel = "counts",
  title = "counts per gene")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of count matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{geom}{\code{string}. Type of ggplot2 geometric object to use.}

\item{trans}{\code{string}. Apply a log transformation (e.g. \code{log2(x + 1L)}) to
the count matrix prior to melting, if desired. Use \code{"identity"} to return
unmodified (default).}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}. To set the discrete color
palette globally, use \code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}. To set the discrete fill palette
globally, use \code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for quality control
plots containing many samples.}

\item{countsAxisLabel}{\code{string}. Counts axis label.}

\item{title}{\code{string}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\examples{
data(rse_small, sce_small)

# SummarizedExperiment ====
plotCountsPerGene(rse_small, geom = "density")
plotCountsPerGene(rse_small, geom = "boxplot")
plotCountsPerGene(rse_small, geom = "violin")

# SingleCellExperiment ====
plotCountsPerGene(sce_small)
}
\seealso{
Other Quality Control Functions: \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotTotalCounts}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
