% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerGene-methods.R
\docType{methods}
\name{plotCountsPerGene}
\alias{plotCountsPerGene}
\alias{plotCountsPerGene,SummarizedExperiment-method}
\alias{plotCountsPerGene,SingleCellExperiment-method}
\title{Plot counts per gene}
\usage{
\S4method{plotCountsPerGene}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, geom = c("boxplot", "density", "violin"),
  trans = c("identity", "log2", "log10"),
  color = getOption("basejump.color.discrete", default = NULL),
  fill = getOption("basejump.fill.discrete", default = NULL),
  flip = getOption("basejump.flip", default = TRUE),
  countsAxisLabel = "counts", title = "counts per gene")

\S4method{plotCountsPerGene}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, geom = c("boxplot", "density", "violin"),
  trans = c("identity", "log2", "log10"),
  color = getOption("basejump.color.discrete", default = NULL),
  fill = getOption("basejump.fill.discrete", default = NULL),
  flip = getOption("basejump.flip", default = TRUE),
  countsAxisLabel = "counts", title = "counts per gene")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{geom}{\code{character(1)}.
Type of ggplot2 geometric object to use.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{countsAxisLabel}{\code{character(1)}.
Counts axis label.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\examples{
data(rse, sce, package = "acidData")

## SummarizedExperiment ====
plotCountsPerGene(rse, geom = "boxplot")
plotCountsPerGene(rse, geom = "density")

## SingleCellExperiment ====
plotCountsPerGene(sce)
}
