% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}
}
\value{
\code{pheatmap}.

\code{pheatmap}.
}
\description{
Construct a simple heatmap.
}
\details{
This function is a simplified version of \code{plotHeatmap()} that is optimized
for automatic handling differentially expressed genes, rather than requiring
manual input of a gene vector. All of the optional parameters for
\code{plotHeatmap} are also available to this function.

To adjust the annotation columns, modify the \code{colData} of the \code{counts}
argument, which must contain/extend a \code{SummarizedExperiment}.
}
\note{
Post hoc alpha level and LFC threshold cutoffs are not recommended.
}
\section{Hierarchical clustering}{


By default, row- and column-wise hierarchical clustering is performed using
the Ward method, but this behavior can be overrided by setting \code{clusterRows}
or \code{clusterCols} to \code{FALSE}.
}

\examples{
data(rse, sce)

## SummarizedExperiment ====
plotHeatmap(rse)
plotCorrelationHeatmap(rse)
plotQuantileHeatmap(rse)

## Disable column clustering.
plotHeatmap(rse, clusterCols = FALSE)

## Using pheatmap default colors.
plotHeatmap(rse, color = NULL, legendColor = NULL)

## Using hexadecimal color input.
library(RColorBrewer)
purple_orange <- colorRampPalette(brewer.pal(n = 11L, name = "PuOr"))(256L)
plotHeatmap(rse, color = purple_orange)

## SingleCellExperiment ====
plotHeatmap(sce)
plotCorrelationHeatmap(sce)
plotQuantileHeatmap(sce)
}
\seealso{
\itemize{
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
\item \code{\link[RColorBrewer:brewer.pal]{RColorBrewer::brewer.pal()}}.
}

\itemize{
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
}
}
