% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotDEGPCA}
\alias{plotDEGPCA}
\title{Plot differentially expressed gene principal component analysis}
\usage{
plotDEGPCA(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} or \code{DataFrame}.
}
\description{
This function is an extension of \code{plotPCA} that is optimized for automatic
handling of differentially expressed genes, rather than requiring manual
input of a gene vector or subset object. All of the optional parameters for
\code{plotPCA} are also supported by this function.
}
\details{
To adjust the annotation columns, modify the \code{colData} of the \code{counts}
argument, which must contain/extend a \code{SummarizedExperiment}.
}
\examples{
isS4(plotDEGPCA)
}
\seealso{
DESeqAnalysis.
}
