% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\title{Plot Exonic Mapping Rate}
\usage{
plotExonicMappingRate(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot Exonic Mapping Rate
}
