% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGenderMarkers-methods.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,RangedSummarizedExperiment-method}
\title{Plot Sexually Dimorphic Gender Marker Genes}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{RangedSummarizedExperiment}(object,
  style = "wide", interestingGroups = NULL,
  countsAxisLabel = "counts", medianLine = TRUE,
  color = getOption("basejump.discrete.color", NULL),
  legend = getOption("basejump.legend", TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{style}{\code{string}. Plot style.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{countsAxisLabel}{\code{string}. Label to use for the counts axis.}

\item{medianLine}{\code{boolean}. Include median line for each group. Disabled if
samples are colored by sample name.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Show plot legend.}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that wraps \code{\link[=plotGene]{plotGene()}} to quickly plot known
sexually dimorphic genes, and overall gene experssion for the X and Y
chromosomes. Currently only \emph{Homo sapiens} and \emph{Mus musculus}
genomes are supported.
}
\examples{
# bcbioRNASeq ====
object <- bcb_small
plotGenderMarkers(object, normalized = "vst")

# DESeqTransform ====
object <- as(deseq_small, "DESeqTransform")
plotGenderMarkers(object)
}
\seealso{
\code{\link[=plotGene]{plotGene()}}.
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
