% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGenderMarkers-methods.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,RangedSummarizedExperiment-method}
\title{Plot Sexually Dimorphic Gender Marker Genes}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{RangedSummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that wraps \code{\link[=plotGene]{plotGene()}} to quickly plot known
sexually dimorphic genes, and overall gene experssion for the X and Y
chromosomes. Currently only \emph{Homo sapiens} and \emph{Mus musculus}
genomes are supported.
}
\examples{
# bcbioRNASeq ====
object <- bcb_small
plotGenderMarkers(object, normalized = "vst")

# DESeqTransform ====
object <- as(deseq_small, "DESeqTransform")
plotGenderMarkers(object)
}
\seealso{
\code{\link[=plotGene]{plotGene()}}.
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
