% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,SummarizedExperiment-method}
\title{Plot Gene Expression}
\usage{
plotGene(object, ...)

\S4method{plotGene}{SummarizedExperiment}(object, genes, assay = 1L,
  interestingGroups = NULL, countsAxisLabel = "counts",
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE),
  style = c("facet", "wide"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{countsAxisLabel}{\code{string}. Label to use for the counts axis.}

\item{medianLine}{\code{boolean}. Include median line for each group. Disabled if
samples are colored by sample name.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}. To set the discrete color
palette globally, use \code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{style}{\code{string}. Plot style.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot Gene Expression
}
\examples{
object <- rse_small

rownames <- head(rownames(object))
print(rownames)
g2s <- gene2symbol(object)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

# Rownames, gene IDs, and gene names (symbols) are supported.
plotGene(object, genes = geneIDs, style = "facet")
plotGene(object, genes = geneNames, style = "wide")
}
\seealso{
Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountsPerBiotype}},
  \code{\link{plotHeatmap}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
