% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,SummarizedExperiment-method}
\title{Plot Gene Expression}
\usage{
plotGene(object, ...)

\S4method{plotGene}{SummarizedExperiment}(object, genes,
  interestingGroups = NULL, countsAxisLabel = "counts",
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE), headerLevel = 2L,
  return = c("facet", "wide"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{countsAxisLabel}{\code{string}. Label to use for the counts axis.}

\item{medianLine}{\code{boolean}. Include median line for each group. Disabled if
samples are colored by sample name.}

\item{headerLevel}{\code{scalar integer}. Markdown header level (1-7).}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\value{
\itemize{
\item "\code{facet}": \code{ggplot} grouped by \code{sampleName}, with \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}
applied to panel the samples.
\item "\code{wide}": Show \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot Gene Expression
}
\examples{
object <- rse_small

rownames <- head(rownames(object))
print(rownames)
g2s <- gene2symbol(object)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

# Rownames, gene IDs, and gene names (symbols) are supported.
plotGene(object, genes = geneIDs, return = "facet")
plotGene(object, genes = geneNames, return = "wide")
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.

Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
