% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGenesDetected-methods.R
\docType{methods}
\name{plotGenesDetected}
\alias{plotGenesDetected}
\alias{plotGenesDetected,SummarizedExperiment-method}
\alias{plotGenesDetected,SingleCellExperiment-method}
\title{Plot Genes Detected}
\usage{
plotGenesDetected(object, ...)

\S4method{plotGenesDetected}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, limit = 0L, minCounts = 1L,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), title = "genes detected")

\S4method{plotGenesDetected}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, limit = 0L, minCounts = 1L,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), title = "genes detected")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of count matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{limit}{\code{scalar numeric}. Threshold to denote on the plot, using a
dashed line.}

\item{minCounts}{\code{scalar integer}. Minimum number of counts per gene in the
count matrix.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}. To set the discrete fill palette
globally, use \code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for quality control
plots containing many samples.}

\item{title}{\code{string}. Plot title.}
}
\description{
Plot Genes Detected
}
\examples{
data(rse_small, sce_small)
plotGenesDetected(rse_small)
plotGenesDetected(sce_small)
}
\seealso{
Other Plots: \code{\link{plotCountsPerBiotype}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotGenderMarkers}}, \code{\link{plotGene}},
  \code{\link{plotHeatmap}}, \code{\link{plotQC}},
  \code{\link{plotTotalCounts}},
  \code{\link{plotZerosVsDepth}}
}
\concept{Plots}
