% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesDetected-methods.R
\docType{methods}
\name{plotGenesDetected}
\alias{plotGenesDetected}
\alias{plotGenesDetected,SummarizedExperiment-method}
\alias{plotGenesDetected,SingleCellExperiment-method}
\title{Plot genes detected}
\usage{
\S4method{plotGenesDetected}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, limit = 0L, minCounts = 1L,
  fill = getOption("basejump.fill.discrete", default = NULL),
  flip = getOption("basejump.flip", default = TRUE),
  title = "genes detected")

\S4method{plotGenesDetected}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, limit = 0L, minCounts = 1L,
  fill = getOption("basejump.fill.discrete", default = NULL),
  flip = getOption("basejump.flip", default = TRUE),
  title = "genes detected")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{limit}{\code{numeric(1)}.
Threshold to denote on the plot, using a dashed line.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per gene in the count matrix.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot genes detected
}
\examples{
data(rse, sce, package = "acidData")
plotGenesDetected(rse)
plotGenesDetected(sce)
}
