% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotHeatmap-methods.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotHeatmap,SummarizedExperiment-method}
\alias{plotHeatmap,SingleCellExperiment-method}
\alias{plotCorrelationHeatmap,SummarizedExperiment-method}
\alias{plotCorrelationHeatmap,SingleCellExperiment-method}
\alias{plotQuantileHeatmap,SummarizedExperiment-method}
\alias{plotQuantileHeatmap,SingleCellExperiment-method}
\title{Plot Heatmap}
\usage{
plotHeatmap(object, ...)

plotCorrelationHeatmap(object, ...)

plotQuantileHeatmap(object, ...)

\S4method{plotHeatmap}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, scale = c("none", "row", "column"),
  clusterRows = TRUE, clusterCols = TRUE, showRownames = FALSE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = viridis::viridis,
  borderColor = NULL, title = NULL, ...)

\S4method{plotHeatmap}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, scale = c("none", "row", "column"),
  clusterRows = TRUE, clusterCols = TRUE, showRownames = FALSE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = viridis::viridis,
  borderColor = NULL, title = NULL, ...)

\S4method{plotCorrelationHeatmap}{SummarizedExperiment}(object,
  assay = 1L, interestingGroups = NULL, method = c("pearson",
  "spearman"), clusteringMethod = "ward.D2", showRownames = TRUE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = viridis::viridis,
  borderColor = NULL, title = TRUE, ...)

\S4method{plotCorrelationHeatmap}{SingleCellExperiment}(object,
  assay = 1L, interestingGroups = NULL, method = c("pearson",
  "spearman"), clusteringMethod = "ward.D2", showRownames = TRUE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = viridis::viridis,
  borderColor = NULL, title = TRUE, ...)

\S4method{plotQuantileHeatmap}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, n = 10L, clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, color = viridis::viridis,
  legendColor = viridis::viridis, legend = FALSE, borderColor = NULL,
  title = NULL, ...)

\S4method{plotQuantileHeatmap}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, n = 10L, clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, color = viridis::viridis,
  legendColor = viridis::viridis, legend = FALSE, borderColor = NULL,
  title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{assay}{\code{scalar}. Name or index of count matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{scale}{\code{string}. Whether the values should be centered and scaled in
either the row or column direction ("\code{row}", "\code{column}"), or remain
unscaled ("\code{none}").}

\item{clusterRows, clusterCols}{\code{boolean}. Arrange with hierarchical
clustering.}

\item{showRownames, showColnames}{\code{boolean}. Show row or column names.}

\item{treeheightRow, treeheightCol}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{color}{\code{function}, \code{character}, or \code{NULL}. Hexadecimal color function
or values to use for plot. We generally recommend these hexadecimal
functions from the viridis package:
\itemize{
\item \code{\link[viridis:viridis]{viridis::viridis()}} (\emph{default}).
\item \code{\link[viridis:inferno]{viridis::inferno()}}.
\item \code{\link[viridis:magma]{viridis::magma()}}.
\item \code{\link[viridis:plasma]{viridis::plasma()}}.
Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.
}}

\item{legendColor}{\code{function} or \code{NULL}. Hexadecimal color function to use
for legend labels. Note that hexadecimal values are not supported. If set
\code{NULL}, will use the default pheatmap colors.}

\item{borderColor}{\code{string} or \code{NULL}. Border color. Disabled by default for
improved aesthetics.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{method}{\code{string}. Correlation coefficient (or covariance) method to be
computed. Defaults to "\code{pearson}" but "\code{spearman}" can also be used.
Consult the \code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{clusteringMethod}{\code{string}. Clustering method. Accepts the same values
as \code{\link[stats:hclust]{stats::hclust()}}.}

\item{n}{\code{scalar integer}. The number of quantile breaks to create.}

\item{legend}{\code{boolean}. Show the color legend.}
}
\value{
\code{pheatmap}.
}
\description{
Construct a simple heatmap.
}
\section{Hierarchical clustering}{


By default, row- and column-wise hierarchical clustering is performed using
the Ward method, but this behavior can be overrided by setting \code{clusterRows}
or \code{clusterCols} to \code{FALSE}.
}

\section{Correlation heatmap}{


Construct a correlation heatmap comparing the columns of the matrix.
}

\section{Quantile heatmap}{


Apply quantile breaks.
}

\examples{
# SummarizedExperiment ====
plotHeatmap(rse_small)
plotCorrelationHeatmap(rse_small)
plotQuantileHeatmap(rse_small)

# Disable column clustering.
plotHeatmap(rse_small, clusterCols = FALSE)

# Using pheatmap default colors.
plotHeatmap(rse_small, color = NULL, legendColor = NULL)

# Using hexadecimal color input.
library(RColorBrewer)
purple_orange <- colorRampPalette(brewer.pal(n = 11L, name = "PuOr"))(256L)
plotHeatmap(rse_small, color = purple_orange)

# SingleCellExperiment ====
plotHeatmap(sce_small)
plotCorrelationHeatmap(sce_small)
plotQuantileHeatmap(sce_small)
}
\seealso{
\itemize{
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
\item \code{\link[RColorBrewer:brewer.pal]{RColorBrewer::brewer.pal()}}.
}

\itemize{
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
}

Other Plots: \code{\link{plotCountsPerBiotype}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotQC}}, \code{\link{plotTotalCounts}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Plots}
