% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,matrix-method}
\title{Plot Heatmap}
\usage{
plotHeatmap(object, ...)

\S4method{plotHeatmap}{matrix}(object, genes = NULL, gene2symbol = TRUE,
  annotationCol = NULL, scale = "row", color = viridis::viridis(256),
  legendColor = viridis::viridis, title = NULL, quiet = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}

\item{genes}{\emph{Optional}. Character vector of gene identifiers to plot. These
must be the stable identifiers (e.g. ENSG00000000003) used on Ensembl and
not the gene symbols.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. If set \code{TRUE},
the function will attempt to automatically map gene identifiers to symbols
from Ensembl using \code{\link[=annotable]{annotable()}}. If set \code{FALSE}/\code{NULL}, then gene2symbol
mapping will be disabled. This is useful when working with a poorly
annotated genome. Alternatively, a gene2symbol \link{data.frame} can be passed
in, and must contain the columns \code{ensgene} and \code{symbol}. then the Ensembl
gene identifiers will be labeled in place of gene symbols.}

\item{annotationCol}{\emph{Optional}. \link{data.frame} that specifies the annotations
shown on the right side of the heatmap. Each row of this \link{data.frame}
defines the features of the heatmap columns.}

\item{scale}{Character indicating if the values should be centered and scaled
in either the row direction or the column direction, or none. Corresponding
values are "row", "column" and "none".}

\item{color}{Colors to use for plot. Defaults to the \code{\link[viridis:viridis]{viridis::viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[viridis:viridis]{viridis::viridis()}} palette.}

\item{title}{\emph{Optional}. Plot title.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Graphical output only.
}
\description{
Construct a heatmap by plotting of a specified set of genes. By default, row-
and column-wise hierarchical clustering is performed using the Ward method,
but this behavior can be overrided by setting \code{cluster_rows} or
\code{cluster_cols} to \code{FALSE}.
}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}
