% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,dgCMatrix-method}
\alias{plotHeatmap,matrix-method}
\title{Plot Heatmap}
\usage{
plotHeatmap(object, ...)

\S4method{plotHeatmap}{dgCMatrix}(object, scale = "row",
  annotationCol = NULL, clusterCols = TRUE, clusterRows = TRUE,
  color = viridis, legendColor = viridis, borderColor = NULL,
  title = NULL, ...)

\S4method{plotHeatmap}{matrix}(object, scale = "row", annotationCol = NULL,
  clusterCols = TRUE, clusterRows = TRUE, color = viridis,
  legendColor = viridis, borderColor = NULL, title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}

\item{scale}{Character indicating if the values should be centered and scaled
in either the row direction or the column direction, or none. Corresponding
values are "row", "column" and "none".}

\item{annotationCol}{\emph{Optional.} \link{data.frame} that defines annotation
mappings for the columns.}

\item{clusterCols}{Logical determining if columns should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{clusterRows}{Logical determining if rows should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[=viridis]{viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[=viridis]{viridis()}} palette.}

\item{borderColor}{Border color.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} return \link{list}, containing \code{gtable}.
}
\description{
Construct a simple heatmap. By default, row- and column-wise hierarchical
clustering is performed using the Ward method, but this behavior can be
overrided by setting \code{cluster_rows} or \code{cluster_cols} to \code{FALSE}.
}
\examples{
mat <- as.matrix(mtcars)
plotHeatmap(mat)
}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}
