% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotIntronicMappingRate}
\alias{plotIntronicMappingRate}
\title{Plot Intronic Mapping Rate}
\usage{
plotIntronicMappingRate(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The majority of reads should map to exons and not introns.
}
\examples{
isS4(plotIntronicMappingRate)
}
\seealso{
bcbioRNASeq.
}
