% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotMeanSD}
\alias{plotMeanSD}
\title{Plot Row Standard Deviations vs. Row Means}
\usage{
plotMeanSD(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot Row Standard Deviations vs. Row Means
}
