% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotMitoRatio}
\alias{plotMitoRatio}
\title{Plot mitochondrial transcript abundance}
\usage{
plotMitoRatio(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial transcript abundance
}
\examples{
isS4(plotMitoRatio)
}
\seealso{
\itemize{
\item bcbioSingleCell.
\item Chromium.
}
}
