% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\title{Plot Mitochondrial vs. Coding Counts}
\usage{
plotMitoVsCoding(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial vs. Coding Counts
}
