% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\title{Plot mitochondrial vs. coding counts}
\usage{
plotMitoVsCoding(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial vs. coding counts
}
\examples{
isS4(plotMitoVsCoding)
}
\seealso{
\itemize{
\item \href{http://bioinformatics.sph.harvard.edu/bcbioSingleCell/}{bcbioSingleCell}.
\item \href{https://github.com/steinbaugh/Chromium/}{Chromium}.
}
}
