% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SummarizedExperiment-method}
\title{Plot Quality Control}
\usage{
plotQC(object, ...)

\S4method{plotQC}{SummarizedExperiment}(object, assay = 1L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}
}
\description{
Plot Quality Control
}
\examples{
plotQC(rse_small)
}
\seealso{
Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountsPerBiotype}},
  \code{\link{plotGene}}, \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
