% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SummarizedExperiment-method}
\title{Plot Quality Control}
\usage{
plotQC(object, ...)

\S4method{plotQC}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot Quality Control
}
\examples{
plotQC(rse_small)
}
\seealso{
Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotGene}}, \code{\link{plotHeatmap}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
