% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotQuantileHeatmap.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,dgCMatrix-method}
\alias{plotQuantileHeatmap,matrix-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
plotQuantileHeatmap(object, ...)

\S4method{plotQuantileHeatmap}{dgCMatrix}(object, n = 5L,
  annotationCol = NULL, clusterCols = TRUE, clusterRows = TRUE,
  color = viridis, legendColor = viridis, borderColor = NULL,
  title = NULL, ...)

\S4method{plotQuantileHeatmap}{matrix}(object, n = 5L, annotationCol = NULL,
  clusterCols = TRUE, clusterRows = TRUE, color = viridis,
  legendColor = viridis, borderColor = NULL, title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}

\item{n}{The number of breaks to create.}

\item{annotationCol}{\emph{Optional.} \link{data.frame} that defines annotation
mappings for the columns.}

\item{clusterCols}{Logical determining if columns should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{clusterRows}{Logical determining if rows should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[=viridis]{viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[=viridis]{viridis()}} palette.}

\item{borderColor}{Border color.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap. Invisibly return \link{list} containing breaks and
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} \code{gtable}.
}
\description{
Plot Heatmap with Quantile Breaks
}
\examples{
mat <- matrix(1:10000, nrow = 100, ncol = 100)
plotQuantileHeatmap(mat)
}
\seealso{
Other Plot Functions: \code{\link{dynamicPlotlist}},
  \code{\link{midnightTheme}}, \code{\link{plotHeatmap}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
