% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotQuantileHeatmap-methods.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,SummarizedExperiment-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
plotQuantileHeatmap(object, ...)

\S4method{plotQuantileHeatmap}{SummarizedExperiment}(object,
  interestingGroups = NULL, n = 10L, clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, color = viridis::viridis,
  legend = FALSE, legendColor = NULL, borderColor = NULL,
  title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{n}{\code{scalar integer}. The number of breaks to create.}

\item{clusterRows}{\code{boolean}. Arrange with hierarchical
clustering.}

\item{clusterCols}{\code{boolean}. Arrange with hierarchical
clustering.}

\item{showRownames}{\code{boolean}. Show row or column names.}

\item{showColnames}{\code{boolean}. Show row or column names.}

\item{treeheightRow}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{treeheightCol}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{color}{\code{function}, \code{character}, or \code{NULL}. Hexadecimal color function
or values to use for plot. We generally recommend these hexadecimal
functions from the viridis package:
\itemize{
\item \code{\link[viridis:viridis]{viridis::viridis()}} (\emph{default}).
\item \code{\link[viridis:inferno]{viridis::inferno()}}.
\item \code{\link[viridis:magma]{viridis::magma()}}.
\item \code{\link[viridis:plasma]{viridis::plasma()}}.
Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.
}}

\item{legend}{\code{boolean}. Show the color legend.}

\item{legendColor}{\code{function} or \code{NULL}. Hexadecimal color function to use
for legend labels. Note that hexadecimal values are not supported. If set
\code{NULL}, will use the default pheatmap colors.}

\item{borderColor}{\code{string} or \code{NULL}. Border color. Disabled by default for
improved aesthetics.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{pheatmap}.
}
\description{
Construct a simple heatmap. By default, row- and column-wise hierarchical
clustering is performed using the Ward method, but this behavior can be
overrided by setting \code{clusterRows} or \code{clusterCols} to \code{FALSE}.
}
\examples{
# SummarizedExperiment ====
plotQuantileHeatmap(rse_small)

# Using viridis color palettes.
plotQuantileHeatmap(
    object = rse_small,
    color = viridis::plasma,
    legendColor = viridis::viridis
)
}
\seealso{
Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotGene}}, \code{\link{plotHeatmap}},
  \code{\link{plotQC}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Plot Functions}
