% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTotalCounts-methods.R
\docType{methods}
\name{plotTotalCounts}
\alias{plotTotalCounts}
\alias{plotTotalCounts,SummarizedExperiment-method}
\title{Plot Total Counts}
\usage{
plotTotalCounts(object, ...)

\S4method{plotTotalCounts}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, perMillion = FALSE,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), title = "total counts")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{perMillion}{\code{boolean}. Display as counts per million.}

\item{title}{\code{string}. Plot title.}
}
\description{
Plot Total Counts
}
\examples{
plotTotalCounts(bcb_small)
}
\seealso{
Other Quality Control Functions: \code{\link{plotCountsPerGene}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
