% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTotalCounts-methods.R
\docType{methods}
\name{plotTotalCounts}
\alias{plotTotalCounts}
\alias{plotTotalCounts,SummarizedExperiment-method}
\alias{plotTotalCounts,SingleCellExperiment-method}
\title{Plot Total Counts}
\usage{
plotTotalCounts(object, ...)

\S4method{plotTotalCounts}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, perMillion = FALSE,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), title = "total counts")

\S4method{plotTotalCounts}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, perMillion = FALSE,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), title = "total counts")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{perMillion}{\code{boolean}. Display as counts per million.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}. To set the discrete fill palette
globally, use \code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for quality control
plots containing many samples.}

\item{title}{\code{string}. Plot title.}
}
\description{
Plot Total Counts
}
\examples{
plotTotalCounts(rse_small)
plotTotalCounts(sce_small)
}
\seealso{
Other Quality Control Functions: \code{\link{plotCountsPerGene}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
