% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalCounts-methods.R
\docType{methods}
\name{plotTotalCounts}
\alias{plotTotalCounts}
\alias{plotTotalCounts,SummarizedExperiment-method}
\alias{plotTotalCounts,SingleCellExperiment-method}
\title{Plot total read counts}
\usage{
\S4method{plotTotalCounts}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL, perMillion = FALSE,
  fill = getOption("basejump.fill.discrete", default = NULL),
  flip = getOption("basejump.flip", default = TRUE),
  title = "total counts")

\S4method{plotTotalCounts}{SingleCellExperiment}(object, assay = 1L,
  interestingGroups = NULL, perMillion = FALSE,
  fill = getOption("basejump.fill.discrete", default = NULL),
  flip = getOption("basejump.flip", default = TRUE),
  title = "total counts")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot total read counts
}
\examples{
data(rse, sce, package = "acidData")
plotTotalCounts(rse)
plotTotalCounts(sce)
}
