% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\title{Plot UMI and gene correlation}
\usage{
plotUMIsVsGenes(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot UMI and gene correlation
}
\examples{
isS4(plotUMIsVsGenes)
}
\seealso{
\itemize{
\item \href{http://bioinformatics.sph.harvard.edu/bcbioSingleCell/}{bcbioSingleCell}.
\item \href{https://github.com/steinbaugh/Chromium/}{Chromium}.
}
}
