% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotZerosVsDepth-methods.R
\docType{methods}
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{zerosVsDepth}
\alias{zerosVsDepth,matrix-method}
\alias{zerosVsDepth,sparseMatrix-method}
\alias{plotZerosVsDepth,matrix-method}
\alias{plotZerosVsDepth,sparseMatrix-method}
\title{Plot Percentage of Zeros vs. Library Depth}
\usage{
plotZerosVsDepth(object, ...)

zerosVsDepth(object, ...)

\S4method{zerosVsDepth}{matrix}(object)

\S4method{zerosVsDepth}{sparseMatrix}(object)

\S4method{plotZerosVsDepth}{matrix}(object, title = "zeros vs. depth")

\S4method{plotZerosVsDepth}{sparseMatrix}(object,
  title = "zeros vs. depth")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{title}{\code{string}. Plot title.}
}
\value{
\code{DataFrame} or \code{ggplot}.
}
\description{
Calculate the dropout rate.
}
\examples{
# sparseMatrix ====
single_cell_counts <- counts(sce_small)
x <- zerosVsDepth(single_cell_counts)
print(x)
plotZerosVsDepth(single_cell_counts)
}
\seealso{
Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountsPerBiotype}},
  \code{\link{plotGene}}, \code{\link{plotHeatmap}},
  \code{\link{plotQC}}, \code{\link{plotQuantileHeatmap}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Plot Functions}
