% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotZerosVsDepth-methods.R
\docType{methods}
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,SummarizedExperiment-method}
\title{Plot Percentage of Zeros vs. Library Depth}
\usage{
plotZerosVsDepth(object, ...)

\S4method{plotZerosVsDepth}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL,
  color = getOption("basejump.discrete.color", NULL),
  title = "zeros vs. depth")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}. To set the discrete color
palette globally, use \code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{title}{\code{string}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Calculate and visualize the dropout rate.
}
\examples{
# SingleCellExperiment ====
plotZerosVsDepth(sce_small)
}
\seealso{
Other Plot Functions: \code{\link{plotCountsPerBiotype}},
  \code{\link{plotGene}}, \code{\link{plotHeatmap}},
  \code{\link{plotQC}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Plot Functions}
