% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZerosVsDepth-methods.R
\docType{methods}
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,SummarizedExperiment-method}
\title{Plot percentage of zeros vs. library depth}
\usage{
\S4method{plotZerosVsDepth}{SummarizedExperiment}(object, assay = 1L,
  interestingGroups = NULL,
  color = getOption("basejump.color.discrete", default = NULL),
  title = "zeros vs. depth")
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot percentage of zeros vs. library depth
}
\examples{
data(sce, package = "acidData")
plotZerosVsDepth(sce)
}
