% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-prepareAnnotable.R
\docType{methods}
\name{prepareAnnotable}
\alias{prepareAnnotable}
\alias{prepareAnnotable,data.frame-method}
\title{Prepare Annotable}
\usage{
prepareAnnotable(object, ...)

\S4method{prepareAnnotable}{data.frame}(object, dropExtraCols = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{dropExtraCols}{Drop extra columns.}
}
\value{
\link{data.frame}.
}
\description{
Standardize a user-defined annotable:
\enumerate{
\item Define \code{broadClass} column, based on \code{biotype} column.
\item Ensure rownames are set to Ensembl gene identifier.
}

This gets used automatically by the \code{\link[=annotable]{annotable()}} function, but is also
useful for standardizing a user-defined annotable, such as the output
from Stephen Turner's annotables package. The annotables package has support
for Homo sapiens GRCh37/hg19, which currently isn't supported in
AnnotationHub.
}
\author{
Broad class definitions by Rory Kirchner
}
