% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-prepareSummarizedExperiment.R
\docType{methods}
\name{prepareSummarizedExperiment}
\alias{prepareSummarizedExperiment}
\alias{prepareSummarizedExperiment}
\alias{prepareSummarizedExperiment,list-method}
\title{Prepare Summarized Experiment}
\usage{
prepareSummarizedExperiment(assays, ...)

\S4method{prepareSummarizedExperiment}{list}(assays, rowData, colData, metadata)
}
\arguments{
\item{assays}{List containing RNA-seq count matrices with matching
dimensions. Counts can be passed in either dense (\code{matrix}) or sparse
(\code{dgCMatrix}, \code{dgTMatrix}) format.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{rowData}{Object describing assay matrix rows. Must support
\code{\link[base:dim]{base::dim()}}.}

\item{colData}{Object describing assay matrix columns. Must support
\code{\link[base:dim]{base::dim()}}.}

\item{metadata}{\emph{Optional}. Metadata list.}
}
\value{
\link{SummarizedExperiment}.
}
\description{
This is a utility wrapper for \code{SummarizedExperiment()} that provides
automatic subsetting for \code{colData} and \code{rowData}.
}
\details{
This function also provides automatic metadata slotting of multiple useful
environment parameters:
\itemize{
\item \code{date}: Today's date.
\item \code{wd}: Working directory.
\item \code{utilsSessionInfo}: \code{\link[utils:sessionInfo]{utils::sessionInfo()}} return.
\item \code{devtoolsSessionInfo}: \code{\link[devtools:session_info]{devtools::session_info()}} return.
}
}
\examples{
mat <- matrix(
    seq(1L:16L),
    nrow = 4L,
    ncol = 4L,
    dimnames = list(
        c("ENSMUSG00000000001",
          "ENSMUSG00000000003",
          "ENSMUSG00000000028",
          "ENSMUSG00000000031"),
        c("sample_1",
          "sample_2",
          "sample_3",
          "sample_4")))
rowData <- data.frame(
    ensgene = c(
        "ENSMUSG00000000001",
        "ENSMUSG00000000003",
        "ENSMUSG00000000028",
        "ENSMUSG00000000031"),
    biotype = c(
        "coding",
        "coding",
        "coding",
        "coding"),
    row.names = rownames(mat))
colData <- data.frame(
    genotype = c(
        "wildtype",
        "wildtype",
        "knockout",
        "knockout"),
    age = c(3, 6, 3, 6),
    row.names = colnames(mat))
prepareSummarizedExperiment(
    assays = list(assay = mat),
    rowData = rowData,
    colData = colData)
}
\seealso{
\itemize{
\item \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}}.
\item \code{\link[base:Sys.Date]{base::Sys.Date()}}.
\item \code{\link[base:getwd]{base::getwd()}}.
\item \code{\link[utils:sessionInfo]{utils::sessionInfo()}}.
}

Other bcbio Utilities: \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAMLMetrics}}, \code{\link{sampleYAML}}
}
