% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTemplate.R
\name{prepareTemplate}
\alias{prepareTemplate}
\title{Prepare R Markdown Template File}
\usage{
prepareTemplate(package, overwrite = FALSE, ...)
}
\arguments{
\item{package}{\code{string}. Name of package containing the R Markdown template.}

\item{overwrite}{\code{boolean}. Should existing destination files be overwritten?}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{logical} indicating which files were copied.
}
\description{
If the required template dependency files aren't present, copy them from the
requested package. Existing files are not overwritten by default. This
function will copy dependency files from a requested package inside the
\code{rmarkdown/shared} directory. If a package doesn't contain this
subdirectory, the function will return an error.
}
\details{
This code is used internally by:
\itemize{
\item \code{bcbioRNASeq::prepareRNASeqTemplate()}.
\item \code{bcbioSingleCell::prepareSingleCellTemplate()}.
}
}
\examples{
# RNA-seq template
\dontrun{
prepareTemplate(package = "bcbioRNASeq")
}

# Single-cell RNA-seq template
\dontrun{
prepareTemplate(package = "bcbioSingleCell")
}
}
\author{
Michael Steinbaugh
}
\concept{Prepare Functions}
