% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-prepareTemplate.R
\docType{methods}
\name{prepareTemplate}
\alias{prepareTemplate}
\alias{prepareTemplate}
\alias{prepareTemplate,missing-method}
\alias{prepareTemplate,character-method}
\title{Download External Dependency File}
\usage{
prepareTemplate(object, ...)

\S4method{prepareTemplate}{missing}(object, package)

\S4method{prepareTemplate}{character}(object, package)
}
\arguments{
\item{object}{\emph{Optional}. File name. If \code{NULL} (default), download the
default dependency files for a new experiment.}

\item{...}{Additional arguments.}

\item{package}{Package name.}
}
\value{
No value.
}
\description{
If the required template dependency files aren't present, download latest
versions from the package website.
}
\details{
File download utility for RMarkdown knit reports.
}
\examples{
\dontrun{
prepareTemplate("setup.R", package = "bcbioSinglecell")
}
}
\seealso{
Other Infrastructure Utilities: \code{\link{detectHPC}}
}
