% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-prepareTemplate.R
\docType{methods}
\name{prepareTemplate}
\alias{prepareTemplate}
\alias{prepareTemplate,missing-method}
\alias{prepareTemplate,character-method}
\title{Prepare R Markdown Template File}
\usage{
prepareTemplate(object, ...)

\S4method{prepareTemplate}{missing}(object, sourceDir, quiet = FALSE)

\S4method{prepareTemplate}{character}(object, sourceDir, overwrite = FALSE,
  quiet = FALSE)
}
\arguments{
\item{object}{\emph{Optional}. File name. If \code{NULL} (default), download the
default dependency files for a new experiment.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{sourceDir}{Source directory, typically a URL, where the dependency
files are located.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}

\item{overwrite}{Overwrite files if they already exist. \emph{This option is used
primarily for unit testing, and is not generally recommended.}}
}
\value{
No value.
}
\description{
If the required template dependency files aren't present, download latest
versions from the package website.
}
\details{
File download utility for RMarkdown knit reports.
}
\examples{
# Load the shared files from basejump
prepareTemplate()

# Request individual files
prepareTemplate(c("setup.R", "_header.Rmd"))

# Load the shared files from bcbioSingleCell
prepareTemplate(
    sourceDir = system.file("rmarkdown/shared",
                            package = "bcbioSingleCell"))
}
\seealso{
Other Infrastructure Utilities: \code{\link{detectHPC}}
}
\concept{Infrastructure Utilities}
