% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTemplate.R
\name{prepareTemplate}
\alias{prepareTemplate}
\title{Prepare R Markdown Template File}
\usage{
prepareTemplate(package, overwrite = FALSE, sourceDir = NULL, ...)
}
\arguments{
\item{package}{\code{string}. Name of package containing the R Markdown template.}

\item{overwrite}{\code{boolean}. Should existing destination files be overwritten?}

\item{sourceDir}{\code{string}. File path to shared source file directory.
Normally this can be left \code{NULL} when working in a standard interactive
session, but is necessary when developing code in a devtools package
environment loaded with \code{devtools::load_all()}.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{logical}. Was the file copied?.
}
\description{
If the required template dependency files aren't present, copy them from the
requested package. Existing files are not overwritten by default. This
function will copy dependency files from a requested package inside the
\code{rmarkdown/shared} directory. If a package doesn't contain this
subdirectory, the function will return an error.
}
\details{
This code is used internally by:
\itemize{
\item \code{bcbioRNASeq::prepareRNASeqTemplate()}.
\item \code{bcbioSingleCell::prepareSingleCellTemplate()}.
}
}
\examples{
# RNA-seq template
\dontrun{
prepareTemplate(package = "bcbioRNASeq")
}

# Single-cell RNA-seq template
\dontrun{
prepareTemplate(package = "bcbioSingleCell")
}
}
\seealso{
Other Markdown Functions: \code{\link{markdownHeader}},
  \code{\link{markdownLink}}, \code{\link{markdownList}},
  \code{\link{markdownPlotlist}},
  \code{\link{markdownTables}}, \code{\link{markdown}}
}
\author{
Michael Steinbaugh
}
\concept{Markdown Functions}
