% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTemplate.R
\name{prepareTemplate}
\alias{prepareTemplate}
\title{Prepare R Markdown Template File}
\usage{
prepareTemplate(file = c("_output.yaml", "_footer.Rmd", "_header.Rmd",
  "_setup.R", "bibliography.bib"), sourceDir, overwrite = FALSE)
}
\arguments{
\item{file}{\code{character}. File name(s).}

\item{sourceDir}{\code{string}. Source directory path.}

\item{overwrite}{\code{boolean}. Should existing destination files be overwritten?}
}
\value{
Invisible \code{logical} indicating which files were copied.
}
\description{
If the required template dependency files aren't present, download latest
versions from the package website. Existing files are never overwritten.
}
\details{
By default, this function will create local copies of these files:
\itemize{
\item \code{_footer.Rmd}
\item \code{_header.Rmd}
\item \code{_output.yaml}
\item \code{_setup.R}
\item \code{bibliography.bib}
}

This code is used internally by:
\itemize{
\item \code{bcbioRNASeq::prepareRNASeqTemplate()}.
\item \code{bcbioSingleCell::prepareSingleCellTemplate()}.
}
}
\examples{
# RNA-seq pipeline
\dontrun{
prepareTemplate(
    sourceDir = system.file("rmarkdown/shared", package = "bcbioRNASeq")
)
}
}
\author{
Michael Steinbaugh
}
\concept{Prepare Functions}
