% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printString.R
\name{printString}
\alias{printString}
\title{Print String}
\usage{
printString(x, max = 100L)
}
\arguments{
\item{x}{\code{atomic}.}

\item{max}{\code{scalar integer}. Maximum length of vector.}
}
\value{
\code{string}.
}
\description{
Capture \code{\link[base:print]{base::print()}} output of an \code{atomic} vector. Useful for returning
informative messages inside a function.
}
\examples{
printString(c("hello", "world"))
}
\seealso{
\code{\link[base:cat]{base::cat()}}.

Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{standardizeCall}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
