% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTable.R
\name{printTable}
\alias{printTable}
\title{Smart data.frame print}
\usage{
printTable(df, rownames = FALSE, ...)
}
\arguments{
\item{df}{\code{data.frame}}

\item{rownames}{Print rownames (\code{TRUE/FALSE})}

\item{...}{Passthrough to \code{kable()} during \code{RMarkdown} rendering
with \code{knitr}}
}
\value{
\code{tibble} or \code{kable} for chunk output, depending on the call
}
\description{
Display as \code{data.frame} in RMarkdown chunk or \code{kable()} during knit
}
\author{
Michael Steinbaugh
}
\keyword{report}
