% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{createProjectDirs}
\alias{createProjectDirs}
\alias{create_project_dirs}
\alias{detectHPC}
\alias{detect_hpc}
\alias{packageProject}
\alias{package_project}
\alias{renderProject}
\alias{render_project}
\title{Project management utilities}
\usage{
createProjectDirs()

detectHPC()

packageProject(install = FALSE)

renderProject(today = TRUE)
}
\arguments{
\item{install}{Install package.}

\item{today}{Render to a subdirectory with today's date.}
}
\description{
Create necessary directory structure for an
\href{http://rmarkdown.rstudio.com}{RMarkdown} report in
\href{https://www.rstudio.com}{RStudio}.

Detect if R is running on an HPC cluster.

Package project and build website.

Render all RMarkdown files in working directory.
}
