% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{packageProj}
\alias{packageProj}
\alias{renderProj}
\alias{clearWarnings}
\title{R project management utilities.}
\usage{
packageProj(install = FALSE, test = FALSE)

renderProj(outputDir = file.path("docs", Sys.Date()), recursive = FALSE)

clearWarnings()
}
\arguments{
\item{install}{Install package.}

\item{test}{Run tests with \code{\link[devtools:test]{devtools::test()}}.}

\item{outputDir}{Output directory.}

\item{recursive}{Find files recursively.}
}
\description{
Build package and website.

Render all RMarkdown files in working directory.

Clear warnings.
}
\examples{
\dontrun{
renderProj()
}
\dontrun{
clearWarnings()
}
}
