% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{createProjectDirs}
\alias{createProjectDirs}
\alias{detectHPC}
\alias{packageProject}
\alias{renderProject}
\title{Project management utilities}
\usage{
createProjectDirs()

detectHPC()

packageProject(install = FALSE)

renderProject(output_dir = file.path("docs", Sys.Date()), recursive = FALSE)
}
\arguments{
\item{install}{Install package.}

\item{output_dir}{Output directory.}

\item{recursive}{Find files recursively.}
}
\description{
Create necessary directory structure for an
\href{http://rmarkdown.rstudio.com}{RMarkdown} report in
\href{https://www.rstudio.com}{RStudio}.

Detect if R is running on an HPC cluster.

Package project and build website.

Render all RMarkdown files in working directory.
}
