% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{createProjectDirs}
\alias{createProjectDirs}
\alias{detectHPC}
\title{Project Management Utilities}
\usage{
createProjectDirs()

detectHPC()
}
\description{
Create necessary directory structure for an
\href{http://rmarkdown.rstudio.com}{RMarkdown} report in
\href{https://www.rstudio.com}{RStudio}.

Detect if R is running on an HPC cluster.
}
