% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk-methods.R
\name{pseudobulk}
\alias{pseudobulk}
\alias{pseudobulk,SingleCellExperiment-method}
\title{Pseudobulk}
\usage{
pseudobulk(object, ...)

\S4method{pseudobulk}{SingleCellExperiment}(object, FUN)
}
\arguments{
\item{object}{Object.}

\item{FUN}{\code{function}.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with cells aggregated to samples, like a bulk
RNA-seq dataset.
}
\description{
Pseudobulk
}
\note{
Updated 2019-08-11.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
x <- pseudobulk(object)
print(x)
}
