% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedMatrix.R
\name{rankedMatrix}
\alias{rankedMatrix}
\title{Calculate a ranked matrix}
\usage{
rankedMatrix(object, MARGIN = 2L)
}
\arguments{
\item{object}{Object.}

\item{MARGIN}{a vector giving the subscripts which the function will
    be applied over.  E.g., for a matrix \code{1} indicates rows,
    \code{2} indicates columns, \code{c(1, 2)} indicates rows and
    columns. Where \code{X} has named dimnames, it can be a character
    vector selecting dimension names.}
}
\description{
Calculate a ranked matrix
}
\seealso{
Not to be confused with \code{\link[Matrix:rankMatrix]{Matrix::rankMatrix()}}.
}
