% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedMatrix.R
\name{rankedMatrix}
\alias{rankedMatrix}
\title{Calculate a ranked matrix}
\usage{
rankedMatrix(object, method = c("decreasing", "increasing",
  "bidirectional"), MARGIN = 2L)
}
\arguments{
\item{object}{Object.}

\item{MARGIN}{a vector giving the subscripts which the function will
    be applied over.  E.g., for a matrix \code{1} indicates rows,
    \code{2} indicates columns, \code{c(1, 2)} indicates rows and
    columns. Where \code{X} has named dimnames, it can be a character
    vector selecting dimension names.}

\item{decreasing}{\code{logical(1)}.
Rank the values from positive to negative.}
}
\description{
Calculate a ranked matrix
}
\note{
Ties are resolved automatically by calculating the average. See the
\code{ties.method} parameter in \code{\link[=`rank]{`rank()}}][base::rank] for details.
}
\examples{
## Order was randomized using `sample()`.
lfc <- matrix(
    data = c(
        -2.00, -1.25,  0.00,  0.25,  0.25,  0.50,  2.00,  4.00,
         2.00,  0.25, -1.25, -2.00,  0.50,  0.25,  4.00,  0.00,
         0.50,  4.00, -1.25,  0.25,  0.00,  0.25,  2.00, -2.00,
         4.00,  0.50, -2.00,  2.00,  0.25, -1.25,  0.25,  0.00
    ),
    nrow = 8L,
    ncol = 4L,
    byrow = FALSE,
    dimnames = list(
        paste0("gene", seq_len(8L)),
        paste0("contrast", seq_len(4L))
    )
)
}
\seealso{
Not to be confused with \code{\link[Matrix:rankMatrix]{Matrix::rankMatrix()}}.
}
