% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedMatrix.R
\name{rankedMatrix}
\alias{rankedMatrix}
\title{Calculate a ranked matrix}
\usage{
rankedMatrix(object, MARGIN = 2L, method = c("increasing",
  "decreasing", "bidirectional"))
}
\arguments{
\item{object}{Object.}

\item{MARGIN}{a vector giving the subscripts which the function will
    be applied over.  E.g., for a matrix \code{1} indicates rows,
    \code{2} indicates columns, \code{c(1, 2)} indicates rows and
    columns. Where \code{X} has named dimnames, it can be a character
    vector selecting dimension names.}

\item{method}{\code{character(1)}.
Rank the values in \code{increasing}, \code{decreasing}, or \code{bidirectional} order.}
}
\value{
\code{matrix}.
}
\description{
Calculate a ranked matrix
}
\note{
Ties are resolved automatically by calculating the average. See the
\code{ties.method} parameter in \code{\link[base:rank]{rank()}} for details.

Updated 2019-08-11.
}
\examples{
data(matrix_lfc, package = "acidtest")
lfc <- matrix_lfc

## Increasing (negative to positive)
rankedMatrix(lfc, method = "increasing")

## Decreasing (positive to negative)
rankedMatrix(lfc, method = "decreasing")

## Bidirectional
rankedMatrix(lfc, method = "bidirectional")
}
\seealso{
\itemize{
\item \code{\link[base:rank]{rank()}}.
\item Not to be confused with \code{\link[Matrix:rankMatrix]{Matrix::rankMatrix()}}.
}
}
