% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{rdataExtPattern}
\alias{rdataExtPattern}
\title{R data extension pattern}
\format{An object of class \code{character} of length 1.}
\usage{
rdataExtPattern
}
\description{
R data extension pattern
}
\examples{
rdataExtPattern
}
\keyword{datasets}
