% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RdTags.R
\name{RdTags}
\alias{RdTags}
\title{R Documentation Tags}
\usage{
RdTags(object)
}
\arguments{
\item{object}{\code{Rd}. R documentation, returned from \code{\link[tools:Rd_db]{tools::Rd_db()}}.}
}
\description{
Modified version of the unexported \code{tools:::RdTags()} function.
}
\examples{
db <- tools::Rd_db("base")
Rd <- db[["nrow.Rd"]]
RdTags(Rd)
}
\seealso{
Other Documentation Functions: \code{\link{parseRd}},
  \code{\link{saveRdExamples}}
}
\author{
Michael Steinbaugh
}
\concept{Documentation Functions}
