% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readDataRaw}
\alias{readDataRaw}
\alias{readYAML}
\title{Read utilities}
\usage{
readDataRaw()

readYAML(file)
}
\arguments{
\item{file}{YAML file.}
}
\description{
Read CSV files inside \code{data-raw} directory.

Read YAML file into a list.
}
