% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readDataRaw.R
\docType{methods}
\name{readDataRaw}
\alias{readDataRaw}
\alias{readDataRaw}
\alias{readDataRaw,missing-method}
\title{Read CSV Files Inside \code{data-raw/} Directory}
\usage{
readDataRaw(object, ...)

\S4method{readDataRaw}{missing}()
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
This function automatically loads raw CSVs using the readr package and
saves the data as R binary data into the \code{data/} directory.
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadData}},
  \code{\link{loadRemoteData}}, \code{\link{packageSE}},
  \code{\link{readFileByExtension}},
  \code{\link{readYAML}}, \code{\link{transmit}}

Other Data Import and Project Utilities: \code{\link{loadData}},
  \code{\link{loadRemoteData}}, \code{\link{packageSE}},
  \code{\link{readFileByExtension}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
