% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readDataVersions.R
\docType{methods}
\name{readDataVersions}
\alias{readDataVersions}
\alias{readDataVersions}
\alias{readDataVersions,character-method}
\title{Read Data Versions}
\usage{
readDataVersions(object, ...)

\S4method{readDataVersions}{character}(object, quiet = FALSE)
}
\arguments{
\item{object}{Project directory path (character vector).}

\item{...}{Object names as symbols.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Read Data Versions
}
\examples{
url <- file.path(
    "http://basejump.seq.cloud",
    "bcbio",
    "data_versions.csv")
readDataVersions(url)
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAMLMetrics}}, \code{\link{sampleYAML}}
}
\keyword{internal}
