% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readDataVersions.R
\docType{methods}
\name{readDataVersions}
\alias{readDataVersions}
\alias{readDataVersions,character-method}
\title{Read Data Versions}
\usage{
readDataVersions(object, ...)

\S4method{readDataVersions}{character}(object)
}
\arguments{
\item{object}{Project directory path (character vector).}
}
\value{
\link{data.frame}.
}
\description{
Read Data Versions
}
\seealso{
Other bcbio Utilities: \code{\link{.sampleMetadata}},
  \code{\link{prepareSummarizedExperiment}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAMLMetrics}}, \code{\link{sampleYAML}}
}
\concept{bcbio Utilities}
\keyword{internal}
