% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readFileByExtension.R
\docType{methods}
\name{readFileByExtension}
\alias{readFileByExtension}
\alias{readFileByExtension}
\alias{readFileByExtension,character-method}
\title{Read File by Extension}
\usage{
readFileByExtension(object, ...)

\S4method{readFileByExtension}{character}(object, makeNames = "camel",
  quiet = FALSE, ...)
}
\arguments{
\item{object}{File path.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{makeNames}{Make syntactically valid names. Supports \strong{\code{camel}},
\code{snake}, or \code{FALSE}.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{tibble} by default, or a sparse matrix for \code{.mtx} files.
}
\description{
Supports automatic loading of standard \code{.csv}, \code{.mtx}, \code{.tsv}, and \code{.xlsx}
files. Also supports bcbio-nextgen pipeline-specific \code{.counts}, \code{.colnames},
and \code{.rownames} files.
}
\examples{
readFileByExtension("http://basejump.seq.cloud/mtcars.csv")
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \code{\link[Matrix:readMM]{Matrix::readMM()}}: Read a MatrixMarket file.
}

Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{readGFF}}, \code{\link{readYAML}},
  \code{\link{transmit}}
}
