% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileByExtension.R
\name{readFileByExtension}
\alias{readFileByExtension}
\title{Read File by Extension}
\usage{
readFileByExtension(object, makeNames = c("camel", "snake", "upperCamel",
  "dotted"), severity = c("stop", "warning", "message", "none"), ...)
}
\arguments{
\item{object}{File path.}

\item{makeNames}{Make syntactically valid names. Supports \strong{\code{camel}},
\code{dotted}, \code{snake}, or \code{upperCamel}. This is always enforced.}

\item{severity}{Return \code{stop} (default), \code{warning}, or \code{message} if file
doesn't exist.}

\item{...}{Additional arguments.}
}
\value{
\code{tibble} by default, or a sparse matrix for \code{.mtx} files.
}
\description{
Supports automatic loading of standard \code{.csv}, \code{.mtx}, \code{.tsv}, and \code{.xlsx}
files. Also supports bcbio-nextgen pipeline-specific \code{.counts}, \code{.colnames},
and \code{.rownames} files.
}
\examples{
# Comma separated values
csv <- readFileByExtension("http://basejump.seq.cloud/mtcars.csv")
glimpse(csv)

# Microsoft Excel Worksheet
xlsx <- readFileByExtension("http://basejump.seq.cloud/mtcars.xlsx")
glimpse(xlsx)
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \code{\link[Matrix:readMM]{Matrix::readMM()}}: Read a MatrixMarket file.
}

Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{localOrRemoteFile}}, \code{\link{readGFF}},
  \code{\link{readYAML}}
}
