% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileByExtension.R
\name{readFileByExtension}
\alias{readFileByExtension}
\title{Read File by Extension}
\usage{
readFileByExtension(file, makeNames = c("camel", "snake", "upperCamel",
  "dotted"), ...)
}
\arguments{
\item{file}{File path.}

\item{makeNames}{Make syntactically valid names. Supports \strong{\code{camel}},
\code{dotted}, \code{snake}, or \code{upperCamel}. This is always enforced.}

\item{...}{Additional arguments.}
}
\value{
\code{tbl_df} by default, or a sparse matrix for \code{.mtx} files.
}
\description{
Supports automatic loading of standard \code{.csv}, \code{.mtx}, \code{.tsv}, and \code{.xlsx}
files. Also supports bcbio-nextgen pipeline-specific \code{.counts}, \code{.colnames},
and \code{.rownames} files.
}
\examples{
# Comma separated values
csv <- readFileByExtension("http://basejump.seq.cloud/mtcars.csv")
glimpse(csv)

# Microsoft Excel Worksheet
xlsx <- readFileByExtension("http://basejump.seq.cloud/mtcars.xlsx")
glimpse(xlsx)
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \href{https://cran.r-project.org/web/packages/Matrix/index.html}{Matrix}.
}

Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{readGFF}}, \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
