% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileByExtension.R
\name{readFileByExtension}
\alias{readFileByExtension}
\title{Read File by Extension}
\usage{
readFileByExtension(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
Varies depending on the file extension.
}
\description{
Supports automatic loading of common file extensions:
}
\details{
\itemize{
\item \code{csv}: Comma Separated Values.
\item \code{gff}/\code{gff3}/\code{gtf}: General Feature Format.
\item \code{mtx}: MatrixMarket sparse matrix.
\item \code{rda}/\code{RData}: R Data. Must contain a single object. Doesn't require
internal object name to match, like with the \code{\link[=loadData]{loadData()}} function.
\item \code{rds}: R Data Serialized.
\item \code{tsv} Tab Separated Values.
\item \code{xlsx}: Excel workbook.
\item \code{yaml}/\code{yml}: YAML.
}

Also supports some additional extensions commonly used with the
\href{https://bcbio-nextgen.readthedocs.io}{bcbio} pipeline:
\itemize{
\item \code{counts}: Counts \code{table}.
\item \code{colnames}: Sidecar file containing column names.
\item \code{rownames}: Sidecar file containing row names.
}

If the file format isn't supported natively (or blacklisted), the
\href{https://cran.r-project.org/web/packages/rio/index.html}{rio} package will
be used as a fallback attempt.
}
\note{
Reading a MatrixMarket ("\code{mtx}") file now requires "\code{colnames}" and
\code{"rownames"} sidecar files containing the \code{\link[=colnames]{colnames()}} and \code{\link[=rownames]{rownames()}} of
the sparse matrix. Legacy support for manual loading of these sidecar files
is provided.
}
\examples{
# R Data
x <- readFileByExtension("http://basejump.seq.cloud/rnaseq_counts.rda")
glimpse(x)

# Comma Separated Values
x <- readFileByExtension("http://basejump.seq.cloud/mtcars.csv")
glimpse(x)

# Microsoft Excel Worksheet
x <- readFileByExtension("http://basejump.seq.cloud/mtcars.xlsx")
glimpse(x)

# bcbio Counts Table
x <- readFileByExtension("http://basejump.seq.cloud/example.counts")
glimpse(x)
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \href{https://cran.r-project.org/web/packages/Matrix/index.html}{Matrix}.
}

Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{readGFF}}, \code{\link{readJSON}},
  \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
