% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aliases.R, R/methods-readGFF.R
\docType{methods}
\name{readGFF}
\alias{readGFF}
\alias{readGTF}
\alias{readGFF}
\alias{readGFF,character-method}
\title{Read GFF/GTF Annotations}
\usage{
readGFF(object, ...)

readGTF(...)

\S4method{readGFF}{character}(object, quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Read GFF/GTF Annotations
}
\details{
The GFF (General Feature Format) format consists of one line per
feature, each containing 9 columns of data, plus optional track definition
lines. The GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
readGFF("http://basejump.seq.cloud/mmusculus.gtf") \%>\% str()
}
\seealso{
\itemize{
\item \href{http://www.ensembl.org/info/website/upload/gff.html}{Ensembl}
\item \href{http://www.gencodegenes.org/gencodeformat.html}{Gencode}
}

Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}},
  \code{\link{readSampleMetadataFile}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
