% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFF.R
\name{readGFF}
\alias{readGFF}
\alias{readGTF}
\title{Read GFF/GTF Annotations}
\usage{
readGFF(file)

readGTF(file)
}
\arguments{
\item{file}{File path.}
}
\value{
\code{data.frame}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
readGFF("http://basejump.seq.cloud/mmusculus.gtf") \%>\% glimpse()
}
\seealso{
\itemize{
\item \href{http://www.ensembl.org/info/website/upload/gff.html}{Ensembl}
\item \href{http://www.gencodegenes.org/gencodeformat.html}{Gencode}
}

Other GFF Functions: \code{\link{gene2symbolFromGFF}},
  \code{\link{rowRangesFromGFF}},
  \code{\link{tx2geneFromGFF}}
}
