% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFF.R
\name{readGFF}
\alias{readGFF}
\alias{readGTF}
\title{Read GFF/GTF Annotations}
\usage{
readGFF(object)

readGTF(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}.
}
\description{
Read GFF/GTF Annotations
}
\details{
The GFF (General Feature Format) format consists of one line per
feature, each containing 9 columns of data, plus optional track definition
lines. The GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
readGFF("http://basejump.seq.cloud/mmusculus.gtf") \%>\% glimpse()
}
\seealso{
\itemize{
\item \href{http://www.ensembl.org/info/website/upload/gff.html}{Ensembl}
\item \href{http://www.gencodegenes.org/gencodeformat.html}{Gencode}
}

Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{localOrRemoteFile}},
  \code{\link{readFileByExtension}}, \code{\link{readYAML}}
}
