% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readGFF.R
\docType{methods}
\name{readGFF}
\alias{readGFF}
\alias{readGFF,character-method}
\alias{readGTF}
\title{Read GFF Annotations}
\usage{
readGFF(object, ...)

\S4method{readGFF}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame}.
}
\description{
Read GFF (General Feature Format) file.
}
\examples{
# GTF is GFF-compatible
file.path(testDataURL, "mmusculus.gtf") \%>\%
    readGFF() \%>\%
    str()
}
\seealso{
\itemize{
\item http://www.ensembl.org/info/website/upload/gff.html
\item http://www.gencodegenes.org/gencodeformat.html
}

Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{prepareSummarizedExperiment}},
  \code{\link{readFileByExtension}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
\concept{Data Import and Project Utilities}
