% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFF.R
\name{readGFF}
\alias{readGFF}
\alias{readGTF}
\title{Read GFF Annotations}
\usage{
readGFF(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{GRanges}.
}
\description{
The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines. The
GTF (General Transfer Format) is identical to GFF version 2.
}
\details{
Column names follow the
\href{https://bit.ly/2K6EBla}{conventions defined by Ensembl}.
}
\examples{
x <- readGFF("http://basejump.seq.cloud/example.gtf")
summary(x)
as.data.frame(x) \%>\% glimpse()
}
\seealso{
\itemize{
\item \href{http://www.ensembl.org/info/website/upload/gff.html}{Ensembl}
\item \href{http://www.gencodegenes.org/gencodeformat.html}{Gencode}
}

Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}},
  \code{\link{readJSON}}, \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
