% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readGTF.R
\docType{methods}
\name{readGTF}
\alias{readGTF}
\alias{readGTF,character-method}
\title{Read GTF Annotations}
\usage{
readGTF(object, ...)

\S4method{readGTF}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame}.
}
\description{
Read GTF Annotations
}
\examples{
file.path(testDataURL, "mmusculus.gtf") \%>\%
    readGTF \%>\%
    str
}
\seealso{
http://www.gencodegenes.org/gencodeformat.html

Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{prepareSummarizedExperiment}},
  \code{\link{readFileByExtension}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
\concept{Data Import and Project Utilities}
