% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readGTF.R
\docType{methods}
\name{readGTF}
\alias{readGTF}
\alias{readGTF}
\alias{readGTF,character-method}
\title{Read GTF Annotations}
\usage{
readGTF(object, ...)

\S4method{readGTF}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Read GTF Annotations
}
\examples{
file.path(testDataURL, "mmusculus.gtf") \%>\%
    readGTF \%>\%
    glimpse
}
\seealso{
http://www.gencodegenes.org/gencodeformat.html

Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{prepareSE}},
  \code{\link{readFileByExtension}},
  \code{\link{readYAML}}, \code{\link{transmit}}
}
