% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readLogFile.R
\docType{methods}
\name{readLogFile}
\alias{readLogFile}
\alias{readLogFile}
\alias{readLogFile,character-method}
\title{Read Log File}
\usage{
readLogFile(object, ...)

\S4method{readLogFile}{character}(object, quiet = FALSE)
}
\arguments{
\item{object}{Log file.}

\item{...}{Object names as symbols.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Character vector.
}
\description{
Read Log File
}
\examples{
url <- file.path(
    "http://basejump.seq.cloud",
    "bcbio",
    "bcbio-nextgen.log")
readLogFile(url) \%>\% head()
}
\keyword{internal}
