% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readYAML.R
\name{readYAML}
\alias{readYAML}
\title{Read YAML File}
\usage{
readYAML(object)
}
\arguments{
\item{object}{YAML file path.}
}
\value{
\link{list}.
}
\description{
Read YAML File
}
\examples{
# bcbioRNASeq example project summary YAML
url <- paste(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "project-summary.yaml",
    sep = "/")
yaml <- readYAML(url)
names(yaml)
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{transmit}}
}
