% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readYAML.R
\name{readYAML}
\alias{readYAML}
\title{Read YAML File}
\usage{
readYAML(file)
}
\arguments{
\item{file}{File path.}
}
\value{
\code{list}.
}
\description{
Read YAML File
}
\examples{
yaml <- readYAML("http://basejump.seq.cloud/summary.yaml")
names(yaml)
}
\seealso{
Other Read Functions: \code{\link{loadDataAsName}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{localOrRemoteFile}},
  \code{\link{readFileByExtension}}
}
