% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readYAML.R
\docType{methods}
\name{readYAML}
\alias{readYAML}
\alias{readYAML,character-method}
\title{Read YAML File}
\usage{
readYAML(object, ...)

\S4method{readYAML}{character}(object)
}
\arguments{
\item{object}{YAML file path.}

\item{...}{Additional arguments.}
}
\value{
\link{list}.
}
\description{
Read YAML File
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadData}},
  \code{\link{loadRemoteData}}, \code{\link{packageSE}},
  \code{\link{readFileByExtension}}, \code{\link{transmit}}
}
