% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readYAML.R
\docType{methods}
\name{readYAML}
\alias{readYAML}
\alias{readYAML,character-method}
\title{Read YAML File}
\usage{
readYAML(object, ...)

\S4method{readYAML}{character}(object)
}
\arguments{
\item{object}{YAML file path.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{list}.
}
\description{
Read YAML File
}
\examples{
# bcbioRNASeq example project summary YAML
yamlFile <- file.path(testDataURL, "project-summary.yaml")
yaml <- readYAML(yamlFile)
names(yaml)
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{readSampleMetadataFile}},
  \code{\link{transmit}}
}
\concept{Data Import and Project Utilities}
