% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readYAML.R
\docType{methods}
\name{readYAML}
\alias{readYAML}
\alias{readYAML,character-method}
\title{Read YAML File}
\usage{
readYAML(object, ...)

\S4method{readYAML}{character}(object)
}
\arguments{
\item{object}{YAML file path.}

\item{...}{Additional arguments.}
}
\value{
\link{list}.
}
\description{
Read YAML File
}
\examples{
# bcbioRnaseq example project summary YAML
yamlFile <- file.path(
    "https://raw.githubusercontent.com",
    "hbc",
    "bcbioRnaseq",
    "master",
    "inst",
    "extra",
    "bcbio",
    "2017-05-23_rnaseq",
    "project-summary.yaml")
yaml <- readYAML(yamlFile)
names(yaml)
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{prepareSummarizedExperiment}},
  \code{\link{readFileByExtension}}, \code{\link{readGTF}},
  \code{\link{transmit}}
}
\concept{Data Import and Project Utilities}
