% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readYAML.R
\name{readYAML}
\alias{readYAML}
\title{Read YAML File}
\usage{
readYAML(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{list}.
}
\description{
Read YAML File
}
\examples{
x <- readYAML("http://basejump.seq.cloud/example.yml")
names(x)
}
\seealso{
Other Read Functions: \code{\link{import}},
  \code{\link{loadDataAsName}}, \code{\link{loadData}},
  \code{\link{loadRemoteData}}, \code{\link{readGFF}},
  \code{\link{readJSON}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
