% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-bioverbs.R, R/reexports-brio.R,
%   R/reexports-syntactic.R, R/reexports-transformer.R, R/reexports.R
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGPCA}
\alias{atomize}
\alias{decode}
\alias{encode}
\alias{export}
\alias{factorize}
\alias{removeNA}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{camel}
\alias{capitalize}
\alias{dotted}
\alias{kebab}
\alias{snake}
\alias{upperCamel}
\alias{flatFiles}
\alias{as.data.frame}
\alias{reexports-S4}
\title{S4 generics re-exported from other packages}
\usage{
plotDEGHeatmap(object, ...)

plotDEGPCA(object, ...)

atomize(object, ...)

decode(x, ...)

encode(x, ...)

export(x, ...)

factorize(object, ...)

removeNA(object, ...)

sanitizeNA(object, ...)

sanitizePercent(object, ...)

camel(object, ...)

capitalize(object, ...)

dotted(object, ...)

kebab(object, ...)

snake(object, ...)

upperCamel(object, ...)

flatFiles(object, ...)

as.data.frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{row.names, optional}{Refer to base \code{\link[base:as.data.frame]{as.data.frame()}} for details.}

\item{object}{Object.}

\item{x}{Object.}

\item{...}{Additional arguments.}
}
\description{
These generics are imported from other packages.
}
