% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-bioverbs.R, R/reexports-brio.R,
%   R/reexports-syntactic.R, R/reexports-transformer.R, R/reexports.R
\docType{import}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGPCA}
\alias{export}
\alias{removeNA}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{camelCase}
\alias{capitalize}
\alias{dottedCase}
\alias{kebabCase}
\alias{mcolnames}
\alias{mcolnames<-}
\alias{snakeCase}
\alias{upperCamelCase}
\alias{as.data.frame}
\alias{atomize}
\alias{decode}
\alias{encode}
\alias{factorize}
\alias{flatFiles}
\alias{reexports-S4}
\title{S4 generics exported from other packages}
\usage{
plotDEGHeatmap(object, ...)

plotDEGPCA(object, ...)

export(object, ...)

removeNA(object, ...)

sanitizeNA(object, ...)

sanitizePercent(object, ...)

camelCase(object, ...)

capitalize(object, ...)

dottedCase(object, ...)

kebabCase(object, ...)

mcolnames(x, ...)

snakeCase(object, ...)

upperCamelCase(object, ...)

as.data.frame(x, row.names = NULL, optional = FALSE, ...)

atomize(object, ...)

decode(x, ...)

encode(x, ...)

factorize(object, ...)

flatFiles(object, ...)
}
\arguments{
\item{row.names, optional}{Refer to base \code{\link[base:as.data.frame]{as.data.frame()}} for details.}

\item{object, x}{Object.}

\item{...}{Additional arguments.}
}
\value{
No return.
}
\description{
These generics are imported from other packages.
}
\section{bioverbs}{

\itemize{
\item \code{plotDEGHeatmap}
\item \code{plotDEGPCA}
}
}

\section{brio}{

\itemize{
\item \code{export}
\item \code{removeNA}
\item \code{sanitizeNA}
\item \code{sanitizePercent}
}
}

\section{syntactic}{

\itemize{
\item \code{camelCase}
\item \code{capitalize}
\item \code{dottedCase}
\item \code{kebabCase}
\item \code{mcolnames}, \code{mcolnames<-}
\item \code{snakeCase}
\item \code{upperCamelCase}
}
}

\section{transformer}{

\itemize{
\item \code{as.data.frame}
\item \code{atomize}
\item \code{decode}
\item \code{encode}
\item \code{factorize}
\item \code{flatFiles}
}
}

\examples{
showMethods("camelCase")
}
\keyword{internal}
