% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-brio.R, R/reexports-syntactic.R,
%   R/reexports.R
\name{atomize}
\alias{atomize}
\alias{decode}
\alias{encode}
\alias{export}
\alias{factorize}
\alias{removeNA}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{camel}
\alias{capitalize}
\alias{dotted}
\alias{kebab}
\alias{snake}
\alias{upperCamel}
\alias{reexports-s4}
\title{S4 generics exported from other packages}
\usage{
atomize(object, ...)

decode(x, ...)

encode(x, ...)

export(x, ...)

factorize(object, ...)

removeNA(object, ...)

sanitizeNA(object, ...)

sanitizePercent(object, ...)

camel(object, ...)

capitalize(object, ...)

dotted(object, ...)

kebab(object, ...)

snake(object, ...)

upperCamel(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
These generics are imported from other packages.
}
