% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R, R/reexports-bioverbs.R,
%   R/reexports-brio.R, R/reexports-syntactic.R, R/reexports-transformer.R
\docType{import}
\name{reexports-S4}
\alias{reexports-S4}
\alias{plotDEGHeatmap}
\alias{plotDEGPCA}
\alias{export}
\alias{removeNA}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{camelCase}
\alias{capitalize}
\alias{dottedCase}
\alias{kebabCase}
\alias{mcolnames}
\alias{mcolnames<-}
\alias{snakeCase}
\alias{upperCamelCase}
\alias{as.data.frame}
\alias{atomize}
\alias{decode}
\alias{encode}
\alias{factorize}
\alias{flatFiles}
\title{S4 generics exported from other packages}
\usage{
plotDEGHeatmap(object, ...)

plotDEGPCA(object, ...)

export(object, ...)

removeNA(object, ...)

sanitizeNA(object, ...)

sanitizePercent(object, ...)

camelCase(object, ...)

capitalize(object, ...)

dottedCase(object, ...)

kebabCase(object, ...)

mcolnames(x, ...)

mcolnames(x ...) <- value

snakeCase(object, ...)

upperCamelCase(object, ...)

as.data.frame(x, row.names = NULL, optional = FALSE, ...)

atomize(object, ...)

decode(x, ...)

encode(x, ...)

factorize(object, ...)

flatFiles(object, ...)
}
\arguments{
\item{object, x}{Object.}

\item{...}{Additional arguments.}

\item{row.names, optional}{Refer to base \code{\link[base:as.data.frame]{as.data.frame()}} for details.}
}
\description{
These generics are imported from other packages.
}
\examples{
showMethods("camel")
}
\keyword{internal}
