% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-methods.R,
%   R/aggregateCellsToSamples-methods.R, R/as-methods.R,
%   R/autopadZeros-methods.R, R/cell2sample-methods.R,
%   R/coerceS4ToList.R, R/collapseToString-methods.R, R/combine-methods.R,
%   R/convertGenesToSymbols-methods.R, R/convertSampleIDsToNames-methods.R,
%   R/convertTranscriptsToGenes-methods.R, R/counts-methods.R,
%   R/decode-methods.R, R/encode-methods.R, R/geneNames-methods.R,
%   R/geometricMean-methods.R, R/globals.R, R/headtail-methods.R,
%   R/humanize-methods.R, R/interestingGroups-methods.R,
%   R/logRatio-methods.R, R/makeNames-methods.R, R/mapGenes-methods.R,
%   R/markdown-methods.R, R/meltCounts-methods.R, R/metrics-methods.R,
%   R/organism-methods.R, R/package.R, R/plotCountsPerBiotype-methods.R,
%   R/plotCountsPerGene-methods.R, R/plotGene-methods.R,
%   R/plotGenderMarkers-methods.R, R/plotGenesDetected-methods.R,
%   R/plotHeatmap-methods.R, R/plotPCA-methods.R, R/plotQC-methods.R,
%   R/plotTotalCounts-methods.R, R/plotZerosVsDepth-methods.R,
%   R/removeNA-methods.R, R/sampleData-methods.R, R/sampleNames-methods.R,
%   R/sanitizeNA-methods.R, R/sanitizePercent-methods.R,
%   R/selectSamples-methods.R, R/show-methods.R,
%   R/stripTranscriptVersions-methods.R, R/subsetPerSample-methods.R,
%   R/topCellsPerSample-methods.R, R/tpm-methods.R,
%   R/uniteInterestingGroups-methods.R, R/zerosVsDepth-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.frame}
\alias{as_tibble}
\alias{extPattern}
\alias{naStrings}
\alias{dots}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Biobase}{\code{\link[Biobase]{sampleNames}}, \code{\link[Biobase]{sampleNames<-}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{as.data.frame}}, \code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{organism}}, \code{\link[BiocGenerics]{plotPCA}}}

  \item{bioverbs}{\code{\link[bioverbs]{aggregateRows}}, \code{\link[bioverbs]{aggregateCols}}, \code{\link[bioverbs]{aggregateCellsToSamples}}, \code{\link[bioverbs]{autopadZeros}}, \code{\link[bioverbs]{cell2sample}}, \code{\link[bioverbs]{flatFiles}}, \code{\link[bioverbs]{collapseToString}}, \code{\link[bioverbs]{convertGenesToSymbols}}, \code{\link[bioverbs]{convertSymbolsToGenes}}, \code{\link[bioverbs]{convertSampleIDsToNames}}, \code{\link[bioverbs]{convertTranscriptsToGenes}}, \code{\link[bioverbs]{encode}}, \code{\link[bioverbs]{geneNames}}, \code{\link[bioverbs]{geometricMean}}, \code{\link[bioverbs]{headtail}}, \code{\link[bioverbs]{humanize}}, \code{\link[bioverbs]{interestingGroups}}, \code{\link[bioverbs]{interestingGroups<-}}, \code{\link[bioverbs]{foldChangeToLogRatio}}, \code{\link[bioverbs]{logRatioToFoldChange}}, \code{\link[bioverbs]{camel}}, \code{\link[bioverbs]{dotted}}, \code{\link[bioverbs]{snake}}, \code{\link[bioverbs]{upperCamel}}, \code{\link[bioverbs]{mapGenesToRownames}}, \code{\link[bioverbs]{mapGenesToIDs}}, \code{\link[bioverbs]{mapGenesToSymbols}}, \code{\link[bioverbs]{markdown}}, \code{\link[bioverbs]{meltCounts}}, \code{\link[bioverbs]{metrics}}, \code{\link[bioverbs]{metricsPerSample}}, \code{\link[bioverbs]{plotCountsPerBiotype}}, \code{\link[bioverbs]{plotCountsPerBroadClass}}, \code{\link[bioverbs]{plotCountsPerGene}}, \code{\link[bioverbs]{plotGene}}, \code{\link[bioverbs]{plotGenderMarkers}}, \code{\link[bioverbs]{plotGenesDetected}}, \code{\link[bioverbs]{plotHeatmap}}, \code{\link[bioverbs]{plotCorrelationHeatmap}}, \code{\link[bioverbs]{plotQuantileHeatmap}}, \code{\link[bioverbs]{plotQC}}, \code{\link[bioverbs]{plotTotalCounts}}, \code{\link[bioverbs]{plotZerosVsDepth}}, \code{\link[bioverbs]{removeNA}}, \code{\link[bioverbs]{sampleData}}, \code{\link[bioverbs]{sampleData<-}}, \code{\link[bioverbs]{sanitizeNA}}, \code{\link[bioverbs]{sanitizePercent}}, \code{\link[bioverbs]{selectSamples}}, \code{\link[bioverbs]{stripTranscriptVersions}}, \code{\link[bioverbs]{subsetPerSample}}, \code{\link[bioverbs]{topCellsPerSample}}, \code{\link[bioverbs]{uniteInterestingGroups}}, \code{\link[bioverbs]{zerosVsDepth}}}

  \item{brio}{\code{\link[brio]{extPattern}}, \code{\link[brio]{naStrings}}, \code{\link[brio]{dots}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{decode}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{tpm}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

