% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-methods.R, R/combine-methods.R,
%   R/counts-methods.R, R/decode-methods.R, R/organism-methods.R, R/package.R,
%   R/plotPCA-methods.R, R/sampleNames-methods.R, R/show-methods.R,
%   R/tpm-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.frame}
\alias{as_tibble}
\alias{capitalize}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Biobase}{\code{\link[Biobase]{sampleNames}}, \code{\link[Biobase]{sampleNames<-}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{as.data.frame}}, \code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{organism}}, \code{\link[BiocGenerics]{plotPCA}}}

  \item{goalie}{\code{\link[goalie]{capitalize}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{decode}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{tpm}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

