% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-generics.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{aggregate}
\alias{aggregateCols}
\alias{aggregateRows}
\alias{autopadZeros}
\alias{as.SummarizedExperiment}
\alias{geneNames}
\alias{mcols}
\alias{mcols<-}
\alias{metadata}
\alias{metadata<-}
\alias{Ensembl2Entrez}
\alias{Entrez2Ensembl}
\alias{Gene2Symbol}
\alias{HGNC}
\alias{HGNC2Ensembl}
\alias{MGI2Ensembl}
\alias{Tx2Gene}
\alias{appendToBody}
\alias{barcodePattern}
\alias{basenameSansExt}
\alias{classContains}
\alias{compress}
\alias{compressExtPattern}
\alias{decompress}
\alias{dots}
\alias{download}
\alias{extPattern}
\alias{fileDepth}
\alias{fileExt}
\alias{forceDetach}
\alias{formalsList}
\alias{genomeMetadataNames}
\alias{initDir}
\alias{lanePattern}
\alias{majorVersion}
\alias{metadataBlacklist}
\alias{methodFormals}
\alias{methodFunction}
\alias{metricsCols}
\alias{minorVersion}
\alias{pasteURL}
\alias{printString}
\alias{realpath}
\alias{requireNamespaces}
\alias{sanitizeVersion}
\alias{shell}
\alias{showSlotInfo}
\alias{standardizeCall}
\alias{updateMessage}
\alias{annotable}
\alias{currentEnsemblGenomeBuild}
\alias{currentEnsemblVersion}
\alias{currentFlyBaseVersion}
\alias{currentGencodeGenomeBuild}
\alias{currentGencodeVersion}
\alias{currentRefSeqGenomeBuild}
\alias{currentRefSeqVersion}
\alias{currentUCSCGenomeBuild}
\alias{currentWormBaseVersion}
\alias{detectOrganism}
\alias{downloadEnsemblGenome}
\alias{downloadGencodeGenome}
\alias{downloadRefSeqGenome}
\alias{downloadUCSCGenome}
\alias{emptyRanges}
\alias{geneSynonyms}
\alias{importTx2Gene}
\alias{makeGRangesFromEnsDb}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromGFF}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromGFF}
\alias{makeProtein2GeneFromEnsembl}
\alias{makeTx2GeneFromEnsDb}
\alias{makeTx2GeneFromEnsembl}
\alias{makeTx2GeneFromFASTA}
\alias{makeTx2GeneFromGFF}
\alias{mapEnsemblBuildToUCSC}
\alias{mapEnsemblReleaseToURL}
\alias{mapHumanOrthologs}
\alias{mapUCSCBuildToEnsembl}
\alias{organism}
\alias{organism<-}
\alias{stripGeneVersions}
\alias{stripTranscriptVersions}
\alias{antiJoin}
\alias{fullJoin}
\alias{innerJoin}
\alias{leftJoin}
\alias{mutateAll}
\alias{mutateAt}
\alias{mutateIf}
\alias{rightJoin}
\alias{selectIf}
\alias{semiJoin}
\alias{splitByLevel}
\alias{transmuteAt}
\alias{transmuteIf}
\alias{GRanges}
\alias{GRangesList}
\alias{IRanges}
\alias{DataFrame}
\alias{SimpleList}
\alias{assayNames}
\alias{assay}
\alias{assay<-}
\alias{assays}
\alias{assays<-}
\alias{colData}
\alias{colData<-}
\alias{rowData}
\alias{rowData<-}
\alias{rowRanges}
\alias{rowRanges<-}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{as.DataFrame}
\alias{as.data.frame}
\alias{as.data.table}
\alias{as_tibble}
\alias{assignAndSaveData}
\alias{atomize}
\alias{cacheURL}
\alias{column_to_rownames}
\alias{decode}
\alias{encode}
\alias{export}
\alias{factorize}
\alias{getURLDirList}
\alias{import}
\alias{loadData}
\alias{loadDataAsName}
\alias{loadRemoteData}
\alias{localOrRemoteFile}
\alias{metadata2}
\alias{metadata2<-}
\alias{naStrings}
\alias{removeNA}
\alias{rownames_to_column}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{saveData}
\alias{tibble}
\alias{transmit}
\alias{camel}
\alias{capitalize}
\alias{dotted}
\alias{kebab}
\alias{kebabCase}
\alias{makeDimnames}
\alias{makeLabel}
\alias{makeNames}
\alias{makeTitle}
\alias{makeWords}
\alias{snake}
\alias{upperCamel}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{appendToBody}}, \code{\link[AcidBase]{barcodePattern}}, \code{\link[AcidBase]{basenameSansExt}}, \code{\link[AcidBase]{classContains}}, \code{\link[AcidBase]{compress}}, \code{\link[AcidBase:reexports]{compressExtPattern}}, \code{\link[AcidBase:compress]{decompress}}, \code{\link[AcidBase]{dots}}, \code{\link[AcidBase]{download}}, \code{\link[AcidBase:reexports]{extPattern}}, \code{\link[AcidBase]{fileDepth}}, \code{\link[AcidBase]{fileExt}}, \code{\link[AcidBase]{forceDetach}}, \code{\link[AcidBase]{formalsList}}, \code{\link[AcidBase]{genomeMetadataNames}}, \code{\link[AcidBase]{initDir}}, \code{\link[AcidBase]{lanePattern}}, \code{\link[AcidBase]{majorVersion}}, \code{\link[AcidBase]{metadataBlacklist}}, \code{\link[AcidBase:methodDefinition]{methodFormals}}, \code{\link[AcidBase:methodDefinition]{methodFunction}}, \code{\link[AcidBase]{metricsCols}}, \code{\link[AcidBase]{minorVersion}}, \code{\link[AcidBase]{pasteURL}}, \code{\link[AcidBase:reexports]{printString}}, \code{\link[AcidBase]{realpath}}, \code{\link[AcidBase]{requireNamespaces}}, \code{\link[AcidBase]{sanitizeVersion}}, \code{\link[AcidBase]{shell}}, \code{\link[AcidBase]{showSlotInfo}}, \code{\link[AcidBase]{standardizeCall}}, \code{\link[AcidBase]{updateMessage}}}

  \item{AcidGenerics}{\code{\link[AcidGenerics]{aggregate}}, \code{\link[AcidGenerics]{aggregateCols}}, \code{\link[AcidGenerics]{aggregateRows}}, \code{\link[AcidGenerics]{as.SummarizedExperiment}}, \code{\link[AcidGenerics]{autopadZeros}}, \code{\link[AcidGenerics]{geneNames}}, \code{\link[AcidGenerics:reexports]{mcols}}, \code{\link[AcidGenerics:reexports]{mcols<-}}, \code{\link[AcidGenerics:reexports]{metadata}}, \code{\link[AcidGenerics:reexports]{metadata<-}}}

  \item{AcidGenomes}{\code{\link[AcidGenomes]{Ensembl2Entrez}}, \code{\link[AcidGenomes:Ensembl2Entrez]{Entrez2Ensembl}}, \code{\link[AcidGenomes]{Gene2Symbol}}, \code{\link[AcidGenomes]{HGNC}}, \code{\link[AcidGenomes]{HGNC2Ensembl}}, \code{\link[AcidGenomes]{MGI2Ensembl}}, \code{\link[AcidGenomes]{Tx2Gene}}, \code{\link[AcidGenomes:makeGRangesFromEnsembl]{annotable}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentEnsemblGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentEnsemblVersion}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentFlyBaseVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentGencodeGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentGencodeVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentRefSeqGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentRefSeqVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentUCSCGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentWormBaseVersion}}, \code{\link[AcidGenomes]{detectOrganism}}, \code{\link[AcidGenomes]{downloadEnsemblGenome}}, \code{\link[AcidGenomes]{downloadGencodeGenome}}, \code{\link[AcidGenomes]{downloadRefSeqGenome}}, \code{\link[AcidGenomes]{downloadUCSCGenome}}, \code{\link[AcidGenomes]{emptyRanges}}, \code{\link[AcidGenomes]{geneSynonyms}}, \code{\link[AcidGenomes]{importTx2Gene}}, \code{\link[AcidGenomes:makeGRangesFromEnsembl]{makeGRangesFromEnsDb}}, \code{\link[AcidGenomes]{makeGRangesFromEnsembl}}, \code{\link[AcidGenomes]{makeGRangesFromGFF}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromEnsDb}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromEnsembl}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromGFF}}, \code{\link[AcidGenomes:makeProtein2Gene]{makeProtein2GeneFromEnsembl}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromEnsDb}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromEnsembl}}, \code{\link[AcidGenomes]{makeTx2GeneFromFASTA}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromGFF}}, \code{\link[AcidGenomes:mapGenomeBuild]{mapEnsemblBuildToUCSC}}, \code{\link[AcidGenomes]{mapEnsemblReleaseToURL}}, \code{\link[AcidGenomes]{mapHumanOrthologs}}, \code{\link[AcidGenomes:mapGenomeBuild]{mapUCSCBuildToEnsembl}}, \code{\link[AcidGenomes]{organism}}, \code{\link[AcidGenomes:organism]{organism<-}}, \code{\link[AcidGenomes:stripVersions]{stripGeneVersions}}, \code{\link[AcidGenomes:stripVersions]{stripTranscriptVersions}}}

  \item{AcidPlyr}{\code{\link[AcidPlyr:join]{antiJoin}}, \code{\link[AcidPlyr:join]{fullJoin}}, \code{\link[AcidPlyr:join]{innerJoin}}, \code{\link[AcidPlyr:join]{leftJoin}}, \code{\link[AcidPlyr:mutate]{mutateAll}}, \code{\link[AcidPlyr:mutate]{mutateAt}}, \code{\link[AcidPlyr:mutate]{mutateIf}}, \code{\link[AcidPlyr:join]{rightJoin}}, \code{\link[AcidPlyr:select]{selectIf}}, \code{\link[AcidPlyr:join]{semiJoin}}, \code{\link[AcidPlyr]{splitByLevel}}, \code{\link[AcidPlyr:mutate]{transmuteAt}}, \code{\link[AcidPlyr:mutate]{transmuteIf}}}

  \item{GenomicRanges}{\code{\link[GenomicRanges:GRanges-class]{GRanges}}, \code{\link[GenomicRanges:GRangesList-class]{GRangesList}}}

  \item{IRanges}{\code{\link[IRanges:IRanges-constructor]{IRanges}}}

  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:tee]{\%T>\%}}}

  \item{pipette}{\code{\link[pipette:coerce-DataFrame]{as.DataFrame}}, \code{\link[pipette:coerce-data.frame]{as.data.frame}}, \code{\link[pipette:coerce-data.table]{as.data.table}}, \code{\link[pipette:coerce-tbl_df]{as_tibble}}, \code{\link[pipette]{assignAndSaveData}}, \code{\link[pipette]{atomize}}, \code{\link[pipette]{cacheURL}}, \code{\link[pipette:reexports]{column_to_rownames}}, \code{\link[pipette]{decode}}, \code{\link[pipette:reexports]{download}}, \code{\link[pipette]{encode}}, \code{\link[pipette]{export}}, \code{\link[pipette]{factorize}}, \code{\link[pipette]{getURLDirList}}, \code{\link[pipette]{import}}, \code{\link[pipette]{loadData}}, \code{\link[pipette]{loadDataAsName}}, \code{\link[pipette]{loadRemoteData}}, \code{\link[pipette]{localOrRemoteFile}}, \code{\link[pipette]{metadata2}}, \code{\link[pipette:metadata2]{metadata2<-}}, \code{\link[pipette]{naStrings}}, \code{\link[pipette]{removeNA}}, \code{\link[pipette:reexports]{rownames_to_column}}, \code{\link[pipette]{sanitizeNA}}, \code{\link[pipette]{sanitizePercent}}, \code{\link[pipette]{saveData}}, \code{\link[pipette:reexports]{tibble}}, \code{\link[pipette]{transmit}}}

  \item{S4Vectors}{\code{\link[S4Vectors:DataFrame-class]{DataFrame}}, \code{\link[S4Vectors:SimpleList-class]{SimpleList}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData<-}}, \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{rowRanges}}, \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{rowRanges<-}}}

  \item{syntactic}{\code{\link[syntactic:deprecated]{camel}}, \code{\link[syntactic]{capitalize}}, \code{\link[syntactic:deprecated]{dotted}}, \code{\link[syntactic:deprecated]{kebab}}, \code{\link[syntactic]{kebabCase}}, \code{\link[syntactic]{makeDimnames}}, \code{\link[syntactic]{makeLabel}}, \code{\link[syntactic]{makeNames}}, \code{\link[syntactic]{makeTitle}}, \code{\link[syntactic]{makeWords}}, \code{\link[syntactic:deprecated]{snake}}, \code{\link[syntactic:deprecated]{upperCamel}}}
}}

