% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rle-methods.R, R/as-methods.R,
%   R/combine-methods.R, R/counts-methods.R, R/organism-methods.R, R/package.R,
%   R/plotPCA-methods.R, R/sampleNames-methods.R, R/show-methods.R,
%   R/tpm-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.frame}
\alias{as_tibble}
\alias{\%>\%}
\alias{DataFrame}
\alias{isSpike}
\alias{isSpike<-}
\alias{reducedDimNames}
\alias{reducedDims}
\alias{spikeNames}
\alias{assay}
\alias{colData}
\alias{colData<-}
\alias{mcols}
\alias{rowRanges}
\alias{rowRanges<-}
\alias{seqnames}
\alias{assayNames}
\alias{assayNames<-}
\alias{assays}
\alias{assays<-}
\alias{rowData}
\alias{rowData<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Biobase}{\code{\link[Biobase]{sampleNames}}, \code{\link[Biobase]{sampleNames<-}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{as.data.frame}}, \code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{organism}}, \code{\link[BiocGenerics]{plotPCA}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{decode}}, \code{\link[S4Vectors]{DataFrame}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{isSpike}}, \code{\link[SingleCellExperiment]{isSpike<-}}, \code{\link[SingleCellExperiment]{reducedDimNames}}, \code{\link[SingleCellExperiment]{reducedDims}}, \code{\link[SingleCellExperiment]{spikeNames}}, \code{\link[SingleCellExperiment]{tpm}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment]{assay}}, \code{\link[SummarizedExperiment]{colData}}, \code{\link[SummarizedExperiment]{colData<-}}, \code{\link[SummarizedExperiment]{mcols}}, \code{\link[SummarizedExperiment]{rowRanges}}, \code{\link[SummarizedExperiment]{rowRanges<-}}, \code{\link[SummarizedExperiment]{seqnames}}, \code{\link[SummarizedExperiment]{assayNames}}, \code{\link[SummarizedExperiment]{assayNames<-}}, \code{\link[SummarizedExperiment]{assays}}, \code{\link[SummarizedExperiment]{assays<-}}, \code{\link[SummarizedExperiment]{colData}}, \code{\link[SummarizedExperiment]{colData<-}}, \code{\link[SummarizedExperiment]{rowData}}, \code{\link[SummarizedExperiment]{rowData<-}}, \code{\link[SummarizedExperiment]{rowRanges}}, \code{\link[SummarizedExperiment]{rowRanges<-}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

