% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-methods.R, R/combine-methods.R,
%   R/counts-methods.R, R/goalie.R, R/organism-methods.R, R/package.R,
%   R/sampleNames-methods.R, R/show-methods.R, R/summary-methods.R,
%   R/tpm-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.frame}
\alias{as_tibble}
\alias{areDirs}
\alias{areFiles}
\alias{areNonExisting}
\alias{areURLs}
\alias{areUniqueGeneNames}
\alias{assertAllAreNonExisting}
\alias{assertAllAreValidNames}
\alias{assertAreDirs}
\alias{assertAreFiles}
\alias{assertAreGeneAnnotations}
\alias{assertAreNonExisting}
\alias{assertAreTranscriptAnnotations}
\alias{assertAreURLs}
\alias{assertAreUniqueGeneNames}
\alias{assertAreValidNames}
\alias{assertFormalCompress}
\alias{assertHasRownames}
\alias{assertHasValidDimnames}
\alias{assertHasValidNames}
\alias{assertIsAHeaderLevel}
\alias{assertIsANumberOrNULL}
\alias{assertIsAStringOrNULL}
\alias{assertIsAlpha}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsAnIntegerOrNULL}
\alias{assertIsAnnotable}
\alias{assertIsCharacterOrNULL}
\alias{assertIsColorScaleContinuousOrNULL}
\alias{assertIsColorScaleDiscreteOrNULL}
\alias{assertIsDataFrameOrNULL}
\alias{assertIsDir}
\alias{assertIsFile}
\alias{assertIsFillScaleContinuousOrNULL}
\alias{assertIsFillScaleDiscreteOrNULL}
\alias{assertIsGFF}
\alias{assertIsGene2symbol}
\alias{assertIsHeaderLevel}
\alias{assertIsHexColorFunctionOrNULL}
\alias{assertIsImplicitInteger}
\alias{assertIsImplicitIntegerOrNULL}
\alias{assertIsStringOrNULL}
\alias{assertIsTx2gene}
\alias{assertIsURL}
\alias{hasRownames}
\alias{isAlpha}
\alias{isAnImplicitInteger}
\alias{isDir}
\alias{isFile}
\alias{isHeaderLevel}
\alias{isImplicitInteger}
\alias{isURL}
\alias{validDimnames}
\alias{validNames}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Biobase}{\code{\link[Biobase]{sampleNames}}, \code{\link[Biobase]{sampleNames<-}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{as.data.frame}}, \code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{organism}}}

  \item{goalie}{\code{\link[goalie]{areDirs}}, \code{\link[goalie]{areFiles}}, \code{\link[goalie]{areNonExisting}}, \code{\link[goalie]{areURLs}}, \code{\link[goalie]{areUniqueGeneNames}}, \code{\link[goalie]{assertAllAreNonExisting}}, \code{\link[goalie]{assertAllAreValidNames}}, \code{\link[goalie]{assertAreDirs}}, \code{\link[goalie]{assertAreFiles}}, \code{\link[goalie]{assertAreGeneAnnotations}}, \code{\link[goalie]{assertAreNonExisting}}, \code{\link[goalie]{assertAreTranscriptAnnotations}}, \code{\link[goalie]{assertAreURLs}}, \code{\link[goalie]{assertAreUniqueGeneNames}}, \code{\link[goalie]{assertAreValidNames}}, \code{\link[goalie]{assertFormalCompress}}, \code{\link[goalie]{assertHasRownames}}, \code{\link[goalie]{assertHasValidDimnames}}, \code{\link[goalie]{assertHasValidNames}}, \code{\link[goalie]{assertIsAHeaderLevel}}, \code{\link[goalie]{assertIsANumberOrNULL}}, \code{\link[goalie]{assertIsAStringOrNULL}}, \code{\link[goalie]{assertIsAlpha}}, \code{\link[goalie]{assertIsAnImplicitInteger}}, \code{\link[goalie]{assertIsAnImplicitIntegerOrNULL}}, \code{\link[goalie]{assertIsAnIntegerOrNULL}}, \code{\link[goalie]{assertIsAnnotable}}, \code{\link[goalie]{assertIsCharacterOrNULL}}, \code{\link[goalie]{assertIsColorScaleContinuousOrNULL}}, \code{\link[goalie]{assertIsColorScaleDiscreteOrNULL}}, \code{\link[goalie]{assertIsDataFrameOrNULL}}, \code{\link[goalie]{assertIsDir}}, \code{\link[goalie]{assertIsFile}}, \code{\link[goalie]{assertIsFillScaleContinuousOrNULL}}, \code{\link[goalie]{assertIsFillScaleDiscreteOrNULL}}, \code{\link[goalie]{assertIsGFF}}, \code{\link[goalie]{assertIsGene2symbol}}, \code{\link[goalie]{assertIsHeaderLevel}}, \code{\link[goalie]{assertIsHexColorFunctionOrNULL}}, \code{\link[goalie]{assertIsImplicitInteger}}, \code{\link[goalie]{assertIsImplicitIntegerOrNULL}}, \code{\link[goalie]{assertIsStringOrNULL}}, \code{\link[goalie]{assertIsTx2gene}}, \code{\link[goalie]{assertIsURL}}, \code{\link[goalie]{hasRownames}}, \code{\link[goalie]{isAlpha}}, \code{\link[goalie]{isAnImplicitInteger}}, \code{\link[goalie]{isDir}}, \code{\link[goalie]{isFile}}, \code{\link[goalie]{isHeaderLevel}}, \code{\link[goalie]{isImplicitInteger}}, \code{\link[goalie]{isURL}}, \code{\link[goalie]{validDimnames}}, \code{\link[goalie]{validNames}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{summary}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{tpm}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

