% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-coerce.R, R/reexports-generics.R,
%   R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{coerce}
\alias{aggregate}
\alias{aggregateCellsToSamples}
\alias{aggregateCols}
\alias{aggregateRows}
\alias{antiJoin}
\alias{as.DataFrame}
\alias{as.SummarizedExperiment}
\alias{as.data.frame}
\alias{atomize}
\alias{autopadZeros}
\alias{calculateMetrics}
\alias{camelCase}
\alias{capitalize}
\alias{cell2sample}
\alias{coerceToList}
\alias{collapseToString}
\alias{combine}
\alias{convertGenesToSymbols}
\alias{convertSampleIDsToNames}
\alias{convertSymbolsToGenes}
\alias{convertTranscriptsToGenes}
\alias{correlation}
\alias{counts}
\alias{counts<-}
\alias{decode}
\alias{dottedCase}
\alias{droplevels}
\alias{encode}
\alias{estimateSizeFactors}
\alias{export}
\alias{factorize}
\alias{filterCells}
\alias{foldChangeToLogRatio}
\alias{fullJoin}
\alias{geneNames}
\alias{geometricMean}
\alias{headtail}
\alias{humanize}
\alias{innerJoin}
\alias{integerCounts}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{intersectAll}
\alias{intersectionMatrix}
\alias{kebabCase}
\alias{leftJoin}
\alias{logRatioToFoldChange}
\alias{makeDimnames}
\alias{makeLabel}
\alias{makeNames}
\alias{makeSampleData}
\alias{makeTitle}
\alias{makeWords}
\alias{mapGenesToIDs}
\alias{mapGenesToRownames}
\alias{mapGenesToSymbols}
\alias{matchSampleColumn}
\alias{mcols}
\alias{mcols<-}
\alias{melt}
\alias{metadata}
\alias{metadata<-}
\alias{metadata2}
\alias{metadata2<-}
\alias{metrics}
\alias{metricsPerSample}
\alias{mutateAll}
\alias{mutateAt}
\alias{mutateIf}
\alias{nonzeroRowsAndCols}
\alias{organism}
\alias{organism<-}
\alias{rankedMatrix}
\alias{removeNA}
\alias{rightJoin}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{selectIf}
\alias{selectSamples}
\alias{sem}
\alias{semiJoin}
\alias{showHeader}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{snakeCase}
\alias{splitByLevel}
\alias{stripGeneVersions}
\alias{stripTranscriptVersions}
\alias{subsetPerSample}
\alias{topCellsPerSample}
\alias{tpm}
\alias{transmuteAt}
\alias{transmuteIf}
\alias{uniteInterestingGroups}
\alias{unlistToDataFrame}
\alias{upperCamelCase}
\alias{zerosVsDepth}
\alias{Ensembl2Entrez}
\alias{Entrez2Ensembl}
\alias{Gene2Symbol}
\alias{Tx2Gene}
\alias{appendToBody}
\alias{barcodePattern}
\alias{basenameSansExt}
\alias{classContains}
\alias{compress}
\alias{compressExtPattern}
\alias{decompress}
\alias{dots}
\alias{download}
\alias{extPattern}
\alias{fileDepth}
\alias{fileExt}
\alias{forceDetach}
\alias{formalsList}
\alias{genomeMetadataNames}
\alias{initDir}
\alias{lanePattern}
\alias{majorVersion}
\alias{metadataBlacklist}
\alias{methodFormals}
\alias{methodFunction}
\alias{metricsCols}
\alias{minorVersion}
\alias{pasteURL}
\alias{printString}
\alias{realpath}
\alias{requireNamespaces}
\alias{sanitizeVersion}
\alias{shell}
\alias{showSlotInfo}
\alias{standardizeCall}
\alias{updateMessage}
\alias{alert}
\alias{alertDanger}
\alias{alertInfo}
\alias{alertSuccess}
\alias{alertWarning}
\alias{dl}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{h7}
\alias{ol}
\alias{separator}
\alias{toInlineString}
\alias{txt}
\alias{ul}
\alias{verbatim}
\alias{DataFrame}
\alias{Rle}
\alias{SimpleList}
\alias{SplitDataFrameList}
\alias{SummarizedExperiment}
\alias{assay}
\alias{assay<-}
\alias{assayNames}
\alias{assayNames<-}
\alias{assays}
\alias{assays<-}
\alias{colData}
\alias{colData<-}
\alias{detectLanes}
\alias{importSampleData}
\alias{makeSummarizedExperiment}
\alias{matchInterestingGroups}
\alias{matchesGene2Symbol}
\alias{matchesInterestingGroups}
\alias{microplate}
\alias{minimalSampleData}
\alias{rowData}
\alias{rowData<-}
\alias{rowRanges}
\alias{rowRanges<-}
\alias{sanitizeSampleData}
\alias{session_info}
\alias{HGNC}
\alias{HGNC2Ensembl}
\alias{MGI2Ensembl}
\alias{annotable}
\alias{currentEnsemblGenomeBuild}
\alias{currentEnsemblVersion}
\alias{currentFlyBaseVersion}
\alias{currentGencodeGenomeBuild}
\alias{currentGencodeVersion}
\alias{currentRefSeqGenomeBuild}
\alias{currentRefSeqVersion}
\alias{currentUCSCGenomeBuild}
\alias{currentWormBaseVersion}
\alias{detectOrganism}
\alias{downloadEnsemblGenome}
\alias{downloadGencodeGenome}
\alias{downloadRefSeqGenome}
\alias{downloadUCSCGenome}
\alias{emptyRanges}
\alias{geneSynonyms}
\alias{importTx2Gene}
\alias{makeGRangesFromEnsDb}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromGFF}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromGFF}
\alias{makeProtein2GeneFromEnsembl}
\alias{makeTx2GeneFromEnsDb}
\alias{makeTx2GeneFromEnsembl}
\alias{makeTx2GeneFromFASTA}
\alias{makeTx2GeneFromGFF}
\alias{mapEnsemblBuildToUCSC}
\alias{mapEnsemblReleaseToURL}
\alias{mapHumanOrthologs}
\alias{mapUCSCBuildToEnsembl}
\alias{markdownHeader}
\alias{markdownLink}
\alias{markdownList}
\alias{markdownPlots}
\alias{markdownTables}
\alias{mdHeader}
\alias{mdLink}
\alias{mdList}
\alias{mdPlots}
\alias{mdTables}
\alias{prepareTemplate}
\alias{SingleCellExperiment}
\alias{makeSingleCellExperiment}
\alias{mapCellsToSamples}
\alias{reducedDim}
\alias{reducedDim<-}
\alias{reducedDimNames}
\alias{reducedDimNames<-}
\alias{reducedDims}
\alias{reducedDims<-}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{as.data.table}
\alias{as_tibble}
\alias{assignAndSaveData}
\alias{cacheURL}
\alias{column_to_rownames}
\alias{getURLDirList}
\alias{import}
\alias{loadData}
\alias{loadDataAsName}
\alias{loadRemoteData}
\alias{localOrRemoteFile}
\alias{naStrings}
\alias{rownames_to_column}
\alias{saveData}
\alias{tibble}
\alias{transmit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{appendToBody}}, \code{\link[AcidBase]{barcodePattern}}, \code{\link[AcidBase]{basenameSansExt}}, \code{\link[AcidBase]{classContains}}, \code{\link[AcidBase]{compress}}, \code{\link[AcidBase:reexports]{compressExtPattern}}, \code{\link[AcidBase:compress]{decompress}}, \code{\link[AcidBase]{dots}}, \code{\link[AcidBase]{download}}, \code{\link[AcidBase:reexports]{extPattern}}, \code{\link[AcidBase]{fileDepth}}, \code{\link[AcidBase]{fileExt}}, \code{\link[AcidBase]{forceDetach}}, \code{\link[AcidBase]{formalsList}}, \code{\link[AcidBase]{genomeMetadataNames}}, \code{\link[AcidBase]{initDir}}, \code{\link[AcidBase]{lanePattern}}, \code{\link[AcidBase]{majorVersion}}, \code{\link[AcidBase]{metadataBlacklist}}, \code{\link[AcidBase:methodDefinition]{methodFormals}}, \code{\link[AcidBase:methodDefinition]{methodFunction}}, \code{\link[AcidBase]{metricsCols}}, \code{\link[AcidBase]{minorVersion}}, \code{\link[AcidBase]{pasteURL}}, \code{\link[AcidBase:reexports]{printString}}, \code{\link[AcidBase]{realpath}}, \code{\link[AcidBase]{requireNamespaces}}, \code{\link[AcidBase]{sanitizeVersion}}, \code{\link[AcidBase]{shell}}, \code{\link[AcidBase]{showSlotInfo}}, \code{\link[AcidBase]{standardizeCall}}, \code{\link[AcidBase]{updateMessage}}}

  \item{AcidCLI}{\code{\link[AcidCLI]{alert}}, \code{\link[AcidCLI:alert]{alertDanger}}, \code{\link[AcidCLI:alert]{alertInfo}}, \code{\link[AcidCLI:alert]{alertSuccess}}, \code{\link[AcidCLI:alert]{alertWarning}}, \code{\link[AcidCLI:list]{dl}}, \code{\link[AcidCLI:header]{h1}}, \code{\link[AcidCLI:header]{h2}}, \code{\link[AcidCLI:header]{h3}}, \code{\link[AcidCLI:header]{h4}}, \code{\link[AcidCLI:header]{h5}}, \code{\link[AcidCLI:header]{h6}}, \code{\link[AcidCLI:header]{h7}}, \code{\link[AcidCLI:list]{ol}}, \code{\link[AcidCLI]{separator}}, \code{\link[AcidCLI]{toInlineString}}, \code{\link[AcidCLI:text]{txt}}, \code{\link[AcidCLI:list]{ul}}, \code{\link[AcidCLI:text]{verbatim}}}

  \item{AcidExperiment}{\code{\link[AcidExperiment:reexports]{DataFrame}}, \code{\link[AcidExperiment:reexports]{Rle}}, \code{\link[AcidExperiment:reexports]{SimpleList}}, \code{\link[AcidExperiment:reexports]{SplitDataFrameList}}, \code{\link[AcidExperiment:reexports]{SummarizedExperiment}}, \code{\link[AcidExperiment:reexports]{assay}}, \code{\link[AcidExperiment:reexports]{assay<-}}, \code{\link[AcidExperiment:reexports]{assayNames}}, \code{\link[AcidExperiment:reexports]{assayNames<-}}, \code{\link[AcidExperiment:reexports]{assays}}, \code{\link[AcidExperiment:reexports]{assays<-}}, \code{\link[AcidExperiment:reexports]{colData}}, \code{\link[AcidExperiment:reexports]{colData<-}}, \code{\link[AcidExperiment]{detectLanes}}, \code{\link[AcidExperiment]{importSampleData}}, \code{\link[AcidExperiment]{makeSummarizedExperiment}}, \code{\link[AcidExperiment]{matchInterestingGroups}}, \code{\link[AcidExperiment]{matchesGene2Symbol}}, \code{\link[AcidExperiment]{matchesInterestingGroups}}, \code{\link[AcidExperiment]{microplate}}, \code{\link[AcidExperiment]{minimalSampleData}}, \code{\link[AcidExperiment:reexports]{rowData}}, \code{\link[AcidExperiment:reexports]{rowData<-}}, \code{\link[AcidExperiment:reexports]{rowRanges}}, \code{\link[AcidExperiment:reexports]{rowRanges<-}}, \code{\link[AcidExperiment]{sanitizeSampleData}}, \code{\link[AcidExperiment:reexports]{session_info}}}

  \item{AcidGenerics}{\code{\link[AcidGenerics]{aggregate}}, \code{\link[AcidGenerics]{aggregateCellsToSamples}}, \code{\link[AcidGenerics]{aggregateCols}}, \code{\link[AcidGenerics]{aggregateRows}}, \code{\link[AcidGenerics:join]{antiJoin}}, \code{\link[AcidGenerics]{as.DataFrame}}, \code{\link[AcidGenerics]{as.SummarizedExperiment}}, \code{\link[AcidGenerics]{as.data.frame}}, \code{\link[AcidGenerics]{atomize}}, \code{\link[AcidGenerics]{autopadZeros}}, \code{\link[AcidGenerics]{calculateMetrics}}, \code{\link[AcidGenerics]{camelCase}}, \code{\link[AcidGenerics]{capitalize}}, \code{\link[AcidGenerics]{cell2sample}}, \code{\link[AcidGenerics]{coerceToList}}, \code{\link[AcidGenerics]{collapseToString}}, \code{\link[AcidGenerics]{combine}}, \code{\link[AcidGenerics]{convertGenesToSymbols}}, \code{\link[AcidGenerics]{convertSampleIDsToNames}}, \code{\link[AcidGenerics]{convertSymbolsToGenes}}, \code{\link[AcidGenerics]{convertTranscriptsToGenes}}, \code{\link[AcidGenerics]{correlation}}, \code{\link[AcidGenerics]{counts}}, \code{\link[AcidGenerics:counts]{counts<-}}, \code{\link[AcidGenerics]{decode}}, \code{\link[AcidGenerics]{dottedCase}}, \code{\link[AcidGenerics]{droplevels}}, \code{\link[AcidGenerics]{encode}}, \code{\link[AcidGenerics]{estimateSizeFactors}}, \code{\link[AcidGenerics]{export}}, \code{\link[AcidGenerics]{factorize}}, \code{\link[AcidGenerics]{filterCells}}, \code{\link[AcidGenerics]{foldChangeToLogRatio}}, \code{\link[AcidGenerics:join]{fullJoin}}, \code{\link[AcidGenerics]{geneNames}}, \code{\link[AcidGenerics]{geometricMean}}, \code{\link[AcidGenerics]{headtail}}, \code{\link[AcidGenerics]{humanize}}, \code{\link[AcidGenerics:join]{innerJoin}}, \code{\link[AcidGenerics]{integerCounts}}, \code{\link[AcidGenerics]{interestingGroups}}, \code{\link[AcidGenerics:interestingGroups]{interestingGroups<-}}, \code{\link[AcidGenerics]{intersectAll}}, \code{\link[AcidGenerics]{intersectionMatrix}}, \code{\link[AcidGenerics]{kebabCase}}, \code{\link[AcidGenerics:join]{leftJoin}}, \code{\link[AcidGenerics]{logRatioToFoldChange}}, \code{\link[AcidGenerics]{makeDimnames}}, \code{\link[AcidGenerics]{makeLabel}}, \code{\link[AcidGenerics]{makeNames}}, \code{\link[AcidGenerics]{makeSampleData}}, \code{\link[AcidGenerics]{makeTitle}}, \code{\link[AcidGenerics]{makeWords}}, \code{\link[AcidGenerics]{mapGenesToIDs}}, \code{\link[AcidGenerics]{mapGenesToRownames}}, \code{\link[AcidGenerics]{mapGenesToSymbols}}, \code{\link[AcidGenerics]{matchSampleColumn}}, \code{\link[AcidGenerics:reexports]{mcols}}, \code{\link[AcidGenerics:reexports]{mcols<-}}, \code{\link[AcidGenerics]{melt}}, \code{\link[AcidGenerics:reexports]{metadata}}, \code{\link[AcidGenerics]{metadata2}}, \code{\link[AcidGenerics:metadata2]{metadata2<-}}, \code{\link[AcidGenerics:reexports]{metadata<-}}, \code{\link[AcidGenerics]{metrics}}, \code{\link[AcidGenerics]{metricsPerSample}}, \code{\link[AcidGenerics:mutate]{mutateAll}}, \code{\link[AcidGenerics:mutate]{mutateAt}}, \code{\link[AcidGenerics:mutate]{mutateIf}}, \code{\link[AcidGenerics]{nonzeroRowsAndCols}}, \code{\link[AcidGenerics]{organism}}, \code{\link[AcidGenerics:organism]{organism<-}}, \code{\link[AcidGenerics]{rankedMatrix}}, \code{\link[AcidGenerics]{removeNA}}, \code{\link[AcidGenerics:join]{rightJoin}}, \code{\link[AcidGenerics]{sampleData}}, \code{\link[AcidGenerics:sampleData]{sampleData<-}}, \code{\link[AcidGenerics]{sampleNames}}, \code{\link[AcidGenerics:sampleNames]{sampleNames<-}}, \code{\link[AcidGenerics]{sanitizeNA}}, \code{\link[AcidGenerics]{sanitizePercent}}, \code{\link[AcidGenerics:select]{selectIf}}, \code{\link[AcidGenerics]{selectSamples}}, \code{\link[AcidGenerics]{sem}}, \code{\link[AcidGenerics:join]{semiJoin}}, \code{\link[AcidGenerics]{showHeader}}, \code{\link[AcidGenerics]{sizeFactors}}, \code{\link[AcidGenerics:sizeFactors]{sizeFactors<-}}, \code{\link[AcidGenerics]{snakeCase}}, \code{\link[AcidGenerics]{splitByLevel}}, \code{\link[AcidGenerics]{stripGeneVersions}}, \code{\link[AcidGenerics]{stripTranscriptVersions}}, \code{\link[AcidGenerics]{subsetPerSample}}, \code{\link[AcidGenerics]{topCellsPerSample}}, \code{\link[AcidGenerics]{tpm}}, \code{\link[AcidGenerics:mutate]{transmuteAt}}, \code{\link[AcidGenerics:mutate]{transmuteIf}}, \code{\link[AcidGenerics]{uniteInterestingGroups}}, \code{\link[AcidGenerics]{unlistToDataFrame}}, \code{\link[AcidGenerics]{upperCamelCase}}, \code{\link[AcidGenerics]{zerosVsDepth}}}

  \item{AcidGenomes}{\code{\link[AcidGenomes]{Ensembl2Entrez}}, \code{\link[AcidGenomes:Ensembl2Entrez]{Entrez2Ensembl}}, \code{\link[AcidGenomes]{Gene2Symbol}}, \code{\link[AcidGenomes]{HGNC}}, \code{\link[AcidGenomes]{HGNC2Ensembl}}, \code{\link[AcidGenomes]{MGI2Ensembl}}, \code{\link[AcidGenomes]{Tx2Gene}}, \code{\link[AcidGenomes:makeGRangesFromEnsembl]{annotable}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentEnsemblGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentEnsemblVersion}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentFlyBaseVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentGencodeGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentGencodeVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentRefSeqGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentRefSeqVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentUCSCGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentWormBaseVersion}}, \code{\link[AcidGenomes]{detectOrganism}}, \code{\link[AcidGenomes]{downloadEnsemblGenome}}, \code{\link[AcidGenomes]{downloadGencodeGenome}}, \code{\link[AcidGenomes]{downloadRefSeqGenome}}, \code{\link[AcidGenomes]{downloadUCSCGenome}}, \code{\link[AcidGenomes]{emptyRanges}}, \code{\link[AcidGenomes]{geneSynonyms}}, \code{\link[AcidGenomes]{importTx2Gene}}, \code{\link[AcidGenomes:makeGRangesFromEnsembl]{makeGRangesFromEnsDb}}, \code{\link[AcidGenomes]{makeGRangesFromEnsembl}}, \code{\link[AcidGenomes]{makeGRangesFromGFF}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromEnsDb}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromEnsembl}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromGFF}}, \code{\link[AcidGenomes:makeProtein2Gene]{makeProtein2GeneFromEnsembl}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromEnsDb}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromEnsembl}}, \code{\link[AcidGenomes]{makeTx2GeneFromFASTA}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromGFF}}, \code{\link[AcidGenomes:mapGenomeBuild]{mapEnsemblBuildToUCSC}}, \code{\link[AcidGenomes]{mapEnsemblReleaseToURL}}, \code{\link[AcidGenomes]{mapHumanOrthologs}}, \code{\link[AcidGenomes:mapGenomeBuild]{mapUCSCBuildToEnsembl}}}

  \item{AcidMarkdown}{\code{\link[AcidMarkdown]{markdownHeader}}, \code{\link[AcidMarkdown]{markdownLink}}, \code{\link[AcidMarkdown]{markdownList}}, \code{\link[AcidMarkdown]{markdownPlots}}, \code{\link[AcidMarkdown]{markdownTables}}, \code{\link[AcidMarkdown:markdownHeader]{mdHeader}}, \code{\link[AcidMarkdown:markdownLink]{mdLink}}, \code{\link[AcidMarkdown:markdownList]{mdList}}, \code{\link[AcidMarkdown:markdownPlots]{mdPlots}}, \code{\link[AcidMarkdown:markdownTables]{mdTables}}, \code{\link[AcidMarkdown]{prepareTemplate}}}

  \item{AcidSingleCell}{\code{\link[AcidSingleCell:reexports]{SingleCellExperiment}}, \code{\link[AcidSingleCell]{makeSingleCellExperiment}}, \code{\link[AcidSingleCell]{mapCellsToSamples}}, \code{\link[AcidSingleCell:reexports]{reducedDim}}, \code{\link[AcidSingleCell:reexports]{reducedDim<-}}, \code{\link[AcidSingleCell:reexports]{reducedDimNames}}, \code{\link[AcidSingleCell:reexports]{reducedDimNames<-}}, \code{\link[AcidSingleCell:reexports]{reducedDims}}, \code{\link[AcidSingleCell:reexports]{reducedDims<-}}}

  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:tee]{\%T>\%}}}

  \item{methods}{\code{\link[methods:setAs]{coerce}}}

  \item{pipette}{\code{\link[pipette:coerce]{as.data.table}}, \code{\link[pipette:coerce]{as_tibble}}, \code{\link[pipette]{assignAndSaveData}}, \code{\link[pipette]{cacheURL}}, \code{\link[pipette:reexports]{column_to_rownames}}, \code{\link[pipette]{getURLDirList}}, \code{\link[pipette]{import}}, \code{\link[pipette]{loadData}}, \code{\link[pipette]{loadDataAsName}}, \code{\link[pipette]{loadRemoteData}}, \code{\link[pipette]{localOrRemoteFile}}, \code{\link[pipette]{naStrings}}, \code{\link[pipette:reexports]{rownames_to_column}}, \code{\link[pipette]{saveData}}, \code{\link[pipette:reexports]{tibble}}, \code{\link[pipette]{transmit}}}
}}

