% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plotGene}
\alias{relevel}
\alias{GRanges}
\alias{GRangesList}
\alias{IRanges}
\alias{DataFrame}
\alias{SimpleList}
\alias{mcols}
\alias{mcols<-}
\alias{metadata}
\alias{metadata<-}
\alias{reducedDims}
\alias{reducedDims<-}
\alias{assayNames}
\alias{assay}
\alias{assay<-}
\alias{assays}
\alias{assays<-}
\alias{colData}
\alias{colData<-}
\alias{rowData}
\alias{rowData<-}
\alias{rowRanges}
\alias{rowRanges<-}
\alias{assignAndSaveData}
\alias{basenameSansExt}
\alias{bunzip2}
\alias{bzip2}
\alias{compressFile}
\alias{decompressFile}
\alias{dots}
\alias{export}
\alias{fileExt}
\alias{gunzip}
\alias{gzip}
\alias{import}
\alias{initDir}
\alias{loadData}
\alias{loadDataAsName}
\alias{loadRemoteData}
\alias{localOrRemoteFile}
\alias{naStrings}
\alias{pasteURL}
\alias{realpath}
\alias{removeNA}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{saveData}
\alias{tar}
\alias{transmit}
\alias{untar}
\alias{unzip}
\alias{writeCounts}
\alias{zip}
\alias{annotable}
\alias{convertUCSCBuildToEnsembl}
\alias{detectOrganism}
\alias{emptyRanges}
\alias{makeGRangesFromEnsDb}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\alias{matchArgsToDoCall}
\alias{methodFormals}
\alias{methodFunction}
\alias{printString}
\alias{standardizeCall}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{camel}
\alias{capitalize}
\alias{dotted}
\alias{kebab}
\alias{kebabCase}
\alias{makeDimnames}
\alias{makeLabel}
\alias{makeNames}
\alias{makeTitle}
\alias{makeWords}
\alias{snake}
\alias{upperCamel}
\alias{tibble}
\alias{antiJoin}
\alias{as.SummarizedExperiment}
\alias{as.data.frame}
\alias{as.data.table}
\alias{as_tibble}
\alias{atomize}
\alias{coerceS4ToList}
\alias{decode}
\alias{encode}
\alias{factorize}
\alias{flatFiles}
\alias{fullJoin}
\alias{innerJoin}
\alias{leftJoin}
\alias{mutateAll}
\alias{mutateAt}
\alias{mutateIf}
\alias{rightJoin}
\alias{selectIf}
\alias{semiJoin}
\alias{transmuteAt}
\alias{transmuteIf}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{plotGene}}}

  \item{brio}{\code{\link[brio]{assignAndSaveData}}, \code{\link[brio]{basenameSansExt}}, \code{\link[brio]{bunzip2}}, \code{\link[brio]{bzip2}}, \code{\link[brio]{compressFile}}, \code{\link[brio]{decompressFile}}, \code{\link[brio]{dots}}, \code{\link[brio]{export}}, \code{\link[brio]{fileExt}}, \code{\link[brio]{gunzip}}, \code{\link[brio]{gzip}}, \code{\link[brio]{import}}, \code{\link[brio]{initDir}}, \code{\link[brio]{loadData}}, \code{\link[brio]{loadDataAsName}}, \code{\link[brio]{loadRemoteData}}, \code{\link[brio]{localOrRemoteFile}}, \code{\link[brio]{naStrings}}, \code{\link[brio]{pasteURL}}, \code{\link[brio]{realpath}}, \code{\link[brio]{removeNA}}, \code{\link[brio]{sanitizeNA}}, \code{\link[brio]{sanitizePercent}}, \code{\link[brio]{saveData}}, \code{\link[brio]{tar}}, \code{\link[brio]{transmit}}, \code{\link[brio]{untar}}, \code{\link[brio]{unzip}}, \code{\link[brio]{writeCounts}}, \code{\link[brio]{zip}}}

  \item{freerange}{\code{\link[freerange]{annotable}}, \code{\link[freerange]{convertUCSCBuildToEnsembl}}, \code{\link[freerange]{detectOrganism}}, \code{\link[freerange]{emptyRanges}}, \code{\link[freerange]{makeGRangesFromEnsDb}}, \code{\link[freerange]{makeGRangesFromEnsembl}}, \code{\link[freerange]{makeGRangesFromGFF}}, \code{\link[freerange]{makeGRangesFromGTF}}}

  \item{GenomicRanges}{\code{\link[GenomicRanges]{GRanges}}, \code{\link[GenomicRanges]{GRangesList}}}

  \item{goalie}{\code{\link[goalie]{matchArgsToDoCall}}, \code{\link[goalie]{methodFormals}}, \code{\link[goalie]{methodFunction}}, \code{\link[goalie]{printString}}, \code{\link[goalie]{standardizeCall}}}

  \item{IRanges}{\code{\link[IRanges]{IRanges}}}

  \item{magrittr}{\code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%$\%}}, \code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%T>\%}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{SimpleList}}, \code{\link[S4Vectors]{mcols}}, \code{\link[S4Vectors]{mcols<-}}, \code{\link[S4Vectors]{metadata}}, \code{\link[S4Vectors]{metadata<-}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{reducedDims}}, \code{\link[SingleCellExperiment]{reducedDims<-}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment]{assayNames}}, \code{\link[SummarizedExperiment]{assay}}, \code{\link[SummarizedExperiment]{assay<-}}, \code{\link[SummarizedExperiment]{assays}}, \code{\link[SummarizedExperiment]{assays<-}}, \code{\link[SummarizedExperiment]{colData}}, \code{\link[SummarizedExperiment]{colData<-}}, \code{\link[SummarizedExperiment]{rowData}}, \code{\link[SummarizedExperiment]{rowData<-}}, \code{\link[SummarizedExperiment]{rowRanges}}, \code{\link[SummarizedExperiment]{rowRanges<-}}}

  \item{syntactic}{\code{\link[syntactic]{camel}}, \code{\link[syntactic]{capitalize}}, \code{\link[syntactic]{dotted}}, \code{\link[syntactic]{kebab}}, \code{\link[syntactic]{kebabCase}}, \code{\link[syntactic]{makeDimnames}}, \code{\link[syntactic]{makeLabel}}, \code{\link[syntactic]{makeNames}}, \code{\link[syntactic]{makeTitle}}, \code{\link[syntactic]{makeWords}}, \code{\link[syntactic]{snake}}, \code{\link[syntactic]{upperCamel}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{transformer}{\code{\link[transformer]{relevel}}, \code{\link[transformer]{antiJoin}}, \code{\link[transformer]{as.SummarizedExperiment}}, \code{\link[transformer]{as.data.frame}}, \code{\link[transformer]{as.data.table}}, \code{\link[transformer]{as_tibble}}, \code{\link[transformer]{atomize}}, \code{\link[transformer]{coerceS4ToList}}, \code{\link[transformer]{decode}}, \code{\link[transformer]{encode}}, \code{\link[transformer]{factorize}}, \code{\link[transformer]{flatFiles}}, \code{\link[transformer]{fullJoin}}, \code{\link[transformer]{innerJoin}}, \code{\link[transformer]{leftJoin}}, \code{\link[transformer]{mutateAll}}, \code{\link[transformer]{mutateAt}}, \code{\link[transformer]{mutateIf}}, \code{\link[transformer]{rightJoin}}, \code{\link[transformer]{selectIf}}, \code{\link[transformer]{semiJoin}}, \code{\link[transformer]{transmuteAt}}, \code{\link[transformer]{transmuteIf}}}
}}

