% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGRangesFromEnsembl}
\alias{makeGene2SymbolFromGTF}
\alias{\%>\%}
\alias{makeTx2GeneFromEnsembl}
\alias{emptyRanges}
\alias{makeTx2GeneFromGTF}
\alias{makeGene2SymbolFromGFF}
\alias{geneSynonyms}
\alias{makeGRangesFromGTF}
\alias{makeGRangesFromGFF}
\alias{makeTx2GeneFromEnsDb}
\alias{makeTx2GeneFromGFF}
\alias{makeGRangesFromEnsDb}
\alias{convertUCSCBuildToEnsembl}
\alias{annotable}
\alias{assertIsDir}
\alias{assertFormalCompress}
\alias{assertHasRownames}
\alias{assertIsImplicitInteger}
\alias{assertAreGeneAnnotations}
\alias{assertAllAreURL}
\alias{assertHasAggregateInfo}
\alias{assertHasValidDimnames}
\alias{assertIsHexColorFunctionOrNULL}
\alias{assertIsAnImplicitIntegerOrNULL}
\alias{assertIsAnIntegerOrNULL}
\alias{assertIsAnImplicitInteger}
\alias{hasAggregateInfo}
\alias{assertAllAreValidNames}
\alias{assertIsFillScaleDiscreteOrNULL}
\alias{assertAllAreNonExisting}
\alias{isURL}
\alias{validDimnames}
\alias{isImplicitInteger}
\alias{assertIsHeaderLevel}
\alias{assertAllAreUniqueGeneNames}
\alias{assertIsANumberOrNULL}
\alias{assertFormalGene2Symbol}
\alias{assertIsAlpha}
\alias{assertHasValidNames}
\alias{assertAreTranscriptAnnotations}
\alias{assertIsImplicitIntegerOrNULL}
\alias{assertIsAStringOrNULL}
\alias{assertIsColorScaleDiscreteOrNULL}
\alias{assertFormalInterestingGroups}
\alias{isAnImplicitInteger}
\alias{assertIsFillScaleContinuousOrNULL}
\alias{validNames}
\alias{hasRownames}
\alias{assertIsColorScaleContinuousOrNULL}
\alias{Ensembl2Entrez}
\alias{show}
\alias{Tx2Gene}
\alias{PANTHER}
\alias{Gene2Symbol}
\alias{HGNC2Ensembl}
\alias{MGI2Ensembl}
\alias{EggNOG}
\alias{coerce}
\alias{as.DataFrame}
\alias{as.SummarizedExperiment}
\alias{coerceS4ToList}
\alias{as_tibble}
\alias{as.data.frame}
\alias{flatFiles}
\alias{parseRd}
\alias{printString}
\alias{matchArgsToDoCall}
\alias{checkClasses}
\alias{standardizeCall}
\alias{multiassignAsEnvir}
\alias{hasLocal}
\alias{RdTags}
\alias{methodFormals}
\alias{headtail}
\alias{separator}
\alias{showSlotInfo}
\alias{methodFunction}
\alias{extractLocal}
\alias{detectHPC}
\alias{dots}
\alias{convertGenesToSymbols}
\alias{selectSamples}
\alias{aggregateCols}
\alias{detectLanes}
\alias{convertTranscriptsToGenes}
\alias{aggregateRows}
\alias{interestingGroups}
\alias{meltCounts}
\alias{convertSampleIDsToNames}
\alias{combine}
\alias{tpm}
\alias{geometricMean}
\alias{minimalSampleData}
\alias{sampleData}
\alias{sampleNames}
\alias{aggregateCellsToSamples}
\alias{mapGenesToSymbols}
\alias{mapGenesToRownames}
\alias{subsetPerSample}
\alias{topCellsPerSample}
\alias{logRatioToFoldChange}
\alias{microplate}
\alias{sampleData<-}
\alias{sampleNames<-}
\alias{foldChangeToLogRatio}
\alias{mapCellsToSamples}
\alias{metrics}
\alias{matchInterestingGroups}
\alias{mapGenesToIDs}
\alias{convertSymbolsToGenes}
\alias{counts}
\alias{uniteInterestingGroups}
\alias{makeSingleCellExperiment}
\alias{makeSummarizedExperiment}
\alias{metricsPerSample}
\alias{cell2sample}
\alias{zerosVsDepth}
\alias{detectOrganism}
\alias{organism}
\alias{interestingGroups<-}
\alias{counts<-}
\alias{plotCorrelationHeatmap}
\alias{plotZerosVsDepth}
\alias{plotHeatmap}
\alias{collapseToString}
\alias{plotQuantileHeatmap}
\alias{upperCamel}
\alias{plotCountsPerBroadClass}
\alias{snake}
\alias{export}
\alias{plotGenderMarkers}
\alias{camel}
\alias{plotGene}
\alias{stripTranscriptVersions}
\alias{dotted}
\alias{plotTotalCounts}
\alias{plotCountsPerGene}
\alias{plotCountsPerBiotype}
\alias{plotQC}
\alias{sanitizePercent}
\alias{sanitizeNA}
\alias{removeNA}
\alias{markdown}
\alias{plotGenesDetected}
\alias{basejumpCacheURL}
\alias{barcodePattern}
\alias{rdataExtPattern}
\alias{naStrings}
\alias{rdataLoadError}
\alias{lanePattern}
\alias{extPattern}
\alias{genomeMetadataNames}
\alias{import}
\alias{loadRemoteData}
\alias{loadDataAsName}
\alias{assignAndSaveData}
\alias{saveData}
\alias{writeCounts}
\alias{localOrRemoteFile}
\alias{loadData}
\alias{realpath}
\alias{transmit}
\alias{initDir}
\alias{mdPlotlist}
\alias{mdLink}
\alias{mdList}
\alias{markdownTables}
\alias{markdownPlotlist}
\alias{markdownList}
\alias{markdownLink}
\alias{markdownHeader}
\alias{prepareTemplate}
\alias{mdHeader}
\alias{basejump_geom_label}
\alias{basejump_geom_label_average}
\alias{theme_midnight}
\alias{basejump_geom_label_repel}
\alias{basejump_geom_abline}
\alias{theme_paperwhite}
\alias{makeDimnames}
\alias{sanitizeSampleData}
\alias{sanitizeRowData}
\alias{sortUnique}
\alias{makeNames}
\alias{toStringUnique}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump.annotations}{\code{\link[basejump.annotations]{makeGene2SymbolFromEnsembl}}, \code{\link[basejump.annotations]{makeGene2SymbolFromEnsDb}}, \code{\link[basejump.annotations]{makeGRangesFromEnsembl}}, \code{\link[basejump.annotations]{makeGene2SymbolFromGTF}}, \code{\link[basejump.annotations]{\%>\%}}, \code{\link[basejump.annotations]{makeTx2GeneFromEnsembl}}, \code{\link[basejump.annotations]{emptyRanges}}, \code{\link[basejump.annotations]{makeTx2GeneFromGTF}}, \code{\link[basejump.annotations]{makeGene2SymbolFromGFF}}, \code{\link[basejump.annotations]{geneSynonyms}}, \code{\link[basejump.annotations]{makeGRangesFromGTF}}, \code{\link[basejump.annotations]{makeGRangesFromGFF}}, \code{\link[basejump.annotations]{makeTx2GeneFromEnsDb}}, \code{\link[basejump.annotations]{makeTx2GeneFromGFF}}, \code{\link[basejump.annotations]{makeGRangesFromEnsDb}}, \code{\link[basejump.annotations]{convertUCSCBuildToEnsembl}}, \code{\link[basejump.annotations]{annotable}}}

  \item{basejump.assertions}{\code{\link[basejump.assertions]{assertIsDir}}, \code{\link[basejump.assertions]{assertFormalCompress}}, \code{\link[basejump.assertions]{assertHasRownames}}, \code{\link[basejump.assertions]{assertIsImplicitInteger}}, \code{\link[basejump.assertions]{assertAreGeneAnnotations}}, \code{\link[basejump.assertions]{assertAllAreURL}}, \code{\link[basejump.assertions]{assertHasAggregateInfo}}, \code{\link[basejump.assertions]{assertHasValidDimnames}}, \code{\link[basejump.assertions]{assertIsHexColorFunctionOrNULL}}, \code{\link[basejump.assertions]{assertIsAnImplicitIntegerOrNULL}}, \code{\link[basejump.assertions]{assertIsAnIntegerOrNULL}}, \code{\link[basejump.assertions]{assertIsAnImplicitInteger}}, \code{\link[basejump.assertions]{hasAggregateInfo}}, \code{\link[basejump.assertions]{assertAllAreValidNames}}, \code{\link[basejump.assertions]{assertIsFillScaleDiscreteOrNULL}}, \code{\link[basejump.assertions]{assertAllAreNonExisting}}, \code{\link[basejump.assertions]{isURL}}, \code{\link[basejump.assertions]{validDimnames}}, \code{\link[basejump.assertions]{isImplicitInteger}}, \code{\link[basejump.assertions]{assertIsHeaderLevel}}, \code{\link[basejump.assertions]{assertAllAreUniqueGeneNames}}, \code{\link[basejump.assertions]{assertIsANumberOrNULL}}, \code{\link[basejump.assertions]{assertFormalGene2Symbol}}, \code{\link[basejump.assertions]{assertIsAlpha}}, \code{\link[basejump.assertions]{assertHasValidNames}}, \code{\link[basejump.assertions]{assertAreTranscriptAnnotations}}, \code{\link[basejump.assertions]{assertIsImplicitIntegerOrNULL}}, \code{\link[basejump.assertions]{assertIsAStringOrNULL}}, \code{\link[basejump.assertions]{assertIsColorScaleDiscreteOrNULL}}, \code{\link[basejump.assertions]{\%>\%}}, \code{\link[basejump.assertions]{assertFormalInterestingGroups}}, \code{\link[basejump.assertions]{isAnImplicitInteger}}, \code{\link[basejump.assertions]{assertIsFillScaleContinuousOrNULL}}, \code{\link[basejump.assertions]{validNames}}, \code{\link[basejump.assertions]{hasRownames}}, \code{\link[basejump.assertions]{assertIsColorScaleContinuousOrNULL}}}

  \item{basejump.classes}{\code{\link[basejump.classes]{Ensembl2Entrez}}, \code{\link[basejump.classes]{\%>\%}}, \code{\link[basejump.classes]{show}}, \code{\link[basejump.classes]{Tx2Gene}}, \code{\link[basejump.classes]{PANTHER}}, \code{\link[basejump.classes]{Gene2Symbol}}, \code{\link[basejump.classes]{HGNC2Ensembl}}, \code{\link[basejump.classes]{MGI2Ensembl}}, \code{\link[basejump.classes]{EggNOG}}}

  \item{basejump.coercion}{\code{\link[basejump.coercion]{coerce}}, \code{\link[basejump.coercion]{as.DataFrame}}, \code{\link[basejump.coercion]{as.SummarizedExperiment}}, \code{\link[basejump.coercion]{coerceS4ToList}}, \code{\link[basejump.coercion]{as_tibble}}, \code{\link[basejump.coercion]{as.data.frame}}, \code{\link[basejump.coercion]{flatFiles}}}

  \item{basejump.developer}{\code{\link[basejump.developer]{parseRd}}, \code{\link[basejump.developer]{\%>\%}}, \code{\link[basejump.developer]{printString}}, \code{\link[basejump.developer]{matchArgsToDoCall}}, \code{\link[basejump.developer]{checkClasses}}, \code{\link[basejump.developer]{standardizeCall}}, \code{\link[basejump.developer]{multiassignAsEnvir}}, \code{\link[basejump.developer]{hasLocal}}, \code{\link[basejump.developer]{RdTags}}, \code{\link[basejump.developer]{methodFormals}}, \code{\link[basejump.developer]{headtail}}, \code{\link[basejump.developer]{separator}}, \code{\link[basejump.developer]{showSlotInfo}}, \code{\link[basejump.developer]{methodFunction}}, \code{\link[basejump.developer]{extractLocal}}, \code{\link[basejump.developer]{detectHPC}}, \code{\link[basejump.developer]{dots}}}

  \item{basejump.experiment}{\code{\link[basejump.experiment]{convertGenesToSymbols}}, \code{\link[basejump.experiment]{selectSamples}}, \code{\link[basejump.experiment]{aggregateCols}}, \code{\link[basejump.experiment]{detectLanes}}, \code{\link[basejump.experiment]{convertTranscriptsToGenes}}, \code{\link[basejump.experiment]{aggregateRows}}, \code{\link[basejump.experiment]{interestingGroups}}, \code{\link[basejump.experiment]{meltCounts}}, \code{\link[basejump.experiment]{convertSampleIDsToNames}}, \code{\link[basejump.experiment]{combine}}, \code{\link[basejump.experiment]{tpm}}, \code{\link[basejump.experiment]{geometricMean}}, \code{\link[basejump.experiment]{minimalSampleData}}, \code{\link[basejump.experiment]{sampleData}}, \code{\link[basejump.experiment]{sampleNames}}, \code{\link[basejump.experiment]{aggregateCellsToSamples}}, \code{\link[basejump.experiment]{mapGenesToSymbols}}, \code{\link[basejump.experiment]{mapGenesToRownames}}, \code{\link[basejump.experiment]{subsetPerSample}}, \code{\link[basejump.experiment]{topCellsPerSample}}, \code{\link[basejump.experiment]{logRatioToFoldChange}}, \code{\link[basejump.experiment]{microplate}}, \code{\link[basejump.experiment]{sampleData<-}}, \code{\link[basejump.experiment]{sampleNames<-}}, \code{\link[basejump.experiment]{foldChangeToLogRatio}}, \code{\link[basejump.experiment]{mapCellsToSamples}}, \code{\link[basejump.experiment]{metrics}}, \code{\link[basejump.experiment]{matchInterestingGroups}}, \code{\link[basejump.experiment]{mapGenesToIDs}}, \code{\link[basejump.experiment]{convertSymbolsToGenes}}, \code{\link[basejump.experiment]{counts}}, \code{\link[basejump.experiment]{uniteInterestingGroups}}, \code{\link[basejump.experiment]{makeSingleCellExperiment}}, \code{\link[basejump.experiment]{makeSummarizedExperiment}}, \code{\link[basejump.experiment]{\%>\%}}, \code{\link[basejump.experiment]{metricsPerSample}}, \code{\link[basejump.experiment]{cell2sample}}, \code{\link[basejump.experiment]{zerosVsDepth}}, \code{\link[basejump.experiment]{detectOrganism}}, \code{\link[basejump.experiment]{organism}}, \code{\link[basejump.experiment]{interestingGroups<-}}, \code{\link[basejump.experiment]{counts<-}}}

  \item{basejump.generics}{\code{\link[basejump.generics]{plotCorrelationHeatmap}}, \code{\link[basejump.generics]{meltCounts}}, \code{\link[basejump.generics]{tpm}}, \code{\link[basejump.generics]{plotZerosVsDepth}}, \code{\link[basejump.generics]{plotHeatmap}}, \code{\link[basejump.generics]{collapseToString}}, \code{\link[basejump.generics]{cell2sample}}, \code{\link[basejump.generics]{plotQuantileHeatmap}}, \code{\link[basejump.generics]{logRatioToFoldChange}}, \code{\link[basejump.generics]{upperCamel}}, \code{\link[basejump.generics]{plotCountsPerBroadClass}}, \code{\link[basejump.generics]{headtail}}, \code{\link[basejump.generics]{geometricMean}}, \code{\link[basejump.generics]{subsetPerSample}}, \code{\link[basejump.generics]{interestingGroups}}, \code{\link[basejump.generics]{uniteInterestingGroups}}, \code{\link[basejump.generics]{aggregateCols}}, \code{\link[basejump.generics]{zerosVsDepth}}, \code{\link[basejump.generics]{sampleData}}, \code{\link[basejump.generics]{interestingGroups<-}}, \code{\link[basejump.generics]{convertSymbolsToGenes}}, \code{\link[basejump.generics]{snake}}, \code{\link[basejump.generics]{convertTranscriptsToGenes}}, \code{\link[basejump.generics]{export}}, \code{\link[basejump.generics]{plotGenderMarkers}}, \code{\link[basejump.generics]{topCellsPerSample}}, \code{\link[basejump.generics]{camel}}, \code{\link[basejump.generics]{aggregateCellsToSamples}}, \code{\link[basejump.generics]{metrics}}, \code{\link[basejump.generics]{plotGene}}, \code{\link[basejump.generics]{stripTranscriptVersions}}, \code{\link[basejump.generics]{convertGenesToSymbols}}, \code{\link[basejump.generics]{dotted}}, \code{\link[basejump.generics]{aggregateRows}}, \code{\link[basejump.generics]{plotTotalCounts}}, \code{\link[basejump.generics]{plotCountsPerGene}}, \code{\link[basejump.generics]{mapGenesToSymbols}}, \code{\link[basejump.generics]{mapGenesToRownames}}, \code{\link[basejump.generics]{plotCountsPerBiotype}}, \code{\link[basejump.generics]{plotQC}}, \code{\link[basejump.generics]{sanitizePercent}}, \code{\link[basejump.generics]{sanitizeNA}}, \code{\link[basejump.generics]{removeNA}}, \code{\link[basejump.generics]{metricsPerSample}}, \code{\link[basejump.generics]{convertSampleIDsToNames}}, \code{\link[basejump.generics]{selectSamples}}, \code{\link[basejump.generics]{foldChangeToLogRatio}}, \code{\link[basejump.generics]{sampleData<-}}, \code{\link[basejump.generics]{markdown}}, \code{\link[basejump.generics]{mapGenesToIDs}}, \code{\link[basejump.generics]{plotGenesDetected}}}

  \item{basejump.globals}{\code{\link[basejump.globals]{basejumpCacheURL}}, \code{\link[basejump.globals]{barcodePattern}}, \code{\link[basejump.globals]{rdataExtPattern}}, \code{\link[basejump.globals]{naStrings}}, \code{\link[basejump.globals]{rdataLoadError}}, \code{\link[basejump.globals]{lanePattern}}, \code{\link[basejump.globals]{extPattern}}, \code{\link[basejump.globals]{genomeMetadataNames}}}

  \item{basejump.io}{\code{\link[basejump.io]{import}}, \code{\link[basejump.io]{loadRemoteData}}, \code{\link[basejump.io]{\%>\%}}, \code{\link[basejump.io]{loadDataAsName}}, \code{\link[basejump.io]{assignAndSaveData}}, \code{\link[basejump.io]{saveData}}, \code{\link[basejump.io]{export}}, \code{\link[basejump.io]{writeCounts}}, \code{\link[basejump.io]{localOrRemoteFile}}, \code{\link[basejump.io]{loadData}}, \code{\link[basejump.io]{realpath}}, \code{\link[basejump.io]{transmit}}, \code{\link[basejump.io]{initDir}}}

  \item{basejump.markdown}{\code{\link[basejump.markdown]{\%>\%}}, \code{\link[basejump.markdown]{mdPlotlist}}, \code{\link[basejump.markdown]{markdown}}, \code{\link[basejump.markdown]{mdLink}}, \code{\link[basejump.markdown]{mdList}}, \code{\link[basejump.markdown]{markdownTables}}, \code{\link[basejump.markdown]{markdownPlotlist}}, \code{\link[basejump.markdown]{markdownList}}, \code{\link[basejump.markdown]{markdownLink}}, \code{\link[basejump.markdown]{markdownHeader}}, \code{\link[basejump.markdown]{prepareTemplate}}, \code{\link[basejump.markdown]{mdHeader}}}

  \item{basejump.plots}{\code{\link[basejump.plots]{plotZerosVsDepth}}, \code{\link[basejump.plots]{basejump_geom_label}}, \code{\link[basejump.plots]{basejump_geom_label_average}}, \code{\link[basejump.plots]{plotCountsPerGene}}, \code{\link[basejump.plots]{plotGene}}, \code{\link[basejump.plots]{plotHeatmap}}, \code{\link[basejump.plots]{plotCorrelationHeatmap}}, \code{\link[basejump.plots]{plotCountsPerBroadClass}}, \code{\link[basejump.plots]{plotGenesDetected}}, \code{\link[basejump.plots]{plotQC}}, \code{\link[basejump.plots]{plotTotalCounts}}, \code{\link[basejump.plots]{plotQuantileHeatmap}}, \code{\link[basejump.plots]{theme_midnight}}, \code{\link[basejump.plots]{\%>\%}}, \code{\link[basejump.plots]{plotCountsPerBiotype}}, \code{\link[basejump.plots]{basejump_geom_label_repel}}, \code{\link[basejump.plots]{basejump_geom_abline}}, \code{\link[basejump.plots]{plotGenderMarkers}}, \code{\link[basejump.plots]{theme_paperwhite}}}

  \item{basejump.sanitization}{\code{\link[basejump.sanitization]{sanitizeNA}}, \code{\link[basejump.sanitization]{makeDimnames}}, \code{\link[basejump.sanitization]{sanitizeSampleData}}, \code{\link[basejump.sanitization]{sanitizePercent}}, \code{\link[basejump.sanitization]{dotted}}, \code{\link[basejump.sanitization]{snake}}, \code{\link[basejump.sanitization]{camel}}, \code{\link[basejump.sanitization]{stripTranscriptVersions}}, \code{\link[basejump.sanitization]{collapseToString}}, \code{\link[basejump.sanitization]{sanitizeRowData}}, \code{\link[basejump.sanitization]{removeNA}}, \code{\link[basejump.sanitization]{sortUnique}}, \code{\link[basejump.sanitization]{makeNames}}, \code{\link[basejump.sanitization]{upperCamel}}, \code{\link[basejump.sanitization]{\%>\%}}, \code{\link[basejump.sanitization]{toStringUnique}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

