% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reexports-bioverbs.R,
%   R/reexports-brio.R, R/reexports-freerange.R, R/reexports-goalie.R,
%   R/reexports-syntactic.R, R/reexports-transformer.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plotGene}
\alias{plotDEGHeatmap}
\alias{plotDEGPCA}
\alias{assignAndSaveData}
\alias{basenameSansExt}
\alias{decode}
\alias{dots}
\alias{encode}
\alias{export}
\alias{extPattern}
\alias{factorize}
\alias{fileExt}
\alias{import}
\alias{initDir}
\alias{loadData}
\alias{loadDataAsName}
\alias{loadRemoteData}
\alias{localOrRemoteFile}
\alias{naStrings}
\alias{pasteURL}
\alias{rdataExtPattern}
\alias{rdataLoadError}
\alias{realpath}
\alias{removeNA}
\alias{saveData}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{transmit}
\alias{writeCounts}
\alias{sanitizeColData}
\alias{sanitizeRowData}
\alias{sanitizeRowRanges}
\alias{bunzip2}
\alias{bzip2}
\alias{compressFile}
\alias{decompressFile}
\alias{gunzip}
\alias{gzip}
\alias{tar}
\alias{untar}
\alias{unzip}
\alias{zip}
\alias{annotable}
\alias{convertUCSCBuildToEnsembl}
\alias{detectOrganism}
\alias{emptyRanges}
\alias{makeGRangesFromEnsDb}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromGFF}
\alias{makeGRangesFromGTF}
\alias{assert}
\alias{bapply}
\alias{cause}
\alias{extractLocal}
\alias{matchArgsToDoCall}
\alias{methodFormals}
\alias{methodFunction}
\alias{printString}
\alias{standardizeCall}
\alias{validate}
\alias{validateClasses}
\alias{hasLocal}
\alias{hasRownames}
\alias{makeDimnames}
\alias{makeNames}
\alias{coerceS4ToList}
\alias{flatFiles}
\alias{as.DataFrame}
\alias{as.SummarizedExperiment}
\alias{as.data.frame}
\alias{as.data.table}
\alias{as_tibble}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{set_colnames}
\alias{set_names}
\alias{set_rownames}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{plotGene}}, \code{\link[bioverbs]{plotDEGHeatmap}}, \code{\link[bioverbs]{plotDEGPCA}}}

  \item{brio}{\code{\link[brio]{assignAndSaveData}}, \code{\link[brio]{basenameSansExt}}, \code{\link[brio]{decode}}, \code{\link[brio]{dots}}, \code{\link[brio]{encode}}, \code{\link[brio]{export}}, \code{\link[brio]{extPattern}}, \code{\link[brio]{factorize}}, \code{\link[brio]{fileExt}}, \code{\link[brio]{import}}, \code{\link[brio]{initDir}}, \code{\link[brio]{loadData}}, \code{\link[brio]{loadDataAsName}}, \code{\link[brio]{loadRemoteData}}, \code{\link[brio]{localOrRemoteFile}}, \code{\link[brio]{naStrings}}, \code{\link[brio]{pasteURL}}, \code{\link[brio]{rdataExtPattern}}, \code{\link[brio]{rdataLoadError}}, \code{\link[brio]{realpath}}, \code{\link[brio]{removeNA}}, \code{\link[brio]{saveData}}, \code{\link[brio]{sanitizeNA}}, \code{\link[brio]{sanitizePercent}}, \code{\link[brio]{transmit}}, \code{\link[brio]{writeCounts}}, \code{\link[brio]{sanitizeColData}}, \code{\link[brio]{sanitizeRowData}}, \code{\link[brio]{sanitizeRowRanges}}, \code{\link[brio]{bunzip2}}, \code{\link[brio]{bzip2}}, \code{\link[brio]{compressFile}}, \code{\link[brio]{decompressFile}}, \code{\link[brio]{gunzip}}, \code{\link[brio]{gzip}}, \code{\link[brio]{tar}}, \code{\link[brio]{untar}}, \code{\link[brio]{unzip}}, \code{\link[brio]{zip}}}

  \item{freerange}{\code{\link[freerange]{annotable}}, \code{\link[freerange]{convertUCSCBuildToEnsembl}}, \code{\link[freerange]{detectOrganism}}, \code{\link[freerange]{emptyRanges}}, \code{\link[freerange]{makeGRangesFromEnsDb}}, \code{\link[freerange]{makeGRangesFromEnsembl}}, \code{\link[freerange]{makeGRangesFromGFF}}, \code{\link[freerange]{makeGRangesFromGTF}}}

  \item{goalie}{\code{\link[goalie]{assert}}, \code{\link[goalie]{bapply}}, \code{\link[goalie]{cause}}, \code{\link[goalie]{extractLocal}}, \code{\link[goalie]{matchArgsToDoCall}}, \code{\link[goalie]{methodFormals}}, \code{\link[goalie]{methodFunction}}, \code{\link[goalie]{printString}}, \code{\link[goalie]{standardizeCall}}, \code{\link[goalie]{validate}}, \code{\link[goalie]{validateClasses}}, \code{\link[goalie]{hasLocal}}, \code{\link[goalie]{hasRownames}}}

  \item{magrittr}{\code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%$\%}}, \code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%T>\%}}, \code{\link[magrittr]{set_colnames}}, \code{\link[magrittr]{set_names}}, \code{\link[magrittr]{set_rownames}}}

  \item{syntactic}{\code{\link[syntactic]{makeDimnames}}, \code{\link[syntactic]{makeNames}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{transformer}{\code{\link[transformer]{coerceS4ToList}}, \code{\link[transformer]{flatFiles}}, \code{\link[transformer]{as.DataFrame}}, \code{\link[transformer]{as.SummarizedExperiment}}, \code{\link[transformer]{as.data.frame}}, \code{\link[transformer]{as.data.table}}, \code{\link[transformer]{as_tibble}}}
}}

