% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-01-functions.R,
%   R/reexports-02-generics.R, R/reexports-03-classes.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{CharacterList}
\alias{DataFrame}
\alias{DataFrameList}
\alias{FactorList}
\alias{GRanges}
\alias{GRangesList}
\alias{HGNC}
\alias{HGNC2Ensembl}
\alias{IRanges}
\alias{IntegerList}
\alias{LLint}
\alias{List}
\alias{LogicalList}
\alias{Matrix}
\alias{MGI2Ensembl}
\alias{NumericList}
\alias{Rle}
\alias{RleList}
\alias{SimpleList}
\alias{SingleCellExperiment}
\alias{SplitDataFrameList}
\alias{SummarizedExperiment}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{alert}
\alias{alertDanger}
\alias{alertInfo}
\alias{alertSuccess}
\alias{alertWarning}
\alias{annotable}
\alias{appendToBody}
\alias{arrow}
\alias{as}
\alias{as.formula}
\alias{assignAndSaveData}
\alias{axisTicks}
\alias{barcodePattern}
\alias{basenameSansExt}
\alias{cacheURL}
\alias{capture.output}
\alias{classContains}
\alias{colorRampPalette}
\alias{column_to_rownames}
\alias{compress}
\alias{compressExtPattern}
\alias{currentEnsemblGenomeBuild}
\alias{currentEnsemblVersion}
\alias{currentFlyBaseVersion}
\alias{currentGencodeGenomeBuild}
\alias{currentGencodeVersion}
\alias{currentRefSeqGenomeBuild}
\alias{currentRefSeqVersion}
\alias{currentUCSCGenomeBuild}
\alias{currentWormBaseVersion}
\alias{decompress}
\alias{detectLanes}
\alias{detectOrganism}
\alias{dist}
\alias{dl}
\alias{dots}
\alias{download}
\alias{downloadEnsemblGenome}
\alias{downloadGencodeGenome}
\alias{downloadRefSeqGenome}
\alias{downloadUCSCGenome}
\alias{ecdf}
\alias{emptyRanges}
\alias{extPattern}
\alias{fileDepth}
\alias{fileExt}
\alias{forceDetach}
\alias{formalArgs}
\alias{formalsList}
\alias{geneSynonyms}
\alias{genomeMetadataNames}
\alias{getS3method}
\alias{getURL}
\alias{getURLDirList}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{h7}
\alias{hclust}
\alias{import}
\alias{importSampleData}
\alias{importTx2Gene}
\alias{initDir}
\alias{is}
\alias{lanePattern}
\alias{loadData}
\alias{loadDataAsName}
\alias{loadRemoteData}
\alias{localOrRemoteFile}
\alias{makeGRangesFromEnsDb}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromGFF}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromGFF}
\alias{makeProtein2GeneFromEnsembl}
\alias{makeSingleCellExperiment}
\alias{makeSummarizedExperiment}
\alias{makeTx2GeneFromEnsDb}
\alias{makeTx2GeneFromEnsembl}
\alias{makeTx2GeneFromFASTA}
\alias{makeTx2GeneFromGFF}
\alias{majorVersion}
\alias{mapCellsToSamples}
\alias{mapEnsemblBuildToUCSC}
\alias{mapEnsemblReleaseToURL}
\alias{mapHumanOrthologs}
\alias{mapUCSCBuildToEnsembl}
\alias{markdownHeader}
\alias{markdownLink}
\alias{markdownList}
\alias{markdownPlots}
\alias{markdownTables}
\alias{matchInterestingGroups}
\alias{matchesGene2Symbol}
\alias{matchesInterestingGroups}
\alias{mdHeader}
\alias{mdLink}
\alias{mdList}
\alias{mdPlots}
\alias{mdTables}
\alias{metadataBlacklist}
\alias{methodFormals}
\alias{methodFunction}
\alias{metricsCols}
\alias{microplate}
\alias{minimalSampleData}
\alias{minorVersion}
\alias{model.matrix}
\alias{naStrings}
\alias{ol}
\alias{packageName}
\alias{packageVersion}
\alias{pasteURL}
\alias{prcomp}
\alias{prepareTemplate}
\alias{printString}
\alias{quantile}
\alias{rbindlist}
\alias{realpath}
\alias{relevel}
\alias{reorder}
\alias{requireNamespaces}
\alias{rgb}
\alias{rownames_to_column}
\alias{sanitizeSampleData}
\alias{sanitizeVersion}
\alias{saveData}
\alias{separator}
\alias{session_info}
\alias{sessionInfo}
\alias{setdiff}
\alias{shell}
\alias{showSlotInfo}
\alias{sparseMatrix}
\alias{standardizeCall}
\alias{tibble}
\alias{toInlineString}
\alias{transmit}
\alias{txt}
\alias{ul}
\alias{unit}
\alias{updateMessage}
\alias{verbatim}
\alias{as_tibble}
\alias{as.data.table}
\alias{Ensembl2Entrez}
\alias{Entrez2Ensembl}
\alias{Gene2Symbol}
\alias{Tx2Gene}
\alias{\%in\%}
\alias{aggregate}
\alias{aggregateCellsToSamples}
\alias{aggregateCols}
\alias{aggregateRows}
\alias{antiJoin}
\alias{as.DataFrame}
\alias{as.SummarizedExperiment}
\alias{as.data.frame}
\alias{as.list}
\alias{assay}
\alias{assay<-}
\alias{assayNames}
\alias{assayNames<-}
\alias{assays}
\alias{assays<-}
\alias{atomize}
\alias{autopadZeros}
\alias{calculateMetrics}
\alias{camelCase}
\alias{capitalize}
\alias{cbind}
\alias{cell2sample}
\alias{coerce}
\alias{coerceToList}
\alias{colData}
\alias{colData<-}
\alias{colSums}
\alias{collapseToString}
\alias{combine}
\alias{complete.cases}
\alias{convertGenesToSymbols}
\alias{convertSampleIDsToNames}
\alias{convertSymbolsToGenes}
\alias{convertTranscriptsToGenes}
\alias{cor}
\alias{correlation}
\alias{counts}
\alias{counts<-}
\alias{decode}
\alias{do.call}
\alias{dottedCase}
\alias{droplevels}
\alias{encode}
\alias{end}
\alias{estimateSizeFactors}
\alias{expand}
\alias{expand.grid}
\alias{export}
\alias{factorize}
\alias{filterCells}
\alias{foldChangeToLogRatio}
\alias{fullJoin}
\alias{geneNames}
\alias{geometricMean}
\alias{head}
\alias{headtail}
\alias{humanize}
\alias{innerJoin}
\alias{integerCounts}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{intersectAll}
\alias{intersectionMatrix}
\alias{is.unsorted}
\alias{kebabCase}
\alias{lapply}
\alias{leftJoin}
\alias{logRatioToFoldChange}
\alias{makeDimnames}
\alias{makeLabel}
\alias{makeNames}
\alias{makeSampleData}
\alias{makeTitle}
\alias{makeWords}
\alias{mapGenesToIDs}
\alias{mapGenesToRownames}
\alias{mapGenesToSymbols}
\alias{match}
\alias{matchSampleColumn}
\alias{mcols}
\alias{mcols<-}
\alias{melt}
\alias{merge}
\alias{metadata}
\alias{metadata<-}
\alias{metadata2}
\alias{metadata2<-}
\alias{metrics}
\alias{metricsPerSample}
\alias{mutateAll}
\alias{mutateAt}
\alias{mutateIf}
\alias{na.omit}
\alias{nonzeroRowsAndCols}
\alias{order}
\alias{organism}
\alias{organism<-}
\alias{rankedMatrix}
\alias{rbind}
\alias{reducedDim}
\alias{reducedDim<-}
\alias{reducedDimNames}
\alias{reducedDimNames<-}
\alias{reducedDims}
\alias{reducedDims<-}
\alias{removeNA}
\alias{rightJoin}
\alias{rowData}
\alias{rowData<-}
\alias{rowMeans}
\alias{rowRanges}
\alias{rowRanges<-}
\alias{rowSums}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{selectIf}
\alias{selectSamples}
\alias{sem}
\alias{semiJoin}
\alias{showHeader}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{snakeCase}
\alias{sort}
\alias{split}
\alias{splitByLevel}
\alias{start}
\alias{stripGeneVersions}
\alias{stripTranscriptVersions}
\alias{subsetPerSample}
\alias{summary}
\alias{topCellsPerSample}
\alias{t}
\alias{table}
\alias{tail}
\alias{tpm}
\alias{transmuteAt}
\alias{transmuteIf}
\alias{unique}
\alias{uniteInterestingGroups}
\alias{unlist}
\alias{unlistToDataFrame}
\alias{unsplit}
\alias{upperCamelCase}
\alias{var}
\alias{width}
\alias{zerosVsDepth}
\alias{Annotated-class}
\alias{AsIs-class}
\alias{AtomicList-class}
\alias{CharacterList-class}
\alias{DataFrame-class}
\alias{DataFrameList-class}
\alias{Ensembl2Entrez-class}
\alias{Entrez2Ensembl-class}
\alias{EnsemblGenes-class}
\alias{EnsemblTranscripts-class}
\alias{Factor-class}
\alias{FactorList-class}
\alias{FlyBaseGenes-class}
\alias{FlyBaseTranscripts-class}
\alias{GRanges-class}
\alias{GRangesList-class}
\alias{GencodeGenes-class}
\alias{GencodeTranscripts-class}
\alias{Gene2Symbol-class}
\alias{HGNC-class}
\alias{HGNC2Ensembl-class}
\alias{IRanges-class}
\alias{IntegerList-class}
\alias{LLint-class}
\alias{List-class}
\alias{LogicalList-class}
\alias{Matrix-class}
\alias{MGI2Ensembl-class}
\alias{NumericList-class}
\alias{Protein2Gene-class}
\alias{RangedSummarizedExperiment-class}
\alias{RectangularData-class}
\alias{RefSeqGenes-class}
\alias{RefSeqTranscripts-class}
\alias{Rle-class}
\alias{RleList-class}
\alias{SimpleList-class}
\alias{SingleCellExperiment-class}
\alias{SplitDataFrameList-class}
\alias{SummarizedExperiment-class}
\alias{Tx2Gene-class}
\alias{UCSCGenes-class}
\alias{UCSCTranscripts-class}
\alias{WormBaseGenes-class}
\alias{WormBaseTranscripts-class}
\alias{Vector-class}
\alias{atomic-class}
\alias{missingOrNULL-class}
\alias{sparseMatrix-class}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{appendToBody}}, \code{\link[AcidBase:reexports]{arrow}}, \code{\link[AcidBase]{barcodePattern}}, \code{\link[AcidBase]{basenameSansExt}}, \code{\link[AcidBase]{classContains}}, \code{\link[AcidBase]{compress}}, \code{\link[AcidBase:reexports]{compressExtPattern}}, \code{\link[AcidBase:compress]{decompress}}, \code{\link[AcidBase]{dots}}, \code{\link[AcidBase]{download}}, \code{\link[AcidBase:reexports]{extPattern}}, \code{\link[AcidBase]{fileDepth}}, \code{\link[AcidBase]{fileExt}}, \code{\link[AcidBase]{forceDetach}}, \code{\link[AcidBase]{formalsList}}, \code{\link[AcidBase]{genomeMetadataNames}}, \code{\link[AcidBase]{initDir}}, \code{\link[AcidBase]{lanePattern}}, \code{\link[AcidBase]{majorVersion}}, \code{\link[AcidBase]{metadataBlacklist}}, \code{\link[AcidBase:methodDefinition]{methodFormals}}, \code{\link[AcidBase:methodDefinition]{methodFunction}}, \code{\link[AcidBase]{metricsCols}}, \code{\link[AcidBase]{minorVersion}}, \code{\link[AcidBase]{pasteURL}}, \code{\link[AcidBase:reexports]{printString}}, \code{\link[AcidBase]{realpath}}, \code{\link[AcidBase]{requireNamespaces}}, \code{\link[AcidBase]{sanitizeVersion}}, \code{\link[AcidBase]{shell}}, \code{\link[AcidBase]{showSlotInfo}}, \code{\link[AcidBase]{standardizeCall}}, \code{\link[AcidBase:reexports]{unit}}, \code{\link[AcidBase]{updateMessage}}}

  \item{AcidCLI}{\code{\link[AcidCLI]{alert}}, \code{\link[AcidCLI:alert]{alertDanger}}, \code{\link[AcidCLI:alert]{alertInfo}}, \code{\link[AcidCLI:alert]{alertSuccess}}, \code{\link[AcidCLI:alert]{alertWarning}}, \code{\link[AcidCLI:list]{dl}}, \code{\link[AcidCLI:header]{h1}}, \code{\link[AcidCLI:header]{h2}}, \code{\link[AcidCLI:header]{h3}}, \code{\link[AcidCLI:header]{h4}}, \code{\link[AcidCLI:header]{h5}}, \code{\link[AcidCLI:header]{h6}}, \code{\link[AcidCLI:header]{h7}}, \code{\link[AcidCLI:list]{ol}}, \code{\link[AcidCLI]{separator}}, \code{\link[AcidCLI]{toInlineString}}, \code{\link[AcidCLI:text]{txt}}, \code{\link[AcidCLI:list]{ul}}, \code{\link[AcidCLI:text]{verbatim}}}

  \item{AcidExperiment}{\code{\link[AcidExperiment:reexports]{SummarizedExperiment}}, \code{\link[AcidExperiment:reexports]{assay}}, \code{\link[AcidExperiment:reexports]{assay<-}}, \code{\link[AcidExperiment:reexports]{assayNames}}, \code{\link[AcidExperiment:reexports]{assayNames<-}}, \code{\link[AcidExperiment:reexports]{assays}}, \code{\link[AcidExperiment:reexports]{assays<-}}, \code{\link[AcidExperiment:reexports]{colData}}, \code{\link[AcidExperiment:reexports]{colData<-}}, \code{\link[AcidExperiment]{detectLanes}}, \code{\link[AcidExperiment]{importSampleData}}, \code{\link[AcidExperiment]{makeSummarizedExperiment}}, \code{\link[AcidExperiment]{matchInterestingGroups}}, \code{\link[AcidExperiment]{matchesGene2Symbol}}, \code{\link[AcidExperiment]{matchesInterestingGroups}}, \code{\link[AcidExperiment]{microplate}}, \code{\link[AcidExperiment]{minimalSampleData}}, \code{\link[AcidExperiment:reexports]{rowData}}, \code{\link[AcidExperiment:reexports]{rowData<-}}, \code{\link[AcidExperiment:reexports]{rowRanges}}, \code{\link[AcidExperiment:reexports]{rowRanges<-}}, \code{\link[AcidExperiment]{sanitizeSampleData}}, \code{\link[AcidExperiment:reexports]{session_info}}}

  \item{AcidGenerics}{\code{\link[AcidGenerics:reexports]{\%in\%}}, \code{\link[AcidGenerics:reexports]{DataFrame}}, \code{\link[AcidGenerics:reexports]{LLint}}, \code{\link[AcidGenerics:reexports]{List}}, \code{\link[AcidGenerics:reexports]{Rle}}, \code{\link[AcidGenerics:reexports]{SimpleList}}, \code{\link[AcidGenerics]{aggregate}}, \code{\link[AcidGenerics]{aggregateCellsToSamples}}, \code{\link[AcidGenerics]{aggregateCols}}, \code{\link[AcidGenerics]{aggregateRows}}, \code{\link[AcidGenerics:join]{antiJoin}}, \code{\link[AcidGenerics]{as.DataFrame}}, \code{\link[AcidGenerics]{as.SummarizedExperiment}}, \code{\link[AcidGenerics]{as.data.frame}}, \code{\link[AcidGenerics]{as.list}}, \code{\link[AcidGenerics]{atomize}}, \code{\link[AcidGenerics]{autopadZeros}}, \code{\link[AcidGenerics]{calculateMetrics}}, \code{\link[AcidGenerics]{camelCase}}, \code{\link[AcidGenerics]{capitalize}}, \code{\link[AcidGenerics:reexports]{cbind}}, \code{\link[AcidGenerics]{cell2sample}}, \code{\link[AcidGenerics]{coerce}}, \code{\link[AcidGenerics]{coerceToList}}, \code{\link[AcidGenerics:reexports]{colSums}}, \code{\link[AcidGenerics]{collapseToString}}, \code{\link[AcidGenerics]{combine}}, \code{\link[AcidGenerics:reexports]{complete.cases}}, \code{\link[AcidGenerics]{convertGenesToSymbols}}, \code{\link[AcidGenerics]{convertSampleIDsToNames}}, \code{\link[AcidGenerics]{convertSymbolsToGenes}}, \code{\link[AcidGenerics]{convertTranscriptsToGenes}}, \code{\link[AcidGenerics:reexports]{cor}}, \code{\link[AcidGenerics]{correlation}}, \code{\link[AcidGenerics]{counts}}, \code{\link[AcidGenerics:counts]{counts<-}}, \code{\link[AcidGenerics]{decode}}, \code{\link[AcidGenerics:reexports]{do.call}}, \code{\link[AcidGenerics]{dottedCase}}, \code{\link[AcidGenerics]{droplevels}}, \code{\link[AcidGenerics]{encode}}, \code{\link[AcidGenerics:reexports]{end}}, \code{\link[AcidGenerics]{estimateSizeFactors}}, \code{\link[AcidGenerics:reexports]{expand}}, \code{\link[AcidGenerics:reexports]{expand.grid}}, \code{\link[AcidGenerics]{export}}, \code{\link[AcidGenerics]{factorize}}, \code{\link[AcidGenerics]{filterCells}}, \code{\link[AcidGenerics]{foldChangeToLogRatio}}, \code{\link[AcidGenerics:join]{fullJoin}}, \code{\link[AcidGenerics]{geneNames}}, \code{\link[AcidGenerics]{geometricMean}}, \code{\link[AcidGenerics]{head}}, \code{\link[AcidGenerics]{headtail}}, \code{\link[AcidGenerics]{humanize}}, \code{\link[AcidGenerics:join]{innerJoin}}, \code{\link[AcidGenerics]{integerCounts}}, \code{\link[AcidGenerics]{interestingGroups}}, \code{\link[AcidGenerics:interestingGroups]{interestingGroups<-}}, \code{\link[AcidGenerics]{intersectAll}}, \code{\link[AcidGenerics]{intersectionMatrix}}, \code{\link[AcidGenerics:reexports]{is.unsorted}}, \code{\link[AcidGenerics]{kebabCase}}, \code{\link[AcidGenerics:reexports]{lapply}}, \code{\link[AcidGenerics:join]{leftJoin}}, \code{\link[AcidGenerics]{logRatioToFoldChange}}, \code{\link[AcidGenerics]{makeDimnames}}, \code{\link[AcidGenerics]{makeLabel}}, \code{\link[AcidGenerics]{makeNames}}, \code{\link[AcidGenerics]{makeSampleData}}, \code{\link[AcidGenerics]{makeTitle}}, \code{\link[AcidGenerics]{makeWords}}, \code{\link[AcidGenerics]{mapGenesToIDs}}, \code{\link[AcidGenerics]{mapGenesToRownames}}, \code{\link[AcidGenerics]{mapGenesToSymbols}}, \code{\link[AcidGenerics:reexports]{match}}, \code{\link[AcidGenerics]{matchSampleColumn}}, \code{\link[AcidGenerics:reexports]{mcols}}, \code{\link[AcidGenerics:reexports]{mcols<-}}, \code{\link[AcidGenerics]{melt}}, \code{\link[AcidGenerics:reexports]{merge}}, \code{\link[AcidGenerics:reexports]{metadata}}, \code{\link[AcidGenerics]{metadata2}}, \code{\link[AcidGenerics:metadata2]{metadata2<-}}, \code{\link[AcidGenerics:reexports]{metadata<-}}, \code{\link[AcidGenerics]{metrics}}, \code{\link[AcidGenerics]{metricsPerSample}}, \code{\link[AcidGenerics:mutate]{mutateAll}}, \code{\link[AcidGenerics:mutate]{mutateAt}}, \code{\link[AcidGenerics:mutate]{mutateIf}}, \code{\link[AcidGenerics:reexports]{na.omit}}, \code{\link[AcidGenerics]{nonzeroRowsAndCols}}, \code{\link[AcidGenerics:reexports]{order}}, \code{\link[AcidGenerics]{organism}}, \code{\link[AcidGenerics:organism]{organism<-}}, \code{\link[AcidGenerics]{rankedMatrix}}, \code{\link[AcidGenerics:reexports]{rbind}}, \code{\link[AcidGenerics]{removeNA}}, \code{\link[AcidGenerics:join]{rightJoin}}, \code{\link[AcidGenerics:reexports]{rowMeans}}, \code{\link[AcidGenerics:reexports]{rowSums}}, \code{\link[AcidGenerics]{sampleData}}, \code{\link[AcidGenerics:sampleData]{sampleData<-}}, \code{\link[AcidGenerics]{sampleNames}}, \code{\link[AcidGenerics:sampleNames]{sampleNames<-}}, \code{\link[AcidGenerics]{sanitizeNA}}, \code{\link[AcidGenerics]{sanitizePercent}}, \code{\link[AcidGenerics:select]{selectIf}}, \code{\link[AcidGenerics]{selectSamples}}, \code{\link[AcidGenerics]{sem}}, \code{\link[AcidGenerics:join]{semiJoin}}, \code{\link[AcidGenerics:reexports]{setdiff}}, \code{\link[AcidGenerics]{showHeader}}, \code{\link[AcidGenerics]{sizeFactors}}, \code{\link[AcidGenerics:sizeFactors]{sizeFactors<-}}, \code{\link[AcidGenerics]{snakeCase}}, \code{\link[AcidGenerics:reexports]{sort}}, \code{\link[AcidGenerics:reexports]{split}}, \code{\link[AcidGenerics]{splitByLevel}}, \code{\link[AcidGenerics:reexports]{start}}, \code{\link[AcidGenerics]{stripGeneVersions}}, \code{\link[AcidGenerics]{stripTranscriptVersions}}, \code{\link[AcidGenerics]{subsetPerSample}}, \code{\link[AcidGenerics]{summary}}, \code{\link[AcidGenerics:reexports]{t}}, \code{\link[AcidGenerics:reexports]{table}}, \code{\link[AcidGenerics]{tail}}, \code{\link[AcidGenerics]{topCellsPerSample}}, \code{\link[AcidGenerics]{tpm}}, \code{\link[AcidGenerics:mutate]{transmuteAt}}, \code{\link[AcidGenerics:mutate]{transmuteIf}}, \code{\link[AcidGenerics:reexports]{unique}}, \code{\link[AcidGenerics]{uniteInterestingGroups}}, \code{\link[AcidGenerics:reexports]{unlist}}, \code{\link[AcidGenerics]{unlistToDataFrame}}, \code{\link[AcidGenerics:reexports]{unsplit}}, \code{\link[AcidGenerics]{upperCamelCase}}, \code{\link[AcidGenerics:reexports]{var}}, \code{\link[AcidGenerics:reexports]{width}}, \code{\link[AcidGenerics]{zerosVsDepth}}}

  \item{AcidGenomes}{\code{\link[AcidGenomes]{Ensembl2Entrez}}, \code{\link[AcidGenomes:Ensembl2Entrez]{Entrez2Ensembl}}, \code{\link[AcidGenomes]{Gene2Symbol}}, \code{\link[AcidGenomes]{HGNC}}, \code{\link[AcidGenomes]{HGNC2Ensembl}}, \code{\link[AcidGenomes]{MGI2Ensembl}}, \code{\link[AcidGenomes]{Tx2Gene}}, \code{\link[AcidGenomes:makeGRangesFromEnsembl]{annotable}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentEnsemblGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentEnsemblVersion}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentFlyBaseVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentGencodeGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentGencodeVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentRefSeqGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentRefSeqVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentUCSCGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentWormBaseVersion}}, \code{\link[AcidGenomes]{detectOrganism}}, \code{\link[AcidGenomes]{downloadEnsemblGenome}}, \code{\link[AcidGenomes]{downloadGencodeGenome}}, \code{\link[AcidGenomes]{downloadRefSeqGenome}}, \code{\link[AcidGenomes]{downloadUCSCGenome}}, \code{\link[AcidGenomes]{emptyRanges}}, \code{\link[AcidGenomes]{geneSynonyms}}, \code{\link[AcidGenomes]{importTx2Gene}}, \code{\link[AcidGenomes:makeGRangesFromEnsembl]{makeGRangesFromEnsDb}}, \code{\link[AcidGenomes]{makeGRangesFromEnsembl}}, \code{\link[AcidGenomes]{makeGRangesFromGFF}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromEnsDb}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromEnsembl}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromGFF}}, \code{\link[AcidGenomes:makeProtein2Gene]{makeProtein2GeneFromEnsembl}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromEnsDb}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromEnsembl}}, \code{\link[AcidGenomes]{makeTx2GeneFromFASTA}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromGFF}}, \code{\link[AcidGenomes:mapGenomeBuild]{mapEnsemblBuildToUCSC}}, \code{\link[AcidGenomes]{mapEnsemblReleaseToURL}}, \code{\link[AcidGenomes]{mapHumanOrthologs}}, \code{\link[AcidGenomes:mapGenomeBuild]{mapUCSCBuildToEnsembl}}}

  \item{AcidMarkdown}{\code{\link[AcidMarkdown]{markdownHeader}}, \code{\link[AcidMarkdown]{markdownLink}}, \code{\link[AcidMarkdown]{markdownList}}, \code{\link[AcidMarkdown]{markdownPlots}}, \code{\link[AcidMarkdown]{markdownTables}}, \code{\link[AcidMarkdown:markdownHeader]{mdHeader}}, \code{\link[AcidMarkdown:markdownLink]{mdLink}}, \code{\link[AcidMarkdown:markdownList]{mdList}}, \code{\link[AcidMarkdown:markdownPlots]{mdPlots}}, \code{\link[AcidMarkdown:markdownTables]{mdTables}}, \code{\link[AcidMarkdown]{prepareTemplate}}}

  \item{AcidSingleCell}{\code{\link[AcidSingleCell:reexports]{SingleCellExperiment}}, \code{\link[AcidSingleCell]{makeSingleCellExperiment}}, \code{\link[AcidSingleCell]{mapCellsToSamples}}, \code{\link[AcidSingleCell:reexports]{reducedDim}}, \code{\link[AcidSingleCell:reexports]{reducedDim<-}}, \code{\link[AcidSingleCell:reexports]{reducedDimNames}}, \code{\link[AcidSingleCell:reexports]{reducedDimNames<-}}, \code{\link[AcidSingleCell:reexports]{reducedDims}}, \code{\link[AcidSingleCell:reexports]{reducedDims<-}}}

  \item{grDevices}{\code{\link[grDevices]{axisTicks}}, \code{\link[grDevices:colorRamp]{colorRampPalette}}, \code{\link[grDevices]{rgb}}}

  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:tee]{\%T>\%}}}

  \item{methods}{\code{\link[methods]{as}}, \code{\link[methods:methodUtilities]{formalArgs}}, \code{\link[methods]{is}}}

  \item{pipette}{\code{\link[pipette:reexports]{CharacterList}}, \code{\link[pipette:reexports]{DataFrameList}}, \code{\link[pipette:reexports]{FactorList}}, \code{\link[pipette:reexports]{GRanges}}, \code{\link[pipette:reexports]{GRangesList}}, \code{\link[pipette:reexports]{IRanges}}, \code{\link[pipette:reexports]{IntegerList}}, \code{\link[pipette:reexports]{LogicalList}}, \code{\link[pipette:reexports]{Matrix}}, \code{\link[pipette:reexports]{NumericList}}, \code{\link[pipette:reexports]{RleList}}, \code{\link[pipette:reexports]{SplitDataFrameList}}, \code{\link[pipette:coerce]{as.data.table}}, \code{\link[pipette:coerce]{as_tibble}}, \code{\link[pipette]{assignAndSaveData}}, \code{\link[pipette]{cacheURL}}, \code{\link[pipette:reexports]{column_to_rownames}}, \code{\link[pipette:reexports]{getURL}}, \code{\link[pipette]{getURLDirList}}, \code{\link[pipette]{import}}, \code{\link[pipette]{loadData}}, \code{\link[pipette]{loadDataAsName}}, \code{\link[pipette]{loadRemoteData}}, \code{\link[pipette]{localOrRemoteFile}}, \code{\link[pipette]{naStrings}}, \code{\link[pipette:reexports]{rbindlist}}, \code{\link[pipette:reexports]{rownames_to_column}}, \code{\link[pipette]{saveData}}, \code{\link[pipette:reexports]{sparseMatrix}}, \code{\link[pipette:reexports]{tibble}}, \code{\link[pipette]{transmit}}}

  \item{stats}{\code{\link[stats:formula]{as.formula}}, \code{\link[stats]{dist}}, \code{\link[stats]{ecdf}}, \code{\link[stats]{hclust}}, \code{\link[stats]{model.matrix}}, \code{\link[stats]{prcomp}}, \code{\link[stats]{quantile}}, \code{\link[stats]{relevel}}, \code{\link[stats:reorder.factor]{reorder}}}

  \item{utils}{\code{\link[utils]{capture.output}}, \code{\link[utils]{getS3method}}, \code{\link[utils]{packageName}}, \code{\link[utils:packageDescription]{packageVersion}}, \code{\link[utils]{sessionInfo}}}
}}

