% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-methods.R,
%   R/aggregateCellsToSamples-methods.R, R/autopadZeros-methods.R,
%   R/cell2sample-methods.R, R/collapseToString-methods.R, R/combine-methods.R,
%   R/convertGenesToSymbols-methods.R, R/convertSampleIDsToNames-methods.R,
%   R/convertTranscriptsToGenes-methods.R, R/counts-methods.R,
%   R/geneNames-methods.R, R/geometricMean-methods.R, R/headtail-methods.R,
%   R/humanize-methods.R, R/interestingGroups-methods.R,
%   R/logRatio-methods.R, R/mapGenes-methods.R, R/markdown-methods.R,
%   R/meltCounts-methods.R, R/metrics-methods.R, R/organism-methods.R,
%   R/plotCountsPerBiotype-methods.R, R/plotCountsPerGene-methods.R,
%   R/plotGene-methods.R, R/plotGenderMarkers-methods.R,
%   R/plotGenesDetected-methods.R, R/plotHeatmap-methods.R,
%   R/plotPCA-methods.R, R/plotQC-methods.R, R/plotTotalCounts-methods.R,
%   R/plotZerosVsDepth-methods.R, R/reexports.R, R/removeNA-methods.R,
%   R/sampleData-methods.R, R/sampleNames-methods.R, R/sanitizeNA-methods.R,
%   R/sanitizePercent-methods.R, R/selectSamples-methods.R,
%   R/show-methods.R, R/stripTranscriptVersions-methods.R,
%   R/subsetPerSample-methods.R, R/topCellsPerSample-methods.R, R/tpm-methods.R,
%   R/uniteInterestingGroups-methods.R, R/zerosVsDepth-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assignAndSaveData}
\alias{atomize}
\alias{basenameSansExt}
\alias{decode}
\alias{dots}
\alias{encode}
\alias{export}
\alias{extPattern}
\alias{factorize}
\alias{import}
\alias{initDir}
\alias{loadData}
\alias{loadDataAsName}
\alias{loadRemoteData}
\alias{localOrRemoteFile}
\alias{naStrings}
\alias{pasteURL}
\alias{rdataExtPattern}
\alias{rdataLoadError}
\alias{realpath}
\alias{saveData}
\alias{transmit}
\alias{writeCounts}
\alias{extractLocal}
\alias{hasLocal}
\alias{matchArgsToDoCall}
\alias{methodFormals}
\alias{methodFunction}
\alias{standardizeCall}
\alias{\%>\%}
\alias{camel}
\alias{dotted}
\alias{makeDimnames}
\alias{makeNames}
\alias{snake}
\alias{upperCamel}
\alias{as_tibble}
\alias{tibble}
\alias{as.SummarizedExperiment}
\alias{coerceS4ToList}
\alias{flatFiles}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Biobase}{\code{\link[Biobase]{sampleNames}}, \code{\link[Biobase]{sampleNames<-}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{organism}}, \code{\link[BiocGenerics]{plotPCA}}}

  \item{bioverbs}{\code{\link[bioverbs]{aggregateRows}}, \code{\link[bioverbs]{aggregateCols}}, \code{\link[bioverbs]{aggregateCellsToSamples}}, \code{\link[bioverbs]{autopadZeros}}, \code{\link[bioverbs]{cell2sample}}, \code{\link[bioverbs]{collapseToString}}, \code{\link[bioverbs]{convertGenesToSymbols}}, \code{\link[bioverbs]{convertSymbolsToGenes}}, \code{\link[bioverbs]{convertSampleIDsToNames}}, \code{\link[bioverbs]{convertTranscriptsToGenes}}, \code{\link[bioverbs]{geneNames}}, \code{\link[bioverbs]{geometricMean}}, \code{\link[bioverbs]{headtail}}, \code{\link[bioverbs]{humanize}}, \code{\link[bioverbs]{interestingGroups}}, \code{\link[bioverbs]{interestingGroups<-}}, \code{\link[bioverbs]{foldChangeToLogRatio}}, \code{\link[bioverbs]{logRatioToFoldChange}}, \code{\link[bioverbs]{mapGenesToRownames}}, \code{\link[bioverbs]{mapGenesToIDs}}, \code{\link[bioverbs]{mapGenesToSymbols}}, \code{\link[bioverbs]{markdown}}, \code{\link[bioverbs]{meltCounts}}, \code{\link[bioverbs]{metrics}}, \code{\link[bioverbs]{metricsPerSample}}, \code{\link[bioverbs]{plotCountsPerBiotype}}, \code{\link[bioverbs]{plotCountsPerBroadClass}}, \code{\link[bioverbs]{plotCountsPerGene}}, \code{\link[bioverbs]{plotGene}}, \code{\link[bioverbs]{plotGenderMarkers}}, \code{\link[bioverbs]{plotGenesDetected}}, \code{\link[bioverbs]{plotHeatmap}}, \code{\link[bioverbs]{plotCorrelationHeatmap}}, \code{\link[bioverbs]{plotQuantileHeatmap}}, \code{\link[bioverbs]{plotQC}}, \code{\link[bioverbs]{plotTotalCounts}}, \code{\link[bioverbs]{plotZerosVsDepth}}, \code{\link[bioverbs]{removeNA}}, \code{\link[bioverbs]{sampleData}}, \code{\link[bioverbs]{sampleData<-}}, \code{\link[bioverbs]{sanitizeNA}}, \code{\link[bioverbs]{sanitizePercent}}, \code{\link[bioverbs]{selectSamples}}, \code{\link[bioverbs]{stripTranscriptVersions}}, \code{\link[bioverbs]{subsetPerSample}}, \code{\link[bioverbs]{topCellsPerSample}}, \code{\link[bioverbs]{uniteInterestingGroups}}, \code{\link[bioverbs]{zerosVsDepth}}}

  \item{brio}{\code{\link[brio]{assignAndSaveData}}, \code{\link[brio]{atomize}}, \code{\link[brio]{basenameSansExt}}, \code{\link[brio]{decode}}, \code{\link[brio]{dots}}, \code{\link[brio]{encode}}, \code{\link[brio]{export}}, \code{\link[brio]{extPattern}}, \code{\link[brio]{factorize}}, \code{\link[brio]{import}}, \code{\link[brio]{initDir}}, \code{\link[brio]{loadData}}, \code{\link[brio]{loadDataAsName}}, \code{\link[brio]{loadRemoteData}}, \code{\link[brio]{localOrRemoteFile}}, \code{\link[brio]{naStrings}}, \code{\link[brio]{pasteURL}}, \code{\link[brio]{rdataExtPattern}}, \code{\link[brio]{rdataLoadError}}, \code{\link[brio]{realpath}}, \code{\link[brio]{saveData}}, \code{\link[brio]{transmit}}, \code{\link[brio]{writeCounts}}}

  \item{goalie}{\code{\link[goalie]{extractLocal}}, \code{\link[goalie]{hasLocal}}, \code{\link[goalie]{matchArgsToDoCall}}, \code{\link[goalie]{methodFormals}}, \code{\link[goalie]{methodFunction}}, \code{\link[goalie]{standardizeCall}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{tpm}}}

  \item{syntactic}{\code{\link[syntactic]{camel}}, \code{\link[syntactic]{dotted}}, \code{\link[syntactic]{makeDimnames}}, \code{\link[syntactic]{makeNames}}, \code{\link[syntactic]{snake}}, \code{\link[syntactic]{upperCamel}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}}

  \item{transformer}{\code{\link[transformer]{as.SummarizedExperiment}}, \code{\link[transformer]{coerceS4ToList}}, \code{\link[transformer]{flatFiles}}}
}}

