% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/functions.R, R/generics.R
\docType{import}
\name{Annotated-class}
\alias{Annotated-class}
\alias{AsIs-class}
\alias{AtomicList-class}
\alias{BiocFile-class}
\alias{CellCycleMarkers-class}
\alias{CellTypeMarkers-class}
\alias{CharacterList-class}
\alias{DataFrame-class}
\alias{DataFrameList-class}
\alias{Ensembl2Entrez-class}
\alias{EnsemblGenes-class}
\alias{EnsemblTranscripts-class}
\alias{Entrez2Ensembl-class}
\alias{EntrezGeneInfo-class}
\alias{Factor-class}
\alias{FactorList-class}
\alias{FlyBaseGenes-class}
\alias{FlyBaseTranscripts-class}
\alias{GencodeGenes-class}
\alias{GencodeTranscripts-class}
\alias{Gene2Symbol-class}
\alias{GenomicRanges-class}
\alias{GenomicRangesList-class}
\alias{HGNC-class}
\alias{HGNC2Ensembl-class}
\alias{IntegerList-class}
\alias{IntegerRanges-class}
\alias{KnownMarkers-class}
\alias{LLint-class}
\alias{List-class}
\alias{LogicalList-class}
\alias{Matrix-class}
\alias{MGI2Ensembl-class}
\alias{NumericList-class}
\alias{Protein2Gene-class}
\alias{RangedSummarizedExperiment-class}
\alias{Ranges-class}
\alias{RectangularData-class}
\alias{RefSeqGenes-class}
\alias{RefSeqTranscripts-class}
\alias{Rle-class}
\alias{RleList-class}
\alias{SimpleList-class}
\alias{SingleCellExperiment-class}
\alias{SplitDataFrameList-class}
\alias{SummarizedExperiment-class}
\alias{Tx2Gene-class}
\alias{UCSCGenes-class}
\alias{UCSCTranscripts-class}
\alias{Vector-class}
\alias{WormBaseGenes-class}
\alias{WormBaseTranscripts-class}
\alias{atomic-class}
\alias{missingOrNULL-class}
\alias{sparseMatrix-class}
\alias{reexports}
\alias{CellCycleMarkers}
\alias{CellTypeMarkers}
\alias{CharacterList}
\alias{DataFrame}
\alias{DataFrameList}
\alias{EntrezGeneInfo}
\alias{Factor}
\alias{FactorList}
\alias{GRanges}
\alias{GRangesList}
\alias{HGNC}
\alias{HGNC2Ensembl}
\alias{Matrix}
\alias{MGI2Ensembl}
\alias{NumericList}
\alias{Rle}
\alias{RleList}
\alias{SimpleList}
\alias{SingleCellExperiment}
\alias{SplitDataFrameList}
\alias{SummarizedExperiment}
\alias{abort}
\alias{alert}
\alias{alertDanger}
\alias{alertInfo}
\alias{alertSuccess}
\alias{alertWarning}
\alias{appendToBody}
\alias{as.formula}
\alias{assignAndSaveData}
\alias{barcodePattern}
\alias{basenameSansExt}
\alias{cacheURL}
\alias{classContains}
\alias{collapseToPathString}
\alias{column_to_rownames}
\alias{compress}
\alias{currentEnsemblGenomeBuild}
\alias{currentEnsemblVersion}
\alias{currentFlyBaseVersion}
\alias{currentGencodeGenomeBuild}
\alias{currentGencodeVersion}
\alias{currentRefSeqGenomeBuild}
\alias{currentRefSeqVersion}
\alias{currentUCSCGenomeBuild}
\alias{currentWormBaseVersion}
\alias{data.table}
\alias{decompress}
\alias{detectLanes}
\alias{detectOrganism}
\alias{dist}
\alias{dl}
\alias{dots}
\alias{download}
\alias{downloadEnsemblGenome}
\alias{downloadGencodeGenome}
\alias{downloadRefSeqGenome}
\alias{downloadUCSCGenome}
\alias{ecdf}
\alias{emptyRanges}
\alias{fileDepth}
\alias{fileExt}
\alias{forceDetach}
\alias{formula}
\alias{geneSynonyms}
\alias{genomeMetadataNames}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{h7}
\alias{hclust}
\alias{importCellCycleMarkers}
\alias{importCellTypeMarkers}
\alias{importSampleData}
\alias{importTx2Gene}
\alias{inform}
\alias{initDir}
\alias{lanePattern}
\alias{loadData}
\alias{loadDataAsName}
\alias{loadRemoteData}
\alias{makeGRangesFromEnsDb}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromGFF}
\alias{makeGene2SymbolFromEnsDb}
\alias{makeGene2SymbolFromEnsembl}
\alias{makeGene2SymbolFromGFF}
\alias{makeProtein2GeneFromEnsembl}
\alias{makeSingleCellExperiment}
\alias{makeSummarizedExperiment}
\alias{makeTx2GeneFromEnsDb}
\alias{makeTx2GeneFromEnsembl}
\alias{makeTx2GeneFromFASTA}
\alias{makeTx2GeneFromGFF}
\alias{majorVersion}
\alias{mapCellsToSamples}
\alias{mapEnsemblReleaseToURL}
\alias{mapHumanOrthologs}
\alias{markdownHeader}
\alias{markdownLink}
\alias{markdownList}
\alias{markdownPlots}
\alias{markdownTables}
\alias{matchInterestingGroups}
\alias{mdHeader}
\alias{mdLink}
\alias{mdList}
\alias{mdPlots}
\alias{mdTables}
\alias{metadataDenylist}
\alias{methodFormals}
\alias{methodFunction}
\alias{metricsCols}
\alias{microplate}
\alias{minimalSampleData}
\alias{minorVersion}
\alias{model.matrix}
\alias{naStrings}
\alias{ol}
\alias{parentDir}
\alias{parentDirectory}
\alias{pasteURL}
\alias{prcomp}
\alias{prepareTemplate}
\alias{printString}
\alias{realpath}
\alias{relevel}
\alias{reorder}
\alias{requireNamespaces}
\alias{rownames_to_column}
\alias{sanitizeSampleData}
\alias{sanitizeVersion}
\alias{saveData}
\alias{separator}
\alias{simpleClass}
\alias{shell}
\alias{showSlotInfo}
\alias{sparseMatrix}
\alias{splitPathString}
\alias{standardizeCall}
\alias{tibble}
\alias{toInlineString}
\alias{transmit}
\alias{txt}
\alias{ul}
\alias{updateMessage}
\alias{verbatim}
\alias{warn}
\alias{as_tibble}
\alias{as.data.table}
\alias{Ensembl2Entrez}
\alias{Entrez2Ensembl}
\alias{Filter}
\alias{Find}
\alias{Gene2Symbol}
\alias{KnownMarkers}
\alias{Map}
\alias{Position}
\alias{Reduce}
\alias{Tx2Gene}
\alias{\%in\%}
\alias{aggregate}
\alias{aggregateCellsToSamples}
\alias{aggregateCols}
\alias{aggregateRows}
\alias{antiJoin}
\alias{anyDuplicated}
\alias{append}
\alias{as.DataFrame}
\alias{as.SummarizedExperiment}
\alias{as.data.frame}
\alias{as.factor}
\alias{as.list}
\alias{as.matrix}
\alias{as.table}
\alias{as.vector}
\alias{assay}
\alias{assay<-}
\alias{assayNames}
\alias{assayNames<-}
\alias{assays}
\alias{assays<-}
\alias{atomize}
\alias{autopadZeros}
\alias{barcodeRanksPerSample}
\alias{basename}
\alias{calculateMetrics}
\alias{camelCase}
\alias{capitalize}
\alias{cbind}
\alias{cell2sample}
\alias{cellCountsPerCluster}
\alias{cellTypesPerCluster}
\alias{clusters}
\alias{coerceToList}
\alias{colData}
\alias{colData<-}
\alias{colSums}
\alias{collapseToString}
\alias{colnames}
\alias{colnames<-}
\alias{combine}
\alias{complete.cases}
\alias{convertGenesToSymbols}
\alias{convertSampleIDsToNames}
\alias{convertSymbolsToGenes}
\alias{convertTranscriptsToGenes}
\alias{cor}
\alias{correlation}
\alias{counts}
\alias{counts<-}
\alias{cpm}
\alias{decode}
\alias{design}
\alias{diffExp}
\alias{diffExpPerCluster}
\alias{dims}
\alias{dirname}
\alias{do.call}
\alias{dottedCase}
\alias{droplevels}
\alias{droplevels2}
\alias{duplicated}
\alias{encode}
\alias{end}
\alias{end<-}
\alias{estimateDispersions}
\alias{estimateSizeFactors}
\alias{eval}
\alias{expand}
\alias{expand.grid}
\alias{export}
\alias{factorize}
\alias{filterCells}
\alias{findMarkers}
\alias{foldChangeToLogRatio}
\alias{fullJoin}
\alias{geneNames}
\alias{genome}
\alias{genome<-}
\alias{geometricMean}
\alias{get}
\alias{getListElement}
\alias{grep}
\alias{grepl}
\alias{gsub}
\alias{head}
\alias{headtail}
\alias{humanize}
\alias{import}
\alias{innerJoin}
\alias{integerCounts}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{intersect}
\alias{intersectAll}
\alias{intersectionMatrix}
\alias{is.unsorted}
\alias{kebabCase}
\alias{lapply}
\alias{leftJoin}
\alias{logRatioToFoldChange}
\alias{logcounts}
\alias{logcounts<-}
\alias{makeDimnames}
\alias{makeLabel}
\alias{makeNames}
\alias{makeSampleData}
\alias{makeTitle}
\alias{makeWords}
\alias{mapGenesToIDs}
\alias{mapGenesToRownames}
\alias{mapGenesToSymbols}
\alias{mapToDataFrame}
\alias{mapply}
\alias{match}
\alias{matchSampleColumn}
\alias{mcols}
\alias{mcols<-}
\alias{mean}
\alias{median}
\alias{melt}
\alias{merge}
\alias{metadata}
\alias{metadata<-}
\alias{metadata2}
\alias{metadata2<-}
\alias{metrics}
\alias{metricsPerSample}
\alias{mget}
\alias{mutateAll}
\alias{mutateAt}
\alias{mutateIf}
\alias{na.omit}
\alias{ncol}
\alias{nonzeroRowsAndCols}
\alias{normalize}
\alias{normcounts}
\alias{normcounts<-}
\alias{nrow}
\alias{order}
\alias{organism}
\alias{organism<-}
\alias{paste}
\alias{pmax}
\alias{pmax.int}
\alias{pmin}
\alias{pmin.int}
\alias{pos}
\alias{quantile}
\alias{ranges}
\alias{rank}
\alias{rankedMatrix}
\alias{rbind}
\alias{rbindToDataFrame}
\alias{reducedDim}
\alias{reducedDim<-}
\alias{reducedDimNames}
\alias{reducedDimNames<-}
\alias{reducedDims}
\alias{reducedDims<-}
\alias{relist}
\alias{removeNA}
\alias{rep.int}
\alias{rightJoin}
\alias{rowData}
\alias{rowData<-}
\alias{rowMeans}
\alias{rowRanges}
\alias{rowRanges<-}
\alias{rowSums}
\alias{rownames}
\alias{rownames<-}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{sanitizeNA}
\alias{sanitizePercent}
\alias{sapply}
\alias{sd}
\alias{selectIf}
\alias{selectSamples}
\alias{sem}
\alias{semiJoin}
\alias{sentenceCase}
\alias{seqinfo}
\alias{seqinfo<-}
\alias{seqlevels}
\alias{seqlevels<-}
\alias{seqnames}
\alias{seqnames<-}
\alias{setdiff}
\alias{showHeader}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{snakeCase}
\alias{sort}
\alias{split}
\alias{splitByLevel}
\alias{start}
\alias{start<-}
\alias{strand}
\alias{strand<-}
\alias{stripGeneVersions}
\alias{stripTranscriptVersions}
\alias{sub}
\alias{subset}
\alias{subsetPerSample}
\alias{summary}
\alias{t}
\alias{table}
\alias{tail}
\alias{tapply}
\alias{topCellsPerSample}
\alias{trim}
\alias{tpm}
\alias{transmuteAt}
\alias{transmuteIf}
\alias{union}
\alias{unique}
\alias{uniteInterestingGroups}
\alias{unlist}
\alias{unsplit}
\alias{updateObject}
\alias{upperCamelCase}
\alias{var}
\alias{weights}
\alias{which}
\alias{which.max}
\alias{which.min}
\alias{width}
\alias{width<-}
\alias{zerosVsDepth}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{appendToBody}}, \code{\link[AcidBase]{barcodePattern}}, \code{\link[AcidBase]{basenameSansExt}}, \code{\link[AcidBase]{classContains}}, \code{\link[AcidBase]{collapseToPathString}}, \code{\link[AcidBase]{compress}}, \code{\link[AcidBase:compress]{decompress}}, \code{\link[AcidBase]{dots}}, \code{\link[AcidBase]{download}}, \code{\link[AcidBase]{fileDepth}}, \code{\link[AcidBase]{fileExt}}, \code{\link[AcidBase]{forceDetach}}, \code{\link[AcidBase]{genomeMetadataNames}}, \code{\link[AcidBase]{initDir}}, \code{\link[AcidBase]{lanePattern}}, \code{\link[AcidBase]{majorVersion}}, \code{\link[AcidBase]{metadataDenylist}}, \code{\link[AcidBase:methodDefinition]{methodFormals}}, \code{\link[AcidBase:methodDefinition]{methodFunction}}, \code{\link[AcidBase]{metricsCols}}, \code{\link[AcidBase]{minorVersion}}, \code{\link[AcidBase:parentDirectory]{parentDir}}, \code{\link[AcidBase]{parentDirectory}}, \code{\link[AcidBase]{pasteURL}}, \code{\link[AcidBase]{printString}}, \code{\link[AcidBase]{realpath}}, \code{\link[AcidBase]{requireNamespaces}}, \code{\link[AcidBase]{sanitizeVersion}}, \code{\link[AcidBase]{shell}}, \code{\link[AcidBase]{showSlotInfo}}, \code{\link[AcidBase]{simpleClass}}, \code{\link[AcidBase]{splitPathString}}, \code{\link[AcidBase]{standardizeCall}}, \code{\link[AcidBase]{updateMessage}}}

  \item{AcidCLI}{\code{\link[AcidCLI]{abort}}, \code{\link[AcidCLI]{alert}}, \code{\link[AcidCLI:alert]{alertDanger}}, \code{\link[AcidCLI:alert]{alertInfo}}, \code{\link[AcidCLI:alert]{alertSuccess}}, \code{\link[AcidCLI:alert]{alertWarning}}, \code{\link[AcidCLI:list]{dl}}, \code{\link[AcidCLI:header]{h1}}, \code{\link[AcidCLI:header]{h2}}, \code{\link[AcidCLI:header]{h3}}, \code{\link[AcidCLI:header]{h4}}, \code{\link[AcidCLI:header]{h5}}, \code{\link[AcidCLI:header]{h6}}, \code{\link[AcidCLI:header]{h7}}, \code{\link[AcidCLI:abort]{inform}}, \code{\link[AcidCLI:list]{ol}}, \code{\link[AcidCLI]{separator}}, \code{\link[AcidCLI]{toInlineString}}, \code{\link[AcidCLI:text]{txt}}, \code{\link[AcidCLI:list]{ul}}, \code{\link[AcidCLI:text]{verbatim}}, \code{\link[AcidCLI:abort]{warn}}}

  \item{AcidExperiment}{\code{\link[AcidExperiment]{detectLanes}}, \code{\link[AcidExperiment]{importSampleData}}, \code{\link[AcidExperiment]{makeSummarizedExperiment}}, \code{\link[AcidExperiment]{matchInterestingGroups}}, \code{\link[AcidExperiment]{microplate}}, \code{\link[AcidExperiment]{minimalSampleData}}, \code{\link[AcidExperiment]{sanitizeSampleData}}}

  \item{AcidGenomes}{\code{\link[AcidGenomes]{EntrezGeneInfo}}, \code{\link[AcidGenomes]{HGNC}}, \code{\link[AcidGenomes]{HGNC2Ensembl}}, \code{\link[AcidGenomes]{MGI2Ensembl}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentEnsemblGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentEnsemblVersion}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentFlyBaseVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentGencodeGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentGencodeVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentRefSeqGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentRefSeqVersion}}, \code{\link[AcidGenomes:currentGenomeBuild]{currentUCSCGenomeBuild}}, \code{\link[AcidGenomes:currentGenomeVersion]{currentWormBaseVersion}}, \code{\link[AcidGenomes]{detectOrganism}}, \code{\link[AcidGenomes]{downloadEnsemblGenome}}, \code{\link[AcidGenomes]{downloadGencodeGenome}}, \code{\link[AcidGenomes]{downloadRefSeqGenome}}, \code{\link[AcidGenomes]{downloadUCSCGenome}}, \code{\link[AcidGenomes]{emptyRanges}}, \code{\link[AcidGenomes]{geneSynonyms}}, \code{\link[AcidGenomes]{importTx2Gene}}, \code{\link[AcidGenomes:makeGRangesFromEnsembl]{makeGRangesFromEnsDb}}, \code{\link[AcidGenomes]{makeGRangesFromEnsembl}}, \code{\link[AcidGenomes]{makeGRangesFromGFF}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromEnsDb}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromEnsembl}}, \code{\link[AcidGenomes:makeGene2Symbol]{makeGene2SymbolFromGFF}}, \code{\link[AcidGenomes:makeProtein2Gene]{makeProtein2GeneFromEnsembl}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromEnsDb}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromEnsembl}}, \code{\link[AcidGenomes]{makeTx2GeneFromFASTA}}, \code{\link[AcidGenomes:makeTx2Gene]{makeTx2GeneFromGFF}}, \code{\link[AcidGenomes]{mapEnsemblReleaseToURL}}, \code{\link[AcidGenomes]{mapHumanOrthologs}}}

  \item{AcidMarkdown}{\code{\link[AcidMarkdown]{markdownHeader}}, \code{\link[AcidMarkdown]{markdownLink}}, \code{\link[AcidMarkdown]{markdownList}}, \code{\link[AcidMarkdown]{markdownPlots}}, \code{\link[AcidMarkdown]{markdownTables}}, \code{\link[AcidMarkdown:markdownHeader]{mdHeader}}, \code{\link[AcidMarkdown:markdownLink]{mdLink}}, \code{\link[AcidMarkdown:markdownList]{mdList}}, \code{\link[AcidMarkdown:markdownPlots]{mdPlots}}, \code{\link[AcidMarkdown:markdownTables]{mdTables}}, \code{\link[AcidMarkdown]{prepareTemplate}}}

  \item{AcidSingleCell}{\code{\link[AcidSingleCell]{CellCycleMarkers}}, \code{\link[AcidSingleCell]{CellTypeMarkers}}, \code{\link[AcidSingleCell:CellCycleMarkers]{importCellCycleMarkers}}, \code{\link[AcidSingleCell:CellTypeMarkers]{importCellTypeMarkers}}, \code{\link[AcidSingleCell]{makeSingleCellExperiment}}, \code{\link[AcidSingleCell]{mapCellsToSamples}}}

  \item{data.table}{\code{\link[data.table]{data.table}}}

  \item{GenomicRanges}{\code{\link[GenomicRanges:GRanges-class]{GRanges}}, \code{\link[GenomicRanges:GRanges-class]{GRanges}}, \code{\link[GenomicRanges:GRangesList-class]{GRangesList}}, \code{\link[GenomicRanges:GRangesList-class]{GRangesList}}}

  \item{IRanges}{\code{\link[IRanges:AtomicList-class]{CharacterList}}, \code{\link[IRanges:DataFrameList-class]{DataFrameList}}, \code{\link[IRanges:AtomicList-class]{FactorList}}, \code{\link[IRanges:AtomicList-class]{NumericList}}, \code{\link[IRanges:AtomicList-class]{RleList}}, \code{\link[IRanges:DataFrameList-class]{SplitDataFrameList}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}, \code{\link[Matrix]{sparseMatrix}}}

  \item{pipette}{\code{\link[pipette]{assignAndSaveData}}, \code{\link[pipette]{cacheURL}}, \code{\link[pipette]{loadData}}, \code{\link[pipette]{loadDataAsName}}, \code{\link[pipette]{loadRemoteData}}, \code{\link[pipette]{naStrings}}, \code{\link[pipette]{saveData}}, \code{\link[pipette]{transmit}}}

  \item{S4Vectors}{\code{\link[S4Vectors:DataFrame-class]{DataFrame}}, \code{\link[S4Vectors:Factor-class]{Factor}}, \code{\link[S4Vectors:Rle-class]{Rle}}, \code{\link[S4Vectors:SimpleList-class]{SimpleList}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{SingleCellExperiment}}}

  \item{stats}{\code{\link[stats:formula]{as.formula}}, \code{\link[stats]{dist}}, \code{\link[stats]{ecdf}}, \code{\link[stats]{formula}}, \code{\link[stats]{hclust}}, \code{\link[stats]{model.matrix}}, \code{\link[stats]{prcomp}}, \code{\link[stats]{relevel}}, \code{\link[stats:reorder.factor]{reorder}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}}}

  \item{tibble}{\code{\link[tibble:rownames]{column_to_rownames}}, \code{\link[tibble:rownames]{rownames_to_column}}, \code{\link[tibble]{tibble}}}
}}

