% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{!!!}
\alias{reexports}
\alias{!!}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{is_string}
\alias{reexports}
\alias{left_join}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{read_csv}
\alias{reexports}
\alias{read_delim}
\alias{reexports}
\alias{read_excel}
\alias{reexports}
\alias{read_tsv}
\alias{reexports}
\alias{remove_rownames}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{set_colnames}
\alias{reexports}
\alias{set_names}
\alias{reexports}
\alias{set_rownames}
\alias{reexports}
\alias{sym}
\alias{reexports}
\alias{syms}
\alias{reexports}
\alias{top_n}
\alias{reexports}
\alias{write_csv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{top_n}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{set_colnames}}, \code{\link[magrittr]{set_names}}, \code{\link[magrittr]{set_rownames}}}

  \item{readr}{\code{\link[readr]{read_csv}}, \code{\link[readr]{read_delim}}, \code{\link[readr]{read_tsv}}, \code{\link[readr]{write_csv}}}

  \item{readxl}{\code{\link[readxl]{read_excel}}}

  \item{rlang}{\code{\link[rlang]{!!!}}, \code{\link[rlang]{!!}}, \code{\link[rlang]{is_string}}, \code{\link[rlang]{sym}}, \code{\link[rlang]{syms}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{remove_rownames}}}
}}

