% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-genes.R, R/methods-transcripts.R,
%   R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{genes}
\alias{reexports}
\alias{transcripts}
\alias{reexports}
\alias{Matrix}
\alias{reexports}
\alias{DataFrame}
\alias{reexports}
\alias{mcols}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{set_rownames}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{is_tibble}
\alias{reexports}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GenomicFeatures}{\code{\link[GenomicFeatures]{genes}}, \code{\link[GenomicFeatures]{transcripts}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{set_rownames}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{mcols}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{is_tibble}}, \code{\link[tibble]{tibble}}}
}}

