% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pull}
\alias{reexports}
\alias{path_expand}
\alias{reexports}
\alias{path_join}
\alias{reexports}
\alias{path_real}
\alias{reexports}
\alias{aes}
\alias{reexports}
\alias{ggplot}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{set_rownames}
\alias{reexports}
\alias{Matrix}
\alias{reexports}
\alias{DataFrame}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{is_tibble}
\alias{reexports}
\alias{tibble}
\alias{reexports}
\alias{viridis}
\alias{reexports}
\alias{scale_color_viridis}
\alias{reexports}
\alias{scale_fill_viridis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{fs}{\code{\link[fs]{path_expand}}, \code{\link[fs]{path_join}}, \code{\link[fs]{path_real}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{ggplot}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{set_rownames}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{is_tibble}}, \code{\link[tibble]{tibble}}}

  \item{viridis}{\code{\link[viridis]{viridis}}, \code{\link[viridis]{scale_color_viridis}}, \code{\link[viridis]{scale_fill_viridis}}}
}}

