% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R, R/reexports.R,
%   R/tidyverse-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{pull}
\alias{seqnames}
\alias{\%>\%}
\alias{set_rownames}
\alias{Matrix}
\alias{DataFrame}
\alias{mcols}
\alias{SummarizedExperiment}
\alias{glimpse}
\alias{tibble}
\alias{arrange}
\alias{filter}
\alias{left_join}
\alias{mutate}
\alias{rename}
\alias{select}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{pull}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}}

  \item{GenomeInfoDb}{\code{\link[GenomeInfoDb]{seqnames}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{set_rownames}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{mcols}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment]{SummarizedExperiment}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{glimpse}}, \code{\link[tibble]{tibble}}}
}}

