% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pull}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{Matrix}
\alias{reexports}
\alias{DataFrame}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tibble}}}
}}

