% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{colMeans}
\alias{reexports}
\alias{colSums}
\alias{reexports}
\alias{rowMeans}
\alias{reexports}
\alias{rowSums}
\alias{reexports}
\alias{biocValid}
\alias{reexports}
\alias{getBM}
\alias{reexports}
\alias{listMarts}
\alias{reexports}
\alias{useEnsembl}
\alias{reexports}
\alias{build}
\alias{reexports}
\alias{build_vignettes}
\alias{reexports}
\alias{check}
\alias{reexports}
\alias{document}
\alias{reexports}
\alias{install}
\alias{reexports}
\alias{load_all}
\alias{reexports}
\alias{test}
\alias{reexports}
\alias{case_when}
\alias{reexports}
\alias{everything}
\alias{reexports}
\alias{funs}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{left_join}
\alias{reexports}
\alias{mutate_all}
\alias{reexports}
\alias{summarise_all}
\alias{reexports}
\alias{top_n}
\alias{reexports}
\alias{ungroup}
\alias{reexports}
\alias{hist}
\alias{reexports}
\alias{asis_output}
\alias{reexports}
\alias{kable}
\alias{reexports}
\alias{opts_chunk}
\alias{reexports}
\alias{opts_knit}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{set_colnames}
\alias{reexports}
\alias{set_names}
\alias{reexports}
\alias{set_rownames}
\alias{reexports}
\alias{as}
\alias{reexports}
\alias{show}
\alias{reexports}
\alias{gunzip}
\alias{reexports}
\alias{gzip}
\alias{reexports}
\alias{getURL}
\alias{reexports}
\alias{read_csv}
\alias{reexports}
\alias{read_delim}
\alias{reexports}
\alias{read_lines}
\alias{reexports}
\alias{read_table}
\alias{reexports}
\alias{read_tsv}
\alias{reexports}
\alias{write_csv}
\alias{reexports}
\alias{write_lines}
\alias{reexports}
\alias{write_tsv}
\alias{reexports}
\alias{read_excel}
\alias{reexports}
\alias{!!!}
\alias{reexports}
\alias{!!}
\alias{reexports}
\alias{.data}
\alias{reexports}
\alias{is_string}
\alias{reexports}
\alias{quo}
\alias{reexports}
\alias{quos}
\alias{reexports}
\alias{sym}
\alias{reexports}
\alias{syms}
\alias{reexports}
\alias{render}
\alias{reexports}
\alias{aggregate}
\alias{reexports}
\alias{median}
\alias{reexports}
\alias{na.omit}
\alias{reexports}
\alias{setNames}
\alias{reexports}
\alias{str_detect}
\alias{reexports}
\alias{str_extract}
\alias{reexports}
\alias{str_extract_all}
\alias{reexports}
\alias{str_match}
\alias{reexports}
\alias{str_match_all}
\alias{reexports}
\alias{str_pad}
\alias{reexports}
\alias{str_replace}
\alias{reexports}
\alias{str_replace_all}
\alias{reexports}
\alias{str_split}
\alias{reexports}
\alias{str_subset}
\alias{reexports}
\alias{as_tibble}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{remove_rownames}
\alias{reexports}
\alias{tibble}
\alias{reexports}
\alias{expand_}
\alias{reexports}
\alias{nest}
\alias{reexports}
\alias{separate_}
\alias{reexports}
\alias{unnest}
\alias{reexports}
\alias{download.file}
\alias{reexports}
\alias{globalVariables}
\alias{reexports}
\alias{sessionInfo}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AnnotationDbi}{\code{\link[AnnotationDbi]{select}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{colMeans}}, \code{\link[BiocGenerics]{colSums}}, \code{\link[BiocGenerics]{rowMeans}}, \code{\link[BiocGenerics]{rowSums}}}

  \item{BiocInstaller}{\code{\link[BiocInstaller]{biocValid}}}

  \item{biomaRt}{\code{\link[biomaRt]{getBM}}, \code{\link[biomaRt]{listMarts}}, \code{\link[biomaRt]{useEnsembl}}}

  \item{devtools}{\code{\link[devtools]{build}}, \code{\link[devtools]{build_vignettes}}, \code{\link[devtools]{check}}, \code{\link[devtools]{document}}, \code{\link[devtools]{install}}, \code{\link[devtools]{load_all}}, \code{\link[devtools]{test}}}

  \item{dplyr}{\code{\link[dplyr]{case_when}}, \code{\link[dplyr]{everything}}, \code{\link[dplyr]{funs}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{summarise_all}}, \code{\link[dplyr]{top_n}}, \code{\link[dplyr]{ungroup}}}

  \item{graphics}{\code{\link[graphics]{hist}}}

  \item{knitr}{\code{\link[knitr]{asis_output}}, \code{\link[knitr]{kable}}, \code{\link[knitr]{opts_chunk}}, \code{\link[knitr]{opts_knit}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{set_colnames}}, \code{\link[magrittr]{set_names}}, \code{\link[magrittr]{set_rownames}}}

  \item{methods}{\code{\link[methods]{as}}, \code{\link[methods]{show}}}

  \item{R.utils}{\code{\link[R.utils]{gunzip}}, \code{\link[R.utils]{gzip}}}

  \item{RCurl}{\code{\link[RCurl]{getURL}}}

  \item{readr}{\code{\link[readr]{read_csv}}, \code{\link[readr]{read_delim}}, \code{\link[readr]{read_lines}}, \code{\link[readr]{read_table}}, \code{\link[readr]{read_tsv}}, \code{\link[readr]{write_csv}}, \code{\link[readr]{write_lines}}, \code{\link[readr]{write_tsv}}}

  \item{readxl}{\code{\link[readxl]{read_excel}}}

  \item{rlang}{\code{\link[rlang]{!!!}}, \code{\link[rlang]{!!}}, \code{\link[rlang]{.data}}, \code{\link[rlang]{is_string}}, \code{\link[rlang]{quo}}, \code{\link[rlang]{quos}}, \code{\link[rlang]{sym}}, \code{\link[rlang]{syms}}}

  \item{rmarkdown}{\code{\link[rmarkdown]{render}}}

  \item{stats}{\code{\link[stats]{aggregate}}, \code{\link[stats]{median}}, \code{\link[stats]{na.omit}}, \code{\link[stats]{setNames}}}

  \item{stringr}{\code{\link[stringr]{str_detect}}, \code{\link[stringr]{str_extract}}, \code{\link[stringr]{str_extract_all}}, \code{\link[stringr]{str_match}}, \code{\link[stringr]{str_match_all}}, \code{\link[stringr]{str_pad}}, \code{\link[stringr]{str_replace}}, \code{\link[stringr]{str_replace_all}}, \code{\link[stringr]{str_split}}, \code{\link[stringr]{str_subset}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{glimpse}}, \code{\link[tibble]{remove_rownames}}, \code{\link[tibble]{tibble}}}

  \item{tidyr}{\code{\link[tidyr]{expand_}}, \code{\link[tidyr]{nest}}, \code{\link[tidyr]{separate_}}, \code{\link[tidyr]{unnest}}}

  \item{utils}{\code{\link[utils]{download.file}}, \code{\link[utils]{globalVariables}}, \code{\link[utils]{sessionInfo}}}
}}

