% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pull}
\alias{seqnames}
\alias{\%>\%}
\alias{set_rownames}
\alias{Matrix}
\alias{DataFrame}
\alias{mcols}
\alias{SummarizedExperiment}
\alias{glimpse}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{GenomeInfoDb}{\code{\link[GenomeInfoDb]{seqnames}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{set_rownames}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{mcols}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment]{SummarizedExperiment}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tibble}}}
}}

