% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{Matrix}
\alias{DataFrame}
\alias{tibble}
\alias{file_path_sans_ext}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{tools}{\code{\link[tools]{file_path_sans_ext}}}
}}

