% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{relativeLogExpression}
\alias{relativeLogExpression}
\title{Relative log expression}
\usage{
relativeLogExpression(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Relative log expression
}
\references{
Anders and Huber (2010).
}
\seealso{
\itemize{
\item \code{edgeR::calcNormFactors}.
\item \href{https://bioinformatics.sph.harvard.edu/bcbioRNASeq/}{bcbioRNASeq}.
}
}
