% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeNA}
\alias{removeNA}
\title{Remove rows and columns containing only \code{NA} values}
\usage{
removeNA(x)
}
\arguments{
\item{x}{Object with column data (e.g. data frame, matrix).}
}
\value{
Sanitized data.
}
\description{
Remove rows and columns containing only \code{NA} values
}
\examples{
# Remove NA only rows and columns
data.frame(a = c(1,  NA, 3),
           b = c(NA, NA, NA),
           c = c(4,  NA, 6)) \%>\% removeNA
}
