% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/removeNA-methods.R
\docType{methods}
\name{removeNA}
\alias{removeNA}
\alias{removeNA,atomic-method}
\alias{removeNA,matrix-method}
\alias{removeNA,sparseMatrix-method}
\alias{removeNA,data.frame-method}
\alias{removeNA,DataFrame-method}
\title{Remove Rows and Columns Containing Only \code{NA} Values}
\usage{
removeNA(object, ...)

\S4method{removeNA}{atomic}(object)

\S4method{removeNA}{matrix}(object)

\S4method{removeNA}{sparseMatrix}(object)

\S4method{removeNA}{data.frame}(object)

\S4method{removeNA}{DataFrame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Sanitized object.
}
\description{
Remove Rows and Columns Containing Only \code{NA} Values
}
\examples{
# atomic ====
removeNA(c("hello", "world", NA))
removeNA(c(1, 2, NA))

# matrix ====
from <- matrix(
    data = c(1, NA, 3, NA, NA, NA, 2, NA, 4),
    nrow = 3,
    ncol = 3
)
print(from)
to <- removeNA(from)
print(to)

# DataFrame ====
from <- DataFrame(
    a = c("A", NA, "C"),
    b = c(NA, NA, NA),
    c = c("B", NA, "D")
)
print(from)
to <- removeNA(from)
print(to)
}
\seealso{
Other Sanitization Functions: \code{\link{camel}},
  \code{\link{collapseToString}}, \code{\link{sanitizeNA}},
  \code{\link{sanitizePercent}},
  \code{\link{sanitizeRowData}},
  \code{\link{sanitizeSampleData}}
}
\author{
Michael Steinbaugh
}
\concept{Sanitization Functions}
