% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-removeNA.R
\docType{methods}
\name{removeNA}
\alias{removeNA}
\alias{removeNA}
\alias{removeNA,ANY-method}
\alias{removeNA,character-method}
\alias{removeNA,numeric-method}
\alias{removeNA,matrix-method}
\alias{removeNA,data.frame-method}
\alias{removeNA,DataFrame-method}
\alias{removeNA,tbl_df-method}
\title{Remove Rows and Columns Containing Only \code{NA} Values}
\usage{
removeNA(object, ...)

\S4method{removeNA}{ANY}(object)

\S4method{removeNA}{character}(object)

\S4method{removeNA}{numeric}(object)

\S4method{removeNA}{matrix}(object)

\S4method{removeNA}{data.frame}(object)

\S4method{removeNA}{DataFrame}(object)

\S4method{removeNA}{tbl_df}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Sanitized object.
}
\description{
Remove Rows and Columns Containing Only \code{NA} Values
}
\examples{
# Remove NA only rows and columns
matrix(
    c(1, NA, 3, NA, NA, NA, 2, NA, 4),
    nrow = 3,
    ncol = 3
) \%>\%
    removeNA()

data.frame(
    a = c("A", NA, "C"),
    b = c(NA, NA, NA),
    c = c("B", NA, "D"),
    stringsAsFactors = FALSE
) \%>\%
    removeNA()

tibble(
    a = c("A", NA, "C"),
    b = c(NA, NA, NA),
    c = c("B", NA, "D")
) \%>\%
    removeNA()

# Support for vectors
removeNA(c("hello", "world", NA))
removeNA(c(1, 2, NA))
}
\seealso{
Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{fixNA}}, \code{\link{sanitizeRowData}},
  \code{\link{sanitizeSampleData}}
}
\author{
Michael Steinbaugh
}
