% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-removeNA.R
\docType{methods}
\name{removeNA}
\alias{removeNA}
\alias{removeNA}
\alias{removeNA,ANY-method}
\alias{removeNA,atomic-method}
\alias{removeNA,matrix-method}
\alias{removeNA,data.frame-method}
\alias{removeNA,DataFrame-method}
\title{Remove Rows and Columns Containing Only \code{NA} Values}
\usage{
removeNA(object, ...)

\S4method{removeNA}{ANY}(object)

\S4method{removeNA}{atomic}(object)

\S4method{removeNA}{matrix}(object)

\S4method{removeNA}{data.frame}(object)

\S4method{removeNA}{DataFrame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Sanitized object.
}
\description{
Remove Rows and Columns Containing Only \code{NA} Values
}
\examples{
# matrix ====
x <- matrix(
    data = c(1, NA, 3, NA, NA, NA, 2, NA, 4),
    nrow = 3,
    ncol = 3
)
print(x)
removeNA(x)

# data.frame ====
x <- data.frame(
    a = c("A", NA, "C"),
    b = c(NA, NA, NA),
    c = c("B", NA, "D"),
    stringsAsFactors = FALSE
)
print(x)
removeNA(x)

# atomic ====
removeNA(c("hello", "world", NA))
removeNA(c(1, 2, NA))
}
\seealso{
Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{fixNA}}, \code{\link{sanitizeRowData}},
  \code{\link{sanitizeSampleData}}
}
\author{
Michael Steinbaugh
}
