% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-removeNA.R
\docType{methods}
\name{removeNA}
\alias{removeNA}
\alias{removeNA,character-method}
\alias{removeNA,numeric-method}
\alias{removeNA,matrix-method}
\alias{removeNA,data.frame-method}
\alias{removeNA,DataFrame-method}
\alias{removeNA,tbl_df-method}
\alias{removeNA,ANY-method}
\title{Remove Rows and Columns Containing Only \code{NA} Values}
\usage{
removeNA(object, ...)

\S4method{removeNA}{character}(object)

\S4method{removeNA}{numeric}(object)

\S4method{removeNA}{matrix}(object)

\S4method{removeNA}{data.frame}(object)

\S4method{removeNA}{DataFrame}(object)

\S4method{removeNA}{tbl_df}(object)

\S4method{removeNA}{ANY}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Sanitized data.
}
\description{
Remove Rows and Columns Containing Only \code{NA} Values
}
\examples{
# Remove NA only rows and columns
matrix(c(1, NA, 3,
         NA, NA, NA,
         2, NA, 4),
       nrow = 3, ncol = 3) \%>\% removeNA

data.frame(a = c("A", NA, "C"),
           b = c(NA, NA, NA),
           c = c("B", NA, "D")) \%>\% removeNA

tibble(a = c("A", NA, "C"),
       b = c(NA, NA, NA),
       c = c("B", NA, "D")) \%>\% removeNA


# Support for vectors (using `stats::na.omit()`)
removeNA(c("hello", "world", NA))
removeNA(c(1, 2, NA))
}
\seealso{
Other Cleanup Utilities: \code{\link{camel}},
  \code{\link{fixNA}}, \code{\link{sortUnique}}
}
\concept{Cleanup Utilities}
