% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderall.R
\name{renderall}
\alias{renderall}
\title{Render all RMarkdown files}
\usage{
renderall(...)
}
\arguments{
\item{...}{Optional rmarkdown parameters}
}
\description{
Render all RMarkdown files
}
\examples{
renderall(output_dir = "docs")
}

