% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderAll.R
\name{renderAll}
\alias{renderAll}
\title{Render all RMarkdown files in working directory}
\usage{
renderAll(outputDir = NULL, ...)
}
\arguments{
\item{outputDir}{Output directory}

\item{...}{Passthrough \code{rmarkdown::render()} parameters}
}
\description{
Render all RMarkdown files in working directory
}
\author{
Michael Steinbaugh
}
\keyword{bcbio}
\keyword{report}
