% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderAll.R
\name{renderAll}
\alias{renderAll}
\title{Render all RMarkdown files in working directory}
\usage{
renderAll(outputDir = "render", ...)
}
\arguments{
\item{outputDir}{Output directory}

\item{...}{Passthrough \code{rmarkdown::render()} parameters}
}
\description{
Render all RMarkdown files in working directory
}
