% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderProject.R
\name{renderProject}
\alias{renderProject}
\title{Render all RMarkdown files in current project}
\usage{
renderProject(dir = "docs", ...)
}
\arguments{
\item{dir}{Output directory.}

\item{...}{Optional \code{rmarkdown} parameters}
}
\description{
Render all RMarkdown files in current project
}
