% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderProject.R
\name{renderProject}
\alias{renderProject}
\title{Render all RMarkdown files in working directory}
\usage{
renderProject(outputDir = "docs", recursive = FALSE)
}
\arguments{
\item{outputDir}{Output directory}

\item{recursive}{Find files recursively}
}
\description{
Render all RMarkdown files in working directory
}
\examples{
\dontrun{
renderProject()
}
}
\author{
Michael Steinbaugh
}
