% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderProject.R
\name{renderProject}
\alias{renderProject}
\title{Render all RMarkdown files in working directory}
\usage{
renderProject(outputDir = NULL, ...)
}
\arguments{
\item{outputDir}{Output directory}

\item{...}{Passthrough \code{rmarkdown::render()} parameters}
}
\description{
Render all RMarkdown files in working directory
}
\author{
Michael Steinbaugh
}
