% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTibble.R
\name{renderTibble}
\alias{renderTibble}
\title{Render a \code{tibble} when necessary}
\usage{
renderTibble(tibble, caption = NULL)
}
\arguments{
\item{tibble}{\code{tibble}}

\item{caption}{Caption, to be used during a \code{knitr} call}
}
\value{
\code{tibble} or \code{kable} depending on the calling environment
}
\description{
Render a \code{tibble} when necessary
}
\keyword{report}
