% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTibble.R
\name{renderTibble}
\alias{renderTibble}
\title{Render a tibble when necessary}
\usage{
renderTibble(tibble, caption = NULL)
}
\arguments{
\item{tibble}{Input tibble.}

\item{caption}{Caption, which will be used with `knitr` call.}
}
\value{
Tibble or kable depending on the call.
}
\description{
Render a tibble when necessary
}

