% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaseq_counts}
\alias{rnaseq_counts}
\alias{rnaseqCounts}
\title{Example RNA-Seq Counts}
\format{An object of class \code{matrix} with 100 rows and 4 columns.}
\usage{
rnaseq_counts
}
\description{
Generated with \code{DESeq2::makeExampleDESeqDataSet()}.
}
\examples{
summary(rnaseq_counts)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
