% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse}
\alias{rse}
\title{Example SummarizedExperiment}
\format{An object of class \code{RangedSummarizedExperiment} with 500 rows and 12 columns.}
\usage{
rse
}
\description{
Gene-level RNA-seq data set.
}
\details{
Generated with \code{DESeq2::makeExampleDESeqDataSet()}. Contains actual gene
identifiers and Ensembl annotations.
}
\examples{
data(rse)
rse
}
\keyword{datasets}
