% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse_dds}
\alias{rse_dds}
\title{\code{RangedSummarizedExperiment} Coerced from \code{DESeqDataSet}}
\format{An object of class \code{RangedSummarizedExperiment} with 1000 rows and 12 columns.}
\usage{
rse_dds
}
\description{
\code{RangedSummarizedExperiment} Coerced from \code{DESeqDataSet}
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
