% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sampleData-methods.R,
%   R/sampleNames-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleData}
\alias{sampleData,SummarizedExperiment-method}
\alias{sampleData<-,SummarizedExperiment,DataFrame-method}
\alias{sampleData<-,SummarizedExperiment,data.frame-method}
\alias{sampleNames,SummarizedExperiment-method}
\title{Sample Data}
\usage{
sampleData(object, ...)

sampleData(object, ...) <- value

\S4method{sampleData}{SummarizedExperiment}(object, clean = TRUE,
  interestingGroups, return = c("DataFrame", "data.frame", "kable"))

\S4method{sampleData}{SummarizedExperiment,DataFrame}(object) <- value

\S4method{sampleData}{SummarizedExperiment,data.frame}(object) <- value

\S4method{sampleNames}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}

\item{clean}{\code{logical}. Only return \code{factor} columns not defined in
\link{metadataBlacklist}.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\value{
Data describing the samples.
}
\description{
Return the sample metadata. Columns are always sanitized to factor.
}
\note{
This is a complement to the standard \code{\link[=colData]{colData()}} function, but improves
support for accessing sample metadata for datasets where multiple items in
the columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\examples{
# SummarizedExperiment ====
sampleData(rse_bcb, clean = TRUE) \%>\% glimpse()
sampleData(rse_bcb, clean = FALSE) \%>\% glimpse()

# Assignment support
x <- rse_dds
sampleData(x)[["test"]] <- seq_len(ncol(x))
# `test` column should be now defined
glimpse(sampleData(x))
}
\seealso{
\link{metadataBlacklist}.

Other Data Functions: \code{\link{counts}},
  \code{\link{gene2symbol}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
