% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sampleData-methods.R,
%   R/sampleNames-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleData,SummarizedExperiment-method}
\alias{sampleData<-,SummarizedExperiment,DataFrame-method}
\alias{sampleData,SingleCellExperiment-method}
\alias{sampleData<-,SingleCellExperiment,DataFrame-method}
\alias{sampleNames,SummarizedExperiment-method}
\title{Sample Data}
\usage{
sampleData(object, ...)

sampleData(object, ...) <- value

\S4method{sampleData}{SummarizedExperiment}(object)

\S4method{sampleData}{SummarizedExperiment,DataFrame}(object) <- value

\S4method{sampleData}{SingleCellExperiment}(object,
  blacklist = c("^ident$", "^origIdent$", "^res[.0-9]+$", "^sScore$",
  "^g2mScore$", "^phase$"))

\S4method{sampleData}{SingleCellExperiment,DataFrame}(object) <- value

\S4method{sampleNames}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}

\item{blacklist}{\code{character}. Column names that should never be treated as
sample-level metadata. Applies to objects where the columns don't map to
samples (e.g. \code{SingleCellExperiment}), and we need to collapse the
\code{colData()} dynamically.}
}
\value{
\code{DataFrame}.
}
\description{
Metadata that describes the samples.
}
\details{
This is a complement to the standard \code{\link[=colData]{colData()}} function, but improves
support for accessing sample metadata for datasets where multiple items in
the columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\examples{
data(rse_small, sce_small)

# SummarizedExperiment ====
x <- rse_small
sampleData(x)

# Assignment support
sampleData(x)[["batch"]] <- 1L
# `batch` column should be now defined.
sampleData(x)

# SingleCellExperiment ====
x <- sce_small
sampleData(x)

# Assignment support.
sampleData(x)[["batch"]] <- 1L
# `batch` column should be now defined.
sampleData(x)
}
\seealso{
Other Data Functions: \code{\link{aggregate}},
  \code{\link{combine}}, \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{minimalSampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
