% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sampleData-methods.R,
%   R/sampleNames-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleData,SummarizedExperiment-method}
\alias{sampleData<-,SummarizedExperiment,DataFrame-method}
\alias{sampleNames,SummarizedExperiment-method}
\title{Sample Data}
\usage{
sampleData(object, ...)

sampleData(object, ...) <- value

\S4method{sampleData}{SummarizedExperiment}(object, ...)

\S4method{sampleData}{SummarizedExperiment,DataFrame}(object) <- value

\S4method{sampleNames}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\code{DataFrame} containing metadata that describes the samples.
}
\description{
Return the sample metadata. Columns are always sanitized to factor.
}
\note{
This is a complement to the standard \code{\link[=colData]{colData()}} function, but improves
support for accessing sample metadata for datasets where multiple items in
the columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\examples{
# SummarizedExperiment ====
sampleData(rse_bcb) \%>\% glimpse()

# Assignment support
x <- rse_dds
sampleData(x)[["test"]] <- seq_len(ncol(x))
# `test` column should be now defined
glimpse(sampleData(x))
}
\seealso{
Other Data Functions: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}}, \code{\link{counts}},
  \code{\link{gene2symbol}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
