% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNames-methods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{sampleNames<-,SummarizedExperiment,character-method}
\title{Sample names}
\usage{
\S4method{sampleNames}{SummarizedExperiment,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Named \code{character}.
}
\description{
This function will return the human readable sample names if defined
in the \code{sampleName} column of \code{sampleData}. Otherwise it will return
the syntactically valid names defined as the rownames of \code{sampleData}.
}
\examples{
data(rse, package = "acidtest")
object <- rse
x <- sampleNames(object)
print(x)

## Assignment support.
value <- sampleNames(object)
value <- toupper(value)
print(value)
sampleNames(object) <- value
x <- sampleNames(object)
print(x)
}
