% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNames-methods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames<-,SummarizedExperiment,character-method}
\title{Sample Names}
\usage{
\S4method{sampleNames}{SummarizedExperiment,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Named \code{character} vector of the sample names.
}
\description{
This function will return the human readable sample names if defined
in the \code{sampleName} column of \code{\link[=sampleData]{sampleData()}}. Otherwise it will return
the syntactically valid names defined as the rownames of \code{\link[=sampleData]{sampleData()}}.
}
\examples{
## SummarizedExperiment ====
data(rse_small)
object <- rse_small
x <- sampleNames(object)
print(x)

## Assignment support
value <- sampleNames(object)
value <- toupper(value)
print(value)
sampleNames(object) <- value
x <- sampleNames(object)
print(x)
colData(object)$sampleName
}
\seealso{
Other Metadata Functions: \code{\link{detectHPC}},
  \code{\link{detectLanes}}, \code{\link{detectOrganism}},
  \code{\link{interestingGroups}},
  \code{\link{microplate}},
  \code{\link{minimalSampleData}}, \code{\link{organism}},
  \code{\link{sampleData}},
  \code{\link{uniteInterestingGroups}}

Other SummarizedExperiment Functions: \code{\link{counts}},
  \code{\link{interestingGroups}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{sampleData}},
  \code{\link{selectSamples}}, \code{\link{tpm}}
}
\concept{Metadata Functions}
\concept{SummarizedExperiment Functions}
