% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNames-methods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames<-,SummarizedExperiment,character-method}
\title{Sample Names}
\usage{
\S4method{sampleNames}{SummarizedExperiment,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Named \code{character} vector of the sample names.
}
\description{
This function will return the human readable sample names if defined
in the \code{sampleName} column of \code{\link[=sampleData]{sampleData()}}. Otherwise it will return
the syntactically valid names defined as the rownames of \code{\link[=sampleData]{sampleData()}}.
}
\examples{
# SummarizedExperiment ====
object <- rse_bcb
x <- sampleNames(object)
print(x)

# Assignment support
value <- sampleNames(object)
value <- gsub("_", " ", value)
print(value)
sampleNames(object) <- value
x <- sampleNames(object)
print(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}}, \code{\link{counts}},
  \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{sampleData}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
