% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNames-methods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames<-,SummarizedExperiment,character-method}
\title{Sample Names}
\usage{
\S4method{sampleNames}{SummarizedExperiment,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Named \code{character} vector of the sample names.
}
\description{
This function will return the human readable sample names if defined
in the \code{sampleName} column of \code{\link[=sampleData]{sampleData()}}. Otherwise it will return
the syntactically valid names defined as the rownames of \code{\link[=sampleData]{sampleData()}}.
}
\examples{
# SummarizedExperiment ====
object <- rse_small
x <- sampleNames(object)
print(x)

# Assignment support
value <- sampleNames(object)
value <- toupper(value)
print(value)
sampleNames(object) <- value
x <- sampleNames(object)
print(x)
colData(object)$sampleName
}
\seealso{
Other Data Functions: \code{\link{aggregate}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapCellsToSamples}}, \code{\link{mapGenes}},
  \code{\link{meltCounts}}, \code{\link{metrics}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}},
  \code{\link{topCellsPerSample}}, \code{\link{tpm}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
