% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAMLMetadata.R
\docType{methods}
\name{sampleYAMLMetadata}
\alias{sampleYAMLMetadata}
\alias{sampleYAMLMetadata,list-method}
\title{Sample Metadata from YAML}
\usage{
sampleYAMLMetadata(yaml, ...)

\S4method{sampleYAMLMetadata}{list}(yaml)
}
\arguments{
\item{yaml}{YAML \link{list}.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata from YAML
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetrics}}, \code{\link{sampleYAML}}
}
\concept{bcbio Utilities}
\keyword{internal}
