% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAMLMetrics.R
\docType{methods}
\name{sampleYAMLMetrics}
\alias{sampleYAMLMetrics}
\alias{sampleYAMLMetrics,list-method}
\title{Sample Metrics from YAML File}
\usage{
sampleYAMLMetrics(yaml, ...)

\S4method{sampleYAMLMetrics}{list}(yaml)
}
\arguments{
\item{yaml}{Project summary YAML list.}
}
\value{
\link{tibble}.
}
\description{
Sample Metrics from YAML File
}
\examples{
yaml <- readYAML("http://basejump.seq.cloud/project-summary.yaml")
sampleYAMLMetrics(yaml) \%>\% str()
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAML}}
}
\concept{bcbio Utilities}
\keyword{internal}
