% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAMLMetrics.R
\docType{methods}
\name{sampleYAMLMetrics}
\alias{sampleYAMLMetrics}
\alias{sampleYAMLMetrics}
\alias{sampleYAMLMetrics,list-method}
\title{Sample Metrics from YAML File}
\usage{
sampleYAMLMetrics(yaml, ...)

\S4method{sampleYAMLMetrics}{list}(yaml)
}
\arguments{
\item{yaml}{Project summary YAML list.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metrics from YAML File
}
\examples{
url <- file.path(
    "http://basejump.seq.cloud",
    "bcbio",
    "project-summary.yaml")
yaml <- readYAML(url)
sampleYAMLMetrics(yaml) \%>\% glimpse()
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAML}}
}
\keyword{internal}
