% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeAnnotable.R
\name{sanitizeAnnotable}
\alias{sanitizeAnnotable}
\title{Sanitize Ensembl Annotations}
\usage{
sanitizeAnnotable(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Data frame without nested list columns.
}
\description{
Drops any nested list columns (e.g. \code{entrez}) that prevents an annotable
data frame from being written to disk as a CSV.
}
\examples{
human <- annotable("Homo sapiens")
sanitizeAnnotable(human) \%>\% glimpse()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{ensembl}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{genes}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{transcripts}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
