% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeColData.R
\name{sanitizeColData}
\alias{sanitizeColData}
\title{Sanitize Column Data}
\usage{
sanitizeColData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DataFrame}.
}
\description{
Sanitize Column Data
}
\examples{
# data.frame ====
from <- data.frame(
    "genotype" = c("wt", "ko", "wt", "ko"),
    "batch" = c(1L, 1L, 2L, 2L),
    row.names = c("sample_1", "sample_2", "sample_3", "sample_4"),
    stringsAsFactors = FALSE
)
glimpse(from)
to <- sanitizeColData(from)
glimpse(to)

# DataFrame ====
from <- DataFrame(
    "genotype" = c("wt", "ko", "wt", "ko"),
    "batch" = c(1L, 1L, 2L, 2L),
    row.names = c("sample_1", "sample_2", "sample_3", "sample_4")
)
glimpse(from)
to <- sanitizeColData(from)
glimpse(to)
}
\seealso{
Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{fixNA}}, \code{\link{removeNA}},
  \code{\link{sanitizeRowData}}
}
