% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeColData.R
\name{sanitizeColData}
\alias{sanitizeColData}
\title{Sanitize Column Data Frame}
\usage{
sanitizeColData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{DataFrame}.
}
\description{
Sanitize Column Data Frame
}
\examples{
colData <- DataFrame(
    row.names = c("sample_1", "sample_2", "sample_3", "sample_4"),
    genotype = c("wt", "ko", "wt", "ko"),
    batch = c(1L, 1L, 2L, 2L)
)
colData
# Note that all columns will be sanitized to factor
sanitizeColData(colData)
}
\seealso{
Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{fixNA}}, \code{\link{removeNA}},
  \code{\link{sanitizeGeneAnnotations}}
}
