% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeNA-methods.R
\docType{methods}
\name{sanitizeNA}
\alias{sanitizeNA}
\alias{sanitizeNA,atomic-method}
\alias{sanitizeNA,character-method}
\alias{sanitizeNA,factor-method}
\alias{sanitizeNA,data.frame-method}
\alias{sanitizeNA,DataFrame-method}
\title{Sanitize NA values}
\usage{
\S4method{sanitizeNA}{atomic}(object)

\S4method{sanitizeNA}{character}(object)

\S4method{sanitizeNA}{factor}(object)

\S4method{sanitizeNA}{data.frame}(object)

\S4method{sanitizeNA}{DataFrame}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
Sanitized to contain proper \code{NA} values.
}
\description{
Standardize empty strings (\code{""}), character NAs (\code{"NA"}), and \code{NULL} values
inside a character vector to \code{NA_character_}. Other \code{atomic} data types are
returned unmodified.
}
\examples{
## character ====
from <- as.character(c(1L, "x", "", "NA", "NULL"))
print(from)
to <- sanitizeNA(from)
print(to)

## DataFrame ====
from <- S4Vectors::DataFrame(
    a = c("foo", ""),
    b = c(NA, "bar"),
    row.names = c("c", "d")
)
print(from)
to <- sanitizeNA(from)
print(to)
}
