% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeNames.R
\name{sanitizeNames}
\alias{sanitizeNames}
\title{Remove dashes in column and row names}
\usage{
sanitizeNames(df, ...)
}
\arguments{
\item{df}{\code{data.frame}}

\item{...}{Passthrough to \code{datatable} function}
}
\description{
This function removes dashes, which can cause unwanted wrapping in RMarkdown reports.
}
