% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizePercent-methods.R
\docType{methods}
\name{sanitizePercent}
\alias{sanitizePercent}
\alias{sanitizePercent,atomic-method}
\alias{sanitizePercent,character-method}
\title{Sanitize percentage}
\usage{
\S4method{sanitizePercent}{atomic}(object)

\S4method{sanitizePercent}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
All percentage strings containing \code{"\%"} (e.g. \code{"50\%"}) are converted to
numeric proportions (e.g. \code{0.5}).
}
\description{
Take a \code{character} vector of percentages (e.g. \code{"50\%"}) and convert it to a
\code{numeric} vector (e.g. \code{0.5}).
}
\section{Intended usage}{



This function is designed primarily to sanitize data imported from Microsoft
Excel.

}

\examples{
object <- c("100\%", "10.0\%", "1\%", "0.1\%", "0.01\%")
class(object)
print(object)
x <- sanitizePercent(object)
class(x)
print(x)
}
