% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeRowData.R
\name{sanitizeRowData}
\alias{sanitizeRowData}
\title{Sanitize Row Data}
\usage{
sanitizeRowData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{data.frame}, without any nested list columns.
}
\description{
Coerce Ensembl \code{rowData} returned from the \code{\link[=genes]{genes()}} or \code{\link[=transcripts]{transcripts()}}
functions to a \code{data.frame}, and drop any nested list columns (e.g.
\code{entrezID}). Nested columns will fail to write to disk as CSVs.
}
\details{
Supports \code{GRanges}, \code{data.frame}, and \code{DataFrame} objects.
}
\examples{
# Gene annotations
genes <- genes("Homo sapiens")
sanitizeRowData(genes) \%>\% glimpse()

# Transcript annotations
transcripts <- transcripts("Homo sapiens")
sanitizeRowData(transcripts) \%>\% glimpse()
}
\seealso{
Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{fixNA}}, \code{\link{removeNA}},
  \code{\link{sanitizeColData}}
}
