% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeRowData.R
\name{sanitizeRowData}
\alias{sanitizeRowData}
\title{Sanitize Row Data}
\usage{
sanitizeRowData(object, dropNested = TRUE)
}
\arguments{
\item{object}{Object.}

\item{dropNested}{Drop any nested list columns (e.g. \code{entrez}). Otherwise, a
user may run into issues attempting to write results to flat CSV files.}
}
\value{
\code{data.frame}, without nested list columns.
}
\description{
Coerce Ensembl \code{rowData} returned from the \code{\link[=genes]{genes()}} or \code{\link[=transcripts]{transcripts()}}
functions to a data frame, and drop any nested list columns (e.g. \code{entrez}),
if desired.
}
\details{
Supports \code{GRanges}, \code{data.frame}, and \code{DataFrame} objects.
}
\examples{
# Gene annotations
genes <- genes("Homo sapiens")
sanitizeRowData(genes) \%>\% glimpse()

# Transcript annotations
transcripts <- transcripts("Homo sapiens")
sanitizeRowData(transcripts) \%>\% glimpse()
}
\seealso{
Other Sanitization Functions: \code{\link{collapseToString}},
  \code{\link{fixNA}}, \code{\link{removeNA}},
  \code{\link{sanitizeColData}}
}
