% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeRowData.R
\name{sanitizeRowData}
\alias{sanitizeRowData}
\title{Sanitize Row Data}
\usage{
sanitizeRowData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{tbl_df}, containing only \code{atomic} columns.
}
\description{
Coerce gene annotations to a \code{tibble}, and keep only \code{atomic} columns.
Complex columns (e.g. Entrez ID \code{list}) will fail to write to disk as CSVs.
}
\details{
Supports \code{GRanges} and \code{DataFrame} class objects.
}
\examples{
from <- SummarizedExperiment::rowRanges(rse_small)
colnames(S4Vectors::mcols(from))
to <- sanitizeRowData(from)
vapply(to, is.atomic, logical(1L))
print(to)
}
\seealso{
Other Sanitization Functions: \code{\link{camel}},
  \code{\link{collapseToString}}, \code{\link{removeNA}},
  \code{\link{sanitizeNA}}, \code{\link{sanitizePercent}},
  \code{\link{sanitizeSampleData}}
}
\author{
Michael Steinbaugh
}
\concept{Sanitization Functions}
