% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeRowData.R
\name{sanitizeRowData}
\alias{sanitizeRowData}
\title{Sanitize Row Data}
\usage{
sanitizeRowData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DataFrame}. Contains only \code{character} and \code{factor} columns.
}
\description{
Coerce gene annotations to \code{DataFrame}, and keep only \code{atomic} columns.
Complex columns (e.g. Entrez ID \code{list}) will fail to write to disk as CSVs.
}
\note{
Supports \code{GRanges} and \code{DataFrame} class objects.
}
\examples{
data(rse_small)
from <- SummarizedExperiment::rowRanges(rse_small)
colnames(S4Vectors::mcols(from))
to <- sanitizeRowData(from)
vapply(to, is.atomic, logical(1L))
print(to)
}
\author{
Michael Steinbaugh
}
