% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSampleData.R
\name{sanitizeSampleData}
\alias{sanitizeSampleData}
\title{Sanitize sample data}
\usage{
sanitizeSampleData(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Sanitize sample data
}
\examples{
data(rse)
from <- sampleData(rse)
print(from)
to <- sanitizeSampleData(from)
all(vapply(to, is.factor, logical(1L)))
print(to)
}
