% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSampleData.R
\name{sanitizeSampleData}
\alias{sanitizeSampleData}
\title{Sanitize Sample Data}
\usage{
sanitizeSampleData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DataFrame}.
}
\description{
This function will standardize sample metadata.
}
\details{
The following conventions are enforced:
\itemize{
\item Column names will be converted to camel case, using \code{\link[=camel]{camel()}}.
\item Required columns: \code{sampleName}.
\item Blacklisted columns: \code{interestingGroups}, \code{sampleID}. These columns are
generated internally, and should not be user defined.
\item All columns will be coerced to factor.
\item Non-atomic columns will be dropped.
}

Currently supports \code{DataFrame} or \code{data.frame} input.
}
\examples{
data(rse_small)
from <- sampleData(rse_small)
print(from)
to <- sanitizeSampleData(from)
all(vapply(to, is.factor, logical(1L)))
print(to)
}
