% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSampleData.R
\name{sanitizeSampleData}
\alias{sanitizeSampleData}
\title{Sanitize Sample Data}
\usage{
sanitizeSampleData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DataFrame}.
}
\description{
This function will standardize sample metadata.
}
\details{
The following conventions are enforced:
\itemize{
\item Column names will be converted to camel case, using \code{\link[=camel]{camel()}}.
\item Data must contain a \code{sampleName} column.
\item All columns will be coerced to factor.
\item Non-atomic columns will be dropped.
}

Currently supports \code{DataFrame} or \code{data.frame} input.
}
\examples{
object <- sampleData(rse_bcb)
x <- sanitizeSampleData(object)
}
\seealso{
Other Sanitization Functions: \code{\link{camel}},
  \code{\link{collapseToString}}, \code{\link{fixNA}},
  \code{\link{removeNA}}, \code{\link{sanitizeRowData}}
}
\author{
Michael Steinbaugh
}
\concept{Sanitization Functions}
