% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSampleData.R
\name{sanitizeSampleData}
\alias{sanitizeSampleData}
\title{Sanitize sample data}
\usage{
sanitizeSampleData(object)
}
\arguments{
\item{object}{\code{DataFrame} (recommended) or \code{data.frame} (legacy).
Note that legacy \code{data.frame} support will be removed in a future update.}
}
\value{
\code{DataFrame}.
Sanitized data frame containing only non-blacklisted columns and all
\code{character} columns coerced to \code{factor} (i.e. \code{stringsAsFactors}).
}
\description{
Sanitize sample data
}
\note{
Updated 2021-01-14.
}
\section{Blacklist}{


Here's the current column blacklist:
\itemize{
\item interestingGroups.
\item sampleId.
}
}

\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
from <- sampleData(rse)
print(from)
to <- sanitizeSampleData(from)
all(vapply(to, is.factor, logical(1L)))
print(to)
}
