% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveData}
\alias{saveData}
\alias{saveDataRaw}
\alias{save_data}
\alias{save_data_raw}
\title{Save data objects.}
\usage{
saveData(..., dir = "data")

saveDataRaw(...)

save_data(..., dir = "data")

save_data_raw(...)
}
\arguments{
\item{...}{Objects.}

\item{dir}{Save directory.}
}
\description{
Quick save to \code{data} (default) or \code{data-raw} directory.
}
