% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveData}
\alias{saveData}
\alias{saveDataRaw}
\title{Save data objects}
\usage{
saveData(..., dir = "data")

saveDataRaw(...)
}
\arguments{
\item{...}{Objects.}

\item{dir}{Save directory.}
}
\description{
Quick save to \code{data} (default) or \code{data-raw} directory.
}
