% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveData}
\alias{saveData}
\alias{save_data}
\alias{saveDataRaw}
\title{Save Data Objects}
\usage{
saveData(..., dir = "data", compress = TRUE)

saveDataRaw(...)
}
\arguments{
\item{...}{Object names as symbols.}

\item{dir}{Output directory. Defaults to \strong{data}.}

\item{compress}{Compression method, supporting \code{xz} (\strong{preferred}), \code{bzip2},
or \code{gzip}. Compression can be disabled by setting as \code{FALSE}, although this
is not generally recommended.}
}
\description{
Quickly save objects in a project.
}
\details{
\itemize{
\item \code{\link[=saveData]{saveData()}}: Wrapper for \code{\link[base:save]{base::save()}} supporting quick saving of object
names passed as symbols. This function saves each object into a separate
\code{.rda} file rather than combining into a single file.
}

\itemize{
\item \code{\link[=saveDataRaw]{saveDataRaw()}}: Wrapper for \code{\link[=saveData]{saveData()}} that enables quick saving of
objects to the \code{data-raw} directory.
}
}
\note{
These functions will \emph{overwrite} existing saved data, following the
same conventions as \code{\link[base:save]{base::save()}}. Conversely, \code{\link[devtools:use_data]{devtools::use_data()}} does
not overwrite by default if that behavior is preferred.
}
\seealso{
\itemize{
\item \code{\link[base:save]{base::save()}}.
\item \code{\link[devtools:use_data]{devtools::use_data()}}.
\item \code{\link[devtools:use_data_raw]{devtools::use_data_raw()}}.
}
}
