% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveData}
\alias{saveData}
\alias{saveDataRaw}
\title{Save data objects}
\usage{
saveData(..., dir = "data", compress = "xz")

saveDataRaw(...)
}
\arguments{
\item{...}{Objects.}

\item{dir}{Save directory.}

\item{compress}{Compression method, supporting \code{xz} (preferred), \code{bzip2}, or
\code{gzip}. Compression can be disabled by setting as \code{FALSE}, although this is
not generally recommended.}
}
\description{
Quick save to \code{data} (default) or \code{data-raw} directory.
}
\seealso{
\code{\link[base:save]{base::save()}}.
}
