% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Save Data}
\usage{
saveData(..., dir = "data", compress = TRUE)
}
\arguments{
\item{...}{Object names as symbols.}

\item{dir}{Output directory. Defaults to \strong{data}.}

\item{compress}{Compression method, supporting \code{xz} (\strong{preferred}), \code{bzip2},
or \code{gzip}. Compression can be disabled by setting as \code{FALSE}, although this
is not generally recommended.}
}
\value{
No value.
}
\description{
Wrapper for \code{\link[base:save]{base::save()}} supporting quick saving of object names passed as
symbols. This function saves each object into a separate \code{.rda} file rather
than combining into a single file.
}
\note{
These function will \emph{overwrite} existing saved data, following the
same conventions as \code{\link[base:save]{base::save()}}. Conversely, \code{\link[devtools:use_data]{devtools::use_data()}} does
not overwrite by default if that behavior is preferred.
}
\seealso{
\itemize{
\item \code{\link[base:save]{base::save()}}
\item \code{\link[devtools:use_data]{devtools::use_data()}}.
}

Other Save Utilities: \code{\link{saveDataRaw}}
}
