% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\alias{saveDataRaw}
\title{Save data objects}
\usage{
saveData(..., dir = "data")

saveDataRaw(...)
}
\arguments{
\item{...}{Data objects}

\item{dir}{Output directory}
}
\description{
Quick save to \code{data} directory

Quick save to \code{data-raw} directory
}
