% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Save Data}
\usage{
saveData(..., dir = ".", ext = c("rda", "rds", "RData"), overwrite = TRUE,
  compress = TRUE)
}
\arguments{
\item{...}{Object names. Note that these arguments are interpreted as symbols
using non-standard evaluation for convenience during interactive use, and
\emph{should not be quoted}.}

\item{dir}{Output directory. Defaults to the current working directory.}

\item{ext}{R data ("\code{rda}", "\code{RData}") or R data serialized (RDS; "\code{rds}").
RDS is preferred when saving single objects per file, which is always the
convention of \code{\link[=saveData]{saveData()}}, regardless of the extension used.}

\item{overwrite}{Overwrite existing file.}

\item{compress}{logical or character string specifying whether saving
    to a named file is to use compression.  \code{TRUE} corresponds to
    \command{gzip} compression, and character strings \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} specify the type of
    compression.  Ignored when \code{file} is a connection and
    for workspace format version 1.}
}
\value{
Invisible named \code{character} vector of file paths.
}
\description{
Wrapper for \code{\link[base:save]{base::save()}} supporting quick saving of object names passed as
symbols. This function saves each object into a separate \code{.rda} file rather
than combining into a single file.
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation.

These function will \emph{overwrite} existing saved data, following the
same conventions as \code{\link[base:save]{base::save()}}. Conversely, \code{\link[devtools:use_data]{devtools::use_data()}} does
not overwrite by default if that behavior is preferred.
}
\examples{
saveData(rnaseq_counts, single_cell_counts, dir = "example")
list.files("example")

# Clean up
unlink("example", recursive = TRUE)
}
\seealso{
\itemize{
\item \code{\link[base:save]{base::save()}}.
\item \code{usethis::use_data()}.
}

Other Write Functions: \code{\link{assignAndSaveData}},
  \code{\link{initializeDirectory}},
  \code{\link{transmit}}, \code{\link{writeCounts}}
}
\author{
Michael Steinbaugh
}
