% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Save Data}
\usage{
saveData(..., dir = getOption("basejump.save.dir", "."),
  ext = getOption("basejump.save.ext", "rds"),
  overwrite = getOption("basejump.save.overwrite", TRUE),
  compress = getOption("basejump.save.compress", TRUE))
}
\arguments{
\item{...}{Object names. Note that these arguments are interpreted as symbols
using non-standard evaluation for convenience during interactive use, and
\emph{must not be quoted}.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}

\item{ext}{\code{string}. R data serialized (RDS; "\code{rds}") or R data ("\code{rda}",
"\code{RData}"). RDS is preferred when saving single objects per file, which is
always the convention of \code{\link[=saveData]{saveData()}}, regardless of the extension used.}

\item{overwrite}{\code{boolean}. Overwrite existing file.}

\item{compress}{logical or character string specifying whether saving
    to a named file is to use compression.  \code{TRUE} corresponds to
    \command{gzip} compression, and character strings \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} specify the type of
    compression.  Ignored when \code{file} is a connection and
    for workspace format version 1.}
}
\value{
Invisible named \code{character}. File paths.
}
\description{
Wrapper for \code{\link[=save]{save()}} supporting quick, interactive saving of object names
passed as symbols.
}
\details{
This function always saves each object into a separate file rather than
combining multiple objects into a single file.
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation. It will \strong{overwrite} existing files
on disk, following the same conventions as \code{\link[=save]{save()}}.
}
\examples{
data(rse_small, sce_small)
saveData(rse_small, sce_small, dir = "example")
list.files("example")

## Clean up.
unlink("example", recursive = TRUE)
}
\seealso{
\link{save}, \link{saveRDS}.

Other Import/Export Functions: \code{\link{assignAndSaveData}},
  \code{\link{export}}, \code{\link{import}},
  \code{\link{loadDataAsName}}, \code{\link{loadData}},
  \code{\link{loadRemoteData}}, \code{\link{transmit}},
  \code{\link{writeCounts}}
}
\concept{Import/Export Functions}
