% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Save data files}
\usage{
saveData(..., dir = "data")
}
\arguments{
\item{...}{Unquoted names of existing objects to save}

\item{dir}{Target directory}
}
\description{
This function makes it easy to save data in the correct format. Based off
\code{use_data()} from the \code{devtools} package. Note that this function
will overwrite existing data files.
}
\examples{
\dontrun{
x <- 1:10
y <- 1:100
saveData(x, y)
}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
