% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveDataRaw}
\alias{saveDataRaw}
\title{Save Raw Data}
\usage{
saveDataRaw(...)
}
\arguments{
\item{...}{Object names as symbols.}
}
\value{
No value.
}
\description{
Wrapper for \code{\link[=saveData]{saveData()}} that enables quick saving of objects to the
\code{data-raw/} directory.
}
\seealso{
\code{\link[devtools:use_data_raw]{devtools::use_data_raw()}}.

Other Save Utilities: \code{\link{saveData}}
}
