% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFunctionExamples.R
\name{saveFunctionExamples}
\alias{saveFunctionExamples}
\title{Save Functions Examples}
\usage{
saveFunctionExamples(fun, package, dir = ".")
}
\arguments{
\item{fun}{\code{character}. Function(s) from which to parse and save the working
examples.}

\item{package}{\code{string}. Package name.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}
}
\value{
\code{character}. Invisible file path(s).
}
\description{
Parse the documentation for a function and save the working examples to an
R script. Note that the \code{fun} argument is parameterized and can handle
multiple requests in a single call.
}
\examples{
saveFunctionExamples(
    fun = c("do.call", "droplevels"),
    package = "base",
    dir = "XXX"
)

# Cleanup
unlink("XXX", recursive = TRUE)
}
\seealso{
Other Developer Functions: \code{\link{calls}},
  \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
