% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRdExamples.R
\name{saveRdExamples}
\alias{saveRdExamples}
\title{Save R Documentation Examples}
\usage{
saveRdExamples(Rd = NULL, package, dir = ".")
}
\arguments{
\item{Rd}{\code{character} or \code{NULL}. R documentation name(s) from which to parse
and save the working examples. If \code{NULL}, all documentation files
containing examples will be saved.}

\item{package}{\code{string}. Package name.}

\item{dir}{\code{string}. Output directory. Defaults to the current working
directory.}
}
\value{
Invisible \code{character}. File path(s).
}
\description{
Parse the documentation for a function and save the working examples to an R
script. Note that the \code{f} argument is parameterized and can handle multiple
requests in a single call.
}
\examples{
saveRdExamples(
    Rd = c("do.call", "droplevels"),
    package = "base",
    dir = "example"
)

# Clean up
unlink("example", recursive = TRUE)
}
\seealso{
Other Documentation Functions: \code{\link{RdTags}},
  \code{\link{parseRd}}
}
\concept{Documentation Functions}
