% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{Example SingleCellExperiment}
\format{An object of class \code{SingleCellExperiment} with 500 rows and 100 columns.}
\usage{
sce
}
\description{
Gene-level single-cell RNA-seq data set.
}
\examples{
data(sce)
sce
}
\seealso{
\code{scater::sc_example_counts}.
}
\keyword{datasets}
