% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_small}
\alias{sce_small}
\title{Example Single-Cell RNA-Seq Data Set}
\format{An object of class \code{SingleCellExperiment} with 100 rows and 200 columns.}
\usage{
sce_small
}
\description{
Example Single-Cell RNA-Seq Data Set
}
\examples{
print(sce_small)
}
\seealso{
\code{scater::sc_example_counts}.

Other Minimal Example Data: \code{\link{rse_small}},
  \code{\link{tx_se_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
