% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSamples-methods.R
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,SummarizedExperiment-method}
\alias{selectSamples,SingleCellExperiment-method}
\title{Select samples}
\usage{
selectSamples(object, ...)

\S4method{selectSamples}{SummarizedExperiment}(object, ...)

\S4method{selectSamples}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Selection arguments:
\itemize{
\item Key value pairs that must map to the column names of \code{\link[acidgenerics:sampleData]{sampleData()}}.
\item \code{atomic} values are supported.
\item Avoid using \code{logical} or \code{numeric} indices
(e.g. \code{\link[base:which]{which()}} return), thereby making your code more
human readable, and easier to interpret.
}}
}
\value{
Modified object.
Subset (extraction) containing the desired samples.
}
\description{
Utility function that enables quick and easy sample selection.
}
\details{
Internally, pattern matching against sample and file names is applied using
logical grep matching.
}
\note{
Updated 2019-08-19.
}
\section{Intended usage}{



This function is intended to be used for sample extraction from Bioconductor
container classes where the column names don't correspond to samples (e.g.
\code{SingleCellExperiment}).

}

\section{Bracket-based subsetting}{



Bracket-based subsetting with \code{[} also works on \code{SingleCellExperiment}
objects but it's not intuitive. In this case, provide cellular barcode
identifiers for columns and gene identifiers for rows.

}

\examples{
data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)
rse <- RangedSummarizedExperiment
sce <- SingleCellExperiment

## SummarizedExperiment ====
object <- rse
sample <- sampleNames(object) \%>\% head(1L)
print(sample)
subset <- selectSamples(object, sampleName = sample)
print(subset)

## SingleCellExperiment ====
object <- sce
sample <- sampleNames(object) \%>\% head(1L)
print(sample)
subset <- selectSamples(object, sampleName = sample)
print(subset)
}
\seealso{
\itemize{
\item \code{sampleData()}.
\item \code{S4Vectors::split()}.
}
}
