% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/selectSamples-methods.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,SingleCellExperiment-method}
\title{Select Samples}
\usage{
selectSamples(object, ...)

\S4method{selectSamples}{SingleCellExperiment}(object, ...)
}
\value{
\code{SingleCellExperiment}.
}
\description{
Utility function that enables quick an easy sample selection in Bioconductor
contains that don't correspond to samples in the columns
(e.g. \code{SingleCellExperiment}).
}
\details{
Internally, pattern matching against sample and file names is
applied using logical grep matching.
}
\note{
Bracket based subsetting with \code{[} also works on \code{SingleCellExperiment}
objects but it's not intuitive. In this case, provide cellular barcode
identifiers for columns and gene identifiers for rows.
}
\examples{
object <- sce_small
sample <- sampleNames(object) \%>\% head(1L)
print(sample)
subset <- selectSamples(object, sampleName = sample)
print(subset)
}
\seealso{
\itemize{
\item \code{\link[=sampleData]{sampleData()}}.
\item \code{\link[S4Vectors:split]{S4Vectors::split()}}.
}

Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{tpm}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
