% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/selectSamples-methods.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,SummarizedExperiment-method}
\title{Select Samples}
\usage{
selectSamples(object, ...)

\S4method{selectSamples}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Selection arguments that map to the column names of
\code{\link[=sampleData]{sampleData()}}. \code{atomic} values are supported. Avoid using \code{logical} or
\code{numeric} indices (e.g. \code{\link[base:which]{base::which()}} calls) here, for improved code
readability.}
}
\value{
\code{SummarizedExperiment}.
}
\description{
Sample selection utility function designed for interactive use.
}
\examples{
# SummarizedExperiment ====
x <- selectSamples(rse_dds, condition = "A")
show(x)
colnames(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{counts}}, \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
