% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separator.R
\name{separator}
\alias{separator}
\title{Separator Bar}
\usage{
separator(sep = c("─", "=", "-", "+"),
  times = min(c(getOption("width", 72L), 72L)))
}
\arguments{
\item{sep}{\code{string} of length 1. Separator character.}

\item{times}{\code{scalar integer}. Number of times to repeat.}
}
\value{
\code{string}.
}
\description{
Maximum of 72 characters wide.
}
\note{
Bioconductor HTML vignettes don't render correctly when printing > 76
characters, even though the default width is set at 80.
}
\examples{
cat(separator())
}
\seealso{
Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}, \code{\link{standardizeCall}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
