% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{separatorBar}
\alias{separatorBar}
\title{Separator Bar}
\format{An object of class \code{character} of length 1.}
\usage{
separatorBar
}
\description{
Separator Bar
}
\examples{
cat(separatorBar)
}
\keyword{internal}
