% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setArgsToDoCall.R
\name{setArgsToDoCall}
\alias{setArgsToDoCall}
\title{Set Arguments to Do Call}
\usage{
setArgsToDoCall(args, removeArgs = NULL, call, fun)
}
\value{
\code{list}. Arguments to pass to \code{\link[=do.call]{do.call()}}.
}
\description{
Set Arguments to Do Call
}
\seealso{
Other Developer Functions: \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{matchS4Call}}, \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
