% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeNames.R, R/setNames.R
\name{sanitizeNames}
\alias{sanitizeNames}
\alias{setNamesCamel}
\alias{setNamesDot}
\alias{setNamesSnake}
\alias{setColnames}
\alias{setRownames}
\title{setNames utility functions}
\usage{
sanitizeNames(data)

setNamesCamel(data)

setNamesDot(data)

setNamesSnake(data)
}
\arguments{
\item{data}{Data type that supports name assignments}
}
\value{
Unmodified data with reformatted names
}
\description{
setNames utility functions
}
\examples{
sanitizeNames(head(iris))
setNamesCamel(head(iris))
setNamesDot(head(iris))
setNamesSnake(head(iris))
}
\author{
Michael Steinbaugh
}
\keyword{internal}
