% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNamesCamel}
\alias{setNamesCamel}
\alias{setNamesDot}
\alias{setNamesSnake}
\alias{sanitizeNames}
\alias{setColnames}
\alias{setRownames}
\title{setNames utility functions}
\usage{
setNamesCamel(data)

setNamesDot(data)

setNamesSnake(data)

sanitizeNames(data)
}
\arguments{
\item{data}{Data type that supports name assignments}
}
\value{
Unmodified data with reformatted names
}
\description{
setNames utility functions
}
\examples{
setNamesCamel(head(iris))
setNamesDot(head(iris))
setNamesSnake(head(iris))
sanitizeNames(head(iris))
}
\author{
Michael Steinbaugh
}
\keyword{internal}
