% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNamesCamel}
\alias{setNamesCamel}
\alias{set_names_camel}
\alias{setNamesDot}
\alias{set_names_dot}
\alias{setNamesSnake}
\alias{set_names_snake}
\alias{sanitizeNames}
\alias{sanitiseNames}
\alias{sanitize_names}
\alias{sanitise_names}
\alias{setColnames}
\alias{setRownames}
\title{Naming utility functions.}
\usage{
setNamesCamel(object)

set_names_camel(object)

setNamesDot(object)

set_names_dot(object)

setNamesSnake(object)

set_names_snake(object)

sanitizeNames(object)

sanitiseNames(object)

sanitize_names(object)

sanitise_names(object)
}
\arguments{
\item{object}{Object for which a \code{\link[=names]{names()}} attribute will be meaningful.}
}
\value{
Object with unmodified data and reformatted names.
}
\description{
The \code{\link[=setNames]{setNames()}} family of functions manipulate the \code{\link[=names]{names()}} of an object,
supporting either \code{camelCase} or \code{snake_case}. \code{\link[=sanitizeNames]{sanitizeNames()}} modifies
both \code{\link[=colnames]{colnames()}} and \code{\link[=rownames]{rownames()}} to \code{snake_case} format, and is generally
recommended for clean-up of RNA-seq count matrices containing gene names.
}
\examples{
setNamesCamel(head(starwars))
setNamesDot(head(starwars))
setNamesSnake(head(starwars))
sanitizeNames(head(starwars))
}
\seealso{
\itemize{
\item \code{\link[magrittr:set_names]{magrittr::set_names()}}.
\item \code{\link[stats:setNames]{stats::setNames()}}.
}
}
