% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNamesCamel}
\alias{setNamesCamel}
\alias{set_names_camel}
\alias{setNamesDot}
\alias{set_names_dot}
\alias{setNamesSnake}
\alias{set_names_snake}
\alias{sanitizeNames}
\alias{sanitiseNames}
\alias{sanitize_names}
\alias{sanitise_names}
\alias{setColnames}
\alias{setRownames}
\title{Naming utility functions.}
\usage{
setNamesCamel(object)

setNamesDot(object)

setNamesSnake(object)

sanitizeNames(object)
}
\arguments{
\item{object}{Object for which a \code{\link[=names]{names()}} attribute will be meaningful.}
}
\value{
Object with unmodified data and reformatted names.
}
\description{
Naming utility functions.
}
\examples{
setNamesCamel(head(starwars))
setNamesDot(head(starwars))
setNamesSnake(head(starwars))
sanitizeNames(head(starwars))
}
\seealso{
\code{\link[stats:setNames]{stats::setNames()}}
}
