% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R, R/setNamesCamel.R, R/setNamesDot.R,
%   R/setNamesSnake.R, R/setRownames.R
\name{setNames}
\alias{setNames}
\alias{setNamesCamel}
\alias{setNamesDot}
\alias{setNamesSnake}
\alias{setRownames}
\title{Set the names in an object}
\usage{
setNamesCamel(data)

setNamesDot(data)

setNamesSnake(data)

setRownames(data, column)
}
\arguments{
\item{data}{Data type that supports name assignments}

\item{column}{Column to use for \code{rownames}}
}
\value{
data Unmodified data with reformatted names
}
\description{
Set the names in an object
}
\examples{
setNamesCamel(head(iris))
setNamesDot(head(iris))
setNamesSnake(head(iris))
\dontrun{
setRownames(data, column = "ensembl_gene_id")
}
}
