% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNames}
\alias{setNames}
\title{setNames}
\arguments{
\item{object}{an object for which a names attribute will be meaningful}

\item{nm}{a character vector of names to assign to the object}
}
\description{
setNames
}

