% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamesCamel.R
\name{setNamesCamel}
\alias{setNamesCamel}
\title{Set names as camelCase}
\usage{
setNamesCamel(data)
}
\arguments{
\item{data}{Data frame, list, tibble, etc.}
}
\value{
data Same data but with reformatted camelCase names
}
\description{
Set names as camelCase
}
\author{
Michael Steinbaugh
}
