% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamesCamel.R
\name{setNamesCamel}
\alias{setNamesCamel}
\title{Set names of tibble as camelCase}
\usage{
setNamesCamel(tibble)
}
\arguments{
\item{tibble}{}
}
\value{
tibble
}
\description{
Set names of tibble as camelCase
}

