% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamesCamel.R
\name{setNamesCamel}
\alias{setNamesCamel}
\title{Set names as camelCase}
\usage{
setNamesCamel(data)
}
\arguments{
\item{data}{\code{data.frame}, \code{list}, or \code{tibble}}
}
\value{
data Same data but with reformatted camelCase names
}
\description{
Set names as camelCase
}
\keyword{general}

