% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamesSnake.R
\name{setNamesSnake}
\alias{setNamesSnake}
\title{Set names as snake_case}
\usage{
setNamesSnake(data)
}
\arguments{
\item{data}{Data type that supports name assignments}
}
\value{
data Unmodified data with reformatted snake_case names
}
\description{
Set names as snake_case
}
\author{
Michael Steinbaugh
}
