% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showHeader.R
\name{showHeader}
\alias{showHeader}
\title{Header for object show method}
\usage{
showHeader(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output, via \code{\link[base:cat]{cat()}}.
}
\description{
Header for object show method
}
\examples{
data(rse, package = "acidtest")
showHeader(rse)
}
\seealso{
\code{\link[methods:show]{methods::show()}}.
}
