% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSlotInfo.R
\name{showSlotInfo}
\alias{showSlotInfo}
\title{Show Slot Information}
\usage{
showSlotInfo(list)
}
\arguments{
\item{list}{\code{list}. Named list containing slot metadata.}
}
\description{
Show Slot Information
}
