% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake.R
\name{snake}
\alias{snake}
\title{snake_case}
\usage{
snake(string)
}
\arguments{
\item{string}{String}
}
\value{
snake_case formatted string
}
\description{
snake_case
}
\examples{
snake("RNAi clone")
}
\author{
Michael Steinbaugh
}
