% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake.R
\name{snake}
\alias{snake}
\title{snake_case}
\usage{
snake(string)
}
\arguments{
\item{string}{\code{string}.}
}
\value{
\code{string} with snake_case formatting.
}
\description{
snake_case
}
\examples{
snake("RNAi clone")
}

