% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-snake.R
\docType{methods}
\name{snake}
\alias{snake}
\alias{snake}
\alias{snake,ANY-method}
\alias{snake,character-method}
\alias{snake,data.frame-method}
\alias{snake,DataFrame-method}
\alias{snake,factor-method}
\alias{snake,GRanges-method}
\alias{snake,list-method}
\alias{snake,List-method}
\alias{snake,matrix-method}
\alias{snake,SimpleList-method}
\alias{snake,tbl_df-method}
\title{Snake Case}
\usage{
snake(object, ...)

\S4method{snake}{ANY}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{character}(object)

\S4method{snake}{data.frame}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{DataFrame}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{factor}(object)

\S4method{snake}{GRanges}(object)

\S4method{snake}{list}(object)

\S4method{snake}{List}(object)

\S4method{snake}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{SimpleList}(object)

\S4method{snake}{tbl_df}(object)
}
\arguments{
\item{object}{Character vector or an object for which \code{\link[=names]{names()}} assignment
will be meaningful.}

\item{...}{Additional arguments.}

\item{rownames}{Apply sanitization on row names. This is not recommended
by default, since rownames commonly contain gene identifiers that should
not be modified.}

\item{colnames}{Apply sanitization on column names. This is generally
recommended by default.}
}
\value{
Object with syntatically valid names. For objects supporting
\code{\link[base:names]{base::names()}}, the underlying data returns unchanged.
}
\description{
Snake Case
}
\details{
For unnamed character vectors, this function will sanitize the
underlying values. Otherwise, it will set \code{\link[=names]{names()}} and/or \code{\link[=rownames]{rownames()}} on
objects supporting name assignments. They return the object without
modification of the underlying data.
}
\examples{
load(system.file("extdata/makeNames.rda", package = "basejump"))

# Character vector
character <- makeNames$character
print(character)
snake(character)

# Named character vector
namedCharacter <- makeNames$namedCharacter
print(namedCharacter)
snake(namedCharacter)

# Factor
factor <- makeNames$factor
print(factor)
snake(factor)

# data.frame
dataFrame <- makeNames$dataFrame
print(dataFrame)
snake(dataFrame, rownames = TRUE)
snake(dataFrame, rownames = FALSE)

# Named list
list <- makeNames$list
print(list)
snake(list)
}
\seealso{
Other Sanitization Functions: \code{\link{camel}},
  \code{\link{dotted}}, \code{\link{fixNA}},
  \code{\link{removeNA}}, \code{\link{sanitizeAnnotable}},
  \code{\link{sanitizeColData}}
}
