% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-snake.R
\docType{methods}
\name{snake}
\alias{snake}
\alias{snake}
\alias{snake,ANY-method}
\alias{snake,character-method}
\alias{snake,data.frame-method}
\alias{snake,DataFrame-method}
\alias{snake,factor-method}
\alias{snake,GRanges-method}
\alias{snake,list-method}
\alias{snake,List-method}
\alias{snake,matrix-method}
\alias{snake,SimpleList-method}
\alias{snake,tbl_df-method}
\title{Snake Case}
\usage{
snake(object, ...)

\S4method{snake}{ANY}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{character}(object)

\S4method{snake}{data.frame}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{DataFrame}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{factor}(object)

\S4method{snake}{GRanges}(object)

\S4method{snake}{list}(object)

\S4method{snake}{List}(object)

\S4method{snake}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{SimpleList}(object)

\S4method{snake}{tbl_df}(object)
}
\arguments{
\item{object}{Character vector or an object for which \code{\link[=names]{names()}} assignment
will be meaningful.}

\item{...}{Additional arguments.}

\item{rownames}{Apply sanitization on row names. This is not recommended
by default, since rownames commonly contain gene identifiers that should
not be modified.}

\item{colnames}{Apply sanitization on column names. This is generally
recommended by default.}
}
\value{
Object with syntatically valid names. For objects supporting
\code{\link[base:names]{base::names()}}, the underlying data returns unchanged.
}
\description{
Snake Case
}
\examples{
load(system.file("extdata/mn.rda", package = "basejump"))

# character ====
character <- mn$character
print(character)
snake(character)

namedCharacter <- mn$namedCharacter
print(namedCharacter)
snake(namedCharacter)

# factor ====
factor <- mn$factor
print(factor)
snake(factor)

# data.frame ====
dataFrame <- mn$dataFrame
print(dataFrame)
snake(dataFrame, rownames = TRUE)
snake(dataFrame, rownames = FALSE)

# list ====
list <- mn$list
print(list)
snake(list)
}
\seealso{
Other Name Functions: \code{\link{camel}},
  \code{\link{dotted}}, \code{\link{makeNames}}
}
\author{
Michael Steinbaugh
}
