% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortUnique.R
\name{sortUnique}
\alias{sortUnique}
\title{Quickly Perform Sort Unique on a Vector}
\usage{
sortUnique(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Unique vector.
}
\description{
The function also strips \code{NA} values. This is useful for gene list server
queries, for example.
}
\examples{
sortUnique(c(NA, NA, "milk", "eggs", "eggs"))
}
\seealso{
Other Vector Functions: \code{\link{collapseToString}},
  \code{\link{grepString}}, \code{\link{toStringUnique}}
}
