% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sortUnique.R
\docType{methods}
\name{sortUnique}
\alias{sortUnique}
\alias{sortUnique}
\alias{sortUnique,factor-method}
\alias{sortUnique,vector-method}
\title{Quickly Perform Sort Unique on a Vector}
\usage{
sortUnique(object, ...)

\S4method{sortUnique}{factor}(object)

\S4method{sortUnique}{vector}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Unique vector.
}
\description{
The function also strips \code{NA} values. This is useful for gene list server
queries, for example.
}
\examples{
sortUnique(c(NA, NA, "milk", "eggs", "eggs"))
}
\seealso{
Other Cleanup Utilities: \code{\link{fixNA}},
  \code{\link{removeNA}}
}
