% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortUnique.R
\name{sortUnique}
\alias{sortUnique}
\title{Sort and Make Unique}
\usage{
sortUnique(atomic)
}
\arguments{
\item{atomic}{\code{atomic} vector.}
}
\value{
\code{atomic} vector.
}
\description{
This is a convenience function to quickly sort and atomic vector and make the
values unique. The function also strips \code{NA} values. This is useful for
repetitive gene vector operations, for example.
}
\examples{
sortUnique(c(NA, NA, "milk", "eggs", "eggs"))
}
\seealso{
Other Atomic Vector Functions: \code{\link{grepString}},
  \code{\link{toStringUnique}}
}
\author{
Michael Steinbaugh
}
