% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortUnique}
\alias{sortUnique}
\title{Quickly perform sort unique on a vector}
\usage{
sortUnique(vector)
}
\arguments{
\item{vector}{Vector with duplicates, \code{NA} values.}
}
\value{
Unique vector.
}
\description{
The function also strips \code{NA} values. This is useful for gene list server
queries, for example.
}
\examples{
sortUnique(c("milk", "eggs", "eggs", NA))
}
