% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-sortUnique.R
\docType{methods}
\name{sortUnique}
\alias{sortUnique}
\alias{sortUnique,character-method}
\title{Quickly Perform Sort Unique on a Vector}
\usage{
sortUnique(object)

\S4method{sortUnique}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Unique vector.
}
\description{
The function also strips \code{NA} values. This is useful for gene list server
queries, for example.
}
\examples{
sortUnique(c("milk", "eggs", "eggs", NA))
}
