% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-internal.R
\name{standardizeCall}
\alias{standardizeCall}
\title{Standardize Call}
\usage{
standardizeCall(definition = sys.function(sys.parent()),
  call = sys.call(sys.parent()), expand.dots = TRUE,
  envir = parent.frame(n = 2L), verbose = FALSE)
}
\description{
Standardize Call
}
\examples{
a <- "AAA"
b <- "BBB"
c <- "CCC"

# Standard call
testing <- function(x, y) {
    print(match.call())
    print(standardizeCall())
}
testing(a, b)

# S4 mode
setGeneric(
    name = "testing",
    def = function(x, y, ...) {
        standardGeneric("testing")
    }
)

setMethod(
    f = "testing",
    signature = signature("character"),
    definition = function(x, y, ...) {
        print(match.call())
        print(standardizeCall())
    }
)
testing(a, b)

setMethod(
    f = "testing",
    signature = signature("character"),
    definition = function(x, y, z) {
        print(match.call())
        print(match.call(call = sys.call(sys.parent())))
        print(standardizeCall())
    }
)
testing(a, b, c)
}
\seealso{
Other Developer Functions: \code{\link{MethodDefinition}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchArgsToDoCall}},
  \code{\link{matchInterestingGroups}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
