% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{starwars}
\alias{starwars}
\title{Starwars Characters}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 87 rows and 13 columns.}
\usage{
starwars
}
\value{
\code{tibble}.
}
\description{
Starwars Characters
}
\examples{
glimpse(starwars)
}
\keyword{internal}
