% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{starwars}
\alias{starwars}
\title{Starwars characters}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 87 rows and 13 columns.}
\usage{
starwars
}
\description{
This data comes from SWAPI, the Star Wars API, \url{http://swapi.co/}
}
\keyword{internal}
