% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-stripTranscriptVersions.R
\docType{methods}
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions,character-method}
\alias{stripTranscriptVersions,dgCMatrix-method}
\alias{stripTranscriptVersions,matrix-method}
\title{Strip Transcript Versions}
\usage{
stripTranscriptVersions(object, ...)

\S4method{stripTranscriptVersions}{character}(object)

\S4method{stripTranscriptVersions}{dgCMatrix}(object)

\S4method{stripTranscriptVersions}{matrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Transcript identifiers without version numbers.
}
\description{
Strip Transcript Versions
}
\note{
This function is strict, and will only strip Ensembl transcript IDs
beginning with "ENS".
}
\examples{
# Ensembl
stripTranscriptVersions("ENSMUST00000000001.1")

# WormBase (unmodified)
stripTranscriptVersions("cTel79B.1")
}
\seealso{
Other Gene Functions: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}
}
\author{
Michael Steinbaugh
}
