% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/stripTranscriptVersions-methods.R
\docType{methods}
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions,character-method}
\alias{stripTranscriptVersions,matrix-method}
\alias{stripTranscriptVersions,sparseMatrix-method}
\title{Strip Transcript Versions}
\usage{
stripTranscriptVersions(object, ...)

\S4method{stripTranscriptVersions}{character}(object)

\S4method{stripTranscriptVersions}{matrix}(object)

\S4method{stripTranscriptVersions}{sparseMatrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, containing transcript identifiers without version
numbers.
}
\description{
Strip Transcript Versions
}
\note{
This function is strict, and will only strip Ensembl transcript IDs
beginning with "ENS".
}
\examples{
# Ensembl (modify; contains versions)
stripTranscriptVersions(c(
    "ENSMUST00000000001.1",
    "ENSMUST00000000001-1",
    "ENSMUST00000000001_1"
))

# WormBase (keep; doesn't contain versions)
stripTranscriptVersions("cTel79B.1")
}
\seealso{
Other Annotation Functions: \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2ensembl}}, \code{\link{makeGRanges}},
  \code{\link{makeGene2symbol}}, \code{\link{makeTx2gene}},
  \code{\link{mgi2ensembl}}, \code{\link{panther}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
