% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/stripTranscriptVersions-methods.R
\docType{methods}
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions,character-method}
\alias{stripTranscriptVersions,matrix-method}
\alias{stripTranscriptVersions,dgCMatrix-method}
\title{Strip Transcript Versions}
\usage{
stripTranscriptVersions(object, ...)

\S4method{stripTranscriptVersions}{character}(object)

\S4method{stripTranscriptVersions}{matrix}(object)

\S4method{stripTranscriptVersions}{dgCMatrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Object containing transcript identifiers without version numbers.
}
\description{
Strip Transcript Versions
}
\note{
This function is strict, and will only strip Ensembl transcript IDs
beginning with "ENS".
}
\examples{
# Ensembl
stripTranscriptVersions("ENSMUST00000000001.1")
stripTranscriptVersions("ENSMUST00000000001-1")
stripTranscriptVersions("ENSMUST00000000001_1")

# WormBase (unmodified)
stripTranscriptVersions("cTel79B.1")
}
\seealso{
Other Annotation Functions: \code{\link{annotable}},
  \code{\link{broadClass}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{detectOrganism}}, \code{\link{eggnog}},
  \code{\link{emptyRanges}}, \code{\link{geneSynonyms}},
  \code{\link{hgnc2gene}},
  \code{\link{makeGRangesFromEnsembl}},
  \code{\link{makeGRangesFromGFF}},
  \code{\link{makeGene2symbolFromEnsembl}},
  \code{\link{makeGene2symbolFromGFF}},
  \code{\link{makeTx2geneFromEnsembl}},
  \code{\link{makeTx2geneFromGFF}}, \code{\link{mgi2gene}},
  \code{\link{panther}}
}
\author{
Michael Steinbaugh
}
\concept{Annotation Functions}
