% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-stripTranscriptVersions.R
\docType{methods}
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions,character-method}
\title{Strip Transcript Versions}
\usage{
stripTranscriptVersions(object, ...)

\S4method{stripTranscriptVersions}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Transcript identifiers without version numbers.
}
\description{
Strip Transcript Versions
}
\examples{
stripTranscriptVersions("ENSMUST00000000001.1")
}
\seealso{
Other Gene Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{ensembl}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{genes}},
  \code{\link{transcripts}}, \code{\link{tx2geneFromGFF}},
  \code{\link{tx2gene}}
}
