% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-stripTranscriptVersions.R
\docType{methods}
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions,character-method}
\title{Strip Transcript Versions}
\usage{
stripTranscriptVersions(object, ...)

\S4method{stripTranscriptVersions}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Transcript identifiers without version numbers.
}
\description{
Strip Transcript Versions
}
\author{
Michael Steinbaugh
}
