% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripTranscriptVersions-methods.R
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions,character-method}
\alias{stripTranscriptVersions,matrix-method}
\alias{stripTranscriptVersions,Matrix-method}
\alias{stripTranscriptVersions,SummarizedExperiment-method}
\alias{stripGeneVersions}
\title{Strip transcript versions}
\usage{
stripTranscriptVersions(object, ...)

\S4method{stripTranscriptVersions}{character}(object)

\S4method{stripTranscriptVersions}{matrix}(object)

\S4method{stripTranscriptVersions}{Matrix}(object)

\S4method{stripTranscriptVersions}{SummarizedExperiment}(object)

stripGeneVersions(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains transcript identifiers without version numbers.
}
\description{
Strip transcript versions
}
\note{
Updated 2019-10-09.

This method is strict, and will only strip Ensembl transcript IDs
beginning with "ENS".
}
\examples{
## Ensembl (modify; contains versions)
stripTranscriptVersions(c(
    "ENSMUST00000000001.1",
    "ENSMUST00000000001-1",
    "ENSMUST00000000001_1"
))

## WormBase (keep; doesn't contain versions)
stripTranscriptVersions("cTel79B.1")
}
