% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subsetPerSample-methods.R
\name{subsetPerSample}
\alias{subsetPerSample}
\alias{subsetPerSample,SingleCellExperiment-method}
\title{Subset per sample}
\usage{
subsetPerSample(object, ...)

\S4method{subsetPerSample}{SingleCellExperiment}(
  object,
  minCells = 1L,
  assignAndSave = FALSE,
  envir = parent.frame(),
  dir = "."
)
}
\arguments{
\item{object}{Object.}

\item{minCells}{\code{integer(1)}.
Minimum number of cells required per sample.}

\item{assignAndSave}{\code{logical(1)}.
Assign and save the individual datasets.}

\item{envir}{\code{environment}.
Where to assign the subsets.
Only applicable when \code{assignAndSave = TRUE}.}

\item{dir}{\code{character(1)}.
Output directory.
Only applicable when \code{assignAndSave = TRUE}.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{assignAndSave = FALSE}: Per sample objects in a \code{list}.
\item \code{assignAndSave = TRUE}: Subset file paths.
}
}
\description{
Subset per sample
}
\note{
Updated 2019-08-11.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")
sce <- SingleCellExperiment

## SingleCellExperiment ====
object <- sce

## List mode (default).
list <- subsetPerSample(object, assignAndSave = FALSE)
names(list)

## Assign and save mode (useful for large datasets).
subsetPerSample(
    object = object,
    assignAndSave = TRUE,
    envir = parent.frame(),
    dir = "subsetPerSample"
)
sort(list.files("subsetPerSample"))

## Clean up.
unlink("subsetPerSample", recursive = TRUE)
}
