% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary}
\alias{summary}
\alias{summary,Gene2Symbol-method}
\alias{summary,Tx2Gene-method}
\title{Object summaries}
\usage{
summary(object, ...)

\S4method{summary}{Gene2Symbol}(object)

\S4method{summary}{Tx2Gene}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Object summaries
}
\note{
Updated 2019-11-19.
}
\examples{
data(
    RangedSummarizedExperiment,
    SummarizedExperiment_transcripts,
    package = "acidtest"
)

## Gene2Symbol ====
object <- Gene2Symbol(RangedSummarizedExperiment)
summary(object)

## Tx2Gene ====
object <- Tx2Gene(SummarizedExperiment_transcripts)
summary(object)
}
\keyword{internal}
