% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary}
\alias{summary}
\alias{summary,Gene2Symbol-method}
\alias{summary,Tx2Gene-method}
\title{Object summaries}
\usage{
summary(object, ...)

\S4method{summary}{Gene2Symbol}(object)

\S4method{summary}{Tx2Gene}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
\code{summary} is a generic function used to produce result summaries
  of the results of various model fitting functions.  The function
  invokes particular \code{\link[base]{methods}} which depend on the
  \code{\link[base]{class}} of the first argument.
}
\note{
Updated 2019-07-28.
}
\examples{
data(
    RangedSummarizedExperiment,
    SummarizedExperiment_transcripts,
    package = "acidtest"
)

## Gene2Symbol ====
object <- Gene2Symbol(RangedSummarizedExperiment)
summary(object)

## Tx2Gene ====
object <- Tx2Gene(SummarizedExperiment_transcripts)
summary(object)
}
\references{
Chambers, J. M. and Hastie, T. J. (1992)
  \emph{Statistical Models in S}.
  Wadsworth & Brooks/Cole.
}
\keyword{internal}
