% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,Gene2Symbol-method}
\alias{summary,Tx2Gene-method}
\title{Object summaries}
\usage{
\S4method{summary}{Gene2Symbol}(object)

\S4method{summary}{Tx2Gene}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Invisible \code{NULL}.
}
\description{
\code{summary} is a generic function used to produce result summaries
  of the results of various model fitting functions.  The function
  invokes particular \code{\link{methods}} which depend on the
  \code{\link{class}} of the first argument.
}
\examples{
data(rse, tx_se, package = "acidData")

## Gene2Symbol ====
x <- Gene2Symbol(rse)
summary(x)

## Tx2Gene ====
x <- Tx2Gene(tx_se)
summary(x)
}
\references{
Chambers, J. M. and Hastie, T. J. (1992)
  \emph{Statistical Models in S}.
  Wadsworth & Brooks/Cole.
}
