% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-symbol2gene.R
\docType{methods}
\name{symbol2gene}
\alias{symbol2gene}
\alias{symbol2gene}
\alias{symbol2gene,character-method}
\alias{symbol2gene,data.frame-method}
\alias{symbol2gene,DataFrame-method}
\alias{symbol2gene,dgCMatrix-method}
\alias{symbol2gene,dgTMatrix-method}
\alias{symbol2gene,matrix-method}
\title{Convert Gene Symbol to Ensembl Identifier}
\usage{
symbol2gene(object, ...)

\S4method{symbol2gene}{character}(object, organism)

\S4method{symbol2gene}{data.frame}(object, organism)

\S4method{symbol2gene}{DataFrame}(object, organism)

\S4method{symbol2gene}{dgCMatrix}(object, organism)

\S4method{symbol2gene}{dgTMatrix}(object, organism)

\S4method{symbol2gene}{matrix}(object, organism)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{organism}{Organism name.}
}
\value{
Same class as original object.
}
\description{
Convert Gene Symbol to Ensembl Identifier
}
\examples{
# character
symbol2gene(c("Gnai3", "Pbsn"), organism = "mouse")

# matrix
matrix(
    data = seq(1L:4L),
    byrow = TRUE,
    nrow = 2L,
    ncol = 2L,
    dimnames = list(c("Gnai3", "Pbsn"),
                    c("sample1", "sample2"))) \%>\%
    symbol2gene(organism = "mouse")
}
\seealso{
\code{\link[=detectOrganism]{detectOrganism()}}.

Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGTF}},
  \code{\link{gene2symbol}}, \code{\link{tx2geneFromGTF}},
  \code{\link{tx2gene}}
}
