% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{tabular}
\alias{tabular}
\title{R documentation table}
\usage{
tabular(df)
}
\arguments{
\item{df}{\code{data.frame}.}
}
\value{
Console output.
}
\description{
R documentation table
}
\seealso{
http://r-pkgs.had.co.nz/man.html
}
