% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl}
\alias{tbl}
\title{Example tibble}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 5 columns.}
\usage{
tbl
}
\description{
Example tibble
}
\examples{
data(tbl)
tbl
}
\keyword{datasets}
