% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\alias{theme_midnight}
\alias{theme_paperwhite}
\alias{midnightTheme}
\alias{paperwhiteTheme}
\title{Complete Themes}
\usage{
theme_midnight(base_size = 14L, base_family = "", face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("right", "bottom",
  "none"), grid = FALSE)

theme_paperwhite(base_size = 14L, base_family = "", face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("right", "bottom",
  "none"), grid = FALSE)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{face}{Font face ("\code{bold}", "\code{plain}").}

\item{aspect_ratio}{Numeric scalar specifying the plot proportions. Use \code{1}
for a perfectly square plot (including the axis labels).}

\item{legend_position}{Legend key position. We're being a little more
restrictive here, only allowing "\code{bottom}", "\code{right}", or "\code{none}".
Including the legend at the top or the left side of the plot rarely makes
sense and is discouraged.}

\item{grid}{Label the major panel grids. If \code{TRUE}, the axis lines will also
be removed.}
}
\value{
\code{theme}.
}
\description{
\href{http://ggplot2.tidyverse.org}{ggplot2} themes.
}
\section{Midnight Theme}{

Blackout theme that sets the plot background as black, with white text.
Inspired by \code{Seurat::DarkTheme()}, with some color modifications. Useful
for visualizing many points with a high dynamic color range, such as t-SNE
expression plots.
}

\section{Paperwhite Theme}{

High contrast black and white theme optimized for print. Recommended for
scientific manuscripts and website tutorials.
}

\examples{
library(ggplot2)

# Example ggplot object
p <- ggplot(
    data = mpg,
    mapping = aes(
        x = manufacturer,
        y = displ,
        colour = manufacturer
    )
) +
    geom_point()
# Midnight Theme
p + theme_midnight(aspect_ratio = 1L, legend_position = "none")
p + theme_paperwhite(aspect_ratio = 1L, legend_position = "none")
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\author{
Michael Steinbaugh
}
