% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\alias{midnightTheme}
\alias{theme_paperwhite}
\alias{theme_midnight}
\alias{paperwhiteTheme}
\title{Complete Themes}
\usage{
midnightTheme(baseSize = 14L, baseFamily = "", baseFace = c("bold",
  "plain"), aspectRatio = NULL, legendPosition = c("bottom", "right",
  "none"), grid = TRUE)

theme_paperwhite(base_size = 14L, base_family = "", base_face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("bottom", "right",
  "none"), grid = FALSE)

theme_midnight(base_size = 14L, base_family = "", base_face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("bottom", "right",
  "none"), grid = TRUE)

paperwhiteTheme(base_size = 14L, base_family = "", base_face = c("bold",
  "plain"), aspect_ratio = NULL, legend_position = c("bottom", "right",
  "none"), grid = FALSE)
}
\value{
\code{theme}.
}
\description{
\href{http://ggplot2.tidyverse.org}{ggplot2} themes.
}
\section{Midnight}{

Blackout theme that sets the plot background as black, with white text.
Inspired by \code{Seurat::DarkTheme()}, with some color modifications. Useful
for visualizing many points with a high dynamic color range, such as t-SNE
expression plots.
}

\section{Paperwhite}{

High contrast black and white theme optimized for print. Recommended for
scientific manuscripts and website tutorials.
}

\examples{
library(ggplot2)

# Example ggplot object
p <- ggplot(
    data = mpg,
    mapping = aes(
        x = manufacturer,
        y = displ,
        color = manufacturer
    )
) +
    geom_point()
# Midnight Theme
p + theme_midnight(aspect_ratio = 1L, legend_position = "none")
p + theme_paperwhite(aspect_ratio = 1L, legend_position = "none")
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.

Other Plot Functions: \code{\link{dynamicPlotlist}}
}
\author{
Michael Steinbaugh
}
