% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-tidy.R,
%   R/reexports.R, R/zzz.R
\docType{methods}
\name{arrange}
\alias{arrange}
\alias{asTibble}
\alias{filter}
\alias{mutate}
\alias{arrange,data.frame-method}
\alias{.asTibble}
\alias{asTibble,data.frame-method}
\alias{asTibble,DataFrame-method}
\alias{asTibble,matrix-method}
\alias{content,response-method}
\alias{expand,data.frame-method}
\alias{filter,data.frame-method}
\alias{first,data.frame-method}
\alias{intersect,data.frame,data.frame-method}
\alias{mutate,data.frame-method}
\alias{rename,data.frame-method}
\alias{select,data.frame-method}
\alias{setdiff,data.frame,data.frame-method}
\alias{setequal,data.frame,data.frame-method}
\alias{union,data.frame,data.frame-method}
\alias{select}
\alias{content}
\alias{intersect}
\alias{setdiff}
\alias{union}
\alias{tidy_select}
\alias{expand}
\alias{first}
\alias{rename}
\alias{as_tibble}
\title{\href{http://tidyverse.org/}{tidyverse} S4 methods}
\usage{
\S4method{arrange}{data.frame}(object, ...)

\S4method{asTibble}{data.frame}(from)

\S4method{asTibble}{DataFrame}(from)

\S4method{asTibble}{matrix}(from)

\S4method{content}{response}(object)

\S4method{expand}{data.frame}(x, ...)

\S4method{filter}{data.frame}(object, ...)

\S4method{first}{data.frame}(x, ...)

\S4method{intersect}{data.frame,data.frame}(x, y, ...)

\S4method{mutate}{data.frame}(object, ...)

\S4method{rename}{data.frame}(x, ...)

\S4method{setdiff}{data.frame,data.frame}(x, y, ...)

\S4method{setequal}{data.frame,data.frame}(x, y)

\S4method{union}{data.frame,data.frame}(x, y, ...)
}
\arguments{
\item{object}{Primary object.}

\item{...}{Additional parameters.}

\item{from}{Object to coerce.}

\item{x}{Primary object.}

\item{y}{Secondary object.}
}
\description{
Avoid NAMESPACE collisions with \href{https://www.bioconductor.org/}{Bioconductor}
generics by using \code{\link[=setMethod]{setMethod()}} to relevant \code{signature}.
}
\note{
For tibble coercion with \code{\link[=as]{as()}} or \code{\link[=as_tibble]{as_tibble()}}, rownames are always
moved to the \code{rowname} column, using \code{\link[=rownames_to_column]{rownames_to_column()}} internally.
This provides more consistent behavior in the tidyverse, which can
inadvertently strip rownames during filtering operations.
}
\seealso{
\itemize{
\item \href{https://www.bioconductor.org/}{Bioconductor}:
\link{Biobase}, \link{BiocGenerics}, S4Vectors.
\item \href{http://tidyverse.org/}{tidyverse}:
\link{dplyr}, \link{httr}, \link{tidyr}.
}
}
