% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/tidyverse.R
\name{tidy_arrange}
\alias{tidy_arrange}
\alias{tidy_filter}
\alias{tidy_mutate}
\alias{tidy_select}
\alias{tidyverse}
\title{Prefixed \href{http://tidyverse.org/}{tidyverse} function alternates.}
\usage{
tidy_arrange(.data, ...)

tidy_filter(.data, ...)

tidy_mutate(.data, ...)

tidy_select(.data, ...)
}
\arguments{
\item{.data}{Tibble. All main verbs are S3 generics.}

\item{...}{Comma separated list of unquoted variable names.}
}
\description{
Arrange rows by variables.

Return rows with matching conditions.

Add new variables.

Select/rename variables by name.

Prefixed \href{http://tidyverse.org/}{tidyverse} function alternates.
}
\details{
Avoid NAMESPACE collisions with
\href{https://www.bioconductor.org}{Bioconductor} \link{BiocGenerics}.
}
\seealso{
\itemize{
\item \href{http://style.tidyverse.org/}{tidyverse style guide}
\item \code{\link[dplyr:arrange]{dplyr::arrange()}}
\item \code{\link[dplyr:filter]{dplyr::filter()}}
\item \code{\link[dplyr:mutate]{dplyr::mutate()}}
\item \code{\link[dplyr:select]{dplyr::select()}}
}
}
