% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{tidyverse}
\alias{tidyverse}
\alias{tbl_arrange}
\alias{tbl_filter}
\alias{tbl_mutate}
\alias{tbl_select}
\alias{tidy_verbs}
\title{Prefixed \href{http://tidyverse.org/}{tidyverse} grammar alternates.}
\usage{
tbl_arrange(.data, ...)

tbl_filter(.data, ...)

tbl_mutate(.data, ...)

tbl_select(.data, ...)

tidy_verbs()
}
\arguments{
\item{.data}{Tibble. All main verbs are S3 generics.}

\item{...}{Comma separated list of unquoted variable names.}
}
\description{
Prefixed \href{http://tidyverse.org/}{tidyverse} grammar alternates.

Arrange rows by variables.

Return rows with matching conditions.

Add new variables.

Select/rename variables by name.

Assign generic tidyverse verbs to the working environment.
}
\details{
Avoid NAMESPACE collisions with
\href{https://www.bioconductor.org}{Bioconductor} and other S4 generics.
}
\seealso{
\itemize{
\item \href{http://style.tidyverse.org/}{tidyverse style guide}
\item \code{\link[dplyr:arrange]{dplyr::arrange()}}
\item \code{\link[dplyr:filter]{dplyr::filter()}}
\item \code{\link[dplyr:mutate]{dplyr::mutate()}}
\item \code{\link[dplyr:select]{dplyr::select()}}
}
}
