% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{tidy_arrange}
\alias{tidy_arrange}
\alias{tidy_filter}
\alias{tidy_mutate}
\alias{tidy_select}
\title{tidyverse functions}
\usage{
tidy_arrange(.data, ...)

tidy_filter(.data, ...)

tidy_mutate(.data, ...)

tidy_select(.data, ...)
}
\description{
Arrange rows by variables.

Return rows with matching conditions.

Add new variables.

Select/rename variables by name.
}
