% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-tidyverse.R
\docType{methods}
\name{tidyverse}
\alias{tidyverse}
\alias{content}
\alias{content,response-method}
\alias{expand}
\alias{expand,data.frame-method}
\alias{first}
\alias{first,data.frame-method}
\alias{intersect}
\alias{intersect,data.frame,data.frame-method}
\alias{rename}
\alias{rename,data.frame-method}
\alias{setdiff}
\alias{setdiff,data.frame,data.frame-method}
\alias{setequal}
\alias{setequal,data.frame,data.frame-method}
\alias{union}
\alias{union,data.frame,data.frame-method}
\title{\href{http://tidyverse.org/}{tidyverse} S4 Methods}
\usage{
content(object)

\S4method{content}{response}(object)

expand(x, ...)

\S4method{expand}{data.frame}(x, ...)

first(x, ...)

\S4method{first}{data.frame}(x, ...)

intersect(x, y, ...)

\S4method{intersect}{data.frame,data.frame}(x, y, ...)

rename(x, ...)

\S4method{rename}{data.frame}(x, ...)

setdiff(x, y, ...)

\S4method{setdiff}{data.frame,data.frame}(x, y, ...)

setequal(x, y)

\S4method{setequal}{data.frame,data.frame}(x, y)

union(x, y, ...)

\S4method{union}{data.frame,data.frame}(x, y, ...)
}
\arguments{
\item{object}{Object.}

\item{x}{Primary object.}

\item{...}{Additional arguments.}

\item{y}{Secondary object.}

\item{.data}{Data object.}
}
\value{
Tidy data.
}
\description{
Avoid \code{NAMESPACE} collisions with
\href{https://www.bioconductor.org/}{Bioconductor} generics by using \code{\link[=setMethod]{setMethod()}}
to relevant \code{signature}.
}
\note{
For tibble coercion with \code{\link[=as]{as()}}, rownames are always moved to the
\code{rowname} column, using \code{\link[=rownames_to_column]{rownames_to_column()}} internally. This provides
more consistent behavior in the tidyverse, which can inadvertently strip
rownames during filtering operations.
}
\seealso{
\itemize{
\item \href{https://www.bioconductor.org/}{Bioconductor}:
Biobase, BiocGenerics, S4Vectors.
\item \href{http://tidyverse.org/}{tidyverse}:
dplyr, httr, tidyr.
}
}
