% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCellsPerSample-methods.R
\name{topCellsPerSample}
\alias{topCellsPerSample}
\alias{topCellsPerSample,SingleCellExperiment-method}
\title{Top cells per sample}
\usage{
topCellsPerSample(object, ...)

\S4method{topCellsPerSample}{SingleCellExperiment}(object, n = 100L)
}
\arguments{
\item{object}{Object.}

\item{n}{\code{integer(1)}.
Number of barcodes to return per sample.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
Top barcodes as \code{character}, split by \code{sampleID}.
}
\description{
Obtain the top cellular barcodes, based on counts.
}
\note{
Updated 2020-01-30.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment
x <- topCellsPerSample(object)
lapply(x, head)
}
