% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/topCellsPerSample-methods.R
\docType{methods}
\name{topCellsPerSample}
\alias{topCellsPerSample}
\alias{topCellsPerSample,SingleCellExperiment-method}
\title{Top Cells per Sample}
\usage{
topCellsPerSample(object, ...)

\S4method{topCellsPerSample}{SingleCellExperiment}(object, n = 100L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{\code{scalar integer}. Number of barcodes to return per sample.}
}
\value{
\code{list}. Top barcodes as \code{character}, split by \code{sampleID}.
}
\description{
Obtain the top cellular barcodes, based on counts.
}
\examples{
x <- topCellsPerSample(sce_small)
lapply(x, head)
}
\seealso{
Other SingleCellExperiment Functions: \code{\link{aggregateCellsToSamples}},
  \code{\link{cell2sample}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{subsetPerSample}}
}
\concept{SingleCellExperiment Functions}
