% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{topTables}
\alias{topTables}
\title{Top Tables of Differential Expression Results}
\usage{
topTables(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{kable}. Markdown tables.
}
\description{
Top Tables of Differential Expression Results
}
