% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString.R
\name{toStringUnique}
\alias{toStringUnique}
\alias{toStringSortUnique}
\alias{toStringSummarize}
\title{Convert an R object to a character string}
\usage{
toStringUnique(x, sep = ", ")

toStringSortUnique(x, sep = ", ")

toStringSummarize(x, sort = TRUE)
}
\arguments{
\item{x}{The object to be converted.}

\item{sep}{Separator used for collapse.}

\item{sort}{Optional arguments passed to or from methods.}
}
\value{
A character vector of length 1 is returned.
}
\description{
Helper functions for to produce a character string output for vectors and
data frames (via \code{\link[=toStringSummarize]{toStringSummarize()}}).
}
\examples{
toStringUnique(c("milk", "eggs", "eggs", NA))
toStringSortUnique(c("milk", "eggs", "eggs", NA))

mtcars \%>\% head \%>\% toStringSummarize
}
