% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString.R
\name{toString}
\alias{toString}
\title{toString call that only outputs uniques.}
\usage{
toString(x)
}
\arguments{
\item{x}{vector.}
}
\value{
string vector.
}
\description{
toString call that only outputs uniques.
}

