% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStringSortUnique.R, R/toStringSummarize.R,
%   R/toStringUnique.R
\name{toStringSortUnique}
\alias{toStringSortUnique}
\alias{toStringSummarize}
\alias{toStringUnique}
\title{Utility functions for toString}
\usage{
toStringSortUnique(character)

toStringSummarize(data)

toStringUnique(character)
}
\arguments{
\item{character}{Character vector}

\item{data}{Data with rows and columns (e.g. data frame, matrix)}
}
\value{
Sorted unique character vector

Summarized data frame that has been collapsed to a single
  \code{toString()}-formatted row, separated by commas

Unique character vector
}
\description{
Utility functions for toString
}
\examples{
toStringSortUnique(c("milk", "eggs", "eggs", NA))
toStringSummarize(head(iris))
toStringUnique(c("milk", "eggs", "eggs", NA))
}
\author{
Michael Steinbaugh
}
