% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString.R
\name{toString}
\alias{toString}
\alias{toStringUnique}
\alias{to_string_unique}
\alias{toStringSortUnique}
\alias{to_string_sort_unique}
\title{Convert an R object to a character string}
\usage{
toStringUnique(x)

to_string_unique(x)

toStringSortUnique(x)

to_string_sort_unique(x)
}
\arguments{
\item{x}{The object to be converted.}

\item{...}{Optional arguments passed to or from methods.}
}
\value{
A character vector of length 1 is returned.
}
\description{
Helper functions for \code{\link[=format]{format()}} to produce a single character string
describing an R object.
}
\examples{
toStringUnique(c("milk", "eggs", "eggs", NA))
toStringSortUnique(c("milk", "eggs", "eggs", NA))
}
