% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStringSummarize.R
\name{toStringSummarize}
\alias{toStringSummarize}
\title{Summarize all rows in a \code{tibble} using \code{toString()}}
\usage{
toStringSummarize(tibble)
}
\arguments{
\item{tibble}{Long \code{tibble}}
}
\value{
Collapsed \code{tibble}
}
\description{
Summarize all rows in a \code{tibble} using \code{toString()}
}
\keyword{general}

