% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStringSummarize.R
\name{toStringSummarize}
\alias{toStringSummarize}
\title{Summarize rows using \code{toString}}
\usage{
toStringSummarize(tibble)
}
\arguments{
\item{tibble}{Long \code{tibble}}
}
\value{
Collapsed \code{tibble}
}
\description{
Summarize rows using \code{toString}
}
\keyword{general}

