% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString.R
\name{toStringSummarize}
\alias{toStringSummarize}
\alias{toStringSummarise}
\alias{to_string_summarize}
\alias{to_string_summarise}
\title{Summarize columns using \code{\link[=toString]{toString()}}.}
\usage{
toStringSummarize(data)
}
\arguments{
\item{data}{Data with rows and columns (e.g. data frame, matrix)}
}
\value{
Summarized data frame that has been collapsed to a single
\code{\link[=toString]{toString()}}-formatted row, separated by commas.
}
\description{
Collapse column data into a single row summary columnwize using
\code{\link[=toString]{toString()}}.
}
\examples{
toStringSummarize(head(starwars))
}
