% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-methods.R
\name{toStringUnique}
\alias{toStringUnique}
\title{Convert to a Unique Character String}
\usage{
toStringUnique(object)
}
\arguments{
\item{object}{\code{atomic}.}
}
\value{
\code{character(1)}.
}
\description{
Convert to a Unique Character String
}
\examples{
toStringUnique(c("hello", "world", NA, "hello", "world", NA))
}
\seealso{
\code{toString}.
}
