% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-toStringUnique.R
\docType{methods}
\name{toStringUnique}
\alias{toStringUnique}
\alias{toStringUnique}
\alias{toStringUnique,character-method}
\title{Convert to a Unique Character String}
\usage{
toStringUnique(object, ...)

\S4method{toStringUnique}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
string.
}
\description{
Convert to a Unique Character String
}
\seealso{
\code{\link[base:toString]{base::toString()}}.
}
