% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-toStringUnique.R
\docType{methods}
\name{toStringUnique}
\alias{toStringUnique}
\alias{toStringUnique}
\alias{toStringUnique,factor-method}
\alias{toStringUnique,vector-method}
\title{Convert to a Unique Character String}
\usage{
toStringUnique(object, ...)

\S4method{toStringUnique}{factor}(object)

\S4method{toStringUnique}{vector}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
String.
}
\description{
Convert to a Unique Character String
}
\examples{
vec <- c("hello", "world", NA, "hello", "world", NA)
toStringUnique(vec)
}
\seealso{
\code{\link[base:toString]{base::toString()}}.
}
