% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStringUnique.R
\name{toStringUnique}
\alias{toStringUnique}
\title{toString call that outputs uniques.}
\usage{
toStringUnique(x)
}
\arguments{
\item{x}{vector.}
}
\value{
vector.
}
\description{
toString call that outputs uniques.
}

