% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStringUnique.R
\name{toStringUnique}
\alias{toStringUnique}
\title{toString call that outputs sorted uniques}
\usage{
toStringUnique(vector)
}
\arguments{
\item{vector}{\code{vector}}
}
\value{
Sorted unique \code{string}
}
\description{
toString call that outputs sorted uniques
}
\keyword{general}

