% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStringUnique.R
\name{toStringUnique}
\alias{toStringUnique}
\title{Convert to a Unique Character String}
\usage{
toStringUnique(object)
}
\arguments{
\item{object}{\code{atomic}.}
}
\value{
\code{string}.
}
\description{
Convert to a Unique Character String
}
\examples{
toStringUnique(c("hello", "world", NA, "hello", "world", NA))
}
\seealso{
\code{\link[base:toString]{base::toString()}}.

Other Atomic Vector Functions: \code{\link{detectLanes}},
  \code{\link{sortUnique}}
}
\author{
Michael Steinbaugh
}
\concept{Atomic Vector Functions}
