% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-toStringUnique.R
\docType{methods}
\name{toStringUnique}
\alias{toStringUnique}
\alias{toStringUnique}
\alias{toStringUnique,character-method}
\title{toString Call that Only Outputs Uniques.}
\usage{
toStringUnique(object, ...)

\S4method{toStringUnique}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
string.
}
\description{
toString Call that Only Outputs Uniques.
}
