% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStringUnique.R
\name{toStringUnique}
\alias{toStringUnique}
\title{Convert to a Unique Character String}
\usage{
toStringUnique(object)
}
\arguments{
\item{object}{Object.}
}
\value{
string.
}
\description{
Convert to a Unique Character String
}
\examples{
vec <- c("hello", "world", NA, "hello", "world", NA)
toStringUnique(vec)
}
\seealso{
\code{\link[base:toString]{base::toString()}}.

Other Atomic Vector Functions: \code{\link{grepString}},
  \code{\link{sortUnique}}
}
