% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalReads.R
\name{totalReads}
\alias{totalReads}
\title{Total reads plot}
\usage{
totalReads(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Total reads plot
}
\examples{
\dontrun{
totalReads(summary)
}
}
\author{
Michael Steinbaugh
}
