% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm-methods.R
\docType{methods}
\name{tpm}
\alias{tpm}
\alias{tpm,SummarizedExperiment-method}
\title{Transcripts Per Million (TPM)}
\usage{
\S4method{tpm}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
}
\description{
Transcripts Per Million (TPM)
}
\note{
Both gene- and transcript-level counts are supported, as long as they
were imported using a tximport caller (e.g. salmon, kallisto).
}
\examples{
\dontrun{
x <- tpm(se)
summary(x)
}
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}},
  \code{\link{topCellsPerSample}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
