% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-transcripts.R
\docType{methods}
\name{transcripts-method}
\alias{transcripts-method}
\alias{transcripts,character-method}
\title{Transcript Annotations}
\usage{
\S4method{transcripts}{character}(x, genomeBuild = NULL, release = NULL,
  return = "GRanges")
}
\arguments{
\item{x}{Primary object.}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}

\item{return}{Class of the returned object. Only applies when \code{format} equals
"\code{genes}" or "\code{transcripts}". Can be "\code{GRanges}" (default), "\code{DataFrame}",
or "\code{data.frame}". See \code{help("genes", "ensembldb")} for additional
information. "\code{gene2symbol}" and "\code{tx2gene}" are always returned as
\code{data.frame}.}
}
\value{
\code{GRanges}, \code{DataFrame}, or \code{data.frame} containing gene or
transcript annotations.
}
\description{
Transcript Annotations
}
\details{
Simply specify the desired organism, using the full latin name. For example,
we can obtain human annotations with \code{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.
}
\section{Broad Class Definitions}{

For gene and transcript tables, when \code{broadClass = TRUE}, a \code{broadClass}
column is added, which generalizes the gene types into a smaller number of
semantically-meaningful groups:
\itemize{
\item \code{coding}
\item \code{noncoding}
\item \code{pseudo}
\item \code{small}
\item \code{decaying}
\item \code{ig} (immunoglobulin)
\item \code{tcr} (T cell receptor)
\item \code{other}
}
}

\examples{
data <- transcripts("Homo sapiens")
summary(data)
colnames(mcols(data))
}
\seealso{
Other Gene Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{ensembl}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{genes-method}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{tx2geneFromGFF}}, \code{\link{tx2gene}}
}
