% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-transmit.R
\docType{methods}
\name{transmit}
\alias{transmit}
\alias{transmit,character-method}
\title{Transmit (Download) Files from a Remote Server}
\usage{
transmit(object, ...)

\S4method{transmit}{character}(object, pattern = NULL, rename = NULL,
  compress = FALSE, localDir = "data-raw")
}
\arguments{
\item{object}{Remote directory URL.}

\item{...}{Additional arguments.}

\item{pattern}{Pattern to match against remote file names.}

\item{rename}{Rename the local file (including suffix), if desired.}

\item{compress}{Compress the file with \code{\link[=gzip]{gzip()}} after download.
(\code{TRUE}/\code{FALSE})}

\item{localDir}{Directory where to save file locally.}
}
\value{
List of local files.
}
\description{
Utility function that supports file matching on a remote server. Also enables
users to rename and compress on the fly.
}
