% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-transmit.R
\docType{methods}
\name{transmit}
\alias{transmit}
\alias{transmit,character-method}
\title{Transmit Files from a Remote FTP Server}
\usage{
transmit(object, ...)

\S4method{transmit}{character}(object, pattern, rename = NULL,
  compress = FALSE, localDir = "data-raw", quiet = FALSE)
}
\arguments{
\item{object}{Remote directory URL. Currently supports FTP.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{pattern}{Pattern to match against remote file names.}

\item{rename}{Rename the local file (including suffix), if desired.}

\item{compress}{Compress the file with \code{\link[=gzip]{gzip()}} after download.
(\code{TRUE}/\code{FALSE})}

\item{localDir}{Directory where to save file locally.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
List of local files.
}
\description{
Utility function that supports easy file matching and download from a remote
FTP server. Also enables on-the-fly file renaming and compression.
}
\examples{
transmit("ftp://ftp.ensembl.org/pub/release-90",
         pattern = "README",
         rename = "ensembl_readme.txt",
         compress = TRUE)
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{readSampleMetadataFile}},
  \code{\link{readYAML}}
}
\concept{Data Import and Project Utilities}
