% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{transmit}
\alias{transmit}
\title{Transmit Files from a Remote FTP Server}
\usage{
transmit(remoteDir, localDir = ".", pattern, rename = NULL,
  compress = FALSE)
}
\arguments{
\item{remoteDir}{Remote directory URL. Currently supports FTP. Works either
with or without the trailing slash.}

\item{localDir}{Directory where to save files locally.}

\item{pattern}{Pattern to match against remote file names.}

\item{rename}{Rename the local files (including suffix), if desired.}

\item{compress}{Compress the files with \code{\link[R.utils:gzip]{R.utils::gzip()}} after download.
(\code{TRUE}/\code{FALSE})}
}
\value{
Invisible vector of local file paths.
}
\description{
Utility function that supports easy file matching and download from a remote
FTP server. Also enables on-the-fly file renaming and compression.
}
\examples{
readme <- transmit(
    remoteDir = "ftp://ftp.ensembl.org/pub/release-90",
    pattern = "README",
    rename = "ensembl_readme.txt",
    compress = TRUE
)
basename(readme)
file.exists(readme)

# Clean up
unlink("ensembl_readme.txt.gz")
}
\seealso{
Other Write Functions: \code{\link{assignAndSaveData}},
  \code{\link{initializeDirectory}},
  \code{\link{saveData}}, \code{\link{writeCounts}}
}
\author{
Michael Steinbaugh
}
