% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{transmit}
\alias{transmit}
\title{Transmit Files from a Remote FTP Server}
\usage{
transmit(remoteDir, localDir = getwd(), pattern, rename = NULL,
  compress = FALSE, quiet = FALSE)
}
\arguments{
\item{remoteDir}{Remote directory URL. Currently supports FTP. Works either
with or without the trailing slash.}

\item{localDir}{Directory where to save file locally.}

\item{pattern}{Pattern to match against remote file names.}

\item{rename}{Rename the local file (including suffix), if desired.}

\item{compress}{Compress the file with \code{\link[=gzip]{gzip()}} after download.
(\code{TRUE}/\code{FALSE})}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Invisibly return a list of the files downloaded.
}
\description{
Utility function that supports easy file matching and download from a remote
FTP server. Also enables on-the-fly file renaming and compression.
}
\examples{
transmit(
    remoteDir = "ftp://ftp.ensembl.org/pub/release-90",
    pattern = "README",
    rename = "ensembl_readme.txt",
    compress = TRUE)
}
\seealso{
Other Data Import and Project Utilities: \code{\link{loadRemoteData}},
  \code{\link{readFileByExtension}}, \code{\link{readGFF}},
  \code{\link{readYAML}}
}
