% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{transmit}
\alias{transmit}
\title{Transmit (download) files from a remote server}
\usage{
transmit(remote_dir, pattern = NULL, rename = NULL, compress = FALSE,
  local_dir = "data-raw")
}
\arguments{
\item{remote_dir}{Remote directory URL.}

\item{pattern}{Pattern to match against remote file names.}

\item{rename}{Rename the local file (including suffix), if desired.}

\item{compress}{Compress the file with \code{\link[=gzip]{gzip()}} after download.
(\code{TRUE}/\code{FALSE})}

\item{local_dir}{Directory where to save file locally.}
}
\value{
List of local files.
}
\description{
Utility function that supports file matching on a remote server. Also enables
users to rename and compress on the fly.
}
