% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transmit-methods.R
\docType{methods}
\name{transmit}
\alias{transmit}
\alias{transmit,character-method}
\title{Transmit Files from a Remote FTP Server}
\usage{
transmit(object, ...)

\S4method{transmit}{character}(object, pattern = NULL, rename = NULL,
  compress = FALSE, localDir = "data-raw")
}
\arguments{
\item{object}{Remote directory URL. Currently supports FTP.}

\item{...}{Additional arguments.}

\item{pattern}{Pattern to match against remote file names.}

\item{rename}{Rename the local file (including suffix), if desired.}

\item{compress}{Compress the file with \code{\link[=gzip]{gzip()}} after download.
(\code{TRUE}/\code{FALSE})}

\item{localDir}{Directory where to save file locally.}
}
\value{
List of local files.
}
\description{
Utility function that supports easy file matching and download from a remote
FTP server. Also enables on-the-fly file renaming and compression.
}
\examples{
transmit("ftp://ftp.ensembl.org/pub/release-89",
         pattern = "README",
         rename = "ensembl_readme.txt",
         compress = TRUE)
}
\seealso{
Other Data Import and Project Utilities: \code{\link{createProjectDirs}},
  \code{\link{loadData}}, \code{\link{loadRemoteData}},
  \code{\link{packageSE}}, \code{\link{readDataRaw}},
  \code{\link{readFileByExtension}}, \code{\link{readYAML}}
}
