% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{transmit}
\alias{transmit}
\title{Transmit files from a remote FTP server}
\usage{
transmit(remoteDir, localDir = ".", pattern, rename = NULL,
  compress = FALSE)
}
\arguments{
\item{remoteDir}{\code{character(1)}.
Remote FTP directory path.}

\item{localDir}{\code{character(1)}.
Directory where to save files locally.}

\item{pattern}{\code{character(1)}.
Pattern to match against remote file names.}

\item{rename}{\code{character(1)} or \code{NULL}.
Rename the local files (including suffix), if desired.}

\item{compress}{\code{logical(1)}.
gzip compress the files after download.}
}
\value{
Invisible \code{character}.
Local file paths.
}
\description{
Utility function that supports easy file matching and download from a remote
FTP server. Also enables on-the-fly file renaming and compression.
}
\examples{
remoteDir <- paste(
    "ftp://ftp.pantherdb.org",
    "sequence_classifications",
    "current_release",
    sep = "/"
)
readme <- transmit(
    remoteDir = remoteDir,
    pattern = "README",
    rename = "panther_readme.txt",
    compress = TRUE
)
basename(readme)
file.exists(readme)

## Clean up.
unlink(readme)
}
