% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/Tx2Gene-methods.R
\docType{class}
\name{Tx2Gene-class}
\alias{Tx2Gene-class}
\alias{Tx2Gene}
\alias{Tx2Gene,DataFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,SummarizedExperiment-method}
\title{Transcript-to-gene identifier mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{DataFrame}(object)

\S4method{Tx2Gene}{GRanges}(object)

\S4method{Tx2Gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}.
Contains a \code{DataFrame} with \code{transcriptID} and \code{geneID} columns.
}
\description{
Transcript-to-gene identifier mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{metadata}.
}

\examples{
## SummarizedExperiment ====
data(SummarizedExperiment_transcripts, package = "acidtest")
txse <- SummarizedExperiment_transcripts

## SummarizedExperiment ====
x <- Tx2Gene(txse)
print(x)
}
\seealso{
\code{makeTx2Gene()}.
}
\author{
Michael Steinbaugh
}
