% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerators.R
\docType{class}
\name{Tx2Gene-class}
\alias{Tx2Gene-class}
\alias{Tx2Gene,DataFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,SummarizedExperiment-method}
\title{Transcript-to-gene identifier mappings}
\usage{
\S4method{Tx2Gene}{DataFrame}(object)

\S4method{Tx2Gene}{GRanges}(object)

\S4method{Tx2Gene}{SummarizedExperiment}(object)
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\details{
Contains a \code{DataFrame} with \code{transcriptID} and \code{geneID} columns.
}
\note{
Updated 2019-08-08.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{metadata}.
}

