% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Tx2Gene-class}
\alias{Tx2Gene-class}
\title{Transcript-to-gene identifier mappings}
\value{
\code{Tx2Gene}. Contains a \code{DataFrame} with \code{transcriptID} and \code{geneID}
columns.
}
\description{
Transcript-to-gene identifier mappings
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{metadata}.
}

\seealso{
\code{makeTx2Gene}.
}
\author{
Michael Steinbaugh
}
