% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Tx2Gene-class}
\alias{Tx2Gene-class}
\title{\code{Tx2Gene} Class}
\description{
Transcript-to-gene mappings.
}
\details{
Contains a \code{DataFrame} with \code{transcriptID} and \code{geneID} columns.
}
\seealso{
\itemize{
\item \link{makeTx2Gene}.
\item \code{\link[=tx2gene]{tx2gene()}}.
}

Other S4 Classes: \code{\link{EggNOG-class}},
  \code{\link{Ensembl2Entrez-class}},
  \code{\link{Gene2Symbol-class}},
  \code{\link{HGNC2Ensembl-class}},
  \code{\link{MGI2Ensembl-class}},
  \code{\link{PANTHER-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
