% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{tx2gene-class}
\alias{tx2gene-class}
\title{\code{tx2gene} Class}
\description{
Transcript-to-gene mappings.
}
\details{
Contains a \code{DataFrame} with \code{transcriptID} and \code{geneID} columns.
}
\examples{
x <- new(
    "tx2gene",
    DataFrame(
        transcriptID = "ENST00000000233",
        geneID = "ENSG00000004059",
        row.names = "ENST00000000233"
    )
)
print(x)
}
