% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Tx2Gene-class}
\alias{Tx2Gene-class}
\title{Transcript-to-gene identifier mappings}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\details{
Contains a \code{DataFrame} with \code{transcriptID} and \code{geneID} columns.
}
\note{
Updated 2020-10-05.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{metadata}.
}

