% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Tx2Gene-methods.R
\docType{methods}
\name{Tx2Gene}
\alias{Tx2Gene}
\alias{Tx2Gene,DataFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,SummarizedExperiment-method}
\title{Transcript-to-Gene Identifier Mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{DataFrame}(object)

\S4method{Tx2Gene}{GRanges}(object)

\S4method{Tx2Gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}. Contains a \code{DataFrame} with \code{transcriptID} and \code{geneID}
columns.
}
\description{
Transcript-to-Gene Identifier Mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{S4Vectors::metadata()}.
}

\examples{
## SummarizedExperiment ====
data(tx_se)
x <- Tx2Gene(tx_se)
print(x)
}
\seealso{
\link{makeTx2Gene}.
}
