% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllGenerators.R
\docType{methods}
\name{Tx2Gene}
\alias{Tx2Gene}
\alias{Tx2Gene,DataFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,SummarizedExperiment-method}
\title{Transcript-to-gene identifier mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{DataFrame}(object)

\S4method{Tx2Gene}{GRanges}(object)

\S4method{Tx2Gene}{SummarizedExperiment}(object)
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.

Updated 2019-08-13.
}
\examples{
## SummarizedExperiment ====
data(SummarizedExperiment_transcripts, package = "acidtest")
txse <- SummarizedExperiment_transcripts

## SummarizedExperiment ====
x <- Tx2Gene(txse)
print(x)
}
\seealso{
\code{\link[=makeTx2Gene]{makeTx2Gene()}}.
}
