% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Tx2Gene-methods.R
\docType{methods}
\name{Tx2Gene}
\alias{Tx2Gene}
\alias{Tx2Gene,matrix-method}
\alias{Tx2Gene,data.frame-method}
\alias{Tx2Gene,DataFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,SummarizedExperiment-method}
\title{Transcript-to-gene identifier mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{matrix}(object)

\S4method{Tx2Gene}{data.frame}(object)

\S4method{Tx2Gene}{DataFrame}(object, metadata = TRUE)

\S4method{Tx2Gene}{GRanges}(object)

\S4method{Tx2Gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{metadata}{\code{logical(1)}.
Slot genome metadata.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.

Updated 2019-10-24.
}
\examples{
## SummarizedExperiment ====
data(SummarizedExperiment_transcripts, package = "acidtest")
txse <- SummarizedExperiment_transcripts

## SummarizedExperiment ====
x <- Tx2Gene(txse)
print(x)
}
\seealso{
\code{\link[=makeTx2Gene]{makeTx2Gene()}}.
}
