% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{Tx2Gene}
\alias{Tx2Gene}
\title{Transcript-to-gene identifier mappings}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.

Updated 2019-08-08.
}
\examples{
## SummarizedExperiment ====
data(SummarizedExperiment_transcripts, package = "acidtest")
txse <- SummarizedExperiment_transcripts

## SummarizedExperiment ====
x <- Tx2Gene(txse)
print(x)
}
\seealso{
\code{\link[=makeTx2Gene]{makeTx2Gene()}}.
}
