% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx2gene-methods.R
\docType{methods}
\name{tx2gene}
\alias{tx2gene}
\alias{tx2gene,SummarizedExperiment-method}
\title{Transcript-to-Gene Mappings}
\usage{
\S4method{tx2gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DataFrame}.
}
\description{
Transcript-to-Gene Mappings
}
\examples{
# SummarizedExperiment ====
x <- tx2gene(tx_rse_small)
print(x)
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{gene2symbol}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{sampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
