% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tx2gene-methods.R
\docType{methods}
\name{tx2gene}
\alias{tx2gene}
\alias{tx2gene,SummarizedExperiment-method}
\title{#' @rdname subsetPerSample
#' @export
setGeneric(
"subsetPerSample",
function(object, ...) {
standardGeneric("subsetPerSample")
}
)}
\usage{
tx2gene(object, ...)

\S4method{tx2gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{tx2gene}.
}
\description{
#' @rdname subsetPerSample
#' @export
setGeneric(
"subsetPerSample",
function(object, ...) {
standardGeneric("subsetPerSample")
}
)

Transcript-to-Gene Mappings
}
\examples{
# SummarizedExperiment ====
x <- tx2gene(tx_se_small)
print(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
