% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tx2gene-methods.R
\docType{methods}
\name{tx2gene}
\alias{tx2gene}
\alias{tx2gene,SummarizedExperiment-method}
\title{Transcript-to-Gene Mappings}
\usage{
tx2gene(object, ...)

\S4method{tx2gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-Gene Mappings
}
\examples{
# SummarizedExperiment ====
x <- tx2gene(tx_se_small)
print(x)
}
\seealso{
Other S4 Generators: \code{\link{eggnog}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{hgnc2ensembl}}, \code{\link{mgi2ensembl}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
