% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Tx2Gene-methods.R,
%   R/deprecated.R
\docType{methods}
\name{Tx2Gene}
\alias{Tx2Gene}
\alias{Tx2Gene,data.frame-method}
\alias{Tx2Gene,DataFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,SummarizedExperiment-method}
\alias{tx2gene}
\title{Transcript-to-Gene Identifier Mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{data.frame}(object)

\S4method{Tx2Gene}{DataFrame}(object)

\S4method{Tx2Gene}{GRanges}(object)

\S4method{Tx2Gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-Gene Identifier Mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.
}
\examples{
## SummarizedExperiment ====
data(tx_se_small)
x <- Tx2Gene(tx_se_small)
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{Ensembl2Entrez}},
  \code{\link{Gene2Symbol}}, \code{\link{HGNC2Ensembl}},
  \code{\link{MGI2Ensembl}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{mapGenes}}

Other Transcript-Level Functions: \code{\link{convertTranscriptsToGenes}},
  \code{\link{makeGRanges}}, \code{\link{makeTx2Gene}},
  \code{\link{stripTranscriptVersions}}
}
\concept{Identifier Mapping Functions}
\concept{Transcript-Level Functions}
