% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tx2gene.R
\docType{methods}
\name{tx2gene}
\alias{tx2gene}
\alias{tx2gene,character-method}
\title{Convert Ensembl Transcripts to Genes}
\usage{
tx2gene(object)

\S4method{tx2gene}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Same class as original object.
}
\description{
Convert Ensembl Transcripts to Genes
}
\examples{
# character
tx2gene(c("ENSMUST00000000001",
          "ENSMUST00000000003",
          "ENSMUST00000114041"))
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}}, \code{\link{gene2entrez}},
  \code{\link{gene2symbol}}
}
