% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-tx2gene.R
\docType{methods}
\name{tx2gene}
\alias{tx2gene}
\alias{tx2gene}
\alias{tx2gene,character-method}
\title{Transcript to Gene Mappings}
\usage{
tx2gene(object, ...)

\S4method{tx2gene}{character}(object, genomeBuild = NULL, release = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genomeBuild}{\emph{Optional.} Genome assembly name (e.g. GRCh38).}

\item{release}{\emph{Optional.} Ensembl release version. Defaults to the most
current release available on AnnotationHub.}
}
\value{
\link{data.frame}.
}
\description{
Transcript to Gene Mappings
}
\examples{
tx2gene("Homo sapiens") \%>\% glimpse()
}
\seealso{
Other Gene Functions: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{ensembl}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{genes}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{transcripts}}, \code{\link{tx2geneFromGFF}}
}
