% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tx2gene-methods.R
\docType{methods}
\name{tx2gene}
\alias{tx2gene}
\alias{Tx2Gene}
\alias{tx2gene,data.frame-method}
\alias{tx2gene,DataFrame-method}
\alias{tx2gene,GRanges-method}
\alias{tx2gene,SummarizedExperiment-method}
\title{Transcript-to-Gene Identifier Mappings}
\usage{
tx2gene(object, ...)

\S4method{tx2gene}{data.frame}(object)

\S4method{tx2gene}{DataFrame}(object)

\S4method{tx2gene}{GRanges}(object)

\S4method{tx2gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-Gene Identifier Mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.
}
\examples{
# SummarizedExperiment ====
data(tx_se_small)
x <- tx2gene(tx_se_small)
print(x)
}
\seealso{
Other Identifier Mapping Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{convertUCSCBuildToEnsembl}},
  \code{\link{ensembl2entrez}}, \code{\link{gene2symbol}},
  \code{\link{hgnc2ensembl}}, \code{\link{mapGenes}},
  \code{\link{mgi2ensembl}}

Other Transcript-Level Functions: \code{\link{convertTranscriptsToGenes}},
  \code{\link{makeGRanges}}, \code{\link{makeTx2Gene}},
  \code{\link{stripTranscriptVersions}}
}
\author{
Michael Steinbaugh
}
\concept{Identifier Mapping Functions}
\concept{Transcript-Level Functions}
