% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aliases.R,
%   R/methods-tx2geneFromGFF.R
\docType{methods}
\name{tx2geneFromGFF}
\alias{tx2geneFromGFF}
\alias{tx2geneFromGTF}
\alias{tx2geneFromGFF,character-method}
\alias{tx2geneFromGFF,data.frame-method}
\title{Generate tx2gene from GFF/GTF File}
\usage{
tx2geneFromGFF(object, ...)

tx2geneFromGTF(...)

\S4method{tx2geneFromGFF}{character}(object, quiet = FALSE)

\S4method{tx2geneFromGFF}{data.frame}(object, quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Generate tx2gene from GFF/GTF File
}
\details{
The GFF (General Feature Format) format consists of one line per
feature, each containing 9 columns of data, plus optional track definition
lines. The GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
# From URL (recommended)
url <- file.path(testDataURL, "mmusculus.gtf")
tx2geneFromGFF(url) \%>\% str()

# GFF data.frame
gff <- readGFF(url)
tx2geneFromGFF(gff) \%>\% str()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{symbol2gene}},
  \code{\link{tx2gene}}
}
\concept{Gene Annotation Utilities}
