% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-tx2geneFromGFF.R
\docType{methods}
\name{tx2geneFromGFF}
\alias{tx2geneFromGFF}
\alias{tx2geneFromGFF}
\alias{tx2geneFromGFF,character-method}
\alias{tx2geneFromGFF,data.frame-method}
\alias{tx2geneFromGTF}
\title{Define Transcript to Gene Mappings from GFF/GTF File}
\usage{
tx2geneFromGFF(object, ...)

\S4method{tx2geneFromGFF}{character}(object)

\S4method{tx2geneFromGFF}{data.frame}(object)

tx2geneFromGTF(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{data.frame}.
}
\description{
Define Transcript to Gene Mappings from GFF/GTF File
}
\details{
The GFF (General Feature Format) format consists of one line per
feature, each containing 9 columns of data, plus optional track definition
lines. The GTF (General Transfer Format) is identical to GFF version 2.
}
\examples{
# From URL (recommended)
url <- "http://basejump.seq.cloud/mmusculus.gtf"
tx2geneFromGFF(url) \%>\% glimpse()

# GFF data.frame
gff <- readGFF(url)
tx2geneFromGFF(gff) \%>\% glimpse()
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{convertGenesToSymbols}},
  \code{\link{convertTranscriptsToGenes}},
  \code{\link{detectOrganism}}, \code{\link{ensembl}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{genes}},
  \code{\link{sanitizeAnnotable}},
  \code{\link{stripTranscriptVersions}},
  \code{\link{transcripts}}, \code{\link{tx2gene}}
}
