% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-tx2geneFromGTF.R
\docType{methods}
\name{tx2geneFromGTF}
\alias{tx2geneFromGTF}
\alias{tx2geneFromGTF}
\alias{tx2geneFromGTF,character-method}
\title{Generate tx2gene from GTF File}
\usage{
tx2geneFromGTF(object, ...)

\S4method{tx2geneFromGTF}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{data.frame}.
}
\description{
Generate tx2gene from GTF File
}
\examples{
# Mouse (Ensembl)
tx2geneFromGTF("http://steinbaugh.com/basejump/tests/mmusculus.gtf")

# Fruitfly (FlyBase)
tx2geneFromGTF("http://steinbaugh.com/basejump/tests/dmelanogaster.gtf")
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGTF}},
  \code{\link{gene2symbol}}, \code{\link{tx2gene}}
}
