% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-tx2geneFromGTF.R
\name{tx2geneFromGTF}
\alias{tx2geneFromGTF}
\alias{tx2geneFromGTF}
\title{Generate tx2gene from GTF File}
\usage{
tx2geneFromGTF(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{save}{Save tx2gene to disk as RData and CSV files.}
}
\value{
\link{data.frame}.
}
\description{
Generate tx2gene from GTF File
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}},
  \code{\link{ensemblVersion}}, \code{\link{gene2entrez}},
  \code{\link{gene2symbol}}, \code{\link{tx2gene}}
}
