% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-tx2geneFromGTF.R
\docType{methods}
\name{tx2geneFromGTF}
\alias{tx2geneFromGTF}
\alias{tx2geneFromGTF}
\alias{tx2geneFromGTF,character-method}
\alias{tx2geneFromGTF,data.frame-method}
\title{Generate tx2gene from GTF File}
\usage{
tx2geneFromGTF(object, ...)

\S4method{tx2geneFromGTF}{character}(object)

\S4method{tx2geneFromGTF}{data.frame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{data.frame}.
}
\description{
Generate tx2gene from GTF File
}
\examples{
# GTF URL
url <- file.path("http://steinbaugh.com",
                 "basejump",
                 "tests",
                 "mmusculus.gtf")
tx2geneFromGTF(url) \%>\%
    glimpse

# GTF data.frame
gtf <- readGTF(url)
tx2geneFromGTF(gtf) \%>\%
    glimpse
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{detectOrganism}},
  \code{\link{gene2symbolFromGTF}},
  \code{\link{gene2symbol}}, \code{\link{tx2gene}}
}
