% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniteInterestingGroups-methods.R
\docType{methods}
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups,DataFrame-method}
\title{Unite interesting groups into a single column}
\usage{
\S4method{uniteInterestingGroups}{DataFrame}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
Modified object.
Now contains an \code{interestingGroups} column.
}
\description{
Create a single interesting groups column (\code{interestingGroups}) used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by a colon.
}
\examples{
data(rse)
object <- rse
from <- sampleData(object)
to <- uniteInterestingGroups(
    object = from,
    interestingGroups = interestingGroups(object)
)
print(to)
}
