% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/uniteInterestingGroups-methods.R
\docType{methods}
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups,data.frame-method}
\alias{uniteInterestingGroups,DataFrame-method}
\alias{uniteInterestingGroups,tbl_df-method}
\title{Unite Interesting Groups}
\usage{
uniteInterestingGroups(object, ...)

\S4method{uniteInterestingGroups}{data.frame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{DataFrame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{tbl_df}(object, interestingGroups)
}
\arguments{
\item{object}{Object containing column data that defines interesting groups.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\code{character}. Groups of interest.}
}
\value{
Modified object, containing an \code{interestingGroups} column.
}
\description{
Create a single interesting groups column ("\code{interestingGroups}") used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by a colon.
}
\examples{
data(rse_small)
object <- rse_small
from <- sampleData(object)
to <- uniteInterestingGroups(
    object = from,
    interestingGroups = interestingGroups(object)
)
print(to)
}
\seealso{
Other Metadata Functions: \code{\link{detectHPC}},
  \code{\link{detectLanes}}, \code{\link{detectOrganism}},
  \code{\link{interestingGroups}},
  \code{\link{microplate}},
  \code{\link{minimalSampleData}}, \code{\link{organism}},
  \code{\link{sampleData}}, \code{\link{sampleNames}}
}
\author{
Michael Steinbaugh
}
\concept{Metadata Functions}
