% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/uniteInterestingGroups-methods.R
\docType{methods}
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups,data.frame-method}
\alias{uniteInterestingGroups,DataFrame-method}
\alias{uniteInterestingGroups,tbl_df-method}
\title{Unite Interesting Groups}
\usage{
uniteInterestingGroups(object, ...)

\S4method{uniteInterestingGroups}{data.frame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{DataFrame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{tbl_df}(object, interestingGroups)
}
\arguments{
\item{object}{Object containing column data that defines interesting groups.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\code{character}. Groups of interest.}
}
\value{
Modified object, containing an \code{interestingGroups} column.
}
\description{
Create a single interesting groups column ("\code{interestingGroups}") used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by a colon.
}
\examples{
object <- sampleData(rse_small)
x <- uniteInterestingGroups(
    object = object,
    interestingGroups = c("genotype", "treatment")
)
print(x)
}
\seealso{
Other Developer Functions: \code{\link{calls}},
  \code{\link{cleanSystemLibrary}},
  \code{\link{detectHPC}}, \code{\link{dots}},
  \code{\link{flatFiles}},
  \code{\link{initializeDirectory}},
  \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
