% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/uniteInterestingGroups-methods.R
\docType{methods}
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups,data.frame-method}
\alias{uniteInterestingGroups,DataFrame-method}
\alias{uniteInterestingGroups,tbl_df-method}
\title{Unite Interesting Groups}
\usage{
uniteInterestingGroups(object, ...)

\S4method{uniteInterestingGroups}{data.frame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{DataFrame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{tbl_df}(object, interestingGroups)
}
\arguments{
\item{object}{Object containing multiple interesting groups.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\code{character}. Groups of interest.}
}
\value{
Data frame containing an \code{interestingGroups} column.
}
\description{
Create a single interesting groups column ("\code{interestingGroups}") used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by a colon.
}
\examples{
x <- sampleData(rse_bcb)
y <- uniteInterestingGroups(x, interestingGroups = c("treatment", "day"))
y[["interestingGroups"]]
}
\seealso{
Other Developer Functions: \code{\link{detectHPC}},
  \code{\link{dots}}, \code{\link{localOrRemoteFile}},
  \code{\link{matchInterestingGroups}},
  \code{\link{methodFormals}},
  \code{\link{multiassignAsEnvir}},
  \code{\link{printString}}
}
\author{
Michael Steinbaugh
}
\concept{Developer Functions}
