% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/uniteInterestingGroups-methods.R
\docType{methods}
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups,data.frame-method}
\alias{uniteInterestingGroups,DataFrame-method}
\alias{uniteInterestingGroups,tbl_df-method}
\title{Unite Interesting Groups}
\usage{
uniteInterestingGroups(object, ...)

\S4method{uniteInterestingGroups}{data.frame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{DataFrame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{tbl_df}(object, interestingGroups)
}
\arguments{
\item{object}{Object containing interesting groups in multiple columns.}

\item{...}{Additional arguments.}
}
\value{
Data frame, now containing an \code{interestingGroups} column.
}
\description{
Create a single interesting groups column ("\code{interestingGroups}") used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by a colon.
}
\examples{
x <- sampleData(rse_bcb, interestingGroups = NULL)
y <- uniteInterestingGroups(x, interestingGroups = c("treatment", "day"))
y[["interestingGroups"]]
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{gene2symbol}}, \code{\link{sampleData}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
