% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniteInterestingGroups.R
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\title{Unite Interesting Groups}
\usage{
uniteInterestingGroups(object, interestingGroups)
}
\arguments{
\item{object}{Object (e.g. \link{data.frame}) containing interesting groups
columns.}

\item{interestingGroups}{Character vector of interesting groups.}
}
\value{
\link{data.frame}.
}
\description{
Create a single interesting groups column (\code{interestingGroups}) used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by \code{:}.
}
