% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{updateMessage}
\alias{updateMessage}
\title{Update Message}
\format{An object of class \code{character} of length 1.}
\usage{
updateMessage
}
\description{
Update Message
}
\examples{
message(updateMessage)
}
\keyword{internal}
