% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/zzz.R
\name{updatePackages}
\alias{updatePackages}
\alias{update_packages}
\title{Update All Installed Packages}
\usage{
updatePackages()
}
\description{
Ensure that all \href{https://bioconductor.org}{Bioconductor},
\href{https://cran.r-project.org}{CRAN}, and \href{https://github.com}{GitHub} packages
are up to date.
}
\details{
Internaly this is a wrapper for \code{\link[devtools:update_packages]{devtools::update_packages()}} that adds
Bioconductor support.
}
