% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updatePackages}
\alias{updatePackages}
\alias{update_packages}
\title{Update all installed packages}
\usage{
updatePackages()
}
\description{
Ensure that all GitHub, Bioconductor, and CRAN packages are up to date.
}
\keyword{internal}
