% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/upperCamelCase-methods.R
\name{upperCamelCase}
\alias{upperCamelCase}
\alias{upperCamelCase,atomic-method}
\alias{upperCamelCase,factor-method}
\alias{upperCamelCase,list-method}
\alias{upperCamelCase,matrix-method}
\alias{upperCamelCase,data.frame-method}
\alias{upperCamelCase,Vector-method}
\alias{upperCamelCase,DataFrame-method}
\alias{upperCamelCase,Ranges-method}
\alias{upperCamelCase,Matrix-method}
\alias{upperCamelCase,SummarizedExperiment-method}
\title{Upper camel case}
\usage{
upperCamelCase(object, ...)

\S4method{upperCamelCase}{atomic}(object, names = TRUE, strict = FALSE)

\S4method{upperCamelCase}{factor}(object, names = TRUE, strict = FALSE)

\S4method{upperCamelCase}{list}(object, names = TRUE, strict = FALSE)

\S4method{upperCamelCase}{matrix}(object, rownames = FALSE, colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{data.frame}(object, rownames = FALSE, colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{Vector}(
  object,
  names = TRUE,
  mcols = TRUE,
  metadata = TRUE,
  strict = FALSE
)

\S4method{upperCamelCase}{DataFrame}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  mcols = TRUE,
  metadata = TRUE,
  strict = FALSE
)

\S4method{upperCamelCase}{Ranges}(
  object,
  names = FALSE,
  mcols = TRUE,
  metadata = TRUE,
  strict = FALSE
)

\S4method{upperCamelCase}{Matrix}(object, rownames = FALSE, colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{SummarizedExperiment}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  assayNames = TRUE,
  rowData = TRUE,
  colData = TRUE,
  metadata = TRUE,
  strict = FALSE
)
}
\arguments{
\item{object}{Object.}

\item{names}{\code{logical(1)}.
Sanitize names.

Only applies to string mode (\code{rename = FALSE}).}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{mcols}{\code{logical(1)}.
Sanitize names of metadata columns (i.e. \code{DataFrame}).}

\item{metadata}{\code{logical(1)}.
Sanitize metadata names.}

\item{assayNames}{\code{logical(1)}.
Sanitize assay names.}

\item{rowData}{\code{logical(1)}.
Sanitize the row data names.}

\item{colData}{\code{logical(1)}.
Sanitize column names of column data.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names.
For objects with \code{\link[base:names]{names()}} defined, the underlying data
returns unchanged, except for \code{character} or \code{vector} class.
}
\description{
Format character strings to use upper camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{ThisIsCamelCase}).
}
\details{
Note that lower camel case is generally recommended in R over the use of
upper camel case. However, upper camel case is recommended by Bioconductor
for S4 class names and corresponding generators, but not variables or
functions.
}
\note{
Updated 2020-05-11.
}
\examples{
data(syntactic, package = "AcidTest")
lapply(syntactic, upperCamelCase)
}
