% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validateClasses}
\alias{validateClasses}
\title{Validate Classes}
\usage{
validateClasses(object, expected, subset = FALSE)
}
\arguments{
\item{object}{Object.}

\item{expected}{\code{list}. Named list of expected classes per slot.}

\item{subset}{\code{boolean}. Only check a subset of slots in the object.}
}
\value{
\code{boolean} (TRUE) on sucess or \code{string} containing informative
message on failure.
}
\description{
Utility function to check multiple classes in a single call.
}
\details{
To be used inside S4 \code{\link[methods:setValidity]{methods::setValidity()}} call or with
\code{\link[assertthat:validate_that]{assertthat::validate_that()}}.
}
\examples{
data(rse)
object <- rse
metadata <- S4Vectors::metadata(object)
validateClasses(
    object = metadata,
    expected = list(
        version = c("package_version", "numeric_version"),
        date = "Date",
        interestingGroups = "character"
    )
)
}
\seealso{
\code{\link[assertthat:validate_that]{assertthat::validate_that()}}.
}
