% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-wash.R
\docType{methods}
\name{wash}
\alias{wash}
\alias{.wash}
\alias{wash,data.frame-method}
\alias{wash,DataFrame-method}
\alias{wash,tbl_df-method}
\title{Wash Column Data}
\usage{
wash(object)

.wash(object)

\S4method{wash}{data.frame}(object)

\S4method{wash}{DataFrame}(object)

\S4method{wash}{tbl_df}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Clean object.
}
\description{
This function cleans poorly formed separators, leading and trailing commas or
spaces, empty cells, and sets \code{NA} values if necessary.
}
