% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wash.R
\name{wash}
\alias{wash}
\title{Wash a \code{data.frame}}
\usage{
wash(data)
}
\arguments{
\item{data}{A \code{data.frame} with leading/trailing spaces/commas, empty cells.}
}
\value{
A reformatted, clean \code{data.frame}.
}
\description{
Wash a \code{data.frame}
}
