% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wash.R
\name{wash}
\alias{wash}
\title{Wash}
\usage{
wash(df)
}
\arguments{
\item{df}{Dirty \code{data.frame}}
}
\value{
Washed \code{data.frame}
}
\description{
This function cleans poorly formed separators, leading and trailing commas or
spaces, empty cells, and sets \code{NA} values if necessary.
}
\author{
Michael Steinbaugh
}
\keyword{dev}
\keyword{text}
