% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCounts.R
\name{writeCounts}
\alias{writeCounts}
\title{Write Counts}
\usage{
writeCounts(..., dir = ".", gzip = TRUE)
}
\arguments{
\item{...}{Count matrices, passed in as dots.}

\item{dir}{Output directory.}

\item{gzip}{Compress the counts file using gzip.}
}
\value{
Invisible named character vector containing file paths to count
matrices written to disk.
}
\description{
Supports both bulk and single-cell RNA-seq output. Bulk RNA-seq counts are
written to disk as comma separated values (\code{.csv}). Single-cell RNA-seq
counts are written to disk in MatrixMarket format (\code{.mtx}), along with the
sample barcodes (\code{.colnames}), and gene identifiers (\code{.rownames}).
}
\details{
Automatic gzip compression is offered as a user-defined option. This setting
is enabled by default to save disk space. Note that the
\href{http://readr.tidyverse.org/}{readr} package, built into
\href{https://www.rstudio.com/}{RStudio}, now natively supports compressed files.
}
\examples{
writeCounts(mtcars, gzip = TRUE)

# Clean up
file_delete("mtcars.csv.gz")
}
\seealso{
Other Write Functions: \code{\link{assignAndSaveData}},
  \code{\link{initializeDirectory}},
  \code{\link{saveData}}, \code{\link{transmit}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
