% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCounts.R
\name{writeCounts}
\alias{writeCounts}
\title{Write Counts}
\usage{
writeCounts(..., dir = ".", gzip = FALSE)
}
\arguments{
\item{...}{Count matrices, passed in as dots.}

\item{dir}{\code{string}. Output directory.}

\item{gzip}{\code{boolean}. Compress the counts file using gzip.}
}
\value{
Invisible \code{list}. File paths.
}
\description{
Supports both bulk and single-cell RNA-seq count matrices. Bulk RNA-seq
counts are written to disk as comma separated values ("\code{.csv}"). Single-cell
RNA-seq counts are written to disk in MatrixMarket format ("\code{.mtx}"), along
with the sample barcodes ("\code{.colnames}"), and gene identifiers
("\code{.rownames}").
}
\details{
Automatic gzip compression is offered as a user-defined option. This setting
is enabled by default to save disk space. Note that the
\href{http://readr.tidyverse.org/}{readr} package, built into
\href{https://www.rstudio.com/}{RStudio}, now natively supports compressed files.
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation.
}
\examples{
rnaseq_counts <- counts(rse_small)
single_cell_counts <- counts(sce_small)

writeCounts(rnaseq_counts, single_cell_counts, dir = "example")
list.files("example")

# Clean up
unlink("example", recursive = TRUE)
}
\seealso{
Other Write Functions: \code{\link{assignAndSaveData}},
  \code{\link{saveData}}, \code{\link{transmit}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Write Functions}
