% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-writeCounts.R
\docType{methods}
\name{writeCounts}
\alias{writeCounts}
\alias{writeCounts,ANY-method}
\title{Write Counts}
\usage{
writeCounts(..., dir = file.path("results", "counts"), gzip = TRUE)

\S4method{writeCounts}{ANY}(..., dir = file.path("results", "counts"),
  gzip = TRUE)
}
\arguments{
\item{...}{Count matrices, passed in as dots.}

\item{dir}{Output directory.}

\item{gzip}{Compress the counts file using gzip.}
}
\value{
No value.
}
\description{
Supports both bulk and single-cell RNA-seq output. Bulk RNA-seq counts are
written to disk as comma separated values (\code{.csv). Single-cell RNA-seq counts are written to disk in MatrixMarket format (}.mtx\code{), along with the sample barcodes (}.colnames\code{), and gene identifiers (}.rownames`).
}
\details{
Automatic gzip compression is offered as a user-defined option. This setting
is enabled by default to save disk space. Note that the
\href{http://readr.tidyverse.org/}{readr} package, built into
\href{https://www.rstudio.com/}{RStudio}, now natively supports compressed files.
}
\seealso{
Other Write Utilities: \code{\link{saveDataRaw}},
  \code{\link{saveData}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
