% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zerosVsDepth-methods.R
\docType{methods}
\name{zerosVsDepth}
\alias{zerosVsDepth}
\alias{zerosVsDepth,matrix-method}
\alias{zerosVsDepth,sparseMatrix-method}
\alias{zerosVsDepth,SummarizedExperiment-method}
\alias{zerosVsDepth,SingleCellExperiment-method}
\title{Percentage of zeros vs. library depth}
\usage{
\S4method{zerosVsDepth}{matrix}(object)

\S4method{zerosVsDepth}{sparseMatrix}(object)

\S4method{zerosVsDepth}{SummarizedExperiment}(object, assay = 1L)

\S4method{zerosVsDepth}{SingleCellExperiment}(object, assay = 1L)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}
}
\value{
\code{matrix}.
}
\description{
Calculate and visualize the dropout rate.
}
\examples{
data(sce, package = "acidtest")
x <- zerosVsDepth(sce)
summary(x)
colnames(x)
}
\author{
Rory Kirchner, Michael Steinbaugh
}
