% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zerosVsDepth-methods.R
\name{zerosVsDepth}
\alias{zerosVsDepth}
\alias{zerosVsDepth,matrix-method}
\alias{zerosVsDepth,Matrix-method}
\alias{zerosVsDepth,SummarizedExperiment-method}
\alias{zerosVsDepth,SingleCellExperiment-method}
\title{Percentage of zeros vs. library depth}
\usage{
zerosVsDepth(object, ...)

\S4method{zerosVsDepth}{matrix}(object)

\S4method{zerosVsDepth}{Matrix}(object)

\S4method{zerosVsDepth}{SummarizedExperiment}(object, assay = 1L)

\S4method{zerosVsDepth}{SingleCellExperiment}(object, assay = 1L)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Calculate and visualize the dropout rate.
}
\note{
Updated 2020-01-30.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")
sce <- SingleCellExperiment

## SingleCellExperiment ====
x <- zerosVsDepth(sce)
summary(x)
colnames(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
